% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_tile_api.R
\name{band_accessors}
\alias{band_accessors}
\alias{.data_type}
\alias{.miss_value}
\alias{.min_value}
\alias{.max_value}
\alias{.scale}
\alias{.offset}
\alias{.cloud_interp_values}
\alias{.cloud_bit_mask}
\alias{.band_cloud}
\title{Band configuration accessors}
\usage{
.data_type(conf)

.miss_value(conf)

.min_value(conf)

.max_value(conf)

.scale(conf)

.offset(conf)

.cloud_interp_values(conf)

.cloud_bit_mask(conf)

.band_cloud()
}
\arguments{
\item{conf}{A band definition value from config. Can be retrieved by
\code{.conf_eo_band()} or \code{.conf_derived_band()}.}
}
\value{
Respective configuration value.
}
\description{
These functions are read-only accessors of \code{band_conf} objects. A
\code{band_conf} is an entry of band definition in config. It can be
from an \code{eo_cube} or \code{derived_cube}.
}
\section{Functions}{
\itemize{
\item \code{.data_type()}: Get \code{data_type} entry.

\item \code{.miss_value()}: Get \code{missing_value} entry.

\item \code{.min_value()}: Get \code{minimum_value} entry.

\item \code{.max_value()}: Get \code{maximum_value} entry.

\item \code{.scale()}: Get \code{scale_factor} entry.

\item \code{.offset()}: Get \code{offset_value} entry.

\item \code{.cloud_interp_values()}: Get \code{interp_values} entry.

\item \code{.cloud_bit_mask()}: Get \code{bit_mask} entry.

\item \code{.band_cloud()}: Return name of the cloud band.

}}
\examples{
\dontrun{
# Get configuration band
x <- .conf_eo_band(
  source = "BDC",
  collection = "MOD13Q1-6",
  band = "NIR"
)

.
}

}
\seealso{
Other accessors: 
\code{\link{bbox_accessors}},
\code{\link{block_accessors}},
\code{\link{chunk_accessors}},
\code{\link{tile_accessors}}
}
\concept{accessors}
\keyword{internal}
