% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_tile_api.R
\name{band_api}
\alias{band_api}
\alias{.band_cloud}
\alias{.band_eo}
\alias{.band_derived}
\title{Band API}
\usage{
.band_cloud()

.band_eo(band)

.band_derived(band)
}
\arguments{
\item{band}{Band name.}
}
\description{
We use term \code{band} to refer either to spectral bands as index generated
from images of actual sensor bands. To organize internal metadata and work
properly there are some restrictions in band names. These functions aims
to impose band name restrictions.
}
\section{Functions}{
\itemize{
\item \code{.band_cloud()}: Returns the name of cloud band.

\item \code{.band_eo()}: Returns a well formatted band name for \code{eo_cube}.

\item \code{.band_derived()}: Returns a well formatted band name for
\code{derived_cube}.

}}
\examples{
\dontrun{
.band_cloud() # 'CLOUD'
# eo bands name are uppercase
.band_eo("nDvI") # 'NDVI'
# derived bands name are lowercase
.band_derived("PrObS") # 'probs'
# bands name cannot have '_' (underscore)
.band_eo("NDVI_2") # 'NDVI-2'
}

}
\seealso{
\link{band_accessors}
}
\keyword{internal}
