% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_tile_api.R
\name{bbox_accessors}
\alias{bbox_accessors}
\alias{.xmin}
\alias{.xmin<-}
\alias{.xmax}
\alias{.xmax<-}
\alias{.ymin}
\alias{.ymin<-}
\alias{.ymax}
\alias{.ymax<-}
\alias{.as_crs}
\alias{.crs}
\alias{.crs<-}
\title{bbox accessors}
\usage{
.xmin(x)

.xmin(x) <- value

.xmax(x)

.xmax(x) <- value

.ymin(x)

.ymin(x) <- value

.ymax(x)

.ymax(x) <- value

.as_crs(x)

.crs(x)

.crs(x) <- value
}
\arguments{
\item{x}{Object to get/set field value.}

\item{value}{Value to set on object field.}
}
\value{
\code{numeric}.

Updated object \code{x}.

\code{numeric}.

Updated object \code{x}.

\code{numeric}.

Updated object \code{x}.

\code{numeric}.

Updated object \code{x}.

\code{character}.

\code{character}.

Updated object \code{x}.
}
\description{
These functions are accessors of bbox fields inside \code{vectors}.
Getters functions returns the respective field values with the expected
data type. Setters functions convert value to expected data type and
store it in respective fields on a given object. If value has no length,
and the field is not \code{atomic} it is removed from the object.
}
\section{Functions}{
\itemize{
\item \code{.xmin()}: Get \code{'xmin'} field.

\item \code{.xmin(x) <- value}: Set \code{'xmin'} field.

\item \code{.xmax()}: Get \code{'xmax'} field.

\item \code{.xmax(x) <- value}: Set \code{'xmax'} field.

\item \code{.ymin()}: Get \code{'ymin'} field.

\item \code{.ymin(x) <- value}: Set \code{'ymin'} field.

\item \code{.ymax()}: Get \code{'ymax'} field.

\item \code{.ymax(x) <- value}: Set \code{'ymax'} field.

\item \code{.as_crs()}: Convert a CRS value to \code{character}.

\item \code{.crs()}: Get \code{'crs'} field.

\item \code{.crs(x) <- value}: Set \code{'crs'} field.

}}
\examples{
\dontrun{

}
}
\keyword{internal}
