% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_tile_api.R
\name{bbox_api}
\alias{bbox_api}
\alias{.bbox}
\alias{.bbox_as_sf}
\title{Bbox API}
\usage{
.bbox(x, ..., default_crs = NULL)

.bbox_as_sf(bbox, ..., default_crs = NULL, as_crs = NULL)
}
\arguments{
\item{x}{Any object to extract a \code{bbox}.}

\item{...}{Additional parameters.}

\item{default_crs}{If no CRS is present in \code{x}, which CRS should be
used? If \code{NULL} default CRS will be \code{'EPSG:4326'}.}

\item{bbox}{A \code{bbox}.}

\item{as_crs}{A CRS to project \code{bbox}. Useful if bbox has multiples
CRS.}
}
\value{
\code{.bbox()}: \code{bbox}.

\code{.bbox_as_sf()}: \code{sf}.
}
\description{
A bounding box represents a rectangular geographical region in a certain
projection. A \code{bbox} is any \code{list} or \code{tibble} containing
\code{xmin}, \code{xmax}, \code{ymin}, \code{ymax}, and \code{crs} fields.
A \code{bbox} may contains multiple entries.
}
\section{Functions}{
\itemize{
\item \code{.bbox()}: extract a \code{bbox} from any given
\code{vector}.

\item \code{.bbox_as_sf()}: Convert a \code{bbox} into a
\code{sf} polygon object.

}}
\examples{
\dontrun{
x <- list(a = 0, z = 0)
.bbox(x) # NULL
x <- list(
  a = 0, xmin = 1:3, xmax = 2:4, ymin = 3:5, ymax = 4:6,
  crs = 4326, z = 0
)
.bbox(x)
.bbox_as_sf(x) # 3 features
.bbox_as_sf(x, as_crs = "EPSG:3857")
}

}
\seealso{
\link{bbox_accessors}

Other region objects API: 
\code{\link{.intersects}()},
\code{\link{block_api}},
\code{\link{chunks_api}},
\code{\link{roi_api}}
}
\concept{region objects API}
\keyword{internal}
