% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_bbox.R
\name{bbox_functions}
\alias{bbox_functions}
\alias{.sits_bbox_intersection}
\title{Bounding box (bbox)}
\usage{
.sits_bbox_intersection(bbox, cube)
}
\arguments{
\item{bbox}{a `bbox` object (inherited from `tbl`) containing
`("xmin", "xmax", "ymin", "ymax", "crs")` variables}

\item{cube}{a sits data cube object (inherited from `tbl`)}
}
\value{
`.sits_bbox_intersection():` a `bbox` object (inherited from `tbl`)
}
\description{
Bounding box functions

`.sits_bbox_intersection()`: compute intersection between a bbox and a
`cube` object (inherited from `tbl`).
}
\keyword{internal}
