% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_tile_api.R
\name{block_accessors}
\alias{block_accessors}
\alias{.col}
\alias{.col<-}
\alias{.row}
\alias{.row<-}
\alias{.ncols}
\alias{.ncols<-}
\alias{.nrows}
\alias{.nrows<-}
\title{Block accessors}
\usage{
.col(x)

.col(x) <- value

.row(x)

.row(x) <- value

.ncols(x)

.ncols(x) <- value

.nrows(x)

.nrows(x) <- value
}
\arguments{
\item{x}{Object to get field value.}

\item{value}{Value to set on object field.}
}
\value{
Getters return respective field value or \code{NULL}, if it doesn't
  exist. Setters return the updated \code{x} object.
}
\description{
These functions are accessors of \code{block} fields in a \code{vector}.
Getters functions returns the respective field values with the expected
data type. Setters functions convert value to expected data type and
store it in respective fields on a given object. If value has no length
and the \code{vector} is not \code{atomic}, it is removed from the object.
}
\details{
\code{.col()}, \code{.row()}, \code{.ncols()}, and \code{.nrows()} get/set,
respectively, \code{"col"}, \code{"row"}, \code{"ncols"}, and
\code{"nrows"} fields.
}
\section{Functions}{
\itemize{
\item \code{.col()}: Get \code{'col'} field.

\item \code{.col(x) <- value}: Set \code{'col'} field as integer.

\item \code{.row()}: Get \code{'row'} field.

\item \code{.row(x) <- value}: Set \code{'row'} field as integer.

\item \code{.ncols()}: Get \code{'ncols'} field.

\item \code{.ncols(x) <- value}: Set \code{'ncols'} field as integer.

\item \code{.nrows()}: Get \code{'nrows'} field.

\item \code{.nrows(x) <- value}: Set \code{'nrows'} field as integer.

}}
\examples{
\dontrun{
x <- c(row = 3.45)
.row(x) # 3 as integer
x <- list(col = 1, row = 2, ncols = 3)
.nrows(x) # NULL
.nrows(x) <- 4
x
}

}
\seealso{
Other accessors: 
\code{\link{band_accessors}},
\code{\link{bbox_accessors}},
\code{\link{chunk_accessors}},
\code{\link{tile_accessors}}
}
\concept{accessors}
\keyword{internal}
