% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cerrado_13classes_modis_col5}
\alias{cerrado_13classes_modis_col5}
\title{Samples of 13 classes for the Cerrado biome in Brazil}
\format{A tibble with 11,744 rows and 7 variables: (a) longitude: East-west coordinate of the time series sample (WGS 84);
  latitude (North-south coordinate of the time series sample in WGS 84), start_date (initial date of the time series),
  end_date (final date of the time series), label (the class label associated to the sample),
  coverage (the name of the coverage associated with the data),
  time_series ( list containing a tibble with the values of the time series).}
\usage{
data(cerrado_13classes_modis_col5)
}
\description{
A dataset containing a tibble with time series samples for the Cerrado biome in Brazil.
  The time series comes from MOD13Q1 collection 6 images with six bands
  (ndvi, evi, blue, red, nir, mir)
}
\keyword{datasets}
