% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cerrado_2classes}
\alias{cerrado_2classes}
\title{Samples of classes Cerrado and Pasture}
\format{A tibble with 736 rows and 7 variables: (a) longitude: East-west coordinate of the time series sample (WGS 84);
  latitude (North-south coordinate of the time series sample in WGS 84), start_date (initial date of the time series),
  end_date (final date of the time series), label (the class label associated to the sample),
  coverage (the name of the coverage associated with the data),
  time_series (list containing a tibble with the values of the time series).}
\usage{
data(cerrado_2classes)
}
\description{
A dataset containing a tibble with time series samples for the Cerrado and Pasture areas of the Mato Grosso state.
  The time series come from MOD13Q1 collection 5 images.
}
