% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_compare.R
\name{comparison_functions}
\alias{comparison_functions}
\alias{.is_eq}
\alias{.is_lt}
\alias{.is_gt}
\alias{.is_int}
\title{Comparison functions}
\usage{
.is_eq(x, y, tolerance = 0)

.is_lt(x, y, tolerance = 0)

.is_gt(x, y, tolerance = 0)

.is_int(x, tolerance = 0)
}
\arguments{
\item{x, y}{a \code{numeric} value to compare.}

\item{tolerance}{a unique positive \code{numeric} value. Default is 0.}
}
\value{
a \code{logical} value.
}
\description{
Functions to compare two numeric vectors using tolerance
 parameter.
}
\keyword{internal}
