% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_tile_api.R
\name{config_api}
\alias{config_api}
\alias{.conf_exists}
\alias{.conf}
\title{Basic access config functions}
\usage{
.conf_exists(..., throw_error = FALSE)

.conf(...)
}
\arguments{
\item{...}{Set of \code{character} values representing a key to access
some hierarchical config entry.}

\item{throw_error}{Should an error be thrown if test fails?}
}
\value{
Configuration value.
}
\description{
These are basic functions to access config options.
}
\section{Functions}{
\itemize{
\item \code{.conf_exists()}: Tests if a key provided as \code{character} values
in \code{...} parameter exists in the config. If \code{throws_error} is
\code{TRUE} and the test failed, an error is raised.

\item \code{.conf()}: Get a config value located in a key provided as
\code{character} values in \code{...} parameter. If a key does not
exists, throws an error. Use \code{.conf_exists()} to test for a key
existence.

}}
\examples{
if (sits_run_examples()) {
.conf_exists("run_tests") # TRUE
.conf("run_tests")
.conf_exists("not_existing_entry") # FALSE
}

}
\seealso{
Other config functions: 
\code{\link{derived_cube_config}},
\code{\link{eo_cube_config}}
}
\concept{config functions}
\keyword{internal}
