% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_config.R
\name{config_colors_file}
\alias{config_colors_file}
\alias{.config_colors_file}
\title{Return the default configuration file for colors}
\usage{
.config_colors_file()
}
\value{
default color configuration file
}
\description{
Return the default configuration file for colors
}
\keyword{internal}
