% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_config.R
\name{config_file}
\alias{config_file}
\alias{.config_file}
\title{Return the default configuration file}
\usage{
.config_file()
}
\value{
default configuration file
}
\description{
Return the default configuration file
}
\keyword{internal}
