% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_config.R
\name{config_functions}
\alias{config_functions}
\alias{.config_file}
\alias{.config_user_file}
\alias{.config_get}
\alias{.config_names}
\alias{.config_gtiff_default_options}
\alias{.config_local_file_extensions}
\alias{.config_local_s3_class}
\alias{.config_palettes}
\alias{.config_palette_colors}
\alias{.config_processing_bloat}
\alias{.config_rstac_limit}
\alias{.config_raster_pkg}
\title{Get values from config file}
\usage{
.config_file()

.config_user_file()

.config_get(key, default = NULL)

.config_names(key)

.config_gtiff_default_options()

.config_local_file_extensions()

.config_local_s3_class()

.config_palettes()

.config_palette_colors(labels, ..., palette = "default")

.config_processing_bloat()

.config_rstac_limit()

.config_raster_pkg()
}
\arguments{
\item{labels}{Vector with labels.}

\item{...}{Additional parameters.}

\item{collection}{Collection to be searched in the data source.}

\item{data}{A sits data cube.}

\item{pallete}{The palette that should be chosen based on the
configuration file.}

\item{simplify}{A logical value that specifies whether the return should be
in vector form, if true, or list form, if false. Default value is FALSE.}

\item{source}{Source of data cube}
}
\value{
Functions that search for values from a key or collection
 return atomic values. Check functions return invisible null values or give
 an error.
}
\description{
Functions that get values from config file.
}
\keyword{internal}
