% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_config.R
\name{config_internals_file}
\alias{config_internals_file}
\alias{.config_internals_file}
\title{Return the internal configuration file (only for developers)}
\usage{
.config_internals_file()
}
\value{
default internal configuration file
}
\description{
Return the internal configuration file (only for developers)
}
\keyword{internal}
