% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_conf.R
\name{config_show}
\alias{config_show}
\alias{sits_config_show}
\title{Shows the contents of the sits configuration file}
\usage{
sits_config_show()
}
\value{
List with the configuration parameters used by sits.
}
\description{
Displays the contents of sits configuration file. For details
on how to set the configuration file, use \code{\link[sits]{sits_config}}.
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
