% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_transitions.R
\name{Convert}
\alias{Convert}
\title{Convert transition function}
\usage{
Convert(source_class, target_class)
}
\arguments{
\item{source_class}{Source class name for the transition}

\item{target_class}{Target class name for the transition}
}
\value{
expression that creates a CONVERT transition call
}
\description{
Convert transition function
}
\author{
Felipe Carlos, \email{efelipecarlos@gmail.com}

Felipe Carvalho, \email{felipe.carvalho@inpe.br}

Gilberto Camara, \email{gilberto.camara@inpe.br}
}
\keyword{internal}
