% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_cube_aux_functions.R
\name{.cube_bands}
\alias{.cube_bands}
\alias{.cube_bands_check}
\alias{cube_functions}
\alias{.cube_band_missing_value}
\alias{.cube_band_minimum_value}
\alias{.cube_band_maximum_value}
\alias{.cube_band_scale_factor}
\alias{.cube_band_offset_value}
\alias{.cube_check}
\alias{.cube_collection}
\alias{.cube_has_cloud}
\alias{.cube_intersects}
\alias{.cube_is_regular}
\alias{.cube_labels}
\alias{.cube_resolution}
\alias{.cube_s3class}
\alias{.cube_size}
\alias{.cube_timeline}
\alias{.cube_start_date}
\alias{.cube_end_date}
\alias{.cube_tiles}
\alias{.cube_tile_check}
\alias{.cube_tile_get_fields}
\alias{.cube_tile_bbox}
\title{Auxiliary functions for sits cubes}
\usage{
.cube_bands(cube, ..., add_cloud = TRUE)

.cube_bands_check(cube, bands, ..., add_cloud = TRUE)

.cube_band_missing_value(cube, band)

.cube_band_minimum_value(cube, band)

.cube_band_maximum_value(cube, band)

.cube_band_scale_factor(cube, band)

.cube_band_offset_value(cube, band)

.cube_check(cube)

.cube_collection(cube)

.cube_has_cloud(cube)

.cube_intersects(cube, roi = NULL)

.cube_is_regular(cube)

.cube_labels(cube)

.cube_resolution(cube, ..., bands = NULL)

.cube_s3class(cube)

.cube_size(cube, ..., bands = NULL)

.cube_timeline(cube)

.cube_start_date(cube)

.cube_end_date(cube)

.cube_tiles(cube)

.cube_tile_check(cube, tile)

.cube_tile_get_fields(cube, tile, fields = NULL)

.cube_tile_bbox(cube, ..., tile = 1)
}
\arguments{
\item{cube}{Data cube from where data is to be retrieved.}

\item{bands}{Bands to be included.}

\item{roi}{a vector containing either `("xmin", "xmax", "ymin", "ymax")`
or `("lon_min", "lat_min", "lon_max", "lat_max")` variables, representing
a spatial region of interest}

\item{tile}{Tiles from the collection to be included in the data cube}

\item{....}{Additional parameters to be included.}
}
\value{
a \code{vector} for get attributes functions and NULL or error for
check parameters functions.

for `.cube_intersects()`: a logical vector
}
\description{
Auxiliary functions for extracting information from sits cubes.

`.cube_intersects()`: inform which tiles of a given cube intersects
a region of interest.

verify is the cube is a regular one
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
\keyword{internal}
