% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_WTSPS.R
\docType{methods}
\name{describeAlgorithm}
\alias{describeAlgorithm}
\alias{describeAlgorithm,WTSPS-method}
\title{Returns a description of a queried algorithm contained in WTSPS}
\usage{
describeAlgorithm(object, algorithm)

\S4method{describeAlgorithm}{WTSPS}(object, algorithm)
}
\arguments{
\item{object}{a WTSPS object}

\item{algorithm}{a specific algorithm}
}
\description{
Returns a description of a queried algorithm contained in WTSPS
}
