% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_transitions.R
\name{Dominates}
\alias{Dominates}
\title{Dominates transition function}
\usage{
Dominates(target_class, size)
}
\arguments{
\item{target_class}{Target class name for the transition}

\item{size}{Neighborhood size}
}
\value{
expression that creates a DOMINATES transition call
}
\description{
Dominates transition function
}
\author{
Felipe Carlos, \email{efelipecarlos@gmail.com}

Felipe Carvalho, \email{felipe.carvalho@inpe.br}

Gilberto Camara, \email{gilberto.camara@inpe.br}
}
\keyword{internal}
