% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_cube.R
\name{...}
\alias{...}
\alias{sits_cube.gdalcubes_cube}
\title{...}
\usage{
\method{sits_cube}{gdalcubes_cube}(
  type = "GDALCUBES",
  ...,
  cube,
  path_db,
  path_images,
  period = NULL,
  method = NULL,
  resampling = "bilinear",
  cloud_mask = TRUE
)
}
\arguments{
\item{type}{...}

\item{cube}{...}

\item{path_db}{...}

\item{period}{...}

\item{method}{...}

\item{resampling}{...}

\item{cloud_mask}{...}
}
\value{
cube ...
}
\description{
...
}
\details{
TODO: documentar
TODO: explicitar a maneira em que o 'period' deve ser fornecido.
}
