% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_check.R
\name{..check_pred_ref_match}
\alias{..check_pred_ref_match}
\alias{.check_pred_ref_match}
\title{Check if data contains predicted and reference values}
\usage{
.check_pred_ref_match(reference, predicted)
}
\arguments{
\item{reference}{vector with reference labels}

\item{predicted}{vector with predicted labels}
}
\value{
No return value, called for side effects.
}
\description{
Check if data contains predicted and reference values
}
\keyword{internal}
