% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_apply.R
\name{.apply_across}
\alias{.apply_across}
\title{Apply a function to a set of time series}
\usage{
.apply_across(data, fn, ...)

.apply_across(data, fn, ...)
}
\description{
Apply a function to a set of time series

Apply a function to a set of time series
}
\keyword{internal}
