% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_apply.R
\name{.apply_capture_expression}
\alias{.apply_capture_expression}
\title{Captures a band expression}
\usage{
.apply_capture_expression(...)
}
\arguments{
\item{tile_name}{Tile name.}
}
\value{
Named list with one expression
}
\description{
Captures a band expression
}
\keyword{internal}
