% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_apply.R
\name{.apply_get_all_names}
\alias{.apply_get_all_names}
\title{Returns all names in an expression}
\usage{
.apply_get_all_names(expr)
}
\arguments{
\item{expr}{Expression.}
}
\value{
Character vector with all names in expression.
}
\description{
Returns all names in an expression
}
\keyword{internal}
