% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_apply.R
\name{.apply_input_bands}
\alias{.apply_input_bands}
\title{Finds out all existing bands in an expression}
\usage{
.apply_input_bands(cube, expr)
}
\arguments{
\item{expr}{Band expression.}

\item{tile}{Data cube tile.}
}
\value{
List of combination among tiles, bands, and dates
                  that are missing from the cube.
}
\description{
Finds out all existing bands in an expression
}
\keyword{internal}
