% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_apply.R
\name{.apply_input_kernels}
\alias{.apply_input_kernels}
\alias{.apply_get_kernel_window_size}
\title{Returns all kernel functions in an expression}
\usage{
.apply_get_kernel_window_size(expr)
}
\arguments{
\item{expr}{Expression.}
}
\value{
List with all used kernel functions.
}
\description{
Returns all kernel functions in an expression
}
\keyword{internal}
