% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_mixture_model.R
\name{.apply_missing_band}
\alias{.apply_missing_band}
\alias{.mesma_get_jobs_lst}
\title{Create a list of jobs}
\usage{
.mesma_get_jobs_lst(cube, fractions)
}
\arguments{
\item{cube}{Data cube.}

\item{fractions}{Name of output fractions.}
}
\value{
List of combination among tiles, endmembers, and dates
                  that are missing from the cube.
}
\description{
Create a list of jobs
}
\keyword{internal}
