% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_apply.R
\name{.apply_out_file_name}
\alias{.apply_out_file_name}
\title{Returns a new file name}
\usage{
.apply_out_file_name(tile_name, band, date, output_dir)
}
\arguments{
\item{tile_name}{Tile name.}

\item{band}{Band name.}

\item{date}{Observation date.}

\item{output_dir}{Base directory.}
}
\value{
File path.
}
\description{
Returns a new file name
}
\keyword{internal}
