% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_source_api_aws.R
\name{.aws_bands_band_name}
\alias{.aws_bands_band_name}
\title{Get aws source bands by resolution}
\usage{
.aws_bands_band_name(source, collection, s2_resolution)
}
\arguments{
\item{source}{Data source (one of "SATVEG", "LOCAL", "BDC", "AWS",
"USGS", "DEAFRICA", "PROBS").}

\item{collection}{Collection to be searched in the data source.}

\item{s2_resolution}{sentinel band resolution}
}
\value{
sentinel bands that corresponds a specified resolution.
}
\description{
retrieve aws bands by resolution
}
\keyword{internal}
