% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_check.R
\name{.aws_check_environment}
\alias{.aws_check_environment}
\alias{.check_aws_environment}
\title{Check access rights on AWS}
\usage{
.check_aws_environment(source, collection)
}
\arguments{
\item{source}{data source}

\item{collection}{...}
}
\value{
TRUE if access is granted
}
\description{
Check access rights on AWS
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
\keyword{internal}
