% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baggio_utils.R
\name{.bg_raster_preprocess}
\alias{.bg_raster_preprocess}
\title{TODO: document}
\usage{
.bg_raster_preprocess(
  cube,
  file,
  file_cld,
  block,
  missing_value,
  minimum_value,
  maximum_value,
  scale_factor,
  offset_value,
  ...,
  impute_fn = sits_impute_linear(),
  filter_fn = NULL,
  normalize_fn = NULL
)
}
\description{
TODO: document
}
