% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_check.R
\name{.check_band_in_cube}
\alias{.check_band_in_cube}
\title{Check if band is present in the cube}
\usage{
.check_band_in_cube(band, cube)
}
\arguments{
\item{band}{name of the band}

\item{cube}{data cube}
}
\value{
No return value, called for side effects.
}
\description{
Check if band is present in the cube
}
\keyword{internal}
