% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_check.R
\name{.check_bands_in_cube}
\alias{.check_bands_in_cube}
\title{Check bands are present in the cube}
\usage{
.check_bands_in_cube(bands, cube)
}
\arguments{
\item{bands}{names of the bands}

\item{cube}{data cube}
}
\value{
No return value, called for side effects.
}
\description{
Check bands are present in the cube
}
\keyword{internal}
