% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_check.R
\name{.check_chr_parameter}
\alias{.check_chr_parameter}
\title{Check is integer parameter is valid using reasonable defaults}
\usage{
.check_chr_parameter(param, len_min = 1, len_max = 1)
}
\arguments{
\item{len_min}{minimum length of vector}

\item{len_max}{maximum length of vector}

\item{x}{parameter to be checked}
}
\value{
No return value, called for side effects.
}
\description{
Check is integer parameter is valid using reasonable defaults
}
\keyword{internal}
