% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_cube.R
\name{.check_collection}
\alias{.check_collection}
\alias{.check_bands}
\title{Check cube collection}
\usage{
.check_bands(source, collection, bands, s2_resolution = NULL)
}
\arguments{
\item{source}{Data source}

\item{collection}{Collection to be searched in the data source.}

\item{bands}{Bands to be included.}

\item{s2_resolution}{Resolution of S2 images ("10m", "20m" or "60m").
used to build cubes (only for AWS cubes).}
}
\value{
An invisible null
}
\description{
A suite of check to verify collection in cube.
}
