% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_check.R
\name{.check_csv}
\alias{.check_csv}
\title{Check if CSV data is correct}
\usage{
.check_csv(csv)
}
\arguments{
\item{csv}{data frame extracted from CSV file}
}
\value{
No return value, called for side effects.
}
\description{
Check if CSV data is correct
}
\keyword{internal}
