% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_check.R
\name{.check_cube_bands}
\alias{.check_cube_bands}
\title{Check if bands are part of a data cube}
\usage{
.check_cube_bands(cube, bands, add_cloud = TRUE)
}
\arguments{
\item{cube}{Data cube}

\item{bands}{Bands to be check}

\item{add_cloud}{Include the cloud band?}
}
\value{
No return value, called for side effects.
}
\description{
Check if bands are part of a data cube
}
\keyword{internal}
