% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_check.R
\name{.check_cube_has_one_tile}
\alias{.check_cube_has_one_tile}
\title{Check if cube has only one tile}
\usage{
.check_cube_has_one_tile(cube)
}
\arguments{
\item{cube}{input data cube}
}
\description{
Check if cube has only one tile
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
\keyword{internal}
