% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_check.R
\name{.check_cube_has_unique_tile_size}
\alias{.check_cube_has_unique_tile_size}
\title{Check if sizes of all tiles of the cube are the same}
\usage{
.check_cube_has_unique_tile_size(cube)
}
\arguments{
\item{cube}{input data cube}
}
\value{
TRUE/FALSE
}
\description{
Check if sizes of all tiles of the cube are the same
}
\keyword{internal}
