% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_check.R
\name{.check_cube_is_classified_image}
\alias{.check_cube_is_classified_image}
\title{Check if cube is a classified image}
\usage{
.check_cube_is_classified_image(cube)
}
\arguments{
\item{cube}{a sits cube to be tested}
}
\value{
No return value, called for side effects
}
\description{
Check if cube is a classified image
}
\keyword{internal}
