% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_check.R
\name{.check_cube_is_probs_cube}
\alias{.check_cube_is_probs_cube}
\title{Check if cube is a probs cube}
\usage{
.check_cube_is_probs_cube(cube)
}
\arguments{
\item{cube}{input data cube}
}
\description{
Check if cube is a probs cube
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
\keyword{internal}
