% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_check.R
\name{.check_cube_is_regular}
\alias{.check_cube_is_regular}
\alias{.cube_is_regular.default}
\title{Check that cube is regular}
\usage{
.check_cube_is_regular(cube)

\method{.cube_is_regular}{default}(cube)
}
\arguments{
\item{cube}{datacube}
}
\value{
TRUE/FALSE
}
\description{
Check that cube is regular
}
\keyword{internal}
