% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_check.R
\name{.check_cube_list_match}
\alias{.check_cube_list_match}
\title{Check if cubes have the same organization}
\usage{
.check_cube_list_match(cubes)
}
\arguments{
\item{cubes}{list of input data cubes}
}
\value{
No return value, called for side effects.
}
\description{
Check if cubes have the same organization
}
\keyword{internal}
