% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_check.R
\name{.check_cube_list_probs_match}
\alias{.check_cube_list_probs_match}
\alias{.check_cube_list_match}
\title{Check if uncert cubes have the same organization}
\usage{
.check_cube_list_match(cubes)
}
\arguments{
\item{cubes}{list of input data cubes}
}
\value{
TRUE/FALSE
}
\description{
Check if uncert cubes have the same organization
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
\keyword{internal}
