% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_check.R
\name{.check_cube_model}
\alias{.check_cube_model}
\title{Check classification parameters}
\usage{
.check_cube_model(cube, ml_model)
}
\arguments{
\item{cube}{Tibble with information about a data cube.}

\item{ml_model}{An R model trained by \code{\link[sits]{sits_train}}.}
}
\value{
No value called for side effects
}
\description{
Verify that required parameters are correct.
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
\keyword{internal}
