% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_check.R
\name{.check_cubes_match}
\alias{.check_cubes_match}
\title{Check if two cubes have the same organization}
\usage{
.check_cubes_match(cube1, cube2)
}
\arguments{
\item{cube1}{input data cube}

\item{cube2}{input data cube}
}
\value{
No return value, called for side effects.
}
\description{
Check if two cubes have the same organization
}
\keyword{internal}
