% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_check.R
\name{.check_cubes_same_size}
\alias{.check_cubes_same_size}
\title{Check if cubes have the same size}
\usage{
.check_cubes_same_size(cube1, cube2)
}
\arguments{
\item{cubes}{list input data cubes}
}
\value{
TRUE/FALSE
}
\description{
Check if cubes have the same size
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
\keyword{internal}
