% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_check.R
\name{.check_cubes_same_timeline}
\alias{.check_cubes_same_timeline}
\title{Check if cubes have the same timeline}
\usage{
.check_cubes_same_timeline(cube1, cube2)
}
\arguments{
\item{cubes}{list of input data cubes}
}
\value{
TRUE/FALSE
}
\description{
Check if cubes have the same timeline
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
\keyword{internal}
