% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_check.R
\name{.check_date_parameter}
\alias{.check_date_parameter}
\title{Check is date is valid}
\usage{
.check_date_parameter(
  x,
  len_min = 1L,
  len_max = 1L,
  allow_null = FALSE,
  msg = NULL
)
}
\arguments{
\item{x}{parameter to be checked}

\item{len_min}{minimum length of vector}

\item{len_max}{maximum length of vector}

\item{allow_null}{allow NULL?}

\item{msg}{Error message}
}
\value{
Called for side effects.
}
\description{
Check is date is valid
}
\keyword{internal}
