% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_check.R
\name{.check_dates_parameter}
\alias{.check_dates_parameter}
\title{Check is dates parameter are valid}
\usage{
.check_dates_parameter(
  param,
  len_min = 1,
  len_max = 2^31 - 1,
  allow_null = TRUE
)
}
\arguments{
\item{param}{parameter to be checked}

\item{len_min}{minimum length of vector}

\item{len_max}{maximum length of vector}

\item{allow_null}{allow NULL?}
}
\value{
No return value, called for side effects.
}
\description{
Check is dates parameter are valid
}
\keyword{internal}
