% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_check.R
\name{.check_distances_pred}
\alias{.check_distances_pred}
\title{Are the prediction distances valid?}
\usage{
.check_distances_pred(distances, samples)
}
\arguments{
\item{distances}{a data.table with distances values}

\item{samples}{samples from where the distances have been calculated}
}
\value{
No return value, called for side effects.
}
\description{
Are the prediction distances valid?
}
\keyword{internal}
