% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_check.R
\name{.check_documentation}
\alias{.check_documentation}
\title{Checks if working in documentation mode}
\usage{
.check_documentation(progress)
}
\arguments{
\item{progress}{flag set to show progress bar}
}
\value{
TRUE/FALSE
}
\description{
Checks if working in documentation mode
}
\keyword{internal}
