% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_check.R
\name{.check_empty_char}
\alias{.check_empty_char}
\title{Checks if the character parameter is empty}
\usage{
.check_empty_char(x, msg, ...)
}
\arguments{
\item{x}{a character vector}
}
\value{
No return value, called for side effects.
}
\description{
Checks if the character parameter is empty
}
\keyword{internal}
