% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_check.R
\name{.check_endmembers_tbl}
\alias{.check_endmembers_tbl}
\title{Checks if the endmembers data is in a valid parameter}
\usage{
.check_endmembers_tbl(em)
}
\arguments{
\item{em}{Reference spectra endmembers.}

\item{cube}{A sits cube}
}
\value{
No return value, called for side effects.
}
\description{
Checks if the endmembers data is in a valid parameter
}
\keyword{internal}
