% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_check.R
\name{.check_expression}
\alias{.check_expression}
\title{Check is expression parameter is valid using reasonable defaults}
\usage{
.check_expression(list_expr)
}
\arguments{
\item{list_expr}{expression parameter}
}
\value{
No return value, called for side effects.
}
\description{
Check is expression parameter is valid using reasonable defaults
}
\keyword{internal}
