% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_check.R
\name{.check_file_csv}
\alias{.check_file_csv}
\title{Check if file is a CSV}
\usage{
.check_file_csv(file)
}
\arguments{
\item{file}{filename}
}
\value{
No return value, called for side effects.
}
\description{
Check if file is a CSV
}
\keyword{internal}
