% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_check.R
\name{.check_has_one_tile}
\alias{.check_has_one_tile}
\title{Check if cube has only one tile}
\usage{
.check_has_one_tile(cube)
}
\arguments{
\item{cube}{input data cube}
}
\value{
No return value, called for side effects.
}
\description{
Check if cube has only one tile
}
\keyword{internal}
