% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_check.R
\name{.check_has_unique_bands}
\alias{.check_has_unique_bands}
\title{Check if the bands of all tiles of the cube are the same}
\usage{
.check_has_unique_bands(cube)
}
\arguments{
\item{cube}{input data cube}
}
\value{
TRUE/FALSE
}
\description{
Check if the bands of all tiles of the cube are the same
}
\keyword{internal}
