% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_check.R
\name{.check_has_unique_timeline}
\alias{.check_has_unique_timeline}
\title{Check if timelines all tiles of the cube are the same}
\usage{
.check_has_unique_timeline(cube)
}
\arguments{
\item{cube}{input data cube}
}
\value{
TRUE/FALSE
}
\description{
Check if timelines all tiles of the cube are the same
}
\keyword{internal}
