% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_check.R
\name{.check_inherits}
\alias{.check_inherits}
\title{Checks if the object inherits classes}
\usage{
.check_inherits(x, inherits, msg = "invalid inheritance")
}
\arguments{
\item{x}{input object}

\item{inherits}{a character vector with the classes}
}
\value{
No return value, called for side effects.
}
\description{
Checks if the object inherits classes
}
\keyword{internal}
