% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_check.R
\name{.check_is_probs_cube}
\alias{.check_is_probs_cube}
\title{Check if cube is a probs cube}
\usage{
.check_is_probs_cube(cube)
}
\arguments{
\item{cube}{input data cube}
}
\value{
No return value, called for side effects.
}
\description{
Check if cube is a probs cube
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
\keyword{internal}
