% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_check.R
\name{.check_is_raster_cube}
\alias{.check_is_raster_cube}
\title{Does the input data contain a raster cube?}
\usage{
.check_is_raster_cube(data)
}
\arguments{
\item{data}{a sits cube to be tested}
}
\value{
No return value, called for side effects.
}
\description{
Does the input data contain a raster cube?
}
\keyword{internal}
