% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_check.R
\name{.check_is_regular}
\alias{.check_is_regular}
\title{Check that cube is regular}
\usage{
.check_is_regular(cube)
}
\arguments{
\item{cube}{datacube}
}
\value{
TRUE/FALSE
}
\description{
Check that cube is regular
}
\keyword{internal}
