% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_check.R
\name{.check_is_sits_accuracy}
\alias{.check_is_sits_accuracy}
\title{Does the input data contain a sits accuracy object?}
\usage{
.check_is_sits_accuracy(data)
}
\arguments{
\item{data}{a sits accuracy object}
}
\value{
No return value, called for side effects.
}
\description{
Does the input data contain a sits accuracy object?
}
\keyword{internal}
