% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_check.R
\name{.check_is_sits_cube}
\alias{.check_is_sits_cube}
\title{Does the input data contain a sits cube?}
\usage{
.check_is_sits_cube(data)
}
\arguments{
\item{data}{a sits cube}
}
\value{
No return value, called for side effects.
}
\description{
Does the input data contain a sits cube?
}
\keyword{internal}
