% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_check.R
\name{.check_is_sits_model}
\alias{.check_is_sits_model}
\title{Does the input data contain a sits model?}
\usage{
.check_is_sits_model(model)
}
\arguments{
\item{model}{a sits model}
}
\value{
No return value, called for side effects.
}
\description{
Does the input data contain a sits model?
}
\keyword{internal}
