% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_check.R
\name{.check_is_sits_tibble}
\alias{.check_is_sits_tibble}
\title{Does the input data contain a sits tibble?}
\usage{
.check_is_sits_tibble(data)
}
\arguments{
\item{data}{a sits tibble}
}
\value{
No return value, called for side effects.
}
\description{
Does the input data contain a sits tibble?
}
\keyword{internal}
