% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_check.R
\name{.check_labels}
\alias{.check_labels}
\title{Does the input data contains valid reference labels?}
\usage{
.check_labels(data)
}
\arguments{
\item{data}{vector with labels}
}
\value{
No return value, called for side effects.
}
\description{
Does the input data contains valid reference labels?
}
\keyword{internal}
