% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_check.R
\name{.check_memsize}
\alias{.check_memsize}
\title{Check is memsize parameter is valid using reasonable defaults}
\usage{
.check_memsize(memsize)
}
\arguments{
\item{memsize}{size of memory in GB}
}
\value{
No return value, called for side effects.
}
\description{
Check is memsize parameter is valid using reasonable defaults
}
\keyword{internal}
