% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_check.R
\name{.check_multicores}
\alias{.check_multicores}
\title{Check is multicores parameter is valid using reasonable defaults}
\usage{
.check_multicores(multicores)
}
\arguments{
\item{multicores}{number of cores to be used}
}
\value{
No return value, called for side effects.
}
\description{
Check is multicores parameter is valid using reasonable defaults
}
\keyword{internal}
