% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_check.R
\name{.check_num_parameter}
\alias{.check_num_parameter}
\title{Check is numerical parameter is valid using reasonable defaults}
\usage{
.check_num_parameter(
  param,
  min = -Inf,
  max = Inf,
  len_min = 1,
  len_max = 1,
  allow_na = FALSE,
  exclusive_min = -Inf,
  tolerance = 0
)
}
\arguments{
\item{min}{minimum value}

\item{max}{maximum value}

\item{len_min}{minimum length of vector}

\item{len_max}{maximum length of vector}

\item{allow_na}{allow NA?}

\item{exclusive_min}{is there an exclusive minimum?}

\item{tolerance}{tolerance for equality comparison}

\item{x}{parameter to be checked}
}
\value{
No return value, called for side effects.
}
\description{
Check is numerical parameter is valid using reasonable defaults
}
\keyword{internal}
