% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_check.R
\name{.check_output_dir}
\alias{.check_output_dir}
\title{Check is output_dir parameter is valid using reasonable defaults}
\usage{
.check_output_dir(output_dir)
}
\arguments{
\item{output_dir}{name of output directory}
}
\value{
No return value, called for side effects.
}
\description{
Check is output_dir parameter is valid using reasonable defaults
}
\keyword{internal}
