% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_check.R
\name{.check_predictors}
\alias{.check_predictors}
\title{Are the predictors valid?}
\usage{
.check_predictors(pred, samples)
}
\arguments{
\item{pred}{a tibble with predictors values}

\item{samples}{samples from where the predictors have been calculated}
}
\value{
No return value, called for side effects.
}
\description{
Are the predictors valid?
}
\keyword{internal}
