% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_check.R
\name{.check_require_packages}
\alias{.check_require_packages}
\title{Checks if the required packages are installed}
\usage{
.check_require_packages(x, ..., msg = "Please install package(s)")
}
\arguments{
\item{x}{the name of the required package}
}
\value{
the name of the package
}
\description{
Checks if the required packages are installed
}
\keyword{internal}
