% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_check.R
\name{.check_samples}
\alias{.check_samples}
\title{Does the input data contain a valid sits tibble?}
\usage{
.check_samples(data)
}
\arguments{
\item{data}{a sits tibble}
}
\value{
No return value, called for side effects.
}
\description{
Does the input data contain a valid sits tibble?
}
\keyword{internal}
