% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_check.R
\name{.check_samples_cluster}
\alias{.check_samples_cluster}
\title{Do the samples contain a cluster column?}
\usage{
.check_samples_cluster(data)
}
\arguments{
\item{data}{a sits tibble with cluster col}
}
\value{
No return value, called for side effects.
}
\description{
Do the samples contain a cluster column?
}
\keyword{internal}
