% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_check.R
\name{.check_samples_tile_match}
\alias{.check_samples_tile_match}
\title{Do the samples and tile match?}
\usage{
.check_samples_tile_match(samples, tile)
}
\arguments{
\item{samples}{samples organised as a tibble}

\item{tile}{one tile of a data cube}
}
\value{
No return value, called for side effects.
}
\description{
Do the samples and tile match?
}
\keyword{internal}
