% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_check.R
\name{.check_samples_train}
\alias{.check_samples_train}
\title{Can the input data be used for training?}
\usage{
.check_samples_train(data)
}
\arguments{
\item{data}{a sits tibble}
}
\value{
No return value, called for side effects.
}
\description{
Can the input data be used for training?
}
\keyword{internal}
