% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_check.R
\name{.check_samples_validation}
\alias{.check_samples_validation}
\title{Is the samples_validation object valid?}
\usage{
.check_samples_validation(samples_validation, labels, timeline, bands)
}
\arguments{
\item{samples_validation}{a sits tibble with validation samples}

\item{labels}{labels of the data cube to be validated}

\item{timeline}{timeline of the data cube to be validated}

\item{bands}{bands of the data cube to be validated}
}
\value{
No return value, called for side effects.
}
\description{
Is the samples_validation object valid?
}
\keyword{internal}
