% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_check.R
\name{.check_smoothness}
\alias{.check_smoothness}
\title{Does the data contain the cols of sample data and is not empty?}
\usage{
.check_smoothness(smoothness, nlabels)
}
\arguments{
\item{smoothness}{a matrix or numeric value}

\item{nlabels}{a numeric value}
}
\value{
No return value, called for side effects.
}
\description{
Does the data contain the cols of sample data and is not empty?
}
\keyword{internal}
