% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_check.R
\name{.check_version}
\alias{.check_version}
\title{Check is version parameter is valid using reasonable defaults}
\usage{
.check_version(version)
}
\arguments{
\item{version}{version parameter}
}
\value{
No return value, called for side effects.
}
\description{
Check is version parameter is valid using reasonable defaults
}
\keyword{internal}
