% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_check.R
\name{.check_window_size}
\alias{.check_window_size}
\title{Check is window size is valid using reasonable defaults}
\usage{
.check_window_size(window_size, min = 1, max = 2^32 - 1)
}
\arguments{
\item{window_size}{size of the local window}

\item{min}{minimum value}

\item{max}{maximum value}
}
\value{
No return value, called for side effects.
}
\description{
Check is window size is valid using reasonable defaults
}
\keyword{internal}
