% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_classify_cube.R
\name{.classify_data_read}
\alias{.classify_data_read}
\title{Read a block of values retrieved from a set of raster images}
\usage{
.classify_data_read(tile, block, ml_model, impute_fn, filter_fn)
}
\arguments{
\item{tile}{Input tile to read data.}

\item{block}{Bounding box in (col, row, ncols, nrows).}

\item{ml_model}{Model trained by \code{\link[sits]{sits_train}}.}

\item{impute_fn}{Impute function to replace NA.}

\item{filter_fn}{Smoothing filter function to be applied to the data.}
}
\value{
A matrix with values for classification.
}
\description{
Read a block of values retrieved from a set of raster images
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}

Rolf Simoes, \email{rolf.simoes@inpe.br}
}
\keyword{internal}
