% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_combine_predictions.R
\name{.comb_estimate_block_size}
\alias{.comb_estimate_block_size}
\title{Estimate the number of blocks to run .sits_split_cluster}
\usage{
.comb_estimate_block_size(cubes, multicores, memsize, uncert = FALSE)
}
\arguments{
\item{cubes}{List of input data cube}

\item{multicores}{number of processes to split up the data}

\item{memsize}{maximum overall memory size (in GB)}

\item{uncert}{include uncertainty cubes?}
}
\value{
returns a list with following information:
            - multicores theoretical upper bound;
            - block x_size (horizontal) and y_size (vertical)
}
\description{
Estimate the number of blocks to run .sits_split_cluster
}
\keyword{internal}
