% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_config.R
\name{.config_check_bands}
\alias{.config_check_bands}
\title{Check cube collection}
\usage{
.config_check_bands(source, collection, bands)
}
\arguments{
\item{source}{Data source}

\item{collection}{Collection to be searched in the data source.}

\item{bands}{Bands to be included.}
}
\value{
An invisible null
}
\description{
A suite of check to verify collection in cube.
}
