% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_config.R
\name{.config_colors}
\alias{.config_colors}
\title{Get colors associated to the labels}
\usage{
.config_colors(labels, palette = "Harmonic", rev = TRUE)
}
\arguments{
\item{labels}{labels associated to the training classes}

\item{palette}{palette from `grDevices::hcl.pals()`
replaces default colors
when labels are not included in the config palette}

\item{rev}{revert the order of colors?}
}
\value{
colors required to display the labels
}
\description{
Get colors associated to the labels
}
\keyword{internal}
