% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_cube.R
\name{.config_data_meta_type}
\alias{.config_data_meta_type}
\title{meta-type for data}
\usage{
.config_data_meta_type(data)
}
\arguments{
\item{data}{tibble (time series or cube)}
}
\value{
file path to the appended to data_dir
}
\description{
meta-type for data
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
\keyword{internal}
