% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_config.R
\name{.config_gdalcubes_chunk_size}
\alias{.config_gdalcubes_chunk_size}
\title{Retrieve the gdalcubes chunk size}
\usage{
.config_gdalcubes_chunk_size()
}
\value{
a numeric vector with chunk size
}
\description{
Retrieve the gdalcubes chunk size
}
\keyword{internal}
