% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_config.R
\name{.config_gdalcubes_max_threads}
\alias{.config_gdalcubes_max_threads}
\title{Retrieve the maximum number of threads in gdalcubes}
\usage{
.config_gdalcubes_max_threads()
}
\value{
a numeric with the number of threads
}
\description{
Retrieve the maximum number of threads in gdalcubes
}
\keyword{internal}
