% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_config.R
\name{.config_gdalcubes_min_files_for_parallel}
\alias{.config_gdalcubes_min_files_for_parallel}
\title{Retrieve the minimum requests to do in parallel}
\usage{
.config_gdalcubes_min_files_for_parallel()
}
\value{
get minimum times the parallel requests will be done
}
\description{
Retrieve the minimum requests to do in parallel
}
\keyword{internal}
