% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_config.R
\name{.config_gdalcubes_open_connections}
\alias{.config_gdalcubes_open_connections}
\title{Retrieve the parallel requests number}
\usage{
.config_gdalcubes_open_connections()
}
\value{
get parallel requests
}
\description{
Retrieve the parallel requests number
}
\keyword{internal}
