% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_config.R
\name{.config_get}
\alias{.config_get}
\title{Given a key, get config values}
\usage{
.config_get(key, default = NULL)
}
\value{
config values associated to a key
}
\description{
Given a key, get config values
}
\keyword{internal}
