% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_config.R
\name{.config_gtiff_default_options}
\alias{.config_gtiff_default_options}
\title{Check GEOTIFF creation options}
\usage{
.config_gtiff_default_options()
}
\value{
the creation options associated to the configuration
}
\description{
Check GEOTIFF creation options
}
\keyword{internal}
