% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_conf.R
\name{.config_info}
\alias{.config_info}
\alias{sits_config_info}
\title{Information about configuration file}
\usage{
sits_config_info()
}
\value{
a message
}
\description{
Displays the local of sits configuration file. For details
on how to set the configuration file, use \code{\link[sits]{sits_config}}.
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
