% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_config.R
\name{.config_local_file_extension}
\alias{.config_local_file_extension}
\alias{.config_local_file_extensions}
\title{Get local file extensions}
\usage{
.config_local_file_extensions()
}
\value{
local file extensions known to sits
}
\description{
Get local file extensions
}
\keyword{internal}
