% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_config.R
\name{.config_local_s3_class}
\alias{.config_local_s3_class}
\title{Get local S3 class}
\usage{
.config_local_s3_class()
}
\value{
classes associated to local cubes
}
\description{
Get local S3 class
}
\keyword{internal}
