% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_config.R
\name{.config_metadata_search_strategies}
\alias{.config_metadata_search_strategies}
\title{Retrieve the valid types of metadata search}
\usage{
.config_metadata_search_strategies()
}
\value{
metadata search strategies used by sits.
}
\description{
Retrieve the valid types of metadata search
}
\keyword{internal}
