% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_config.R
\name{.config_names}
\alias{.config_names}
\title{Get names associated to a configuration key}
\usage{
.config_names(key)
}
\arguments{
\item{key}{key combination to access config information}
}
\value{
names associated to the chosen access key
}
\description{
Get names associated to a configuration key
}
\keyword{internal}
