% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_config.R
\name{.config_new_cloud_band}
\alias{.config_new_cloud_band}
\title{Include a new cloud band in the configuration}
\usage{
.config_new_cloud_band(
  bit_mask,
  values,
  interp_values,
  resolution,
  band_name,
  ...
)
}
\arguments{
\item{bit_mask}{bit mask to describe clouds (if applicable)}

\item{interp_values}{pixel values that need to be replaced by interpolation}

\item{resolution}{spatial resolution (in meters)}

\item{band_name}{name of the band}

\item{...}{other relevant parameters}

\item{value}{values of the cloud band}
}
\value{
list with the configuration associated to the new band
}
\description{
creates a description associated to a new cloud band
}
\keyword{internal}
