% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_config.R
\name{.config_new_collection}
\alias{.config_new_collection}
\title{Include a new collection in the configuration}
\usage{
.config_new_collection(bands, ..., satellite = NULL, sensor = NULL)
}
\arguments{
\item{bands}{bands associated to the collection}

\item{...}{other relevant parameters}

\item{satellite}{satellite associated to the collection}

\item{sensor}{sensor associated to the collection}
}
\value{
list with the configuration associated to the new collection
}
\description{
creates a new collection associated to a source
}
\keyword{internal}
