% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_config.R
\name{.config_new_source}
\alias{.config_new_source}
\title{Include a new source in the configuration}
\usage{
.config_new_source(s3_class, collections, ..., service = NULL, url = NULL)
}
\arguments{
\item{s3_class}{s3 class associated to the source}

\item{collections}{collections associated to the source}

\item{...}{other parameters associated to the new source}

\item{service}{protocol associated to the source (e.g., "STAC")}

\item{url}{url associated to the endpoint of the service}
}
\value{
list with the configuration associated to the new source
}
\description{
creates a new data source in the config file
}
\keyword{internal}
