% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_config.R
\name{.config_palettes}
\alias{.config_palettes}
\title{Retrieve the color palettes}
\usage{
.config_palettes()
}
\value{
names of the available color palettes
}
\description{
Retrieve the color palettes
}
\keyword{internal}
