% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_config.R
\name{.config_processing_bloat}
\alias{.config_processing_bloat}
\title{Retrieve the processing bloat}
\usage{
.config_processing_bloat()
}
\value{
estimated processing bloat
}
\description{
Retrieve the processing bloat
}
\keyword{internal}
