% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_config.R
\name{.config_raster_pkg}
\alias{.config_raster_pkg}
\title{Retrieve the raster package to be used}
\usage{
.config_raster_pkg()
}
\value{
the raster package used to process raster data
}
\description{
Retrieve the raster package to be used
}
\keyword{internal}
