% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_config.R
\name{.config_rstac_limit}
\alias{.config_rstac_limit}
\title{Retrieve the rstac pagination limit}
\usage{
.config_rstac_limit()
}
\value{
pagination limit to rstac output
}
\description{
Retrieve the rstac pagination limit
}
\keyword{internal}
