% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_config.R
\name{.config_user_file}
\alias{.config_user_file}
\title{Return the user configuration file}
\usage{
.config_user_file()
}
\value{
user configuration file
}
\description{
Return the user configuration file
}
\keyword{internal}
