% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_utils.R
\name{.create_chunk_file}
\alias{.create_chunk_file}
\title{Create chunk file}
\usage{
.create_chunk_file(output_dir, pattern, ext = ".tif")
}
\arguments{
\item{output_dir}{A character with the output directory to be concatenated.}

\item{pattern}{File name pattern.}

\item{ext}{File extension.}
}
\value{
A character with the file name.
}
\description{
Create a temp file name.
}
\keyword{internal}
