% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_utils.R
\name{.create_filename}
\alias{.create_filename}
\title{Create file name}
\usage{
.create_filename(
  ...,
  filenames = NULL,
  sep = "_",
  ext = NULL,
  output_dir = NULL,
  create_dir = FALSE
)
}
\arguments{
\item{...}{A vector of characters that will be concatenated.}

\item{filenames}{Named parameter to create file name. Note: There is
a difference between \code{dots} and \code{filenames}
parameters. The \code{filenames} parameter just
concatenated the provided names, while \code{dots}
recycled values for each provided vector.
A vector of characters that will be concatenated.}

\item{sep}{A character with a file name separator.}

\item{ext}{A character with the extension of file.}

\item{output_dir}{A character with the output directory to be concatenated.}

\item{create_dir}{A boolean indicating if directory should be created.}
}
\value{
A character with the file name.
}
\description{
Create a file name from a character vectors.
}
\keyword{internal}
