% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_cube_aux_functions.R
\name{.cube_area_freq}
\alias{.cube_area_freq}
\title{Given a labelled cube, return the band information}
\usage{
.cube_area_freq(cube)
}
\arguments{
\item{cube}{Data cube}
}
\value{
Frequency of each label in the data cube
}
\description{
Given a labelled cube, return the band information
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
\keyword{internal}
