% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_cube_aux_functions.R
\name{.cube_band_maximum_value}
\alias{.cube_band_maximum_value}
\title{Maximum value of a cube band}
\usage{
.cube_band_maximum_value(cube, band)
}
\arguments{
\item{cube}{Data cube}

\item{band}{Band}
}
\value{
Maximum value for the band in the data cube.
}
\description{
Maximum value of a cube band
}
\keyword{internal}
