% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_cube_aux_functions.R
\name{.cube_band_minimum_value}
\alias{.cube_band_minimum_value}
\title{Minimum value of a cube band}
\usage{
.cube_band_minimum_value(cube, band)
}
\arguments{
\item{cube}{Data cube}

\item{band}{Band}
}
\value{
minimum value for the band in the data cube
}
\description{
Minimum value of a cube band
}
\keyword{internal}
