% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_cube_aux_functions.R
\name{.cube_band_scale_factor}
\alias{.cube_band_scale_factor}
\title{Scale factor of a cube band}
\usage{
.cube_band_scale_factor(cube, band)
}
\arguments{
\item{cube}{Data cube}

\item{band}{Band}
}
\value{
Scale factor for the band in the data cube.
}
\description{
Scale factor of a cube band
}
\keyword{internal}
