% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_cube_aux_functions.R
\name{.cube_bands}
\alias{.cube_bands}
\title{Return bands of a data cube}
\usage{
.cube_bands(cube, add_cloud = TRUE)
}
\arguments{
\item{cube}{Data cube}

\item{add_cloud}{Include the cloud band?}
}
\value{
a \code{vector} with the bands.
}
\description{
Return bands of a data cube
}
\keyword{internal}
