% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_cube_aux_functions.R
\name{.cube_check}
\alias{.cube_check}
\title{Check if R object is a data cube}
\usage{
.cube_check(cube)
}
\arguments{
\item{cube}{R object to be checked}
}
\value{
TRUE/FALSE
}
\description{
Check if R object is a data cube
}
\keyword{internal}
