% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_cube_aux_functions.R
\name{.cube_collection}
\alias{.cube_collection}
\title{Return collection of a data cube}
\usage{
.cube_collection(cube)
}
\arguments{
\item{cube}{data cube}
}
\value{
collection associated to the cube
}
\description{
Return collection of a data cube
}
\keyword{internal}
