% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_cube_aux_functions.R
\name{.cube_crs}
\alias{.cube_crs}
\title{Return crs of a data cube}
\usage{
.cube_crs(cube)
}
\arguments{
\item{cube}{data cube}
}
\value{
crs associated to the cube
}
\description{
Return crs of a data cube
}
\keyword{internal}
