% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_cube_aux_functions.R
\name{.cube_derived_create_probs}
\alias{.cube_derived_create_probs}
\title{Create a data cube derived from another}
\usage{
.cube_derived_create_probs(
  cube,
  cube_class,
  band_name,
  labels,
  start_date,
  end_date,
  bbox,
  output_dir,
  version
)
}
\arguments{
\item{cube}{input data cube}

\item{cube_class}{class to be attributed to created cube}

\item{band_name}{name of band in created cube}

\item{labels}{labels of derived cube}

\item{start_date}{start date of the cube interval}

\item{end_date}{end date of the cube interval}

\item{bbox}{bounding box of the ROI}

\item{output_dir}{prefix of the output files.}

\item{version}{version of the output files}
}
\value{
output data cube
}
\description{
Take a tibble containing metadata about a data cube
containing time series and create a
set of files to store the result of a classification or smoothing
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
\keyword{internal}
