% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_tile_api.R
\name{.cube_during}
\alias{.cube_during}
\title{Cube API}
\usage{
.cube_during(cube, start_date, end_date)
}
\arguments{
\item{cube}{A cube.}

\item{start_date, end_date}{Date of start and end.}
}
\value{
logical
}
\description{
What tiles have file_info entries between 'start_date'
and 'end_date'?
}
\keyword{internal}
