% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_cube_aux_functions.R
\name{.cube_extract}
\alias{.cube_extract}
\title{Given a band, return a set of values for chosen location}
\usage{
.cube_extract(cube, band_cube, xy)
}
\arguments{
\item{cube}{Metadata about a data cube}

\item{band_cube}{Name of the band to the retrieved}

\item{xy}{Matrix with XY location}
}
\value{
Numeric matrix with raster values for each coordinate
}
\description{
Given a data cube, retrieve the time series
                      of XY locations
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
\keyword{internal}
