% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_cube_aux_functions.R
\name{.cube_file_info}
\alias{.cube_file_info}
\title{Get cube file info}
\usage{
.cube_file_info(cube, bands = NULL)
}
\arguments{
\item{cube}{input data cube}

\item{bands}{bands to be filtered}
}
\value{
A file_info tibble
}
\description{
Get cube file info
}
\author{
Rolf Simoes, \email{rolf.simoes@inpe.br}
}
\keyword{internal}
