% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_cube_aux_functions.R
\name{.cube_filter}
\alias{.cube_filter}
\title{Given a tile, fid, and band, return a new cube with bounding box
values updated.}
\usage{
.cube_filter(cube, tile = NULL, fid = NULL, band = NULL)
}
\arguments{
\item{cube}{Metadata about a data cube}

\item{tile}{A tile name}

\item{fid}{A feature fid}

\item{band}{A band name}
}
\value{
a sits cube with bounding box values updated.
}
\description{
Given a data cube and a tile, fid, and band, return a new
cube with bounding box updated
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
\keyword{internal}
