% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_tile_api.R
\name{.cube_intersects}
\alias{.cube_intersects}
\title{Cube API}
\usage{
.cube_intersects(cube, roi)
}
\arguments{
\item{cube}{A cube.}

\item{roi}{A region of interest (ROI).}
}
\value{
logical
}
\description{
What tiles intersect \code{roi} parameter?
}
