% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_cube_aux_functions.R
\name{.cube_is_equal}
\alias{.cube_is_equal}
\title{Verify if two cubes are equal}
\usage{
.cube_is_equal(x, y)
}
\arguments{
\item{x, y}{a sits cube}
}
\value{
a \code{logical} value.
}
\description{
Given two cubes verify if they are equal
}
\keyword{internal}
