% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_cube_aux_functions.R
\name{.cube_is_equal_bbox}
\alias{.cube_is_equal_bbox}
\title{Checks if the local raster and the cube have exactly the same bbox}
\usage{
.cube_is_equal_bbox(cube)
}
\arguments{
\item{cube}{a sits cube}
}
\value{
a \code{logical} value.
}
\description{
Given a cube this function checks if the cube and the local
raster have the exact box.
}
\keyword{internal}
