% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_cube_aux_functions.R
\name{.cube_is_regular}
\alias{.cube_is_regular}
\title{Check if cube is regular}
\usage{
.cube_is_regular(cube)
}
\arguments{
\item{cube}{input data cube}
}
\value{
TRUE/FALSE
}
\description{
Check if cube is regular
}
\author{
Rolf Simoes, \email{rolf.simoes@inpe.br}
}
\keyword{internal}
