% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_cube_aux_functions.R
\name{.cube_is_token_expires}
\alias{.cube_is_token_expires}
\alias{.cube_is_token_expired}
\title{Check if a cube token was expired}
\usage{
.cube_is_token_expired(cube)
}
\arguments{
\item{cube}{input data cube}
}
\value{
a boolean value.
}
\description{
Check if a cube token was expired
}
\keyword{internal}
