% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_cube_aux_functions.R
\name{.cube_labels}
\alias{.cube_labels}
\title{Return the labels of the cube}
\usage{
.cube_labels(cube)
}
\arguments{
\item{cube}{input data cube}
}
\value{
a vector with the labels of the cube
}
\description{
Return the labels of the cube
}
\author{
Rolf Simoes, \email{rolf.simoes@inpe.br}
}
\keyword{internal}
