% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_cube_aux_functions.R
\name{.cube_params_block}
\alias{.cube_params_block}
\title{Determine the block spatial parameters of a given cube}
\usage{
.cube_params_block(cube, block)
}
\arguments{
\item{cube}{A valid cube}

\item{block}{A block inside the cube}
}
\value{
A tibble with the cube parameters
}
\description{
Based on the R object associated to a raster object,
                determine its parameters
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
\keyword{internal}
