% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_cube_aux_functions.R
\name{.cube_probs_label}
\alias{.cube_probs_label}
\title{Clone a probs or label cube with a new extension}
\usage{
.cube_probs_label(cube, ext, output_dir, version)
}
\arguments{
\item{cube}{input data cube}

\item{ext}{file extension}

\item{output_dir}{prefix of the output files}

\item{version}{version of the output files}
}
\value{
output data cube
}
\description{
Clone a probs or label cube with a new extension
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
\keyword{internal}
