% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_cube_aux_functions.R
\name{.cube_resolution}
\alias{.cube_resolution}
\title{Return the resolution of the cube}
\usage{
.cube_resolution(cube, bands = NULL)
}
\arguments{
\item{cube}{input data cube}
}
\value{
a vector with the x and y resolution
}
\description{
Return the resolution of the cube
}
\author{
Rolf Simoes, \email{rolf.simoes@inpe.br}
}
\keyword{internal}
