% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_cube_aux_functions.R
\name{.cube_resolution_template}
\alias{.cube_resolution_template}
\title{Return the resolution by the template of a tile}
\usage{
.cube_resolution_template(cube, bands = NULL)
}
\arguments{
\item{cube}{input data cube}
}
\value{
a vector with the x and y resolution generate by the raster package.
 By default the terra package is used.
}
\description{
Return the resolution by the template of a tile
}
\author{
Rolf Simoes, \email{rolf.simoes@inpe.br}
}
\keyword{internal}
