% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_cube_aux_functions.R
\name{.cube_s3class}
\alias{.cube_s3class}
\title{Return the S3 class of the cube}
\usage{
.cube_s3class(cube)
}
\arguments{
\item{cube}{input data cube}
}
\value{
class of the cube
}
\description{
Return the S3 class of the cube
}
\author{
Rolf Simoes, \email{rolf.simoes@inpe.br}
}
\keyword{internal}
