% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_cube_aux_functions.R
\name{.cube_size}
\alias{.cube_size}
\title{Return the size of the cube (nrows x ncols)}
\usage{
.cube_size(cube, ..., bands = NULL)
}
\arguments{
\item{cube}{input data cube}
}
\value{
a vector with the labels of the cube
}
\description{
Return the size of the cube (nrows x ncols)
}
\author{
Rolf Simoes, \email{rolf.simoes@inpe.br}
}
\keyword{internal}
