% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_tile_api.R
\name{.cube_spatial_filter}
\alias{.cube_spatial_filter}
\title{Cube API}
\usage{
.cube_spatial_filter(cube, roi)
}
\arguments{
\item{cube}{A cube.}

\item{roi}{A region of interest (ROI).}
}
\value{
cube
}
\description{
Filter tiles that intersect \code{roi} parameter.
}
