% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_tile_api.R
\name{.cube_temporal_filter}
\alias{.cube_temporal_filter}
\title{Cube API}
\usage{
.cube_temporal_filter(cube, start_date, end_date)
}
\arguments{
\item{cube}{A cube.}

\item{start_date, end_date}{Date of start and end.}
}
\value{
cube
}
\description{
Filter tiles with 'file_info' entries between 'start_date'
and 'end_date'.
}
