% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_cube_aux_functions.R
\name{.cube_tile_bbox}
\alias{.cube_tile_bbox}
\title{Get bbox of a cube (single tile)}
\usage{
.cube_tile_bbox(cube)
}
\arguments{
\item{cube}{input data cube}
}
\value{
A vector with tile names
}
\description{
Get bbox of a cube (single tile)
}
\author{
Rolf Simoes, \email{rolf.simoes@inpe.br}
}
\keyword{internal}
