% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_cube_aux_functions.R
\name{.cube_tiles}
\alias{.cube_tiles}
\title{Get cube tiles}
\usage{
.cube_tiles(cube)
}
\arguments{
\item{cube}{input data cube}
}
\value{
A vector with tile names
}
\description{
Get cube tiles
}
\author{
Rolf Simoes, \email{rolf.simoes@inpe.br}
}
\keyword{internal}
