% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_tile_api.R
\name{.cube_timeline}
\alias{.cube_timeline}
\title{Cube API}
\usage{
.cube_timeline(cube)
}
\value{
date or list(date)
}
\description{
Get timeline from each cube. If there are at least two
different timelines, all timelines will be returned in a list).
}
