% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_tile_api.R
\name{.cube_timeline_acquisiton}
\alias{.cube_timeline_acquisiton}
\title{Cube API}
\usage{
.cube_timeline_acquisiton(cube, period, origin)
}
\value{
tibble
}
\description{
Compute how many images were acquired in different periods
and different tiles.
}
