% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_cube_aux_functions.R
\name{.cube_tolerance}
\alias{.cube_tolerance}
\title{Get tolerance of a cube for comparison purposes}
\usage{
.cube_tolerance(cube)
}
\arguments{
\item{cube}{input data cube}
}
\value{
The numerical tolerance value
}
\description{
Get tolerance of a cube for comparison purposes
}
\author{
Rolf Simoes, \email{rolf.simoes@inpe.br}
}
\keyword{internal}
