% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_cube_aux_functions.R
\name{.cube_unique_timeline}
\alias{.cube_unique_timeline}
\title{Check if timelines all tiles of the cube are the same}
\usage{
.cube_unique_timeline(cube)
}
\description{
Check if timelines all tiles of the cube are the same
}
