% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_cube_aux_functions.R
\name{.cube_xres}
\alias{.cube_xres}
\title{Return the cube resolution of the x axis}
\usage{
.cube_xres(cube, bands = NULL)
}
\arguments{
\item{cube}{input data cube}
}
\value{
a numeric with the x resolution
}
\description{
Return the cube resolution of the x axis
}
\author{
Rolf Simoes, \email{rolf.simoes@inpe.br}
}
\keyword{internal}
