% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_cube_aux_functions.R
\name{.cube_yres}
\alias{.cube_yres}
\title{Return the cube resolution of the y axis}
\usage{
.cube_yres(cube, bands = NULL)
}
\arguments{
\item{cube}{input data cube}
}
\value{
a numeric with the y resolution
}
\description{
Return the cube resolution of the y axis
}
\author{
Rolf Simoes, \email{rolf.simoes@inpe.br}
}
\keyword{internal}
