% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_config.R
\name{.data_type}
\alias{.data_type}
\title{Get the data type from a band configuration}
\usage{
.data_type(conf)
}
\arguments{
\item{conf}{A band definition value from config.}
}
\value{
Data type associated to the configuration
}
\description{
Get the data type from a band configuration
}
