% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_gdalcubes.R
\name{.gc_arrange_images}
\alias{.gc_arrange_images}
\title{Images arrangement in sits cube}
\usage{
.gc_arrange_images(cube, timeline, period, ...)
}
\arguments{
\item{cube}{A sits data cube}

\item{timeline}{A timeline of regularized cube}

\item{period}{A period of interval to aggregate images}

\item{...}{Additional parameters.}
}
\value{
A sits cube with the images arranged by cloud
}
\description{
Images arrangement in sits cube
}
\keyword{internal}
