% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_gdalcubes.R
\name{.gc_cloud_mask}
\alias{.gc_cloud_mask}
\title{Create an object image_mask with information about mask band}
\usage{
.gc_cloud_mask(tile)
}
\arguments{
\item{tile}{Data cube tile from where data is to be retrieved.}
}
\value{
A \code{object} 'image_mask' from gdalcubes containing information
 about the mask band.
}
\description{
Create an object image_mask with information about mask band
}
\keyword{internal}
