% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_gdalcubes.R
\name{.gc_create_cloud_mask}
\alias{.gc_create_cloud_mask}
\title{Create an gdalcubes::image_mask object}
\usage{
.gc_create_cloud_mask(cube)
}
\arguments{
\item{cube}{a sits cube object}
}
\value{
A \code{gdalcubes::image_mask} with information about mask band.
}
\description{
Create an gdalcubes::image_mask object
}
\keyword{internal}
