% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_gdalcubes.R
\name{.gc_create_cube_view}
\alias{.gc_create_cube_view}
\title{Create a cube_view object}
\usage{
.gc_create_cube_view(tile, period, res, roi, toi, agg_method, resampling)
}
\arguments{
\item{tile}{A data cube tile}

\item{period}{A \code{character} with the The period of time in which it
is desired to apply in the cube, must be provided based on ISO8601, where 1
number and a unit are provided, for example "P16D".}

\item{res}{A \code{numeric} with spatial resolution of the image that
will be aggregated.}

\item{roi}{A region of interest.}

\item{toi}{A timeline of intersection}

\item{agg_method}{A \code{character} with the method that will be applied in
the aggregation, the following are available: "min", "max", "mean",
"median" or "first".}

\item{resampling}{A \code{character} with method to be used by
\code{gdalcubes} for resampling in mosaic operation.
Options: \code{near}, \code{bilinear}, \code{bicubic} or others supported by
gdalwarp (see https://gdal.org/programs/gdalwarp.html).
By default is bilinear.}
}
\value{
a \code{cube_view} object from gdalcubes.
}
\description{
Create a cube_view object
}
\keyword{internal}
