% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_gdalcubes.R
\name{.gc_create_database_stac}
\alias{.gc_create_database_stac}
\title{Create an image_collection object}
\usage{
.gc_create_database_stac(cube, path_db)
}
\arguments{
\item{cube}{Data cube from where data is to be retrieved.}

\item{path_db}{Path and name for gdalcubes database.}
}
\value{
Image_collection containing information on the
                 images metadata.
}
\description{
Create an image_collection object
}
\keyword{internal}
