% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_gdalcubes.R
\name{.gc_create_pack}
\alias{.gc_create_pack}
\title{Create a gdalcubes::pack object}
\usage{
.gc_create_pack(cube, band)
}
\arguments{
\item{cube}{a sits cube object}

\item{band}{a \code{character} band name}
}
\value{
an \code{gdalcube::pack} object.
}
\description{
Create a gdalcubes::pack object
}
\keyword{internal}
