% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_gdalcubes.R
\name{.gc_create_raster_cube}
\alias{.gc_create_raster_cube}
\title{Create an gdalcubes::raster_cube object}
\usage{
.gc_create_raster_cube(cube_view, path_db, band, mask_band)
}
\arguments{
\item{cube_view}{A \code{gdalcubes::cube_view} object.}

\item{path_db}{A \code{character} path to a gdalcubes database.}

\item{band}{A \code{character} with band name to be generated}

\item{mask_band}{An \code{gdalcubes::image_mask} object with metadata
about the band to be used to mask clouds}
}
\value{
A \code{gdalcubes::image_mask} with information about mask band.
}
\description{
Create an gdalcubes::raster_cube object
}
\keyword{internal}
