% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_gdalcubes.R
\name{.gc_detect_crs_type}
\alias{.gc_detect_crs_type}
\title{Detect the type of cube crs}
\usage{
.gc_detect_crs_type(cube_crs)
}
\arguments{
\item{cube_crs}{A vector of characters with cube crs.}
}
\value{
A character with the type of crs: "proj:wkt2" or "proj:epsg"
}
\description{
Detect the type of cube crs
}
\keyword{internal}
