% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_gdalcubes.R
\name{.gc_format_col}
\alias{.gc_format_col}
\title{Internal function to handle with different file collection formats
 for each provider.}
\usage{
.gc_format_col(source, collection, ...)
}
\arguments{
\item{source}{A \code{character} value referring to a valid data source.}

\item{collection}{A \code{character} value referring to a valid collection.}

\item{...}{Additional parameters.}
}
\value{
A \code{character} path with format collection.
}
\description{
Generic function with the goal that each source implements its own way of
localizing the collection format file.
}
\keyword{internal}
