% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_gdalcubes.R
\name{.gc_get_date}
\alias{.gc_get_date}
\title{Extracted date from aggregated cubes}
\usage{
.gc_get_date(dir_images)
}
\arguments{
\item{dir_images}{A \code{character}  corresponds to the path on which the
images will be saved.}
}
\value{
a \code{character} vector with the dates extracted.
}
\description{
Extracted date from aggregated cubes
}
\keyword{internal}
