% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_gdalcubes.R
\name{.gc_get_valid_interval}
\alias{.gc_get_valid_interval}
\title{Get the interval of intersection in all tiles}
\usage{
.gc_get_valid_interval(cube, period)
}
\arguments{
\item{cube}{Data cube from where data is to be retrieved.}

\item{period}{A \code{character} with ISO8601 time period for regular
data cubes produced by \code{gdalcubes}, with number and unit, e.g., "P16D"
for 16 days. Use "D", "M" and "Y" for days, month and year.}
}
\value{
a \code{list} object with max_min_date and min_max_date.
}
\description{
Get the interval of intersection in all tiles
}
\keyword{internal}
