% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_gdalcubes.R
\name{.gc_get_valid_timeline}
\alias{.gc_get_valid_timeline}
\title{Get the timeline of intersection in all tiles}
\usage{
.gc_get_valid_timeline(cube, period)
}
\arguments{
\item{cube}{Data cube.}

\item{period}{ISO8601 time period.}
}
\value{
a \code{vector} with all timeline values.
}
\description{
Get the timeline of intersection in all tiles
}
\keyword{internal}
