% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_gdalcubes.R
\name{.gc_missing_tiles}
\alias{.gc_missing_tiles}
\title{Finds the missing tiles in a regularized cube}
\usage{
.gc_missing_tiles(cube, local_cube, timeline)
}
\arguments{
\item{cube}{Original cube to be regularized.}

\item{timeline}{Timeline used by gdalcubes for regularized cube}

\item{gc_cube}{Regularized cube (may be missing tiles).}

\item{period}{Period of timeline regularization.}
}
\value{
Tiles that are missing from the regularized cube.
}
\description{
Finds the missing tiles in a regularized cube
}
\keyword{internal}
