% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_gdalcubes.R
\name{.gc_new_cube}
\alias{.gc_new_cube}
\title{Save the images based on an aggregation method.}
\usage{
.gc_new_cube(
  tile,
  res,
  start_date,
  resampling,
  roi,
  output_dir,
  multithreads,
  ...
)
}
\arguments{
\item{tile}{A data cube tile}

\item{res}{...}

\item{resampling}{...}

\item{roi}{...}

\item{output_dir}{Directory where the aggregated images will be written.}

\item{multithreads}{A \code{numeric} with the number of cores will be used in
the regularize. By default is used 1 core.}

\item{...}{...}
}
\value{
A data cube tile with information used in its creation.
}
\description{
Save the images based on an aggregation method.
}
\keyword{internal}
