% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_gdalcubes.R
\name{.gc_save_raster_cube}
\alias{.gc_save_raster_cube}
\title{Saves the images of a raster cube.}
\usage{
.gc_save_raster_cube(raster_cube, pack, output_dir, files_prefix, ...)
}
\arguments{
\item{raster_cube}{\code{gdalcubes::raster_cube} object.}

\item{pack}{\code{gdalcubes::pack} object.}

\item{output_dir}{Directory where the aggregated images will be written.}

\item{files_prefix}{File names prefix.}

\item{...}{Additional parameters that can be included. See
'?gdalcubes::write_tif'.}
}
\value{
A list of generated images.
}
\description{
Saves the images of a raster cube.
}
\keyword{internal}
