% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_gdalcubes.R
\name{.gc_update_metadata}
\alias{.gc_update_metadata}
\title{Update metadata from sits cube using gdalcubes metadata}
\usage{
.gc_update_metadata(cube, cube_view)
}
\arguments{
\item{cube}{Data cube from where data is to be retrieved.}

\item{cv}{A \code{object} 'cube_view' with values from cube.
for aggregation.}
}
\value{
a \code{sits_cube} object with updated metadata.
}
\description{
Update metadata from sits cube using gdalcubes metadata
}
\keyword{internal}
