% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_active_learning.R
\name{.get_top_values}
\alias{.get_top_values}
\title{Get top values of a raster.}
\usage{
.get_top_values(r_obj, band, n, min_dist_pixels)
}
\arguments{
\item{n}{Number of values to extract.}

\item{min_dist_pixels}{Minimum distance among values (in pixels).}

\item{raster}{A `terra` raster object.}
}
\value{
A point `tibble` object.
}
\description{
Get the top values of a raster as a point `sf` object. The values
locations are guaranteed to be separated by a certain number of pixels.
}
\author{
Alber Sanchez, \email{alber.ipia@inpe.br}
}
