% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_tile_api.R
\name{.intersects}
\alias{.intersects}
\title{Spatial intersects}
\usage{
.intersects(x, y)
}
\arguments{
\item{x, y}{\code{sf} geometries.}
}
\value{
A \code{logical} vector indicating which geometries of \code{x}
intersects \code{y} geometries.
}
\description{
This function is based on \code{sf::st_intersects()}. It projects \code{y}
to the CRS of \code{x} before compute intersection. For each geometry of
\code{x}, the function returns the \code{TRUE} if it intersects with any
geometry of \code{y}, otherwise it returns \code{FALSE}.
}
\examples{
\dontrun{
x <- .bbox_as_sf(c(xmin=1, xmax=2, ymin=3, ymax=4, crs=4326))
y <- .roi_as_sf(c(lon_min=1.5, lon_max=3, lat_min=3.5, lat_max=5))
.intersects(x, y) # TRUE
}

}
\seealso{
Other utility functions: 
\code{\link{.try}()}

Other region objects API: 
\code{\link{bbox_api}},
\code{\link{block_api}},
\code{\link{chunks_api}},
\code{\link{roi_api}}
}
\concept{region objects API}
\concept{utility functions}
\keyword{internal}
