% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_mixture_model.R
\name{.mesma_raster_blocks_estimate}
\alias{.mesma_raster_blocks_estimate}
\title{Estimate the number of blocks}
\usage{
.mesma_raster_blocks_estimate(nbands, sub_image, memsize, multicores)
}
\arguments{
\item{nbands}{Number of bands to open.}

\item{sub_image}{Area of interest in the image.}

\item{memsize}{Memory available for classification (in GB).}

\item{multicores}{Number of cores to process the time series.}
}
\value{
Number of blocks to be read.
}
\description{
Defines the number of blocks of a Raster Brick
             to be read into memory.
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
\keyword{internal}
