% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_mixture_model.R
\name{.mesma_scale_endmembers}
\alias{.mesma_scale_endmembers}
\alias{.mm_scale_endmembers}
\title{Scale the endmembers spectra}
\usage{
.mm_scale_endmembers(endmembers, cube)
}
\arguments{
\item{endmembers}{Tibble with endmembers spectra values.}

\item{cube}{Data cube.}
}
\value{
a matrix with spectral values scaled.
}
\description{
Scale the endmembers spectra
}
\keyword{internal}
