% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_metrics_generate.R
\name{.metrics_blocks_estimate}
\alias{.metrics_blocks_estimate}
\title{...}
\usage{
.metrics_blocks_estimate(cube, n_metrics, memsize)
}
\arguments{
\item{cube}{...}

\item{n_metrics}{...}

\item{memsize}{...}
}
\value{
a list ...
}
\description{
...
}
