% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_mixture_model.R
\name{.mm_get_steps_lst}
\alias{.mm_get_steps_lst}
\title{Create a list of steps}
\usage{
.mm_get_steps_lst(cube, output_fractions)
}
\arguments{
\item{cube}{Data cube.}

\item{output_fractions}{Name of output fractions.}
}
\value{
List of combination among tiles, endmembers, and dates
                  that are missing from the cube.
}
\description{
Create a list of steps
}
\keyword{internal}
