% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_plot.R
\name{.plot_allyears}
\alias{.plot_allyears}
\title{Plot all intervals of one time series for the same lat/long together}
\usage{
.plot_allyears(data)
}
\arguments{
\item{data}{One or more time series.}
}
\value{
A plot object produced by the ggplot2 package
               showing an individual time series.
}
\description{
For each lat/long location in the data, join temporal
instances of the same place together for plotting.
}
\keyword{internal}
