% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_plot.R
\name{.plot_false_color}
\alias{.plot_false_color}
\title{Plot a false color image}
\usage{
.plot_false_color(file, band, palette, rev)
}
\arguments{
\item{file}{File to be plotted.}

\item{band}{Band to be plotted.}

\item{palette}{A sequential RColorBrewer palette}

\item{rev}{Reverse the color palette?}
}
\value{
A plot object
}
\description{
plots a set of false color image
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
\keyword{internal}
