% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_plot.R
\name{.plot_ggplot_series_na}
\alias{.plot_ggplot_series_na}
\title{Plot one time series with NAs using ggplot}
\usage{
.plot_ggplot_series_na(row)
}
\arguments{
\item{row}{row of a sits tibble with the time series to be plotted.}
}
\value{
A plot object produced by the ggplot2 package
                   which shows the NA values of a time series.
}
\description{
Plots a set of time series using ggplot, showing where NAs are.
}
\keyword{internal}
