% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_plot.R
\name{.plot_rgb}
\alias{.plot_rgb}
\title{Plot a RGB image}
\usage{
.plot_rgb(red_file, green_file, blue_file)
}
\arguments{
\item{red_file}{File with the band to be plotted in red}

\item{green_file}{File with the band to be plotted in green}

\item{blue_file}{File with the band to be plotted in blue}
}
\value{
A plot object
}
\description{
Plot a RGB image
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
\keyword{internal}
