% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_plot.R
\name{.plot_together}
\alias{.plot_together}
\title{Plot a set of time series for the same spatiotemporal reference}
\usage{
.plot_together(data)
}
\arguments{
\item{data}{A sits tibble with the list of time series to be plotted.}
}
\value{
A set of plots produced by the ggplot2 package
                  each containing all time series associated to one band
                  and one label.
}
\description{
Plots all time series for the same label together.
This function is useful to find out the spread of the values of
the time series for a given label.
}
\keyword{internal}
