% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_raster_api.R
\name{.raster_check_bbox}
\alias{.raster_check_bbox}
\title{Check for bbox object consistency}
\usage{
.raster_check_bbox(bbox)
}
\value{
No value, called for side effects.
}
\description{
Check for bbox object consistency
}
\author{
Rolf Simoes, \email{rolf.simoes@inpe.br}
}
\keyword{internal}
