% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_raster_api.R
\name{.raster_colrow}
\alias{.raster_colrow}
\alias{.raster_col}
\alias{.raster_row}
\title{Raster package internal frequency values function}
\usage{
.raster_col(r_obj, x)

.raster_row(r_obj, y)
}
\arguments{
\item{r_obj}{raster package object}

\item{x, y}{coordinates (either x or y) in raster projection}
}
\value{
integer with column or row
}
\description{
Raster package internal frequency values function
}
\author{
Rolf Simoes, \email{rolf.simoes@inpe.br}
}
\keyword{internal}
