% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_raster_api.R
\name{.raster_crop_metadata}
\alias{.raster_crop_metadata}
\title{Raster package internal crop raster function}
\usage{
.raster_crop_metadata(r_obj, ..., block = NULL, bbox = NULL)
}
\arguments{
\item{r_obj}{raster package object to be written}

\item{...}{additional parameters to be passed to raster package}

\item{block}{a valid block with (\code{col}, \code{row},
\code{ncols}, \code{nrows}).}

\item{bbox}{numeric vector with (\code{xmin}, \code{xmax},
\code{ymin}, \code{ymax}).}
}
\value{
Subset of a raster object as defined by either block
               or bbox parameters
}
\description{
Raster package internal crop raster function
}
\note{
block starts at (1, 1)
}
\author{
Rolf Simoes, \email{rolf.simoes@inpe.br}
}
\keyword{internal}
