% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_raster_api.R
\name{.raster_data_type}
\alias{.raster_data_type}
\title{Raster package internal data type representation}
\usage{
.raster_data_type(data_type)
}
\arguments{
\item{data_type}{sits internal raster data type.}
}
\value{
internal data type used by raster package
}
\description{
Raster package internal data type representation
}
\author{
Rolf Simoes, \email{rolf.simoes@inpe.br}
}
\keyword{internal}
