% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_raster_api.R
\name{.raster_ext_as_sf}
\alias{.raster_ext_as_sf}
\alias{.raster_file_blocksize}
\title{Raster package internal extract values function}
\usage{
.raster_ext_as_sf(r_obj)

.raster_file_blocksize(r_obj)
}
\arguments{
\item{r_obj}{raster package object}
}
\value{
An object with raster extent.

An vector with the file block size.
}
\description{
Raster package internal extract values function

Raster package internal extract values function
}
\author{
Rolf Simoes, \email{rolf.simoes@inpe.br}
}
\keyword{internal}
