% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_raster_api.R
\name{.raster_gdal_datatype}
\alias{.raster_gdal_datatype}
\alias{.raster_gdal_datatypes}
\title{Convert internal data type to gdal data type}
\usage{
.raster_gdal_datatype(data_type)

.raster_gdal_datatypes()
}
\value{
character string
}
\description{
Convert internal data type to gdal data type
}
\author{
Rolf Simoes, \email{rolf.simoes@inpe.br}
}
\keyword{internal}
