% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_raster_api.R
\name{.raster_gdal_datatypes}
\alias{.raster_gdal_datatypes}
\title{Match sits data types to GDAL data types}
\usage{
.raster_gdal_datatypes(sits_names = TRUE)
}
\description{
Match sits data types to GDAL data types
}
