% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_raster_api.R
\name{.raster_get_values}
\alias{.raster_get_values}
\title{Raster package internal get values function}
\usage{
.raster_get_values(r_obj, ...)
}
\arguments{
\item{r_obj}{raster package object}

\item{...}{additional parameters to be passed to raster package}
}
\value{
Numeric matrix
}
\description{
Raster package internal get values function
}
\author{
Rolf Simoes, \email{rolf.simoes@inpe.br}
}
\keyword{internal}
