% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_raster_api.R
\name{.raster_merge}
\alias{.raster_merge}
\alias{.raster_merge_blocks}
\title{Merge all input files into one raster file}
\usage{
.raster_merge_blocks(
  out_files,
  base_file,
  block_files,
  data_type,
  missing_value,
  multicores = 2
)
}
\arguments{
\item{out_files}{Output raster files path.}

\item{base_file}{Raster file path to be used as template. If \code{NULL},
all block_files will be merged without a template.}

\item{block_files}{Input block files paths.}

\item{data_type}{sits internal raster data type. One of "INT1U",
"INT2U", "INT2S", "INT4U", "INT4S", "FLT4S", "FLT8S".}

\item{missing_value}{Missing value of out_files.}

\item{multicores}{Number of cores to process merging}
}
\value{
No return value, called for side effects.
}
\description{
Merge all input files into one raster file
}
\author{
Rolf Simoes, \email{rolf.simoes@inpe.br}
}
\keyword{internal}
