% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_raster_api.R
\name{.raster_missing_value}
\alias{.raster_missing_value}
\title{Raster package internal open raster function}
\usage{
.raster_missing_value(file)
}
\arguments{
\item{file}{raster file to be opened}
}
\value{
a numeric with no data value
}
\description{
Raster package internal open raster function
}
\author{
Rolf Simoes, \email{rolf.simoes@inpe.br}
}
\keyword{internal}
