% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_raster_api.R
\name{.raster_new_rast}
\alias{.raster_new_rast}
\title{Raster package internal create raster object function}
\usage{
.raster_new_rast(nrows, ncols, xmin, xmax, ymin, ymax, nlayers, crs, ...)
}
\arguments{
\item{nrows}{Number of rows in the raster}

\item{ncols}{Number of columns in the raster}

\item{xmin}{X minimum of raster origin}

\item{xmax}{X maximum of raster origin}

\item{ymin}{Y minimum of raster origin}

\item{ymax}{Y maximum of raster origin}

\item{nlayers}{Number of layers of the raster}

\item{crs}{Coordinate Reference System of the raster}

\item{...}{additional parameters to be passed to raster package}
}
\value{
A raster object.
}
\description{
Raster package internal create raster object function
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}

Rolf Simoes, \email{rolf.simoes@inpe.br}
}
\keyword{internal}
