% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_raster_api.R
\name{.raster_properties}
\alias{.raster_properties}
\alias{.raster_nrows}
\alias{.raster_ncols}
\alias{.raster_nlayers}
\alias{.raster_xmax}
\alias{.raster_xmin}
\alias{.raster_ymax}
\alias{.raster_ymin}
\alias{.raster_xres}
\alias{.raster_yres}
\alias{.raster_crs}
\alias{.raster_bbox}
\alias{.raster_res}
\alias{.raster_size}
\title{Raster package internal object properties}
\usage{
.raster_nrows(r_obj, ...)

.raster_ncols(r_obj, ...)

.raster_nlayers(r_obj, ...)

.raster_xmax(r_obj, ...)

.raster_xmin(r_obj, ...)

.raster_ymax(r_obj, ...)

.raster_ymin(r_obj, ...)

.raster_xres(r_obj, ...)

.raster_yres(r_obj, ...)

.raster_crs(r_obj, ...)

.raster_bbox(r_obj, ..., block = NULL)

.raster_res(r_obj, ...)

.raster_size(r_obj, ...)
}
\arguments{
\item{r_obj}{raster package object}

\item{...}{additional parameters to be passed to raster package}
}
\value{
raster object spatial properties
}
\description{
Raster package internal object properties
}
\author{
Rolf Simoes, \email{rolf.simoes@inpe.br}
}
\keyword{internal}
