% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_raster_api.R
\name{.raster_read_rast}
\alias{.raster_read_rast}
\alias{.raster_write_rast}
\title{Raster package internal read raster file function}
\usage{
.raster_read_rast(file, ..., block = NULL)

.raster_write_rast(
  r_obj,
  file,
  format,
  data_type,
  gdal_options,
  overwrite,
  ...
)
}
\arguments{
\item{file}{file path to save raster file}

\item{...}{additional parameters to be passed to raster package}

\item{block}{numeric vector with names "first_col", "ncols", "first_row", "nrows".}

\item{r_obj}{raster package object to be written}

\item{format}{GDAL file format string (e.g. GTiff)}

\item{data_type}{sits internal raster data type. One of "INT1U",
"INT2U", "INT2S", "INT4U", "INT4S", "FLT4S", "FLT8S".}

\item{gdal_options}{GDAL creation option string (e.g. COMPRESS=LZW)}

\item{overwrite}{logical indicating if file can be overwritten}
}
\value{
numeric matrix

numeric matrix
}
\description{
Raster package internal read raster file function

Raster package internal write raster file function
}
\author{
Rolf Simoes, \email{rolf.simoes@inpe.br}

Rolf Simoes, \email{rolf.simoes@inpe.br}
}
\keyword{internal}
