% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_raster_api.R
\name{.raster_resampling}
\alias{.raster_resampling}
\title{Raster package internal resampling method}
\usage{
.raster_resampling(method)
}
\arguments{
\item{method}{sits internal raster resampling method.}
}
\value{
resampling method (if valid)
}
\description{
Raster package internal resampling method
}
\author{
Rolf Simoes, \email{rolf.simoes@inpe.br}
}
\keyword{internal}
