% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_raster_api.R
\name{.raster_set_values}
\alias{.raster_set_values}
\title{Raster package internal set values function}
\usage{
.raster_set_values(r_obj, values, ...)
}
\arguments{
\item{r_obj}{raster package object}

\item{values}{Numeric matrix to copy to raster object}

\item{...}{additional parameters to be passed to raster package}
}
\value{
raster package object
}
\description{
Raster package internal set values function
}
\author{
Rolf Simoes, \email{rolf.simoes@inpe.br}
}
\keyword{internal}
