% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_raster_api.R
\name{.raster_supported_packages}
\alias{.raster_supported_packages}
\title{Supported raster packages}
\usage{
.raster_supported_packages()
}
\value{
Names of raster packages supported by sits
}
\description{
Supported raster packages
}
\keyword{internal}
