% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_raster_api.R
\name{.raster_write_rast}
\alias{.raster_write_rast}
\title{Raster package internal write raster file function}
\usage{
.raster_write_rast(
  r_obj,
  file,
  format,
  data_type,
  overwrite,
  ...,
  missing_value = NA
)
}
\arguments{
\item{r_obj}{raster package object to be written}

\item{file}{file path to save raster file}

\item{format}{GDAL file format string (e.g. GTiff)}

\item{data_type}{sits internal raster data type. One of "INT1U",
"INT2U", "INT2S", "INT4U", "INT4S", "FLT4S", "FLT8S".}

\item{overwrite}{logical indicating if file can be overwritten}

\item{...}{additional parameters to be passed to raster package}

\item{missing_value}{A \code{integer} with image's missing value}
}
\value{
No value, called for side effects.
}
\description{
Raster package internal write raster file function
}
\author{
Rolf Simoes, \email{rolf.simoes@inpe.br}
}
\keyword{internal}
