% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_reclassify.R
\name{.rclasf_get_jobs}
\alias{.rclasf_get_jobs}
\title{List all jobs to be computed}
\usage{
.rclasf_get_jobs(cube)
}
\arguments{
\item{cube}{Data cube.}
}
\value{
List of combination among tiles and dates.
}
\description{
List all jobs to be computed
}
\keyword{internal}
