% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_gdalcubes.R
\name{.reg_aggregate_chunk}
\alias{.reg_aggregate_chunk}
\title{Aggregate raster chunk}
\usage{
.reg_aggregate_chunk(rast_stack, filename, datatype, ...)
}
\arguments{
\item{rast_stack}{A \code{SpatRast} object}

\item{filename}{A \code{character} with filename of the rast.}

\item{datatype}{A \code{character} with terra datatype.}

\item{...}{additional paramters for terra merge methods.}
}
\value{
An invisible null
}
\description{
Aggregate raster chunk
}
\keyword{internal}
