% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_gdalcubes.R
\name{.reg_apply_mask_rast}
\alias{.reg_apply_mask_rast}
\title{Apply a cloud mask in a raster}
\usage{
.reg_apply_mask_rast(rast, mask_rast, interp_values, ...)
}
\arguments{
\item{rast}{A \code{SpatRast} object}

\item{mask_rast}{A \code{numeric}}

\item{interp_values}{A \code{numeric} with}

\item{...}{additional paramters for terra mask methods.}
}
\value{
A \code{SpatRast} object masked.
}
\description{
Apply a cloud mask in a raster
}
\keyword{internal}
