% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_gdalcubes.R
\name{.reg_create_filaname}
\alias{.reg_create_filaname}
\title{Create the merge image filename}
\usage{
.reg_create_filaname(tile, date, output_dir, block = NULL)
}
\arguments{
\item{tile}{A unique tile from \code{sits_cube} object}

\item{output_dir}{A \code{character} with a valid directory where the
regularized images will be written.}

\item{block}{A \code{numeric} vector with information about a block}

\item{period}{A \code{character} vector with two position, first one is
the start date and second one is the end date.}
}
\value{
A \code{character} with the file name of resampled image.
}
\description{
Create the merge image filename
}
\keyword{internal}
