% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_regularize_aux_functions.R
\name{.reg_filename}
\alias{.reg_filename}
\title{Create the merge image filename}
\usage{
.reg_filename(tile, band, date, output_dir, ..., block = NULL)
}
\arguments{
\item{tile}{...}

\item{band}{....}

\item{output_dir}{A \code{character} with a valid directory where the
regularized images will be written.}

\item{block}{A \code{numeric} vector with information about a block}

\item{date_period}{A \code{character} vector with two position, first one is
the start date and second one is the end date.}
}
\value{
A \code{character} with the file name of resampled image.
}
\description{
Create the merge image filename
}
\keyword{internal}
