% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_gdalcubes.R
\name{.reg_get_chunk_rast}
\alias{.reg_get_chunk_rast}
\title{Get chunks from rasters}
\usage{
.reg_get_chunk_rast(rast, tile, block, filename, datatype)
}
\arguments{
\item{rast}{A \code{SpatRast} object}

\item{block}{A \code{numeric} vector with information about a block}
}
\value{
A \code{SpatRast} object cropped.
}
\description{
Get chunks from rasters
}
\keyword{internal}
