% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_gdalcubes.R
\name{.reg_merge_chunks}
\alias{.reg_merge_chunks}
\title{Merge raster chunks}
\usage{
.reg_merge_chunks(rast, filename, datatype, ...)
}
\arguments{
\item{rast}{A \code{SpatRast} object}

\item{filename}{A \code{character} with filename of the rast.}

\item{datatype}{A \code{character} with terra datatype.}

\item{...}{additional paramters for terra merge methods.}
}
\value{
An invisible null
}
\description{
Merge raster chunks
}
\keyword{internal}
