% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_regularize.R
\name{.reg_preprocess_block}
\alias{.reg_preprocess_block}
\title{Preprocessing steps of sits regularize}
\usage{
.reg_preprocess_block(
  tile_fid,
  band,
  out_size,
  band_block,
  cloud_block,
  output_block,
  ratio_band_out,
  ratio_cloud_out,
  data_type,
  output_file
)
}
\arguments{
\item{tile}{A unique tile from \code{sits_cube} object}

\item{band_paths}{A \code{character} with paths for a unique band}

\item{resolution}{A \code{numeric} with spatial resolution of the image
that will be aggregated.}

\item{resampling}{A \code{character} with method to be used  for resampling
in mosaic operation. Options: \code{near}, \code{bilinear}, \code{bicubic},
\code{cubicspline}, and \code{lanczos}. Default is bilinear.}

\item{block}{A \code{numeric} vector with information about a block}
}
\value{
A \code{SpatRast} object resampled
}
\description{
Preprocessing steps of sits regularize
}
\keyword{internal}
