% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_regularize.R
\name{.reg_regularize}
\alias{.reg_regularize}
\title{Finds the missing files in a regularized cube}
\usage{
.reg_regularize(
  cube,
  period,
  res,
  output_dir,
  multicores = 1,
  memsize = 4,
  progress = TRUE
)
}
\arguments{
\item{...}{...}
}
\value{
tiles that are missing from the regularized cube
}
\description{
Finds the missing files in a regularized cube
}
\keyword{internal}
