% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_gdalcubes.R
\name{.reg_resample_rast}
\alias{.reg_resample_rast}
\title{Resample raster to a resolution}
\usage{
.reg_resample_rast(rast, resolution, resampling, ...)
}
\arguments{
\item{rast}{A \code{SpatRast} object}

\item{resolution}{A \code{numeric} with spatial resolution of the image
that will be aggregated.}

\item{resampling}{A \code{character} with method to be used  for resampling
in mosaic operation. Options: \code{near}, \code{bilinear}, \code{bicubic},
\code{cubicspline}, and \code{lanczos}. Default is bilinear.}

\item{...}{additional paramters for terra resample methods.}
}
\value{
A \code{SpatRast} object resample
}
\description{
Resample raster to a resolution
}
\keyword{internal}
