% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_regularize_aux_functions.R
\name{.reg_timeline}
\alias{.reg_timeline}
\title{Compute the timeline of the regularized cube}
\usage{
.reg_timeline(cube, period)
}
\arguments{
\item{cube}{Data cube from where data is to be retrieved.}

\item{period}{ISO8601 time period.}
}
\value{
Timeline values.
}
\description{
Compute the timeline of the regularized cube
}
\keyword{internal}
