% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_metrics.R
\name{.set_source_metrics}
\alias{.set_source_metrics}
\title{...}
\usage{
.set_source_metrics(cube, metrics)
}
\arguments{
\item{cube}{input data cube.}

\item{metrics}{metrics list. melhorar a documentacao}
}
\value{
an invisible null
}
\description{
change config on the fly
}
\keyword{internal}
