% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_tibble_aux_functions.R
\name{.sits_align_dates}
\alias{.sits_align_dates}
\title{Aligns dates of time series to a reference date}
\usage{
.sits_align_dates(data, ref_dates)
}
\arguments{
\item{data}{Input sits tibble (useful for chaining functions).}

\item{ref_dates}{Dates to align the time series.}
}
\value{
The converted sits tibble
}
\description{
Converts the time indexes of a set of sits
tibble to a single reference year.
This function is useful to join many time series from
different years to a single year,
which is required by methods that combine many time series,
such as clustering methods.
The reference year is taken from the date of the start of the time series
available in the data cube.
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
\keyword{internal}
