% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_aws.R
\name{.sits_aws_check_access}
\alias{.sits_aws_check_access}
\title{Check access rights on AWS}
\usage{
.sits_aws_check_access(source)
}
\arguments{
\item{source}{data source}
}
\value{
TRUE if access is granted
}
\description{
Check access rights on AWS
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
\keyword{internal}
