% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_sentinel.R
\name{.sits_aws_check_bands}
\alias{.sits_aws_check_bands}
\title{Check bands by resolution}
\usage{
.sits_aws_check_bands(bands, s2_resolution)
}
\arguments{
\item{bands}{a \code{character} vector with the bands name.}

\item{s2_resolution}{a \code{character} withh resolution of S2 images
("10m", "20m" or "60m")}
}
\value{
a \code{character} vector with the selected bands.
}
\description{
Check bands by resolution
}
\keyword{internal}
