% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_sentinel.R
\name{.sits_aws_items}
\alias{.sits_aws_items}
\title{Get information from items}
\usage{
.sits_aws_items(
  url = NULL,
  collection = NULL,
  tiles = NULL,
  roi = NULL,
  start_date = NULL,
  end_date = NULL,
  bands = NULL,
  ...
)
}
\arguments{
\item{url}{a \code{character} representing a URL for the AWS catalog.}

\item{collection}{a \code{character} with the collection to be searched.}

\item{tiles}{a \code{character} with the names of the tiles.}

\item{roi}{selects images (tiles) that intersect according to the
region of interest provided. Expressed either as an \code{sfc} or \code{sf}
object from sf package, a \code{character} with GeoJSON following the rules
from RFC 7946, or a \code{vector} with bounding box named XY values in
WGS 84 ("xmin", "xmax", "ymin", "ymax").}

\item{start_date}{a \code{character} corresponds to the initial date
when the cube will be created.}

\item{end_date}{a \code{character} corresponds to the final date when the
cube will be created.}

\item{bands}{a \code{character} vector with the bands name.}

\item{...}{other parameters to be passed for specific types.}
}
\value{
a \code{STACItemCollection} object representing the search
                  by rstac.
}
\description{
Get information from items
}
\keyword{internal}
