% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_sentinel.R
\name{.sits_aws_tile_cube}
\alias{.sits_aws_tile_cube}
\title{Get the STAC information corresponding to a tile.}
\usage{
.sits_aws_tile_cube(url, name, items, cube, resolution, file_info)
}
\arguments{
\item{url}{a \code{character} representing URL for the AWS STAC.}

\item{name}{a \code{character} representing the output data cube.}

\item{items}{a \code{STACItemCollection} object returned by rstac.}

\item{cube}{a \code{character} with name input data cube in AWS}

\item{resolution}{a \code{character} withh resolution of S2 images
("10m", "20m" or "60m")}

\item{file_info}{a \code{tbl_df} with the information from STAC.}
}
\value{
a \code{tibble} with metadata information about a
                  raster data set.
}
\description{
Get the STAC information corresponding to a tile.
}
\keyword{internal}
