% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_kohonen.R
\name{.sits_bayesian_som_neighbourhood}
\alias{.sits_bayesian_som_neighbourhood}
\title{Paint neurons}
\usage{
.sits_bayesian_som_neighbourhood(kohonen_obj, neurons_labelled.tb,
  result.tb)
}
\arguments{
\item{kohonen_obj}{Object Kohonen, this object contains all parameters of SOM provided by package Kohonen}

\item{neurons_labelled.tb}{A tibble containing informations about each neuron.}

\item{result.tb}{A tibble with samples.}
}
\value{
Returns the probability of a sample belongs to a cluster based on class of neuron and its neighborhood.
}
\description{
This function compute the probability of a sample belongs to a cluster using bayesian filter.
}
\author{
Lorena Santos, \email{lorena.santos@inpe.br}
}
