% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_bbox.R
\name{.sits_bbox_intersect}
\alias{.sits_bbox_intersect}
\title{Intersection between a bounding box and a cube}
\usage{
.sits_bbox_intersect(bbox, cube)
}
\arguments{
\item{bbox}{Bounding box for a region of interest.}

\item{cube}{Data cube.}
}
\value{
Vector the bounding box intersection.
}
\description{
Intersection between a bounding box and a cube
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
\keyword{internal}
