% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_space_time_operations.R
\name{.sits_bbox_time_series}
\alias{.sits_bbox_time_series}
\title{Find the bounding box for a set of time series}
\usage{
.sits_bbox_time_series(data)
}
\arguments{
\item{data}{A tibble with a set of time series}
}
\value{
A vector the bounding box
}
\description{
Given a set of time series, find the bounding box.
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
