% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_bbox.R
\name{.sits_bbox_to_sf}
\alias{.sits_bbox_to_sf}
\title{Convert a bounding box to a sf object (polygon)}
\usage{
.sits_bbox_to_sf(xmin, xmax, ymin, ymax, crs)
}
\arguments{
\item{xmin, xmax, ymin, ymax}{Bounding box values.}

\item{crs}{Valid crs value.}
}
\value{
An sf object.
}
\description{
Convert a bounding box to a sf object (polygon)
}
\keyword{internal}
