% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_bdc.R
\name{.sits_bdc_access_check}
\alias{.sits_bdc_access_check}
\title{Check access rigths on BDC}
\usage{
.sits_bdc_access_check()
}
\value{
access_key
}
\description{
Check access rigths on BDC
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
\keyword{internal}
