% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_bdc.R
\name{.sits_bdc_access_info}
\alias{.sits_bdc_access_info}
\title{Include BDC access info}
\usage{
.sits_bdc_access_info(cube, access_key)
}
\arguments{
\item{cube}{BDC data cube}

\item{access_key}{BDC access key}
}
\value{
cube with additional information
}
\description{
Include BDC access info
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
\keyword{internal}
