% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_bdc.R
\name{.sits_bdc_check_tiles}
\alias{.sits_bdc_check_tiles}
\title{Check if the BDC tiles are working}
\usage{
.sits_bdc_check_tiles(
  satellite,
  sensor,
  bands,
  cube,
  tile,
  data_access,
  start_date,
  end_date
)
}
\arguments{
\item{satellite}{satellite}

\item{sensor}{sensor}

\item{bands}{bands}

\item{cube}{input cube}

\item{tile}{tile}

\item{data_access}{type of access}

\item{start_date}{start_date of the cube}

\item{end_date}{end date of the cube}
}
\description{
Check if the BDC tiles are working
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
