% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_bdc.R
\name{.sits_bdc_info_tiles}
\alias{.sits_bdc_info_tiles}
\title{Get information on BDC tiles}
\usage{
.sits_bdc_info_tiles(
  satellite,
  sensor,
  cube,
  tile,
  data_access,
  start_date,
  end_date,
  .local,
  .web
)
}
\arguments{
\item{satellite}{satellite}

\item{sensor}{sensor}

\item{cube}{input cube}

\item{tile}{tile}

\item{data_access}{type of access}

\item{start_date}{start_date of the cube}

\item{end_date}{end date of the cube}

\item{.local}{local address (if different from default)}

\item{.web}{web address (if different from default)}
}
\description{
Get information on BDC tiles
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
