% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_bdc.R
\name{.sits_bdc_stack_info}
\alias{.sits_bdc_stack_info}
\title{Obtain the file information about the BDC data}
\usage{
.sits_bdc_stack_info(
  satellite,
  sensor,
  start_date,
  end_date,
  bands,
  data_dir,
  data_access
)
}
\arguments{
\item{satellite}{satellite}

\item{sensor}{sensor}

\item{start_date}{starting date for the cube}

\item{end_date}{end date for the cube}

\item{bands}{bands}

\item{data_dir}{base directory for the search}

\item{data_access}{type of access}
}
\value{
tibble with metadata information
}
\description{
Obtain the file information about the BDC data
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
\keyword{internal}
