% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_cloud.R
\name{.sits_bitmask_values}
\alias{.sits_bitmask_values}
\title{Retrieve the indexes to be removed in time series}
\usage{
.sits_bitmask_values(cld_values, cld_index)
}
\arguments{
\item{cld_values}{Cloud band values.}

\item{cld_index}{Indexes to be removed in cloud values.}
}
\value{
The values with indexes to be removed in time series.
}
\description{
Retrieve the indexes to be removed in time series
}
\keyword{internal}
