% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_tibble_aux_functions.R
\name{.sits_break}
\alias{.sits_break}
\title{Breaks a set of time series into equal intervals}
\usage{
.sits_break(data, timeline, start_date, end_date, interval = "12 month")
}
\arguments{
\item{data}{A sits tibble.}

\item{timeline}{Timeline associated with the data cube.}

\item{start_date}{Starting date within an interval.}

\item{end_date}{Ending date within an interval.}

\item{interval}{Interval for breaking the series.}
}
\value{
A sits tibble broken into equal intervals.
}
\description{
This function breaks a set of time series
into equal time intervals. This function is useful to produce
a set of time series with the same number of samples, which is
required for building a set of samples for classification.
}
\examples{
points.tb <- sits:::.sits_break(point_ndvi, timeline_modis_392,
                        "2000-08-28", "2016-08-12")

}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
\keyword{internal}
