% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_brewer.R
\name{.sits_brewer_color_name}
\alias{.sits_brewer_color_name}
\title{Brewer color schemes}
\usage{
.sits_brewer_color_name(name = NULL)
}
\arguments{
\item{name}{name of the brewer color set.}
}
\value{
A string with the common color set name.
}
\description{
Brewer color schemes names
}
\keyword{internal}
