% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_raster_utils.R
\name{.sits_check_classify_params}
\alias{.sits_check_classify_params}
\title{Check clasification parameters}
\usage{
.sits_check_classify_params(file, coverage, ml_model)
}
\arguments{
\item{file}{vector of file names to store the output (one file per classified year)}

\item{coverage}{tibble with information about a set of space-time raster bricks}

\item{ml_model}{an R model trained by \code{\link[sits]{sits_train}}}
}
\value{
OK              (logical) tests succeeded?
}
\description{
Verify that required parameters are correct
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
