% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_raster_utils.R
\name{.sits_check_results}
\alias{.sits_check_results}
\title{Check the results of the classification with the input data}
\usage{
.sits_check_results(prediction, nrows_DT)
}
\arguments{
\item{prediction}{prototype with predictions}

\item{nrows_DT}{number of rows of distance matrix}
}
\value{
check           TRUE if check is OK
}
\description{
verify that classification results have the right size
}
\author{
Rolf Simoes, \email{rolf.simoes@inpe.br}
}
