% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_roi.R
\name{.sits_check_roi_cube}
\alias{.sits_check_roi_cube}
\title{Check is a ROI is valid for an existing data cube}
\usage{
.sits_check_roi_cube(roi, cube)
}
\arguments{
\item{roi}{spatial region of interest}

\item{cube}{input data cube.}
}
\value{
vector with information on the subimage
}
\description{
Check is a ROI is valid for an existing data cube
}
\keyword{internal}
