% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_roi.R
\name{.sits_check_roi_cube.xy}
\alias{.sits_check_roi_cube.xy}
\title{Check is a ROI defined as XY is valid for an existing data cube}
\usage{
\method{.sits_check_roi_cube}{xy}(roi)
}
\arguments{
\item{roi}{spatial region of interest}
}
\value{
roi in WGS84 projection
}
\description{
Check is a ROI defined as XY is valid for an existing data cube
}
\keyword{internal}
