% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_config.R
\name{.sits_check_service}
\alias{.sits_check_service}
\title{Check that the service is valid}
\usage{
.sits_check_service(service)
}
\arguments{
\item{service}{name of the time series service}
}
\description{
Check that the service is valid
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
