% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_cube.R
\name{.sits_class_band_name}
\alias{.sits_class_band_name}
\title{Define a name for classified band}
\usage{
.sits_class_band_name(name, type, start_date, end_date)
}
\arguments{
\item{name}{Original cube name (without temporal information).}

\item{type}{Type of output}

\item{start_date}{Starting date of the time series classification.}

\item{end_date}{End date of the time series classification.}
}
\value{
Name of the classification file for the required interval.
}
\description{
Creates a name for a raster layer with associated temporal information,
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
