% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_classification.R
\name{.sits_classify_check_params}
\alias{.sits_classify_check_params}
\title{Check clasification parameters}
\usage{
.sits_classify_check_params(cube, ml_model)
}
\arguments{
\item{cube}{Tibble with information about a data cube.}

\item{ml_model}{An R model trained by \code{\link[sits]{sits_train}}.}
}
\value{
Tests succeeded?
}
\description{
Verify that required parameters are correct.
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
