% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_cube_classification.R
\name{.sits_classify_cube}
\alias{.sits_classify_cube}
\title{Classify a data cube using multicore machines}
\usage{
.sits_classify_cube(cube = NULL, ml_model = NULL,
  interval = "12 month", filter = NULL, memsize = 4,
  multicores = 2, out_prefix = "cube-class")
}
\arguments{
\item{cube}{Tibble with information about a data cube.}

\item{ml_model}{An R model trained by \code{\link[sits]{sits_train}}.}

\item{interval}{Interval between two sucessive classifications, expressed in months.}

\item{filter}{Smoothing filter to be applied (if desired).}

\item{memsize}{Memory available for classification (in GB).}

\item{multicores}{Number of cores to be used for classification.}

\item{out_prefix}{Prefix of the output files. For each time interval, one file will be created.}
}
\value{
A tibble with the metadata for the vector of classified RasterLayers.
}
\description{
Takes a data cube, whose metadata is
             described by tibble (created by \code{\link[sits]{sits_cube}}),
             a set of samples used for training a classification model,
             a prediction model (created by \code{\link[sits]{sits_train}}),
             and produces a classified set of RasterLayers. These
              parameters are "memsize" and "multicores". The "multicores" parameter defines the
              number of cores used for processing. The "memsize" parameter  controls
              the amount of memory available for classification.
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
