% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_raster_data.R
\name{.sits_classify_data_read}
\alias{.sits_classify_data_read}
\title{Read a block of values retrieved from a set of raster images}
\usage{
.sits_classify_data_read(
  tile,
  block,
  ml_model,
  impute_fn,
  filter_fn,
  output_dir
)
}
\arguments{
\item{tile}{Input tile to read data.}

\item{block}{Bounding box in (col, row, ncols, nrows).}

\item{ml_model}{Model trained by \code{\link[sits]{sits_train}}.}

\item{impute_fn}{Impute function to replace NA.}

\item{filter_fn}{Smoothing filter function to be applied to the data.}

\item{output_dir}{Output directory.}
}
\value{
A matrix with values for classification.
}
\description{
Read a block of values retrieved from a set of raster images
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}

Rolf Simoes, \email{rolf.simoes@inpe.br}
}
\keyword{internal}
