% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_classification.R
\name{.sits_classify_estimate_processing_time}
\alias{.sits_classify_estimate_processing_time}
\title{Estimate the processing time}
\usage{
.sits_classify_estimate_processing_time(start_time, select.lst, bs, block,
  time)
}
\arguments{
\item{start_time}{Initial processing time.}

\item{select.lst}{List of time intervals.}

\item{bs}{Raster block parameters.}

\item{block}{Current block.}

\item{time}{Current interval.}
}
\value{
Scaled matrix.
}
\description{
This function normalizes one band of the values read from a raster brick.
}
\author{
Rolf Simoes, \email{rolf.simoes@inpe.br}
}
