% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_classify_cube.R
\name{.sits_classify_multicores}
\alias{.sits_classify_multicores}
\title{Classify a chunk of raster data  using multicores}
\usage{
.sits_classify_multicores(
  tile,
  ml_model,
  roi,
  filter_fn,
  impute_fn,
  interp_fn,
  compose_fn,
  memsize,
  multicores,
  output_dir,
  version,
  verbose
)
}
\arguments{
\item{tile}{a single tile of a data cube.}

\item{ml_model}{model trained by \code{\link[sits]{sits_train}}.}

\item{roi}{region of interest}

\item{filter_fn}{smoothing filter function to be applied to the data.}

\item{impute_fn}{impute function to replace NA}

\item{interp_fn}{function to interpolate points from cube to match
samples}

\item{compose_fn}{function to compose points from cube to match samples}

\item{memsize}{memory available for classification (in GB).}

\item{multicores}{number of cores.}

\item{output_dir}{output directory}

\item{version}{version of result}

\item{verbose}{print processing information?}
}
\value{
List of the classified raster layers.
}
\description{
Classifies a block of data using multicores. It breaks
the data into horizontal blocks and divides them between the available cores.

Reads data using terra, cleans the data for NAs and missing values.
The clean data is stored in a data table with the time instances
for all pixels of the block. The algorithm then classifies data on
an year by year basis. For each year, extracts the sub-blocks for each band.

After all cores process their blocks, it joins the result and then writes it
in the classified images for each corresponding year.
}
\author{
Rolf Simoes, \email{rolf.simoes@inpe.br}

Gilberto Camara, \email{gilberto.camara@inpe.br}
}
\keyword{internal}
