% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_ts_classification.R
\name{.sits_classify_ts}
\alias{.sits_classify_ts}
\title{Classify a set of time series using machine learning models}
\usage{
.sits_classify_ts(data.tb, ml_model, interval, multicores)
}
\arguments{
\item{data.tb}{Tibble with time series metadata and data.}

\item{ml_model}{Pre-built machine learning model (see \code{\link[sits]{sits_train}}).}

\item{interval}{Interval used for classification (in months).}

\item{multicores}{Number of cores to be used for classification.}
}
\value{
A tibble with the predicted labels for each input segment.
}
\description{
This function classifies a set of time series, given
a set of training samples, an inference model, and an interval.
To perform the classification, users should provide a set of
labelled samples. Each samples should be associated to one spatial location
(latitude/longitude), one time interval and a label.

After defining the training samples, the users need to provide a machine learning model.
Currenly, sits supports the following models:
'svm' (see \code{\link[sits]{sits_svm}}), 'random forest' (see \code{\link[sits]{sits_rfor}}),
'lda' (see \code{\link[sits]{sits_lda}}),
'qda' (see \code{\link[sits]{sits_qda}}), multinomial logit' (see \code{\link[sits]{sits_mlr}}),
'lasso' (see \code{\link[sits]{sits_mlr}}), and 'ridge' (see \code{\link[sits]{sits_mlr}}).

The model should be precomputed by the user. This model should be
passed to the function using the parameter "ml_model".
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
