% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_clouds.R
\name{.sits_cloud_blocks_estimate}
\alias{.sits_cloud_blocks_estimate}
\title{Estimate the number of blocks to correct for clouds}
\usage{
.sits_cloud_blocks_estimate(cube, memsize)
}
\arguments{
\item{cube}{input data cube}

\item{memsize}{size of memory available}
}
\value{
list with:
               n (number of blocks),
               row (vector of starting rows)
               nrow (vector with number of rows for each block)
               size (vector with size of each block)
}
\description{
Estimate the number of blocks to correct for clouds
}
\keyword{internal}
