% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_clouds.R
\name{.sits_cloud_create_output}
\alias{.sits_cloud_create_output}
\title{Create the bricks that will be the output of the cloud estimation procedure}
\usage{
.sits_cloud_create_output(cube, bands = NULL, interval = "16 day", data_dir)
}
\arguments{
\item{cube}{input data cube}

\item{bands}{bands to be included in the output}

\item{interval}{temporal interval between images of the cube}

\item{data_dir}{directory where data is to be stored}
}
\value{
A list of brick objects pointing to output files
}
\description{
Create the bricks that will be the output of the cloud estimation procedure
}
