% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_clouds.R
\name{.sits_clouds_interpolate}
\alias{.sits_clouds_interpolate}
\title{Create the bricks that will be the output of the cloud estimation procedure}
\usage{
.sits_clouds_interpolate(
  cube,
  data_dir,
  bricks,
  blocks,
  impute_ts,
  multicores,
  verbose
)
}
\arguments{
\item{cube}{input data cube}

\item{data_dir}{directory where data is to be stored}

\item{bricks}{terra object pointing to output bricks}

\item{blocks}{block information}

\item{impute_ts}{imputation function to remove NA}

\item{multicores}{number of cores to use}

\item{.verbose}{print information about the processing}
}
\value{
a tibble with date, band and path information
}
\description{
Create the bricks that will be the output of the cloud estimation procedure
}
\keyword{internal}
