% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_clouds.R
\name{.sits_clouds_shds_estimate}
\alias{.sits_clouds_shds_estimate}
\title{Estimate clouds and shadows for a CBERS AWFI image}
\usage{
.sits_clouds_shds_estimate(
  cube,
  extent,
  ref_date,
  values,
  t1,
  t2,
  t3,
  t4,
  t5,
  t6,
  multicores
)
}
\arguments{
\item{cube}{Data cube}

\item{extent}{Image extent to be read}

\item{ref_date}{Reference date for estimation}

\item{values}{Vector of values for cloud detection}

\item{t1}{controls difference btw visible and infrared bands}

\item{t2}{controls brightness properties of cloud}

\item{t3}{controls dark property of cloud shadows.}

\item{t4}{remove influence of water in shadow detection}

\item{t5}{size of window to search for clouds near shadows}

\item{t6}{size of window of median filter to remove outliers}

\item{multicores}{number of cores}
}
\value{
Matrix of cloud and shadow values
}
\description{
Estimate clouds and shadows for a CBERS AWFI image
}
\keyword{internal}
