% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_clouds.R
\name{.sits_clouds_thres_estimate}
\alias{.sits_clouds_thres_estimate}
\title{Retrieve the values associated to cloud detection for an extent}
\usage{
.sits_clouds_thres_estimate(cube, extent, ref_date)
}
\arguments{
\item{cube}{Data cube}

\item{extent}{Image extent to be read.}

\item{ref_date}{Reference date to be processed}
}
\value{
Vector of values for cloud detection
}
\description{
Retrieve the values associated to cloud detection for an extent
}
\keyword{internal}
