% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_clouds.R
\name{.sits_clouds_values_combine}
\alias{.sits_clouds_values_combine}
\title{Estimate values associated to cloud detection for a whole CBERS image}
\usage{
.sits_clouds_values_combine(values_lst)
}
\arguments{
\item{values}{List of values estimated for each extent}
}
\value{
Combined vector of estimated values
}
\description{
Estimate values associated to cloud detection for a whole CBERS image
}
\keyword{internal}
