% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_kohonen.R
\name{.sits_cluster_probability}
\alias{.sits_cluster_probability}
\title{Get cluster probability}
\usage{
.sits_cluster_probability(data)
}
\arguments{
\item{data}{A sits tibble with info of samples provided by sits_cluster_som.}
}
\value{
Returns a sits tibble with a new column indicating the probability of a sample belongs to a cluster and the majority cluster of a sample
}
\description{
This function summarize the probability a sample belongs to a cluster from the label of the neuron
of several iterations.
}
\author{
Lorena Santos, \email{lorena.santos@inpe.br}
}
