% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_dendrogram.R
\name{.sits_cluster_validity}
\alias{.sits_cluster_validity}
\title{Cluster validity indices}
\usage{
.sits_cluster_validity(data)
}
\arguments{
\item{data}{A tibble with `cluster` column.}
}
\value{
A vector with four external validity indices.
}
\description{
Compute different cluster validity indices. This function needs
as input a sits tibble with `cluster` column.
It is a front-end to `dtwclust::cvi` function. That function computes five indices:
1) adjusted Rand index; 2) Rand index; 3) Jaccard index; 4) Fowlkes-Mallows; and 5) Variation of Information index
Please refer to the documentation in that package for more details.
}
\references{
"dtwclust" package (https://CRAN.R-project.org/package=dtwclust)
}
\author{
Rolf Simoes, \email{rolf.simoes@inpe.br}
}
