% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_conf_matrix.R
\name{.sits_conf_matrix_show}
\alias{.sits_conf_matrix_show}
\title{Print the values of a confusion matrix}
\usage{
.sits_conf_matrix_show(
  x,
  mode = "sens_spec",
  digits = max(3, getOption("digits") - 3)
)
}
\arguments{
\item{x}{An object of class \code{confusionMatrix}.}

\item{mode}{A single character string either "sens_spec",
"prec_recall", or "everything".}

\item{digits}{Number of significant digits when printed.}
}
\value{
\code{x}   is invisibly returned.
}
\description{
Adaptation of the caret::print.confusionMatrix method
             for the more common usage in Earth Observation.
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
\keyword{internal}
