% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_config.R
\name{.sits_config_aws_default_region}
\alias{.sits_config_aws_default_region}
\title{Read the AWS default region from configuration file}
\usage{
.sits_config_aws_default_region(type)
}
\arguments{
\item{type}{Type of data cube}
}
\value{
directory where BDC is accessible on the web
}
\description{
Read the AWS default region from configuration file
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
\keyword{internal}
