% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_config.R
\name{.sits_config_aws_stac}
\alias{.sits_config_aws_stac}
\title{Directory to read the AWS STAC catalogue}
\usage{
.sits_config_aws_stac()
}
\value{
directory where DEAFRICA is accessible on the web
}
\description{
Directory to read the AWS STAC catalogue
}
\keyword{internal}
