% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_config.R
\name{.sits_config_band_names}
\alias{.sits_config_band_names}
\title{Obtain the name of the bands used by a cube or by SITS}
\usage{
.sits_config_band_names(sensor, type)
}
\arguments{
\item{sensor}{Name of the sensor}

\item{type}{Type of the data cube (or "SITS")}
}
\value{
Name of the bands used in that data cube or by SITS
}
\description{
Obtain the name of the bands used by a cube or by SITS
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
\keyword{internal}
