% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_config.R
\name{.sits_config_band_names_convert_bdc}
\alias{.sits_config_band_names_convert_bdc}
\title{Convert bands names from BDC to SITS}
\usage{
.sits_config_band_names_convert_bdc(satellite, sensor)
}
\arguments{
\item{satellite}{Name of the satellite}

\item{sensor}{Name of sensor}
}
\value{
Name of the bands used in SITS (named vector)
}
\description{
Convert the name of the band used by BDC
             to the name used by SITS
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
\keyword{internal}
