% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_config.R
\name{.sits_config_bands_no_cloud}
\alias{.sits_config_bands_no_cloud}
\title{Retrieve all but except the cloud band}
\usage{
.sits_config_bands_no_cloud(cube)
}
\arguments{
\item{cube}{A data cube}
}
\value{
vector with bands that are not cloud bands
}
\description{
Retrieve all but except the cloud band
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
\keyword{internal}
