% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_config.R
\name{.sits_config_bands_res}
\alias{.sits_config_bands_res}
\title{Resolution for S2 bands in DEAFRICA}
\usage{
.sits_config_bands_res(source, collection, bands)
}
\arguments{
\item{source}{cube source}
}
\value{
resolution information
}
\description{
Resolution for S2 bands in DEAFRICA
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
\keyword{internal}
