% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_config.R
\name{.sits_config_bands_stac_read}
\alias{.sits_config_bands_stac_read}
\title{Convert bands names from SITS to cube}
\usage{
.sits_config_bands_stac_read(source, collection, bands)
}
\arguments{
\item{source}{Name of the STAC provider}

\item{collection}{Name of sensor}

\item{bands}{Bands requested to be read}
}
\value{
Name of the bands used in the STAC provider
}
\description{
Convert the name of the band used by SITS to
    the names used by the STAC provider
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
\keyword{internal}
