% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_config.R
\name{.sits_config_bands_stac_write}
\alias{.sits_config_bands_stac_write}
\title{Convert bands names from cube to SITS}
\usage{
.sits_config_bands_stac_write(tile)
}
\arguments{
\item{tile}{Data cube tile}
}
\value{
Data cube tile with SITS bands
}
\description{
Convert the name of the band used by the origin data cube
             to the name used by SITS
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
\keyword{internal}
