% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_config.R
\name{.sits_config_bdc_extension}
\alias{.sits_config_bdc_extension}
\title{File extension used by BDC}
\usage{
.sits_config_bdc_extension()
}
\value{
extension of BDC files
}
\description{
File extension used by BDC
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
\keyword{internal}
