% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_config.R
\name{.sits_config_bdc_stac}
\alias{.sits_config_bdc_stac}
\title{Directory to read the BDC STAC catalogue}
\usage{
.sits_config_bdc_stac()
}
\value{
directory where BDC is accessible on the web
}
\description{
Directory to read the BDC STAC catalogue
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
\keyword{internal}
