% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_config.R
\name{.sits_config_bdc_stac_access}
\alias{.sits_config_bdc_stac_access}
\title{Test if BDC STAC catalogue is available}
\usage{
.sits_config_bdc_stac_access(url = NULL)
}
\arguments{
\item{url}{URL for access to the BDC STAC}
}
\value{
directory where BDC is accessible on the web
}
\description{
Test if BDC STAC catalogue is available
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
\keyword{internal}
