% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_config.R
\name{.sits_config_bdc_web_access}
\alias{.sits_config_bdc_web_access}
\title{Test if the BDC is working}
\usage{
.sits_config_bdc_web_access(url = NULL)
}
\arguments{
\item{url}{URL for access to the BDC}
}
\value{
TRUE/FALSE if BDC can be accessed
}
\description{
Test if the BDC is working
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
\keyword{internal}
