% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_config.R
\name{.sits_config_check}
\alias{.sits_config_check}
\title{Check that the service is valid, based on the configuration file}
\usage{
.sits_config_check(service)
}
\arguments{
\item{service}{Name of the time series service.}
}
\description{
Check that the service is valid, based on the configuration file
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
