% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_config.R
\name{.sits_config_check_type}
\alias{.sits_config_check_type}
\title{Check that the type is valid, based on the configuration file}
\usage{
.sits_config_check_type(type)
}
\arguments{
\item{type}{Type of data cube}
}
\description{
Check that the type is valid, based on the configuration file
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
\keyword{internal}
