% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_config.R
\name{.sits_config_cloud_band}
\alias{.sits_config_cloud_band}
\title{Get the name of the band used for cloud information}
\usage{
.sits_config_cloud_band(satellite)
}
\arguments{
\item{satellite}{satellite}
}
\value{
vector with names of the bands available in AWS for a given resolution
}
\description{
Get the name of the band used for cloud information
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
