% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_config.R
\name{.sits_config_cloud_bitmask}
\alias{.sits_config_cloud_bitmask}
\title{Get the flag of bitmask in cloud information}
\usage{
.sits_config_cloud_bitmask(cube)
}
\arguments{
\item{cube}{data cube}
}
\value{
vector with bands available in AWS for a given resolution
}
\description{
Get the flag of bitmask in cloud information
}
\author{
Gilberto Camara \email{gilberto.camara@inpe.br}
}
\keyword{internal}
