% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_config.R
\name{.sits_config_cube_access}
\alias{.sits_config_cube_access}
\title{Test if cube is available via URL}
\usage{
.sits_config_cube_access(url, name)
}
\arguments{
\item{url}{URL for access to the cube}

\item{name}{service name}
}
\value{
TRUE/FALSE
}
\description{
Test if cube is available via URL
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
\keyword{internal}
