% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_config.R
\name{.sits_config_cube_bdc_tile_local}
\alias{.sits_config_cube_bdc_tile_local}
\title{Directory to read the BDC information as local file}
\usage{
.sits_config_cube_bdc_tile_local()
}
\value{
directory where BDC is accessible on the web
}
\description{
Directory to read the BDC information as local file
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
