% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_config.R
\name{.sits_config_cube_bdc_tile_web}
\alias{.sits_config_cube_bdc_tile_web}
\title{Directory to read the BDC information on the web}
\usage{
.sits_config_cube_bdc_tile_web()
}
\value{
directory where BDC is accessible on the web
}
\description{
Directory to read the BDC information on the web
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
