% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_config.R
\name{.sits_config_cube_check}
\alias{.sits_config_cube_check}
\title{Check that the cube data source is valid, based on the configuration file}
\usage{
.sits_config_cube_check(cube)
}
\arguments{
\item{cube}{Data cube}
}
\value{
true/false is the type is valid
}
\description{
Check that the cube data source is valid, based on the configuration file
}
\author{
Gilberto Camara \email{gilberto.camara@inpe.br}
}
\keyword{internal}
