% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_config.R
\name{.sits_config_cube_class_generic}
\alias{.sits_config_cube_class_generic}
\title{Retrieve the classes associated to data cubes known to SITS}
\usage{
.sits_config_cube_class_generic(type)
}
\arguments{
\item{type}{Data cube type}
}
\value{
Class of data cube
}
\description{
Retrieve the class name associated to a cube type
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
\keyword{internal}
