% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_config.R
\name{.sits_config_cube_file_access}
\alias{.sits_config_cube_file_access}
\title{Test if files in a raster cube are}
\usage{
.sits_config_cube_file_access(cube)
}
\arguments{
\item{cube}{cube}
}
\value{
TRUE/FALSE
}
\description{
Test if files in a raster cube are
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
\keyword{internal}
