% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_config.R
\name{.sits_config_cube_metadata}
\alias{.sits_config_cube_metadata}
\title{Retrieve the metadata class associated to data cubes known to SITS}
\usage{
.sits_config_cube_metadata(type)
}
\arguments{
\item{type}{Data cube type}
}
\value{
Class of data cube metadata
}
\description{
Retrieve the metadata class name associated to a cube type
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
