% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_config.R
\name{.sits_config_cube_robj_class}
\alias{.sits_config_cube_robj_class}
\title{Retrieve the classes of Raster objects associated to data cubes known to SITS}
\usage{
.sits_config_cube_robj_class(cube)
}
\arguments{
\item{cube}{Data cube}
}
\value{
class of robjects in data cubes ("brick", "raster", or "stack")
}
\description{
Retrieve the class name associated to a cube type
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
\keyword{internal}
