% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_config.R
\name{.sits_config_cube_service}
\alias{.sits_config_cube_service}
\title{Check if the cube is a web service}
\usage{
.sits_config_cube_service(cube)
}
\arguments{
\item{cube}{data cube}
}
\value{
TRUE or FALSE
}
\description{
Check if the cube is a web service
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
