% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_config.R
\name{.sits_config_cube_types_chk}
\alias{.sits_config_cube_types_chk}
\title{Check the cube types available in the configuration file}
\usage{
.sits_config_cube_types_chk(type)
}
\arguments{
\item{type}{type of data cube}
}
\value{
List of services supported by SITS
}
\description{
Check the cube types available in the configuration file
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
