% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_config.R
\name{.sits_config_data_delim}
\alias{.sits_config_data_delim}
\title{Delimiter for data type}
\usage{
.sits_config_data_delim(type)
}
\arguments{
\item{type}{cube_type}
}
\value{
delimiter information
}
\description{
Delimiter for data type
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
\keyword{internal}
