% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_config.R
\name{.sits_config_data_dir_path}
\alias{.sits_config_data_dir_path}
\title{Standard files for data directory for cube type}
\usage{
.sits_config_data_dir_path(type)
}
\arguments{
\item{type}{cube_type}
}
\value{
file path to the appended to data_dir
}
\description{
Standard files for data directory for cube type
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
