% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_config.R
\name{.sits_config_data_parse_info}
\alias{.sits_config_data_parse_info}
\title{Standard files for data directory for cube type}
\usage{
.sits_config_data_parse_info(type)
}
\arguments{
\item{type}{cube_type}
}
\value{
parsing information
}
\description{
Standard files for data directory for cube type
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
\keyword{internal}
