% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_config.R
\name{.sits_config_download_maxcores}
\alias{.sits_config_download_maxcores}
\alias{.sits_config_access_maxcores}
\title{Number of parallel data access}
\usage{
.sits_config_access_maxcores(source)
}
\arguments{
\item{type}{Data cube type}
}
\value{
name of the bucket
}
\description{
Number of parallel data access
}
\author{
Rolf Simoes, \email{rolf.simoes@inpe.br}
}
\keyword{internal}
