% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_config.R
\name{.sits_config_gtiff_default_options}
\alias{.sits_config_gtiff_default_options}
\title{GDAL GTiff creation options}
\usage{
.sits_config_gtiff_default_options()
}
\value{
character vector with GDAL creation options
}
\description{
GDAL GTiff creation options
}
\author{
Rolf Simoes, \email{rolf.simoes@inpe.br}
}
\keyword{internal}
