% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_config.R
\name{.sits_config_img_file_ext}
\alias{.sits_config_img_file_ext}
\title{Returns the file extensions known to SITS}
\usage{
.sits_config_img_file_ext()
}
\value{
vector of image file extensions known to SITS
}
\description{
Returns the file extensions known to SITS
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
\keyword{internal}
