% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_config.R
\name{.sits_config_label_missing_value}
\alias{.sits_config_label_missing_value}
\title{Retrieve the missing value for a label cube}
\usage{
.sits_config_label_missing_value()
}
\description{
Retrieve the missing value for a label cube
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
\keyword{internal}
