% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_config.R
\name{.sits_config_minimum_values}
\alias{.sits_config_minimum_values}
\title{Retrieve the minimum values for a given band}
\usage{
.sits_config_minimum_values(cube, bands)
}
\arguments{
\item{cube}{Name of the sensor}

\item{bands}{Bands provided by the sensor}
}
\value{
The minimum values.
}
\description{
Retrieve the minimum values for a given band
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
\keyword{internal}
