% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_config.R
\name{.sits_config_missing_values}
\alias{.sits_config_missing_values}
\title{Retrieve the missing values for bands of a sensor}
\usage{
.sits_config_missing_values(sensor, bands)
}
\arguments{
\item{sensor}{Name of the sensor}

\item{bands}{Vector of bands.}
}
\value{
The missing values.
}
\description{
Retrieve the missing values for bands of a sensor
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
