% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_config.R
\name{.sits_config_probs_minimum_value}
\alias{.sits_config_probs_minimum_value}
\title{Retrieve the minimum value for a probs cube}
\usage{
.sits_config_probs_minimum_value()
}
\description{
Retrieve the minimum value for a probs cube
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
\keyword{internal}
