% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_config.R
\name{.sits_config_probs_scale_factor}
\alias{.sits_config_probs_scale_factor}
\title{Retrieve the scale factor for a probs cube}
\usage{
.sits_config_probs_scale_factor()
}
\description{
Retrieve the scale factor for a probs cube
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
\keyword{internal}
