% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_config.R
\name{.sits_config_processing_bloat}
\alias{.sits_config_processing_bloat}
\title{Retrieve the estimated value of R memory bloat}
\usage{
.sits_config_processing_bloat()
}
\description{
Retrieve the expected memory bloat associated to R.
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
\keyword{internal}
