% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_config.R
\name{.sits_config_raster_api}
\alias{.sits_config_raster_api}
\title{Find out what raster api is being used by SITS}
\usage{
.sits_config_raster_api()
}
\value{
raster API used
}
\description{
Find out what raster api is being used by SITS
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
\keyword{internal}
