% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_config.R
\name{.sits_config_raster_package}
\alias{.sits_config_raster_package}
\title{Raster package to be used}
\usage{
.sits_config_raster_package()
}
\value{
name of the raster package
}
\description{
Raster package to be used
}
\author{
Rolf Simoes, \email{rolf.simoes@inpe.br}
}
\keyword{internal}
