% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_config.R
\name{.sits_config_resampling}
\alias{.sits_config_resampling}
\title{Retrieve the resmapling method for bands of a sensor}
\usage{
.sits_config_resampling(cube, bands)
}
\arguments{
\item{bands}{Vector of bands.}

\item{sensor}{Name of the sensor}
}
\value{
The resampling methods.
}
\description{
Retrieve the resmapling method for bands of a sensor
}
\author{
Rolf Simoes, \email{rolf.simoes@inpe.br}
}
\keyword{internal}
