% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_config.R
\name{.sits_config_rstac_limit}
\alias{.sits_config_rstac_limit}
\title{Get pagination limit}
\usage{
.sits_config_rstac_limit()
}
\value{
number of items to be returned in each page
}
\description{
Get pagination limit
}
\keyword{internal}
