% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_config.R
\name{.sits_config_s2_aws_bucket}
\alias{.sits_config_s2_aws_bucket}
\title{Get the bucket where Sentinel-2 level 2A images are available in AWS}
\usage{
.sits_config_s2_aws_bucket()
}
\value{
name of the bucket
}
\description{
Get the bucket where Sentinel-2 level 2A images are available in AWS
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
\keyword{internal}
