% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_config.R
\name{.sits_config_s2_aws_res}
\alias{.sits_config_s2_aws_res}
\title{Get the the resolutions for Sentinel-2 ARD in AWS}
\usage{
.sits_config_s2_aws_res()
}
\value{
vector with names of the resolutions available in AWS for S2 L2A
}
\description{
Get the the resolutions for Sentinel-2 ARD in AWS
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
\keyword{internal}
