% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_config.R
\name{.sits_config_satellite}
\alias{.sits_config_satellite}
\title{retrieve the satellite associated to a given cube}
\usage{
.sits_config_satellite(cube)
}
\arguments{
\item{cube}{Name of the data cube}
}
\value{
List of providers associated to a service
}
\description{
retrieve the satellite associated to a given cube
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
