% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_config.R
\name{.sits_config_satellite_sensor}
\alias{.sits_config_satellite_sensor}
\alias{.sits_config_local_satellite_sensor}
\title{Tests is satellite and sensor are known to SITS}
\usage{
.sits_config_local_satellite_sensor(satellite, sensor)
}
\arguments{
\item{satellite}{Name of the satellite}

\item{sensor}{Name of the sensor}
}
\value{
A logical value
}
\description{
Tests is satellite and sensor are known to SITS
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
\keyword{internal}
