% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_config.R
\name{.sits_config_satellites}
\alias{.sits_config_satellites}
\title{List the satellites supported by the configuration file}
\usage{
.sits_config_satellites()
}
\value{
List of satellites supported by SITS
}
\description{
List the satellites supported by the configuration file
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
\keyword{internal}
