% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_config.R
\name{.sits_config_satveg_bbox}
\alias{.sits_config_satveg_bbox}
\title{Retrieve the bounding box for the product available at service}
\usage{
.sits_config_satveg_bbox(name)
}
\arguments{
\item{name}{Name of the cube.}
}
\value{
The bounding box.
}
\description{
Retrieve the bounding box for the product available at service
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
