% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_config.R
\name{.sits_config_satveg_projection}
\alias{.sits_config_satveg_projection}
\title{Retrieve the projection for SATVEG service}
\usage{
.sits_config_satveg_projection(name)
}
\arguments{
\item{name}{Name of the cube.}
}
\value{
CRS PROJ4 information.
}
\description{
Retrieve the projection for SATVEG service
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
\keyword{internal}
