% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_config.R
\name{.sits_config_satveg_size}
\alias{.sits_config_satveg_size}
\title{Retrieve the size of the cube for a given service}
\usage{
.sits_config_satveg_size(name, r_obj = NA)
}
\arguments{
\item{name}{Name of the cube.}

\item{r_obj}{R object associated with the cube.}
}
\value{
Vector of (nrows, ncols).
}
\description{
Retrieve the size of the cube for a given service
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
