% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_config.R
\name{.sits_config_satveg_url}
\alias{.sits_config_satveg_url}
\title{Get the URL to be used for SATVEG access}
\usage{
.sits_config_satveg_url()
}
\value{
URL to test SATVEG access
}
\description{
Get the URL to be used for SATVEG access
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
\keyword{internal}
