% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_config.R
\name{.sits_config_scale_factors}
\alias{.sits_config_scale_factors}
\title{Retrieve the scale factor for a given band for a data cube}
\usage{
.sits_config_scale_factors(sensor, bands)
}
\arguments{
\item{sensor}{Name of the sensor.}

\item{bands}{Vector of bands.}
}
\value{
Vector of scale factors.
}
\description{
Retrieve the scale factor for a given band for a data cube
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
\keyword{internal}
