% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_config.R
\name{.sits_config_sensor}
\alias{.sits_config_sensor}
\title{retrieve the sensor associated to a data cube}
\usage{
.sits_config_sensor(cube)
}
\arguments{
\item{cube}{Name of the data cube}
}
\value{
List of providers associated to a service
}
\description{
retrieve the sensor associated to a data cube
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
