% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_config.R
\name{.sits_config_sensor_bands}
\alias{.sits_config_sensor_bands}
\title{Retrieve the bands associated to DEAfrica STAC}
\usage{
.sits_config_sensor_bands(sensor, source)
}
\arguments{
\item{sensor}{Type of sensor of cube}

\item{source}{Data source}
}
\value{
Names of bands available for sensor in data source
}
\description{
Retrieve the cubes associated to the STAC service
}
\keyword{internal}
