% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_config.R
\name{.sits_config_sentinel_aws_bands}
\alias{.sits_config_sentinel_aws_bands}
\title{Get the the bands stored in AWS for Sentinel-2 ARD given the resolution}
\usage{
.sits_config_sentinel_aws_bands(resolution)
}
\arguments{
\item{resolution}{Resolution of the bands}
}
\value{
vector with names of the bands available in AWS for a given resolution
}
\description{
Get the the bands stored in AWS for Sentinel-2 ARD given the resolution
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
