% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_config.R
\name{.sits_config_sentinel_aws_test_file}
\alias{.sits_config_sentinel_aws_test_file}
\title{File to test access to Sentinel-2 level 2A images in AWS}
\usage{
.sits_config_sentinel_aws_test_file()
}
\value{
name of the bucket
}
\description{
File to test access to Sentinel-2 level 2A images in AWS
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
