% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_config.R
\name{.sits_config_stac}
\alias{.sits_config_stac}
\title{Directory to read the DEAFRICA STAC catalogue}
\usage{
.sits_config_stac(source)
}
\value{
directory where DEAFRICA is accessible on the web
}
\description{
Directory to read the DEAFRICA STAC catalogue
}
\keyword{internal}
