% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_config.R
\name{.sits_config_tcap_brightness}
\alias{.sits_config_tcap_brightness}
\title{Retrieve the vector of coeficientes for brightness of tasseled cap}
\usage{
.sits_config_tcap_brightness(sensor = "MODIS")
}
\arguments{
\item{sensor}{Name of sensor}
}
\value{
Named vector of brightness coefficients.
}
\description{
Retrieve the vector of coeficientes for brightness of tasseled cap
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
