% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_config.R
\name{.sits_config_tcap_greenness}
\alias{.sits_config_tcap_greenness}
\title{Retrieve the vector of coeficientes for greeness of tasseled cap}
\usage{
.sits_config_tcap_greenness(sensor = "MODIS")
}
\arguments{
\item{sensor}{Name of ssatellite (or sensor).}
}
\value{
Named vector of greenness coefficients.
}
\description{
Retrieve the vector of coeficientes for greeness of tasseled cap
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
