% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_config.R
\name{.sits_config_tcap_wetness}
\alias{.sits_config_tcap_wetness}
\title{Retrieve the vector of coeficientes for wetness component of tasseled cap}
\usage{
.sits_config_tcap_wetness(sensor = "MODIS")
}
\arguments{
\item{sensor}{Name of sensor.}
}
\value{
Named vector of wetness coefficients.
}
\description{
Retrieve the vector of coeficientes for wetness component of tasseled cap
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
