% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_config.R
\name{.sits_config_test_file}
\alias{.sits_config_test_file}
\title{File to test access to cloud services}
\usage{
.sits_config_test_file(source)
}
\arguments{
\item{source}{Data source}
}
\value{
name of the bucket
}
\description{
File to test access to cloud services
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
\keyword{internal}
