% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_config.R
\name{.sits_config_timeline}
\alias{.sits_config_timeline}
\title{Retrieve the default timeline based on the configuration file}
\usage{
.sits_config_timeline(service, name)
}
\arguments{
\item{service}{Name of the web service.}

\item{name}{Name of the product (a cube).}
}
\value{
The default timeline associated to the service
}
\description{
Retrieve the default timeline based on the configuration file
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
