% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_config.R
\name{.sits_config_types}
\alias{.sits_config_types}
\title{List the data services available in the configuration file}
\usage{
.sits_config_types()
}
\value{
List of services supported by SITS
}
\description{
List the data services available in the configuration file
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
