% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_config.R
\name{.sits_config_usgs_stac}
\alias{.sits_config_usgs_stac}
\title{Directory to read the USGS STAC catalogue}
\usage{
.sits_config_usgs_stac()
}
\value{
directory where USGS is accessible on the web
}
\description{
Directory to read the USGS STAC catalogue
}
\keyword{internal}
