% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_config.R
\name{.sits_config_wtss_bdc}
\alias{.sits_config_wtss_bdc}
\title{Directory to read the BDC WTSS server}
\usage{
.sits_config_wtss_bdc()
}
\value{
directory where BDC is accessible on the web
}
\description{
Directory to read the BDC WTSS server
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
\keyword{internal}
