% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_bbox.R
\name{.sits_coords_to_bbox_wgs84}
\alias{.sits_coords_to_bbox_wgs84}
\title{Convert coordinates to bounding box}
\usage{
.sits_coords_to_bbox_wgs84(xmin, xmax, ymin, ymax, crs)
}
\arguments{
\item{xmin}{Minimum X coordinate}

\item{xmax}{Maximum X coordinate}

\item{ymin}{Minimum Y coordinate}

\item{ymax}{Maximum Y coordinate}

\item{crs}{Projection for X,Y coordinates}
}
\value{
Coordinates in WGS84.
}
\description{
Convert coordinates to bounding box
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
\keyword{internal}
