% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_coverage.R
\name{.sits_coverage_eocubes}
\alias{.sits_coverage_eocubes}
\alias{.sits_coverage_EOCUBES}
\title{Provides information about one coverage of the EOCUBES}
\usage{
.sits_coverage_EOCUBES(remote.obj, service, name, bands, tiles, geom, from,
  to)
}
\arguments{
\item{remote.obj}{Remote object.}

\item{service}{Service string.}

\item{name}{Name of the cube}

\item{bands}{Bands of the coverage.}

\item{tiles}{Filter tiles by prefix name.}

\item{geom}{Geometry to filter tiles.}

\item{from}{Start date to be filtered.}

\item{to}{End date to be filtered.}
}
\description{
Creates a tibble with metadata about a given coverage.
}
