% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_coverage.R
\name{.sits_coverage_satveg}
\alias{.sits_coverage_satveg}
\title{Provides information about one coverage of the SATVEG time series service}
\usage{
.sits_coverage_satveg(name, timeline)
}
\arguments{
\item{name}{Name of the coverage.}

\item{timeline}{Timeline of the coverage.}
}
\description{
Creates a tibble with metadata about a given coverage.
}
