% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_coverage.R
\name{.sits_coverage_wtss}
\alias{.sits_coverage_wtss}
\title{Provides information about one coverage of the WTSS service}
\usage{
.sits_coverage_wtss(wtss.obj, service, name)
}
\arguments{
\item{wtss.obj}{R WTSS object associated to the service.}

\item{service}{Name of the service.}

\item{name}{Name of the coverage.}
}
\description{
Uses the WTSS services to print information and save metadata about a
chosen coverage.
}
