% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_coverage.R
\name{.sits_create_coverage}
\alias{.sits_create_coverage}
\title{Creates a coverage metadata}
\usage{
.sits_create_coverage(r_objs.lst, name, service, bands.vec, labels.vec,
  scale_factors.vec, missing_values.vec, minimum_values.vec, timeline.lst,
  nrows, ncols, xmin, xmax, ymin, ymax, xres, yres, crs, files.vec)
}
\arguments{
\item{r_objs.lst}{list of raster objects contained in the coverage}

\item{name}{name of the coverage}

\item{service}{name of the time series service}

\item{bands.vec}{vector with the names of the bands}

\item{labels.vec}{vector with labels (only valid for classified data)}

\item{scale_factors.vec}{vector with scale factor for each band}

\item{missing_values.vec}{vector with missing values for each band}

\item{minimum_values.vec}{vector with minimum values for each band}

\item{timeline.lst}{list with vectors of valid timelines for each band}

\item{nrows}{number of rows in the coverage}

\item{ncols}{number of columns in the coverage}

\item{xmin}{spatial extent (xmin)}

\item{xmax}{spatial extent (xmax)}

\item{ymin}{spatial extent (ymin)}

\item{ymax}{spatial extent (ymin)}

\item{xres}{spatial resolution (x dimension)}

\item{yres}{spatial resolution (y dimension)}

\item{crs}{CRS for coverage}

\item{files.vec}{vector with associated files}
}
\description{
uses the configuration file to print information and save metadata about a
chosen coverage
}
