% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_cube.R
\name{.sits_create_cube}
\alias{.sits_create_cube}
\title{Creates the description of a data cube}
\usage{
.sits_create_cube(r_objs.lst, name, service, bands.vec, labels.vec,
  scale_factors.vec, missing_values.vec, minimum_values.vec,
  maximum_values.vec, timeline.lst, nrows, ncols, xmin, xmax, ymin, ymax,
  xres, yres, crs, files.vec)
}
\arguments{
\item{r_objs.lst}{List of raster objects contained in the cube.}

\item{name}{Name of the data cube.}

\item{service}{Name of the web service that has provided metadata about the cube.}

\item{bands.vec}{Vector with the names of the bands.}

\item{labels.vec}{Vector with labels (only valid for classified data).}

\item{scale_factors.vec}{Vector with scale factor for each band.}

\item{missing_values.vec}{Vector with missing values for each band.}

\item{minimum_values.vec}{Vector with minimum values for each band.}

\item{maximum_values.vec}{Vector with maximum values for each band.}

\item{timeline.lst}{List with vectors of valid timelines for each band.}

\item{nrows}{Number of rows in the cube.}

\item{ncols}{Number of columns in the cube.}

\item{xmin}{Spatial extent (xmin).}

\item{xmax}{Spatial extent (xmax).}

\item{ymin}{Spatial extent (ymin).}

\item{ymax}{Spatial extent (ymin).}

\item{xres}{Spatial resolution (x dimension).}

\item{yres}{Spatial resolution (y dimension).}

\item{crs}{CRS for cube.}

\item{files.vec}{Vector with associated files.}
}
\description{
Uses the configuration file to print information and save metadata about a
data cube.
}
