% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_tibble_utils.R
\name{.sits_create_folds}
\alias{.sits_create_folds}
\title{Create partitions of a data set}
\usage{
.sits_create_folds(data.tb, folds = 5)
}
\arguments{
\item{data.tb}{SITS tibble to be partitioned}

\item{folds}{number of folds}
}
\description{
Split a SITS table into k groups, based on the label
}
\author{
Rolf Simoes, \email{rolf.simoes@inpe.br}

Alexandre Ywata, \email{alexandre.ywata@ipea.gov.br}

Gilberto Camara, \email{gilberto.camara@inpe.br}
}
