% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_metrics_generate.R
\name{.sits_create_metrics_bands}
\alias{.sits_create_metrics_bands}
\title{Create the output of the cloud estimation procedure}
\usage{
.sits_create_metrics_bands(
  cube,
  data_dir,
  blocks,
  list_metrics,
  impute_fn,
  multicores
)
}
\arguments{
\item{cube}{input data cube}

\item{data_dir}{directory where data is to be stored}

\item{blocks}{block information}

\item{list_metrics}{...}

\item{impute_fn}{imputation function to remove NA}

\item{multicores}{number of cores to use}
}
\value{
a tibble with date, band and path information.
}
\description{
Create the output of the cloud estimation procedure
}
\keyword{internal}
