% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_coverage.R
\name{.sits_create_raster_coverage}
\alias{.sits_create_raster_coverage}
\title{Creates a tibble with information about a set of raster bricks}
\usage{
.sits_create_raster_coverage(raster.lst, service, name, timeline.lst,
  bands.vec, labels.vec, scale_factors.vec, missing_values.vec,
  minimum_values.vec, files.vec)
}
\arguments{
\item{raster.lst}{list of Raster objects associated with the raster coverages}

\item{service}{time series service}

\item{name}{name of the coverage}

\item{timeline.lst}{list - coverage timelines}

\item{bands.vec}{vector - names of bands}

\item{labels.vec}{vector - labels for classified image}

\item{scale_factors.vec}{vector - scale factors}

\item{missing_values.vec}{vector - missing values}

\item{minimum_values.vec}{vector - minimum values}

\item{files.vec}{vector - names of raster files where the data is stored}
}
\description{
creates a tibble with metadata about a given coverage
}
