% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_cube.R
\name{.sits_create_raster_cube}
\alias{.sits_create_raster_cube}
\title{Creates a tibble with information about a data cube based on Raster objects}
\usage{
.sits_create_raster_cube(raster.lst, service, name, timeline.lst,
  bands.vec, labels.vec, scale_factors.vec, missing_values.vec,
  minimum_values.vec, maximum_values.vec, files.vec)
}
\arguments{
\item{raster.lst}{List of Raster objects associated with the raster cubes.}

\item{service}{Time series service.}

\item{name}{Name of the cube.}

\item{timeline.lst}{List of cube timelines.}

\item{bands.vec}{Vector with names of bands.}

\item{labels.vec}{Vector of labels for classified image.}

\item{scale_factors.vec}{Vector of scale factors.}

\item{missing_values.vec}{Vector of missing values.}

\item{minimum_values.vec}{Vector of minimum values.}

\item{maximum_values.vec}{Vector of maximum values.}

\item{files.vec}{Vector of names of raster files where the data is stored.}
}
\description{
Creates a tibble with metadata about a given cube.
}
