% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_csv.R
\name{.sits_csv_check}
\alias{.sits_csv_check}
\title{Check if a CSV tibble is valid}
\usage{
.sits_csv_check(csv.tb)
}
\arguments{
\item{csv.tb}{Tibble read from a CSV file}
}
\value{
TRUE/FALSE
}
\description{
Check if a CSV tibble is valid
}
