% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_csv.R
\name{.sits_csv_from_raster}
\alias{.sits_csv_from_raster}
\title{Read a CSV from a raster}
\usage{
.sits_csv_from_raster(cube, csv.tb, bands)
}
\arguments{
\item{cube}{Data cube}

\item{csv.tb}{Tibble with CSV information}

\item{bands}{Vector with bands to be retrieved}
}
\value{
Tibble with requested samples
}
\description{
Read a CSV from a raster
}
