% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_csv.R
\name{.sits_csv_from_service}
\alias{.sits_csv_from_service}
\title{Read a CSV from a webservice}
\usage{
.sits_csv_from_service(cube, csv.tb, bands, .prefilter)
}
\arguments{
\item{cube}{Data cube}

\item{csv.tb}{Tibble with CSV information}

\item{bands}{Vector with bands to be retrieved}

\item{.prefilter}{Prefilter info (used for SATVEG)}
}
\description{
Read a CSV from a webservice
}
