% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_cube_aux_functions.R
\name{.sits_cube_align_bands}
\alias{.sits_cube_align_bands}
\title{Align the bands of the cube with those of the samples}
\usage{
.sits_cube_align_bands(cube, sample_bands)
}
\arguments{
\item{cube}{Metadata about a data cube}

\item{sample_bands}{Bands of the data sample}
}
\value{
Updated cube
}
\description{
Given a data cube, retrieve the scale factors
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
\keyword{internal}
