% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_cube.R
\name{.sits_cube_all_robjs}
\alias{.sits_cube_all_robjs}
\title{Return the Raster objects associated to a data cube}
\usage{
.sits_cube_all_robjs(cube)
}
\arguments{
\item{cube}{Metadata about a data cube}
}
\value{
list of raster object associated to the indexed file
}
\description{
Given a data cube, retrieve the Raster objects
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
