% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_cube_aux_functions.R
\name{.sits_cube_bands}
\alias{.sits_cube_bands}
\title{Find the bands associated to a cube}
\usage{
.sits_cube_bands(cube)
}
\arguments{
\item{cube}{Metadata about a data cube}
}
\value{
Vector of bands available in the data cube
}
\description{
Given a data cube, retrieves the bands
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
\keyword{internal}
