% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_cube_aux_functions.R
\name{.sits_cube_bands_check}
\alias{.sits_cube_bands_check}
\title{Check that the requested bands exist in the cube}
\usage{
.sits_cube_bands_check(cube, bands = NULL)
}
\arguments{
\item{cube}{Metadata about a data cube}

\item{bands}{Requested bands of the data sample}
}
\value{
Checked bands (cube bands if bands are NULL)
}
\description{
Check that the requested bands exist in the cube
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
