% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_cube_aux_functions.R
\name{.sits_cube_bands_set}
\alias{.sits_cube_bands_set}
\title{Set the bands associated to a cube}
\usage{
.sits_cube_bands_set(cube, bands)
}
\arguments{
\item{cube}{Metadata about a data cube}

\item{bands}{Bands to be assigned to the cube}
}
\value{
Vector of bands available in the data cube
}
\description{
Given a data cube and a set of bands, sets the bands
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
\keyword{internal}
