% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_cube_aux_functions.R
\name{.sits_cube_check_validity}
\alias{.sits_cube_check_validity}
\title{Check that the cube is valid}
\usage{
.sits_cube_check_validity(cube)
}
\arguments{
\item{cube}{Metadata about a data cube}
}
\value{
Boolean value
}
\description{
Given a data cube, retrieve the scale factors
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
\keyword{internal}
