% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_cube_aux_functions.R
\name{.sits_cube_class_band_name}
\alias{.sits_cube_class_band_name}
\title{Define a name for classified band}
\usage{
.sits_cube_class_band_name(name, type, start_date, end_date)
}
\arguments{
\item{name}{original cube name (without temporal information).}

\item{type}{type of output}

\item{start_date}{starting date of the time series classification.}

\item{end_date}{end date of the time series classification.}
}
\value{
name of the classification file for the required interval.
}
\description{
Creates a name for a raster layer based on timeline
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
\keyword{internal}
