% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_cube.R
\name{.sits_cube_classified}
\alias{.sits_cube_classified}
\title{Create a set of RasterLayer objects to store data cube classification results}
\usage{
.sits_cube_classified(raster.tb, samples.tb, file, interval)
}
\arguments{
\item{raster.tb}{Tibble with metadata about the input data cube.}

\item{samples.tb}{Samples used for training the classification model.}

\item{file}{Generic name of the files that will contain the RasterLayers.}

\item{interval}{Classification interval.}
}
\value{
A tibble with metadata about the output RasterLayer objects.
}
\description{
Takes a tibble containing metadata about a data cube
containing time series (each Brick has information for one band) and creates a
set of RasterLayers to store the classification result. Each RasterLayer corresponds
to one time step. The time steps are specified in a list of dates.
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
