% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_cube_aux_functions.R
\name{.sits_cube_clone}
\alias{.sits_cube_clone}
\title{Clone a data cube}
\usage{
.sits_cube_clone(cube, name, ext, output_dir, version)
}
\arguments{
\item{cube}{input data cube}

\item{name}{name of the new cube}

\item{ext}{file extension}

\item{output_dir}{prefix of the output files}

\item{version}{version of the output files}
}
\value{
output data cube
}
\description{
Clone a data cube
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
\keyword{internal}
