% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_cube.R
\name{.sits_cube_file}
\alias{.sits_cube_file}
\title{Return a file associated to a data cube, given an index}
\usage{
.sits_cube_file(cube, index = 1)
}
\arguments{
\item{cube}{Metadata about a data cube}

\item{index}{Index for file to be retrived}
}
\value{
Name of file
}
\description{
Given a data cube and an index, retrieve the file
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
