% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_cube_aux_functions.R
\name{.sits_cube_files}
\alias{.sits_cube_files}
\title{Return all file associated to a data cube}
\usage{
.sits_cube_files(cube)
}
\arguments{
\item{cube}{Metadata about a data cube}
}
\value{
Vector of files
}
\description{
Given a data cube and an index, retrieve the files
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
