% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_cube.R
\name{.sits_cube_files_set}
\alias{.sits_cube_files_set}
\title{Sets the files associated to a data cube}
\usage{
.sits_cube_files_set(cube, files)
}
\arguments{
\item{cube}{Metadata about a data cube}

\item{files}{A vector of files to be assigned to the cube}
}
\value{
Vector of files
}
\description{
Given a data cube and an index, retrieve the files
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
