% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_cube_aux_functions.R
\name{.sits_cube_fix_name}
\alias{.sits_cube_fix_name}
\title{Fix cube name uniqueness}
\usage{
.sits_cube_fix_name(cube)
}
\arguments{
\item{cube}{input data cube}
}
\value{
A data cube
}
\description{
Fix cube name uniqueness
}
\author{
Rolf Simoes, \email{rolf.simoes@inpe.br}
}
\keyword{internal}
