% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_cube.R
\name{.sits_cube_labelled}
\alias{.sits_cube_labelled}
\title{Create a set of RasterLayer objects to store data cube classification results (labelled classes)}
\usage{
.sits_cube_labelled(cube_probs, smoothing, output_dir, version)
}
\arguments{
\item{cube_probs}{Tibble with metadata about the input data cube (probability).}

\item{smoothing}{(optional) smoothing method to be applied ("none", "bayesian", "majority")}

\item{output_dir}{Output directory where to put the files}

\item{version}{Name of the version of the result}
}
\value{
A tibble with metadata about the output RasterLayer objects.
}
\description{
Takes a tibble containing metadata about a data cube wuth
classification probabilites and and creates a
set of RasterLayers to store the classification result. Each RasterLayer corresponds
to one time step. The time steps are specified in a list of dates.
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
