% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_cube.R
\name{.sits_cube_labels}
\alias{.sits_cube_labels}
\title{Return all labels associated to a data cube}
\usage{
.sits_cube_labels(cube)
}
\arguments{
\item{cube}{Metadata about a data cube}
}
\value{
Vector of labels
}
\description{
Given a data cube, retrieve the ;abels
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
