% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_cube_aux_functions.R
\name{.sits_cube_maximum_values}
\alias{.sits_cube_maximum_values}
\title{Retrieve the minimum values for a data cube}
\usage{
.sits_cube_maximum_values(cube)
}
\arguments{
\item{cube}{Metadata about a data cube}
}
\value{
Vector of maximum values.
}
\description{
Given a data cube, retrieve the maximum values
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
