% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_cube.R
\name{.sits_cube_minimum_values}
\alias{.sits_cube_minimum_values}
\title{Retrieve the minimum values for a data cube}
\usage{
.sits_cube_minimum_values(cube)
}
\arguments{
\item{cube}{Metadata about a data cube}
}
\value{
Vector of minimum values.
}
\description{
Given a data cube, retrieve the minimum values
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
