% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_cube_aux_functions.R
\name{.sits_cube_missing_values}
\alias{.sits_cube_missing_values}
\title{Retrieve the missing values for a data cube}
\usage{
.sits_cube_missing_values(cube)
}
\arguments{
\item{cube}{Metadata about a data cube}
}
\value{
Vector of missing values.
}
\description{
Given a data cube, retrieve the missing values
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
\keyword{internal}
