% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_cube.R
\name{.sits_cube_robj}
\alias{.sits_cube_robj}
\title{Given a file index, return the associated Raster object}
\usage{
.sits_cube_robj(cube, index = 1)
}
\arguments{
\item{cube}{Metadata about a data cube}

\item{index}{Index for file to be retrived}
}
\value{
Raster object associated to the indexed file
}
\description{
Given a data cube, retrieve the timeline
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
