% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_cube.R
\name{.sits_cube_robj_band}
\alias{.sits_cube_robj_band}
\title{Given a band, return the associated Raster object for the cube}
\usage{
.sits_cube_robj_band(cube, band_cube)
}
\arguments{
\item{cube}{Metadata about a data cube}

\item{band_cube}{Name of the band to the retrieved}
}
\value{
Raster object associated to the indexed file
}
\description{
Given a data cube, retrieve the timeline
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
