% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_cube.R
\name{.sits_cube_satveg}
\alias{.sits_cube_satveg}
\title{Provides information about one cube of the SATVEG time series service}
\usage{
.sits_cube_satveg(name)
}
\arguments{
\item{name}{Name of the cube.}
}
\description{
Creates a tibble with metadata about a given cube.
}
