% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_cube.R
\name{.sits_cube_scale_factors}
\alias{.sits_cube_scale_factors}
\title{Retrieve the scale factors for a data cube}
\usage{
.sits_cube_scale_factors(cube)
}
\arguments{
\item{cube}{Metadata about a data cube}
}
\value{
Vector of scale factors
}
\description{
Given a data cube, retrieve the scale factors
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
