% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_cube.R
\name{.sits_cube_service}
\alias{.sits_cube_service}
\title{Find the web service associated to a cube}
\usage{
.sits_cube_service(cube)
}
\arguments{
\item{cube}{Metadata about a data cube}
}
\value{
Name of web service
}
\description{
Given a data cube, informs the web services
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
