% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_cube_aux_functions.R
\name{.sits_cube_source}
\alias{.sits_cube_source}
\title{Get cube source}
\usage{
.sits_cube_source(cube)
}
\arguments{
\item{cube}{input data cube}
}
\value{
A character string
}
\description{
Get cube source
}
\author{
Rolf Simoes, \email{rolf.simoes@inpe.br}
}
\keyword{internal}
