% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_cube.R
\name{.sits_cube_stack}
\alias{.sits_cube_stack}
\title{Create a data cube based on a set of Raster Stacks}
\usage{
.sits_cube_stack(name, timeline.vec, bands.vec, scale_factors.vec,
  missing_values.vec, minimum_values.vec, maximum_values.vec, files.lst)
}
\arguments{
\item{name}{Name of the cube file.}

\item{timeline.vec}{Vector of dates with the timeline of the bands.}

\item{bands.vec}{Vector of bands contained in the Raster Brick set (in the same order as the files).}

\item{scale_factors.vec}{Vector of scale factors (one per band).}

\item{missing_values.vec}{Vector of missing values (one per band).}

\item{minimum_values.vec}{Minimum values for each band (only for raster data).}

\item{maximum_values.vec}{Maximum values for each band (only for raster data).}

\item{files.lst}{List of vectors with the file paths of the raster files.}
}
\value{
A tibble with metadata information about a data cube.
}
\description{
This function creates a tibble containing the metadata for
              a set of spatio-temporal raster files, organized as a set of "Raster Bricks".
              These files should be of the same size and
              projection. Each raster brick file should contain one band
              per time step. Different bands are archived in different raster files.
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
