% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_cube_aux_functions.R
\name{.sits_cube_sub_interval}
\alias{.sits_cube_sub_interval}
\title{Extract a temporal interval from the cube to select
only the images to be used for classification}
\usage{
.sits_cube_sub_interval(cube, samples, start_date, end_date)
}
\arguments{
\item{cube}{Metadata about a data cube tile}

\item{samples}{Samples used for training}

\item{start_date}{Starting date for the classification}

\item{end_date}{End date for the classification}
}
\value{
Name of file
}
\description{
Extract a temporal interval from the cube to select
only the images to be used for classification
}
\author{
Rolf Simoes, \email{rolf.simoes@inpe.br}

Gilberto Camara, \email{gilberto.camara@inpe.br}
}
\keyword{internal}
