% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_cube_aux_functions.R
\name{.sits_cube_update_raster_params}
\alias{.sits_cube_update_raster_params}
\title{Update informations of a data cube}
\usage{
.sits_cube_update_raster_params(cube)
}
\arguments{
\item{cube}{input data cube}
}
\value{
output data cube
}
\description{
Update informations of a data cube
}
\author{
Rolf Simoes, \email{rolf.simoes@inpe.br}
}
\keyword{internal}
