% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_cube.R
\name{.sits_cube_wtss}
\alias{.sits_cube_wtss}
\title{Provides information about one cube of the WTSS service}
\usage{
.sits_cube_wtss(wtss.obj, service, name, bands)
}
\arguments{
\item{wtss.obj}{R WTSS object associated to the service.}

\item{service}{Name of the service.}

\item{name}{Name of the cube.}

\item{bands}{Name of the bands.}
}
\description{
Uses the WTSS services to print information and save metadata about a
chosen cube.
}
