% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_db.R
\name{.sits_db_write_ts}
\alias{.sits_db_write_ts}
\title{Write time series on an SQLite database}
\usage{
.sits_db_write_ts(conn, name, data)
}
\arguments{
\item{conn}{SQLite connection}

\item{name}{Name of the object to be storeed}

\item{data}{Data to be stored}
}
\value{
A connection to an RSQLite database
}
\description{
This functions write a set of time series
to an RSQLite database.
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
\keyword{internal}
