% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_deafrica.R
\name{.sits_deafrica_search_tiles}
\alias{.sits_deafrica_search_tiles}
\title{Search items tiles}
\usage{
.sits_deafrica_search_tiles(items, tiles)
}
\arguments{
\item{items}{a \code{STACItemCollection} object returned by rstac package.
grouped.}

\item{tiles}{a \code{character} with the names of the tiles.}
}
\value{
a \code{STACItemCollection} object representing the search
             by rstac.
}
\description{
Search items tiles
}
\keyword{internal}
