% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_deafrica.R
\name{.sits_deafrica_upper}
\alias{.sits_deafrica_upper}
\title{Converts bands name to upper case}
\usage{
.sits_deafrica_upper(items)
}
\arguments{
\item{items}{a \code{STACItemCollection} object returned by rstac package.
grouped.}
}
\value{
a \code{STACItemCollection} object representing the search
             by rstac.
}
\description{
Converts bands name to upper case
}
\keyword{internal}
