% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_dendrogram.R
\name{.sits_dendro_bestcut}
\alias{.sits_dendro_bestcut}
\title{Compute validity indexes to a range of cut height}
\usage{
.sits_dendro_bestcut(data.tb, dendro.obj)
}
\arguments{
\item{data.tb}{Tibble used to generate `dendro.obj`.}

\item{dendro.obj}{Dendrogram object returned from \code{\link[sits]{.sits_dendrogram}}.}
}
\value{
Vector with the best number of clusters (k) and its respective height.
}
\description{
Reads a dendrogram object and its corresponding sits tibble and
computes the best number of clusters that maximizes the adjusted Rand index.
}
\references{
Lawrence Hubert and Phipps Arabie. Comparing partitions.
Journal of Classification, 2, p.193--218, 1985.

See \link[flexclust]{randIndex} for implementation details.
}
\author{
Rolf Simoes, \email{rolf.simoes@inpe.br}
}
