% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_raster_data.R
\name{.sits_derived_data_read}
\alias{.sits_derived_data_read}
\title{Read a block of values retrieved from a derived cube}
\usage{
.sits_derived_data_read(tile, band, block)
}
\arguments{
\item{tile}{Input tile to read data.}

\item{band}{Band to read data from tile.}

\item{block}{Bounding box in (col, row, ncols, nrows).}
}
\value{
A matrix with probability values for labeling.
}
\description{
Read a block of values retrieved from a derived cube
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}

Rolf Simoes, \email{rolf.simoes@inpe.br}
}
\keyword{internal}
