% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_machine_learning.R
\name{.sits_distances}
\alias{.sits_distances}
\title{Use time series values as distances for training patterns}
\usage{
.sits_distances(data)
}
\arguments{
\item{data}{A tibble with time series data and metadata.}
}
\value{
A data.table where columns have the reference label
                   and the time series values as distances.
}
\description{
This function allows using a set of labelled time series as
input to the machine learning models. The attributes used to train the model
are the series themselves. It extracts the time series from a sits tibble
and "spreads" them in time to produce a tibble with distances.
}
\author{
Rolf Simoes, \email{rolf.simoes@inpe.br}

Gilberto Camara, \email{gilberto.camara@inpe.br}
}
