% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_distances.R
\name{.sits_distances_classify}
\alias{.sits_distances_classify}
\title{Classify a distances tibble using machine learning models}
\usage{
.sits_distances_classify(distances, class_info, ml_model, multicores)
}
\arguments{
\item{distances}{data.table with distances.}

\item{class_info}{tibble with classification information.}

\item{ml_model}{model trained by \code{\link[sits]{sits_train}}.}

\item{multicores}{number of threads to process the time series.}
}
\value{
A data.table with the predicted labels.
}
\description{
Returns a sits tibble with the results of the ML classifier.
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
\keyword{internal}
