% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_distances.R
\name{.sits_distances_sample}
\alias{.sits_distances_sample}
\title{Sample a percentage of a time series distance matrix}
\usage{
.sits_distances_sample(distances, frac)
}
\arguments{
\item{distances}{Distances associated to a time series.}

\item{frac}{Percentage of samples to pick.}
}
\value{
Data.table with a fixed quantity of samples
                        of informed labels and all other.
}
\description{
Takes a sits tibble with different labels and
             returns a new tibble. For a given field as a group criterion,
             this new table contains a given number or percentage
             of the total number of samples per group. Parameter n indicates
             the number of random samples with reposition.
             Parameter frac indicates a fraction of random samples
             without reposition. If frac > 1, no sampling is done.
}
\author{
Rolf Simoes, \email{rolf.simoes@inpe.br}
}
\keyword{internal}
