% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_deep_learning.R
\name{.sits_dl_binary_class}
\alias{.sits_dl_binary_class}
\title{Adjust prediction for the binary classification case}
\usage{
.sits_dl_binary_class(prediction_DT)
}
\arguments{
\item{prediction_DT}{Data.table with the predicted values from the keras model
for the binary classification case (one column)}
}
\value{
Data.table with an additional column for multi-class
                         compatibility
}
\description{
For binary classification, the prediction function produces only
one column (the probability of label 1). For compatibility with the rest of the
code in the sits package, this function includes a second column in the prediction
values to match the results of multi-class classification.
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
