% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_deep_learning.R
\name{.sits_dl_prepare_data}
\alias{.sits_dl_prepare_data}
\title{Prepare data for keras model training}
\usage{
.sits_dl_prepare_data(data, validation_split = 0.2, type = "MLP")
}
\arguments{
\item{data}{Time series with the training samples.}

\item{validation_split}{Number between 0 and 1.
Fraction of the training data for validation.}

\item{type}{Type of the model. Valid options:
"MLP" for multi-layer perceptrons and
"1DCNN" for 1D convolutional neural networks.}
}
\value{
List with four elements (training data (X and Y)
                         and test data (X and Y))
}
\description{
Given a training set organised as a SITS tibble, prepare
the data for keras training, providing test and training data,
distinguishing between MLP and 1DCNN models
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
