% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_eocubes.R
\name{.sits_eocubes_check}
\alias{.sits_eocubes_check}
\title{Check that the EOCUBES service parameter are valid}
\usage{
.sits_eocubes_check(URL)
}
\arguments{
\item{URL}{URL of the EOCUBES service.}
}
\value{
eocubes.obj  R object for the EOCUBES service
}
\description{
Check that the EOCUBES service parameter are valid
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
