% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_processing_estimate.R
\name{.sits_est_class_time}
\alias{.sits_est_class_time}
\title{Estimate the processing time}
\usage{
.sits_est_class_time(start_time, n_blocks, block)
}
\arguments{
\item{start_time}{initial processing time.}

\item{n_blocks}{number of blocks}

\item{block}{current block.}
}
\value{
Expected time to complete the function.
}
\description{
This function normalizes image values.
}
\author{
Rolf Simoes, \email{rolf.simoes@inpe.br}
}
\keyword{internal}
