% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_raster_blocks.R
\name{.sits_estimate_nblocks}
\alias{.sits_estimate_nblocks}
\title{Estimate the}
\usage{
.sits_estimate_nblocks(ml_model, nbands, nrows, ncols, timeline, interval,
  memsize, multicores)
}
\arguments{
\item{ml_model}{machine learning model}

\item{nbands}{number of bands}

\item{nrows}{number of rows per brick}

\item{ncols}{number of cols per brick}

\item{timeline}{timeline of the brick}

\item{interval}{classification interval}

\item{memsize}{memory available for classification (in GB)}

\item{multicores}{number of threads to process the time series.}
}
\value{
nblocks         number of blocks to read
}
\description{
Defines the number of blocks of a Raster Brick to be read into memory.
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
