% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_raster_utils.R
\name{.sits_estimate_processing_time}
\alias{.sits_estimate_processing_time}
\title{Estimate the processing time}
\usage{
.sits_estimate_processing_time(start_time, select.lst, bs, block, time)
}
\arguments{
\item{start_time}{initial processing time}

\item{select.lst}{list of time intervals}

\item{bs}{raster block parameters}

\item{block}{current block}

\item{time}{current interval}
}
\value{
values.mx      scaled matrix
}
\description{
this function normalizes one band of the values read from a raster brick
}
\author{
Rolf Simoes, \email{rolf.simoes@inpe.br}
}
