% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_factory.R
\name{.sits_expr_new}
\alias{.sits_expr_new}
\title{Create a closure to generate new data from data and expressions}
\usage{
.sits_expr_new(...)
}
\arguments{
\item{...}{Named parameters containing any valid expression to be
applied later on actual data.}
}
\description{
This internal function implements the factory expression
method.
Its creates a function that process data using expressions

This factory can be used in many situations in the sits package,
to allow user defined expressions to process satellite data.
}
\author{
Rolf Simoes, \email{rolf.simoes@inpe.br}
}
\keyword{internal}
