% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_tibble_utils.R
\name{.sits_extract}
\alias{.sits_extract}
\title{Extract a subset of the data based on dates}
\usage{
.sits_extract(row.tb, start_date, end_date)
}
\arguments{
\item{row.tb}{SITS tibble}

\item{start_date}{starting date of the time series segment}

\item{end_date}{end date of the time series segment}
}
\value{
subset.tb  tibble in SITS format with the chosen subset
}
\description{
returns a vector containing the dates of a sits table
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
