% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_cube.R
\name{.sits_files_robj}
\alias{.sits_files_robj}
\title{Given a vector of files, return the Raster object associated with the first file}
\usage{
.sits_files_robj(files)
}
\arguments{
\item{files}{Vector of files}
}
\value{
Raster object associated to the indexed file
}
\description{
Given a vector of files and index, return the Raster object for the file
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
