% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_utils.R
\name{.sits_filter_intersecting_tiles}
\alias{.sits_filter_intersecting_tiles}
\title{Filter tiles that intersects with samples}
\usage{
.sits_filter_intersecting_tiles(cube, samples)
}
\arguments{
\item{cube}{Data cube from where data is to be retrieved.}

\item{samples}{Samples to be retrieved.}
}
\value{
A cube with filtered tiles.
}
\description{
Filter tiles that intersects with samples.
}
\keyword{internal}
