% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_stac.R
\name{.sits_format_crs}
\alias{.sits_format_crs}
\title{Checks if the crs provided is valid}
\usage{
.sits_format_crs(crs)
}
\arguments{
\item{crs}{a \code{numeric} or \code{character} with CRS provided by
STAC.}
}
\value{
a \code{character} with the formatted CRS.
}
\description{
Checks if the crs provided is valid
}
\keyword{internal}
