% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_eocubes.R
\name{.sits_from_EOCubes}
\alias{.sits_from_EOCubes}
\title{Obtain one timeSeries from EOCubes package}
\usage{
.sits_from_EOCubes(
  cube,
  longitude,
  latitude,
  start_date = NULL,
  end_date = NULL,
  bands = NULL,
  label = "NoClass"
)
}
\arguments{
\item{cube}{Metadata about the data cube to be retrived.}

\item{longitude}{The longitude of the chosen location.}

\item{latitude}{The latitude of the chosen location.}

\item{start_date}{Date with the start of the period.}

\item{end_date}{Date with the end of the period.}

\item{bands}{Names of the bands of the data cube.}

\item{label}{Label to attach to the time series (optional).}
}
\value{
A sits tibble.
}
\description{
Returns one set of time series provided by a EOCubes package
Given a location (lat/long), and start/end period, and the Cube name
retrieve a time series and include it on a sits tibble.
}
