% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_shp.R
\name{.sits_from_point_shp}
\alias{.sits_from_point_shp}
\title{Obtain a tibble with latitude and longitude points from POINT geometry}
\usage{
.sits_from_point_shp(sf_shape, shp_attr, label)
}
\arguments{
\item{sf_shape}{sf object linked to a shapefile.}

\item{shp_attr}{Attribute in the shapefile used as a polygon label}

\item{label}{Label to be assigned to points}
}
\description{
Obtain a tibble with latitude and longitude points from POINT geometry
}
