% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_shp.R
\name{.sits_from_polygon_shp}
\alias{.sits_from_polygon_shp}
\title{Obtain a tibble with latitude and longitude points from POLYGON geometry}
\usage{
.sits_from_polygon_shp(sf_shape, shp_attr, label, .n_shp_pol)
}
\arguments{
\item{sf_shape}{sf object linked to a shapefile}

\item{shp_attr}{Attribute in the shapefile used as a polygon label}

\item{label}{Label to be assigned to points}

\item{.n_shp_pol}{Number of samples per polygon to be read}
}
\description{
Obtain a tibble with latitude and longitude points from POLYGON geometry
}
