% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_satveg.R
\name{.sits_from_satveg}
\alias{.sits_from_satveg}
\title{Obtain one timeSeries from the EMBRAPA SATVEG server and load it on a sits tibble}
\usage{
.sits_from_satveg(cube, longitude, latitude, start_date = NULL,
  end_date = NULL, bands = NULL, prefilter = "1",
  label = "NoClass")
}
\arguments{
\item{cube}{The data cube metadata that describes the SATVEG data.}

\item{longitude}{A double value with the longitude of the chosen location.}

\item{latitude}{A double value with the latitude of the chosen location.}

\item{start_date}{The start date of the period.}

\item{end_date}{The end date of the period.}

\item{bands}{The bands to be retrieved.}

\item{prefilter}{A string ("0" - none, "1" - no data correction, "2" - cloud correction, "3" - no data and cloud correction).}

\item{label}{A string with the label to attach to the time series (optional).}
}
\value{
A sits tibble.
}
\description{
Returns one set of MODIS time series provided by the EMBRAPA server (SATVEG)
Given a location (lat/long), the function retrieves the "ndvi" and "evi" bands from SATVEG
and inclues the data on a stis tibble. If start and end date are given, the function
filter the data to limit the temporal interval.
}
\author{
Julio Esquerdo, \email{julio.esquerdo@embrapa.br}

Gilberto Camara, \email{gilberto.camara@inpe.br}
}
