% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_getdata.R
\name{.sits_from_service}
\alias{.sits_from_service}
\title{Obtain timeSeries from time series service}
\usage{
.sits_from_service(coverage, longitude, latitude, start_date, end_date,
  bands, prefilter = "1", label = "NoClass")
}
\arguments{
\item{coverage}{coverage metadata}

\item{longitude}{double - the longitude of the chosen location}

\item{latitude}{double - the latitude of the chosen location)}

\item{start_date}{(optional) date - the start of the period}

\item{end_date}{(optional) date - the end of the period}

\item{bands}{(optional) string vector - the names of the bands to be retrieved}

\item{prefilter}{string (only for SATVEG) ("0" - none, "1" - no data correction, "2" - cloud correction, "3" - no data and cloud correction)}

\item{label}{string - the label to attach to the time series}
}
\value{
table          a SITS tibble
}
\description{
obtains a time series from a time series service
}
