% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_twdtw.R
\name{.sits_from_twdtw_matches}
\alias{.sits_from_twdtw_matches}
\title{Transform patterns from TWDTW format to sits format}
\usage{
.sits_from_twdtw_matches(match)
}
\arguments{
\item{match}{A TWDTW Matches object of class dtwSat::twdtwMatches}
}
\value{
A list with information on the matches
}
\description{
Reads one TWDTW matches object and transforms it into a
             tibble ready to be stored into a sits tibble column.
}
\keyword{internal}
