% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_wtss.R
\name{.sits_from_wtss}
\alias{.sits_from_wtss}
\title{Obtain one timeSeries from WTSS server and load it on a sits tibble}
\usage{
.sits_from_wtss(coverage, longitude, latitude, start_date = NULL,
  end_date = NULL, bands = NULL, label = "NoClass")
}
\arguments{
\item{coverage}{Metadata about the coverage where the data is to be retrived.}

\item{longitude}{The longitude of the chosen location.}

\item{latitude}{The latitude of the chosen location.}

\item{start_date}{Date with the start of the period.}

\item{end_date}{Date with the end of the period.}

\item{bands}{A list of string with the names of the bands of the coverage.}

\item{label}{A string with the label to attach to the time series (optional).}
}
\value{
A sits tibble.
}
\description{
Returns one set of time series provided by a WTSS server
Given a location (lat/long), and start/end period, and the WTSS server information
retrieve a time series and include it on a stis tibble.
A Web Time Series Service (WTSS) is a light-weight service that
retrieves one or more time series in JSON format from a data base.
}
\references{
Lubia Vinhas, Gilberto Queiroz, Karine Ferreira, Gilberto Camara,
Web Services for Big Earth Observation Data.
In: XVII Brazilian Symposium on Geoinformatics, 2016, Campos do Jordao.
Proceedings of GeoInfo 2016. Sao Jose dos Campos: INPE/SBC, 2016. v.1. p.166-177
}
