% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_getdata.R
\name{.sits_fromSHP}
\alias{.sits_fromSHP}
\title{Obtain timeSeries from WTSS server, based on a SHP file.}
\usage{
.sits_fromSHP(shp_file, coverage, start_date = NULL, end_date = NULL,
  bands = NULL, prefilter = "1", label = "NoClass")
}
\arguments{
\item{shp_file}{string  - name of a SHP file which provides the boundaries of a region of interest}

\item{coverage}{tibble  - metadata about the coverage}

\item{start_date}{date - the start of the period}

\item{end_date}{date - the end of the period}

\item{bands}{string vector - the names of the bands to be retrieved}

\item{prefilter}{string ("0" - none, "1" - no data correction, "2" - cloud correction, "3" - no data and cloud correction)}

\item{label}{string - the label to attach to the time series}
}
\value{
table          a SITS tibble
}
\description{
reads a shapefile and retrieves a SITS tibble
containing time series from a coverage that are inside the SHP file.
The script uses the WTSS service, taking information about coverage, spatial and
temporal resolution from the WTSS configuration.
}
