% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_smooth_aux_functions.R
\name{.sits_gauss_kernel}
\alias{.sits_gauss_kernel}
\title{Compute the 2-D Gaussian kernel}
\usage{
.sits_gauss_kernel(window_size, sigma = 1)
}
\arguments{
\item{window_size}{Size of the neighbourhood.}

\item{sigma}{Standard deviation of the spatial Gaussian kernel}
}
\value{
returns a squared matrix filled with Gaussian function
}
\description{
Compute the 2-D Gaussian kernel
}
\keyword{internal}
