% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_gdalcubes.R
\name{.sits_gc_brick}
\alias{.sits_gc_brick}
\title{Create a raster_cube object}
\usage{
.sits_gc_brick(cube, img_col, cube_view, cloud_mask)
}
\arguments{
\item{cube}{Data cube from where data is to be retrieved.}

\item{img_col}{A \code{object} 'image_collection' containing information
about the images metadata.}

\item{cube_view}{A \code{object} 'cube_view' with values from cube.}

\item{cloud_mask}{A \code{logical} corresponds to the use of the cloud band
for aggregation.}
}
\value{
a \code{object} 'raster_cube' from gdalcubes containing information
 about the cube brick metadata.
}
\description{
Create a raster_cube object
}
\keyword{internal}
