% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_gdalcubes.R
\name{.sits_gc_compose}
\alias{.sits_gc_compose}
\title{Save the images based on an aggregation method.}
\usage{
.sits_gc_compose(
  tile,
  name,
  cv,
  img_col,
  db_file,
  dir_images,
  cloud_mask,
  ...,
  version = "v1"
)
}
\arguments{
\item{tile}{A data cube tile}

\item{name}{Name of the new data cube}

\item{cv}{A \code{list} 'cube_view' with values from cube.}

\item{img_col}{A \code{object} 'image_collection' containing information
about the images metadata.}

\item{db_file}{Database to be created by gdalcubes}

\item{dir_images}{Directory where the aggregated images will be written.}

\item{cloud_mask}{A \code{logical} corresponds to the use of the cloud band
for aggregation.}

\item{...}{Additional parameters that can be included. See
'?gdalcubes::write_tif'.}

\item{version}{A \code{character} with version of the output files.}
}
\value{
A data cube tile with information used in its creation.
}
\description{
Save the images based on an aggregation method.
}
\keyword{internal}
