% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_gdalcubes.R
\name{.sits_gc_database}
\alias{.sits_gc_database}
\title{Create an image_collection object}
\usage{
.sits_gc_database(cube, path_db)
}
\arguments{
\item{cube}{Data cube from where data is to be retrieved.}

\item{path_db}{A \code{character} with the path and name where the
database will be create. E.g. "my/path/gdalcubes.db"}
}
\value{
a \code{object} 'image_collection' containing information about the
 images metadata.
}
\description{
Create an image_collection object
}
\keyword{internal}
