% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_config.R
\name{.sits_get_account}
\alias{.sits_get_account}
\title{Get an account to access a time series service}
\usage{
.sits_get_account(service, name)
}
\arguments{
\item{service}{Name of the time series service.}

\item{name}{Name of the coverage.}
}
\value{
The account for service access.
}
\description{
Get an account to access a time series service
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
