% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_config.R
\name{.sits_get_bands}
\alias{.sits_get_bands}
\title{Retrieve the bands avaliable for the product in the time series service}
\usage{
.sits_get_bands(service, name)
}
\arguments{
\item{service}{Name of the time series service.}

\item{name}{Name of the product.}
}
\value{
The available bands.
}
\description{
Retrieve the bands avaliable for the product in the time series service
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
