% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_config.R
\name{.sits_get_bbox}
\alias{.sits_get_bbox}
\title{Retrieve the bounding box for the product available at service}
\usage{
.sits_get_bbox(service, name, r_obj = NA)
}
\arguments{
\item{service}{Name of the time series service.}

\item{name}{Name of the coverage.}

\item{r_obj}{R object associated with the coverage.}
}
\value{
The bounding box.
}
\description{
Retrieve the bounding box for the product available at service
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
