% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_get_data.R
\name{.sits_get_data_check}
\alias{.sits_get_data_check}
\title{check if all points have been retrieved}
\usage{
.sits_get_data_check(n_rows_input, n_rows_output)
}
\arguments{
\item{n_rows_input}{Number of rows in input}

\item{n_rows_output}{Number of rows in output}
}
\value{
TRUE/FALSE
}
\description{
check if all points have been retrieved
}
\keyword{internal}
