% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_get_data.R
\name{.sits_get_data_from_satveg}
\alias{.sits_get_data_from_satveg}
\title{Obtain one time series from the EMBRAPA SATVEG server}
\usage{
.sits_get_data_from_satveg(
  cube,
  longitude,
  latitude,
  start_date,
  end_date,
  label,
  id
)
}
\arguments{
\item{cube}{The data cube metadata that describes the SATVEG data.}

\item{longitude}{Longitude of the chosen location.}

\item{latitude}{Latitude of the chosen location.}

\item{start_date}{The start date of the period.}

\item{end_date}{The end date of the period.}

\item{label}{Label to attach to the time series (optional).}

\item{id}{Points id in case of shapefile polygons.}
}
\value{
A sits tibble.
}
\description{
Returns one set of MODIS time series provided by the EMBRAPA
Given a location (lat/long), retrieve the "ndvi" and "evi" bands from SATVEG
If start and end date are given, the function
filters the data to limit the temporal interval.
}
\author{
Julio Esquerdo, \email{julio.esquerdo@embrapa.br}

Gilberto Camara, \email{gilberto.camara@inpe.br}
}
\keyword{internal}
