% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_config.R
\name{.sits_get_maximum_values}
\alias{.sits_get_maximum_values}
\title{Retrieve the maximum values for a given band}
\usage{
.sits_get_maximum_values(service, bands)
}
\arguments{
\item{service}{Name of the product.}

\item{bands}{Vector of bands.}
}
\value{
The maximum values.
}
\description{
Retrieve the maximum values for a given band
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
