% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_config.R
\name{.sits_get_minimum_values}
\alias{.sits_get_minimum_values}
\title{Retrieve the minimum values for a given band}
\usage{
.sits_get_minimum_values(service, bands)
}
\arguments{
\item{service}{Name of the product.}

\item{bands}{Vector of bands.}
}
\value{
The minimum values.
}
\description{
Retrieve the minimum values for a given band
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
