% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_config.R
\name{.sits_get_missing_values}
\alias{.sits_get_missing_values}
\title{Retrieve the missing values for a given band for an image product}
\usage{
.sits_get_missing_values(service, name, bands)
}
\arguments{
\item{service}{name of the product}

\item{name}{name of the coverage}

\item{bands}{vector of bands}
}
\value{
missing_values
}
\description{
Retrieve the missing values for a given band for an image product
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
