% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_config.R
\name{.sits_get_protocol}
\alias{.sits_get_protocol}
\title{Retrieve the protocol associated to the time series service}
\usage{
.sits_get_protocol(service)
}
\arguments{
\item{service}{The name of the service.}
}
\value{
The protocol associated to the service.
}
\description{
Retrieve the protocol associated to the time series service
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
