% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_get_data.R
\name{.sits_get_samples_from_csv}
\alias{.sits_get_samples_from_csv}
\title{Transform a shapefile into a samples file}
\usage{
.sits_get_samples_from_csv(csv_file, .n_pts_csv)
}
\arguments{
\item{csv_file}{CSV that describes the data to be retrieved.}

\item{.n_pts_csv}{number of points to be retrived}
}
\value{
A tibble with information the samples to be retrieved
}
\description{
Transform a shapefile into a samples file
}
\author{
Gilberto Camara
}
\keyword{internal}
