% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_sf.R
\name{.sits_get_samples_from_sf}
\alias{.sits_get_samples_from_sf}
\title{Transform an sf object into a samples file}
\usage{
.sits_get_samples_from_sf(
  sf_object,
  label,
  label_attr,
  start_date,
  end_date,
  n_sam_pol,
  pol_id
)
}
\arguments{
\item{sf_object}{sf object that describes the data to be retrieved.}

\item{label}{Default label for samples.}

\item{label_attr}{sf attribute that describes the label.}

\item{start_date}{Start date for the data set.}

\item{end_date}{End date for the data set.}

\item{n_sam_pol}{Number of samples per polygon to be read.}

\item{pol_id}{ID attribute for polygons shapefile.
(for POLYGON or MULTIPOLYGON shapefile).}
}
\value{
A tibble with information the samples to be retrieved.
}
\description{
Transform an sf object into a samples file
}
\author{
Gilberto Camara
}
\keyword{internal}
