% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_get_data.R
\name{.sits_get_samples_from_shp}
\alias{.sits_get_samples_from_shp}
\title{Transform a shapefile into a samples file}
\usage{
.sits_get_samples_from_shp(
  shp_file,
  label,
  shp_attr,
  start_date,
  end_date,
  .n_shp_pol,
  .shp_id
)
}
\arguments{
\item{shp_file}{Shapefile that describes the data to be retrieved.}

\item{label}{Default label for samples.}

\item{shp_attr}{Shapefile attribute that describes the label.}

\item{start_date}{Start date for the data set.}

\item{end_date}{End date for the data set.}

\item{.n_shp_pol}{Number of samples per polygon to be read.}

\item{.shp_id}{ID attribute for polygons shapefile.
(for POLYGON or MULTIPOLYGON shapefile).}
}
\value{
A tibble with information the samples to be retrieved.
}
\description{
Transform a shapefile into a samples file
}
\author{
Gilberto Camara
}
\keyword{internal}
