% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_SATVEG.R
\name{.sits_get_SATVEG_timeline_from_txt}
\alias{.sits_get_SATVEG_timeline_from_txt}
\alias{.sits_getSATVEG_timeline_from_txt}
\title{Retrieve a timeline from the SATVEG service based on text expression}
\usage{
.sits_getSATVEG_timeline_from_txt(satveg.txt)
}
\arguments{
\item{satveg.txt}{Information retrieved from SATVEG (in text format)}
}
\description{
Retrieves a time series from the SATVEG service
}
\author{
Julio Esquerdo, \email{julio.esquerdo@embrapa.br}

Gilberto Camara, \email{gilberto.camara@inpe.br}
}
