% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_config.R
\name{.sits_get_server}
\alias{.sits_get_server}
\title{Retrieve the time series server for the product}
\usage{
.sits_get_server(service, provider = NULL)
}
\arguments{
\item{service}{Name of the data service}

\item{provider}{URL of the service or name of the provider}
}
\value{
A string with the server URL that provides the service.
}
\description{
Retrieve the time series server for the product
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
