% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_config.R
\name{.sits_get_services}
\alias{.sits_get_services}
\title{List the time series services available}
\usage{
.sits_get_services(protocol = NULL)
}
\arguments{
\item{protocol}{(string) protocol used to assess the time series service}
}
\value{
size     vector of (nrows, ncols)
}
\description{
List the time series services available
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
