% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_config.R
\name{.sits_get_size}
\alias{.sits_get_size}
\title{Retrieve the size of the product for a given time series service}
\usage{
.sits_get_size(service, name, r_obj = NA)
}
\arguments{
\item{service}{name of the time series service}

\item{name}{name of the coverage}

\item{r_obj}{R object associated with the coverage}
}
\value{
size          vector of (nrows, ncols)
}
\description{
Retrieve the size of the product for a given time series service
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
