% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_config.R
\name{.sits_get_tcap_brightness}
\alias{.sits_get_tcap_brightness}
\title{Retrieve the vector of coeficientes for brightness component of tasseled cap}
\usage{
.sits_get_tcap_brightness(satellite = "MODIS")
}
\arguments{
\item{satellite}{name of ssatellite (or sensor)}
}
\value{
coef        named vector of brightness coefficients
}
\description{
Retrieve the vector of coeficientes for brightness component of tasseled cap
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
