% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_config.R
\name{.sits_get_tcap_greenness}
\alias{.sits_get_tcap_greenness}
\title{Retrieve the vector of coeficientes for brightness component of tasseled cap}
\usage{
.sits_get_tcap_greenness(satellite = "MODIS")
}
\arguments{
\item{satellite}{Name of ssatellite (or sensor).}
}
\value{
Named vector of greenness coefficients.
}
\description{
Retrieve the vector of coeficientes for brightness component of tasseled cap
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
