% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_timeline.R
\name{.sits_get_time_index}
\alias{.sits_get_time_index}
\title{Find the time index of the blocks to be extracted for classification}
\usage{
.sits_get_time_index(class_info.tb)
}
\arguments{
\item{class_info.tb}{Tibble with information required for classification.}
}
\value{
List with indexes of the input data set associated to each time interval
used for classification.
}
\description{
Obtains the indexes of the blocks to be extract for each time interval
associated with classification.
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
