% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_config.R
\name{.sits_get_timeline}
\alias{.sits_get_timeline}
\title{Retrieve the default timeline for a product for a given data cube}
\usage{
.sits_get_timeline(service, name)
}
\arguments{
\item{service}{Name of the web service.}

\item{name}{Name of the cube.}
}
\value{
Vector of (nrows, ncols).
}
\description{
Retrieve the default timeline for a product for a given data cube
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
