% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_active_learning.R
\name{.sits_get_top_values}
\alias{.sits_get_top_values}
\title{Get top values of a raster.}
\usage{
.sits_get_top_values(r_obj, band, n, sampling_window)
}
\arguments{
\item{r_obj}{A raster object.}

\item{band}{A numeric band index used to read bricks.}

\item{n}{Number of values to extract.}

\item{sampling_window}{Window size to collect a point (in pixels).}
}
\value{
A point `tibble` object.
}
\description{
Get the top values of a raster as a point `sf` object. The values
locations are guaranteed to be separated by a certain number of pixels.
}
\author{
Alber Sanchez, \email{alber.ipia@inpe.br}
}
