% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_get_data.R
\name{.sits_get_ts}
\alias{.sits_get_ts}
\alias{.sits_get_ts.raster_cube}
\alias{.sits_get_ts.classified_image}
\title{Dispatch function to get time series from data cubes and cloud
services}
\usage{
.sits_get_ts(
  cube,
  samples,
  ...,
  bands = NULL,
  impute_fn,
  multicores,
  output_dir,
  progress
)

\method{.sits_get_ts}{raster_cube}(
  cube,
  samples,
  ...,
  bands,
  crs = 4326,
  impute_fn,
  multicores,
  output_dir,
  progress
)

\method{.sits_get_ts}{classified_image}(
  cube,
  samples,
  ...,
  bands,
  crs = 4326,
  impute_fn,
  multicores,
  output_dir,
  progress
)
}
\arguments{
\item{cube}{Data cube from where data is to be retrieved.}

\item{samples}{Samples to be retrieved.}

\item{bands}{Bands to be retrieved (optional).}

\item{impute_fn}{Imputation function for NA values.}

\item{multicores}{Number of threads to process the time series.}

\item{progress}{A logical value indicating if a progress bar
should be shown. Default is \code{FALSE}.}

\item{crs}{A coordinate reference system of samples.
The provided crs could be a character
(e.g, "EPSG:4326" or "WGS84" or a proj4string), or a
a numeric with the EPSG code (e.g. 4326).
This parameter only works for 'csv' or data.frame'
samples. Default is 4326.}
}
\value{
A tibble with a set of time series retrieved
                       from a data cube.
}
\description{
Dispatch function to get time series from data cubes and cloud
services
}
\author{
Gilberto Camara
}
\keyword{internal}
