% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_plot.R
\name{.sits_ggplot_series}
\alias{.sits_ggplot_series}
\title{Plot one timeSeries using ggplot}
\usage{
.sits_ggplot_series(row, colors = "Dark2")
}
\arguments{
\item{row}{a row of a SITS table with the time series to be plotted}

\item{colors}{string - the set of Brewer colors to be used for plotting}
}
\description{
plots a set of time series using ggplot. This function is used
for showing the same lat/long location in a series of time steps.
}
