% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_plot.R
\name{.sits_ggplot_series_na}
\alias{.sits_ggplot_series_na}
\title{Plot one timeSeries wih NAs using ggplot}
\usage{
.sits_ggplot_series_na(row, colors = "Dark2")
}
\arguments{
\item{row}{row of a sits tibble with the time series to be plotted.}

\item{colors}{brewer colors to be used for plotting.}
}
\value{
The plot itself.
}
\description{
Plots a set of time series using ggplot, showing where NAs are.
}
\keyword{internal}
