% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_plot.R
\name{.sits_ggplot_series_no_na}
\alias{.sits_ggplot_series_no_na}
\title{Plot one timeSeries using ggplot (no NAs present)}
\usage{
.sits_ggplot_series_no_na(row, colors = "Dark2")
}
\arguments{
\item{row}{row of a sits tibble with the time series to be plotted.}

\item{colors}{brewer colors to be used for plotting.}
}
\value{
The plot itself.
}
\description{
Plots a set of time series using ggplot in the case the series
             has no NA values.
}
\keyword{internal}
