% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_cube.R
\name{.sits_guess_satellite}
\alias{.sits_guess_satellite}
\title{Try a best guess for the type of sensor/satellite}
\usage{
.sits_guess_satellite(r_obj)
}
\arguments{
\item{r_obj}{A raster object.}
}
\value{
Name of the satellite (or sensor).
}
\description{
Based on the projection, tries to guess what is the satellite.
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
