% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_config.R
\name{.sits_guess_sensor}
\alias{.sits_guess_sensor}
\title{Try a best guess for the type of sensor}
\usage{
.sits_guess_sensor(satellite)
}
\arguments{
\item{satellite}{Name of the sensor}
}
\value{
A best guess for the sensor
}
\description{
Based on the satellite, tries to guess what is the sensor
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
