% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_config.R
\name{.sits_guess_timeline}
\alias{.sits_guess_timeline}
\title{Retrieve the default timeline for a product for a given service}
\usage{
.sits_guess_timeline(service, name)
}
\arguments{
\item{service}{Name of the web service.}

\item{name}{Name of the product (a cube).}
}
\value{
The default timeline associated to the service
}
\description{
Retrieve the default timeline for a product for a given service
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
