% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_clouds.R
\name{.sits_interpolate_clouds}
\alias{.sits_interpolate_clouds}
\title{Interpolate a series of bands based on the cloud information}
\usage{
.sits_interpolate_clouds(cube, data_dir, bricks, blocks, multicores)
}
\arguments{
\item{cube}{input data cube}

\item{data_dir}{data directory where output data is written}

\item{bricks}{list of brick objects pointing to output files}

\item{blocks}{list of blocks to read}

\item{multicores}{number of cores to process}
}
\value{
file_info  file information on the output images
}
\description{
Interpolate a series of bands based on the cloud information
}
