% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_timeline.R
\name{.sits_is_valid_start_date}
\alias{.sits_is_valid_start_date}
\title{Test if starting date fits with the timeline}
\usage{
.sits_is_valid_start_date(date, timeline)
}
\arguments{
\item{date}{a dates}

\item{timeline}{a vector of reference dates}
}
\value{
bool       is this is valid starting date?
}
\description{
A timeline is a list of dates where observations are available. This
functions estimates if a date is valid by comparing it to the timeline. If the date's estimate
is not inside the timeline and the difference between the date and the first date of timeline is
greater than the acquisition interval of the timeline, then we conclude the date is not valid.
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
