% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_keras.R
\name{.sits_keras_prepare_data}
\alias{.sits_keras_prepare_data}
\title{Prepare data for keras model training}
\usage{
.sits_keras_prepare_data(data, validation_split, int_labels, n_bands, n_times)
}
\arguments{
\item{data}{Time series with the training samples.}

\item{validation_split}{Number between 0 and 1.
Fraction of the training data for validation.}

\item{int_labels}{Integer vector named with the labels.}

\item{n_bands}{Number of bands.}

\item{n_times}{Number of time steps.}
}
\value{
List with four elements (training data (X and Y)
                         and test data (X and Y))
}
\description{
Given a training set organised as a SITS tibble, prepare
the data for keras training, providing test and training data.
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
