% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_kohonen.R
\name{.sits_kohonen_pallete}
\alias{.sits_kohonen_pallete}
\title{Configure pallete}
\usage{
.sits_kohonen_pallete()
}
\value{
rreturn a pallete of colors
}
\description{
configure a pallete with 21 colors.
}
\author{
Lorena Santos, \email{lorena.santos@inpe.br}
}
