% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_kohonen.R
\name{.sits_labelling_neurons}
\alias{.sits_labelling_neurons}
\title{Labelling neurons using majority voting}
\usage{
.sits_labelling_neurons(data.tb, grid_size)
}
\arguments{
\item{data.tb}{A SITS tibble with info of samples and kohonen.obj.}

\item{size_grid}{The size of kohonen map.}
}
\value{
returns the labels of each neuron.
}
\description{
Labelling all neurons using the majority voting.
If the neuron is empty, It will labeled as "NoClass".
}
\author{
Lorena Santos, \email{lorena.santos@inpe.br}
}
