% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_labels.R
\name{.sits_labels_list}
\alias{.sits_labels_list}
\title{Sits labels processing function}
\usage{
.sits_labels_list(data, list.lst = list(), fun_label = function(lb) lb)
}
\arguments{
\item{data}{A sits tibble (or a pred_ref tibble)}

\item{list.lst}{Any named list whose names are unique labels
from data input. Non-informed labels will be completed
according to fun_label function.}

\item{fun_label}{A function that will be executed for each label
non listed in list.lst parameter. The result of
 is used as list.lst value for the respective label.}
}
\value{
A list whose non informed values
                    are filled by fun_label for each unique label in data.
}
\description{
Completes list.lst list as a named list
             (names are unique labels from data) according
             to a given function that receives each label as an argument.
}
\author{
Rolf Simoes, \email{rolf.simoes@inpe.br}
}
