% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_space_time_operations.R
\name{.sits_latlong_to_cell}
\alias{.sits_latlong_to_cell}
\title{Coordinate transformation (lat/long to image cell)}
\usage{
.sits_latlong_to_cell(longitude, latitude, cube)
}
\arguments{
\item{longitude}{The longitude of the chosen location.}

\item{latitude}{The latitude of the chosen location.}

\item{cube}{data cube}
}
\value{
Matrix with (x, y) coordinates.
}
\description{
Transform a latitude and longitude coordinate to a XY projection coordinate
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
\keyword{internal}
