% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_space_time_operations.R
\name{.sits_latlong_to_proj}
\alias{.sits_latlong_to_proj}
\title{Coordinate transformation (lat/long to X/Y)}
\usage{
.sits_latlong_to_proj(longitude, latitude, crs)
}
\arguments{
\item{longitude}{The longitude of the chosen location.}

\item{latitude}{The latitude of the chosen location.}

\item{crs}{Projection definition to be converted to.}
}
\value{
Tibble with X and Y coordinates
}
\description{
Transform a latitude and longitude coordinate to XY coordinate
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
\keyword{internal}
