% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_log.R
\name{.sits_log}
\alias{.sits_log}
\title{Log variable content to a file}
\usage{
.sits_log(output_dir = tempdir(), topic = "", ...)
}
\arguments{
\item{output_dir}{Output directory to write log file}

\item{topic}{A description to append to time stamp}

\item{...}{Any valid value to be logged}
}
\description{
Log to a text file the content of a variable passed to values.
Log file name is the same base name of temporary directory for current R
session.
}
\author{
Rolf Simoes, \email{rolf.simoes@inpe.br}
}
\keyword{internal}
