% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_log.R
\name{.sits_log_CSV}
\alias{.sits_log_CSV}
\title{Saves a CSV data set}
\usage{
.sits_log_CSV(csv.tb, file_name = "errors.csv")
}
\arguments{
\item{csv.tb}{Tibble containing CSV data}

\item{file_name}{Name of data to be saved}
}
\description{
Save a CSV data set in the log directory
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
