% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_log.R
\name{.sits_log_data}
\alias{.sits_log_data}
\title{Saves a data set for future use}
\usage{
.sits_log_data(data, file_name = "data_save.rda")
}
\arguments{
\item{data}{Data set to be saved}

\item{file_name}{Name of data to be saved}
}
\description{
Save a data set in the log directory
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
