% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_log.R
\name{.sits_log_elapsed_time}
\alias{.sits_log_elapsed_time}
\title{Compute the duration between the last .sits_log call and now}
\usage{
.sits_log_elapsed_time(time)
}
\arguments{
\item{time}{current time}
}
\value{
A character string informing the elapsed time
}
\description{
Compute the difference between now and the last call to
.sits_log function.
}
\author{
Rolf Simoes, \email{rolf.simoes@inpe.br}
}
\keyword{internal}
