% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_log.R
\name{.sits_log_error}
\alias{.sits_log_error}
\title{Logs an error in the log file}
\usage{
.sits_log_error(message)
}
\arguments{
\item{message}{Message to be logged.}
}
\value{
TRUE if creation succeeds.
}
\description{
Logs an error message in the log file.
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
