% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_classify_cube_aux_functions.R
\name{.sits_map_stack_cluster}
\alias{.sits_map_stack_cluster}
\title{parallel raster brick}
\usage{
.sits_map_stack_cluster(
  cube_in,
  cube_out,
  resolution,
  sits_proc_fn,
  multicores,
  memsize,
  ...
)
}
\arguments{
\item{cube_in}{raster data cube}

\item{cube_out}{output probability data cube}

\item{multicores}{Number of process to run the function in
snow subprocess.}

\item{memsize}{Maximum overall memory (in GB) to run the function.}

\item{...}{optional arguments to merge final raster
(see \link[raster]{writeRaster} function)}

\item{layer_fn}{a function that receives RasterLayer and
returns a RasterLayer.}

\item{args}{additional arguments to pass to \code{fun} function.}
}
\value{
RasterBrick object
}
\description{
Process chunks of raster brick individually in parallel.
}
\keyword{internal}
