% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_plot.R
\name{.sits_max_colors}
\alias{.sits_max_colors}
\title{Brewer color schemes}
\usage{
.sits_max_colors(brewer = NULL)
}
\arguments{
\item{brewer}{name of the brewer color set.}
}
\value{
The number of available colors.
}
\description{
Number of colors available in brewer color schemes.
}
