% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_machine_learning.R
\name{.sits_ml_model_samples}
\alias{.sits_ml_model_samples}
\title{Get the samples from a ml_model}
\usage{
.sits_ml_model_samples(ml_model)
}
\arguments{
\item{ml_model}{A trained model.}
}
\value{
A tibble with samples used to train a machine learning model.
}
\description{
this function get the samples from a ml_model object.
}
\author{
Rolf Simoes, \email{rolf.simoes@inpe.br}
}
\keyword{internal}
