% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_mspc.R
\name{.sits_mspc_filter_bands}
\alias{.sits_mspc_filter_bands}
\title{Filter bands in sentinel collection}
\usage{
.sits_mspc_filter_bands(items, bands, s2_resolution)
}
\arguments{
\item{items}{a \code{STACItemCollection} object returned by rstac package.
grouped.}

\item{bands}{a \code{character} vector with the bands name.}

\item{s2_resolution}{Resolution of S2 images ("10m", "20m" or "60m")
used to build cubes (only for AWS cubes)}
}
\value{
a \code{STACItemCollection} object representing the search
             by rstac with filtered bands.
}
\description{
Filter bands in sentinel collection
}
\keyword{internal}
