% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_mspc.R
\name{.sits_mspc_search_tiles_sentinel}
\alias{.sits_mspc_search_tiles_sentinel}
\title{Search items tiles in sentinel collection}
\usage{
.sits_mspc_search_tiles_sentinel(items, tiles)
}
\arguments{
\item{items}{a \code{STACItemCollection} object returned by rstac package.
grouped.}

\item{tiles}{a \code{character} with the names of the tiles.}
}
\value{
a \code{STACItemCollection} object representing the search
             by rstac.
TODO: improve this function
}
\description{
Search items tiles in sentinel collection
}
\keyword{internal}
