% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_kohonen.R
\name{.sits_neighbor_neurons}
\alias{.sits_neighbor_neurons}
\title{Get the neighbor of neurons}
\usage{
.sits_neighbor_neurons(class_vector, koh, radius = 1)
}
\arguments{
\item{class_vector}{A vector contained the labels of each neuron.}

\item{koh}{An object kohonen.}

\item{radius}{Radius of neigbourhood of each neuron.}
}
\value{
returns a SITS tibble with informations about the vinicity of each neuron.
}
\description{
This function create a table contained the information about all
neighbor of one neuron of radius 1.
}
\author{
Lorena Santos, \email{lorena.santos@inpe.br}
}
