% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_machine_learning.R
\name{.sits_normalize_data}
\alias{.sits_normalize_data}
\title{Normalize the time series in the given sits_tibble}
\usage{
.sits_normalize_data(data, stats, multicores = 1, progress = TRUE)
}
\arguments{
\item{data}{A sits tibble.}

\item{stats}{Statistics for normalization.}

\item{multicores}{Number of cores to process.}

\item{progress}{Show progress bar? Default is TRUE.}
}
\value{
A normalized sits tibble.
}
\description{
This function normalizes the time series using the mean and
standard deviation of all the time series.
}
\author{
Alber Sanchez, \email{alber.ipia@inpe.br}
}
