% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_normalization.R
\name{.sits_normalize_matrix}
\alias{.sits_normalize_matrix}
\title{Normalize the time series values in the case of a matrix}
\usage{
.sits_normalize_matrix(data.mx, stats.tb, band, multicores)
}
\arguments{
\item{data.mx}{matrix of values}

\item{stats.tb}{statistics for normalization}

\item{band}{band to be normalized}

\item{multicores}{number of cores}
}
\value{
data.mx        a normalized matrix
}
\description{
this function normalizes one band of the values read from a raster brick
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
