% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_ops_bands.R
\name{.sits_ops_compute}
\alias{.sits_ops_compute}
\title{Add new sits bands and drops existing.}
\usage{
.sits_ops_compute(data, ...)
}
\arguments{
\item{data}{A sits tibble.}

\item{...}{Pair expressions in the format `name = value`.
See \code{\link[dplyr]{mutate}} help for more details.}
}
\value{
A sits tibble with same samples and the new bands.
}
\description{
Adds new bands and drops existing in the time series
              of a sits tibble using dplyr::transmute function.
}
\author{
Rolf Simoes, \email{rolf.simoes@inpe.br}
}
\keyword{internal}
