% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_sample_functions.R
\name{.sits_oversample_smote}
\alias{.sits_oversample_smote}
\title{Oversample a dataset by SMOTE.}
\usage{
.sits_oversample_smote(data, cls, cls_col, m)
}
\arguments{
\item{data}{Dataset to be oversampled.}

\item{cls}{Class to be oversampled.}

\item{cls_col}{Column containing class information.}

\item{m}{Desired number of samples in the oversampled data.}
}
\value{
The oversampled dataset.
}
\description{
Lifted from R package "scutr".
}
