% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_parallel.R
\name{.sits_parallel_is_open}
\alias{.sits_parallel_is_open}
\title{Check if sits clusters are open or not}
\usage{
.sits_parallel_is_open()
}
\description{
Check if sits clusters are open or not
}
\author{
Rolf Simoes, \email{rolf.simoes@inpe.br}
}
\keyword{internal}
