% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_parallel.R
\name{.sits_parallel_map}
\alias{.sits_parallel_map}
\title{Maps a function to a given list in parallel or sequential processing}
\usage{
.sits_parallel_map(x, fn, ..., progress = FALSE, n_retries = 3, sleep = 0)
}
\arguments{
\item{x}{a given list to be passed to a function}

\item{fn}{a function to be applied to each list element}

\item{progress}{a logical value indicating if a progress bar should
be shown}

\item{n_retries}{a number of retries before fail}

\item{sleep}{a number in seconds to wait before try again}
}
\value{
a list with the function results in the same order
as the input list
}
\description{
Maps a function to a given list in parallel or sequential processing
}
\author{
Rolf Simoes, \email{rolf.simoes@inpe.br}
}
\keyword{internal}
