% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_parallel.R
\name{.sits_parallel_reset_node}
\alias{.sits_parallel_reset_node}
\title{Recreates a cluster worker}
\usage{
.sits_parallel_reset_node(worker_id)
}
\arguments{
\item{worker_id}{id of the cluster work to be recreated}
}
\description{
Recreates a cluster worker
}
\author{
Rolf Simoes, \email{rolf.simoes@inpe.br}
}
\keyword{internal}
