% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_parallel.R
\name{.sits_parallel_start}
\alias{.sits_parallel_start}
\title{Start a new sits cluster for parallel processing}
\usage{
.sits_parallel_start(workers)
}
\arguments{
\item{workers}{number of cluster to instantiate}
}
\description{
Start a new sits cluster for parallel processing
}
\author{
Rolf Simoes, \email{rolf.simoes@inpe.br}
}
\keyword{internal}
