% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_parallel.R
\name{.sits_parallel_stop}
\alias{.sits_parallel_stop}
\title{Stop all workers of the sits cluster}
\usage{
.sits_parallel_stop()
}
\value{
No value, called for side effect.
}
\description{
Stop all workers of the sits cluster
}
\author{
Rolf Simoes, \email{rolf.simoes@inpe.br}
}
\keyword{internal}
