% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_parallel.R
\name{.sits_parallel_workers}
\alias{.sits_parallel_workers}
\title{Return the number of active workers}
\usage{
.sits_parallel_workers()
}
\description{
Return the number of active workers
}
\author{
Rolf Simoes, \email{rolf.simoes@inpe.br}
}
\keyword{internal}
