% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_roi.R
\name{.sits_parse_roi_cube.ll}
\alias{.sits_parse_roi_cube.ll}
\title{Parse a ROI defined as lat-long}
\usage{
\method{.sits_parse_roi_cube}{ll}(roi)
}
\arguments{
\item{roi}{spatial region of interest}
}
\value{
a sf object in WGS84 projection
}
\description{
Parse a ROI defined as lat-long
}
\keyword{internal}
