% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_roi.R
\name{.sits_parse_roi_cube}
\alias{.sits_parse_roi_cube}
\title{Parse is a ROI is valid for an existing data cube}
\usage{
.sits_parse_roi_cube(roi)
}
\arguments{
\item{roi}{spatial region of interest}
}
\value{
roi in WGS84 projection or NULL if error
}
\description{
Parse is a ROI is valid for an existing data cube
}
\keyword{internal}
