% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_roi.R
\name{.sits_parse_roi_cube.sf}
\alias{.sits_parse_roi_cube.sf}
\title{Parse is a ROI defined as sf is valid for an existing data cube}
\usage{
\method{.sits_parse_roi_cube}{sf}(roi)
}
\arguments{
\item{roi}{spatial region of interest}
}
\value{
roi in WGS84 projection
}
\description{
Parse is a ROI defined as sf is valid for an existing data cube
}
\keyword{internal}
