% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_plot.R
\name{.sits_plot}
\alias{.sits_plot}
\title{Plot a set of satellite image time series}
\usage{
.sits_plot(data, bands = c("ndvi"), colors = "Dark2")
}
\arguments{
\item{data}{Data to be plotted (can be a sits tibble, clusters, or TWDTW matches).}

\item{bands}{Bands used for visualisation (optional for sits_plot_classification).}

\item{colors}{Color pallete to be used (based on Color Brewer - default is "Dark2").}
}
\value{
Input sits tibble (useful for chaining functions).
}
\description{
Given a sits tibble with a set of time series, plot them.

The plot function produces different plots based on the input data:
\itemize{
 \item{"all years": }{Plot all samples from the same location together}
 \item{"together": }{Plot all samples of the same band and label together}
}
The sits_plot function makes an educated guess of what plot is required,
based on the input data. If the input data has less than 30 samples, it
will default to "all years". If there is only one sample per class, it will
default to "patterns". If there are more than 30 samples, it will default to
"together".
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
