% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_plot.R
\name{.sits_plot_allyears}
\alias{.sits_plot_allyears}
\title{Plot all time intervals of one time series for the same lat/long together}
\usage{
.sits_plot_allyears(data, colors)
}
\arguments{
\item{data}{One or more time series (stored in a sits tibble).}

\item{colors}{The color pallete to be used (default is "Set2").}
}
\description{
For each lat/long location in the data, join temporal
instances of the same place together for plotting.
}
