% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_plot.R
\name{.sits_plot_dendrogram}
\alias{.sits_plot_dendrogram}
\title{Plot a dendrogram}
\usage{
.sits_plot_dendrogram(data, cluster, cutree_height, palette)
}
\arguments{
\item{data}{sits tibble with data used to extract the dendrogram.}

\item{cluster}{cluster object produced by `sits_cluster` function.}

\item{cutree_height}{dashed horizontal line to be drawn
indicating the height of dendrogram cutting.}

\item{palette}{hcl color palette}
}
\value{
The plot itself.
}
\description{
Plot a dendrogram
}
\author{
Rolf Simoes, \email{rolf.simoes@inpe.br}
}
\keyword{internal}
