% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_plot.R
\name{.sits_plot_ggplot_series}
\alias{.sits_plot_ggplot_series}
\title{Plot one time series using ggplot}
\usage{
.sits_plot_ggplot_series(row)
}
\arguments{
\item{row}{row of a sits tibble with the time series to be plotted.}
}
\value{
A plot object produced by the ggplot2 package showing
                   one time series.
}
\description{
Plots a set of time series using ggplot. This function is used
for showing the same lat/long location in a series of time steps.
}
\keyword{internal}
