% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_plot.R
\name{.sits_plot_patterns}
\alias{.sits_plot_patterns}
\title{Plot classification patterns}
\usage{
.sits_plot_patterns(data)
}
\arguments{
\item{data}{One or more time series containing patterns (stored in a sits tibble).}
}
\description{
Plots the patterns to be used for classification
             this code is reused from the dtwSat package by Victor Maus.
}
\author{
Victor Maus, \email{vwmaus1@gmail.com}
}
