% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_plot.R
\name{.sits_plot_predicted_ts}
\alias{.sits_plot_predicted_ts}
\title{Plot time series classification results}
\usage{
.sits_plot_predicted_ts(data, bands, palette)
}
\arguments{
\item{data}{sits tibble with classified time series.}

\item{bands}{band for plotting the classification.}

\item{palette}{hcl palette used for visualisation}
}
\value{
The plot itself.
}
\description{
plots the classification results
                    (code reused from the dtwSat package by Victor Maus).
}
\author{
Victor Maus, \email{vwmaus1@gmail.com}
}
\keyword{internal}
