% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_plot.R
\name{.sits_plot_raster}
\alias{.sits_plot_raster}
\title{Plot a raster classified images}
\usage{
.sits_plot_raster(cube, time = 1, title = "Classified Image", colors = NULL)
}
\arguments{
\item{cube}{A tibble with the metadata for a labelled data cube.}

\item{time}{Temporal reference for plot.}

\item{title}{A string.}

\item{colors}{Color pallete.}
}
\description{
plots a raster using ggplot. This function is used
for showing the same lat/long location in a series of time steps.
}
