% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_plot.R
\name{.sits_plot_resample_class}
\alias{.sits_plot_resample_class}
\title{Calculate resample params for classified images}
\usage{
.sits_plot_resample_class(nrows, ncols, max_Mbytes)
}
\arguments{
\item{nrows}{number of rows in the input image}

\item{ncols}{number of cols in the input image}

\item{max_Mbytes}{maximum number of MB per plot}
}
\value{
ratio and new size of output plot
}
\description{
Calculate resample params for classified images
}
\keyword{internal}
