% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_plot.R
\name{.sits_plot_som_evaluate_cluster}
\alias{.sits_plot_som_evaluate_cluster}
\title{Plot information about confusion between clusters}
\usage{
.sits_plot_som_evaluate_cluster(
  data,
  cluster_name = NULL,
  title = "Confusion by cluster"
)
}
\arguments{
\item{data}{Percentage of mixture between the clusters}

\item{title}{Title of plot.}

\item{name_cluster}{Choose the cluster to plot}
}
\value{
ggplot2 object
}
\description{
Plot a bar graph with information about each cluster.
The percentage of mixture between the clusters.
}
\author{
Lorena Santos \email{lorena.santos@inpe.br}
}
\keyword{internal}
