% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_plot.R
\name{.sits_plot_som_map}
\alias{.sits_plot_som_map}
\title{Plot the SOM grid with neurons labeled}
\usage{
.sits_plot_som_map(koh, type = "codes", whatmap = 1)
}
\arguments{
\item{koh}{SOM map produced by "sits_som_map" function}

\item{type}{Type of plot ("codes" or "mapping")}

\item{whatmap}{What data layer will be plotted.}
}
\description{
Given a kohonen object with a set of time neurons, plot them.

The plot function produces different plots based on the input data:
\itemize{
 \item{"codes": }{Plot the vector weight for each neuron.}
 \item{"mapping": }{Shows where samples are mapped.}
}
}
\author{
Lorena Santos \email{lorena.santos@inpe.br}
}
