% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_plot.R
\name{.sits_plot_title}
\alias{.sits_plot_title}
\title{Create a plot title to use with ggplot}
\usage{
.sits_plot_title(latitude, longitude, label)
}
\arguments{
\item{latitude}{latitude of the location to be plotted.}

\item{longitude}{longitude of the location to be plotted.}

\item{label}{label of the location to be plotted.}
}
\value{
title to be used in the plot.
}
\description{
Creates a plot title from row information.
}
\keyword{internal}
