% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_plot.R
\name{.sits_plot_TWDTW_classification}
\alias{.sits_plot_TWDTW_classification}
\title{Plot classification results using the dtwSat package}
\usage{
.sits_plot_TWDTW_classification(matches, start_date = NULL,
  end_date = NULL, interval = "12 month", overlap = 0.5)
}
\arguments{
\item{matches}{A list of dtwSat S4 matches objects produced by sits_TWDTW_matches.}

\item{start_date}{Start date of the plot (used for showing classifications).}

\item{end_date}{End date of the plot (used for showing classifications).}

\item{interval}{Interval between classifications (used for showing classifications).}

\item{overlap}{Minimum overlapping between one match and the interval of classification. For details see dtwSat::twdtwApply help.}
}
\description{
Plots the results of TWDTW classification (uses dtwSat).
}
\author{
Victor Maus, \email{vwmaus1@gmail.com}

Gilberto Camara, \email{gilberto.camara@inpe.br}
}
