% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_shp.R
\name{.sits_points_from_shp}
\alias{.sits_points_from_shp}
\title{Obtain a tibble with lat/long points to be retrieved from a SHP}
\usage{
.sits_points_from_shp(sf_shape, shp_attr, label, .n_shp_pol)
}
\arguments{
\item{sf_shape}{sf object that contains a SHP file}

\item{shp_attr}{Attribute in the shapefile used as a polygon label}

\item{label}{Label to be assigned to points}

\item{.n_shp_pol}{Number of samples per polygon to be read
(for POLYGON or MULTIPOLYGON shapes)}
}
\value{
A sits tibble with points to to be read.
}
\description{
reads a shapefile and retrieves a sits tibble
containing a set of lat/long points for data retrieval
}
\keyword{internal}
