% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_conf_matrix.R
\name{.sits_pred_ref}
\alias{.sits_pred_ref}
\title{Obtains the predicted value of a reference set}
\usage{
.sits_pred_ref(class)
}
\arguments{
\item{class}{Tibble with classified samples whose labels are known.}
}
\value{
A tibble with predicted and reference values.
}
\description{
Obtains a tibble of predicted and reference values
from a classified data set.
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
\keyword{internal}
