% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_machine_learning.R
\name{.sits_predict}
\alias{.sits_predict}
\title{Predict class based on the trained models}
\usage{
.sits_predict(distances_DT = NULL, ml_model, ...)
}
\arguments{
\item{distances_DT}{set of distance metrics to each of the classes}

\item{ml_model}{model trained by \code{\link[sits]{sits_train}}}

\item{...}{other parameters to be passed to the model function}
}
\value{
predicted    vector of predicted labels
}
\description{
Given a SITS tibble time series and a model trained by \code{\link[sits]{sits_train}},
  returns a predicted label. Note that the \code{\link[sits]{.sits_predict}} function is
  called inside \code{\link[sits]{sits_classify}},
  and \code{\link[sits]{sits_classify_raster}}, so the user does not need
  to explicitly use it. Please see the above-mentioned classification functions.
}
\author{
Alexandre Xavier Ywata de Carvalho, \email{alexandre.ywata@ipea.gov.br}

Rolf Simoes, \email{rolf.simoes@inpe.br}
}
