% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_raster_classification.R
\name{.sits_predict_interval}
\alias{.sits_predict_interval}
\title{Classify one interval of data}
\usage{
.sits_predict_interval(DT, time, output.lst, ml_model, labels, int_labels,
  first_row, multicores)
}
\arguments{
\item{DT}{data.table with distance values}

\item{time}{time interval to be processed}

\item{output.lst}{list with the raster objects for classification (values and probs)}

\item{ml_model}{machine learning model to be applied}

\item{labels}{class labels}

\item{int_labels}{integer values corresponding to labels}

\item{first_row}{initial row of the output layer to write block}

\item{multicores}{number of cores to process the time series}
}
\value{
layers.lst        list of layers with classification results
}
\description{
Classify one interval of data
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
