% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_raster_read_write.R
\name{.sits_preprocess_data}
\alias{.sits_preprocess_data}
\title{Preprocess a set of values retrived from a raster brick}
\usage{
.sits_preprocess_data(values.mx, band, missing_value, minimum_value,
  scale_factor, stats, filter, multicores)
}
\arguments{
\item{values.mx}{matrix of values retrieved from a brick}

\item{band}{band to be processed}

\item{missing_value}{missing value for the band}

\item{minimum_value}{minimum values for the band}

\item{scale_factor}{scale factor for each band (only for raster data)}

\item{stats}{normalization parameters}

\item{filter}{smoothing filter to be applied}

\item{multicores}{number of cores to process the time series}
}
\value{
values.mx        matrix with pre-processed values
}
\description{
Preprocess a set of values retrived from a raster brick
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
