% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_smooth_aux_functions.R
\name{.sits_probs_blocks_size_estimate}
\alias{.sits_probs_blocks_size_estimate}
\title{Estimate the number of blocks to run .sits_split_cluster}
\usage{
.sits_probs_blocks_size_estimate(cube, multicores, memory)
}
\arguments{
\item{cube}{input data cube}

\item{multicores}{number of processes to split up the data}

\item{memory}{maximum overall memory size (in GB)}
}
\value{
returns a list with following information:
            - multicores theoretical upper bound;
            - block x_size (horizontal) and y_size (vertical)
}
\description{
Estimate the number of blocks to run .sits_split_cluster
}
\keyword{internal}
