% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_processing_estimate.R
\name{.sits_processing_estimate_classification_time}
\alias{.sits_processing_estimate_classification_time}
\title{Estimate the processing time}
\usage{
.sits_processing_estimate_classification_time(
  start_time,
  n_intervals,
  bs,
  block,
  time
)
}
\arguments{
\item{start_time}{initial processing time.}

\item{n_intervals}{number of time intervals}

\item{bs}{raster block parameters.}

\item{block}{current block.}

\item{time}{current interval.}

\item{task}{task to be performed}
}
\value{
Scaled matrix.
}
\description{
This function normalizes image values.
}
\author{
Rolf Simoes, \email{rolf.simoes@inpe.br}
}
\keyword{internal}
