% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_processing_estimate.R
\name{.sits_processing_task_time}
\alias{.sits_processing_task_time}
\title{Estimate the processing time of a task}
\usage{
.sits_processing_task_time(task, start_time)
}
\arguments{
\item{task}{Description of the task}

\item{start_time}{Initial processing time.}
}
\value{
message about processing time for a task
}
\description{
This function normalizes image values.
}
\author{
Rolf Simoes, \email{rolf.simoes@inpe.br}
}
\keyword{internal}
