% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_space_time_operations.R
\name{.sits_proj_to_latlong}
\alias{.sits_proj_to_latlong}
\title{Coordinate transformation (X/Y to lat/long)}
\usage{
.sits_proj_to_latlong(x, y, crs)
}
\arguments{
\item{x}{double - x coordinate of the chosen location}

\item{y}{double - y coordinateof the chosen location}

\item{crs}{projection definition to be converted from}
}
\value{
xy             matrix with latlong coordinates
}
\description{
Transform a latitude and longitude coordinate to a XY projection coordinate
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
