% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_config.R
\name{.sits_providers}
\alias{.sits_providers}
\title{List the data services available}
\usage{
.sits_providers(service)
}
\arguments{
\item{service}{Name of the web service}
}
\value{
List of providers associated to a service
}
\description{
List the data services available
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
