% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_tibble_aux_functions.R
\name{.sits_prune}
\alias{.sits_prune}
\title{Checks that the timeline of all time series of a data set are equal}
\usage{
.sits_prune(data)
}
\arguments{
\item{data}{Either a sits tibble or a raster metadata.}
}
\value{
A pruned sits tibble.
}
\description{
This function tests if all time series in a sits tibble
have the same number of samples, and returns a time series whose indices
match the majority of the samples.
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
\keyword{internal}
