% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_raster_api.R
\name{.sits_raster_api_area_frequency}
\alias{.sits_raster_api_area_frequency}
\title{Given a labelled cube, return the band information}
\usage{
.sits_raster_api_area_frequency(cube)
}
\arguments{
\item{cube}{Metadata about a data cube}
}
\value{
Frequency of each label in the data cube
}
\description{
Given a labelled cube, return the band information
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
\keyword{internal}
