% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_raster_api.R
\name{.sits_raster_api_check_access}
\alias{.sits_raster_api_check_access}
\title{Check if the raster files are accessible by GDAL}
\usage{
.sits_raster_api_check_access(file)
}
\arguments{
\item{file}{file associated to the raster data}
}
\value{
TRUE         true if files are accessible
}
\description{
Check if the raster files are accessible by GDAL
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
\keyword{internal}
