% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_raster_api.R
\name{.sits_raster_api_check_block}
\alias{.sits_raster_api_check_block}
\title{Check for block object consistency}
\usage{
.sits_raster_api_check_block(block)
}
\description{
Check for block object consistency
}
\author{
Rolf Simoes, \email{rolf.simoes@inpe.br}
}
\keyword{internal}
