% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_raster_api.R
\name{.sits_raster_api_check_brick}
\alias{.sits_raster_api_check_brick}
\title{Check if file is a brick}
\usage{
.sits_raster_api_check_brick(file, timeline)
}
\arguments{
\item{file}{A valid raster image}

\item{timeline}{Timeline for the brick}
}
\value{
TRUE/FALSE
}
\description{
Check if file is a valid raster brick
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
\keyword{internal}
