% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_raster_api.R
\name{.sits_raster_api_check_package}
\alias{.sits_raster_api_check_package}
\title{Check for raster package availability}
\usage{
.sits_raster_api_check_package()
}
\value{
package representation class
}
\description{
Check for raster package availability
}
\author{
Rolf Simoes, \email{rolf.simoes@inpe.br}
}
\keyword{internal}
