% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_raster_api.R
\name{.sits_raster_api_data_type}
\alias{.sits_raster_api_data_type}
\title{Raster package internal data type representation}
\usage{
.sits_raster_api_data_type(data_type)
}
\arguments{
\item{data_type}{sits internal raster data type. One of "INT1U", "INT2U",
"INT2S", "INT4U", "INT4S", "FLT4S", "FLT8S".}
}
\value{
character string
}
\description{
Raster package internal data type representation
}
\author{
Rolf Simoes, \email{rolf.simoes@inpe.br}
}
\keyword{internal}
