% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_raster_api.R
\name{.sits_raster_api_extract}
\alias{.sits_raster_api_extract}
\title{Given a band, return a set of values for chosen location}
\usage{
.sits_raster_api_extract(cube, band_cube, xy)
}
\arguments{
\item{cube}{Metadata about a data cube}

\item{band_cube}{Name of the band to the retrieved}

\item{xy}{Matrix with XY location}
}
\value{
Matrix with values extracted from image files
}
\description{
Given a data cube, retrieve the timeline
                      (uses the terra package which is faster than raster)
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
\keyword{internal}
