% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_raster_api.R
\name{.sits_raster_api_focal}
\alias{.sits_raster_api_focal}
\title{Raster package internal moving window function}
\usage{
.sits_raster_api_focal(r_obj, window_size, fn, ...)
}
\arguments{
\item{r_obj}{raster package object to pass a window function}

\item{window_size}{number indicating the length of a squared window's side.}

\item{fn}{a function to be convoluted. Can be either a string or
a R function. Character strings options are: "sum",
"mean", and "modal".}

\item{...}{additional parameters to be passed to raster package}
}
\value{
raster package object
}
\description{
Raster package internal moving window function
}
\author{
Rolf Simoes, \email{rolf.simoes@inpe.br}
}
\keyword{internal}
