% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_raster_api.R
\name{.sits_raster_api_freq}
\alias{.sits_raster_api_freq}
\title{Raster package internal frequency values function}
\usage{
.sits_raster_api_freq(r_obj, ...)
}
\arguments{
\item{r_obj}{raster package object to count values}

\item{...}{additional parameters to be passed to raster package}
}
\value{
matrix with layer, value, and count columns
}
\description{
Raster package internal frequency values function
}
\author{
Rolf Simoes, \email{rolf.simoes@inpe.br}
}
\keyword{internal}
