% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_raster_api.R
\name{.sits_raster_api_new_rast}
\alias{.sits_raster_api_new_rast}
\alias{.sits_raster_api_new_rast.terra_package}
\title{Create a new raster file}
\usage{
\method{.sits_raster_api_new_rast}{terra_package}(
  nrows,
  ncols,
  xmin,
  xmax,
  ymin,
  ymax,
  nlayers,
  crs,
  ...
)
}
\arguments{
\item{nrows}{Number of rows in the raster}

\item{ncols}{Number of columns in the raster}

\item{xmin}{X minimum of raster origin}

\item{xmax}{X maximum of raster origin}

\item{ymin}{Y minimum of raster origin}

\item{ymax}{Y maximum of raster origin}

\item{nlayers}{Number of layers of the raster}

\item{crs}{Coordinate Reference System of the raster}
}
\value{
Data cube values.
}
\description{
Create a new raster file
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}

Rolf Simoes, \email{rolf.simoes@inpe.br}
}
\keyword{internal}
