% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_raster_api.R
\name{.sits_raster_api_open_rast}
\alias{.sits_raster_api_open_rast}
\title{Raster package internal open raster function}
\usage{
.sits_raster_api_open_rast(file, ...)
}
\arguments{
\item{file}{raster file to be opened}

\item{...}{additional parameters to be passed to raster package}
}
\value{
raster package object
}
\description{
Raster package internal open raster function
}
\author{
Rolf Simoes, \email{rolf.simoes@inpe.br}
}
\keyword{internal}
