% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_raster_api.R
\name{.sits_raster_api_open_stack}
\alias{.sits_raster_api_open_stack}
\title{Raster package internal open raster stack function}
\usage{
.sits_raster_api_open_stack(files, ...)
}
\arguments{
\item{files}{raster files to be opened}

\item{...}{additional parameters to be passed to raster package}
}
\value{
raster package object
}
\description{
Raster package internal open raster stack function
}
\author{
Rolf Simoes, \email{rolf.simoes@inpe.br}
}
\keyword{internal}
