% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_raster_api.R
\name{.sits_raster_api_properties}
\alias{.sits_raster_api_properties}
\alias{.sits_raster_api_nrows}
\alias{.sits_raster_api_ncols}
\alias{.sits_raster_api_nlayers}
\alias{.sits_raster_api_xmax}
\alias{.sits_raster_api_xmin}
\alias{.sits_raster_api_ymax}
\alias{.sits_raster_api_ymin}
\alias{.sits_raster_api_xres}
\alias{.sits_raster_api_yres}
\alias{.sits_raster_api_crs}
\title{Raster package internal object properties}
\usage{
.sits_raster_api_nrows(r_obj, ...)

.sits_raster_api_ncols(r_obj, ...)

.sits_raster_api_nlayers(r_obj, ...)

.sits_raster_api_xmax(r_obj, ...)

.sits_raster_api_xmin(r_obj, ...)

.sits_raster_api_ymax(r_obj, ...)

.sits_raster_api_ymin(r_obj, ...)

.sits_raster_api_xres(r_obj, ...)

.sits_raster_api_yres(r_obj, ...)

.sits_raster_api_crs(r_obj, ...)
}
\arguments{
\item{r_obj}{raster package object}

\item{...}{additional parameters to be passed to raster package}
}
\value{
raster object spatial properties
}
\description{
Raster package internal object properties
}
\author{
Rolf Simoes, \email{rolf.simoes@inpe.br}
}
\keyword{internal}
