% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_raster_api.R
\name{.sits_raster_api_rast}
\alias{.sits_raster_api_rast}
\title{Raster package internal object creation}
\usage{
.sits_raster_api_rast(r_obj, nlayers = 1, ...)
}
\arguments{
\item{r_obj}{raster package object to be cloned}

\item{nlayers}{number of raster layers}

\item{...}{additional parameters to be passed to raster package}
}
\value{
raster package object
}
\description{
Raster package internal object creation
}
\author{
Rolf Simoes, \email{rolf.simoes@inpe.br}
}
\keyword{internal}
