% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_raster_api_raster.R,
%   R/sits_raster_api_terra.R
\name{.sits_raster_api_read_extent}
\alias{.sits_raster_api_read_extent}
\alias{.sits_raster_api_read_stack.raster_package}
\alias{.sits_raster_api_read_stack.terra_package}
\title{Read a part of a raster file and return a matrix}
\usage{
\method{.sits_raster_api_read_stack}{raster_package}(files, extent = NULL, ...)

\method{.sits_raster_api_read_stack}{terra_package}(files, extent = NULL, ...)
}
\arguments{
\item{files}{Files associated to the raster object}

\item{extent}{Image extent to be read.}
}
\value{
Data.table of values

Data.table of values
}
\description{
Read a part of a raster file and return a matrix

Read a part of a raster file and return a matrix
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}

Gilberto Camara, \email{gilberto.camara@inpe.br}
}
\keyword{internal}
