% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_raster_api.R
\name{.sits_raster_api_read_file}
\alias{.sits_raster_api_read_file}
\title{Read a raster file and return a matrix}
\usage{
.sits_raster_api_read_file(r_file)
}
\arguments{
\item{r_file}{Files associated to the raster object}
}
\value{
Data.table of values
}
\description{
Read a raster file and return a matrix
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
\keyword{internal}
