% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_raster_api.R
\name{.sits_raster_api_write}
\alias{.sits_raster_api_write}
\title{Set the values of a raster object matrix}
\usage{
.sits_raster_api_write(
  params,
  num_layers,
  values,
  filename,
  datatype,
  format = "GTiff",
  compress = "LZW",
  overwrite = TRUE
)
}
\arguments{
\item{params}{Output file params}

\item{num_layers}{Number of layers}

\item{values}{Data.table with values to be written}

\item{filename}{File name of the raster image file.}

\item{datatype}{Data type}

\item{format}{Format to write the file}

\item{compress}{Compression method to be used}

\item{overwrite}{Overwrite the file?}
}
\value{
Data cube values.
}
\description{
Set the values of a raster object matrix
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
\keyword{internal}
