% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_raster_api.R
\name{.sits_raster_api_write_lab}
\alias{.sits_raster_api_write_lab}
\title{Set the values of a labelled raster object}
\usage{
.sits_raster_api_write_lab(cube, values, smoothing = "none", filename)
}
\arguments{
\item{cube}{Data cube}

\item{values}{Matrix with values to be written}

\item{smoothing}{Smoothing procedure to apply}

\item{filename}{Filename of the raster image file}
}
\value{
Data cube
}
\description{
Set the values of a labelled raster object
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
\keyword{internal}
