% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_raster_data.R
\name{.sits_raster_bdc_tile_cube}
\alias{.sits_raster_bdc_tile_cube}
\title{Create a data cube for a BDC TILE}
\usage{
.sits_raster_bdc_tile_cube(
  satellite,
  sensor,
  name,
  bands,
  cube,
  tile,
  stack_info
)
}
\arguments{
\item{satellite}{Name of satellite}

\item{sensor}{Name of sensor}

\item{name}{Name of the data cube.}

\item{bands}{Vector of bands}

\item{cube}{Input cube}

\item{tile}{Tile}

\item{stack_info}{Tibble with information about the stack.}
}
\value{
A tibble with metadata information about a raster data set.
}
\description{
Builds a BDC_TILE cube
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
