% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_raster_blocks.R
\name{.sits_raster_block_list}
\alias{.sits_raster_block_list}
\title{Calculate a list of blocks to be read from disk to memory}
\usage{
.sits_raster_block_list(nblocks, sub_image)
}
\arguments{
\item{nblocks}{number of blocks to read from each image}

\item{sub_image}{nrea of interest in the image}
}
\value{
a list with n (number of blocks), row (vector of starting rows),
               nrow (vector with number of rows for each block) and
               size (vector with size of each block)
}
\description{
Calculate a list of blocks to be read from disk to memory
}
\keyword{internal}
