% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_raster_blocks.R
\name{.sits_raster_blocks_estimate_regularize}
\alias{.sits_raster_blocks_estimate_regularize}
\title{Estimate the number of blocks for regularize}
\usage{
.sits_raster_blocks_estimate_regularize(
  cube,
  n_images_interval,
  sub_image,
  memsize,
  multicores
)
}
\arguments{
\item{cube}{Tile of data cube.}

\item{n_images_interval}{Number of images per interval.}

\item{sub_image}{Area of interest in the image.}

\item{memsize}{Memory available for classification (in GB).}

\item{multicores}{Number of threads to process the time series.}
}
\value{
Number of blocks to be read.
}
\description{
Defines the number of blocks of a Raster Brick
             to be read into memory.
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
\keyword{internal}
