% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_raster_brick.R
\name{.sits_raster_brick_check}
\alias{.sits_raster_brick_check}
\title{Check if the raster files are bricks}
\usage{
.sits_raster_brick_check(satellite, sensor, name, timeline, bands, files)
}
\arguments{
\item{satellite}{Name of satellite}

\item{sensor}{Name of sensor}

\item{name}{Name of the data cube.}

\item{timeline}{Vector of dates with the timeline of the bands.}

\item{bands}{Vector of bands contained in the Raster Brick
set (in the same order as the files).}

\item{files}{Vector with the file paths of the raster files}
}
\value{
TRUE                  passed the check?
}
\description{
Check if the raster files are bricks
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
