% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_raster_data.R
\name{.sits_raster_brick_cube}
\alias{.sits_raster_brick_cube}
\title{Check if input data is made of raster bricks and create a cube if all is OK}
\usage{
.sits_raster_brick_cube(satellite, sensor, name, timeline, bands, files)
}
\arguments{
\item{satellite}{Name of satellite}

\item{sensor}{Name of sensor}

\item{name}{Name of the data cube.}

\item{timeline}{Vector of dates with the timeline of the bands.}

\item{bands}{Vector of bands contained in the Raster Brick
set (in the same order as the files).}

\item{files}{Vector with the file paths of the raster files.}
}
\value{
A tibble with metadata information about a raster data set.
}
\description{
This function checks if the input data corresponds
              to a set of raster bricks and creates a tibble containing the metadata for
              a set of spatio-temporal raster files,
              These files should be of the same size and
              projection. Each raster brick file should contain one band
              per time step.
              Different bands are archived in different raster files.
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
