% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_raster_data.R
\name{.sits_raster_check_bdc_tiles}
\alias{.sits_raster_check_bdc_tiles}
\title{Check if the BDC tiles are working}
\usage{
.sits_raster_check_bdc_tiles(
  satellite,
  sensor,
  bands,
  cube,
  tile,
  data_access,
  start_date,
  end_date,
  .local,
  .web
)
}
\arguments{
\item{satellite}{satellite}

\item{sensor}{sensor}

\item{bands}{bands}

\item{cube}{input cube}

\item{tile}{tile}

\item{data_access}{type of access}

\item{start_date}{start_date of the cube}

\item{end_date}{end date of the cube}

\item{.local}{local address (if different from default)}

\item{.web}{web address (if different from default)}
}
\description{
Check if the BDC tiles are working
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
