% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_raster.R
\name{.sits_raster_check_gdal_access}
\alias{.sits_raster_check_gdal_access}
\title{Check if the raster files are accessible by GDAL}
\usage{
.sits_raster_check_gdal_access(files)
}
\arguments{
\item{files}{files associated to the raster data}
}
\value{
TRUE         true if filles are acessible
}
\description{
Check if the raster files are accessible by GDAL
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
