% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_raster.R
\name{.sits_raster_cld_remove}
\alias{.sits_raster_cld_remove}
\title{Remove cloud pixels and NA values by imputation}
\usage{
.sits_raster_cld_remove(values.mx, impute_fn, multicores)
}
\arguments{
\item{values.mx}{matrix of values retrieved from a raster object}

\item{impute_fn}{imputing function to be applied to replace NA}

\item{multicores}{number of cores to process the time series.}
}
\value{
Matrix with pre-processed values.
}
\description{
Remove cloud pixels and NA values by imputation
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
\keyword{internal}
