% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_raster_data.R
\name{.sits_raster_data_na_remove}
\alias{.sits_raster_data_na_remove}
\title{Remove cloud pixels and NA values by imputation}
\usage{
.sits_raster_data_na_remove(values, impute_fn, multicores)
}
\arguments{
\item{values}{matrix of values retrieved from a raster object}

\item{impute_fn}{imputing function to be applied to replace NA}

\item{multicores}{number of cores to process the time series.}
}
\value{
Data.table with pre-processed values.
}
\description{
Remove cloud pixels and NA values by imputation
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
\keyword{internal}
