% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_raster_data.R
\name{.sits_raster_data_read}
\alias{.sits_raster_data_read}
\title{Read a block of values retrieved from a set of raster bricks}
\usage{
.sits_raster_data_read(
  cube,
  samples,
  extent,
  stats,
  filter,
  impute_fn,
  multicores,
  .verbose
)
}
\arguments{
\item{cube}{input data cube.}

\item{samples}{tibble with samples.}

\item{extent}{bounding box in (i,j) coordinates}

\item{stats}{normalization parameters.}

\item{filter}{smoothing filter to be applied.}

\item{impute_fn}{impute function to replace NA}

\item{multicores}{number of cores to process the time series.}

\item{.verbose}{print processing information}
}
\value{
A data.table with values for classification.
}
\description{
Read a block of values retrieved from a set of raster bricks
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
\keyword{internal}
