% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_raster_data.R
\name{.sits_raster_data_split}
\alias{.sits_raster_data_split}
\title{Split a data.table or a matrix for multicore processing}
\usage{
.sits_raster_data_split(data, ncores)
}
\arguments{
\item{data}{Data (data.table or matrix).}

\item{ncores}{Number of cores for processing.}
}
\value{
List of pairs of positions (first row, last row)
                        to be assigned to each core.
}
\description{
This function splits a data.table into a
             list of chunks for multicore processing.
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
\keyword{internal}
