% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_raster_data.R
\name{.sits_raster_file_info}
\alias{.sits_raster_file_info}
\title{Create a tibble with file information to include in the cube}
\usage{
.sits_raster_file_info(bands, timeline, files)
}
\arguments{
\item{bands}{List of spectral bands}

\item{timeline}{Cube timeline}

\item{files}{List of files associated to the}
}
\description{
Create a tibble with file information to include in the cube
}
\keyword{internal}
