% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_raster_data.R
\name{.sits_raster_files_robj}
\alias{.sits_raster_files_robj}
\title{Raster object from file}
\usage{
.sits_raster_files_robj(files)
}
\arguments{
\item{files}{Vector of files}
}
\value{
Raster object associated to the first file
}
\description{
Given the vector of files and the name of the service,
                 return the Raster object for the file
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
\keyword{internal}
