% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_raster.R
\name{.sits_raster_filter_data}
\alias{.sits_raster_filter_data}
\title{Filter the time series values in the case of a matrix}
\usage{
.sits_raster_filter_data(values.mx, filter, multicores)
}
\arguments{
\item{values.mx}{Matrix of values.}

\item{filter}{Filter function to apply to matrix.}

\item{multicores}{Number of cores.}
}
\value{
Scaled integer matrix.
}
\description{
This function filters a matrix.
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
