% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_raster.R
\name{.sits_raster_get_ts}
\alias{.sits_raster_get_ts}
\title{Extract a time series from raster}
\usage{
.sits_raster_get_ts(cube, points, bands)
}
\arguments{
\item{cube}{Metadata describing a raster data cube.}

\item{points}{tibble with points}

\item{bands}{Bands to be retrieved.}
}
\value{
A sits tibble with the time series.
}
\description{
Retrieve a set of time series for a raster data cube.
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
\keyword{internal}
