% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_raster_data.R
\name{.sits_raster_metrics}
\alias{.sits_raster_metrics}
\title{. . .}
\usage{
.sits_raster_metrics(metric, values, multicores)
}
\arguments{
\item{metric}{metric to be process}

\item{values}{matrix of values.}

\item{multicores}{Number of cores.}
}
\value{
Filtered matrix.
}
\description{
This function filters a matrix.
}
\keyword{internal}
