% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_raster_data.R
\name{.sits_raster_params}
\alias{.sits_raster_params}
\title{Determine the cube params to write in the metadata}
\usage{
.sits_raster_params(t_obj)
}
\arguments{
\item{t_obj}{A valid raster object}
}
\value{
A tibble with the cube params
}
\description{
Based on the R object associated to a raster object,
                determine its params
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
\keyword{internal}
