% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_raster.R
\name{.sits_raster_preprocess_data}
\alias{.sits_raster_preprocess_data}
\title{Preprocess a set of values retrived from a raster brick}
\usage{
.sits_raster_preprocess_data(
  values.mx,
  band,
  missing_value,
  minimum_value,
  scale_factor,
  stats,
  filter,
  multicores
)
}
\arguments{
\item{values.mx}{Matrix of values retrieved from a brick.}

\item{band}{Band to be processed.}

\item{missing_value}{Missing value for the band.}

\item{minimum_value}{Minimum values for the band.}

\item{scale_factor}{Scale factor for each band (only for raster data).}

\item{stats}{Normalization parameters.}

\item{filter}{Smoothing filter to be applied.}

\item{multicores}{Number of cores to process the time series.}
}
\value{
Matrix with pre-processed values.
}
\description{
Preprocess a set of values retrived from a raster brick
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
\keyword{internal}
