% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_raster_data.R
\name{.sits_raster_satellite_sensor}
\alias{.sits_raster_satellite_sensor}
\title{Tests if satellite and sensor are supported by SITS}
\usage{
.sits_raster_satellite_sensor(satellite, sensor)
}
\arguments{
\item{satellite}{Name of the satellite}

\item{sensor}{Name of the sensor}
}
\value{
TRUE/FALSE
}
\description{
Check if the satellite and sensor are supported by SITS
             looking at the configuration file
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
