% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_raster.R
\name{.sits_raster_scale_matrix_integer}
\alias{.sits_raster_scale_matrix_integer}
\title{Scale the time series values in the case of a matrix}
\usage{
.sits_raster_scale_matrix_integer(values.mx, scale_factor, multicores,
  progress = TRUE)
}
\arguments{
\item{values.mx}{Matrix of values.}

\item{scale_factor}{Scaling factor.}

\item{multicores}{Number of cores.}

\item{progress}{Show progress bar? Default is TRUE.}
}
\value{
Scaled integer matrix.
}
\description{
This function transforms a numerical matrix into an integer one.
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
