% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_raster_stack.R
\name{.sits_raster_stack_cube}
\alias{.sits_raster_stack_cube}
\title{Create a stack cube from a set of files}
\usage{
.sits_raster_stack_cube(satellite, sensor, name, file_info)
}
\arguments{
\item{satellite}{Name of satellite}

\item{sensor}{Name of sensor}

\item{name}{Name of the output data cube}

\item{file_info}{File information}
}
\description{
Create a stack cube from a set of files
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
