% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_raster.R
\name{.sits_raster_stack_info}
\alias{.sits_raster_stack_info}
\title{Obtain the information about files that make up a stack cube}
\usage{
.sits_raster_stack_info(resolution, data_dir, delim, parse_info)
}
\arguments{
\item{resolution}{Resolution of sensor}

\item{data_dir}{Directory where data is located}

\item{delim}{A character to use as delimiter (default = "_")}

\item{parse_info}{The parsing information (see above))}
}
\description{
All image files should have the same spatial resolution
             and same projection. In addition, image file names should
             include information on band and date.
             The timeline and the bands are deduced from this information.
             Examples of valid image names include
             "CB4_64_16D_STK_022024_2018-08-29_2018-09-13_EVI.tif" and
             "B02_2018-07-18.jp2". In each case, the user has to provide
             appropriate parsing information that allows SITS to extract
             the band and the date. In the examples above, the parsing info
             would include "_" as a delimiter. In the first, the names of the
             resulting columns after parsing are "X1, X2, X3, X4, X5, date, X7, band".
             In the second, they are "band, date".
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
