% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_raster_sub_image.R
\name{.sits_raster_sub_image}
\alias{.sits_raster_sub_image}
\title{Find the dimensions and location of a spatial ROI in a data cube}
\usage{
.sits_raster_sub_image(cube, roi)
}
\arguments{
\item{cube}{input data cube.}

\item{roi}{spatial region of interest}
}
\value{
vector with information on the subimage
}
\description{
Find the dimensions and location of a spatial ROI in a data cube
}
\keyword{internal}
