% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_raster_sub_image.R
\name{.sits_raster_sub_image_default}
\alias{.sits_raster_sub_image_default}
\title{Find the dimensions and location of a spatial ROI in a data cube}
\usage{
.sits_raster_sub_image_default(cube)
}
\arguments{
\item{cube}{input data cube.}

\item{sf_region}{spatial region of interest (sf_object)}
}
\value{
vector with information on the subimage
}
\description{
Find the dimensions and location of a spatial ROI in a data cube
}
\keyword{internal}
