% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_raster_sub_image.R
\name{.sits_raster_sub_image_from_block}
\alias{.sits_raster_sub_image_from_block}
\title{Extract a sub_image from a valid block}
\usage{
.sits_raster_sub_image_from_block(block, tile)
}
\arguments{
\item{block}{a valid block with \code{first_col}, \code{first_row},
\code{nrows}, and \code{ncols}.}

\item{tile}{tile of data cube}
}
\value{
sub_image with additional info on first row,
                      first col, nrows, ncols, and crs.
}
\description{
Extract a sub_image from a valid block
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
\keyword{internal}
