% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_raster_sub_image.R
\name{.sits_raster_sub_image_intersects}
\alias{.sits_raster_sub_image_intersects}
\title{Informs if a spatial ROI intersects a data cube}
\usage{
.sits_raster_sub_image_intersects(cube, roi)
}
\arguments{
\item{cube}{input data cube.}

\item{roi}{spatial region of interest}
}
\value{
logical
}
\description{
Informs if a spatial ROI intersects a data cube
}
\keyword{internal}
