% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_raster_data.R
\name{.sits_raster_xy_inside}
\alias{.sits_raster_xy_inside}
\title{Tests if an XY position is inside a ST Raster Brick}
\usage{
.sits_raster_xy_inside(xy, raster.tb)
}
\arguments{
\item{xy}{XY extent compatible with the R raster package.}

\item{raster.tb}{Tibble with metadata information about a raster data set.}
}
\value{
TRUE if XY is inside the raster extent, FALSE otherwise.
}
\description{
Compares an XY position to the extent of a RasterBrick
             described by a raster metadata tibble.
             Returns TRUE if the point is
             inside the extent of the RasterBrick object.
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
