% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_raster_read_write.R
\name{.sits_read_data}
\alias{.sits_read_data}
\title{Read a block of values retrived from a set of raster bricks}
\usage{
.sits_read_data(coverage, samples, ml_model, first_row, n_rows_block,
  stats, filter, multicores)
}
\arguments{
\item{coverage}{input raster coverage}

\item{samples}{tibble with samples}

\item{ml_model}{machine learning model}

\item{first_row}{first row to start reading}

\item{n_rows_block}{number of rows in the block}

\item{stats}{normalization parameters}

\item{filter}{smoothing filter to be applied}

\item{multicores}{number of cores to process the time series}
}
\value{
dist_DT         data.table with values for classification
}
\description{
Read a block of values retrived from a set of raster bricks
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
