% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_config.R
\name{.sits_resolution}
\alias{.sits_resolution}
\title{Retrieve the pixel spatial resolution for a data cube}
\usage{
.sits_resolution(sensor)
}
\arguments{
\item{sensor}{Name of the sensor.}
}
\value{
Vector of (xres, yres).
}
\description{
Retrieve the pixel spatial resolution for a data cube
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
