% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_roi.R
\name{.sits_roi_bbox.ll}
\alias{.sits_roi_bbox.ll}
\title{Find the bounding box for a spatial ROI defined as a lat/lon box}
\usage{
\method{.sits_roi_bbox}{ll}(roi, cube)
}
\arguments{
\item{roi}{spatial region of interest}

\item{cube}{input data cube.}
}
\value{
vector with information on the subimage
}
\description{
Find the bounding box for a spatial ROI defined as a lat/lon box
}
\keyword{internal}
