% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_roi.R
\name{.sits_roi_sf_to_geojson}
\alias{.sits_roi_sf_to_geojson}
\title{Convert a ROI defined as sf object to a geojson object}
\usage{
.sits_roi_sf_to_geojson(roi_sf)
}
\arguments{
\item{roi_sf}{region of interest as sf object}
}
\value{
a geojson object
}
\description{
Convert a ROI defined as sf object to a geojson object
}
\keyword{internal}
