% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_sentinel.R
\name{.sits_s2_aws_add_res}
\alias{.sits_s2_aws_add_res}
\title{Get bands names from items}
\usage{
.sits_s2_aws_add_res(file_info, resolution)
}
\arguments{
\item{resolution}{Resolution of S2 images (10, 20 or 60)}

\item{fileinfo}{Tibble with date, band and path information.}
}
\value{
a \code{tibble} with date, band, res and path information,
 arranged by the date.
}
\description{
Get bands names from items
}
\keyword{internal}
