% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_sentinel.R
\name{.sits_s2_aws_tile_cube}
\alias{.sits_s2_aws_tile_cube}
\title{Get the STAC information corresponding to a tile.}
\usage{
.sits_s2_aws_tile_cube(name, items, collection, resolution, file_info)
}
\arguments{
\item{name}{Name of output data cube.}

\item{items}{\code{STACItemCollection} object returned by rstac.}

\item{collection}{AWS collection}

\item{resolution}{S2 image resolution (10, 20 or 60)}

\item{file_info}{file information information from STAC.}
}
\value{
a \code{tibble} with metadata information about a
                  raster data set.
}
\description{
Get the STAC information corresponding to a tile.
}
\keyword{internal}
