% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_source_api_aws.R
\name{.sits_s2_aws_tiles}
\alias{.sits_s2_aws_tiles}
\alias{.aws_tiles}
\title{Verify items tiles}
\usage{
.aws_tiles(tiles)
}
\arguments{
\item{tiles}{Tile names to be searched.}
}
\value{
a \code{tibble} with information of tiles to be searched in STAC AWS.
}
\description{
Verify items tiles
}
\keyword{internal}
