% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_tibble_aux_functions.R
\name{.sits_samples_bands_check}
\alias{.sits_samples_bands_check}
\title{Check that the requested bands exist in the samples}
\usage{
.sits_samples_bands_check(samples, bands = NULL)
}
\arguments{
\item{samples}{Time series with the samples}

\item{bands}{Requested bands of the data sample}
}
\value{
Checked bands (cube bands if bands are NULL)
}
\description{
Check that the requested bands exist in the samples
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
\keyword{internal}
