% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_tibble.R
\name{.sits_samples_split}
\alias{.sits_samples_split}
\title{Split a sits tibble}
\usage{
.sits_samples_split(samples, validation_split = 0.2)
}
\arguments{
\item{data}{A sits tibble.}
}
\value{
Returns TRUE if data has data.
}
\description{
Add a column to sits tibble indicating if a sample is
training sample or not.
}
\keyword{internal}
