% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_config.R
\name{.sits_satellite_product}
\alias{.sits_satellite_product}
\title{retrieve the satellite associated to a given product}
\usage{
.sits_satellite_product(product)
}
\arguments{
\item{product}{Name of the product}
}
\value{
List of providers associated to a service
}
\description{
retrieve the satellite associated to a given product
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
