% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_config.R
\name{.sits_satellites}
\alias{.sits_satellites}
\title{List the satellites supported}
\usage{
.sits_satellites()
}
\value{
List of satellites supported by SITS
}
\description{
List the satellites supported
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
