% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_satveg.R
\name{.sits_satveg_check}
\alias{.sits_satveg_check}
\title{Check that the SATVEG service is working}
\usage{
.sits_satveg_check()
}
\value{
Boolean that indicates if SATVEG is operating
}
\description{
Check that the SATVEG service is working
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
\keyword{internal}
