% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_satveg.R
\name{.sits_satveg_cube}
\alias{.sits_satveg_cube}
\title{Provides information about one cube of the SATVEG time series service}
\usage{
.sits_satveg_cube(collection)
}
\arguments{
\item{collection}{SATVEG collection to be used.}
}
\description{
Creates a tibble with metadata about a given cube.
}
\keyword{internal}
