% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_SATVEG.R
\name{.sits_satveg_timeline}
\alias{.sits_satveg_timeline}
\title{Retrieve a timeline for the SATVEG service}
\usage{
.sits_satveg_timeline()
}
\description{
Retrieves a time series from the SATVEG service.
}
\author{
Julio Esquerdo, \email{julio.esquerdo@embrapa.br}

Gilberto Camara, \email{gilberto.camara@inpe.br}
}
