% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_satveg.R
\name{.sits_satveg_ts_from_txt}
\alias{.sits_satveg_ts_from_txt}
\title{Retrieve a time series from the SATVEG service}
\usage{
.sits_satveg_ts_from_txt(longitude, latitude, cube)
}
\arguments{
\item{longitude}{The longitude of the chosen location.}

\item{latitude}{The latitude of the chosen location.}

\item{cube}{SATVEG cube}
}
\value{
A tibble containing a time series
}
\description{
Retrieves a time series from the SATVEG service.
}
\author{
Julio Esquerdo, \email{julio.esquerdo@embrapa.br}

Gilberto Camara, \email{gilberto.camara@inpe.br}
}
\keyword{internal}
