% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_raster_read_write.R
\name{.sits_scale_data}
\alias{.sits_scale_data}
\title{Scale the time series values in the case of a matrix}
\usage{
.sits_scale_data(values.mx, scale_factor, multicores)
}
\arguments{
\item{values.mx}{matrix of values}

\item{scale_factor}{scaling factor}

\item{multicores}{number of cores}
}
\value{
values.mx      scaled matrix
}
\description{
this function normalizes one band of the values read from a raster brick
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
