% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_raster_read_write.R
\name{.sits_scale_matrix_integer}
\alias{.sits_scale_matrix_integer}
\title{Scale the time series values in the case of a matrix}
\usage{
.sits_scale_matrix_integer(values.mx, scale_factor, multicores)
}
\arguments{
\item{values.mx}{matrix of values}

\item{scale_factor}{scaling factor}

\item{multicores}{number of cores}
}
\value{
values.mx      scaled integer matrix
}
\description{
this function transforms a numerical matrix into an integer one
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
