% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_tibble_utils.R
\name{.sits_select_bands_}
\alias{.sits_select_bands_}
\title{Filter bands on a sits tibble}
\usage{
.sits_select_bands_(data.tb, bands)
}
\arguments{
\item{data.tb}{A sits tibble metadata and data on time series.}

\item{bands}{The selected bands.}
}
\value{
A tibble in sits format with the selected bands.
}
\description{
Returns a sits tibble with the selected bands.
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
