% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_timeline.R
\name{.sits_select_raster_indexes}
\alias{.sits_select_raster_indexes}
\title{Provide a list of indexes to extract data from a raster-derived data table for classification}
\usage{
.sits_select_raster_indexes(coverage, samples, interval)
}
\arguments{
\item{coverage}{Coverage with input data set}

\item{samples}{tibble with samples used for classification}

\item{interval}{classification interval}
}
\value{
select.lst         list of values to be extracted for each classification interval
}
\description{
Given a list of time indexes that indicate the start and end of the values to
be extracted to classify each band, obtain a list of indexes that will be used to
extract values from a combined distance table (with has all the bands put together)
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
