% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_config.R
\name{.sits_sensor_product}
\alias{.sits_sensor_product}
\title{retrieve the sensor associated to a product}
\usage{
.sits_sensor_product(product)
}
\arguments{
\item{product}{Name of the product}
}
\value{
List of providers associated to a service
}
\description{
retrieve the sensor associated to a product
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
