% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_sentinel.R
\name{.sits_sentinel_aws_check}
\alias{.sits_sentinel_aws_check}
\title{Check access rigths on AWS}
\usage{
.sits_sentinel_aws_check(access_key = NULL, secret_key = NULL, region = NULL)
}
\arguments{
\item{access_key}{AWS access key}

\item{secret_key}{AWS secret key}

\item{region}{AWS region}
}
\value{
TRUE if access is granted
}
\description{
Check access rigths on AWS
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
\keyword{internal}
