% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_sentinel.R
\name{.sits_sentinel_aws_info_tiles}
\alias{.sits_sentinel_aws_info_tiles}
\title{Get information on S2 leval 2A tiles in AWS}
\usage{
.sits_sentinel_aws_info_tiles(tile, resolution, start_date, end_date)
}
\arguments{
\item{tile}{tile}

\item{resolution}{resolution of S2 cube}

\item{start_date}{start_date of the cube}

\item{end_date}{end date of the cube}
}
\value{
tibble with information on s2 tile
}
\description{
Get information on S2 leval 2A tiles in AWS
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
