% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_sentinel.R
\name{.sits_sentinel_aws_tile_cube}
\alias{.sits_sentinel_aws_tile_cube}
\title{Create a data cube for a Sentinel-2 AWS TILE}
\usage{
.sits_sentinel_aws_tile_cube(name, bands, tile, file_info)
}
\arguments{
\item{name}{Name of the data cube.}

\item{bands}{Vector of bands}

\item{tile}{Tile}

\item{file_info}{Tibble with information about the files.}
}
\value{
A tibble with metadata information about a raster data set.
}
\description{
Builds a Sentinel-2 AWS cube
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
\keyword{internal}
