% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_config.R
\name{.sits_services}
\alias{.sits_services}
\title{List the data services available}
\usage{
.sits_services()
}
\value{
List of services supported by SITS
}
\description{
List the data services available
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
