% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_sf.R
\name{.sits_sf_point_to_tibble}
\alias{.sits_sf_point_to_tibble}
\title{Obtain a tibble with latitude/longitude points from POINT geometry}
\usage{
.sits_sf_point_to_tibble(sf_object, label_attr, label)
}
\arguments{
\item{sf_object}{sf object.}

\item{label_attr}{Attribute used as a polygon label}

\item{label}{Label to be assigned if no attribute is provided}
}
\description{
Obtain a tibble with latitude/longitude points from POINT geometry
}
\keyword{internal}
