% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_sf.R
\name{.sits_sf_polygon_to_tibble}
\alias{.sits_sf_polygon_to_tibble}
\alias{.sits_shp_polygon_to_tibble}
\title{Obtain a tibble from POLYGON geometry}
\usage{
.sits_shp_polygon_to_tibble(sf_object, label_attr, label, n_sam_pol, pol_id)
}
\arguments{
\item{sf_object}{sf object linked to a shapefile}

\item{label_attr}{Attribute in the shapefile used as a polygon label}

\item{label}{Label to be assigned to points}

\item{n_sam_pol}{Number of samples per polygon to be read}

\item{pol_id}{ID attribute for polygons shapefile.}
}
\description{
Obtain a tibble from POLYGON geometry
}
\keyword{internal}
