% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_sf.R
\name{.sits_sf_to_tibble}
\alias{.sits_sf_to_tibble}
\title{Obtain a tibble with lat/long points to be retrieved from an sf object}
\usage{
.sits_sf_to_tibble(sf_object, label_attr, label, n_sam_pol, pol_id)
}
\arguments{
\item{sf_object}{sf object .}

\item{label_attr}{Attribute in the sf object used as a polygon label.}

\item{label}{Label to be assigned to points.}

\item{n_sam_pol}{Number of samples per polygon to be read
(for POLYGON or MULTIPOLYGON shapes).}

\item{pol_id}{ID attribute for polygons.}
}
\value{
A sits tibble with points to to be read.
}
\description{
reads a shapefile and retrieves a sits tibble
containing a set of lat/long points for data retrieval
}
\keyword{internal}
