% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_shp.R
\name{.sits_shp_avg_polygon}
\alias{.sits_shp_avg_polygon}
\title{Extracts the time series average by polygon.}
\usage{
.sits_shp_avg_polygon(data)
}
\arguments{
\item{data}{A sits tibble with points time series.}
}
\value{
A sits tibble with the average of all points by each polygon.
}
\description{
This function extracts the average of the automatically
generated points for each polygon in a shapefile.
}
\keyword{internal}
