% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_shp.R
\name{.sits_shp_check_validity}
\alias{.sits_shp_check_validity}
\title{Check the validity of the shape file and return an sf object}
\usage{
.sits_shp_check_validity(shp_file, shp_attr = NULL, label = NULL)
}
\arguments{
\item{shp_file}{SHP file which provides the boundaries of a region.}

\item{shp_attr}{attribute in the shapefile that contains the label}

\item{label}{Label to be used instead of shp_attr}
}
\value{
A valid sf object of POINT or POLYGON geometry.
}
\description{
Check the validity of the shape file and return an sf object
}
\keyword{internal}
