% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_sample_functions.R
\name{.sits_smote}
\alias{.sits_smote}
\title{Oversample a dataset by SMOTE.}
\usage{
.sits_smote(data, target, K = 5, dup_size = 0)
}
\arguments{
\item{data}{Dataset to be oversampled.}

\item{target}{Target data set}

\item{K}{The number of nearest neighbors during sampling process}

\item{dup_size}{The maximum times of synthetic minority instances
over original majority instances in the oversampling.}
}
\value{
A list with the following values.
}
\description{
Lifted from R package "smotefamily"
to reduce number of dependencies in "sits".
}
\references{
Chawla, N., Bowyer, K., Hall, L. and Kegelmeyer, W. 2002.
  SMOTE: Synthetic minority oversampling technique.
  Journal of Artificial Intelligence Research. 16, 321-357.
}
\author{
Wacharasak Siriseriwan <wacharasak.s@gmail.com>
}
\keyword{internal}
