% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_som.R
\name{.sits_som_bayes_est}
\alias{.sits_som_bayes_est}
\title{Probability of a sample belongs to a cluster using bayesian filter}
\usage{
.sits_som_bayes_est(data, kohonen_obj, labelled_neurons, som_radius)
}
\arguments{
\item{data}{A tibble with samples.}

\item{kohonen_obj}{Object that contains all parameters of SOM
provided by "kohonen" package}

\item{labelled_neurons}{A tibble containing information about each neuron.}

\item{som_radius}{Distance in the SOM map to consider neighbours}
}
\value{
Returns the probability of a sample belongs
                        to a cluster based on the class of neuron
                        and its neighborhood.
}
\description{
Computes the probability of a sample belongs
                      to a cluster using bayesian filter.
}
\author{
Lorena Santos, \email{lorena.santos@inpe.br}
}
\keyword{internal}
