% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_kohonen.R
\name{.sits_som_bayesian_neighbourhood}
\alias{.sits_som_bayesian_neighbourhood}
\title{Paint neurons}
\usage{
.sits_som_bayesian_neighbourhood(kohonen_obj, neurons_labelled.tb, result)
}
\arguments{
\item{kohonen_obj}{SOM object from package "kohonen"}

\item{neurons_labelled.tb}{Tibble containing informations about each neuron.}

\item{result}{A tibble with samples.}
}
\value{
Returns the probability that a sample belongs to a
                           cluster based on the class of neuron
                           and its neighborhood.
}
\description{
Compute the probability that a sample belongs to
                      a cluster using bayesian filter.
}
\author{
Lorena Santos, \email{lorena.santos@inpe.br}
}
