% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_kohonen.R
\name{.sits_som_cluster_probability}
\alias{.sits_som_cluster_probability}
\title{Get cluster probability}
\usage{
.sits_som_cluster_probability(data)
}
\arguments{
\item{data}{A sits tibble with info of samples
provided by sits_cluster_som.}
}
\value{
a sits tibble with a new column indicating the probability of a sample belongs to a cluster and the majority cluster of a sample
}
\description{
This function summarizes the probability that a sample belongs
              to a cluster from the label of the neuron
              after several iterations.
}
\author{
Lorena Santos, \email{lorena.santos@inpe.br}
}
