% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_kohonen.R
\name{.sits_som_cluster_unitary}
\alias{.sits_som_cluster_unitary}
\title{SOM neuron unitary}
\usage{
.sits_som_cluster_unitary(som_cluster)
}
\arguments{
\item{som_cluster}{Kohonen map and associated information created by \code{\link[sits]{sits_som_map}}}
}
\value{
Returns a sits tibble with the amount of times a sample was allocated alone in a neuron during the
clustering process.
}
\description{
This function shows how many times a sample was allocated in a neuron alone.
}
\author{
Lorena Santos, \email{lorena.santos@inpe.br}
}
