% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_som.R
\name{.sits_som_label_neurons}
\alias{.sits_som_label_neurons}
\title{Label neurons}
\usage{
.sits_som_label_neurons(data, kohonen_obj)
}
\arguments{
\item{data}{A SITS tibble with info of samples and kohonen.obj.}

\item{kohonen_obj}{The kohonen object returned by kohonen package.}
}
\value{
Tibble with the probability of each neuron belongs to class
and a majority label which is the neuron is labelled.
}
\description{
Compute the probability of a neuron belongs to a class.
The neuron is labelled using the majority voting.
If the neuron is empty, it will labeled as "NoClass".
}
\author{
Lorena Santos, \email{lorena.santos@inpe.br}
}
\keyword{internal}
