% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_kohonen.R
\name{.sits_som_paint_neurons}
\alias{.sits_som_paint_neurons}
\title{Paint neurons}
\usage{
.sits_som_paint_neurons(kohonen_obj)
}
\arguments{
\item{kohonen_obj}{SOM provided by package Kohonen}
}
\value{
SOM object with the colour of the neuron.
}
\description{
Paints all neurons of the last iteration of SOM
}
\author{
Lorena Santos, \email{lorena.santos@inpe.br}
}
