% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_kohonen.R
\name{.sits_som_tie_breaking_neuron_labelling}
\alias{.sits_som_tie_breaking_neuron_labelling}
\title{Tie breaking neuron labelling}
\usage{
.sits_som_tie_breaking_neuron_labelling(
  class_vector,
  kohonen_obj,
  duplicated_id_neuron
)
}
\arguments{
\item{class_vector}{A vector contained the labels of each neuron.}

\item{kohonen_obj}{The kohonen object returned by kohonen package.}

\item{duplicated_id_neuron}{A vector containing the id of each neuron
that has two majority labels.}
}
\value{
Return a new majority label of a neuron based on its neighbourhood
}
\description{
This function does tie breaking for neurons
             that have the same rate to belong a class.
}
\author{
Lorena Santos, \email{lorena.santos@inpe.br}
}
