% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_raster_blocks.R
\name{.sits_split_block_size}
\alias{.sits_split_block_size}
\title{Define the split of the data blocks for multicore processing}
\usage{
.sits_split_block_size(data, ncores)
}
\arguments{
\item{data}{data (data.table or matrix)}

\item{ncores}{number of cores for processing}
}
\value{
block_size.lst  list of pairs of positions (first row, last row) to be assigned to each core
}
\description{
this functions defines the rows of the input data that will be
split to fit to be divided between the different cores
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
