% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_smooth_aux_functions.R
\name{.sits_split_cluster}
\alias{.sits_split_cluster}
\title{parallel raster brick}
\usage{
.sits_split_cluster(
  file,
  block_y_size,
  overlapping_y_size = 0,
  func,
  func_args = NULL,
  out_file,
  cl = NULL,
  ...
)
}
\arguments{
\item{file}{a path to raster (brick) file.}

\item{block_y_size}{number of vertical chunks to break the raster.
For now, only vertical blocking is supported.}

\item{func}{a function that receives RasterBrick and
returns any Raster*.}

\item{out_file}{file path of output raster.}

\item{cl}{snow cluster object}

\item{...}{optional arguments to merge final raster
(see \link[raster]{writeRaster} function)}

\item{overlapping_rows}{number of overlapping rows of each chunk.}

\item{args}{additional arguments to pass to \code{fun} function.}
}
\value{
RasterBrick object
}
\description{
Process chunks of raster brick in parallel.
}
\keyword{internal}
