% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_raster_blocks.R
\name{.sits_split_data}
\alias{.sits_split_data}
\title{Split a data.table or a matrix for multicore processing}
\usage{
.sits_split_data(data, ncores, ml_model)
}
\arguments{
\item{data}{data (data.table or matrix)}

\item{ncores}{number of cores for processing}

\item{ml_model}{machine learning model which is part of the object}
}
\value{
block_size.lst  list of pairs of positions (first row, last row) to be assigned to each core
}
\description{
this functions splits a data.table into a list of chunks for multicore processing
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
