% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_stac.R
\name{.sits_stac_bands}
\alias{.sits_stac_bands}
\title{Get bands names from items}
\usage{
.sits_stac_bands(items, bands, source, collection)
}
\arguments{
\item{items}{a \code{STACItemCollection} object returned by rstac
package.}

\item{bands}{a \code{character} vector with the bands name.}

\item{source}{Data source}

\item{collection}{a \code{character} with the collection to be searched.}

\item{sensor}{a \code{character} with sensor name.}
}
\value{
a \code{STACItemCollection} object representing the search
                  by rstac.
}
\description{
Get bands names from items
}
\keyword{internal}
