% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_stac.R
\name{.sits_stac_collection}
\alias{.sits_stac_collection}
\title{Get information from collection}
\usage{
.sits_stac_collection(url = NULL, collection = NULL, bands = NULL, ...)
}
\arguments{
\item{url}{a \code{character} representing a URL for the BDC catalog.}

\item{collection}{a \code{character} with the collection to be searched.}

\item{bands}{a \code{character} with the bands names to be filtered.}

\item{...}{other parameters to be passed for specific types.}
}
\value{
a \code{STACCollection} object returned by rstac.
}
\description{
Get information from collection
}
\keyword{internal}
