% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_stac.R
\name{.sits_stac_datetime}
\alias{.sits_stac_datetime}
\title{Datetime format}
\usage{
.sits_stac_datetime(start_date, end_date)
}
\arguments{
\item{start_date}{a \code{character} corresponds to the initial date when the
cube will be created.}

\item{end_date}{a \code{character} corresponds to the final date when the
cube will be created.}
}
\value{
a \code{character} formatted as parameter to STAC requisition.
}
\description{
Datetime format
}
\keyword{internal}
