% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_stac.R
\name{.sits_stac_group}
\alias{.sits_stac_group}
\title{Create a group of items}
\usage{
.sits_stac_group(items, fields)
}
\arguments{
\item{items}{a \code{STACItemCollection} object returned by rstac package.}

\item{fields}{a \code{character} vector with the names of fields to be
grouped.}
}
\value{
a \code{list} in which each index corresponds to a group with
 its corresponding \code{STACItemCollection} objects.
}
\description{
Create a group of items
}
\keyword{internal}
