% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_stac.R
\name{.sits_stac_items}
\alias{.sits_stac_items}
\title{Get information from items}
\usage{
.sits_stac_items(
  url = NULL,
  collection = NULL,
  tiles = NULL,
  roi = NULL,
  start_date = NULL,
  end_date = NULL,
  ...
)
}
\arguments{
\item{url}{a \code{character} representing a URL for the BDC catalog.}

\item{collection}{a \code{character} with the collection to be searched.}

\item{tiles}{a \code{character} with the names of the tiles.}

\item{roi}{the "roi" parameter defines a region of interest. It can be
an \code{sfc} or \code{sf} object from sf package, a \code{character} with
a GeoJSON following the rules from RFC 7946, or a \code{vector}
bounding box \code{vector} with named XY values
("xmin", "xmax", "ymin", "ymax").}

\item{start_date}{a \code{character} corresponds to the initial date when the
cube will be created.}

\item{end_date}{a \code{character} corresponds to the final date when the
cube will be created.}
}
\value{
a \code{STACItemCollection} object representing the search
 by rstac.
}
\description{
Get information from items
}
\keyword{internal}
