% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_stac.R
\name{.sits_stac_tile_cube}
\alias{.sits_stac_tile_cube}
\title{Get the STAC information corresponding to a tile.}
\usage{
.sits_stac_tile_cube(name, collection, collection_info, items, file_info)
}
\arguments{
\item{name}{Name of output data cube.}

\item{collection}{Name of image collection in  BDC.}

\item{collection_info}{STACCollection object returned by rstac.}

\item{items}{STACItemCollection object returned by rstac.}

\item{file_info}{Tibble with the information from STAC.}
}
\value{
a \code{tibble} with metadata information about a
 raster data set.
}
\description{
Get the STAC information corresponding to a tile.
}
\keyword{internal}
