% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_stac.R
\name{.sits_stac_to_fileinfo}
\alias{.sits_stac_to_fileinfo}
\title{Format assets}
\usage{
.sits_stac_to_fileinfo(items)
}
\arguments{
\item{items}{a \code{STACItemCollection} object returned by rstac package.}
}
\value{
a \code{tibble} with date, band and path information, arranged
 by the date.
}
\description{
Format assets
}
\keyword{internal}
