% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_stac.R
\name{.sits_stac_toupper}
\alias{.sits_stac_toupper}
\title{Converts bands name to upper case}
\usage{
.sits_stac_toupper(collection_info)
}
\arguments{
\item{collection_info}{a \code{STACCollection} object returned by rstac.
package.}
}
\value{
a \code{STACCollection} object with the bands in upper case.
}
\description{
Converts bands name to upper case
}
\keyword{internal}
