% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_space_time_operations.R
\name{.sits_sub_image_from_bbox}
\alias{.sits_sub_image_from_bbox}
\title{Extract a sub_image from a bounding box and a cube}
\usage{
.sits_sub_image_from_bbox(bbox, cube)
}
\arguments{
\item{bbox}{bounding box for a region of interest}

\item{cube}{data cube}
}
\value{
sub_image with additional info on first row,
                      first col, nrows, ncols
}
\description{
Extract a sub_image from a bounding box and a cube
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
\keyword{internal}
