% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_config.R
\name{.sits_supported_sensors}
\alias{.sits_supported_sensors}
\title{List all the sensors supported}
\usage{
.sits_supported_sensors()
}
\value{
List of sensors supported by SITS
}
\description{
List all the sensors supported
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
