% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_tibble_utils.R
\name{.sits_test_tibble}
\alias{.sits_test_tibble}
\title{Tests if a sits tibble is valid}
\usage{
.sits_test_tibble(data.tb)
}
\arguments{
\item{data.tb}{A sits tibble.}
}
\value{
Returns TRUE if data.tb has data.
}
\description{
Tests if a sits tibble exists or has data inside.
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}

Rolf Simoes, \email{rolf.simoes@inpe.br}
}
