% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_tibble.R
\name{.sits_tibble_csv}
\alias{.sits_tibble_csv}
\title{Store the results of CSV samples that could not be read}
\usage{
.sits_tibble_csv()
}
\value{
A tibble to store the result of classifications.
}
\description{
Create an empty tibble to store the results of classification.
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}

Rolf Simoes, \email{rolf.simoes@inpe.br}
}
