% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_tibble_utils.R
\name{.sits_tibble_prediction}
\alias{.sits_tibble_prediction}
\title{Create an empty tibble to store the results of predictions}
\usage{
.sits_tibble_prediction(data.tb, class_info.tb, pred.mtx, interval)
}
\arguments{
\item{data.tb}{a tibble with the input data}

\item{class_info.tb}{a tibble with the information on classification}

\item{pred.mtx}{the result of the classification (one class per column and one row per interval)}

\item{interval}{the time interval between two classifications}
}
\value{
predic.tb       a tibble to store the predictions
}
\description{
Create a tibble to store the results of predictions
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}

Rolf Simoes, \email{rolf.simoes@inpe.br}
}
