% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_tibble_aux_functions.R
\name{.sits_tibble_rename}
\alias{.sits_tibble_rename}
\title{Rename a tibble to use "cube" instead of "coverage"}
\usage{
.sits_tibble_rename(data)
}
\arguments{
\item{data}{A sits tibble.}
}
\description{
Rename a tibble to use "cube" instead of "coverage"
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
\keyword{internal}
