% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_tibble.R
\name{.sits_tibble_test}
\alias{.sits_tibble_test}
\title{Tests if a sits tibble is valid}
\usage{
.sits_tibble_test(data)
}
\arguments{
\item{data}{A sits tibble.}
}
\value{
Returns TRUE if data has data.
}
\description{
Tests if a sits tibble exists or has data inside.
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}

Rolf Simoes, \email{rolf.simoes@inpe.br}
}
\keyword{internal}
