% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_timeline.R
\name{.sits_time_index}
\alias{.sits_time_index}
\title{Create a list of time indexes from the dates index}
\usage{
.sits_time_index(dates_index.lst, timeline, bands)
}
\arguments{
\item{dates_index.lst}{A list of dates with the subsets of the input data}

\item{timeline}{The timeline of the data set}

\item{bands}{Bands used for classification}
}
\value{
time_index.lst  The subsets of the timeline
}
\description{
Create a list of time indexes from the dates index
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
