% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_timeline.R
\name{.sits_timeline_check}
\alias{.sits_timeline_check}
\title{Checks that the timeline of all time series of a data set are equal}
\usage{
.sits_timeline_check(data)
}
\arguments{
\item{data}{Either a sits tibble}
}
\value{
TRUE if the length of time series is unique
}
\description{
This function tests if all time series in a sits tibble
have the same number of samples
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
\keyword{internal}
