% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_timeline.R
\name{.sits_timeline_check_cube}
\alias{.sits_timeline_check_cube}
\title{Check cube timeline against requested start and end dates}
\usage{
.sits_timeline_check_cube(cube, start_date, end_date)
}
\arguments{
\item{cube}{Data cube metadata.}

\item{start_date}{Start date of the period.}

\item{end_date}{End date of the period.}
}
\value{
A vector with corrected start and end dates
}
\description{
Tests if required start and end dates are available in
             the data cube
}
\keyword{internal}
