% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_timeline.R
\name{.sits_timeline_date_format}
\alias{.sits_timeline_date_format}
\title{Find if the date information is correct}
\usage{
.sits_timeline_date_format(date_band)
}
\arguments{
\item{file_info}{a tibble with date and band information}
}
\value{
Tibble with corrected date information
}
\description{
Given a information about dates, check if the date can be
             interpreted by lubridate
}
\keyword{internal}
