% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_timeline.R
\name{.sits_timeline_dist_indexes}
\alias{.sits_timeline_dist_indexes}
\title{Indexes to extract data from a distance table for classification}
\usage{
.sits_timeline_dist_indexes(class_info, ntimes)
}
\arguments{
\item{class_info}{Tibble with classification information.}

\item{ntimes}{Number of time instances.}
}
\value{
List of indexes to be extracted for each classification interval.
}
\description{
Given a list of time indexes that indicate
             the start and end of the values to
             be extracted to classify each band,
             obtain a list of indexes that will be used to
             extract values from a combined distance tibble
             (which has all the bands put together).
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
\keyword{internal}
