% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_timeline.R
\name{.sits_timeline_during}
\alias{.sits_timeline_during}
\title{Find the subset of a timeline that is contained
       in an interval defined by start_date and end_date}
\usage{
.sits_timeline_during(timeline, start_date = NULL, end_date = NULL)
}
\arguments{
\item{timeline}{A valid timeline}

\item{start_date}{A date which encloses the start of timeline}

\item{end_date}{A date which encloses the end of timeline}
}
\value{
A timeline
}
\description{
Find the subset of a timeline that is contained
       in an interval defined by start_date and end_date
}
\keyword{internal}
