% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_timeline.R
\name{.sits_timeline_index_blocks}
\alias{.sits_timeline_index_blocks}
\title{Find the time index of the blocks to be extracted for classification}
\usage{
.sits_timeline_index_blocks(class_info)
}
\arguments{
\item{class_info}{Tibble with information required for classification.}
}
\value{
Indexes of the input data set associated to each time interval
used for classification.
}
\description{
Obtains the indexes of the blocks for each time interval
associated with classification.
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
\keyword{internal}
