% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_timeline.R
\name{.sits_timeline_indexes}
\alias{.sits_timeline_indexes}
\title{Given a start and end date, find the indexes in a timeline}
\usage{
.sits_timeline_indexes(timeline, start_date = NULL, end_date = NULL)
}
\arguments{
\item{timeline}{A valid timeline}

\item{start_date}{A date inside the timeline}

\item{end_date}{A date inside the timeline}
}
\value{
Named vector with start and end indexes
}
\description{
Given a start and end date, find the indexes in a timeline
}
\keyword{internal}
