% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_timeline.R
\name{.sits_timeline_indexes_interval}
\alias{.sits_timeline_indexes_interval}
\title{Given a timeline and an interval, find the corresponding indexes of a new timeline}
\usage{
.sits_timeline_indexes_interval(timeline, interval = "16 day")
}
\arguments{
\item{timeline}{A valid timeline}

\item{interval}{A valid intervak}
}
\value{
Named vector with indexes for the new timeline
}
\description{
Given a timeline and an interval, find the corresponding indexes of a new timeline
}
\keyword{internal}
