% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_timeline.R
\name{.sits_timeline_match}
\alias{.sits_timeline_match}
\title{Find dates in the input data cube that match those of the patterns}
\usage{
.sits_timeline_match(timeline, ref_start_date, ref_end_date, num_samples)
}
\arguments{
\item{timeline}{Timeline of input observations (vector).}

\item{ref_start_date}{Reference for starting the classification.}

\item{ref_end_date}{Reference for ending the classification.}

\item{num_samples}{Number of samples.}
}
\value{
A list of breaks that will be applied to the input data set.
}
\description{
For correct classification, the input data set
             should be aligned to that of the reference data set.
             This function aligns these data sets.
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
\keyword{internal}
