% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_timeline.R
\name{.sits_timeline_num_samples}
\alias{.sits_timeline_num_samples}
\title{Find number of samples, given a timeline and an interval}
\usage{
.sits_timeline_num_samples(timeline, interval = "12 month")
}
\arguments{
\item{timeline}{Timeline of input observations (vector).}

\item{interval}{Period to match the data to the patterns.}
}
\value{
The number of measures during the chosen interval.
}
\description{
This function retrieves the number of samples
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
