% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_timeline.R
\name{.sits_timeline_valid_date}
\alias{.sits_timeline_valid_date}
\title{Test if date fits with the timeline}
\usage{
.sits_timeline_valid_date(date, timeline)
}
\arguments{
\item{date}{A date.}

\item{timeline}{A vector of reference dates.}
}
\value{
Is this is valid starting date?
}
\description{
A timeline is a list of dates where observations are available.
This function estimates if a date is valid by comparing it to the timeline.
If the date's estimate is not inside the timeline and the difference between
the date and the first date of timeline is greater than the acquisition
interval of the timeline, the date is not valid.
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
\keyword{internal}
