% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_twdtw.R
\name{.sits_to_twdtw}
\alias{.sits_to_twdtw}
\title{Export data to be used by the dtwSat package}
\usage{
.sits_to_twdtw(data.tb)
}
\arguments{
\item{data.tb}{A tibble in sits format with time series to be converted to TWTDW time series.}
}
\value{
A time series in TWDTW format (an object of the twdtwTimeSeries class).
}
\description{
Converts data from a sits tibble to an instance of a TWDTW time series class.
}
\author{
Victor Maus, \email{vwmaus1@gmail.com}

Gilberto Camara, \email{gilberto.camara@inpe.br}
}
