% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_utils.R
\name{.sits_transform_samples}
\alias{.sits_transform_samples}
\title{Transform samples to wgs84}
\usage{
.sits_transform_samples(samples, crs)
}
\arguments{
\item{samples}{Samples to be retrieved.}

\item{crs}{A coordinate reference system of samples.
The provided crs could be a character
(e.g, "EPSG:4326" or "WGS84" or a proj4string), or a
a numeric with the EPSG code (e.g. 4326).
This parameter only works for 'csv' or data.frame'
samples. Default is 4326.}
}
\value{
A tibble with tranformed points.
}
\description{
Transforming samples points to wgs84 and replace the longitude
and latitude values.
}
\keyword{internal}
