% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_metrics_utils.R
\name{.sits_transform_vector_metrics}
\alias{.sits_transform_vector_metrics}
\title{Transforms a character vector of a metrics into a list metrics}
\usage{
.sits_transform_vector_metrics(metrics)
}
\arguments{
\item{metrics}{a \code{character} vector of metrics.}
}
\value{
A named \code{list} of metrics, where name is band and the values the
metrics name.
}
\description{
Transforms a character vector of a metrics into a list metrics
}
\keyword{internal}
