% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_get_data.R
\name{.sits_ts_from_cube}
\alias{.sits_ts_from_cube}
\title{Extract a time series from a cube}
\usage{
.sits_ts_from_cube(
  cube,
  longitude,
  latitude,
  start_date,
  end_date,
  bands,
  label,
  .prefilter
)
}
\arguments{
\item{cube}{Data cube}

\item{longitude}{Longitude of point}

\item{latitude}{Latitude of point}

\item{start_date}{starting date for the time series}

\item{end_date}{end date for the time series}

\item{bands}{Bands to be retrieved}

\item{label}{Label to be assigned to the series}

\item{.prefilter}{Prefilter (used for SATVEG)}
}
\value{
A valid sits tibble
}
\description{
Extract a time series from a cube
}
