% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_get_data.R
\name{.sits_ts_from_raster}
\alias{.sits_ts_from_raster}
\title{Extract a time series from a ST raster data set}
\usage{
.sits_ts_from_raster(cube, sf_object, bands)
}
\arguments{
\item{cube}{Metadata describing a raster data cube.}

\item{sf_object}{sf object.}

\item{bands}{Bands to be retrieved.}
}
\value{
A sits tibble with the time series.
}
\description{
Retrieve a set of time series for a raster data cube.
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
