% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_shp.R
\name{.sits_ts_from_raster_shp}
\alias{.sits_ts_from_raster_shp}
\title{Extract a time series from}
\usage{
.sits_ts_from_raster_shp(cube, sf_shape, bands)
}
\arguments{
\item{cube}{Metadata describing a raster data cube.}

\item{sf_shape}{sf object.}

\item{bands}{Bands to be retrieved.}
}
\value{
A sits tibble with the time series.
}
\description{
Retrieve a set of time series for a raster data cube.
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
\keyword{internal}
