% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_raster_time_series.R
\name{.sits_ts_from_raster_xy}
\alias{.sits_ts_from_raster_xy}
\title{Extract a time series for a set of Raster Layers}
\usage{
.sits_ts_from_raster_xy(raster.tb, xy, longitude, latitude,
  label = "NoClass")
}
\arguments{
\item{raster.tb}{A tibble with metadata information about a raster data set.}

\item{xy}{A matrix with X/Y coordinates.}

\item{longitude}{Longitude of the chosen location.}

\item{latitude}{Latitude of the chosen location.}

\item{label}{Label to attach to the time series.}
}
\value{
A sits tibble with the time series.
}
\description{
This function extracts a sits time series from a set of
              Raster Layers whose metadata is stored in a tibble
              created by the sits_STraster function.
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
