% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_SATVEG.R
\name{.sits_ts_from_SATVEG}
\alias{.sits_ts_from_SATVEG}
\title{Retrieve a time series from the SATVEG service}
\usage{
.sits_ts_from_SATVEG(longitude, latitude, name, prefilter)
}
\arguments{
\item{longitude}{double - the longitude of the chosen location}

\item{latitude}{double - the latitude of the chosen location}

\item{name}{name of the desired coverage in SATVEG (see configuration file)}

\item{prefilter}{string ("0" - none, "1" - no data correction, "2" - cloud correction, "3" - no data and cloud correction)}
}
\value{
status         TRUE if no problems are detected
}
\description{
Retrieves a time series from the SATVEG service
}
\author{
Julio Esquerdo, \email{julio.esquerdo@embrapa.br}

Gilberto Camara, \email{gilberto.camara@inpe.br}
}
