% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_raster_time_series.R
\name{.sits_ts_fromRaster}
\alias{.sits_ts_fromRaster}
\title{Extract a time series for a set of Raster Layers}
\usage{
.sits_ts_fromRaster(raster.tb, longitude, latitude, start_date, end_date,
  label = "NoClass")
}
\arguments{
\item{raster.tb}{A tibble with metadata information about a raster data set}

\item{longitude}{Longitude of the chosen location}

\item{latitude}{Latitude of the chosen location}

\item{start_date}{date - the start of the period}

\item{end_date}{date - the end of the period}

\item{label}{Label to attach to the time series}
}
\value{
data.tb         SITS tibble with the time series
}
\description{
This function extracts a SITS time series from a set of
              Raster Layers whose metadata is stored in a tibble
              created by the sits_STraster function
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
