% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_raster_time_series.R
\name{.sits_ts_fromRasterCSV}
\alias{.sits_ts_fromRasterCSV}
\title{Extract a time series for a set of Raster Layers}
\usage{
.sits_ts_fromRasterCSV(raster.tb, file)
}
\arguments{
\item{raster.tb}{A tibble with metadata describing a spatio-temporal data set}

\item{file}{A CSV file with lat/long locations to be retrieved}
}
\value{
data.tb         SITS tibble with the time series
}
\description{
This function extracts a SITS time series from a set of
              Raster Layers whose metadata is stored in a tibble
              created by the sits_STraster function
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
