% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_twdtw.R
\name{.sits_twdtw_from_tibble}
\alias{.sits_twdtw_from_tibble}
\title{Export data to be used by the dtwSat package}
\usage{
.sits_twdtw_from_tibble(samples)
}
\arguments{
\item{samples}{A tibble in sits format with time series
to be converted to TWTDW time series.}
}
\value{
An object of the twdtwTimeSeries class).
}
\description{
Converts sits tibble to an instance of a TWDTW class.
}
\author{
Victor Maus, \email{vwmaus1@gmail.com}

Gilberto Camara, \email{gilberto.camara@inpe.br}
}
\keyword{internal}
