% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_usgs.R
\name{.sits_usgs_filter_datetime}
\alias{.sits_usgs_filter_datetime}
\title{Filter datetime in STAC items}
\usage{
.sits_usgs_filter_datetime(items, start_date, end_date)
}
\arguments{
\item{items}{a \code{STACItemCollection} object returned by rstac
package.}

\item{start_date}{a \code{character} with the initial date to search in
collection.}

\item{end_date}{a \code{character} with the final date to searched in
collection.}
}
\value{
a \code{STACItemCollection} object with datetime filtered.
}
\description{
Filter datetime in STAC items
}
\keyword{internal}
