% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_usgs.R
\name{.sits_usgs_fix_bands}
\alias{.sits_usgs_fix_bands}
\title{Fix bands name in hrefs provided by USGS}
\usage{
.sits_usgs_fix_bands(items)
}
\arguments{
\item{items}{a \code{STACItemCollection} object returned by rstac
package.}
}
\value{
a \code{STACItemCollection} object with URLs fixed.
}
\description{
Fix bands name in hrefs provided by USGS
}
\keyword{internal}
