% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_landsat.R
\name{.sits_usgs_fix_href}
\alias{.sits_usgs_fix_href}
\title{Fix href on URLs provided by USGS}
\usage{
.sits_usgs_fix_href(items)
}
\arguments{
\item{items}{a \code{STACItemCollection} object returned by rstac
package.}
}
\value{
a \code{STACItemCollection} object with URLs fixed.
}
\description{
Fix href on URLs provided by USGS
}
\keyword{internal}
