% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_landsat.R
\name{.sits_usgs_tile_cube}
\alias{.sits_usgs_tile_cube}
\title{Get the STAC information corresponding to a tile.}
\usage{
.sits_usgs_tile_cube(name, items, collection, file_info)
}
\arguments{
\item{name}{Name of output data cube.}

\item{items}{\code{STACItemCollection} object returned by rstac.}

\item{collection}{Image collection in USGS}

\item{file_info}{file information with date/time.}
}
\value{
a \code{tibble} with metadata information about a
                 raster data set.
}
\description{
Get the STAC information corresponding to a tile.
}
\keyword{internal}
