% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_raster_read_write.R
\name{.sits_write_raster_values}
\alias{.sits_write_raster_values}
\title{Write the values and probs into raster files}
\usage{
.sits_write_raster_values(output.lst, prediction_DT, labels, int_labels,
  time, first_row, multicores)
}
\arguments{
\item{output.lst}{List with value layers and probability bricks.}

\item{prediction_DT}{A data.table with predicted probabilities for each class.}

\item{labels}{Class labels.}

\item{int_labels}{Integer values corresponding to labels.}

\item{time}{Interval to be written to file.}

\item{first_row}{Initial row of the output layer to write block.}

\item{multicores}{Number of cores to process the time series.}
}
\value{
Updated list with value layers and probability bricks.
}
\description{
Write the raster values to the outout files.
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
