% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_wtss.R
\name{.sits_wtss_cube}
\alias{.sits_wtss_cube}
\title{Provides information about one cube of the WTSS service}
\usage{
.sits_wtss_cube(URL, name, collection, bands)
}
\arguments{
\item{URL}{URL of the service provider.}

\item{name}{Name of the cube.}

\item{collection}{Collection to be searched in the data source}

\item{bands}{Bands to be included}
}
\value{
a \code{tibble} with cube information.
}
\description{
Uses the WTSS services to print information and save metadata
about a chosen cube.
}
\keyword{internal}
