% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_smooth_aux_functions.R
\name{.smth_block_filename}
\alias{.smth_block_filename}
\alias{.smth_filename}
\title{Create a filename to store an output block from smoothing operation}
\usage{
.smth_filename(tile, output_dir, block)
}
\arguments{
\item{tile}{Output tile to be written.}

\item{output_dir}{Directory where block will be written.}

\item{block}{Block designation.}
}
\value{
returns a filename witn
}
\description{
Create a filename to store an output block from smoothing operation
}
\keyword{internal}
