% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_source_api.R
\name{.source_bands_to_sits}
\alias{.source_bands_to_sits}
\title{Convert bands names from cube to SITS}
\usage{
.source_bands_to_sits(source, collection, bands)
}
\arguments{
\item{source}{Name of the STAC provider}

\item{collection}{Name of sensor}

\item{bands}{Bands requested to be read}
}
\value{
Data cube tile with SITS bands
}
\description{
Convert the name of the band used by the origin data cube
             to the name used by SITS
}
\keyword{internal}
