% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_source_api.R
\name{.source_bands_to_source}
\alias{.source_bands_to_source}
\title{Convert bands names as source names}
\usage{
.source_bands_to_source(source, collection, bands)
}
\arguments{
\item{source}{Data source (one of "SATVEG", "LOCAL", "BDC", "AWS", "USGS"
"DEAFRICA", "PROBS").}

\item{collection}{Collection to be searched in the data source.}

\item{bands}{Bands requested to be read.}
}
\value{
an \code{character} with converted bands.
}
\description{
Convert bands used by sits to source names.
}
\keyword{internal}
