% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_source_api.R
\name{.source_cube}
\alias{.source_cube}
\title{Convert bands names from cube to SITS}
\usage{
.source_cube(source, ...)
}
\arguments{
\item{source}{Name of the STAC provider}

\item{...}{Aditional parameters}
}
\value{
Data cube tile with SITS bands
}
\description{
TODO: add
}
\keyword{internal}
