% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_source_api.R
\name{.source_items_bands_select}
\alias{.source_items_bands_select}
\title{...}
\usage{
.source_items_bands_select(source, collection, items, bands, ...)
}
\arguments{
\item{source}{Name of the STAC provider}

\item{collection}{...}

\item{items}{...}

\item{bands}{...}

\item{...}{...}
}
\value{
...
}
\description{
TODO: document
}
\keyword{internal}
