% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_source_api.R
\name{.source_items_fileinfo}
\alias{.source_items_fileinfo}
\title{...}
\usage{
.source_items_fileinfo(source, items, ..., collection = NULL)
}
\arguments{
\item{source}{Name of the STAC provider}

\item{items}{...}

\item{collection}{...}
}
\value{
...
}
\description{
TODO: document
}
\keyword{internal}
