% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_source_api.R
\name{.source_items_new}
\alias{.source_items_new}
\title{Create an items object}
\usage{
.source_items_new(source, collection, ...)
}
\arguments{
\item{source}{Data source (one of "SATVEG", "LOCAL", "BDC", "AWS",
"USGS", "DEAFRICA", "PROBS").}

\item{collection}{Collection to be searched in the data source.}

\item{...}{Additional parameters.}
}
\value{
a \code{STACItemCollection} object returned by rstac.
}
\description{
Create an items object. In the case of STAC APIs, this function
 is responsible for making the request to the server.
}
\keyword{internal}
