% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_source_api.R
\name{.source_items_tile_cube}
\alias{.source_items_tile_cube}
\title{...}
\usage{
.source_items_tile_cube(source, collection, name, tile_items, file_info, ...)
}
\arguments{
\item{source}{Name of the STAC provider}

\item{collection}{...}

\item{name}{...}

\item{tile_items}{...}

\item{file_info}{...}

\item{...}{...}
}
\value{
...
}
\description{
TODO: document
}
\keyword{internal}
