% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_source_api.R
\name{.source_new}
\alias{.source_new}
\title{Creates a s3 class}
\usage{
.source_new(source)
}
\arguments{
\item{source}{Data source (one of "SATVEG", "LOCAL", "BDC", "AWS",
"USGS", "DEAFRICA", "PROBS").}
}
\value{
a \code{character} with a specified class.
}
\description{
Create an S3 class with information returned from the
 configuration file.
}
\keyword{internal}
