% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_stac.R
\name{.stac_format_datetime}
\alias{.stac_format_datetime}
\title{Datetime format}
\usage{
.stac_format_datetime(start_date, end_date)
}
\arguments{
\item{start_date}{Initial date for cube.}

\item{end_date}{Final date for the cube.}
}
\value{
a \code{character} formatted as parameter to STAC requisition.
}
\description{
Datetime format
}
\keyword{internal}
