% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_stac.R
\name{.stac_format_platform}
\alias{.stac_format_platform}
\title{Platform format}
\usage{
.stac_format_platform(source, collection, platform)
}
\arguments{
\item{source}{A \code{character} value referring to a valid data source.}

\item{collection}{Image collection.}

\item{platform}{Sensor platform}
}
\value{
a \code{character} formatted as parameter to STAC requisition.
}
\description{
Platform format
}
\keyword{internal}
