% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_stac.R
\name{.stac_get_roi}
\alias{.stac_get_roi}
\title{Get bbox and intersects parameters}
\usage{
.stac_get_roi(roi)
}
\arguments{
\item{roi}{the "roi" parameter defines a region of interest. It can be
an \code{sfc} or \code{sf} object from sf package, or a \code{vector}
bounding box \code{vector} with named XY values
("xmin", "xmax", "ymin", "ymax").}
}
\value{
A named \code{list} with the values of the intersection and bbox
        parameters. If bbox is supplied, the intersection parameter gets
        NULL, otherwise bbox gets NULL if intersects is specified.
}
\description{
Get bbox and intersects parameters
}
\keyword{internal}
