% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_source_api_stac.R
\name{.stac_items_query}
\alias{.stac_items_query}
\title{...}
\usage{
.stac_items_query(
  source,
  collection,
  name,
  bands,
  tiles,
  bbox,
  start_date,
  end_date,
  ...
)
}
\arguments{
\item{source}{Name of the STAC provider}

\item{collection}{...}

\item{name}{...}

\item{bands}{...}

\item{tiles}{...}

\item{bbox}{...}

\item{start_date}{...}

\item{end_date}{...}

\item{...}{...}
}
\value{
...
}
\description{
...
}
