% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_tile_api.R
\name{.tile}
\alias{.tile}
\title{Tile API}
\usage{
.tile(cube)
}
\arguments{
\item{cube}{A \code{cube} or a \code{tile}.}
}
\value{
The first row of a cube.
}
\description{
A \code{cube} consists of multiple tiles stacked together as rows of a
\code{tibble}. Get first tile of a cube. This function should be called
by all tile API function to ensure that only one tile will be processed.
}
