% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_tile_api.R
\name{.tile_as_sf.raster_cube}
\alias{.tile_as_sf.raster_cube}
\title{Convert raster cube \code{bbox} to a sf polygon object.}
\usage{
\method{.tile_as_sf}{raster_cube}(tile)
}
\arguments{
\item{tile}{A tile.}
}
\description{
Convert raster cube \code{bbox} to a sf polygon object.
}
