% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_tile_api.R
\name{.tile_band_conf}
\alias{.tile_band_conf}
\title{Tile API}
\usage{
.tile_band_conf(tile, band)
}
\arguments{
\item{tile}{A tile.}

\item{band}{Band character vector.}
}
\value{
band_conf or band_cloud_conf
}
\description{
Get a band definition from config.
}
