% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_tile_api.R
\name{.tile_band_filter}
\alias{.tile_band_filter}
\title{Tile API}
\usage{
.tile_band_filter(tile, band)
}
\arguments{
\item{tile}{A tile.}

\item{band}{A region of interest (ROI).}
}
\value{
tile
}
\description{
Filter file_info entries of a given \code{band}.
}
