% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_tile_api.R
\name{.tile_bands}
\alias{.tile_bands}
\title{Get sorted unique bands from file_info.}
\usage{
.tile_bands(tile)
}
\arguments{
\item{tile}{A tile.}
}
\value{
names of bands in the tile
}
\description{
Get sorted unique bands from file_info.
}
\keyword{internal}
