% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_tile_api.R
\name{.tile_chunks_create()}
\alias{.tile_chunks_create()}
\alias{.tile_chunks_create}
\title{Create chunks of a tile to be processed}
\usage{
.tile_chunks_create(tile, overlap)
}
\arguments{
\item{tile}{tile to be processed}

\item{overlap}{overlap between tiles}
}
\value{
set of chunks to be read from the file
}
\description{
Create chunks of a tile to be processed
}
\keyword{internal}
