% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_tile_api.R
\name{.tile_cloud_read_block()}
\alias{.tile_cloud_read_block()}
\alias{.tile_cloud_read_block}
\title{Read and preprocess a \code{block} of cloud values from
file_info rasters.}
\usage{
.tile_cloud_read_block(tile, block)
}
\arguments{
\item{tile}{A tile.}

\item{block}{A block list with (col, row, ncols, nrows).}
}
\value{
set of values of a band of a tile in a block
}
\description{
Read and preprocess a \code{block} of cloud values from
file_info rasters.
}
\keyword{internal}
