% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_tile_api.R
\name{.tile_collection}
\alias{.tile_collection}
\title{Tile API
Get tile collection field.}
\usage{
.tile_collection(tile)
}
\arguments{
\item{tile}{A tile.}
}
\value{
character
}
\description{
Tile API
Get tile collection field.
}
