% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_tile_api.R
\name{.tile_derived_class()}
\alias{.tile_derived_class()}
\alias{.tile_derived_class}
\title{Get derived class of a tile.}
\usage{
.tile_derived_class(tile)
}
\arguments{
\item{tile}{A tile.}
}
\value{
derived class
}
\description{
Get derived class of a tile.
}
\keyword{internal}
