% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_tile_api.R
\name{.tile_derived_class}
\alias{.tile_derived_class}
\title{Tile API}
\usage{
.tile_derived_class(tile)
}
\arguments{
\item{tile}{A tile.}
}
\value{
character
}
\description{
Get derived class of a tile.
}
