% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_tile_api.R
\name{.tile_during()}
\alias{.tile_during()}
\alias{.tile_during}
\title{Is any date of tile's timeline between 'start_date'
and 'end_date'?}
\usage{
.tile_during(tile, start_date, end_date)
}
\arguments{
\item{tile}{A tile.}

\item{start_date, end_date}{Date of start and end.}
}
\value{
logical
}
\description{
Tile API
}
\keyword{internal}
