% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_tile_api.R
\name{.tile_during}
\alias{.tile_during}
\title{Tile API}
\usage{
.tile_during(tile, start_date, end_date)
}
\value{
logical
}
\description{
Is any date of tile's timeline between 'start_date'
and 'end_date'?
}
