% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_tile_api.R
\name{.tile_end_date}
\alias{.tile_end_date}
\title{Get end date from file_info.}
\usage{
.tile_end_date(tile)
}
\arguments{
\item{tile}{A tile.}
}
\value{
date
}
\description{
Get end date from file_info.
}
\keyword{internal}
