% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_tile_api.R
\name{.tile_eo_from_files}
\alias{.tile_eo_from_files}
\title{read an EO tile from files}
\usage{
.tile_eo_from_files(files, fid, bands, date, base_tile, update_bbox)
}
\arguments{
\item{files}{files to be read}

\item{fid}{file ID}

\item{bands}{bands to be read in the files}

\item{date}{date associated to the file}

\item{base_tile}{reference tile used in the operation}

\item{update_bbox}{should bbox be updated?}
}
\value{
a base tile
}
\description{
read an EO tile from files
}
\keyword{internal}
