% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_tile_api.R
\name{.tile_eo_merge_blocks}
\alias{.tile_eo_merge_blocks}
\title{Merge block from an EO tile}
\usage{
.tile_eo_merge_blocks(
  files,
  bands,
  base_tile,
  block_files,
  multicores,
  update_bbox
)
}
\arguments{
\item{files}{files to be merged}

\item{bands}{bands to be used in the files}

\item{base_tile}{reference tile used in the operation}

\item{block_files}{files associated with the the blocks}

\item{multicores}{multicores for processing}

\item{update_bbox}{should bbox be updated?}
}
\value{
an EO tile with merged blocks
}
\description{
Merge block from an EO tile
}
\keyword{internal}
