% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_tile_api.R
\name{.tile_file_info}
\alias{.tile_file_info}
\alias{.tile_file_info<-}
\title{Tile API}
\usage{
.tile_file_info(tile)

.tile_file_info(tile) <- value
}
\value{
file_info

tile
}
\description{
Get tile file_info field.

Set tile file_info field.
}
