% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_tile_api.R
\name{.tile_filter_bands}
\alias{.tile_filter_bands}
\title{Tile API}
\usage{
.tile_filter_bands(tile, bands)
}
\arguments{
\item{tile}{A tile.}

\item{bands}{Band names to be filtered.}
}
\value{
tile
}
\description{
Filter file_info entries of a given \code{band}.
}
