% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_tile_api.R
\name{.tile_filter_dates}
\alias{.tile_filter_dates}
\title{Filter file_info entries by date}
\usage{
.tile_filter_dates(tile, dates)
}
\arguments{
\item{tile}{A tile.}

\item{dates}{Desired date}
}
\value{
file_info entries
}
\description{
Filter file_info entries by date
}
\keyword{internal}
