% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_tile_api.R
\name{.tile_filter_interval}
\alias{.tile_filter_interval}
\title{Filter file_info entries by 'start_date' and 'end_date.'}
\usage{
.tile_filter_interval(tile, start_date, end_date)
}
\arguments{
\item{tile}{A tile.}

\item{start_date, end_date}{Date of start and end.}
}
\value{
file_info entries
}
\description{
Filter file_info entries by 'start_date' and 'end_date.'
}
\keyword{internal}
