% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_tile_api.R
\name{.tile_filter_spatial}
\alias{.tile_filter_spatial}
\title{Tile API}
\usage{
.tile_filter_spatial(tile, roi)
}
\arguments{
\item{tile}{A tile.}

\item{roi}{A region of interest (ROI).}
}
\value{
tile
}
\description{
Filter file_info entries that intersect \code{roi} parameter.
}
