% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_tile_api.R
\name{.tile_filter_temporal}
\alias{.tile_filter_temporal}
\title{Tile API}
\usage{
.tile_filter_temporal(tile, start_date, end_date)
}
\arguments{
\item{tile}{A tile.}

\item{start_date, end_date}{Date of start and end.}
}
\value{
tile
}
\description{
Filter file_info entries by 'start_date' and 'end_date.'
}
