% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_tile_api.R
\name{.tile_intersects}
\alias{.tile_intersects}
\title{Tile API}
\usage{
.tile_intersects(tile, roi)
}
\arguments{
\item{tile}{A tile.}

\item{roi}{A region of interest (ROI).}
}
\value{
logical
}
\description{
Does tile \code{bbox} intersect \code{roi} parameter?
}
