% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_tile_api.R
\name{.tile_labels}
\alias{.tile_labels}
\alias{.tile_labels<-}
\title{Tile API}
\usage{
.tile_labels(tile)

.tile_labels(tile) <- value
}
\arguments{
\item{tile}{A tile.}

\item{value}{Label character vector.}
}
\value{
character

tile
}
\description{
Get tile labels field.

Set tile labels field.
}
