% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_tile_api.R
\name{.tile_nrows}
\alias{.tile_nrows}
\title{Tile API}
\usage{
.tile_nrows(tile)
}
\arguments{
\item{tile}{A tile.}
}
\value{
integer
}
\description{
Get number of image columns from \code{nrows} field (if it exists)
or from first \code{nrows} in file_info.
}
