% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_tile_api.R
\name{.tile_read_block}
\alias{.tile_read_block}
\title{Tile API}
\usage{
.tile_read_block(tile, band, block, replace_by_minmax)
}
\arguments{
\item{tile}{A tile.}

\item{band}{Band character vector.}

\item{block}{A block list with (col, row, ncols, nrows).}

\item{replace_by_minmax}{Should values out of minimum and maximum range
be replaced by minimum and maximum value? If \code{FALSE}, values out
of range will be replaced by \code{NA}.}
}
\value{
numeric
}
\description{
Read and preprocess a \code{block} of \code{band} values from
file_info rasters.
}
