% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_tile_api.R
\name{.tile_size}
\alias{.tile_size}
\title{Tile API}
\usage{
.tile_size(tile)
}
\arguments{
\item{tile}{A tile.}
}
\value{
\code{block}
}
\description{
Get the size of tile from \code{ncols} and \code{nrows} fields
(if they exist) or from first entry in file_info.
}
