% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_tile_api.R
\name{.tile_spatial_filter}
\alias{.tile_spatial_filter}
\title{Tile API}
\usage{
.tile_spatial_filter(tile, roi)
}
\value{
logical
}
\description{
Filter file_info entries that intersect 'roi' parameter.
}
