% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_tile_api.R
\name{.tile_start_date}
\alias{.tile_start_date}
\title{Get first date from tile}
\usage{
.tile_start_date(tile)
}
\arguments{
\item{tile}{A tile.}
}
\value{
date
}
\description{
Get first date from tile
}
\keyword{internal}
