% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_tile_api.R
\name{.tile_temporal_filter}
\alias{.tile_temporal_filter}
\title{Tile API}
\usage{
.tile_temporal_filter(tile, start_date, end_date)
}
\value{
tile
}
\description{
Filter file_info entries by 'start_date' and 'end_date.'
}
