% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_tile_api.R
\name{.tile_timeline}
\alias{.tile_timeline}
\title{Get unique timeline from file_info.}
\usage{
.tile_timeline(tile)
}
\arguments{
\item{tile}{A tile.}
}
\value{
a timeline
}
\description{
Get unique timeline from file_info.
}
\keyword{internal}
