% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_torch_linear.R
\name{.torch_linear_batch_norm_relu}
\alias{.torch_linear_batch_norm_relu}
\title{Torch module for linear MLP}
\usage{
.torch_linear_batch_norm_relu(input_dim, output_dim)
}
\arguments{
\item{input_dim}{Input dimension of neural net.}

\item{output_dim}{Output dimension of neural net.}
}
\value{
A linear tensor block.
}
\description{
Defines a torch module composed of; (a) linear transformation;
(b) batch normalization; (c) relu activation
}
\author{
Charlotte Pelletier, \email{charlotte.pelletier@univ-ubs.fr}

Gilberto Camara, \email{gilberto.camara@inpe.br}

Rolf Simoes, \email{rolf.simoes@inpe.br}

Felipe Souza, \email{lipecaso@gmail.com}
}
\keyword{internal}
