% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_tuning.R
\name{.tuning_params_as_tibble}
\alias{.tuning_params_as_tibble}
\title{Convert hyper-parameters list to a tibble}
\usage{
.tuning_params_as_tibble(params)
}
\value{
A named list with provided parameters
}
\description{
Generate a tibble (one row per trial) with all model parameters
}
\keyword{internal}
