% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_tuning.R
\name{.tuning_pick_random}
\alias{.tuning_pick_random}
\title{Get random hyper-parameter}
\usage{
.tuning_pick_random(trial, params)
}
\description{
Evaluate params by returning random numbers according
to params definition returned by \code{sits_tuning_hparams}
}
\keyword{internal}
