% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_source_api_usgs.R
\name{.usgs_filter_datetime}
\alias{.usgs_filter_datetime}
\title{Filter datetime in STAC items}
\usage{
.usgs_filter_datetime(items, datetime)
}
\arguments{
\item{items}{a \code{STACItemCollection} object returned by rstac
package.}

\item{datetime}{a \code{character} ...}
}
\value{
a \code{STACItemCollection} object with datetime filtered.
}
\description{
Filter datetime in STAC items
}
\keyword{internal}
