% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_view.R
\name{.view_class_cube}
\alias{.view_class_cube}
\title{Return the r_object associated to classified cube (RATified)}
\usage{
.view_class_cube(class_cube, tile)
}
\arguments{
\item{class_cube}{object of class "classified image"}

\item{tile}{tile to be plotted (in case of a multi-tile cube)}
}
\value{
a \code{raster} object
}
\description{
Return the r_object associated to classified cube (RATified)
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
\keyword{internal}
