% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_view.R
\name{.view_get_colors}
\alias{.view_get_colors}
\title{Return the colors associated to the classified image}
\usage{
.view_get_colors(labels, legend, palette)
}
\arguments{
\item{labels}{labels of the classified cube}

\item{legend}{named vector that associates labels to colors}

\item{palette}{palette provided in the configuration file}
}
\description{
Return the colors associated to the classified image
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
\keyword{internal}
