% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_view.R
\name{.view_resample_size}
\alias{.view_resample_size}
\title{Return the cell size for the image to be resamples}
\usage{
.view_resample_size(nrows, ncols, ntiles)
}
\arguments{
\item{nrows}{Number of rows in the input image.}

\item{ncols}{Number of cols in the input image.}

\item{ntiles}{Number of tiles in the input image.}
}
\value{
Cell size for x and y coordinates.
}
\description{
Return the cell size for the image to be resamples
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
\keyword{internal}
