% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_view.R
\name{.view_reshape_band}
\alias{.view_reshape_band}
\title{Reduce B/W image for visualisation and load files in tempdir}
\usage{
.view_reshape_band(band_file, date, max_Mbytes)
}
\arguments{
\item{band_file}{file for B/W band.}

\item{date}{date reference for the file}

\item{max_Mbytes}{maximum number of megabytes to be shown in leaflet}
}
\value{
Raster Stack with RGB object
}
\description{
Reduce B/W image for visualisation and load files in tempdir
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
\keyword{internal}
