% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_view.R
\name{.view_reshape_image}
\alias{.view_reshape_image}
\title{Reduce the cube size for visualisation and load files in tempdir}
\usage{
.view_reshape_image(cube, rgb_files, max_Mbytes)
}
\arguments{
\item{rgb_files}{vector with RGB files.}

\item{max_Mbytes}{maximum number of megabytes to be shown in leaflet}

\item{x}{object of "raster_cube" or "classified image"}
}
\description{
Reduce the cube size for visualisation and load files in tempdir
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
\keyword{internal}
