% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_view.R
\name{.view_reshape_rgb}
\alias{.view_reshape_rgb}
\title{Reduce an RGB image for visualisation and load files in tempdir}
\usage{
.view_reshape_rgb(rgb_files, date, max_Mbytes)
}
\arguments{
\item{rgb_files}{vector with RGB files.}

\item{date}{date reference for the file}

\item{max_Mbytes}{maximum number of megabytes to be shown in leaflet}
}
\value{
Raster Stack with RGB object
}
\description{
Reduce an RGB image for visualisation and load files in tempdir
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
\keyword{internal}
