% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_transitions.R
\name{Edges}
\alias{Edges}
\title{Edges transition function (start and end are equal in the series)}
\usage{
Edges(target_class)
}
\arguments{
\item{target_class}{Target class name for the transition}
}
\value{
expression that creates a EDGES transition call
}
\description{
Edges transition function (start and end are equal in the series)
}
\author{
Felipe Carlos, \email{efelipecarlos@gmail.com}

Felipe Carvalho, \email{felipe.carvalho@inpe.br}

Gilberto Camara, \email{gilberto.camara@inpe.br}
}
\keyword{internal}
