% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_transitions.R
\name{Ends}
\alias{Ends}
\title{Ends transition function}
\usage{
Ends(target_class)
}
\arguments{
\item{target_class}{Target class name for the transition}
}
\value{
expression that creates a ENDS transition call
}
\description{
Ends transition function
}
\author{
Felipe Carlos, \email{efelipecarlos@gmail.com}

Felipe Carvalho, \email{felipe.carvalho@inpe.br}

Gilberto Camara, \email{gilberto.camara@inpe.br}
}
\keyword{internal}
