% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_tile_api.R
\name{eo_cube_config}
\alias{eo_cube_config}
\alias{.conf_eo_band_exists}
\alias{.conf_eo_band}
\title{Config functions for \code{eo_cube}}
\usage{
.conf_eo_band_exists(source, collection, band)

.conf_eo_band(source, collection, band)
}
\arguments{
\item{source}{Source name.}

\item{collection}{Collection name.}

\item{band}{Band name.}
}
\value{
Configuration value.
}
\description{
These are syntactic sugar functions to easily access config options for
bands of \code{eo_cube} cubes. \code{eo_cubes} are a S3 class representation
for an Earth Observation cube. It is the primary data used to obtain a
classification map.
}
\details{
The config entries of a \code{eo_cube} are located in
\code{sources -> <SOURCE> -> collections -> <COLLECTION>} key.
Values for \code{source}, \code{collection}, and \code{band} are uppercase.
}
\section{Functions}{
\itemize{
\item \code{.conf_eo_band_exists()}: Tests if a \code{band} entry exists in config
for some \code{source} and \code{collection}. If neither \code{source}
nor \code{collection} entry are found in config, an error is thrown.
Use \code{.conf_exists()} to test for \code{source} and \code{collection}
existence.

\item \code{.conf_eo_band()}: Get a config value of for a \code{band} located
in a \code{source} and \code{collection}. If the \code{band} is not
found, a default value will be returned from config. If neither
\code{source} nor \code{collection} entry are found in config, an
error is thrown. Use \code{.conf_exists()} to test for \code{source} and
\code{collection} existence.

}}
\examples{
if (sits_run_examples()) {
# tests if 'BDC -> MOD13Q1-6 -> NDVI' key exists in config
.conf_eo_band_exists(
  source = "BDC",
  collection = "MOD13Q1-6",
  band = "NDVI"
)
# get configuration for band NDVI of 'BDC -> MOD13Q1-6' collection
x <- .conf_eo_band(
  source = "BDC",
  collection = "MOD13Q1-6",
  band = "NDVI"
)
}

}
\seealso{
Band accessors: \link{band_accessors}

Other config functions: 
\code{\link{config_api}},
\code{\link{derived_cube_config}}
}
\concept{config functions}
\keyword{internal}
