% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_WTSPS.R
\docType{methods}
\name{executeAlgorithm}
\alias{executeAlgorithm}
\alias{executeAlgorithm,WTSPS-method}
\title{Returns the status of a processing algorithm}
\usage{
executeAlgorithm(object, query)

\S4method{executeAlgorithm}{WTSPS}(object, query)
}
\arguments{
\item{object}{A WTSPS object}

\item{query}{query with algorithm and required parameters}
}
\description{
Returns the status of a processing algorithm
}
