% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_tile_api.R
\name{fi_api}
\alias{fi_api}
\title{File info API}
\arguments{
\item{fi}{A \code{file_info}.}

\item{...}{Parameters to be evaluated accordingly to \code{file_info} type.}
}
\description{
A \code{file_info} represents a set of raster references. It is a
\code{tibble} containing metadata describing images (e.g. bbox, band,
size, spatial resolution, date) and a path to access the resource. There
are two types of \code{file_info}: a \code{eo_cube} file info and a
\code{derived_cube} file_info. The main differences are in the metadata
stored by each type.
}
\examples{
\dontrun{
}

}
\concept{file info API}
\keyword{internal}
