% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_file_info.R
\name{file_info_api}
\alias{file_info_api}
\alias{.fi}
\alias{.fi<-}
\title{File info API}
\usage{
.fi(tile)

.fi(tile) <- value
}
\arguments{
\item{cube}{Input data cube.}

\item{bands}{Bands to be filtered}

\item{dates}{Dates to be filtered.}

\item{fid}{Feature id (fid) to be filtered.}
}
\value{
See each function description.
}
\description{
Set of functions to handling \code{file_info}.
}
\section{Functions}{
\itemize{
\item \code{.fi()}: Get \code{file_info} from a \code{tile}. If a
\code{cube} is informed, get the \code{file_info} from the first
\code{tile}. Returns a \code{file_info}.

\item \code{.fi(tile) <- value}: Set \code{'file_info'} field of a \code{tile}.
If a \code{cube} is informed, considers only the first \code{tile}.

}}
\author{
Rolf Simoes, \email{rolf.simoes@inpe.br}
}
\keyword{internal}
