% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_WTSPS.R
\docType{methods}
\name{getServerURL}
\alias{getServerURL}
\alias{getServerURL,WTSPS-method}
\title{Returns the WTSPS object's server URL}
\usage{
getServerURL(object)

\S4method{getServerURL}{WTSPS}(object)
}
\arguments{
\item{object}{A WTSPS object}
}
\value{
serverURL an URL character of a WTSPS server
}
\description{
Returns the WTSPS object's server URL
}
