% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_histogram.R
\name{hist.sits}
\alias{hist.sits}
\title{Histogram}
\usage{
\method{hist}{sits}(x, ...)
}
\arguments{
\item{x}{Object of classes "sits".}

\item{...}{Further specifications for \link{hist}.}
}
\value{
A summary of the sits tibble.
}
\description{
This is a generic function. Parameters depend on the specific
type of input.
}
\examples{
if (sits_run_examples()) {
    # create a data cube from local files
    data_dir <- system.file("extdata/raster/mod13q1", package = "sits")
    cube <- sits_cube(
        source = "BDC",
        collection = "MOD13Q1-6.1",
        data_dir = data_dir
    )
    hist(cube)
}

}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
