% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_transitions.R
\name{Holds}
\alias{Holds}
\title{Holds a class in a given time}
\usage{
Holds(target_class)
}
\arguments{
\item{target_class}{Target class name to keep}
}
\value{
expression that creates a HOLDS transition call
}
\description{
Holds a class in a given time
}
\author{
Felipe Carlos, \email{efelipecarlos@gmail.com}

Felipe Carvalho, \email{felipe.carvalho@inpe.br}

Gilberto Camara, \email{gilberto.camara@inpe.br}
}
\keyword{internal}
