% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_imputation.R
\name{impute_kalman}
\alias{impute_kalman}
\title{Replace NA values by Kalman Smoothing}
\usage{
impute_kalman(data = NULL)
}
\arguments{
\item{data}{A time series vector or matrix}
}
\value{
A set of filtered time series using
                      the imputation function.
}
\description{
Remove NA by Kalman Smoothing
}
