% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_imputation.R
\name{impute_locf}
\alias{impute_locf}
\title{Replace NA values by Last Observation Carried Forward}
\usage{
impute_locf(data = NULL)
}
\arguments{
\item{data}{A time series vector or matrix}
}
\value{
A set of filtered time series using
                      the imputation function.
}
\description{
Remove NA by Last Observation Carried Forward
}
