% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_imputation.R
\name{impute_mean}
\alias{impute_mean}
\title{Remove NA using mean}
\usage{
impute_mean(data = NULL)
}
\arguments{
\item{data}{A time series vector or matrix}
}
\value{
A set of filtered time series using
                      the imputation function.
}
\description{
Remove NA using mean
}
\examples{
# fill NA values in a time series with the mean
impute_mean(c(1, 2, NA, 4, NA, 6))

}
\seealso{
Other imputation:
\code{\link[=impute_linear]{impute_linear()}},
\code{\link[=impute_mean_window]{impute_mean_window()}},
\code{\link[=impute_median]{impute_median()}}
}
\author{
Felipe Carlos, \email{efelipecarlos@gmail.com}
}
\concept{imputation}
