% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_imputation.R
\name{impute_median}
\alias{impute_median}
\title{Remove NA using median}
\usage{
impute_median(data = NULL)
}
\arguments{
\item{data}{A time series vector or matrix}
}
\value{
A set of filtered time series using
                      the imputation function.
}
\description{
Remove NA using median
}
\examples{
# fill NA values in a time series with the median
impute_median(c(1, 2, NA, 4, NA, 6))

}
\seealso{
Other imputation:
\code{\link[=impute_linear]{impute_linear()}},
\code{\link[=impute_mean]{impute_mean()}},
\code{\link[=impute_mean_window]{impute_mean_window()}}
}
\author{
Felipe Carlos, \email{efelipecarlos@gmail.com}
}
\concept{imputation}
