% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_imputation.R
\name{impute_weighted_moving_average}
\alias{impute_weighted_moving_average}
\title{Replace NA values by Weighted Moving Average}
\usage{
impute_weighted_moving_average(data = NULL)
}
\arguments{
\item{data}{A time series vector or matrix}
}
\value{
A set of filtered time series using
                      the imputation function.
}
\description{
Remove NA by Weighted Moving Average
}
