% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_transitions.R
\name{Keeps}
\alias{Keeps}
\title{Keeps transition function}
\usage{
Keeps(target_class)
}
\arguments{
\item{target_class}{Target class name to keep}
}
\value{
expression that creates a KEEPS transition call
}
\description{
Keeps transition function
}
\author{
Felipe Carlos, \email{efelipecarlos@gmail.com}

Felipe Carvalho, \email{felipe.carvalho@inpe.br}

Gilberto Camara, \email{gilberto.camara@inpe.br}
}
\keyword{internal}
