% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_WTSPS.R
\docType{methods}
\name{listAlgorithms}
\alias{listAlgorithms}
\alias{listAlgorithms,WTSPS-method}
\title{Returns a list of algorithms contained in WTSPS}
\usage{
listAlgorithms(object)

\S4method{listAlgorithms}{WTSPS}(object)
}
\arguments{
\item{object}{A WTSPS object}
}
\description{
Returns a list of algorithms contained in WTSPS
}
