% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_config.R
\name{parallel_minimum_requets}
\alias{parallel_minimum_requets}
\alias{.config_parallel_minimum_requests}
\title{Retrieve the minimum requests to do in parallel}
\usage{
.config_parallel_minimum_requests()
}
\value{
get minimum times the parallel requests will be done
}
\description{
Retrieve the minimum requests to do in parallel
}
\keyword{internal}
