% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_config.R
\name{parallel_requests}
\alias{parallel_requests}
\alias{.config_parallel_requests}
\title{Retrieve the parallel requests number}
\usage{
.config_parallel_requests()
}
\value{
get parallel requests
}
\description{
Retrieve the parallel requests number
}
\keyword{internal}
