% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_transitions.R
\name{Peaks}
\alias{Peaks}
\title{Peaks transition function}
\usage{
Peaks(target_class)
}
\arguments{
\item{target_class}{Target class name for the transition}
}
\value{
expression that creates a PEAKS transition call
}
\description{
Peaks transition function
}
\author{
Felipe Carlos, \email{efelipecarlos@gmail.com}

Felipe Carvalho, \email{felipe.carvalho@inpe.br}

Gilberto Camara, \email{gilberto.camara@inpe.br}
}
\keyword{internal}
