% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_tile_api.R
\name{period_api}
\alias{period_api}
\alias{.period_check}
\alias{.period_val}
\alias{.period_unit}
\title{period API}
\usage{
.period_check(period)

.period_val(period)

.period_unit(period)
}
\arguments{
\item{period}{A \code{character}.}
}
\value{
\code{.period_check()}: nothing.

\code{.period_val()}: numeric value of a period.

\code{.period_unit()}: description of unit of a period.
}
\description{
According to ISO-8601 a duration is the amount of intervening time
in a time interval. Here, we use a simplified representation of a duration
that we call \code{period}.
}
\details{
\code{period} is represented by the the regular expression
\code{^P[0-9]+[DMY]$}. \code{P} is the period designator placed at the
start of the string. \code{[0-9]+} is the integer value and is followed
by a \code{period} unit: \code{Y} is the year designator, \code{M} is the
month designator, and \code{D} is the day designator.
}
\section{Functions}{
\itemize{
\item \code{.period_check()}: Check if a character string is a valid
\code{period}.

\item \code{.period_val()}: Return the value part of a
\code{period}.

\item \code{.period_unit()}: Return the unit of a \code{period}.
Can be one of \code{'day'}, \code{'month'}, or \code{'year'}.

}}
\examples{
\dontrun{
.period_check("P16D") # valid
.period_check("P1M10D") # error: invalid period format
.period_val("P16D") # 16
.period_val("P2M") # 2
.period_val("P1Y") # 1
.period_unit("P16D") # day
.period_unit("P2M") # month
.period_unit("P1Y") # year
}

}
\seealso{
Other data types: 
\code{\link{data_type}}
}
\concept{data types}
\keyword{internal}
