% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_transitions.R
\name{Persist}
\alias{Persist}
\title{Persist transition function}
\usage{
Persist(target_class, size)
}
\arguments{
\item{target_class}{Target class name for the transition}

\item{size}{Temporal interval size}
}
\value{
expression that creates a PERSIST transition call
}
\description{
Persist transition function
}
\author{
Felipe Carlos, \email{efelipecarlos@gmail.com}

Felipe Carvalho, \email{felipe.carvalho@inpe.br}

Gilberto Camara, \email{gilberto.camara@inpe.br}
}
\keyword{internal}
