% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipe.R
\name{\%>\%}
\alias{\%>\%}
\title{Pipe}
\arguments{
\item{lhs, rhs}{A visualization and a function to apply to it.}
}
\value{
Apply lhs as input to rhs function
}
\description{
Magrittr compound assignment pipe-operator.
}
