% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_plot.R
\name{plot.brick_cube}
\alias{plot.brick_cube}
\title{Generic interface for plotting brick cubes}
\usage{
\method{plot}{brick_cube}(x, y, ..., red, green, blue, time = 1)
}
\arguments{
\item{x}{object of class "brick_cube"}

\item{y}{ignored}

\item{...}{further specifications for \link{plot}.}

\item{red}{band for red color.}

\item{green}{band for green color.}

\item{blue}{band for blue color.}

\item{time}{temporal instance to be plotted}
}
\value{
mapview object
}
\description{
plots a stack cube using terra
}
\examples{
\dontrun{
# retrieve two files with NDVI and EVI from MODIS
ndvi_file <- c(system.file("extdata/raster/mod13q1/sinop-ndvi-2014.tif",
    package = "sits"
))
evi_file <- c(system.file("extdata/raster/mod13q1/sinop-evi-2014.tif",
    package = "sits"
))
# retrieve the timeline
data("timeline_2013_2014")
# create a data cube
sinop_2014 <- sits_cube(
    type = "BRICK",
    name = "sinop-2014",
    timeline = timeline_2013_2014,
    satellite = "TERRA",
    sensor = "MODIS",
    bands = c("NDVI", "EVI"),
    files = c(ndvi_file, evi_file)
)
# plot the data cube
plot(sinop_2014, red = "EVI", green = "EVI", blue = "EVI")
}

}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
