% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_plot.R
\name{plot.class_cube}
\alias{plot.class_cube}
\title{Plot classified images}
\usage{
\method{plot}{class_cube}(
  x,
  y,
  ...,
  tile = x[["tile"]][[1L]],
  roi = NULL,
  legend = NULL,
  palette = "Spectral",
  scale = 1,
  max_cog_size = 1024L,
  legend_position = "outside"
)
}
\arguments{
\item{x}{Object of class "class_cube".}

\item{y}{Ignored.}

\item{...}{Further specifications for \link{plot}.}

\item{tile}{Tile to be plotted.}

\item{roi}{Spatial extent to plot (see note)}

\item{legend}{Named vector that associates labels to colors.}

\item{palette}{A RColorBrewer or "cols4all" palette}

\item{scale}{Relative scale (0.4 to 1.0) of plot text}

\item{max_cog_size}{Maximum size of COG overviews (lines or columns)}

\item{legend_position}{Where to place the legend (default = "outside")}
}
\value{
A  color map, where each pixel has the color
                        associated to a label, as defined by the legend
                        parameter.
}
\description{
plots a classified raster using tmap.
}
\note{
To see which color palettes are supported, please run cols4all::c4a_gui().
The following optional parameters are available to allow for detailed
      control over the plot output:
\itemize{
\item \code{graticules_labels_size}: size of coord labels (default = 0.8)
\item \code{legend_title_size}: relative size of legend title (default = 1.0)
\item \code{legend_text_size}: relative size of legend text (default = 1.0)
\item \code{legend_bg_color}: color of legend background (default = "white")
\item \code{legend_bg_alpha}: legend opacity (default = 0.5)
}
To define a \code{roi} use one of:
     \itemize{
       \item{A path to a shapefile with polygons;}
       \item{A path to a single-layer geopackage file ("gpkg") with polygons;}
       \item{A \code{sfc} or \code{sf} object from \code{sf} package;}
       \item{A \code{SpatExtent} object from \code{terra} package;}
       \item{A named \code{vector} (\code{"lon_min"},
            \code{"lat_min"}, \code{"lon_max"}, \code{"lat_max"}) in WGS84;}
       \item{A named \code{vector} (\code{"xmin"}, \code{"xmax"},
             \code{"ymin"}, \code{"ymax"}) with XY coordinates.}
}
}
\examples{
if (sits_run_examples()) {
    # create a random forest model
    rfor_model <- sits_train(samples_modis_ndvi, sits_rfor())
    # create a data cube from local files
    data_dir <- system.file("extdata/raster/mod13q1", package = "sits")
    cube <- sits_cube(
        source = "BDC",
        collection = "MOD13Q1-6.1",
        data_dir = data_dir
    )
    # classify a data cube
    probs_cube <- sits_classify(
        data = cube, ml_model = rfor_model, output_dir = tempdir()
    )
    # label cube with the most likely class
    label_cube <- sits_label_classification(
        probs_cube,
        output_dir = tempdir()
    )
    # plot the resulting classified image
    plot(label_cube)
}
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
