% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_plot.R
\name{plot.keras_model}
\alias{plot.keras_model}
\title{Generic interface for plotting a Keras model}
\usage{
\method{plot}{keras_model}(x, y, ...)
}
\arguments{
\item{x}{Object of class "keras_model"}

\item{y}{ignored}

\item{...}{further specifications for \link{plot}.}
}
\value{
The plot itself.
}
\description{
plots a deep learning model developed using keras
}
\examples{
\donttest{
# Get a set of samples
samples_ndvi_evi <- sits_select(samples_modis_4bands,
                                bands = c("NDVI", "EVI"))

# train a deep learning model
dl_model <- sits_train(samples_ndvi_evi, ml_method = sits_mlp())
plot(dl_model)
}

}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
