% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_plot.R
\name{plot.patterns}
\alias{plot.patterns}
\title{Generic interface for ploting patterns}
\usage{
\method{plot}{patterns}(x, y, ...)
}
\arguments{
\item{x}{object of class "patterns"}

\item{y}{ignored}

\item{...}{further specifications for \link{plot}.}
}
\value{
Input sits tibble (useful for chaining functions).
}
\description{
Given a sits tibble with a set of patterns, plot them.
}
\examples{
\donttest{
# Read a set of samples with 2 classes ("Cerrado" and "Pasture")
# Plot the patterns
plot(sits_patterns(cerrado_2classes))
}
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
