% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_plot.R
\name{plot.radd_model}
\alias{plot.radd_model}
\title{...}
\usage{
\method{plot}{radd_model}(x, y)
}
\arguments{
\item{x}{Object of class "patterns".}

\item{y}{Ignored.}
}
\value{
A plot object produced by ggplot2
                      with one average pattern per label.
}
\description{
...
}
\note{
....
}
\examples{
if (sits_run_examples()) {
    # plot patterns

}
}
