% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_plot.R
\name{plot.sits_cluster}
\alias{plot.sits_cluster}
\title{Plot a dendrogram cluster}
\usage{
\method{plot}{sits_cluster}(x, ..., cluster, cutree_height, palette)
}
\arguments{
\item{x}{sits tibble with cluster indexes.}

\item{...}{Further specifications for \link{plot}.}

\item{cluster}{cluster object produced by `sits_cluster` function.}

\item{cutree_height}{dashed horizontal line to be drawn
indicating the height of dendrogram cutting.}

\item{palette}{HCL color palette.}
}
\value{
The dendrogram object.
}
\description{
Plot a dendrogram
}
\examples{
if (sits_run_examples()) {
     samples <- sits_cluster_dendro(cerrado_2classes,
                bands = c("NDVI", "EVI"))
}

}
\author{
Rolf Simoes, \email{rolfsimoes@gmail.com}
}
