% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_plot.R
\name{plot.sits_model}
\alias{plot.sits_model}
\title{Message for models whose plots are not available}
\usage{
\method{plot}{sits_model}(x, ...)
}
\arguments{
\item{x}{Object of class "sits_model".}

\item{...}{Further specifications for \link{plot}.}
}
\value{
Called for side effects
}
\description{
Plots trees in an extreme gradient boosting model.
}
\examples{
if (sits_run_examples()) {
    # Retrieve the samples for Mato Grosso
    # train an extreme gradient boosting
    svm_model <- sits_train(samples_modis_ndvi,
        ml_method = sits_svm()
    )
    plot(svm_model)
}
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
