% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_plot.R
\name{plot.som_confusion}
\alias{plot.som_confusion}
\title{Plot information about confunsion between clusters}
\usage{
\method{plot}{som_confusion}(x, y, ..., data, title = "Confusion by cluster")
}
\arguments{
\item{data}{Table containing the percentage of mixture between the clusters
(produced by \code{\link[sits]{sits_som_evaluate_cluster}})}

\item{title}{Title of plot. default is ""Confusion by cluster"".}
}
\value{

}
\description{
Plot a bar graph with informations about each cluster.
The percentage of mixture between the clusters.
}
\examples{
\donttest{
# Produce a cluster map
som_cluster <- sits_som_map(prodes_226_064)
# Evaluate the clusters
cluster_overall <- sits_som_evaluate_cluster(som_cluster)
# Plot confusion between the clusters
plot(cluster_overall)
}
}
\author{
Lorena Santos \email{lorena.santos@inpe.br}
}
