% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_plot.R
\name{plot.som_map}
\alias{plot.som_map}
\title{Generic interface for plotting a SOM map}
\usage{
\method{plot}{som_map}(x, y, ..., type = "codes", whatmap = 1)
}
\arguments{
\item{x}{Object of class "som_map"}

\item{y}{Ignored}

\item{...}{Further specifications for \link{plot}.}

\item{type}{Type of plot: "codes" for neuron weight (time series) and
"mapping" for the number of samples allocated in a neuron.}

\item{whatmap}{What data layer will be plotted.}
}
\description{
plots a SOM map generated by "sits_som_map"
The plot function produces different plots based on the input data:
\itemize{
 \item{"codes": }{Plot the vector weight for in each neuron.}
 \item{"mapping": }{Shows where samples are mapped.}
}
}
\examples{
\donttest{
# Produce a cluster map
som_map <- sits_som_map(prodes_226_064)
# Plot the clusters
plot(som_map, type = "codes")
# Plot kohonen map showing where the samples were allocated
plot(som_map, type = "mapping")
}
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
