% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_plot.R
\name{plot.torch_model}
\alias{plot.torch_model}
\title{Plot Torch (deep learning) model}
\usage{
\method{plot}{torch_model}(x, y, ...)
}
\arguments{
\item{x}{Object of class "torch_model".}

\item{y}{Ignored.}

\item{...}{Further specifications for \link{plot}.}
}
\value{
The plot itself.
}
\description{
Plots a deep learning model developed using torch.
}
\note{
This code has been lifted from the "keras" package.
}
\examples{
\dontrun{
# Get a set of samples
samples_ndvi_evi <- sits_select(samples_modis_4bands,
  bands = c("NDVI", "EVI")
)

# train a deep learning model
dl_model <- sits_train(samples_ndvi_evi, ml_method = sits_mlp())
plot(dl_model)
}

}
\author{
Felipe Souza, \email{lipecaso@gmail.com}

Rolf Simoes, \email{rolf.simoes@inpe.br}

Alber Sanchez, \email{alber.ipia@inpe.br}
}
