% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_plot.R
\name{plot.tsne}
\alias{plot.tsne}
\alias{plot.sits_tsne}
\title{Plot t-SNE results for sits models}
\usage{
\method{plot}{sits_tsne}(x, y, ...)
}
\arguments{
\item{x}{Object of class \code{"tsne"} returned by \code{sits_tsne()}.}

\item{y}{Ignored (for S3 compatibility with \code{plot()} generic).}

\item{...}{Further plotting options (currently ignored).}
}
\value{
A ggplot2 scatter plot.
}
\description{
Plots a t-SNE projection from a sits_tsne object, coloring samples by class labels.
}
