% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_plot.R
\name{plot.variance_vector_cube}
\alias{plot.variance_vector_cube}
\title{Plot variance vector cubes}
\usage{
\method{plot}{variance_vector_cube}(
  x,
  ...,
  tile = x[["tile"]][[1L]],
  roi = NULL,
  palette = "YlGnBu",
  rev = FALSE,
  quantile = 0.75,
  scale = 1,
  max_cog_size = 1024L,
  seg_color = "black",
  line_width = 0.5,
  legend_position = "inside",
  legend_title = "logvar"
)
}
\arguments{
\item{x}{Object of class "variance_vector_cube".}

\item{...}{Further specifications for \link{plot}.}

\item{tile}{Tile to be plotted.}

\item{roi}{Region of interest (see notes below).}

\item{palette}{RColorBrewer or "cols4all" palette}

\item{rev}{Reverse order of colors in palette?}

\item{quantile}{Upper quantile threshold to plot (default = 0.75)}

\item{scale}{Scale to plot map (0.4 to 1.0)}

\item{max_cog_size}{Maximum size of COG overviews (lines or columns)}

\item{seg_color}{Color for segment borders (default = "black")}

\item{line_width}{Line width for segment borders (default = 0.5)}

\item{legend_position}{Where to place the legend (default = "outside")}

\item{legend_title}{Title of legend (default = "logvar")}
}
\value{
A plot containing local variance with segment overlay.
}
\description{
Plots a variance vector cube, which result from
running a variance estimator on a probability vector cube.
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
