% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_TWDTW_plot.R
\name{plotClassification}
\alias{plotClassification}
\title{Plotting subintervals classification}
\usage{
plotClassification(x, timeseries.labels = NULL, patterns.labels = NULL,
  attr, ...)
}
\arguments{
\item{x}{An object of class \code{\link[dtwSat]{twdtwMatches}}.}

\item{timeseries.labels}{the label or index of the time series.
Default is 1.}

\item{patterns.labels}{a vector with labels of the patterns. If not
declared the function will plot one alignment for each pattern.}

\item{attr}{An \link[base]{integer} vector or \link[base]{character} vector
indicating the attribute for plotting. If not declared the function will plot
all attributes.}

\item{...}{additional arguments passed to \code{\link[dtwSat]{twdtwClassify}}.}
}
\value{
A \link[ggplot2]{ggplot} object.
}
\description{
Method for plotting the classification of each
subinterval of the time series based on TWDTW analysis.
}
\author{
Victor Maus, \email{vwmaus1@gmail.com}
}
