% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_tile_api.R
\name{point_accessors}
\alias{point_accessors}
\alias{.lon}
\alias{.lon<-}
\alias{.lat}
\alias{.lat<-}
\title{Point accessors}
\usage{
.lon(x)

.lon(x) <- value

.lat(x)

.lat(x) <- value
}
\arguments{
\item{x}{Object to get/set field value.}

\item{value}{Value to set on object field.}
}
\value{
Getters return respective field value or \code{NULL}, if it doesn't
  exist. Setters return the updated \code{x} object.
}
\description{
These functions are accessors of \code{point} fields of a \code{vector}.
Getters functions returns the respective field values with the expected
data type. Setters functions convert value to expected data type and
store it in respective fields on a given object. If value has no length
and the \code{vector} is not \code{atomic}, it is removed from the object.
}
\details{
\code{.lon()} and \code{.lat()} get/set, respectively, \code{"longitude"}
and \code{"latitude"} fields.
}
\section{Functions}{
\itemize{
\item \code{.lon()}: Get \code{'longitude'} field.

\item \code{.lon(x) <- value}: Set \code{'longitude'} field as numeric.

\item \code{.lat()}: Get \code{'latitude'} field.

\item \code{.lat(x) <- value}: Set \code{'latitude'} field as numeric.

}}
\examples{
\dontrun{
x <- c(longitude = "123")
.lon(x) # 123 as number
x <- list(longitude = 1:10)
.lat(x) <- 11:20
x # with 'longitude' and 'latitude' fields
}

}
\seealso{
Other accessors: 
\code{\link{band_accessors}},
\code{\link{bbox_accessors}},
\code{\link{block_accessors}},
\code{\link{chunk_accessors}},
\code{\link{tile_accessors}}
}
\concept{accessors}
\keyword{internal}
