% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_tile_api.R
\name{point_api}
\alias{point_api}
\alias{.has_point}
\alias{.point}
\alias{.point_as_sf}
\title{Point API}
\usage{
.has_point(x)

.point(x, ..., crs = NULL)

.point_as_sf(point, ..., crs = NULL, as_crs = NULL)
}
\arguments{
\item{x}{Any object to extract a \code{point}.}

\item{...}{Additional parameters.}

\item{crs}{The point CRS. If not informed, default CRS is \code{'EPSG:4326'}.}

\item{point}{A \code{point}.}

\item{as_crs}{A CRS to project \code{point}.}
}
\value{
\code{.has_point()}: \code{logical}.

\code{.point()}: \code{point}.

\code{.point_as_sf()}: \code{sf}.
}
\description{
A point represents a dimensionless geographical location in a given
projection. A \code{point} is any \code{list} or \code{tibble}
containing \code{longitude} and \code{latitude} fields. A \code{point} may
contains multiple entries.
}
\section{Functions}{
\itemize{
\item \code{.has_point()}: Does vector \code{x} has \code{point} fields?

\item \code{.point()}: Extract a \code{point} from any given
\code{vector}.

\item \code{.point_as_sf()}: Convert a \code{point} into a
\code{sf} point object.

}}
\examples{
\dontrun{
x <- list(a = 0, z = 0)
.point(x) # NULL
x <- list(a = 0, longitude = 1:3, b = 2:4, latitude = 2, z = 0)
.point(x)
.point_as_sf(x) # 3 features
.point_as_sf(x, as_crs = "EPSG:3857") # reprojected features
}

}
\seealso{
\link{point_accessors}

Other region objects API: 
\code{\link{.intersects}()},
\code{\link{bbox_api}},
\code{\link{block_api}},
\code{\link{chunks_api}},
\code{\link{roi_api}}
}
\concept{region objects API}
\keyword{internal}
