% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_accuracy.R
\name{print.sits_area_accuracy}
\alias{print.sits_area_accuracy}
\title{Print the area-weighted accuracy}
\usage{
\method{print}{sits_area_accuracy}(x, ..., digits = 2L)
}
\arguments{
\item{x}{An object of class \code{sits_area_accuracy}.}

\item{\dots}{Other parameters passed to the "print" function}

\item{digits}{Significant digits}
}
\value{
Called for side effects.
}
\description{
Adaptation of the caret::print.confusionMatrix method
             for the more common usage in Earth Observation.
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
\keyword{internal}
