% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_accuracy.R
\name{print.sits_area_assessment}
\alias{print.sits_area_assessment}
\title{Print the area assessment}
\usage{
\method{print}{sits_area_assessment}(x, ..., digits = 2)
}
\arguments{
\item{x}{An object of class \code{sits_area_assessment}.}

\item{\dots}{Other parameters passed to the "print" function}

\item{digits}{Significant digits}
}
\value{
No return value, called for side effects.
}
\description{
Adaptation of the caret::print.confusionMatrix method
             for the more common usage in Earth Observation.
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
\keyword{internal}
