% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{prodes_226_064}
\alias{prodes_226_064}
\title{Samples of deforestation-related classes for the LANDSAT image WRS 226/064}
\format{A tibble with 617 rows and 7 variables}
\usage{
data(prodes_226_064)
}
\description{
A dataset containing a tibble with time series samples for the combination
   of the LANDSAT images for WRS 226/04 with the MOD13Q1 collection 5 images, used
   to fill the gaps where there is too much cloud in the LANDSAT data.
   The data set has the following classes (and samples per class): Deforestation_2014 (146 samples),
   Deforestation_2015 (198 samples), Forest (128 samples), and  Pasture (145 samples).
}
\keyword{datasets}
