% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_transitions.R
\name{Recur}
\alias{Recur}
\title{Recur transition function}
\usage{
Recur(target_class)
}
\arguments{
\item{target_class}{Target class name for the transition}
}
\value{
expression that creates a RECUR transition call
}
\description{
Recur transition function
}
\author{
Felipe Carlos, \email{efelipecarlos@gmail.com}

Felipe Carvalho, \email{felipe.carvalho@inpe.br}

Gilberto Camara, \email{gilberto.camara@inpe.br}
}
\keyword{internal}
