% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_tile_api.R
\name{roi_api}
\alias{roi_api}
\alias{.roi_type}
\alias{.roi_switch}
\alias{.roi_as_sf}
\title{ROI API}
\usage{
.roi_type(roi)

.roi_switch(roi, ...)

.roi_as_sf(roi, default_crs = NULL)
}
\arguments{
\item{roi}{A \code{roi}.}

\item{...}{Parameters to be evaluated accordingly to \code{roi} type.}

\item{default_crs}{If no CRS is present in a \code{bbox} object passed
to \code{crs}, which CRS should be used? If \code{NULL} default CRS will
be \code{'EPSG:4326'}.}
}
\value{
\code{.roi_type()}: \code{character}.

\code{.roi_switch()}: one of the arguments in \code{...}.

\code{.roi_as_sf()}: \code{sf}.
}
\description{
A ROI (or 'Region of Interest') represents an geographic area. There are
three types of ROI objects, namely \code{sf} (from package \code{sf}),
\code{bbox} (from \code{.bbox()}), and \code{lonlat}.
A \code{lonlat} object is any \code{vector} containing \code{lon_min},
\code{lon_max}, \code{lat_min}, and \code{lat_max} fields. Its CRS is
defined to be \code{'EPSG:4326'}.
}
\section{Functions}{
\itemize{
\item \code{.roi_type()}: Tells which type of ROI is in \code{roi}
parameter (One of \code{'sf'}, \code{'bbox'}, or \code{'lonlat'}).

\item \code{.roi_switch()}: Chooses one of the arguments passed in
\code{...} according to which type of \code{roi} parameter.

\item \code{.roi_as_sf()}: Converts \code{roi} to an \code{sf} object.

}}
\examples{
x <- list(lon_min = 1, lon_max = 2, lat_min = 3, lat_max = 4)
.roi_type(x) # lonlat
.roi_as_sf(x)
x <- list(xmin = 1, xmax = 2, ymin = 3, ymax = 4, crs = 3857)
.roi_type(x) # bbox
.roi_as_sf(x)
.roi_switch(
  x,
  lonlat = "It's in WGS 84",
  bbox = paste("It's in", .crs(x))
)

}
\seealso{
\code{\link{.bbox}()}
}
