% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_roi.R
\name{roi_functions}
\alias{roi_functions}
\alias{.sits_roi_bbox}
\title{Region of interest (roi)}
\usage{
.sits_roi_bbox(roi, cube)
}
\arguments{
\item{roi}{a vector containing either `("xmin", "xmax", "ymin", "ymax")`
or `("lon_min", "lat_min", "lon_max", "lat_max")` variables, representing
a spatial region of interest}

\item{cube}{a sits data cube object (inherited from `tbl`)}
}
\value{
a `bbox` object (inherited from `tbl`)
}
\description{
Region of interest functions

`.sits_roi_bbox()`: convert a given `roi` to a bbox with same
crs as `cube` (that should have only one row). If no `roi` is given,
returns the `cube`'s `bbox`.
}
\keyword{internal}
