% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{samples_MT_9classes}
\alias{samples_MT_9classes}
\title{Samples of nine classes for the state of Mato Grosso used for classification}
\format{A tibble with 2115 rows and 7 variables: (a) longitude: East-west coordinate of the time series sample (WGS 84);
  latitude (North-south coordinate of the time series sample in WGS 84), start_date (initial date of the time series),
  end_date (final date of the time series), label (the class label associated to the sample),
  coverage (the name of the coverage associated with the data),
  time_series (list containing a tibble with the values of the time series).}
\usage{
data(samples_MT_9classes)
}
\description{
A dataset containing a tibble with time series samples for the Mato Grosso state in Brasil.
  The time series come from MOD13Q1 collection 6 images. The data set has the following classes:
  Cerrado(400 samples), Fallow_Cotton (34 samples), Forest (138 samples), Pasture (370 samples),
  Soy-Corn (398 samples),  Soy-Cotton (399 samples), Soy_Fallow (88 samples),
  Soy_Millet (235 samples), and Soy_Sunflower (53 samples).
}
\keyword{datasets}
