% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_WTSPS.R
\docType{methods}
\name{setServerURL}
\alias{setServerURL}
\alias{setServerURL,WTSPS-method}
\title{Set WTSPS object's server URL}
\usage{
setServerURL(object, serverURL)

\S4method{setServerURL}{WTSPS}(object, serverURL)
}
\arguments{
\item{object}{A WTSPS object}

\item{serverURL}{URL character}
}
\description{
Set WTSPS object's server URL
}
