% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_assessment.R
\name{sits_accuracy_area}
\alias{sits_accuracy_area}
\title{Area-weighted post-classification accuracy assessment of classified maps}
\usage{
sits_accuracy_area(results.tb, area, conf.int = 0.95, rm.nosample = FALSE)
}
\arguments{
\item{results.tb}{a sits table with a set of lat/long/time locations  with known and trusted labels and
with the result of classification method}

\item{area}{a list with the area of each label}

\item{conf.int}{specifies the confidence level (0-1).}

\item{rm.nosample}{if sum of columns and sum of rows of the error matrix are zero
then remove class. Default is TRUE.}
}
\description{
To use this function, the input table should be a set of results containing
both the label assigned by the user and the classification result.
Accuracy assessment set us a confusion matrix to determine the accuracy of your classified result.
This function uses an area-weighted technique proposed by Olofsson et al. to
produce accuracy estimates that are more reliable

We plan to do an improved version of this function that includes a Raster R object
with the classified map and a vector with the labels of the classified map
(Gilberto-Rolf-05-Jun-2017)

This function calls \code{\link[dtwSat]{twdtwAssess}} from \pkg{dtwSat}.
\code{\link[dtwSat]{twdtwAssess}} performs an accuracy assessment of the classified, including
Overall Accuracy, User's Accuracy, Produce's Accuracy, error matrix (confusion matrix),
and estimated area according to [1-2].
}
\references{
[1] Olofsson, P., Foody, G.M., Stehman, S.V., Woodcock, C.E. (2013).
Making better use of accuracy data in land change studies: Estimating
accuracy and area and quantifying uncertainty using stratified estimation.
Remote Sensing of Environment, 129, pp.122-131.

[2] Olofsson, P., Foody G.M., Herold M., Stehman, S.V., Woodcock, C.E., Wulder, M.A. (2014)
Good practices for estimating area and assessing accuracy of land change. Remote Sensing of
Environment, 148, pp. 42-57.
}
\author{
Victor Maus, \email{vwmaus1@gmail.com}

Gilberto Camara, \email{gilberto.camara@inpe.br}
}
