% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_assessment.R
\name{sits_accuracy_csv}
\alias{sits_accuracy_csv}
\title{Saves the results of accuracy assessment as CSV files}
\usage{
sits_accuracy_csv(conf.mx, file = NULL)
}
\arguments{
\item{conf.mx}{A caret S4 object with a confusion matrix}

\item{file}{The file where the CSV data is to be saved}
}
\value{
conf.mx       The input confusion matrix
}
\description{
Saves the the accuracy of classification in CSV
Returns a confusion matrix used by the "caret" package
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
