% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_accuracy.R
\name{sits_accuracy_summary}
\alias{sits_accuracy_summary}
\title{Print accuracy summary}
\usage{
sits_accuracy_summary(x, digits = NULL)
}
\arguments{
\item{x}{Object of class \code{sits_accuracy}.}

\item{digits}{Number of significant digits when printed.}
}
\value{
Called for side effects.
}
\description{
Adaptation of the caret::print.confusionMatrix method
             for the more common usage in Earth Observation.
}
\seealso{
Other validation and accuracy:
\code{\link[=sits_accuracy]{sits_accuracy()}},
\code{\link[=sits_kfold_validate]{sits_kfold_validate()}},
\code{\link[=sits_validate]{sits_validate()}}
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
\concept{validation and accuracy}
\keyword{internal}
