% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_active_learning.R
\name{sits_active_learning}
\alias{sits_active_learning}
\title{Use the given samples to automatically collect new samples.}
\usage{
sits_active_learning(
  samples_tb,
  sits_method,
  data_cube,
  n_samples = 100,
  min_probability = 0.95,
  min_entropy = 0.5
)
}
\arguments{
\item{samples_tb}{A sits tibble.}

\item{sits_method}{A sits model specification.}

\item{data_cube}{A sits data cube.}

\item{n_samples}{The number of random points to take.}

\item{min_probability}{The minimum probability for automatically including new samples.}

\item{min_entropy}{The minimum entropy for consider a sample for the oracle.}
}
\value{
A list with two sits tibbles: One for new samples and another for a human expert to review.
}
\description{
Active Learning improves the results of a classification by feeding the
classifier with informative samples. This function returns two sits tibbles:
The first (new_samples) contains new samples with certainty above
min_probability. The second (oracle_samples) contains samples of
low confidence (their entropy is above min_entropy). The samples in both
tibbles are selected randomly inside the extent of data_cube.

The new_samples can be merged with samples_tb to increase the number of
training samples while the oracle_samples, after human inspection, would
provide the classifier with extra information where it has low confidence.
}
\examples{
\donttest{
# Prepare a classification.
samples_tb <- sits_select(samples_modis_4bands, bands = c("NDVI", "EVI"))
sits_method <- sits_xgboost(verbose = FALSE)
data_cube <- sits_cube(source = "LOCAL",
                       name = "sinop-2014",
                       satellite = "TERRA",
                       sensor = "MODIS",
                       data_dir = system.file("extdata/raster/mod13q1",
                                              package = "sits"),
                       delim = "_",
                       parse_info = c("X1", "X2", "tile", "band", "date"))

# Run sits' active learning
al <- sits_active_learning(samples_tb, sits_method, data_cube)

# Use new_samples to increase the number available samples
rbind(samples_tb, al$new_samples[colnames(samples_tb)])

# The oracle_samples are meant to be reviewed by humans.
al$oracle_samples
}

}
\author{
Alber Sanchez, \email{alber.ipia@inpe.br}
}
