% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_utils.R
\name{sits_active_tests}
\alias{sits_active_tests}
\title{Informs if sits tests should run}
\usage{
sits_active_tests()
}
\value{
A logical value
}
\description{
This function informs if sits test should run.
This is useful to avoid running slow tests in CRAN environment.
}
