% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_tibble_functions.R
\name{sits_add_row}
\alias{sits_add_row}
\title{Add a new row to a SITS tibble}
\usage{
sits_add_row(data.tb = NULL, longitude = 0, latitude = 0,
  start_date = "1970-01-01", end_date = "1970-01-01", label = "NoClass",
  coverage = "image", time_series = list())
}
\arguments{
\item{data.tb}{A SITS tibble}

\item{longitude}{Longitude of the chosen location}

\item{latitude}{Latitude of the chosen location}

\item{start_date}{Start of the period}

\item{end_date}{End of the period}

\item{label}{Label to attach to the time series (optional)}

\item{coverage}{The name of the coverage}

\item{time_series}{List containing a time series (Index and band values)}
}
\value{
data.tb        An updated SITS tibble
}
\description{
Adds a row to a tibble, with suitable defaults
}
\examples{
# Reads a data set with a ZOO time series
data (ts_zoo)
# convert the data from the zoo format to the SITS format
ts.tb <- tibble::as_tibble (zoo::fortify.zoo (ts_zoo))
# create a list to store the zoo time series
ts.lst <- list()
# Put the tibble in the list
ts.lst[[1]] <- ts.tb
# get the start date
start_date <- ts.tb[1,]$Index
# get the end date
end_date <- ts.tb[NROW(ts.tb),]$Index
# create an empty sits tibble
data.tb <- sits_tibble()
# put the time series data and metadata into the tibble
data.tb <- sits_add_row (data.tb, longitude = -54.2313, latitude = -14.0482,
           start_date = start_date, end_date = end_date,
           label = "Cerrado", coverage = "mod13q1", time_series = ts.lst)
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
