% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_distances.R
\name{sits_adjust}
\alias{sits_adjust}
\title{Adjust distances used as training sets for machine learning}
\usage{
sits_adjust()
}
\value{
adj_fun       Adjustment function to be applied to the data
}
\description{
This function provides the default adjustment of the values of the time series
used to train ML methods. Since most ML methods do not allow for negative data, the default
adjustment is to shift the data by a fixed value. The adjustment value is taken from
the configuration file (config.yml). See \code{\link[sits]{sits_config}} for more details
on how to change the configuration file.
This function should always be used in combination with \code{\link[sits]{sits_distances}}.
}
\examples{
# get one time series
data(cerrado_2classes)
# compute the distance vector for the first time series
distances.tb <- sits_distances(cerrado_2classes[1,], adj_fun = sits_adjust())

}
\author{
Rolf Simoes, \email{rolf.simoes@inpe.br}

Gilberto Camara, \email{gilberto.camara@inpe.br}
}
