% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_WTSPS.R
\name{sits_algorithmWTSPS}
\alias{sits_algorithmWTSPS}
\title{Provides information about one algorithm of the WTSPS service}
\usage{
sits_algorithmWTSPS(URL = "http://www.dpi.inpe.br/tws/wtsps",
  algorithm = NULL)
}
\arguments{
\item{URL}{the URL for the WTSPS service}

\item{algorithm}{the name of the algorithm}
}
\description{
uses the WTSPS services to print information about a
chosen algorithm:
 name
 input_parametes
 output
 description
}
