% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_table.R
\name{sits_align}
\alias{sits_align}
\title{Aligns dates of time series to a reference date}
\usage{
sits_align(data, ref_date)
}
\arguments{
\item{ref_date}{date   - a reference date where all series will start}

\item{data.tb}{tibble - input SITS table (useful for chaining functions)}
}
\value{
data1.tb   tibble - the converted SITS table (useful for chaining functions)
}
\description{
\code{sits_align} convert the index of a sits time to a reference year
The reference year is taken from the date of the start of the time series
available in the coverage.
}
\examples{
data1.tb <- sits_align (data = "mydata.tb", ref_date = "2000-09-02")
}
\seealso{
Other STIS main functions: \code{\link{sits_centroids}},
  \code{\link{sits_dendogram}},
  \code{\link{sits_plot_together}},
  \code{\link{sits_plot}}, \code{\link{sits_save}}
}
\keyword{STIS}
