% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_tibble_functions.R
\name{sits_apply_on}
\alias{sits_apply_on}
\title{Apply a function on each SITS tibble column element}
\usage{
sits_apply_on(data.tb, field, fun)
}
\arguments{
\item{data.tb}{a SITS tibble}

\item{field}{a valid field of data.tb to apply `fun` function}

\item{fun}{a function to apply in field data. The function must have an input parameter to receive each
field data element at a time and output a new value.}
}
\value{
result.tb  a tibble where columns have the reference label and the time series bands as distances
}
\description{
Computes new values for a given SITS tibble field and returns a tibble whose field column have
the new values.
}
\author{
Rolf Simoes, \email{rolf.simoes@inpe.br}
}
