% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_table.R
\name{sits_assert_table}
\alias{sits_assert_table}
\title{Verify if the sits table has been created}
\usage{
sits_assert_table(table_obj)
}
\arguments{
\item{table_name}{a string - name of table (must exist)}
}
\description{
\code{sits_assert_table} tests if the sits table exists
}
