% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_WTSS.R
\name{sits_assert_WTSS}
\alias{sits_assert_WTSS}
\title{Verify if the information concerning WTSS service has been configured}
\usage{
sits_assert_WTSS()
}
\description{
\code{sits_assert_WTSS} tests if the WTSS service has been correctly configured
}
