% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_assertWTSS.R
\name{sits_assertWTSS}
\alias{sits_assertWTSS}
\title{Verify if the information concerning WTSS service has been configured}
\usage{
sits_assertWTSS()
}
\description{
\code{sits_assert_WTSS} tests if the WTSS service has been correctly configured
 The function verifies that the WTSS service is running
}

