% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_getdata.R
\name{sits_band_info}
\alias{sits_band_info}
\title{Obtain the band information from the configuration file}
\usage{
sits_band_info(bands, product)
}
\arguments{
\item{bands}{a list of bands whose information is to be retrieved}

\item{product}{the image product whose information we need to retrieve}
}
\value{
band_info a tibble with information about the bands (missing values and scale factors)
}
\description{
Reads information about missing values and scale factor for image products
}
\examples{
band_info <- sits_band_info(bands = c("ndvi", "evi"), product = "MOD13Q1")
}
