% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_band_names.R
\name{sits_band_names}
\alias{sits_band_names}
\title{Finds the names of the bands of a time series}
\usage{
sits_band_names(ts)
}
\arguments{
\item{ts}{a time series extracted from a sits table}
}
\value{
names a string vector with the names of the bands
}
\description{
\code{sits_band_names} tests if the sits table exists and has data
}

