% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_data_generics.R
\name{sits_bands.cube}
\alias{sits_bands.cube}
\title{Informs the names of the bands of a cube}
\usage{
\method{sits_bands}{cube}(data)
}
\arguments{
\item{data}{Valid sits tibble (time series)}
}
\value{
A string vector with the names of the bands.
}
\description{
Finds the names of the bands of a cube
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}

Rolf Simoes, \email{rolf.simoes@inpe.br}
}
