% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_data_generics.R
\name{sits_bands.patterns}
\alias{sits_bands.patterns}
\title{Informs the names of the bands of a set of timeseries}
\usage{
\method{sits_bands}{patterns}(data)
}
\arguments{
\item{data}{Valid sits tibble (time series)}
}
\value{
A string vector with the names of the bands.
}
\description{
Finds the names of the bands of time series in a sits tibble
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}

Rolf Simoes, \email{rolf.simoes@inpe.br}
}
