% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_config.R
\name{sits_bands_service}
\alias{sits_bands_service}
\alias{.sits_bands_service}
\title{Retrieve the bands associated to a service}
\usage{
.sits_bands_service(service, name)
}
\arguments{
\item{service}{Name of a service.}

\item{name}{Name of a cube}
}
\description{
Retrieve the cubes associated a service.
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
