% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_bayes_postprocess.R
\name{sits_bayes_postprocess}
\alias{sits_bayes_postprocess}
\title{Read a block of values retrived from a set of raster bricks}
\usage{
sits_bayes_postprocess(raster_class, noise, file)
}
\arguments{
\item{raster_class}{output raster coverage}

\item{noise}{bayesian smoothing parameter}

\item{file}{file to save the post processed raster}
}
\value{
raster_layers.tb  tibble with metadata about the output RasterLayer objects
}
\description{
Takes a set of classified raster layers, whose metadata is
             described by tibble (created by \code{\link[sits]{sits_coverage}}),
             and a noise value, to do a bayesian smoothing process.
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
