% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_data_generics.R
\name{sits_bbox.cube}
\alias{sits_bbox.cube}
\title{Get the bounding box of a data cube}
\usage{
\method{sits_bbox}{cube}(data)
}
\arguments{
\item{data}{Valid data cube}
}
\value{
named vector with bounding box ("xmin", "xmax", "ymin", "ymax")
}
\description{
Obtain a vector of limits for the cube
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
