% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_data_generics.R
\name{sits_bbox.sits}
\alias{sits_bbox.sits}
\title{Get the bounding box of a set of time series}
\usage{
\method{sits_bbox}{sits}(data)
}
\arguments{
\item{data}{Valid sits tibble with a set of time series}
}
\value{
named vector with bounding box ("lon_min", "lon_max",
     "lat_min", "lat_max")
}
\description{
Obtain a vector of limits in lat/long for time series
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
