% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_tibble_functions.R
\name{sits_bind}
\alias{sits_bind}
\title{Bind two SITS tibbles}
\usage{
sits_bind(data1.tb, data2.tb)
}
\arguments{
\item{data1.tb}{the first SITS table to be bound}

\item{data2.tb}{the second SITS table to be bound}
}
\value{
result.tb    a merged SITS tibble with a bind set of time series
}
\description{
This function merges two STIS tables.
To merge two series, we consider that they contain different spatio-temporal location
but refer to the same coverage, and have the same attributes.
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
