% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_tibble_functions.R
\name{sits_break}
\alias{sits_break}
\title{Break a data set in segments to match the dates of a set of samples, given an interval}
\usage{
sits_break(data.tb, samples.tb, interval = "12 month")
}
\arguments{
\item{data.tb}{a valid sits tibble}

\item{samples.tb}{the samples used in the classification}

\item{interval}{the interval between the classificationa}
}
\description{
This function aligns an input data set to the dates used by a set of samples
}
\examples{
# Retrieve the set of samples for the Mato Grosso region (provided by EMBRAPA)
data(samples_MT_9classes)
# get a point
data (ts_2000_2016)
# break the point to match the samples (breaks a long time series into intervals)
point2.tb <- sits_break(ts_2000_2016, samples_MT_9classes)
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
