% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_tibble_utils.R
\name{sits_break_ts}
\alias{sits_break_ts}
\title{Break a time series to match the time range of a set of patterns}
\usage{
sits_break_ts(data.tb, patterns.tb)
}
\arguments{
\item{data.tb}{a valid sits tibble with data}

\item{patterns.tb}{a tibble with time series patterns}
}
\value{
newdata.tb  a SITS tibble with the original data cut into subsets of equal lengths
}
\description{
Given a long time series, divide it into segments to match
              the time range of a set of patterns
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}

Rolf Simoes, \email{rolf.simoes@inpe.br}
}
