% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_check.R
\name{sits_check}
\alias{sits_check}
\title{Checks if the data is a valid SITS tibble}
\usage{
sits_check(sits.tb)
}
\arguments{
\item{sits.tb}{a table (tibble) in SITS formal}
}
\value{
TRUE/FALSE
}
\description{
\code{sits_check} tests the names of the columns of a tibble to
verifiy if they include the mandatory columns
<longitude>, <latitude>, <start_date>, <end_date>, <label>, <coverage>
}
