% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_data_generics.R
\name{sits_check_data.sits}
\alias{sits_check_data.sits}
\title{Checks if a set of time series is consistent}
\usage{
\method{sits_check_data}{sits}(data)
}
\arguments{
\item{data}{Valid sits tibble (time series or a cube)}
}
\value{
Messages with warnings
}
\description{
Check is timelines of data sets are consistent
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
