% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_table.R
\name{sits_check_table}
\alias{sits_check_table}
\title{Check if a table exists in the global environment}
\usage{
sits_check_table(table_obj)
}
\arguments{
\item{table_name}{a string - name of table (created if not existent)}
}
\description{
\code{sits_check_table} tests if the sits table exists and creates a new one if the response is negative
The table is created in the global environment
}
