% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_classification.R
\name{sits_class_info}
\alias{sits_class_info}
\title{Define the information required for classifying time series}
\usage{
sits_class_info(bands, labels, timeline, interval, start_date, end_date)
}
\arguments{
\item{bands}{The bands to be used for classification}

\item{labels}{A vector with the labels of the classes to be classified}

\item{timeline}{The timeline of the coverage being classified}

\item{interval}{The interval between two sucessive classification}

\item{start_date}{The start date of the reference period for classification (inside an interval)}

\item{end_date}{The end date of the reference period for classification (inside an interval)}
}
\value{
class_info.tb   A SITS tibble with the classification information
}
\description{
Time series classification requires that users do a series of steps:
(a) Provide labelled samples that will be used as training data.
(b) Provide information on how the classification will be performed, including data timeline,
temporal interval, and start and end dates per interval.
(c) Clean the training data to ensure it meets the specifications of the classification info.
(d) Use the clean data to train a machine learning classifier.
(e) Classify non-labelled data sets

In this set of steps, this function provides support for step (b). It requires the user
to provide a timeline, the classification interval, and the start and end dates of
the reference period. The results is a tibble with information that allows the user
to perform steps (c) to (e)
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
