% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_classification.R
\name{sits_classify.stack_cube}
\alias{sits_classify.stack_cube}
\title{Classify a stack data cube using multicore machines}
\usage{
\method{sits_classify}{stack_cube}(
  data,
  ml_model,
  ...,
  sf_region = NULL,
  filter = NULL,
  memsize = 8,
  multicores = 2,
  output_dir = "./",
  version = "v1"
)
}
\arguments{
\item{data}{data cube}

\item{ml_model}{R model trained by \code{\link[sits]{sits_train}}.}

\item{...}{other parameters to be passed to specific functions}

\item{sf_region}{an sf object with the region of interest}

\item{filter}{smoothing filter to be applied (if desired).}

\item{memsize}{memory available for classification (in GB).}

\item{multicores}{number of cores to be used for classification.}

\item{output_dir}{directory for output file}

\item{version}{version of the output (for multiple classifications)}
}
\value{
cube with the metadata of a brick of probabilities.
}
\description{
Takes a data cube, whose metadata is
   described by tibble (created by \code{\link[sits]{sits_cube}}),
   a set of samples used for training a classification model,
   a prediction model (created by \code{\link[sits]{sits_train}}),
   and produces a classified set of RasterLayers. These
   parameters are "memsize" and "multicores".
   The "multicores" parameter defines the
   number of cores used for processing. The "memsize" parameter  controls
   the amount of memory available for classification.
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
