% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_classification.R
\name{sits_classify2}
\alias{sits_classify2}
\title{Classify a sits tibble using machine learning models}
\usage{
sits_classify2(data.tb = NULL, ml_model = NULL,
  dist_method = sits_distances_from_data2(start_from = "2000-09-01", interval
  = "12 month"))
}
\arguments{
\item{data.tb}{a SITS tibble time series}

\item{ml_model}{a model trained by \code{\link[sits]{sits_train}}}

\item{dist_method}{method to compute distances (e.g., sits_TWDTW_distances)}

\item{interval}{the period between two classifications}

\item{...}{other parameters to be passed to the distance function}
}
\value{
data.tb      a SITS tibble with the predicted labels for each input segment
}
\description{
Returns a sits table with the results of the ML classifier.
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
