% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_classification.R
\name{sits_classify_distances}
\alias{sits_classify_distances}
\title{Classify a distances matrix using machine learning models}
\usage{
sits_classify_distances(data.mx, timeline, time_index.lst, nsamples, attr_names,
  bands, ml_model = NULL, multicores = 1)
}
\arguments{
\item{data.mx}{a matrix with data values}

\item{timeline}{the timeline of the data set}

\item{time_index.lst}{The subsets of the timeline}

\item{nsamples}{number of samples}

\item{attr_names}{Names of columns of distance tibble}

\item{bands}{Bands used for classification}

\item{ml_model}{a model trained by \code{\link[sits]{sits_train}}}

\item{multicores}{Number of threads to process the time series.}
}
\value{
pred.lst        list with the predicted labels for each output time interval
}
\description{
Returns a sits table with the results of the ML classifier.
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
