% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_classification.R
\name{sits_classify_model}
\alias{sits_classify_model}
\title{Classify a sits tibble using pre-built machine learning models}
\usage{
sits_classify_model(data.tb = NULL, train_samples.tb = NULL, model = NULL,
  adj_fun = sits_adjust(), interval = "12 month", multicores = 1)
}
\arguments{
\item{data.tb}{SITS tibble time series (cleaned)}

\item{train_samples.tb}{The samples used for training the classification model}

\item{model}{A pre-built ML model (see \code{\link[sits]{sits_train}})}

\item{adj_fun}{Adjustment function to be applied to the data}

\item{interval}{The interval used for classification}

\item{multicores}{Number of threads to process the time series.}
}
\value{
data.tb           SITS tibble with the predicted labels for each input segment
}
\description{
This function classifies a set of time series, given
a set of training samples, an inference model, and an interval.
To perform the classification, users should provide a set of
labelled samples. Each samples should be associated to one spatial location
(latitude/longitude), one time interval and a label.
}
\examples{
\donttest{
# read a training data set
# Retrieve the set of samples for the Mato Grosso region (provided by EMBRAPA)
data(samples_MT_ndvi)
# select the bands "ndvi", "evi", "nir", and "mir"
samples.tb <- sits_select(samples_MT_9classes, bands = c("ndvi","evi","nir","mir"))
# find a training model based on the distances and on default values
model <- sits_train(samples.tb)
# Retrieve a time series
data("ts_2000_2016")
# select the bands "ndvi", "evi", "nir", and "mir"
point.tb <- sits_select(ts_2000_2016, bands = c("ndvi","evi","nir","mir"))
# classify the point
class.tb <-  sits_classify_model(point.tb, samples.tb, model)
# plot the classification
sits_plot(class.tb)
}

}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
