% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_raster.R
\name{sits_classify_raster}
\alias{sits_classify_raster}
\title{Classify a set of spatio-temporal raster bricks using machine learning models}
\usage{
sits_classify_raster(file = NULL, raster.tb, samples.tb,
  ml_method = sits_svm(), adj_fun = sits_adjust(), interval = "12 month",
  blocksize = 2000, multicores = 2, verbose = FALSE)
}
\arguments{
\item{file}{vector of file names to store the output (one file per classified year)}

\item{raster.tb}{tibble with information about a set of space-time raster bricks}

\item{samples.tb}{tibble with samples used for training the classification model}

\item{ml_method}{an R model trained by \code{\link[sits]{sits_train}}}

\item{adj_fun}{adjustment function to be applied to the data}

\item{interval}{interval between two sucessive classifications, expressed in months}

\item{blocksize}{size of the block to be read to build a block for classification}

\item{multicores}{number of threads to process the time series.}

\item{verbose}{logical: run function in verbose mode? (useful for working with big data sets)}
}
\value{
raster_class.tb tibble with the metadata for the vector of classified RasterLayers
}
\description{
Takes a set of spatio-temporal raster bricks, whose metadata is
             described by tibble (created by \code{\link[sits]{sits_coverage}}),
             a set of samples used for training a classification model,
             a prediction model (created by \code{\link[sits]{sits_train}}),
             and produces a classified set of RasterLayers. This function is similar to
              \code{\link[sits]{sits_classify}} which is applied to time series stored in a SITS tibble.
              There are two parameters for optimizing processing of large data sets. These
              parameters are "blocksize" and "multicores". The first controls the size of the data block to
              be read from disk. When acessing raster data, the block size is a multiple of the number of
              columns in each band. Thus, reading a 1000 x 1000 image with a block size of 250000 will result
              in using blocks of size 250 lines x 1000 columns. The "multicores" parameter defines the
              number of cores used for processing.
}
\examples{
\donttest{
# Retrieve the set of samples for the Mato Grosso region (provided by EMBRAPA)
data(samples_MT_ndvi)

# read a raster file and put it into a vector
files  <- c(system.file ("extdata/raster/mod13q1/sinop-crop-ndvi.tif", package = "sits"))

# define the timeline
data(timeline_mod13q1)
timeline <- lubridate::as_date (timeline_mod13q1$V1)

# create a raster metadata file based on the information about the files
raster.tb <- sits_coverageRaster(product = "MOD13Q1", coverage = "Sinop-Crop",
             timeline = timeline, bands = c("ndvi"), files = files)

# classify the raster file
raster_class.tb <- sits_classify_raster (file = "./raster-class", raster.tb, samples_MT_ndvi,
   ml_method = sits_svm(), blocksize = 250, multicores = 1)
}

}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
