% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_classification.R
\name{sits_classify_ts}
\alias{sits_classify_ts}
\title{Classify a sits tibble using machine learning models}
\usage{
sits_classify_ts(ts.lst, patterns.tb = NULL, ml_model = NULL,
  interval = "12 month")
}
\arguments{
\item{ts.lst}{a list of time series}

\item{patterns.tb}{a set of known temporal signatures for the chosen classes}

\item{ml_model}{a model trained by \code{\link[sits]{sits_train}}}

\item{interval}{the period between two classifications}
}
\value{
class.tb        a SITS tibble with the predicted labels for each input segment
}
\description{
Returns a sits table with the results of the ML classifier.
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
