% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_kohonen.R
\name{sits_clean_samples_som}
\alias{sits_clean_samples_som}
\title{Clean samples}
\usage{
sits_clean_samples_som(som_cluster, prob_label_change = 0.8,
  min_cluster_prob = 0.6)
}
\arguments{
\item{som_cluster}{A sits tibble returned by \code{\link[sits]{sits_som_map}}}

\item{prob_label_change}{Threshold of probability to change the label}

\item{min_cluster_prob}{Minimum probability for a good sample to belong to a cluster}
}
\value{
Returns a sits tibble with a new subset of samples and a new column
presenting the probability of each sample belongs to a class described in column label.
}
\description{
This function remove sample that do not have good quality based
on the statistics of the result of the clustering using SOM.
}
\examples{
\donttest{
# Read a set of samples
# Get a new subset of samples evaluated by clustering methods
som_cluster <- sits_som_map(prodes_226_064, grid_xdim = 10, grid_ydim = 10, rlen = 100,
       distance = "euclidean", iterations = 50)
new_samples.tb <- sits_clean_samples_som(som_cluster)
}
}
\author{
Lorena Santos, \email{lorena.santos@inpe.br}

Karine Ferreira. \email{karine.ferreira@inpe.br}
}
