% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_clouds.R
\name{sits_cloud_cbers}
\alias{sits_cloud_cbers}
\title{Clean data cube to improve quality}
\usage{
sits_cloud_cbers(
  cube,
  cld_band_name = "CMASK",
  data_dir = NULL,
  t1 = 1,
  t2 = 0.11,
  t3 = 0.5,
  t4 = 0.75,
  t5 = 40,
  t6 = 5,
  memsize = 8,
  multicores = 2
)
}
\arguments{
\item{cube}{input data cube}

\item{cld_band_name}{indication of the cloud band to be produced}

\item{data_dir}{directory where cloud band will be written}

\item{t1}{controls the difference btw visible and infrared bands}

\item{t2}{controls the brightness properties of cloud}

\item{t3}{controls the dark property of cloud shadows.}

\item{t4}{remove influence of water in the cloud shadow detection}

\item{t5}{size of window to search for clouds near shadows}

\item{t6}{size of window of median filter to remove outliers}

\item{memsize}{size of memory}

\item{multicores}{number of cores}
}
\value{
new data cube with cloud data
}
\description{
This algorithm is an implementation of the paper by Zhai et al,
"Cloud/shadow detection based on spectral indices for multispectral
 optical remote sensing imagery", ISPRS Journal of Photogrammetry
 and Remote Sensing, volume 144, October 2018, Pages 235-253.

 The authors recommend the following typical values for the parameters:
 SITS supports the following models:
\itemize{
 \item{t1: } - (1,...,10) - default = 1
 \item{t2: } - (0.1,...,0.5) - default = 0.11
 \item{t3: } - (0.25,...,0.75) - default = 0.50
 \item{t4: } - (0.5,...,0.90) - default = 0.75
 \item{t5: } - (30,...,90) - default = 40
 \item{t6: } - (3,...,11) - default = 5
 }

 Please see Zhai et al.'s paper for more detail.
}
