% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_filters.R
\name{sits_cloud_removal}
\alias{sits_cloud_removal}
\title{cloud removal}
\usage{
sits_cloud_removal(
  data = NULL,
  cutoff = 0.25,
  bands_suffix = "",
  apply_whit = TRUE,
  lambda_whit = 1
)
}
\arguments{
\item{data}{Time series data and metadata - only the "ndvi" band.}

\item{cutoff}{Maximum acceptable value of a NDVI difference.}

\item{bands_suffix}{Suffix to rename the filtered bands.}

\item{apply_whit}{Apply the whittaker smoother after filtering?}

\item{lambda_whit}{Lambda parameter of the whittaker smoother.}
}
\value{
A sits tibble with same samples and the new bands.
}
\description{
This function tries to remove clouds in the ndvi band of
a satellite image time series. It looks for points where the NDVI value
band goes down abruptly. These points are taken as those whose difference
is more than a cutoff value set by the user.
Then it applies an spline interpolation.
Finally, the function applies a whitakker smoother.
}
\examples{
\donttest{
# Read a set of samples of forest/non-forest in Amazonia
# This is an area full of clouds
data(prodes_226_064)
# Select the NDVI band of the first point
point_ndvi.tb <- sits_select_bands(prodes_226_064[1,], ndvi)
# Apply the cloud filter
point_cld.tb <- sits_filter(point_ndvi.tb,
                sits_cloud_removal(band_suffix = "cf"))
# Merge the filtered with the raw data
point2.tb <- sits_merge (point_ndvi.tb, point_cld.tb)
# Plot the result
plot (point2.tb)
}
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
