% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_clouds.R
\name{sits_cloud_remove}
\alias{sits_cloud_remove}
\title{Clean data cube to improve quality}
\usage{
sits_cloud_remove(cube, ...)
}
\arguments{
\item{cube}{input data cube}

\item{...}{other parameters to be passed for specific types}
}
\description{
Clean data cube to improve quality
}
