% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_clouds.R
\name{sits_cloud_remove.s2_l2a_aws}
\alias{sits_cloud_remove.s2_l2a_aws}
\title{Clean data cube to improve quality for S2 L2A AWS}
\usage{
\method{sits_cloud_remove}{s2_l2a_aws}(
  cube,
  data_dir,
  bands = NULL,
  interval = "16 day",
  memsize = 8,
  multicores = 2,
  ...
)
}
\arguments{
\item{cube}{input data cube}

\item{data_dir}{data directory where output data is written}

\item{bands}{bands to be processed}

\item{interval}{time between the output images}

\item{memsize}{size of memory}

\item{multicores}{number of cores}

\item{...}{other parameters to be passed for specific types}
}
\description{
Clean data cube to improve quality for S2 L2A AWS
}
