% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_cluster.R
\name{sits_cluster_cleaner}
\alias{sits_cluster_cleaner}
\title{Cluster cleaner}
\usage{
sits_cluster_cleaner(data.tb, min_perc = 0.05, method = "intracluster")
}
\arguments{
\item{data.tb}{a SITS tibble with `cluster` column.}

\item{min_perc}{The minimum percentage of label inside a cluster for the label to remain in cluster.}

\item{method}{Either "intracluster" or "intercluster"}
}
\value{
result.tb        a SITS tibble with all selected samples
}
\description{
Removes SITS tibble samples of labels that are minority in each cluster.
The function sets a percentage threshold for removal of samples of a label
from a cluster to be excluded from that cluster
If the method "intracluster" is chosen, the "min_perc" parameter
controls the relative percentage of labels inside each cluster. If samples
of a given label inside one cluster are below this limit, then the samples of this label are
removed from the clustar. If the method intercluster" is set, the "min_perc"
parameter is used a threshold that controls the minimum
percentage of each label in all clusters. If the percentage of samples of a label is a cluster is less
than this value, all samples from this label are removed from the cluster.
}
\author{
Rolf Simoes, \email{rolf.simoes@inpe.br}
}
