% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_cluster.R
\name{sits_cluster_frequency}
\alias{sits_cluster_frequency}
\title{Show label frequency in each cluster produced by dendrogram analysis}
\usage{
sits_cluster_frequency(samples)
}
\arguments{
\item{samples}{Tibble with input set of time series with additional
cluster information produced
by \code{\link[=sits_cluster_dendro]{sits_cluster_dendro()}}.}
}
\value{
A matrix containing frequencies
                       of labels in clusters.
}
\description{
Show label frequency in each cluster produced by dendrogram analysis
}
\examples{
if (sits_run_examples()) {
    clusters <- sits_cluster_dendro(cerrado_2classes)
    freq <- sits_cluster_frequency(clusters)
    freq
}
}
\seealso{
Other clustering and SOM:
\code{\link[=sits_cluster_clean]{sits_cluster_clean()}},
\code{\link[=sits_cluster_dendro]{sits_cluster_dendro()}},
\code{\link[=sits_som_clean_samples]{sits_som_clean_samples()}},
\code{\link[=sits_som_evaluate_cluster]{sits_som_evaluate_cluster()}},
\code{\link[=sits_som_map]{sits_som_map()}},
\code{\link[=sits_som_remove_samples]{sits_som_remove_samples()}}
}
\author{
Rolf Simoes, \email{rolfsimoes@gmail.com}
}
\concept{clustering and SOM}
