% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_cluster.R
\name{sits_cluster_hierarchical}
\alias{sits_cluster_hierarchical}
\title{Cluster a set of time series using hierarchical clustering}
\usage{
sits_cluster_hierarchical(data.tb, n_clusters = 6)
}
\arguments{
\item{data.tb}{a tibble the list of time series to be clustered}

\item{n_clusters}{the number of clusters to be identified}
}
\value{
none
}
\description{
\code{sits_cluster_hierarchical} cluster time series in hierarchical mode
}
\details{
Hierarchical clustering, as its name suggests, is an algorithm that tries to create a hierarchy
of groups in which, as the level in the hierarchy increases, clusters are created by merging
the clusters from the next lower level, such that an ordered sequence of groupings is obtained
(Hastie et al. 2009). In order to decide how the merging is performed, a similarity measure
between groups should be specied, in addition to the one that is used to calculate pairwise
similarities. However, a specific number of clusters does not need to be specified for the
hierarchy to be created, and the procedure is deterministic, so it will always give the same
result for a chosen set of similarity measures (taken from the DTWCLUST package docs).

Ref: Comparing Time-Series Clustering Algorithms in R Using the dtwclust Package
(dtwclust vignette)
}
\examples{
sits_plot  (ts, "Set1")


}
\seealso{
Other STIS auxiliary functions: \code{\link{sits_cluster_partitional}}
}
\keyword{STIS}
