% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_cluster.R
\name{sits_cluster_remove}
\alias{sits_cluster_remove}
\title{Remove cluster with mixed classes}
\usage{
sits_cluster_remove(data.tb, min_perc = 0.9)
}
\arguments{
\item{data.tb}{a SITS tibble with `cluster` column.}

\item{min_perc}{The minimum percentage of the most frequent label inside a cluster for the cluster not to be deleted.}
}
\value{
result.tb        a SITS tibble with all selected samples
}
\description{
The function removes cluster according to a percentage threshold "min_perc".
A cluster is considered good when the most frequent class has a percentage
of samples greater than the "min_perc" threshold.
}
\author{
Rolf Simoes, \email{rolf.simoes@inpe.br}
}
