% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_cluster.R
\name{sits_cluster_rename}
\alias{sits_cluster_rename}
\title{Cluster rename}
\usage{
sits_cluster_rename(data.tb, values)
}
\arguments{
\item{data.tb}{a SITS tibble with `cluster` column.}

\item{values}{character vector informing all cluster names. If unnamed vector is informed,
the index of each name will be treated as cluster code}
}
\value{
result.tb       SITS tibble with relabeled samples
}
\description{
Rename the SITS tibble clusters names according to a given value.
This function needs as input a SITS tibble with `cluster` column.
}
\author{
Rolf Simoes, \email{rolf.simoes@inpe.br}
}
