% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_cluster.R
\name{sits_clustersMembers}
\alias{sits_clustersMembers}
\title{Returns a tibble of centroids' members and its metadata.}
\usage{
sits_clustersMembers(data.tb, clusters, min_clu_perc)
}
\arguments{
\item{data.tb}{a tibble with input data of dtwclust.}

\item{clusters}{a cluster structure returned from dtwclust.}

\item{min_clu_perc}{the minimum percentage of members for a cluster to be valid}
}
\value{
centroids.tb    a SITS table with the clusters
}
\description{
reads a list of clusters provided by the dtwclust
package and produces a sits table.
}
\references{
"dtwclust" package (https://CRAN.R-project.org/package=dtwclust)
}
\author{
Rolf Simoes, \email{rolf.simoes@inpe.br}
}
