% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_colors.R
\name{sits_color_nama}
\alias{sits_color_nama}
\alias{sits_color_name}
\title{Brewer color schemes}
\usage{
sits_color_name(name = NULL)
}
\arguments{
\item{name}{name of the brewer color set}
}
\value{
name   string - a common color set name
}
\description{
brewer color schemes names to be used with `sits_plot_dendrogram`
}
