% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_colors.R
\name{sits_color_names}
\alias{sits_color_names}
\title{Show names of SITS color palettes}
\usage{
sits_color_names()
}
\value{
Names of available color palettes in SITS
}
\description{
Returns the names of the palettes supported by SITS
}
\examples{
if (sits_run_examples()) {
    # show the names of the colors supported by SITS
    sits_color_names()
    # show a color palette
    sits_colors_show("Blues", n_colors = 16)
}
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
