% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_colors.R
\name{sits_color_value}
\alias{sits_color_value}
\title{Function to retrieve sits color value}
\usage{
sits_color_value(name)
}
\arguments{
\item{name}{Name of color to obtain values}
}
\value{
A color value used in sits
}
\description{
Returns a color value based on name
}
\examples{
if (sits_run_examples()) {
    # show the names of the colors supported by SITS
    sits_color_value("Water")
}
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
