% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_colors.R
\name{sits_colors_reset}
\alias{sits_colors_reset}
\title{Function to reset sits color table}
\usage{
sits_colors_reset()
}
\value{
No return, called for side effects
}
\description{
Resets the color table
}
\examples{
if (sits_run_examples()) {
    # reset the default colors supported by SITS
    sits_colors_reset()
}
}
\seealso{
Other visualization:
\code{\link[=sits_colors]{sits_colors()}},
\code{\link[=sits_colors_qgis]{sits_colors_qgis()}},
\code{\link[=sits_colors_set]{sits_colors_set()}},
\code{\link[=sits_colors_show]{sits_colors_show()}},
\code{\link[=sits_sankey]{sits_sankey()}},
\code{\link[=sits_view]{sits_view()}}
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
\concept{visualization}
