% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_getdata.R
\name{sits_conf_fromJSON}
\alias{sits_conf_fromJSON}
\title{Obtain a confusion matrix from a JSON file.}
\usage{
sits_conf_fromJSON(file)
}
\arguments{
\item{json_file}{string  - name of a JSON file with sits data and metadata}
}
\value{
data.tb    tibble  - a SITS table
}
\description{
reads a set of data and metadata for satellite image time series from a JSON file
}
