% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_config.R
\name{sits_config_bands}
\alias{sits_config_bands}
\alias{.sits_config_bands}
\title{Retrieve the bands associated to a service in the configuration file}
\usage{
.sits_config_bands(service, name)
}
\arguments{
\item{service}{Name of a service.}

\item{name}{Name of a cube}
}
\description{
Retrieve the cubes associated a service.
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
