% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_config.R
\name{sits_config_color}
\alias{sits_config_color}
\alias{.sits_config_color}
\title{Retrieve the color associated to a class in the configuration file}
\usage{
.sits_config_color(label)
}
\arguments{
\item{label}{A class label.}
}
\description{
Retrieve the color associated a class label.
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
