% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_config.R
\name{sits_config_colors}
\alias{sits_config_colors}
\alias{.sits_config_colors}
\title{Retrieve the colors associated to classes}
\usage{
.sits_config_colors(labels)
}
\arguments{
\item{labels}{The class labels.}
}
\value{
list of colors.
}
\description{
Retrieve the color associated a class label.
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
\keyword{internal}
