% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_config.R
\name{sits_config_cubes}
\alias{sits_config_cubes}
\alias{.sits_config_cubes}
\title{Retrieve the cubes associated to a service based on the configuration file}
\usage{
.sits_config_cubes(service)
}
\arguments{
\item{service}{Name of a service.}
}
\description{
Retrieve the cubes associated a service.
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
