% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_config.R
\name{sits_config_satveg_bands}
\alias{sits_config_satveg_bands}
\alias{.sits_config_satveg_bands}
\title{Retrieve the bands associated to SATVEG}
\usage{
.sits_config_satveg_bands()
}
\value{
Names of SATVEG bands
}
\description{
Retrieve the cubes associated to the SATVEG service
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
\keyword{internal}
