% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_config.R
\name{sits_config_satveg_cubes}
\alias{sits_config_satveg_cubes}
\alias{.sits_config_satveg_cubes}
\title{Retrieve the cubes associated to SATVEG}
\usage{
.sits_config_satveg_cubes()
}
\description{
Retrieve the cubes associated to SATVEG.
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
\keyword{internal}
