% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_config.R
\name{sits_config_show}
\alias{sits_config_show}
\title{Show current sits configuration}
\usage{
sits_config_show()
}
\value{
No return value, called for side effects.
}
\description{
Prints the current sits configuration options.
To show specific configuration options for
a source, a collection, or a palette, users can inform the corresponding
keys to \code{source} and \code{collection}.
}
\examples{
sits_config_show()
}
\seealso{
Other utilities:
\code{\link[=sits_config]{sits_config()}},
\code{\link[=sits_config_user_file]{sits_config_user_file()}},
\code{\link[=sits_factory_function]{sits_factory_function()}},
\code{\link[=sits_list_collections]{sits_list_collections()}},
\code{\link[=sits_parallel]{sits_parallel()}},
\code{\link[=sits_run_examples]{sits_run_examples()}},
\code{\link[=sits_run_tests]{sits_run_tests()}}
}
\concept{utilities}
