% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_config.R
\name{sits_config_user_file}
\alias{sits_config_user_file}
\title{Create a user configuration file.}
\usage{
sits_config_user_file(file_path, overwrite = FALSE)
}
\arguments{
\item{file_path}{file to store the user configuration file}

\item{overwrite}{replace current configuration file?}
}
\value{
Called for side effects
}
\description{
Creates a user configuration file.
}
\examples{
user_file <- paste0(tempdir(), "/my_config_file.yml")
sits_config_user_file(user_file)
}
\seealso{
Other utilities:
\code{\link[=sits_config]{sits_config()}},
\code{\link[=sits_config_show]{sits_config_show()}},
\code{\link[=sits_factory_function]{sits_factory_function()}},
\code{\link[=sits_list_collections]{sits_list_collections()}},
\code{\link[=sits_parallel]{sits_parallel()}},
\code{\link[=sits_run_examples]{sits_run_examples()}},
\code{\link[=sits_run_tests]{sits_run_tests()}}
}
\concept{utilities}
