% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_WTSS.R
\name{sits_config_WTSS}
\alias{sits_config_WTSS}
\title{Set important global variables for WTSS access}
\usage{
sits_config_WTSS(URL = "http://www.dpi.inpe.br/tws/wtss",
  coverage = "mod13q1_512", bands = c("ndvi", "evi"))
}
\arguments{
\item{URL}{the URL for the WTSS time series service}

\item{coverage}{name of space-time coverage that contains the time series information}

\item{bands}{a character vector with the names of the bands to be retrieved}
}
\description{
\code{sits_config} stores the information about the WTSS server
and the coverage name to be retrieved in the global environment
The assumption is that in a SITS session, this information is defined one and
used in many applications
}
\details{
This function should be run at the start of each SITS session
}
