% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_config.R
\name{sits_configuration}
\alias{sits_configuration}
\alias{sits_config}
\alias{sits_config_show}
\alias{sits_config_list_collections}
\title{sits configuration}
\usage{
sits_config(
  processing_bloat = NULL,
  rstac_pagination_limit = NULL,
  raster_api_package = NULL,
  gdal_creation_options = NULL,
  reset = FALSE
)

sits_config_show(source = NULL, collection = NULL, palette = NULL)

sits_config_list_collections(source = NULL)
}
\arguments{
\item{processing_bloat}{A \code{numeric} value to estimate
growth size of R memory relative to block size.}

\item{rstac_pagination_limit}{A \code{numeric} value indicating number of
items returned by STAC service.}

\item{raster_api_package}{A \code{character} value indicating a
supported raster handling package.}

\item{gdal_creation_options}{A \code{character} vector specifying GDAL
creation option for GeoTiff image format.}

\item{reset}{A \code{logical} value indicating if current
configuration options must be cleaned before
load config files. Default \code{FALSE}.}

\item{source}{A \code{character} value indicating a source
key entry to be shown in detail.}

\item{collection}{A \code{character} value used in conjunction
with \code{source} parameter to indicate a collection key entry to be shown
in detail.}

\item{palette}{A \code{character} value indicating a palette
to be shown in detail.}
}
\value{
\code{sits_config()} returns a \code{list} containing the final
configuration options.

\code{sits_config_show()} returns a \code{list} containing the respective
configuration printed in the console.

\code{sits_config_list_collection()} prints the collections available in
each cloud service supported by sits.
}
\description{
These functions load and show sits configurations.

\code{sits_config()} loads the default configuration file and
the user provided configuration file. The final configuration is
obtained by overriding the options by the values provided in
\code{processing_bloat}, \code{rstac_pagination_limit},
\code{raster_api_package}, and \code{gdal_creation_options} parameters.

\code{sits_config_show()} prints the current sits
configuration options. To show specific configuration options for
a source, a collection, or a palette, user can inform the corresponding
keys to \code{source}, \code{collection}, and \code{palette} parameters.
}
\details{
Users can provide additional configuration files, by specifying the
location of their file in the environmental variable
\code{SITS_CONFIG_USER_FILE}.

To see the key entries and contents of the current configuration values,
use \code{sits_config_show()}.
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}

Rolf Simoes, \email{rolf.simoes@inpe.br}
}
