% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_WTSS.R
\name{sits_configWTSS}
\alias{sits_configWTSS}
\title{Set important global variables for WTSS access}
\usage{
sits_configWTSS(URL = "http://www.dpi.inpe.br/tws/wtss",
  coverage = "mod13q1_512", bands = c("ndvi", "evi", "nir"))
}
\arguments{
\item{URL}{the URL for the WTSS time series service}

\item{coverage}{name of space-time coverage that contains the time series information}

\item{bands}{a character vector with the names of the bands to be retrieved}
}
\value{
wtss_info       a list containing the information about the WTSS server and coverage
}
\description{
\code{sits_configWTSS} stores the information about the WTSS server
and the coverage name to be retrieved in the global environment
The assumption is that in a SITS session, this information is defined one and
used in many applications
}
\details{
This function should be run at the start of each SITS session.

The Web Time Seris Service is a lightweight web service the allow remote access to satellite
 image time series and provides three operations:

1. list coverages: this operation allows clients to retrieve the capabilities provided
    by any server that implements WTSS. It returns a list of coverage
    names available in a server instance.

2. describe coverage: this operation returns the metadata for a given coverage
    identified by its name. It includes its range in the spatial and temporal dimensions.

3. time series: this operation requests the time series of values of a coverage attribute
   at a given location.

To allow for further processing of time series from a WTSS service, this operation
 creates a list with the following variables

URL            - the URL for the WTSS time series service
 ts_server      - an object that handles the WTSS connection
 coverage       - the name of the coverage that contains the time series for the analysis
 cov_desc       - a table with the parameters of the coverage
 start_date     - the start date for the time series data in the coverage
 end_date       - the end date for the time series data in the coverage
 res_x          - spatial resolution (x dimension)
 res_y          - spatial resolution (y dimension)
 start_date     - initial date of the coverage time series
 end_date       - final date of the coverage time series
 x_min          - spatial extent (xmin)
 y_min          - spatial extent (ymin)
 x_max          - spatial extent (xmax)
 y_max          - spatial extent (ymin)
 bands          - the bands of the data to be retrieved from the WTSS
}

