% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_coverage.R
\name{sits_coverageRaster}
\alias{sits_coverageRaster}
\title{Create a metadata tibble to store the description of a spatio-temporal raster dataset}
\usage{
sits_coverageRaster(product = "MOD13Q1", coverage = NULL, timeline = NULL,
  bands, files)
}
\arguments{
\item{product}{The image product where the files are extracted (e.g. MOD13Q1)}

\item{coverage}{The name of the coverage file}

\item{timeline}{Vector of dates with the timeline of the bands}

\item{bands}{The bands contained in the Raster Brick set (in the same order as the files)}

\item{files}{Vector with the file paths of the raster files}
}
\value{
raster.tb   A tibble with metadata information about a raster data set
}
\description{
This function creates a tibble containing the metadata for
              a set of spatio-temporal raster files, organized as a set of "Raster Bricks".
              These files should be of the same size and
              projection. Each raster brick file should contain one band
              per time step. Different bands are archived in different raster files.
}
\examples{
# read a raster file and put it into a vector
files  <- c(system.file ("extdata/raster/mod13q1/sinop-crop-ndvi.tif", package = "sits"))

# define the timeline
data(timeline_mod13q1)
timeline <- lubridate::as_date (timeline_mod13q1$V1)

# create a raster metadata file based on the information about the files
#raster.tb <- sits_coverageRaster(coverage = "Sinop-crop",
#             timeline = timeline, bands = c("ndvi"), files = files)

}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
