% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_coverage.R
\name{sits_coverageSATVEG}
\alias{sits_coverageSATVEG}
\title{Retrieve a coverage name from the SATVEG service}
\usage{
sits_coverageSATVEG(coverage, bands = c("ndvi", "evi"), product = "MOD13Q1")
}
\arguments{
\item{coverage}{the name of the coverage ("terra", "aqua", "comb")}

\item{bands}{the bands available in SATVEG}

\item{product}{the SATVEG product}
}
\value{
coverage.tb       metadata about the coverage
}
\description{
Retrieves a coverage name based on the capabilities of the SATVEG service
}
\author{
Julio Esquerdo, \email{julio.esquerdo@embrapa.br}

Gilberto Camara, \email{gilberto.camara@inpe.br}
}
