% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_WTSS.R
\name{sits_coverageWTSS}
\alias{sits_coverageWTSS}
\title{Provides information about one coverage of the WTSS service}
\usage{
sits_coverageWTSS(URL = "http://www.dpi.inpe.br/tws/wtss", coverage = NULL)
}
\arguments{
\item{URL}{the URL for the WTSS time series service}

\item{coverage}{the name of the coverage}
}
\description{
\code{sits_coverageWTSS} uses the WTSS services to provide information about a
chosen coverage
}
