% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_validate.R
\name{sits_create_partitions}
\alias{sits_create_partitions}
\alias{.sits_create_partitions}
\title{Create partitions of a data set}
\usage{
.sits_create_partitions(data.tb, times, perc)
}
\arguments{
\item{data.tb}{a SITS table to be partitioned}

\item{times}{number of iterations}

\item{perc}{percentagem of original data to be extracted}
}
\description{
Create a list of partitions of a SITS table, based on a percentage and
a number of iterations
}
\author{
Rolf Simoes, \email{rolf.simoes@inpe.br}
}
