% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_SCIDB.R
\name{sits_createColRowSequence}
\alias{sits_createColRowSequence}
\title{Get unique SciDB columns and rows indexes}
\usage{
sits_createColRowSequence(col_id, row_id)
}
\arguments{
\item{col_id}{a integer index with SciDB column indexes}

\item{row_id}{a integer index with SciDB row indexes}
}
\value{
sits.tb a sits table
}
\description{
This function returns a sits table with unique SciDB columns and rows indexes.
}
