% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_SCIDB.R
\name{sits_createZooObject}
\alias{sits_createZooObject}
\title{Create an indexed ordered obsevartion set within a sits_table.}
\usage{
sits_createZooObject(bands, dates, scale_factor, idx, missing_values)
}
\arguments{
\item{bands}{a character string vector with the interested bands of the satellite imagery product}

\item{dates}{a string vector contaning time series dates}

\item{scale_factor}{a real number representing the scale factor of the attribute values}

\item{idx}{a integer vector containing the respective index of the attribute values}

\item{missing_values}{the missing value of the coverage}
}
\value{
ts.tb
}
\description{
This function returns a sits table with an indexed ordered observation set as a zoo object
}
