% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_table.R
\name{sits_cross}
\alias{sits_cross}
\title{Cross join bands of two satellite image time series}
\usage{
sits_cross(sits1.tb, sits2.tb)
}
\arguments{
\item{sits1.tb}{the first SITS table in wich entries will be crossed with ts2 entries}

\item{sits2.tb}{the second SITS table entries}
}
\value{
new.tb    a cross-joined SITS tibble with a nested set of time series
}
\description{
Cross-join two SITS tables with the same spatio-temporal references. To cross two series' bands,
we consider that they contain different attributes (bands) but refer to the same coverage.
We make no assumptions about their spatio-temporal location.
This function is useful to create different bands of clusters centroids time series.
For example, one may want to put evi and ndvi bands centroids together in an cross joined fashion in order to
generate patterns combinations.
}
\author{
Rolf Simoes, \email{rolf.simoes@inpe.br}
}
