% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_log.R
\name{sits_csv_error_file}
\alias{sits_csv_error_file}
\title{Loads the CSV error file saved in the log directory}
\usage{
sits_csv_error_file(file_name = "errors.csv")
}
\description{
Loads the CSV error file saved in the log directory
}
