% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_cube.R
\name{sits_cube.bdc_cube}
\alias{sits_cube.bdc_cube}
\title{Defines a data cube for a BDC TILE}
\usage{
\method{sits_cube}{bdc_cube}(
  type = "BDC_TILE",
  ...,
  name = NULL,
  satellite = NULL,
  sensor = NULL,
  bands = NULL,
  cube = NULL,
  tiles = NULL,
  version = "v001",
  data_access = "web",
  start_date = NULL,
  end_date = NULL,
  .local = NULL,
  .web = NULL,
  .cloud_band = FALSE
)
}
\arguments{
\item{type}{type of cube}

\item{...}{other parameters to be passed for specific types}

\item{name}{output data cube.}

\item{satellite}{satellite}

\item{sensor}{sensor}

\item{bands}{bands.}

\item{cube}{name input data cube in BDC}

\item{tiles}{names of the tiles}

\item{version}{version of the cube}

\item{data_access}{access (local or web)}

\item{start_date}{starting date of the cube}

\item{end_date}{ending date of the cube}

\item{.local}{directory for local access to the input cube (optional)}

\item{.web}{directory for web access to the input cube (optional)}

\item{.cloud_band}{include cloud band? (TRUE/FALSE)}
}
\value{
A data cube
}
\description{
Defines a cube to retrieve data from the Brazil Data Cube (BDC). The retrieval
             is based on tiles of a given cube. Allows local or web access. For
             local access, the user should be logged in the BDC.
             For more on BDC, please see http://brazildatacube.dpi.inpe.br/
}
\examples{
\donttest{

# create a raster cube file based on the information about the files
cbers_bdc_tile <- sits_cube(type        = "BDC_TILE",
                            name        = "022024",
                            satellite   = "CBERS-4",
                            sensor      = "AWFI",
                            cube        = "CB4_64_16D_STK",
                            tiles       = "022024",
                            version     = "v001",
                            data_access = "web",
                            bands       = c("NDVI", "EVI"),
                            start_date  = as.Date("2018-08-29"),
                            end_date    = as.Date("2019-08-13"))
}

}
