% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_cube.R
\name{sits_cube.brick_cube}
\alias{sits_cube.brick_cube}
\title{Defines a data cube for a BRICK}
\usage{
\method{sits_cube}{brick_cube}(
  type = "BRICK",
  ...,
  name = NULL,
  satellite = NULL,
  sensor = NULL,
  timeline = NULL,
  bands = NULL,
  files = NULL,
  start_date = NULL,
  end_date = NULL
)
}
\arguments{
\item{type}{Type of cube}

\item{...}{Other parameters to be passed for specific types}

\item{name}{Name of the output data cube.}

\item{satellite}{Name of satellite}

\item{sensor}{Name of sensor}

\item{timeline}{Vector with the timeline of the collection}

\item{bands}{Vector of bands.}

\item{files}{Vector of file names for each band}

\item{start_date}{Starting date of the cube}

\item{end_date}{Ending date of the cube}
}
\value{
A brick data cube
}
\description{
Defines a cube to retrieve data from a set of raster bricks.
             Each band has to be organised as a raster brick, and the
             number of input files must match the number of bands.
             All input files must have the same spatial resolution and
             share the same timeline (in order).
             The timeline for the cube must be provided.
}
\examples{
\donttest{
# Create a raster cube based on bricks
# inform the files that make up a raster brick with 392 time instances
files <- c(system.file("extdata/raster/mod13q1/sinop-crop-ndvi.tif",
           package = "sits"))

# create a raster cube file based on the information about the files
raster.tb <- sits_cube(type = "BRICK",
                       name      = "Sinop-crop",
                       satellite = "TERRA",
                       sensor    = "MODIS",
                       timeline  = timeline_modis_392,
                       bands     = "NDVI",
                       files     = files)
}
}
