% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_cube.R
\name{sits_cube.deafrica_cube}
\alias{sits_cube.deafrica_cube}
\title{Defines a data cube for Digital Earth Africa STAC}
\usage{
\method{sits_cube}{deafrica_cube}(
  type = "DEAFRICA",
  ...,
  name = "deafrica_cube",
  url = NULL,
  collection = NULL,
  tiles = NULL,
  bands = NULL,
  roi = NULL,
  start_date = NULL,
  end_date = NULL
)
}
\arguments{
\item{type}{Type of cube.}

\item{...}{Other parameters to be passed for specific types.}

\item{name}{Name of the output data cube (optional).}

\item{url}{URL for the DEAfrica catalog (optional). By default,
'https://explorer.digitalearth.africa/stac/' will be used.}

\item{collection}{DEAFRICA collection to be searched (mandatory).}

\item{tiles}{Tile names to be searched (optional).}

\item{bands}{Bands names to be filtered (optional).}

\item{roi}{Selects images (tiles) that intersect according to the
region of interest provided. Expressed either as an \code{sfc} or \code{sf}
object from sf package, a \code{character} with GeoJSON following the rules
from RFC 7946, or a \code{vector} with bounding box named XY values in
WGS 84 ("xmin", "ymin", "xmax", "ymax").
Obs.: This parameter does not crop a region, but only selects the images
that intersect with it.}

\item{start_date}{Initial date for the cube files (optional).}

\item{end_date}{Final date for the cube files (optional).}
}
\value{
A data cube.
}
\description{
Defines a cube to retrieve data from the Digital Earth Africa
 (DEAFRICA) STAC. The retrieval is based on tiles of a given cube.
 For more on DEAfrica, please see https://www.digitalearthafrica.org/
}
\note{
For now, we only support the collections 'ga_s2_gm' and 's2_l2a'.
We are working on supporting the other products offered by DEAfrica STAC.
}
\examples{
\dontrun{
# this example requires access to an external service, so should not be run
# by CRAN

# Provide your AWS credentials here
# Sys.setenv(
# "AWS_ACCESS_KEY_ID"     = <your_access_key>,
# "AWS_SECRET_ACCESS_KEY" = <your_secret_access_key>,
# "AWS_DEFAULT_REGION"    = <your AWS region>,
# "AWS_ENDPOINT" = <your_end_point>,
# "AWS_REQUEST_PAYER"     = "requester"
# )

# create a raster cube file based on the information about the files
cube_dea <- sits::sits_cube(type = "DEAFRICA",
                           name = "deafrica_cube",
                           collection = "ga_s2_gm",
                           bands = c("B04", "B08"),
                           roi = c("xmin" = 17.379,
                                  "ymin" = 1.1573,
                                  "xmax" = 17.410,
                                  "ymax" = 1.1910),
                           start_date = "2019-01-01",
                           end_date = "2019-10-28")
}
}
\references{
`rstac` package (https://github.com/brazil-data-cube/rstac)
}
