% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_cube.R
\name{sits_cube.default}
\alias{sits_cube.default}
\title{Default methods for sits_cube}
\usage{
\method{sits_cube}{default}(type = NULL, ...)
}
\arguments{
\item{type}{Type of cube}

\item{...}{Other parameters to be passed for specific types}
}
\description{
Default methods for sits_cube
}
