% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_cube.R
\name{sits_cube.probs_cube}
\alias{sits_cube.probs_cube}
\title{Defines a cube from a set of image bricks}
\usage{
\method{sits_cube}{probs_cube}(
  type = "PROBS",
  ...,
  names = "probs_cube",
  satellite,
  sensor,
  timeline,
  labels,
  files
)
}
\arguments{
\item{type}{type of cube}

\item{...}{other parameters}

\item{names}{names of output data cube}

\item{satellite}{satellite}

\item{sensor}{sensor}

\item{timeline}{vector with timeline of the files}

\item{labels}{vector of labels associated to the probs cube}

\item{files}{vector of file names for each band}
}
\value{
data cube
}
\description{
Defines a cube to retrieve data from a set of image files
             that have been classified.
             All image files should have the same spatial resolution
             and same projection. Each probs image
             has to be organised as a raster brick, and the
             number of layers must match the number of labels.
             All input files must have the same spatial resolution and
             share the same timeline (in order).
             The timeline for the cube must be provided.
}
\examples{

# Create a raster cube based on bricks
# inform the files that make up a raster probs brick with 23 time instances
probs_file <- c(system.file("extdata/raster/mod13q1/sinop-2014_probs_2013_9_2014_8_v1.tif",
    package = "sits"
))

# inform the labels
labels <- c("Cerrado", "Fallow_Cotton", "Forest", "Pasture", "Soy_Corn",
"Soy_Cotton", "Soy_Fallow", "Soy_Millet", "Soy_Sunflower")


# create a raster cube file based on the information about the files
probs_cube <- sits_cube(
    type = "PROBS",
    names = "Sinop-crop-probs",
    satellite = "TERRA",
    sensor  = "MODIS",
    timeline = timeline_2013_2014,
    labels = labels,
    files = probs_file
)
}
