% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_cube.R
\name{sits_cube.s2_l2a_aws_cube}
\alias{sits_cube.s2_l2a_aws_cube}
\title{Defines a data cube for a Sentinel-2 L2A AWS cube}
\usage{
\method{sits_cube}{s2_l2a_aws_cube}(
  type = "S2_L2A_AWS",
  ...,
  name = NULL,
  bands = NULL,
  cube = NULL,
  tile = NULL,
  start_date = NULL,
  end_date = NULL,
  s2_aws_resolution = NULL,
  access_key = NULL,
  secret_key = NULL,
  region = NULL
)
}
\arguments{
\item{type}{Type of cube}

\item{...}{Other parameters to be passed for specific types}

\item{name}{Name of the output data cube.}

\item{bands}{Vector of bands.}

\item{cube}{Name of the input data cube (or image collection)}

\item{tile}{Name of the tile}

\item{start_date}{Starting date of the cube}

\item{end_date}{Ending date of the cube}

\item{s2_aws_resolution}{Resolution of Sentinel images in AWS ("10m", "20m" or "60m")}

\item{access_key}{AWS access key}

\item{secret_key}{AWS secret key}

\item{region}{AWS region}
}
\value{
A data cube
}
\description{
Defines a cube to retrieve data from the Sentinel-2 Level 2A data
             available in AWS. To access this data, the user needs to be an AWS
             user and provide her access key and secret key. These keys may
             be passed as environment variables. The bands available in AWS for
             10m resolution are "B02", "B03", "B04", and "B08". The  20m bands
             are "B02", "B03", "B04", "B05", "B06", "BO7", "B08", "B8A", "B11", and "B12".
             All 12 bands are available at 60m resolution.
}
\examples{
\donttest{

# Provide your AWS credentials here
# Sys.setenv(
# "AWS_ACCESS_KEY_ID"     = <your_access_key>,
# "AWS_SECRET_ACCESS_KEY" = <your_secret_access_key>,
# "AWS_DEFAULT_REGION"    = <your AWS region>,
# "AWS_REQUEST_PAYER"     = "requester"
# )

s2_cube <- sits_cube(type       = "S2_L2A_AWS",
                     name       = "T20LKP_2018_2019",
                     satellite  = "SENTINEL-2",
                     sensor     = "MSI",
                     tile       = "20LKP",
                     s2_aws_resolution = "20m",
                     start_date = as.Date("2018-07-18"),
                     end_date   = as.Date("2018-07-23"))
}

}
