% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_cube.R
\name{sits_cube.satveg_cube}
\alias{sits_cube.satveg_cube}
\title{Defines a data cube for the SATVEG service}
\usage{
\method{sits_cube}{satveg_cube}(type = "SATVEG", name = NULL, ...)
}
\arguments{
\item{type}{Type of cube}

\item{name}{Name of the input data ("terra", "aqua", "comb").}

\item{...}{Other parameters to be passed for specific types}
}
\value{
A valid data cube
}
\description{
The SATVEG service is provided by the Embrapa Agricultural
 Informatics Centre and provides access to time series from the MODIS sensor.
 There are three types of time series: "terra" (from the TERRA satellite),
 "aqua" (from the AQUA satellite) and "comb" (combination of both satellites)
}
\examples{
\dontrun{
# Create a data cube based on the SATVEG service
cube_satveg <- sits_cube(
    type = "SATVEG",
    name = "terra"
)
}

}
