% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_cube.R
\name{sits_cube.stack_cube}
\alias{sits_cube.stack_cube}
\title{Defines a cube to retrieve data from a set of image files}
\usage{
\method{sits_cube}{stack_cube}(
  type = "STACK",
  ...,
  name,
  satellite,
  sensor,
  resolution,
  data_dir,
  parse_info,
  delim = NULL
)
}
\arguments{
\item{type}{Type of cube}

\item{...}{Other parameters to be passed for specific types}

\item{name}{Name of the output data cube.}

\item{satellite}{Name of satellite}

\item{sensor}{Name of sensor}

\item{resolution}{Resolution of sensor}

\item{data_dir}{Directory where data is located}

\item{parse_info}{The parsing information (see above)}

\item{delim}{A character to use as delimiter (default = "_")}
}
\value{
A brick data cube
}
\description{
Defines a cube to retrieve data from a set of image files.
             All image files should have the same spatial resolution
             and same projection. In addition, image file names should
             include information on band and date.
             The timeline and the bands are deduced from this information.
             Examples of valid image names include
             "CB4_64_16D_STK_022024_2018-08-29_2018-09-13_EVI.tif" and
             "B02_2018-07-18.jp2". In each case, the user has to provide
             appropriate parsing information that allows SITS to extract
             the band and the date. In the examples above, the parsing info
             would include "_" as a delimiter. In the first, the names of the
             resulting columns after parsing are "X1", "X2", "X3", "X4", "X5",
             "date", "X7", and "band". In the second, they are "band" and "date".
}
\examples{
\donttest{
# Create a raster cube based on CBERS data provided by the inSitu package
data_dir <- system.file("extdata/CBERS/CB4_64_16D_STK/022024", package = "inSitu")

# create a raster cube file based on the information about the files
cbers_stack.tb <- sits_cube(type       = "STACK",
                            name       = "022024",
                            satellite  = "CBERS-4",
                            sensor     = "AWFI",
                            resolution = "64m",
                            data_dir   = data_dir,
                            parse_info = c("X1", "X2", "X3", "X4", "X5", "date", "X7", "band"))
}
}
