% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_cube.R
\name{sits_cube.wtss_cube}
\alias{sits_cube.wtss_cube}
\title{Defines a data cube for the WTSS service}
\usage{
\method{sits_cube}{wtss_cube}(type = "WTSS", ..., name = NULL, URL = NULL)
}
\arguments{
\item{type}{Type of cube}

\item{...}{Other parameters to be passed for specific types}

\item{name}{Name of the output data cube.}

\item{URL}{URL of the service provider.}
}
\value{
A valid data cube
}
\description{
Implements an interface to a web time series service (WTSS)
              that offers time series of remote sensing data using a simple API.
}
\examples{
\donttest{
# Create a data cube based on a WTSS service
cube_wtss <- sits_cube(type = "WTSS",
          name = "MOD13Q1",
          URL = "http://www.esensing.dpi.inpe.br/wtss/")
}
}
