% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_cube_aux_functions.R
\name{sits_cube_timeline}
\alias{sits_cube_timeline}
\title{Return the timeline associated to a data cube, given an index}
\usage{
sits_cube_timeline(cube, index = 1)
}
\arguments{
\item{cube}{Metadata about a data cube}

\item{index}{Index of timeline list}
}
\value{
Vector of times for an index
}
\description{
Given a data cube, retrieve the timeline
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
