% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_config.R
\name{sits_cubes_service}
\alias{sits_cubes_service}
\alias{.sits_cubes_service}
\title{Retrieve the cubes associated to a service}
\usage{
.sits_cubes_service(service)
}
\arguments{
\item{service}{Name of a service.}
}
\description{
Retrieve the cubes associated a service.
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
