% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_csv.R
\name{sits_data_to_csv}
\alias{sits_data_to_csv}
\title{Export a sits tibble data to the CSV format}
\usage{
sits_data_to_csv(data, file)
}
\arguments{
\item{data}{A tibble with time series data and metadata.}

\item{file}{Name of the exported CSV file.}
}
\value{
Status of the operation.
}
\description{
Converts time series data from a sits tibble to a CSV file. The CSV file will not contain the metadata,
but will have the actual time series, with a reference value. This function is useful to
export the data for external applications
}
\examples{
\donttest{
# read a tibble with 400 samples of Cerrado and 346 samples of Pasture
data(cerrado_2classes)
# export a time series to zoo
sits_data_to_csv(cerrado_2classes, file = "cerrado_2classes.csv")
}
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
