% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_db.R
\name{sits_db_create}
\alias{sits_db_create}
\title{Creates a connection to an RSQLite database}
\usage{
sits_db_create(name = NULL)
}
\arguments{
\item{name}{Name of the database}
}
\value{
A connection to an RSQLite database
}
\description{
This functions creates a connection to an RSQLite database
to be able to store the data and metadata tables associated to time series
and data cubes associated to the "sits" suite of packages.
}
\examples{
# create RSQLite connection
conn <- sits_db_create()
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
