% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_db.R
\name{sits_db_read}
\alias{sits_db_read}
\title{Read time series and data cubes information on an SQLite database}
\usage{
sits_db_read(conn, name)
}
\arguments{
\item{conn}{SQLite connection}

\item{name}{Name of the object to be read}
}
\value{
Tibble with time series
}
\description{
This functions write a set of time series or data cubes
to an RSQLite database.
}
\examples{
\dontrun{
# create RSQLite connection
conn <- sits_db_connect()
# write a set of time series
conn <- sits_db_write(conn, "cerrado_2classes", cerrado_2classes)
# read a set of time series
ts <- sits_db_read(conn, "cerrado_2classes")
}
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
