% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_db.R
\name{sits_db_read.sits}
\alias{sits_db_read.sits}
\title{Read time series from an SQLite database}
\usage{
\method{sits_db_read}{sits}(conn, name)
}
\arguments{
\item{conn}{SQLite connection}

\item{name}{Name of the object to be storeed}

\item{data}{Data to be retrived}
}
\value{
A sits tibble
}
\description{
This function reads a set of time series
to an RSQLite database.
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
\keyword{internal}
