% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_db.R
\name{sits_db_write.sits}
\alias{sits_db_write.sits}
\title{Write time series on an SQLite database}
\usage{
\method{sits_db_write}{sits}(conn, name, data)
}
\arguments{
\item{conn}{SQLite connection}

\item{name}{Name of the object to be stored}

\item{data}{Data to be stored}
}
\value{
A connection to an RSQLite database
}
\description{
This functions write a set of time series
to an RSQLite database.
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
\keyword{internal}
