% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_machine_learning.R
\name{sits_deep_learning}
\alias{sits_deep_learning}
\title{Train SITS classifiction models with deep learning}
\usage{
sits_deep_learning(distances.tb = NULL, ...)
}
\arguments{
\item{distances.tb}{a time series with a set of distance measures for each training sample}

\item{...}{other parameters to be passed to `h2o::h2o.deeplearning` function}
}
\value{
result          either an model function to be passed in sits_predict or an function prepared that can be called further to compute multinom training model
}
\description{
Use deep learning algorithm to classify data.
This function is a front-end to the "deeplearning" method in the "randomForest" package.
Please refer to the documentation in that package for more details.
}
\author{
Alexandre Xavier Ywata de Carvalho, \email{alexandre.ywata@ipea.gov.br}

Rolf Simoes, \email{rolf.simoes@inpe.br}
}
