% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_cluster.R
\name{sits_dendogram}
\alias{sits_dendogram}
\title{Cluster a set of time series using hierarchical clustering}
\usage{
sits_dendogram(data.tb, n_clusters = 4)
}
\arguments{
\item{data.tb}{a tibble the list of time series to be clustered}

\item{n_clusters}{the number of clusters to be identified}
}
\value{
clusters.tb a SITS tibble with the clusters
}
\description{
Cluster time series in hierarchical mode. Hierarchical clustering, as its name suggests,
is an algorithm that tries to create a hierarchy
of groups in which, as the level in the hierarchy increases, clusters are created by merging
the clusters from the next lower level, such that an ordered sequence of groupings is obtained.
The similarity measure used to group time series in a cluster is the dtw metric.
The procedure is deterministic, so it will always give the same
result for a chosen set of similarity measures (taken from the DTWCLUST package docs).
}
\references{
"dtwclust" package (https://CRAN.R-project.org/package=dtwclust)
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}

Rolf Simoes, \email{rolf.simoes@inpe.br}
}
