% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_detect_change.R
\name{sits_detect_change}
\alias{sits_detect_change}
\title{Detect changes in time series}
\usage{
sits_detect_change(data, cd_method)
}
\arguments{
\item{data}{Set of time series}

\item{cd_method}{Change detection method (with parameters)}
}
\value{
Set of time series where significant change has been detected
}
\description{
Given a set of time series or an image, this function compares each time series
                        with a set of change/no-change patterns, and indicates places and dates where
                         change has been detected
}
