% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_data_table.R
\name{sits_distance_table}
\alias{sits_distance_table}
\title{Create an empty distance table to store the results of distance metrics}
\usage{
sits_distance_table(patterns.tb)
}
\arguments{
\item{patterns.tb}{a tibble with a set of patterns}
}
\value{
distances.tb   a data.table to store the distances between a time series and a set of patterns
}
\description{
Create an empty distance data.table to store the results of distance metrics
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}

Rolf Simoes, \email{rolf.simoes@inpe.br}
}
