% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_distances.R
\name{sits_distance_tibble_from_data}
\alias{sits_distance_tibble_from_data}
\title{Create an empty distance tibble based on an input data set}
\usage{
sits_distance_tibble_from_data(data.tb)
}
\arguments{
\item{data.tb}{a SITS tibble with a data set}
}
\value{
distances.tb   a tibble to store the distances between a time series and a set of patterns
}
\description{
Create an empty distance tibble to store the results of distance metrics
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}

Rolf Simoes, \email{rolf.simoes@inpe.br}
}
