% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_distances.R
\name{sits_distances}
\alias{sits_distances}
\title{Calculate a set of distance measures for satellite image time series}
\usage{
sits_distances(data.tb, patterns.tb, dist_method = sits_distances_from_data())
}
\arguments{
\item{data.tb}{SITS tibble time series}

\item{patterns.tb}{Set of patterns obtained from training samples}

\item{dist_method}{Method for calculating distances between time series}
}
\value{
result          Set of distance metrics
}
\description{
This function allows the user to select different alternatives to define a set of
distances between a set of satellite image time series and a set of patterns.
 The alternatives are:
"TWTDTW"    - uses the TWDTW (time-weighted dynamic time warping)
other methods as used by the TSdist package
}
\author{
Rolf Simoes, \email{rolf.simoes@inpe.br}

Gilberto Camara, \email{gilberto.camara@inpe.br}
}
