% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_distances.R
\name{sits_distances_from_data2}
\alias{sits_distances_from_data2}
\title{Use time series values from a sits tibble as distances for training patterns}
\usage{
sits_distances_from_data2(data.tb = NULL, fun = NULL, start_from, interval)
}
\arguments{
\item{data.tb}{a SITS tibble with original data}

\item{fun}{a function to be called over all data input. It must takes one argument as input.}

\item{start_from}{start date to break segments}

\item{interval}{an interval to define the lengths of breaked segments}
}
\value{
distances.tb  a tibble where columns have the reference label and the time series values as distances
}
\description{
This function allows using a set of labelled time series as
input to the machine learning models. Instead of first estimating a set
of idealised patterns and then computing distances from these patterns,
the attributes used to train the model are the series themselves.
This function then extracts the time series from a SITS tibble and
"spreads" them in time to produce a tibble with distances.
}
\author{
Rolf Simoes, \email{rolf.simoes@inpe.br}

Gilberto Camara, \email{gilberto.camara@inpe.br}
}
