% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_distances.R
\name{sits_distances_from_ts}
\alias{sits_distances_from_ts}
\title{Use time series values as distances for classification}
\usage{
sits_distances_from_ts(ts.tb, shift = 3)
}
\arguments{
\item{ts.tb}{Tibble with time series}

\item{shift}{Adjustment value to avoid negative pixel vales}
}
\value{
distances.tb  a tibble where columns have the reference label and the time series values as distances
}
\description{
This function allows using a time series as
input to the machine learning classification.
This function extracts the time series and "spreads" them in time to produce a tibble with distances.
}
\author{
Rolf Simoes, \email{rolf.simoes@inpe.br}

Gilberto Camara, \email{gilberto.camara@inpe.br}
}
