% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_distances.R
\name{sits_distances_normalized}
\alias{sits_distances_normalized}
\title{Use time series values from a sits tibble as distances for training patterns}
\usage{
sits_distances_normalized(data.tb = NULL, method = "center", margin = 2)
}
\arguments{
\item{data.tb}{a SITS tibble with original data}

\item{method}{Normalizing method. Available are: "center": Subtract mean.
"scale": Divide by standard deviation. "standardize": Center and scale.
"range": Scale to a given range.}

\item{margin}{1 = rows, 2 = cols}
}
\value{
distances.tb  a tibble where columns have the reference label and the time series values as distances
}
\description{
This function allows using a set of labelled time series as
input to the machine learning models. Instead of first estimating a set
of idealised patterns and then computing distances from these patterns,
the attributes used to train the model are the series themselves.
This function then extracts the time series from a SITS tibble and
"spreads" them in time to produce a tibble with distances.
}
\examples{
\donttest{
# Retrieve the set of samples for the Cerrado and Pasture classes
data(cerrado_2classes)
# estimate distances from the data
distances.tb <- sits_distances_normalized(cerrado_2classes)
}
}
\author{
Rolf Simoes, \email{rolf.simoes@inpe.br}

Gilberto Camara, \email{gilberto.camara@inpe.br}
}
