% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_detect_change_methods.R
\name{sits_dtw}
\alias{sits_dtw}
\title{Dynamic Time Warping for Detect changes.}
\usage{
sits_dtw(
  samples = NULL,
  start_date = NULL,
  end_date = NULL,
  window = NULL,
  step = NULL,
  threshold = 2,
  ...
)
}
\arguments{
\item{samples}{Time series with the training samples.}

\item{start_date}{Initial date of the interval used to calculate DTW.}

\item{end_date}{Final date of the interval used to calculate DTW.}

\item{step}{ISO8601-compliant time period used as step for the
DTW moving window, with number and unit,
where "D", "M" and "Y" stands for days, month and
year; e.g., "P16D" for 16 days.}

\item{threshold}{Distance threshold used to define if an event was
detected.}

\item{...}{Parameters to configure the
\code{\link[sits_som]{sits_detect_change_method}}
used to create the temporal patterns of samples.}
}
\value{
Change detection method prepared
                         to be passed to
                         \code{\link[sits]{sits_detect_change_method}}
}
\description{
Create a Dynamic Time Warping (DTW) method for the
\code{\link[sits]{sits_detect_change_method}}.
}
\author{
Felipe Carlos, \email{efelipecarlos@gmail.com}

Felipe Carvalho, \email{felipe.carvalho@inpe.br}

Gilberto Camara, \email{gilberto.camara@inpe.br}

Rolf Simoes, \email{rolf.simoes@inpe.br}
}
