% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_cluster.R
\name{sits_dtwclust}
\alias{sits_dtwclust}
\title{Cluster a set of time series}
\usage{
sits_dtwclust(data.tb, bands = NULL, method = "dendogram", n_clusters = 4,
  min_clu_perc = 0.1, show = TRUE)
}
\arguments{
\item{data.tb}{a tibble the list of time series to be clustered}

\item{bands}{a vector the bands to be clusterized..}

\item{method}{string - either "dendogram" or "centroids"}

\item{n_clusters}{the number of clusters to be identified}

\item{min_clu_perc}{the minimum percentage of members for a cluster to be valid}

\item{show}{(boolean) should the results be shown?}
}
\value{
clusters.tb a SITS tibble with the clusters
}
\description{
This function provides an interface to the DTWCLUST package.
}
\references{
"dtwclust" package (https://CRAN.R-project.org/package=dtwclust)
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}

Rolf Simoes, \email{rolf.simoes@inpe.br}
}
