% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_embedding.R
\name{sits_embedding}
\alias{sits_embedding}
\title{Pre-train deep learning models to generate embeddings for sits}
\usage{
sits_embedding(
  samples_pre_train,
  samples_fine_tune = NULL,
  method = sits_mae(encoder_model = "lighttae", decoder_model = "mlp", masking_method =
    "contiguous", mask_ratio = 0.6, epochs = 150L, batch_size = 128L, validation_split =
    0.2, optimizer_fn = torch::optim_adamw, loss_fn = torch::nnf_mse_loss, device = if
    (torch::cuda_is_available()) torch::torch_device("cuda") else
    torch::torch_device("cpu"), lr = 0.001, verbose = FALSE, seed = 10L)
)
}
\arguments{
\item{method}{A function defining the deep learning pre-training method.}

\item{samples}{Time series samples (as a sits tibble). Labels are optional and
may or may not be used depending on the pre-training method.}
}
\value{
A \code{sits_model} object containing the pre-trained deep model.
}
\description{
Self-supervised pre-training for Earth observation time series.
Allows users to pre-train deep learning models using unlabeled or weakly labeled data.
}
\examples{
if (sits_run_examples()) {
    # Example of MAE pre-training
    mae_model <- sits_pre_train(samples_modis_ndvi, sits_mae_pretrain(mask_ratio = 0.5))
}

}
\author{
Alexandre Assuncao \email{alexcarssuncao@gmail.com}
}
