% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_kohonen.R
\name{sits_evaluate_cluster}
\alias{sits_evaluate_cluster}
\title{Metrics by cluster}
\usage{
sits_evaluate_cluster(data.tb)
}
\arguments{
\item{data.tb}{Tibble containg information about evaluation of samples.}
}
\value{
Returns the confusion matrix and a table with percentage of mixture between the clusters.
}
\description{
This function extracts metrics about the clusters calculating
the percentage of mixture between a cluster and others.
}
\examples{
\donttest{
#' # perform kohonen to associate a sample to  a cluster
koh <- sits_kohonen(samples_mt_6bands, bands = NULL, grid_xdim = 15, grid_ydim = 15)
# Extract metrics about the clusters
confusion_by_cluster <- sits_evaluate_cluster(koh$info_samples)
# Show confusion matrix
confusion_by_cluster$confusion_matrix
}
}
\author{
Lorena Santos, \email{lorena.santos@inpe.br}
}
