% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_kohonen.R
\name{sits_evaluate_samples}
\alias{sits_evaluate_samples}
\title{Evaluate samples}
\usage{
sits_evaluate_samples(data.tb, grid_xdim = 5, grid_ydim = 5,
  rlen = 100, alpha = 1, radius = 6, distance = "euclidean",
  iterations = 1, mode = "online")
}
\arguments{
\item{data.tb}{A sits tibble with info of samples.}

\item{grid_xdim}{X dimension of the SOM grid (default = 5).}

\item{grid_ydim}{Y dimension of the SOM grid.}

\item{rlen}{Number of times the complete data set will be presented to the SOM grid}

\item{alpha}{Learning rate, a vector of two numbers indicating the amount of change.}

\item{radius}{The radius of kohonen map.}

\item{distance}{The similiraty measure.}

\item{iterations}{The number of time to run the SOM cluster.}

\item{mode}{Type of learning algorithm (online or batch).}
}
\value{
Returns a sits tibble with a new column of label and a table with information about the
confiability of each samples.
}
\description{
This function evaluate the samples and extract metrics from SOM cluster.
The user can choose the amount of time that the kohonen process will be run.
This function allows check if the sample reliable or not
}
\examples{
\donttest{
# Read a set of samples
data(samples_mt_9classes)
# Evaluate the quality of each sample
evaluate_samples <- sits_evaluate_samples( data.tb, grid_xdim = 5, grid_ydim = 5, rlen = 100,
distance = "euclidean", iterations = 100)
# Evaluate the confiability of each sample
metrics_by_samples <- evaluate_samples$metrics_by_samples
}
}
\author{
Lorena Santos, \email{lorena.santos@inpe.br}
}
