% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_tibble_utils.R
\name{sits_extract}
\alias{sits_extract}
\title{Extract a subset of the data based on dates}
\usage{
sits_extract(data.tb, start_date = NULL, end_date = NULL)
}
\arguments{
\item{data.tb}{a tibble in SITS format with time series for different bands}

\item{start_date}{the starting date of the time series segment}

\item{end_date}{the end date of the time series segment}
}
\value{
subset.tb  a tibble in SITS format with the chosen subset
}
\description{
returns a vector containing the dates of a sits table
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
