% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_tibble_functions.R
\name{sits_fast_apply}
\alias{sits_fast_apply}
\title{Apply a function over SITS bands.}
\usage{
sits_fast_apply(data.tb, fun, fun_index = function(index) {     return(index)
  }, bands_suffix = "")
}
\arguments{
\item{data.tb}{a valid sits table}

\item{fun}{a function with one parameter as input and a vector or list of vectors as output.}

\item{fun_index}{a function with one parameter as input and a Date vector as output.}

\item{bands_suffix}{a string informing the resulting bands name's suffix.}
}
\value{
data.tb      a sits tibble with same samples and the new bands
}
\description{
Apply a 1D generic function to a time series and specific methods for
`sits_apply` returns a sits tibble with the same samples points and new bands computed by `fun`,
`fun_index` functions. These functions must be defined inline and are called by `sits_apply` for each band,
whose vector values is passed as the function argument.
The `fun` function may either return a vector or a list of vectors. In the first case, the vector will be the new values
of the corresponding band. In the second case, the returned list must have names, and each element vector will
generate a new band which name composed by concatenating original band name and the corresponding list element name.

If a suffix is provided in `bands_suffix`, all resulting bands names will end with provided suffix separated by a ".".

Differently from `sits_apply` function, this function merge all time series before applying `fun` function.
Functions that depend on time sequence must use `sits_apply` function.
}
\author{
Rolf Simoes, \email{rolf.simoes@inpe.br}
}
