% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_feature_selection.R
\name{sits_feature_selection}
\alias{sits_feature_selection}
\title{Select the most representative bands for spatiotemporal
classification}
\usage{
sits_feature_selection(
  samples,
  ml_method = sits_rfor(),
  impute_fn = impute_linear(),
  filter_fn = NULL,
  folds = 5L,
  metric = "mda",
  n_iter = 5,
  acc_loss = 0.02,
  multicores = 2L
)
}
\arguments{
\item{samples}{Time series with the training samples
(tibble of class "sits").}

\item{ml_method}{Machine learning method (default = RandomForest)}

\item{impute_fn}{Imputation function to remove NA.}

\item{filter_fn}{Smoothing filter to be applied - optional}

\item{folds}{Number of partitions to create.}

\item{metric}{Importance metric. Use \code{"mda"} for mean decrease
accuracy and \code{"mdg"} for mean decrease gini.}

\item{n_iter}{Number of iterations for feature selection.}

\item{acc_loss}{Number of maximum accuracy loss accepted (0 to 1).}

\item{multicores}{Number of cores to be used for cross-validation.
(integer, min = 1, max = folds).}
}
\value{
Samples with selected bands
}
\description{
Use recursive feature elimination for selecting the
most representative bands for each iteration. For each iteration,
one band is removed if the accuracy loss is lower than the threshold.
The feature selection stops in two ways: (1) when the accuracy loss
reaches the threshold value; (2) or when it reaches the maximum
iteration number.
}
\examples{
if (sits_run_examples()) {
 samples_selected <- sits_feature_selection(
     samples    = samples_l8_rondonia_2bands,
     ml_method  = sits_rfor(),
     n_iter     = 1,
     acc_loss   = 0.01,
     multicores = 2L
 )
}

}
\author{
Felipe Carvalho, \email{lipecaso@gmail.com}

Felipe Carlos, \email{efelipecarlos@gmail.com}

Rolf Simoes, \email{rolfsimoes@gmail.com}

Gilberto Camara, \email{gilberto.camara@inpe.br}
}
