% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_table.R
\name{sits_foreach}
\alias{sits_foreach}
\title{apply a function to a grouped SITS table}
\usage{
sits_foreach(data.tb, ..., fun)
}
\arguments{
\item{data.tb}{a sits table with the time series of the selected bands}

\item{...}{one or more sits table field separated by commas that are used to group the data.
See `dplyr::group_by` help for more details.}

\item{fun}{a function that receives as input an sits table and outputs an sits table}
}
\value{
result.tb   a tibble in SITS format with the selected bands
}
\description{
returns a sits table by compound the sits tables apply a function to a grouped SITS table
}
\author{
Rolf Simoes, \email{rolf.simoes@inpe.br}
}
