% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_machine_learning.R
\name{sits_formula_logref}
\alias{sits_formula_logref}
\title{Define a loglinear formula for classification models}
\usage{
sits_formula_logref(predictors_index = -2L:0L)
}
\arguments{
\item{predictors_index}{Index of the valid columns
to compose formula (default: -2:0).}
}
\value{
A function that computes a valid formula using a log function.
}
\description{
A function to be used as a symbolic description
of some fitting models such as svm and random forest.
This function tells the models to do a log transformation of the inputs.
The `predictors_index` parameter informs
the positions of `tb` fields corresponding to formula independent variables.
If no value is given, the default is NULL,
a value indicating that all fields will be used as predictors.
}
\examples{
if (sits_run_examples()) {
    # Example of training a model for time series classification
    # Retrieve the samples for Mato Grosso
    # train an SVM model
    ml_model <- sits_train(samples_modis_ndvi,
        ml_method = sits_svm(formula = sits_formula_logref())
    )
    # classify the point
    point_ndvi <- sits_select(point_mt_6bands, bands = "NDVI")
    # classify the point
    point_class <- sits_classify(
        data = point_ndvi, ml_model = ml_model
    )
    plot(point_class)
}
}
\seealso{
Other machine learning:
\code{\link[=sits_formula_linear]{sits_formula_linear()}},
\code{\link[=sits_lightgbm]{sits_lightgbm()}},
\code{\link[=sits_lighttae]{sits_lighttae()}},
\code{\link[=sits_lstm_fcn]{sits_lstm_fcn()}},
\code{\link[=sits_mlp]{sits_mlp()}},
\code{\link[=sits_resnet]{sits_resnet()}},
\code{\link[=sits_rfor]{sits_rfor()}},
\code{\link[=sits_svm]{sits_svm()}},
\code{\link[=sits_tae]{sits_tae()}},
\code{\link[=sits_tempcnn]{sits_tempcnn()}},
\code{\link[=sits_xgboost]{sits_xgboost()}}
}
\author{
Alexandre Ywata de Carvalho, \email{alexandre.ywata@ipea.gov.br}

Rolf Simoes, \email{rolfsimoes@gmail.com}
}
\concept{machine learning}
