% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_zoo.R
\name{sits_from_matrix}
\alias{sits_from_matrix}
\title{Import a matrix to a sits tibble}
\usage{
sits_from_matrix(values, raster_obj, timeline, initial_row, nrows, band)
}
\arguments{
\item{values}{matrix of values extracted from a raster object}

\item{raster_obj}{raster object from which the matrix of values have been extracted}

\item{timeline}{timeline associated to the data}

\item{initial_row}{initial row in the raster object from which the matrix has been extracted}

\item{nrows}{number of rows read in the raster object}

\item{band}{spectral band or index associated to the data}
}
\description{
Import a matrix to a sits tibble
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
\keyword{internal}
