% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_getdata.R
\name{sits_from_service}
\alias{sits_from_service}
\title{Obtain timeSeries from time series service}
\usage{
sits_from_service(service = "WTSS", latitude = NULL, longitude = NULL,
  start_date = NULL, end_date = NULL, coverage = "mod13q1_512",
  bands = NULL, satellite = "terra", prefilter = "1", label = "NoClass")
}
\arguments{
\item{service}{string - name of the time series service (options are "WTSS" or "SATVEG")}

\item{latitude}{double - the latitude of the chosen location)}

\item{longitude}{double - the longitude of the chosen location}

\item{start_date}{(optional) date - the start of the period}

\item{end_date}{(optional) date - the end of the period}

\item{coverage}{name of the coverage (required for WTSS)}

\item{bands}{(optional) string vector - the names of the bands to be retrieved}

\item{satellite}{(optional) - the same of the satellite (options - "terra", "aqua", "comb")}

\item{prefilter}{string ("0" - none, "1" - no data correction, "2" - cloud correction, "3" - no data and cloud correction)}

\item{label}{string - the label to attach to the time series}
}
\value{
table          a SITS tibble
}
\description{
obtains a time series from a time series service
}
\examples{
\donttest{
# define a shapefile and read from the points inside it from the WTSS service
shp_file <- system.file("extdata/shapefiles/anhanguera/anhanguera.shp", package = "sits")
munic.tb <- sits_fromSHP(shp_file)
}
}
