% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_zoo.R
\name{sits_from_zoo}
\alias{sits_from_zoo}
\title{Import time series in the zoo format to a sits tibble}
\usage{
sits_from_zoo(
  ts_zoo,
  longitude = 0,
  latitude = 0,
  label = "NoClass",
  name = "unknown"
)
}
\arguments{
\item{ts_zoo}{Zoo time series.}

\item{longitude}{Longitude of the chosen location.}

\item{latitude}{Latitude of the chosen location.}

\item{label}{Label to attach to the time series (optional).}

\item{name}{Name of the data cube where data comes from.}
}
\value{
Time series in sits tibble format.
}
\description{
Converts data from an instance of a zoo series to a sits tibble.
}
\examples{
# Read a time series in ZOO format
data(ts_zoo)
# Convert the zoo series into a sits tibble
data <- sits_from_zoo(ts_zoo,
    longitude = -54.2313, latitude = -14.0482,
    label = "Cerrado", name = "mod13q1"
)

}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
