% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_fromClusters.R
\name{sits_fromClusters}
\alias{sits_fromClusters}
\title{Returns a list of centroids and its metadata.}
\usage{
sits_fromClusters(data.tb, clusters.lst)
}
\arguments{
\item{data.tb}{a tibble with input data of dtwclust.}

\item{clusters.lst}{a list of cluster structure returned from dtwclust. Each element correspond to a band attribute.}
}
\value{
table     tibble  - a SITS table
}
\description{
\code{sits_fromClusters} reads a set of centroids from clusters output.
(See: sits_dendogram(...))
}
\details{
A sits table has the metadata and data for each time series centroids
<longitude, latitude, start_date, end_date, label, coverage, time_series>
}
\examples{
sits_fromClusters (savanna_s.tb, clusters.lst)
}
\seealso{
Other SITS main functions: \code{\link{sits_fromCSV}},
  \code{\link{sits_fromSHP}}, \code{\link{sits_fromWTSS}},
  \code{\link{sits_merge}}
}
\keyword{SITS}
