% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_CSV.R
\name{sits_fromCSV}
\alias{sits_fromCSV}
\title{Obtain timeSeries from WTSS server, based on a CSV file.}
\usage{
sits_fromCSV(csv_file, n_max = Inf)
}
\arguments{
\item{csv_file}{string  - name of a CSV file with information <id, latitude, longitude, from, end, label>}

\item{n_max}{integer - the maximum number of samples to be read}
}
\value{
table     the SITS table

an updated version of the global database
}
\description{
\code{sits_fromCSV} returns a list of WTSS time series from a sample data frame and
stores them in a global table for later use
A sits table has the metadata and data for each time series
<longitude, latitude, from, to, lable, time_series>
}
\details{
The CSV file should have the following column names:
"longitude", "latitude", "from", "to", "label"
Longitude/latitude data is mandatory, the other parameters are optional
}
\examples{
sits_fromCSV ("mysamples.csv", table = "samples.db")
}
\seealso{
Other SITS auxiliary functions: \code{\link{sits_fromWTSS}},
  \code{\link{sits_ggplot_series}},
  \code{\link{sits_ggplot_together}},
  \code{\link{sits_make_plot_title}},
  \code{\link{sits_merge}}
}
\keyword{SITS}
