% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_JSON.R
\name{sits_fromJSON}
\alias{sits_fromJSON}
\title{Retrieve a set of time series from a JSON file}
\usage{
sits_fromJSON(json_file)
}
\arguments{
\item{json_file}{name of the JSON file}

\item{table}{a string with the name of a global sits table (created if not existent)}
}
\value{
an updated in-memory database
}
\description{
\code{sits_from_JSON} gets a SITS database from a JSON file
}
\details{
The set of time series from the SITS database can be saved in JSON format
and later retrieved for further use
}
\examples{
sits_fromJSON ("myjson_file.json", table = "my.tb")
}
\seealso{
Other STIS: \code{\link{sits_toJSON}}
}
\keyword{STIS}
