% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_import.R
\name{sits_fromLatLong}
\alias{sits_fromLatLong}
\title{Obtain timeSeries from WTSS server based on a lat/long information.}
\usage{
sits_fromLatLong(longitude, latitude, start_date = NULL, end_date = NULL,
  wtss.obj, cov, bands)
}
\arguments{
\item{longitude}{double - the longitude of the chosen location}

\item{latitude}{double - the latitude of the chosen location)}

\item{start_date}{date - the start of the period}

\item{end_date}{date - the end of the period}

\item{wtss.obj}{an R object that represents the WTSS server}

\item{cov}{a list with the coverage parameters (retrived from the WTSS server)}

\item{bands}{string vector - the names of the bands to be retrieved}
}
\value{
data.tb       tibble  - a SITS table
}
\description{
This function uses the lat/long location to retrive a time seriees
for a WTSS service. A Web Time Series Service is a light-weight service that
retrieves one or more time series in JSON format from a data base.
}
\references{
Lubia Vinhas, Gilberto Queiroz, Karine Ferreira, Gilberto Camara,
Web Services for Big Earth Observation Data.
In: XVII Brazilian Symposium on Geoinformatics, 2016, Campos do Jordao.
Proceedings of GeoInfo 2016. Sao Jose dos Campos: INPE/SBC, 2016. v.1. p.166-177.
}
