% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_fromRData.R
\name{sits_fromRData}
\alias{sits_fromRData}
\title{Transform patterns saved in .RData to a SITS table
\code{sits_fromRData} reads a set of TWDTW patterns saved as RData and
transforms them into a SITS table}
\usage{
sits_fromRData(patterns)
}
\arguments{
\item{patterns}{- a TWDTW object containing a set of patterns to be used for classification}
}
\value{
sits.tb  - a SITS table containing the patterns
}
\description{
Transform patterns saved in .RData to a SITS table
\code{sits_fromRData} reads a set of TWDTW patterns saved as RData and
transforms them into a SITS table
}

