% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_SATVEG.R
\name{sits_fromSATVEG}
\alias{sits_fromSATVEG}
\title{Obtain one timeSeries from the EMBRAPA SATVEG server and load it on a sits tibble}
\usage{
sits_fromSATVEG(longitude = NULL, latitude = NULL, start_date = NULL,
  end_date = NULL, coverage = "terra", prefilter = "1",
  label = "NoClass")
}
\arguments{
\item{longitude}{double - the longitude of the chosen location}

\item{latitude}{double - the latitude of the chosen location}

\item{start_date}{date - the start of the period}

\item{end_date}{date - the end of the period}

\item{coverage}{the coverage ("terra", "aqua", "comb")}

\item{prefilter}{string ("0" - none, "1" - no data correction, "2" - cloud correction, "3" - no data and cloud correction)}

\item{label}{string - the label to attach to the time series (optional)}
}
\value{
data.tb        a SITS tibble
}
\description{
Returns one set of MODIS time series provided by the EMBRAPA server (SATVEG)
Given a location (lat/long), the function retrieves the "ndvi" and "evi" bands from SATVEG
and inclues the data on a stis tibble. If start and end date are given, the function
filter the data to limit the temporal interval.
}
\examples{
\donttest{
# Read a single lat long point from the SATVEG server
point.tb <- sits_fromSATVEG (longitude = -55.50563, latitude = -11.71557)
}
}
\author{
Julio Esquerdo, \email{julio.esquerdo@embrapa.br}

Gilberto Camara, \email{gilberto.camara@inpe.br}
}
