% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_getdata.R
\name{sits_fromtable}
\alias{sits_fromtable}
\title{Obtain timeSeries from WTSS server, based on a SITS table.}
\usage{
sits_fromtable(table, wtss.obj, cov, bands)
}
\arguments{
\item{table}{a  sits_table}

\item{wtss.obj}{an R object that represents the WTSS server}

\item{cov}{a list with the coverage parameters (retrived from the WTSS server)}

\item{bands}{string vector - the names of the bands to be retrieved}
}
\value{
data.tb       tibble  - a SITS table
}
\description{
reads descriptive information about a set of
spatio-temporal locations from a SITS table. Then it uses the WTSS service to
obtain the required data. This function is useful when you have a sits table
but you want to get the time series from a different set of bands.
}
