% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_fromTWDTW.R
\name{sits_fromTWDTW}
\alias{sits_fromTWDTW}
\title{Transform patterns from .RData to JSON}
\usage{
sits_fromTWDTW(patterns, coverage = cov_name.gl)
}
\description{
Transform patterns from .RData to JSON
}

