% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_import.R
\name{sits_fromTWDTW_matches}
\alias{sits_fromTWDTW_matches}
\title{Transform patterns from TWDTW format to SITS format}
\usage{
sits_fromTWDTW_matches(match.twdtw)
}
\arguments{
\item{match.twdtw}{a TWDTW Matches object of class dtwSat::twdtwMatches (S4)}
}
\value{
result.tb    a tibble containing the matches information
}
\description{
reads one TWDTW matches object and transforms it into a tibble ready to be stored into a SITS table column.
}
