% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_import.R
\name{sits_fromTWDTW_timeseries}
\alias{sits_fromTWDTW_timeseries}
\title{Transform patterns from TWDTW format to SITS format}
\usage{
sits_fromTWDTW_timeseries(patterns, coverage)
}
\arguments{
\item{patterns}{- a TWDTW object containing a set of patterns to be used for classification}

\item{coverage}{- the name of the coverage from where the time series have been obtained}
}
\value{
sits.tb  - a SITS table containing the patterns
}
\description{
reads a set of TWDTW patterns and transforms them into a SITS table
}
