% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_WTSS.R
\name{sits_fromWTSS}
\alias{sits_fromWTSS}
\title{Obtain one timeSeries from WTSS server and load it on a sits tibble}
\usage{
sits_fromWTSS(product = "MOD13Q1", coverage = "mod13q1_512", longitude,
  latitude, start_date = NULL, end_date = NULL, bands = NULL,
  label = "NoClass")
}
\arguments{
\item{product}{the image product where the coverage depends}

\item{coverage}{name of the coverage}

\item{longitude}{double - the longitude of the chosen location}

\item{latitude}{double - the latitude of the chosen location}

\item{start_date}{date - the start of the period}

\item{end_date}{date - the end of the period}

\item{bands}{list of string - a list of the names of the bands of the coverage}

\item{label}{string - the label to attach to the time series (optional)}
}
\value{
data.tb        a SITS tibble
}
\description{
Returns one set of time series provided by a WTSS server
Given a location (lat/long), and start/end period, and the WTSS server information
retrieve a time series and include it on a stis tibble.
A Web Time Series Service (WTSS) is a light-weight service that
retrieves one or more time series in JSON format from a data base.
}
\examples{
\donttest{
# Read a single lat long point from a WTSS server
point.tb <- sits_fromWTSS (longitude = -55.50563, latitude = -11.71557)
}
}
\references{
Lubia Vinhas, Gilberto Queiroz, Karine Ferreira, Gilberto Camara,
Web Services for Big Earth Observation Data.
In: XVII Brazilian Symposium on Geoinformatics, 2016, Campos do Jordao.
Proceedings of GeoInfo 2016. Sao Jose dos Campos: INPE/SBC, 2016. v.1. p.166-177
}
