% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_zoo.R
\name{sits_fromZOO}
\alias{sits_fromZOO}
\title{Import time series in the zoo format to a SITS tibble}
\usage{
sits_fromZOO(ts.zoo, longitude = 0, latitude = 0, label = "NoClass",
  name = "unknown")
}
\arguments{
\item{ts.zoo}{zoo time series}

\item{longitude}{longitude of the chosen location}

\item{latitude}{latitude of the chosen location}

\item{label}{label to attach to the time series (optional)}

\item{name}{name of the coverage where data comes from}
}
\value{
data.tb      time series in SITS tibble format
}
\description{
Converts data from an instance of a zoo series to a SITS tibble
}
\examples{
# Read a time series in ZOO format
data(ts_zoo)
# Convert the zoo series into a SITS tibble
data.tb <- sits_fromZOO (ts_zoo, longitude = -54.2313, latitude = -14.0482,
           label = "Cerrado", name = "mod13q1")
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
