% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_machine_learning.R
\name{sits_gbm}
\alias{sits_gbm}
\title{Train SITS classifiction models with Gradient Boosting Machine}
\usage{
sits_gbm(distances.tb = NULL, formula = sits_formula_logref(),
  distribution = "multinomial", n.trees = 500, interaction.depth = 4,
  shrinkage = 0.001, cv.folds = 3, n.cores = 1, ...)
}
\arguments{
\item{distances.tb}{a time series with a set of distance measures for each training sample}

\item{formula}{a symbolic description of the model to be fit. SITS offers a set of such formulas (default: sits_formula_logref)}

\item{distribution}{the name of the distribution. Either "multinomial" for classification)}

\item{n.trees}{Number of trees to fit. This should not be set to too small a number,
to ensure that every input row gets predicted at least a few times. (default: 500)}

\item{interaction.depth}{The maximum depth of variable interactions. 1 implies an additive model, 2 implies a model with up to 2-way interactions.}

\item{shrinkage}{a shrinkage parameter applied to each tree in the expansion.
Also known as the learning rate or step-size reduction.}

\item{cv.folds}{number of cross-validations to run}

\item{n.cores}{number of cores to run}

\item{...}{other parameters to be passed to `gbm::gbm` function}
}
\value{
result          either an model function to be passed in sits_predict or an function prepared that can be called further to compute multinom training model
}
\description{
Use generalized liner model (glm) via penalized maximim likelihood to classify data.
This function is a front-end to the "cv.glmnet" method in the "glmnet" package.
Please refer to the documentation in that package for more details.
}
\author{
Alexandre Xavier Ywata de Carvalho, \email{alexandre.ywata@ipea.gov.br}

Rolf Simoes, \email{rolf.simoes@inpe.br}
}
