% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_gdalcubes.R
\name{sits_gdalcubes_aggregation}
\alias{sits_gdalcubes_aggregation}
\title{Save the images based on an aggregation method.}
\usage{
sits_gdalcubes_aggregation(gdalcubes_list, path_images, ...)
}
\arguments{
\item{gdalcubes_list}{a \code{gdalcubes_list} object returned by
sits_gdalcubes_raster function.}

\item{path_images}{a \code{character} with the path where the aggregated
images will be writed.}

\item{...}{Additional parameters that can be included. See
'?gdalcubes::write_tif'.}
}
\value{
an invisible return of "gdalcubes_list" object.
}
\description{
Based on the defined parameters in 'sits_gdalcubes_raster'
 function, this function materializes the cubes created in the .tif format.
}
\examples{
\dontrun{
# this example requires access to an external service, so should not be run
# by CRAN

# s3://sentinel-cogs/sentinel-s2-l2a-cogs/2017/S2A_35MNR_20171025_0_L2A/

# Provide your AWS credentials here
# Sys.setenv(
# "AWS_ACCESS_KEY_ID"     = <your_access_key>,
# "AWS_SECRET_ACCESS_KEY" = <your_secret_access_key>,
# "AWS_DEFAULT_REGION"    = <your AWS region>,
# "AWS_ENDPOINT" = "sentinel-s2-l2a.s3.amazonaws.com",
# "AWS_REQUEST_PAYER"     = "requester"
# )

s2_cube <- sits_cube(
    type = "S2_L2A_AWS",
    name = "T20LKP_2018_2019",
    satellite = "SENTINEL-2",
    sensor = "MSI",
    tiles = "20LKP",
    s2_aws_resolution = "20m",
    start_date = as.Date("2018-07-18"),
    end_date = as.Date("2018-09-18")
)

r_cube <- sits_gdalcubes_raster(s2_cube, path_db = "my/path/cube.db",
                                period     = "P1M",
                                method     = "median",
                                resampling = "bilinear")

sits_gdalcubes_aggregation(r_cube, "my/path/images/")
}

}
