% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_gdalcubes.R
\name{sits_gdalcubes_raster}
\alias{sits_gdalcubes_raster}
\title{Create a list of a gdal_cubes raster.}
\usage{
sits_gdalcubes_raster(
  cube,
  path_db,
  period = NULL,
  method = NULL,
  resampling = "bilinear"
)
}
\arguments{
\item{cube}{Data cube from where data is to be retrieved.}

\item{path_db}{a \code{character} with the path and name where the
database will be create. E.g. "my/path/gdalcubes.db"}

\item{period}{a \code{character} with the The period of time in which it
is desired to apply in the cube, must be provided based on ISO8601, where 1
number and a unit are provided, for example "P16D".}

\item{method}{a \code{character} with the method that will be applied in
the aggregation, the following are available: "min", "max", "mean",
"median" or "first".}

\item{resampling}{a \code{character} with the method that will be applied
in the resampling in mosaic operation. The following are available: "near",
"bilinear", "bicubic" or others supported by gdalwarp
(see https://gdal.org/programs/gdalwarp.html).}

\item{...}{Aditional parameters that can be included. See
'?gdalcubes::cube_view'.}
}
\value{
a \code{gdalcubes_list} object where each index has a 'raster_cube'
 for each tile.
}
\description{
Creates a list of cubes based on the object provided by the
 package 'gdalcubes' to be later aggregated or applied any operations.
}
\examples{
\dontrun{
# this example requires access to an external service, so should not be run
# by CRAN

# s3://sentinel-cogs/sentinel-s2-l2a-cogs/2017/S2A_35MNR_20171025_0_L2A/

# Provide your AWS credentials here
# Sys.setenv(
# "AWS_ACCESS_KEY_ID"     = <your_access_key>,
# "AWS_SECRET_ACCESS_KEY" = <your_secret_access_key>,
# "AWS_DEFAULT_REGION"    = <your AWS region>,
# "AWS_ENDPOINT" = "sentinel-s2-l2a.s3.amazonaws.com",
# "AWS_REQUEST_PAYER"     = "requester"
# )

s2_cube <- sits_cube(
    type = "S2_L2A_AWS",
    name = "T20LKP_2018_2019",
    satellite = "SENTINEL-2",
    sensor = "MSI",
    tiles = "20LKP",
    s2_aws_resolution = "20m",
    start_date = as.Date("2018-07-18"),
    end_date = as.Date("2018-09-18")
)

r_cube <- sits_gdalcubes_raster(s2_cube, path_db = "my/path/cube.db",
                                period     = "P1M",
                                method     = "median",
                                resampling = "bilinear")

}

}
\references{
`gdalcubes` package (https://github.com/appelmar/gdalcubes_R)
}
