% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_generics.R
\name{sits_generic}
\alias{sits_generic}
\alias{.crs}
\alias{.crs<-}
\alias{.xmin}
\alias{.xmin<-}
\alias{.xmax}
\alias{.xmax<-}
\alias{.ymin}
\alias{.ymin<-}
\alias{.ymax}
\alias{.ymax<-}
\alias{.nrows}
\alias{.nrows<-}
\alias{.ncols}
\alias{.ncols<-}
\alias{.xres}
\alias{.xres<-}
\alias{.yres}
\alias{.yres<-}
\alias{.lon}
\alias{.lon<-}
\alias{.lat}
\alias{.lat<-}
\alias{.start_date}
\alias{.start_date<-}
\alias{.end_date}
\alias{.end_date<-}
\alias{.first_row}
\alias{.first_row<-}
\alias{.first_col}
\alias{.first_col<-}
\title{generic accessors}
\usage{
.crs(x)

.crs(x) <- value

.xmin(x)

.xmin(x) <- value

.xmax(x)

.xmax(x) <- value

.ymin(x)

.ymin(x) <- value

.ymax(x)

.ymax(x) <- value

.nrows(x)

.nrows(x) <- value

.ncols(x)

.ncols(x) <- value

.xres(x)

.xres(x) <- value

.yres(x)

.yres(x) <- value

.lon(x)

.lon(x) <- value

.lat(x)

.lat(x) <- value

.start_date(x)

.start_date(x) <- value

.end_date(x)

.end_date(x) <- value

.first_row(x)

.first_row(x) <- value

.first_col(x)

.first_col(x) <- value
}
\arguments{
\item{x}{any R object from which value will be set or extracted.}

\item{value}{any value to set in the \code{x} object.}
}
\description{
Internal accessors to be used instead of standard ones (`[[` or \code{$}).

This approach has advantage to check the data validity before return
value or before set value.

\code{.crs}: access \code{x[["crs"]]} value

\code{.crs}: set \code{x[["crs"]] <- value}

\code{.xmin}: access \code{x[["xmin"]]} value

\code{.xmin}: access \code{x[["xmin"]]} value

\code{.xmax}: access \code{x[["xmax"]]} value

\code{.xmax<-()}: set \code{x[["xmax"]] <- value}

\code{.ymin}: access \code{x[["ymin"]]} value

\code{.ymin<-()}: set \code{x[["ymin"]] <- value}

\code{.ymax}: access \code{x[["ymax"]]} value

\code{.ymax<-()}: set \code{x[["ymax"]] <- value}

\code{.nrows}: access \code{x[["nrows"]]} value

\code{.nrows<-()}: set \code{x[["nrows"]] <- value}

\code{.ncols}: access \code{x[["ncols"]]} value

\code{.ncols<-()}: set \code{x[["ncols"]] <- value}

\code{.xres}: access \code{x[["xres"]]} value

\code{.xres<-()}: set \code{x[["xres"]] <- value}

\code{.yres}: access \code{x[["yres"]]} value

\code{.yres<-()}: set \code{x[["yres"]] <- value}

\code{.lon}: access \code{x[["lon"]]} value

\code{.lon<-()}: set \code{x[["lon"]] <- value}

\code{.lat}: access \code{x[["lat"]]} value

\code{.lat<-()}: set \code{x[["lat"]] <- value}

\code{.start_date}: access \code{x[["start_date"]]} value

\code{.start_date<-()}: set \code{x[["start_date"]] <- value}

\code{.end_date}: access \code{x[["end_date"]]} value

\code{.end_date<-()}: set \code{x[["end_date"]] <- value}

\code{.first_row}: access \code{x[["first_row"]]} value

\code{.first_row<-()}: set \code{x[["first_row"]] <- value}

\code{.first_col}: access \code{x[["first_col"]]} value

\code{.first_col<-()}: set \code{x[["first_col"]] <- value}
}
\keyword{internal}
