% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_config.R
\name{sits_get_adjustment_shift}
\alias{sits_get_adjustment_shift}
\alias{.sits_get_adjustment_shift}
\title{Retrieve the value of the adjustment shift}
\usage{
.sits_get_adjustment_shift()
}
\description{
Retrieves the value of the shift to adjust entries to have only positive values.
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
