% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_get_data.R
\name{sits_get_data.csv_raster_cube}
\alias{sits_get_data.csv_raster_cube}
\title{Obtain time series from brick based on CSV file}
\usage{
\method{sits_get_data}{csv_raster_cube}(cube, file, ..., bands = NULL, impute_fn = sits_impute_linear())
}
\arguments{
\item{cube}{Data cube from where data is to be retrived.}

\item{file}{File with information on the data to be retrieved}

\item{...}{Other parameters to be passed for specific types}

\item{bands}{Bands to be retrieved (optional)}

\item{impute_fn}{Imputation function for NA values}
}
\value{
A tibble with time series data and metadata.
}
\description{
Obtain time series from brick based on CSV file
}
\examples{
#' Read a CSV in a Raster Brick
# define the file that has the raster brick
files  <- c(system.file ("extdata/raster/mod13q1/sinop-crop-ndvi.tif",
                         package = "sits"))
# define the timeline
data(timeline_modis_392)
# create a data cube based on the information about the files
raster_cube <- sits_cube(type = "RASTER", satellite = "TERRA",
                         sensor = "MODIS", name = "Sinop-crop",
                         timeline = timeline_modis_392,
                         bands = c("NDVI"), files = files)
# read data from a CSV file
csv_file <- system.file ("extdata/samples/samples_sinop_crop.csv",
                         package = "sits")
points.tb <- sits_get_data (raster_cube, file = csv_file)

# show the points
plot(points.tb)

}
