% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_get_data.R
\name{sits_get_data.csv_satveg_cube}
\alias{sits_get_data.csv_satveg_cube}
\title{Obtain time series from SATVEG}
\usage{
\method{sits_get_data}{csv_satveg_cube}(cube, file, ...)
}
\arguments{
\item{cube}{Data cube from where data is to be retrived.}

\item{file}{CSV File with information on the data to be retrieved}

\item{...}{Other parameters to be passed for specific types}
}
\value{
A tibble with time series data and metadata.
}
\description{
Obtain time series from SATVEG
}
