% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_get_data.R
\name{sits_get_data.csv_wtss_cube}
\alias{sits_get_data.csv_wtss_cube}
\title{Obtain time series from wtss based on CSV file}
\usage{
\method{sits_get_data}{csv_wtss_cube}(cube, file, ..., bands = NULL)
}
\arguments{
\item{cube}{Data cube from where data is to be retrived.}

\item{file}{CSV File with information on the data to be retrieved}

\item{...}{Other parameters to be passed for specific types}

\item{bands}{Bands to be retrieved (optional)}
}
\value{
A tibble with time series data and metadata.
}
\description{
Obtain time series from wtss based on CSV file
}
\examples{
\dontrun{
# Read a single lat long point from a WTSS server
# Requires access to external service

wtss_cube <- sits_cube(
    type = "WTSS",
    URL = "http://www.esensing.dpi.inpe.br/wtss/",
    name = "MOD13Q1"
)

# Read a set of points defined in a CSV file from a WTSS server
csv_file <- system.file("extdata/samples/samples_matogrosso.csv",
    package = "sits"
)
points.tb <- sits_get_data(wtss_cube, file = csv_file)
# show the points retrieved for the WTSS server
plot(points.tb[1:3, ])
}

}
