% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_get_data.R
\name{sits_get_data.satveg_cube}
\alias{sits_get_data.satveg_cube}
\title{Obtain time series from satveg}
\usage{
\method{sits_get_data}{satveg_cube}(
  cube,
  file = NULL,
  ...,
  longitude = NULL,
  latitude = NULL,
  start_date = NULL,
  end_date = NULL,
  label = "NoClass"
)
}
\arguments{
\item{cube}{Data cube from where data is to be retrieved.}

\item{file}{File with information on the data to be retrieved.}

\item{...}{Other parameters to be passed for specific types.}

\item{longitude}{Longitude of the chosen location.}

\item{latitude}{Latitude of the chosen location.}

\item{start_date}{Start of the interval for the time series
in "YYYY-MM-DD" format (optional).}

\item{end_date}{End of the interval for the time series in
"YYYY-MM-DD" format (optional).}

\item{label}{Label to be assigned to the time series (optional).}
}
\value{
A tibble with time series data and metadata.
}
\description{
Obtain time series from satveg
}
\examples{
\dontrun{
cube_terra <- sits_cube(type = "SATVEG", name = "terra")
point_terra <- sits_get_data(cube_terra,
    longitude = -55.50563, latitude = -11.71557
)
plot(point_terra)
}
}
