% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_get_data.R
\name{sits_get_data.stack_cube}
\alias{sits_get_data.stack_cube}
\title{Obtain time series from stack cube}
\usage{
\method{sits_get_data}{stack_cube}(
  cube,
  file,
  ...,
  longitude = NULL,
  latitude = NULL,
  start_date = NULL,
  end_date = NULL,
  bands = NULL,
  label = "NoClass"
)
}
\arguments{
\item{cube}{Data cube from where data is to be retrived..}

\item{file}{File with information on the data to be retrieved}

\item{...}{Other parameters to be passed for specific types}

\item{longitude}{Longitude of the chosen location.}

\item{latitude}{Latitude of the chosen location.}

\item{start_date}{Start of the interval for the time series
in "YYYY-MM-DD" format (optional)}

\item{end_date}{End of the interval for the time series in
"YYYY-MM-DD" format (optional).}

\item{bands}{Bands to be retrieved (optional)}

\item{label}{Label to be assigned to the time series (optional)}
}
\value{
A tibble with time series data and metadata.
}
\description{
Obtain time series from stack cube
}
