% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_metrics.R
\name{sits_get_metrics}
\alias{sits_get_metrics}
\alias{sits_get_metrics.sits}
\alias{.extract_metric}
\title{...}
\usage{
sits_get_metrics(data, metrics, ...)

\method{sits_get_metrics}{sits}(data, metrics, ...)

.extract_metric(ts, metrics)
}
\arguments{
\item{data}{documentar}

\item{metrics}{documentar}

\item{...}{documentar}

\item{ts}{documentar}
}
\value{
documentar

documentar
}
\description{
materialize metrics in sits_tibble and sits_cube
}
\keyword{internal}
