% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_raster_utils.R
\name{sits_get_raster}
\alias{sits_get_raster}
\title{Get a raster object from a raster coverage}
\usage{
sits_get_raster(raster.tb, i = NULL)
}
\arguments{
\item{raster.tb}{Raster coverage}

\item{i}{i-th element of the list to retrieve}
}
