% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_raster_classification.R
\name{sits_get_robj}
\alias{sits_get_robj}
\title{Get a raster object from a raster classified coverage}
\usage{
sits_get_robj(raster.tb, i)
}
\arguments{
\item{raster.tb}{raster coverage (classified)}

\item{i}{i-th element of the list to retrieve}
}
\value{
r_obj     a raster layer with classification
}
\description{
This function retrieves one or more raster layer objects stored in a raster coverage.
             It should be used to ensure that the raster objects are returned correctly.
}
\examples{
\donttest{
# Define a raster Brick and retrieve the associated object
# define the file that has the raster brick
files  <- c(system.file ("extdata/raster/mod13q1/sinop-crop-ndvi.tif", package = "sits"))
# define the timeline
data(timeline_modis_392)
# create a raster metadata file based on the information about the files
raster_cov <- sits_coverage(files = files, name = "Sinop-crop",
                            timeline = timeline_modis_392, bands = c("ndvi"))

# plot the first raster object with a selected color pallete
# get the first classified object
r.obj <- sits_get_robj(raster_class.tb,1)

# make a title, define the colors and the labels)
title <- paste0("Classified image of part of SINOP-MT - 2000/2001")
colors <- c("#65AF72", "#d4d6ed", "#006400","#add8e6","#a0522d",
            "#a52a2a","#d2b48c", "#cd853f", "#ff8c00")
labels <- sits_labels(samples_MT_ndvi)$label

sits_plot_raster(r.obj, title, labels, colors)
}

}
