% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_ggplot_together.R
\name{sits_ggplot_together}
\alias{sits_ggplot_together}
\title{Plot many timeSeries together using ggplot}
\usage{
sits_ggplot_together(melted.tb, means.tb, plot_title)
}
\arguments{
\item{melted.tb}{a tibble with the time series (a lot of data already melted)}

\item{means.tb}{a tibble with the means and std deviations of the time series}
}
\description{
\code{sits_ggplot_together} plots a set of  time series together
}
\seealso{
Other SITS auxiliary functions: \code{\link{sits_ggplot_series}},
  \code{\link{sits_plot_title}}
}
\keyword{SITS}

