% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_glcm.R
\name{sits_glcm}
\alias{sits_glcm}
\alias{sits_glcm.raster_cube}
\alias{sits_glcm.derived_cube}
\alias{sits_glcm.default}
\title{Apply a GLCM metric on a data cube}
\usage{
sits_glcm(data, ...)

\method{sits_glcm}{raster_cube}(
  data,
  ...,
  window_size = 3L,
  angles = c(0, pi/2, pi/4, 3 * pi/4),
  memsize = 4L,
  multicores = 2L,
  output_dir,
  progress = FALSE
)

\method{sits_glcm}{derived_cube}(data, ...)

\method{sits_glcm}{default}(data, ...)
}
\arguments{
\item{data}{Valid sits tibble or cube}

\item{...}{Named expressions to be evaluated (see details).}

\item{window_size}{An odd number representing the size of the
sliding window of sits kernel functions
used in expressions (for a list of supported
kernel functions, please see details).}

\item{angles}{...}

\item{memsize}{Memory available for classification (in GB).}

\item{multicores}{Number of cores to be used for classification.}

\item{output_dir}{Directory where files will be saved.}

\item{progress}{Show progress bar?}
}
\value{
A sits cube with new bands, produced
        according to the requested expression.
}
\description{
...
}
\section{Summarizing GLCM functions}{

\itemize{
\item{\code{glcm_contrast()}: ...}
\item{\code{glcm_dissimilarity()}: ...}
\item{\code{glcm_homogeneity()}: ...}
\item{\code{glcm_energy()}: ...}
\item{\code{glcm_asm()}: ...}
\item{\code{glcm_mean()}: ...}
\item{\code{glcm_variance()}: ...}
\item{\code{glcm_std()}: ...}
\item{\code{glcm_correlation()}: ...}
}
}

\examples{
if (sits_run_examples()) {
    data_dir <- system.file("extdata/raster/mod13q1", package = "sits")
    cube <- sits_cube(
        source = "BDC",
        collection = "MOD13Q1-6.1",
        data_dir = data_dir
    )

    # Generate a texture images with variance in NDVI images
    cube_texture <- sits_glcm(
        data = cube,
        NDVIMEAN = glcm_mean(NDVI),
        window_size = 5,
        angles = 0,
        output_dir = tempdir()
    )
}
}
\author{
Rolf Simoes, \email{rolf.simoes@inpe.br}

Felipe Carvalho, \email{felipe.carvalho@inpe.br}

Gilberto Camara, \email{gilberto.camara@inpe.br}
}
