% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_tibble_utils.R
\name{sits_group_by}
\alias{sits_group_by}
\title{Group the contents of a sits tibble by different criteria}
\usage{
sits_group_by(data.tb, ...)
}
\arguments{
\item{data.tb}{a sits table with the time series of the selected bands}

\item{...}{one or more sits table field separated by commas that are used to group the data.
See `dplyr::group_by` help for more details.}
}
\value{
result.tb   a tibble in SITS format with the selected bands
}
\description{
returns a sits table by compound the sits tables apply a function to a grouped SITS table
}
\author{
Rolf Simoes, \email{rolf.simoes@inpe.br}
}
