% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_imputation.R
\name{sits_impute}
\alias{sits_impute}
\title{Replace NA values in time series with imputation function}
\usage{
sits_impute(samples, impute_fn = impute_linear())
}
\arguments{
\item{samples}{A time series tibble}

\item{impute_fn}{Imputation function}
}
\value{
A set of filtered time series using
                      the imputation function.
}
\description{
Remove NA
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
