% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_imputation.R
\name{sits_impute_linear}
\alias{sits_impute_linear}
\title{Replace NA values with linear interpolation}
\usage{
sits_impute_linear(data = NULL)
}
\arguments{
\item{data}{A time series vector or matrix}
}
\value{
A set of filtered time series.
}
\description{
Remove NA by linear interpolation
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
