% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_imputation.R
\name{sits_impute_stine}
\alias{sits_impute_stine}
\title{Imputation of NA values by Stineman interpolation}
\usage{
sits_impute_stine(data = NULL)
}
\arguments{
\item{data}{A time series vector}
}
\value{
A set of filtered time series
}
\description{
Remove NA by linear interpolation
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
