% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_active_learning.R
\name{sits_increase_samples}
\alias{sits_increase_samples}
\title{Suggest high confidence samples to increase the training set.}
\usage{
sits_increase_samples(u_cube, c_cube, n = 100, min_dist_pixels = 10)
}
\arguments{
\item{u_cube}{A `sits` uncertainty cube. See `sits_uncertainty`.}

\item{c_cube}{A `sits` classified cube. See
`sits_label_classification`.}

\item{n}{Number of suggested points.}

\item{min_dist_pixels}{Mininum distance among suggested points (in pixels).}
}
\value{
A data.frame with longitude & latitude in WGS84.
}
\description{
Suggest points for increasing the training set. These points are labelled
with high confidence so they can be added to the training set.
}
\author{
Alber Sanchez, \email{alber.sanchez@inpe.br}

Rolf Simoes, \email{rolf.simoes@inpe.br}

Felipe Carvalho, \email{felipe.carvalho@inpe.br}

Gilberto Camara, \email{gilberto.camara@inpe.br}
}
