% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_eocubes.R
\name{sits_info_EOCubes}
\alias{sits_info_EOCubes}
\title{Provides information about EOCubes service}
\usage{
sits_info_EOCubes()
}
\value{
An R object containing the information about the EOCubes object.
}
\description{
Obtains information about the EOCubes remotes
and about the cubes.

The EOCubes package is a lightweight link data package that allows remote access to satellite
 image time series.
}
\examples{
\donttest{
# Obtain information about the coverages available
sits_info_EOCubes()
}
}
\author{
Rolf Simoes
}
