% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_info.R
\name{sits_info_services}
\alias{sits_info_services}
\title{Provides information about time series service}
\usage{
sits_info_services()
}
\value{
A boolean value indicating success of failure.
}
\description{
Obtains information about the time series servers, their products and their coverages.
}
\author{
Gilberto Camara
}
