% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_WTSS.R
\name{sits_info_WTSS}
\alias{sits_info_WTSS}
\title{Provides information about WTSS service}
\usage{
sits_info_WTSS(URL = "http://www.dpi.inpe.br/tws/wtss")
}
\arguments{
\item{URL}{the URL for the WTSS time series service}
}
\description{
\code{sits_info_WTSS} uses the WTSS services to provide information about the
coverages that are served
}
