% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_WTSPS.R
\name{sits_infoWTSPS}
\alias{sits_infoWTSPS}
\title{Provides information about WTSPS service}
\usage{
sits_infoWTSPS(URL = "http://www.dpi.inpe.br/tws/wtsps")
}
\arguments{
\item{URL}{the URL for the WTSPS service}
}
\value{
wtsps.obj       an R object containing the information about the WTSPS server
}
\description{
uses the WTSPS services to print information about WTSPS
}
