% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_WTSS.R
\name{sits_infoWTSS}
\alias{sits_infoWTSS}
\title{Set variables for WTSS access}
\usage{
sits_infoWTSS(URL = "http://www.dpi.inpe.br/tws/wtss")
}
\arguments{
\item{URL}{the URL for the WTSS time series service}
}
\value{
wtss_info       a list containing the information about the WTSS server and coverage
}
\description{
\code{sits_infoWTSS} obtains information about the WTSS server
and about the coverages. This information is used in many applications
which obtain data from the WTSS
}
\details{
The Web Time Series Service is a lightweight web service the allow remote access to satellite
 image time series and provides three operations:

1. list coverages: this operation allows clients to retrieve the capabilities provided
    by any server that implements WTSS. It returns a list of coverage
    names available in a server instance.

2. describe coverage: this operation returns the metadata for a given coverage
    identified by its name. It includes its range in the spatial and temporal dimensions.

3. time series: this operation requests the time series of values of a coverage attribute
   at a given location.
}
