% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_filters.R
\name{sits_interp}
\alias{sits_interp}
\title{Inerpolation function of sits_table's time series}
\usage{
sits_interp(data.tb, fun = stats::approx, n = 23, ...)
}
\arguments{
\item{data.tb}{a valid sits table}

\item{fun}{the interpolation function to be used}

\item{n}{the number of time series elements to be created between start date and end date}

\item{...}{additional parameters to be used by the fun function}
}
\value{
result.tb    a sits_table with same samples and the new bands
}
\description{
Computes the linearly interpolated bands for a given resolution
using the R base function approx
}
\author{
Rolf Simoes, \email{rolf.simoes@inpe.br}
}
