% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_table.R
\name{sits_is_empty}
\alias{sits_is_empty}
\title{Check if a table is empty}
\usage{
sits_is_empty(table_obj)
}
\arguments{
\item{table_obj}{an object (the sits table)}
}
\description{
\code{sits_is_empty} tests if the sits table exists and has data
}
