% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_filters.R
\name{sits_kalman}
\alias{sits_kalman}
\title{Kalman filter}
\usage{
sits_kalman(data.tb = NULL, bands_suffix = "kf")
}
\arguments{
\item{data.tb}{A sits tibble containing the original time series.}

\item{bands_suffix}{The suffix to be appended to the smoothed filters.}
}
\value{
A tibble with smoothed sits time series.
}
\description{
A simple Kalman filter implementation.
}
