% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_machine_learning.R
\name{sits_keras_diagnostics}
\alias{sits_keras_diagnostics}
\title{Provides access to diagnostic information about a Keras deep learning model}
\usage{
sits_keras_diagnostics(test.x = NULL, test.y = NULL)
}
\arguments{
\item{test.x}{Test data to be run with the model (X variables)}

\item{test.y}{Test data to be run with the model (Y variables)}
}
\value{
NULL   Prints the model diagnostics
}
\description{
After the Keras deeplearning model is compiled and fit, this
             function provides access to the history plot and the evaluation results
}
\examples{
\donttest{
# Retrieve the set of samples for the Mato Grosso region (provided by EMBRAPA)
data(cerrado2classes)
# create a vector of distances
distances <- sits_distances(cerrado2classes)
# obtain a DL model
ml_model = sits_deeplearning(distances)
# run the keras diagnostics
sits_keras_diagnostics()
}
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
