% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_keras.R
\name{sits_keras_load}
\alias{sits_keras_load}
\title{Load a Keras model for processing in sits}
\usage{
sits_keras_load(hdffile, rdsfile)
}
\arguments{
\item{hdffile}{An hdf5 file where the keras model is to be saved.}

\item{rdsfile}{A rds file where the R environment is to be saved.}
}
\value{
An R Keras model trained by \code{\link[sits]{sits_deeplearning}}.
}
\description{
Takes a save Keras model
been saved in two files and loads it in memory for further processing.
One file is the keras model itself (saved as hdf5)
and the other is the R environment required for sits
to work with the model.
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
