% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_filters.R
\name{sits_kf}
\alias{sits_kf}
\title{Kalman filter}
\usage{
sits_kf(data.tb, bands_suffix = "kf")
}
\arguments{
\item{data.tb}{The SITS tibble containing the original time series}

\item{bands_suffix}{The suffix to be appended to the smoothed filters}
}
\value{
output.tb   A tibble with smoothed sits time series
}
\description{
A simple Kalman filter implementation
}
