% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_kohonen.R
\name{sits_kohonen}
\alias{sits_kohonen}
\title{Cluster a set of satellite image time series using SOM}
\usage{
sits_kohonen(data.tb, bands = NULL, grid_xdim = 5, grid_ydim = 5,
  rlen = 100, alpha = c(0.05, 0.01), ...)
}
\arguments{
\item{data.tb}{a SITS tibble the list of time series to be clustered}

\item{bands}{the bands to be clusterized.}

\item{grid_xdim}{x dimension of the SOM grid (used only in `kohonen` or `kohonen-dendogram` methods). Defaul is 5.}

\item{grid_ydim}{y dimension of the SOM grid}

\item{rlen}{the number of times the complete data set will be presented to the SOM grid}

\item{alpha}{learning rate, a vector of two numbers indicating the amount of change.
Default is to decline linearly from 0.05 to 0.01 over rlen updates.}

\item{...}{Additional parameters to be passed to kohonen::supersom function}
}
\value{
clusters.tb a SITS tibble with the clusters time series or cluster' members time series according to return_member parameter.
If return_members are FALSE, the returning SITS tibble will contain a new collumn called `n_members` informing how many members has each cluster.
}
\description{
This function uses package "kohonen" (self-organized maps) to find clusters in
satellite image time series.
}
\references{
`kohonen` package (https://CRAN.R-project.org/package=kohonen)
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}

Rolf Simoes, \email{rolf.simoes@inpe.br}

Lorena Alves, \email{lorena.santos@inpe.br}
}
