% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_table.R
\name{sits_label_sample}
\alias{sits_label_sample}
\title{Sample a percentage of a time series}
\usage{
sits_label_sample(data.tb, frac = 0.1)
}
\arguments{
\item{data.tb}{tibble - input SITS table}

\item{frac}{fraction (value between 0 and 1) of samples of each label to be saved.}
}
\value{
result.tb   tibble - the new SITS table with a fixed percentage of samples per class
}
\description{
takes a sits table with different labels and
returns a new table. For each label, this new table contains a percentage
of the total number of samples per label
}
\author{
Rolf Simoes, \email{rolf.simoes@inpe.br}
}
