% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_labels.R
\name{sits_labels<-.class_vector_cube}
\alias{sits_labels<-.class_vector_cube}
\title{Change the labels of a class vector data cube}
\usage{
\method{sits_labels}{class_vector_cube}(data) <- value
}
\arguments{
\item{data}{Class vector data cube.}

\item{value}{A character vector used to convert labels.}
}
\description{
A "class_vector_cube" is a data cube with a set of segments that
contain the probability values of each class for each polygon
and also list the most probable class.
When a user changes the labels of the class, this function
modifies the labels associated to the cube's metadata and
also changes the names in the segments file. The GPKG file
containing the segments and the probability values is replace
with a new file with the desired labels.
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
