% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_labels.R
\name{sits_labels<-.default}
\alias{sits_labels<-.default}
\title{Change the labels of other data structures}
\usage{
\method{sits_labels}{default}(data) <- value
}
\arguments{
\item{data}{Data cube or time series.}

\item{value}{A character vector used to convert labels. Labels will
be renamed to the respective value positioned at the
labels order returned by \code{\link{sits_labels}}.}
}
\description{
Change the labels of other data structures
}
