% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_labels.R
\name{sits_labels<-.sits}
\alias{sits_labels<-.sits}
\title{Change the labels of a set of time series}
\usage{
\method{sits_labels}{sits}(data) <- value
}
\arguments{
\item{data}{Time series.}

\item{value}{A character vector used to convert labels. Labels will
be renamed to the respective value positioned at the
labels order returned by \code{\link{sits_labels}}.}
}
\description{
Change the labels of a set of time series
}
