% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_labels.R
\name{sits_labels<-}
\alias{sits_labels<-}
\title{Change the labels of a set of time series}
\usage{
sits_labels(data) <- value
}
\arguments{
\item{data}{Data cube or time series.}

\item{value}{A character vector used to convert labels. Labels will
be renamed to the respective value positioned at the
labels order returned by \code{\link{sits_labels}}.}
}
\value{
A sits tibble or data cube with modified labels.
}
\description{
Given a sits tibble with a set of labels, renames the labels
to the specified in value.
}
\examples{
# show original samples ("Cerrado" and "Pasture")
sits_labels(cerrado_2classes)
# rename label samples to "Savanna" and "Grasslands"
sits_labels(cerrado_2classes) <- c("Savanna", "Grasslands")
# see the change
sits_labels(cerrado_2classes)
}
\seealso{
Other metadata:
\code{\link[=sits_bands]{sits_bands()}},
\code{\link[=sits_bbox]{sits_bbox()}},
\code{\link[=sits_labels]{sits_labels()}},
\code{\link[=sits_labels_summary]{sits_labels_summary()}},
\code{\link[=sits_timeline]{sits_timeline()}}
}
\author{
Rolf Simoes, \email{rolfsimoes@gmail.com}
}
\concept{metadata}
