% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_labels.R
\name{sits_labels.cube}
\alias{sits_labels.cube}
\title{Returns the information about labels of a data cube}
\usage{
\method{sits_labels}{cube}(data)
}
\arguments{
\item{data}{A data cube}
}
\value{
A list of labels
}
\description{
Returns the labels and its respective counting and frequency.
}
