% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_labels.R
\name{sits_labels.patterns}
\alias{sits_labels.patterns}
\title{Returns the information about labels of patterns tibble}
\usage{
\method{sits_labels}{patterns}(data)
}
\arguments{
\item{data}{A data cube}
}
\value{
A tibble with the names of the labels and
                their absolute and relative frequency.
}
\description{
Returns the labels and its respective counting and frequency.
}
