% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_labels.R
\name{sits_labels.sits}
\alias{sits_labels.sits}
\title{Returns the information about labels of a set of time series}
\usage{
\method{sits_labels}{sits}(data)
}
\arguments{
\item{data}{A tibble with time series data and metadata.}
}
\value{
A tibble with the names of the labels and
                their absolute and relative frequency.
}
\description{
Returns the labels and its respective counting and frequency.
}
\examples{
# read a tibble with 400 samples of Cerrado and 346 samples of Pasture
data(cerrado_2classes)
# print the labels
sits_labels(cerrado_2classes)
}
