% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_labels.R
\name{sits_labels_sample}
\alias{sits_labels_sample}
\title{Sample a percentage of a time series}
\usage{
sits_labels_sample(data.tb, samples.lst)
}
\arguments{
\item{data.tb}{input SITS table}

\item{samples.lst}{a named list with the nambber of samples to be randomly picked from data input.
The labels are the list names, and the numbers of samples the list values.
If the size of a sample is greater than the number of actual samples of that label in data input, then
sampling procedure will be realized with replacement. If no entry for a given label is informed,
all samples of that label will be returned}
}
\value{
result.tb    the new SITS table with a fixed quantity of samples of informed labels and all other
}
\description{
takes a sits table with different labels and
returns a new table. For each label, this new table contains a percentage
of the total number of samples per label
}
\author{
Rolf Simoes, \email{rolf.simoes@inpe.br}
}
