% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_table.R
\name{sits_lag_diff}
\alias{sits_lag_diff}
\title{Lagged differences of a SITS band.}
\usage{
sits_lag_diff(data.tb, bands = NULL, differences = 1)
}
\arguments{
\item{data.tb}{a valid sits table}

\item{bands}{a vector of strings with band's names.}

\item{differences}{an integer indicating the order of the difference.}
}
\value{
result.tb    a sits_table with same samples and the new bands
}
\description{
Computes the lagged differences of a set of time series.
}
\author{
Rolf Simoes, \email{rolf.simoes@inpe.br}
}
