% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_cluster.R
\name{sits_last_cluster}
\alias{sits_last_cluster}
\title{Last cluster object}
\usage{
sits_last_cluster()
}
\value{
.sits_last_cluster an object created by dtwclust or SOM
}
\description{
Returns the content of global variable `.sits_last_cluster` that stores the last
cluster object generated by sits_cluster function.
}
\author{
Rolf Simoes, \email{rolf.simoes@inpe.br}
}
