% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_projections.R
\name{sits_latlong_to_proj}
\alias{sits_latlong_to_proj}
\title{Coordinate transformation (lat/long to X/Y)}
\usage{
sits_latlong_to_proj(longitude, latitude, crs)
}
\arguments{
\item{longitude}{double - the longitude of the chosen location}

\item{latitude}{double - the latitude of the chosen location}

\item{crs}{projection definition to be converted to}
}
\value{
xy             a matrix with X/Y coordinates
}
\description{
Transform a latitude and longitude coordinate to a XY projection coordinate
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
