% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_machine_learning.R
\name{sits_lda}
\alias{sits_lda}
\title{Train SITS classifiction models}
\usage{
sits_lda(data.tb = NULL, formula = sits_formula_logref(), ...)
}
\arguments{
\item{data.tb}{a SITS tibble time series with an `matches` column. If `data.tb` is NULL, the function returns
a pre-parameterized function to be called further in order to compute `lda` training model}

\item{formula}{a symbolic description of the model to be fit. SITS offers a set of such formulas (default: sits_formula_logref)}

\item{...}{other parameters to be passed to MASS::lda function}
}
\value{
result   either an MASS::lda class or an function prepared that can be called further to compute lda training model
}
\description{
Performs a linear discriminant analysis (lda) to classify data.
This function is a front-end to the "lda" method in the "MASS" package.
Please refer to the documentation in that package for more details.
}
\author{
Alexandre Xavier Ywata de Carvalho, \email{alexandre.ywata@ipea.gov.br}

Victor Maus, \email{vwmaus1@gmail.com}

Gilberto Camara, \email{gilberto.camara@inpe.br}

Rolf Simoes, \email{rolf.simoes@inpe.br}
}
