% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_filters.R
\name{sits_linear_interp}
\alias{sits_linear_interp}
\title{Interpolation function of the time series in a sits tibble}
\usage{
sits_linear_interp(data.tb, n = 23)
}
\arguments{
\item{data.tb}{a valid sits tibble}

\item{n}{the number of time series elements to be created between start date and end date}
}
\value{
result.tb    a sits tibble with same samples and the new bands
}
\description{
Computes the linearly interpolated bands for a given resolution
              using the R base function approx
}
\examples{
# Retrieve a point
point.tb <- readRDS(system.file("extdata/time_series/point.rds", package = "sits"))
# select one band
point.tb <- sits_select (point.tb, bands = c("ndvi"))
# find out how many time instances are there
n_times <- NROW (point.tb$time_series[[1]])
# interpolate three times more points
point_int.tb <- sits_linear_interp(point.tb, n = 3*n_times)
# plot the result
sits_plot (point_int.tb)

}
\author{
Rolf Simoes, \email{rolf.simoes@inpe.br}
}
