% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_config.R
\name{sits_list_collections}
\alias{sits_list_collections}
\title{List the cloud collections supported by sits}
\usage{
sits_list_collections(source = NULL)
}
\arguments{
\item{source}{Data source to be shown in detail.}
}
\value{
Prints collections available in
        each cloud service supported by sits.
}
\description{
Prints the collections available
in each cloud service supported by sits.
Users can select to get information
only for a single service by using the \code{source} parameter.
}
\examples{
if (sits_run_examples()) {
    # show the names of the colors supported by SITS
    sits_list_collections()
}
}
\seealso{
Other utilities:
\code{\link[=sits_config]{sits_config()}},
\code{\link[=sits_config_show]{sits_config_show()}},
\code{\link[=sits_config_user_file]{sits_config_user_file()}},
\code{\link[=sits_factory_function]{sits_factory_function()}},
\code{\link[=sits_parallel]{sits_parallel()}},
\code{\link[=sits_run_examples]{sits_run_examples()}},
\code{\link[=sits_run_tests]{sits_run_tests()}}
}
\concept{utilities}
