% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_log.R
\name{sits_log}
\alias{sits_log}
\title{Creates a directory and logfile for logging information and debug}
\usage{
sits_log()
}
\value{
boolean      TRUE if creation succeeds
}
\description{
Creates a logger file for debugging and information (uses log4r package)
}
\examples{
# Creates a sits logger using the default location
logger <- sits_log()
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
