% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_cluster.R
\name{sits_mclust_compute}
\alias{sits_mclust_compute}
\title{Compute the clusters for a chosen model}
\usage{
sits_mclust_compute(data.tb, nclusters = 9, model_name = NULL)
}
\arguments{
\item{data.tb}{a SITS tibble with labelled samples.}

\item{nclusters}{the maximum number of clusters assumed to exist in the data.}

\item{model_name}{one of the `mclust` model structures.}
}
\value{
data.tb          a SITS tibble with two additional columns: "cluster" (integer) and
                         "uncertainity" (double).
}
\description{
This function is part the support in `sits` for model-based clustering.
The general idea of model-based clustering is that the input data can be modelled as
a finite mixture model of Gaussian distributions, with different covariance structures
and different members of misture components. The package `sits` provides a convenient front-end
to the `mclust` package.

This function computes the cluster values assigned to each input samples, based on
a model chosen by the user. Users should first run \code{\link[sits]{sits_mclust_info}}
the get the bayesian information criteria (BIC) for each different model structure.
Then, they should select one of models suggested by this function to assign a
cluster to each input of the SITS tibble. The function will also assign an uncertainity value
to the predicted cluster.

Usually, the model_name will be one of {"EEI", "VII", "EEI", "VEI", "EVI", "VVI",
"EEE", "EVE", "VEE", "VVE", "EEV", "VEV", "EVV", "VVV"}. See the reference for more details.
}
\examples{
\donttest{
# Load the "mclust" package
library(mclust)
# load a simple data set with two classes
data(cerrado_2classes)
# calculate the BIC for different model shapes
BIC <- sits_mclust_info(cerrado_2classes, nclusters = 4)
# select one model for processing
cluster.tb <- sits_mclust_compute(cerrado_2classes, nclusters = 4, model_name = "EEE")
# show clusters samples frequency
sits_cluster_frequency(clusters.tb)
}
}
\references{
Scrucca L, Fop M, Murphy TB, Raftery AE.
              mclust 5: Clustering, Classification and Density Estimation Using Gaussian
              Finite Mixture Models. The R journal, 8(1):289-317, 2016.
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
