% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_cluster.R
\name{sits_mclust_info}
\alias{sits_mclust_info}
\title{Measure information content on model based clusterinbg}
\usage{
sits_mclust_info(data.tb, nclusters = 9)
}
\arguments{
\item{data.tb}{a SITS tibble with labelled samples}

\item{nclusters}{the maximum number of clusters assumed to exist in the data}
}
\value{
BIC              a `mclust` structure with the bayesian information criterion.
}
\description{
This function is part the support in `sits` for model-based clustering.
The general idea of model-based clustering is that the input data can be modelled as
a finite mixture model of Gaussian distributions, with different covariance structures
and different members of misture components. The package `sits` provides a convenient front-end
to the `mclust` package.

This function measures the bayesian information criterion (BIC) for a set
of parametrized Gaussian mixture models fitted by an expectation-maximization (EM) method.
It plots the different models for the chosen number of clusters and informs the best models
based on the BIC.
}
\examples{
\donttest{
# Load the "mclust" package
library(mclust)
# load a simple data set with two classes
data(cerrado_2classes)
# calculate the BIC for different model shapes
BIC <- sits_mclust_info(cerrado_2classes, nclusters = 4)
}
}
\references{
Scrucca L, Fop M, Murphy TB, Raftery AE.
              mclust 5: Clustering, Classification and Density Estimation Using Gaussian
              Finite Mixture Models. The R journal, 8(1):289-317, 2016.
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
