% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_data_generics.R
\name{sits_merge.cube}
\alias{sits_merge.cube}
\title{Merge two data cubes}
\usage{
\method{sits_merge}{cube}(data1, data2)
}
\arguments{
\item{data1}{The first cube to be merged.}

\item{data2}{The second cube to be merged.}
}
\description{
Merge two data cubes
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
