% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_mesma.R
\name{sits_mesma}
\alias{sits_mesma}
\title{Create a multiple endmember spectral mixture analyses fractions images}
\usage{
sits_mesma(
  cube,
  endmembers,
  memsize = 1,
  multicores = 2,
  output_dir = getwd(),
  progress = TRUE
)
}
\arguments{
\item{cube}{...}

\item{endmembers}{...}

\item{memsize}{...}

\item{multicores}{...}

\item{output_dir}{...}

\item{progress}{...}
}
\value{
a cube ...
}
\description{
A multiple endmember spectral mixture analyses...
}
