% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_kohonen.R
\name{sits_metrics_by_cluster}
\alias{sits_metrics_by_cluster}
\title{Metrics by cluster}
\usage{
sits_metrics_by_cluster(info_sample_cluster.tb)
}
\arguments{
\item{info_sample_cluster.tb}{tibble containg information about evaluation of samples}
}
\value{
returns the confusion matrix and a table with percentage of mixture between the clusters
}
\description{
This function extracts metrics about the clusters calculating
the percentage of mixture between a cluster and others.
}
\author{
Lorena Santos, \email{lorena.santos@inpe.br}
}
