% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_metrics.R
\name{sits_metrics_cube}
\alias{sits_metrics_cube}
\title{...}
\usage{
sits_metrics_cube(cube, metrics)
}
\arguments{
\item{cube}{input data cube.}

\item{metrics}{metrics list. melhorar a documentacao}
}
\value{
the same input data cube
}
\description{
a lazy-eval metrics creation cube
}
