% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_metrics_generate.R
\name{sits_metrics_generate}
\alias{sits_metrics_generate}
\title{Generate bands metrics from cube images}
\usage{
sits_metrics_generate(
  cube,
  output_dir,
  name,
  metrics = list(EVI = "min", NDVI = "min"),
  impute_fn = sits_impute_linear(),
  memsize = 8,
  multicores = 2
)
}
\arguments{
\item{cube}{input data cube}

\item{output_dir}{data directory where output data is written}

\item{name}{name of the output data cube}

\item{metrics}{a \code{character} or \code{list} of metrics. If the
input is a \code{character} vector, the following
specification must be obeyed: "X1_metrics_name", where
"X1" is the name of the band contained in the cube.
Otherwise, a named \code{list} must be provided, where
in the name is contained the band and the values in
character vector form with the name of each metric.}

\item{impute_fn}{imputing function to be applied to replace NA}

\item{memsize}{size of memory}

\item{multicores}{number of cores}
}
\value{
new data cube of metrics
}
\description{
Generate a set of metrics by band
}
\examples{
# define a data cube of CBERS-4 AWFI data
data_dir <- system.file("extdata/raster/cbers", package = "sits")

cbers_022024 <- sits_cube(
    source = "LOCAL",
    name = "cbers_022024",
    satellite = "CBERS-4",
    sensor = "AWFI",
    resolution = "64m",
    data_dir = data_dir,
    parse_info = c("X1", "X2", "band", "date")
)

cbers_022024_no_clds <- sits_metrics_generate(
   cube = cbers_022024,
   output_dir = tempdir(),
   name = "cbers_022024_no_cld",
   metrics = list("B13" = c("max", "min")),
   impute_fn = sits_impute_linear(),
   memsize = 8,
   multicores = 2
)
}
