% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_cube.R
\name{sits_mgrs_to_roi}
\alias{sits_mgrs_to_roi}
\title{Convert MGRS tile information to ROI in WGS84}
\usage{
sits_mgrs_to_roi(tiles)
}
\arguments{
\item{tiles}{Character vector with names of MGRS tiles}
}
\value{
roi                  Valid ROI to use in other SITS functions
}
\description{
Takes a list of MGRS tiles and produces a ROI covering them
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}

Rolf Simoes, \email{rolf.simoes@gmail.com}
}
