% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_filters.R
\name{sits_missing_values}
\alias{sits_missing_values}
\title{Remove missing values}
\usage{
sits_missing_values(data, miss_value)
}
\arguments{
\item{data}{A tibble with time series data and metadata.}

\item{miss_value}{Number indicating missing values in a time series.}
}
\value{
A tibble with time series data and metadata (with missing values removed).
}
\description{
This function removes the missing values from an image time series by substituting them by NA.
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
