% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_filters.R
\name{sits_ndvi_arima}
\alias{sits_ndvi_arima}
\title{NDVI filter with ARIMA model}
\usage{
sits_ndvi_arima(
  data = NULL,
  cutoff = -0.25,
  p = 0,
  d = 0,
  q = 3,
  bands_suffix = "ar",
  apply_whit = TRUE,
  lambda_whit = 1
)
}
\arguments{
\item{data}{Time series data and metadata (only the "NDVI" band).}

\item{cutoff}{Maximum acceptable value of a NDVI difference.}

\item{p}{Order (number of time lags) of the autoregressive model.}

\item{d}{Degree of differencing  (the number of times
the data has had past values subtracted).}

\item{q}{Order of the moving-average model.}

\item{bands_suffix}{Suffix to rename the filtered bands.}

\item{apply_whit}{Apply the whittaker smoother after filtering? (logical)}

\item{lambda_whit}{Lambda parameter of the whittaker smoother.}
}
\value{
A sits tibble with same samples and the new bands.
}
\description{
This function tries to remove clouds in the NDVI band of
a time series. It looks for points where the value of the NDVI
band goes down abruptly. These points are those whose difference is more
than a cutoff value which is set by the user.
Then it applies an spline interploation.
Finally, the function applies a whitakker smoother.
}
\examples{
# Select the NDVI band of a point
# Apply the filter
point_ar <- sits_ndvi_arima(point_ndvi)
# Merge the filtered with the raw data
point2 <- sits_merge(point_ndvi, point_ar)
# Plot the result
plot(point2)
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
