% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_filters.R
\name{sits_ndvi_arima_filter}
\alias{sits_ndvi_arima_filter}
\title{NDVI filter with ARIMA model}
\usage{
sits_ndvi_arima_filter(data.tb = NULL, cutoff = -0.25, p = 0, d = 0,
  q = 3, bands_suffix = "ar", apply_whit = TRUE, lambda_whit = 1)
}
\arguments{
\item{data.tb}{A tibble with time series data and metadata with only the "ndvi" band.}

\item{cutoff}{Numeric value for the maximum acceptable value of a NDVI difference.}

\item{p}{Order (number of time lags) of the autoregressive model.}

\item{d}{Degree of differencing (the number of times the data have had past values subtracted).}

\item{q}{Order of the moving-average model.}

\item{bands_suffix}{Suffix to rename the filtered bands.}

\item{apply_whit}{Apply the whittaker smoother after filtering? (logical)}

\item{lambda_whit}{Lambda parameter of the whittaker smoother.}
}
\value{
A sits tibble with same samples and the new bands.
}
\description{
This function tries to remove clouds in the ndvi band of
a satellite image time series. It looks for points where the value of the NDVI
band goes down abruptly. These points are taken as those whose difference is more
than a cutoff value which is set by the user. Then it applies an spline interploation.
Finally, the function applies a whitakker smoother.
}
\examples{
\donttest{
# Read a set of samples of forest/non-forest in Amazonia
# This is an area full of clouds
data(prodes_226_064)
# Select the NDVI band of the first point
point_ndvi.tb <- sits_select(prodes_226_064[1,], bands = c("ndvi"))
# Apply the cloud filter
point_cld.tb <- sits_ndvi_arima_filter(point_ndvi.tb)
# Merge the filtered with the raw data
point2.tb <- sits_merge (point_ndvi.tb, point_cld.tb)
# Plot the result
sits_plot (point2.tb)
}
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
