% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_indexes.R
\name{sits_ndwi}
\alias{sits_ndwi}
\title{Builds normalized difference water index}
\usage{
sits_ndwi(data)
}
\arguments{
\item{data}{A valid sits tibble.}
}
\value{
A sits tibble with the SAVI index.
}
\description{
Adds new tasseled cap bands.
}
\examples{
\donttest{
# Retrieve data for time series with label samples in Mato Grosso in Brazil
data(samples_mt_6bands)
# Generate a new image with the tasseled cap
ndwi.tb <- sits_ndwi(samples_mt_6bands)
}
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
