% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_patterns.R
\name{sits_no_patterns}
\alias{sits_no_patterns}
\title{Do not create patterns for classification}
\usage{
sits_no_patterns(data.tb = NULL)
}
\arguments{
\item{data.tb}{a SITS tibble time series with an alignment column}
}
\value{
result          a model fitted into input data given by train_method parameter
}
\description{
This function is to be used when oen does not want to create
a set of idealised patterns from the input data, but prefers to use the
input series themselves as training data from estimating a prediction model.
It should be used in connection with sits_spread_time_series.
}
\author{
Rolf Simoes, \email{rolf.simoes@inpe.br}

Gilberto Camara, \email{gilberto.camara@inpe.br}
}
