% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_tibble_functions.R
\name{sits_normalization_param}
\alias{sits_normalization_param}
\title{Normalize the time series in the given sits_tibble}
\usage{
sits_normalization_param(data.tb, use_IQR = TRUE)
}
\arguments{
\item{data.tb}{a tibble in SITS format}

\item{use_IQR}{(logical): should we use inter-quartile ranges?}
}
\value{
result.tb  a tibble with statistics
}
\description{
this function normalizes the time series using the mean and
standard deviation of all the time series.
}
\author{
Alber Sanchez, \email{alber.ipia@inpe.br}
}
