% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_tibble_functions.R
\name{sits_normalize}
\alias{sits_normalize}
\title{Normalize the time series in the given sits_tibble}
\usage{
sits_normalize(data.tb = NULL)
}
\arguments{
\item{data.tb}{a tibble in SITS format}
}
\value{
result.tb  a list of 2: A sits_tibble and a tibble with statistics
}
\description{
this function normalizes the time series using the mean and
standard deviation of all the time series.
}
\examples{
data(point_MT_6bands)
(sits_tibble_normalized <- sits_normalize(point_MT_6bands)[[1]])
(sits_tibble_stats <- sits_normalize(point_MT_6bands)[[2]])
}
\author{
Alber Sánchez, \email{alber.ipia@inpe.br}
}
