% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_normalization.R
\name{sits_normalize_data}
\alias{sits_normalize_data}
\title{Normalize the time series in the given sits_tibble}
\usage{
sits_normalize_data(data.tb, stats.tb, multicores = NULL)
}
\arguments{
\item{data.tb}{a tibble in SITS format}

\item{stats.tb}{statistics for normalization}

\item{multicores}{number of cores to process}
}
\value{
data.tb    a normalized sits tibble
}
\description{
this function normalizes the time series using the mean and
standard deviation of all the time series.
}
\author{
Alber Sanchez, \email{alber.ipia@inpe.br}
}
