% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_parallel.R
\name{sits_parallel_fault_tolerant}
\alias{sits_parallel_fault_tolerant}
\alias{.sits_parallel_recv_one_data}
\alias{.sits_parallel_recv_one_result}
\alias{.sits_parallel_cluster_apply}
\title{Fault tolerant version of some parallel functions}
\usage{
.sits_parallel_recv_one_data()

.sits_parallel_recv_one_result()

.sits_parallel_cluster_apply(x, fn, ...)
}
\arguments{
\item{x}{a given list to be passed to a function}

\item{fn}{a function to be applied to each list element}

\item{...}{additional arguments to fn function}
}
\description{
Fault tolerant version of some parallel functions
}
\author{
Rolf Simoes, \email{rolf.simoes@inpe.br}

Derived from the snow package
}
\keyword{internal}
