% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_patterns.R
\name{sits_patt_dendgam}
\alias{sits_patt_dendgam}
\title{Create time series patterns for classification}
\usage{
sits_patt_dendgam(data.tb = NULL, first_cluster_cleaner = 0.027,
  cluster_a_labels = c("Forest", "Cerrado", "Pasture"),
  second_cluster_cleaner = 0.035, k_cluster_a = 8, k_cluster_b = 6)
}
\arguments{
\item{data.tb}{a SITS tibble time series with an alignment column}

\item{first_cluster_cleaner}{first_cluster_cleaner}

\item{cluster_a_labels}{cluster_a_labels}

\item{second_cluster_cleaner}{second_cluster_cleaner}

\item{k_cluster_a}{k_cluster_a}

\item{k_cluster_b}{k_cluster_b}
}
\value{
result                 result
}
\description{
This function ...
}
\author{
Rolf Simoes, \email{rolf.simoes@inpe.br}

Michelle Picoli, \email{mipicoli@gmail.com}
}
