% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_patterns.R
\name{sits_patterns}
\alias{sits_patterns}
\title{Find temporal patterns associated to a set of time series}
\usage{
sits_patterns(data = NULL, freq = 8L, formula = y ~ s(x), ...)
}
\arguments{
\item{data}{Time series.}

\item{freq}{Interval in days for estimates.}

\item{formula}{Formula to be applied in the estimate.}

\item{...}{Any additional parameters.}
}
\value{
Time series with patterns.
}
\description{
This function takes a set of time series samples as input
estimates a set of patterns. The patterns are calculated using a GAM model.
The idea is to use a formula of type y ~ s(x), where x is a temporal
reference and y if the value of the signal. For each time, there will
be as many predictions as there are sample values.
The GAM model predicts a suitable
approximation that fits the assumptions of the statistical model,
based on a smooth function.

This method is based on the "createPatterns" method of the R dtwSat package,
which is also described in the reference paper.
}
\examples{
if (sits_run_examples()) {
    patterns <- sits_patterns(cerrado_2classes)
    plot(patterns)
}

}
\references{
Maus V, Camara G, Cartaxo R, Sanchez A, Ramos F, Queiroz GR.
A Time-Weighted Dynamic Time Warping Method for Land-Use
and Land-Cover Mapping. IEEE Journal of Selected Topics in Applied
Earth Observations and Remote Sensing, 9(8):3729-3739,
August 2016. ISSN 1939-1404. \doi{10.1109/JSTARS.2016.2517118}.

Maus, V., Câmara, G., Appel, M., & Pebesma, E. (2019).
dtwSat: Time-Weighted Dynamic Time Warping for Satellite Image
Time Series Analysis in R. Journal of Statistical Software, 88(5), 1–31.
\doi{10.18637/jss.v088.i05}.
}
\seealso{
Other training data:
\code{\link[=sits_confidence_sampling]{sits_confidence_sampling()}},
\code{\link[=sits_geo_dist]{sits_geo_dist()}},
\code{\link[=sits_get_data]{sits_get_data()}},
\code{\link[=sits_random_sampling]{sits_random_sampling()}},
\code{\link[=sits_reduce_imbalance]{sits_reduce_imbalance()}},
\code{\link[=sits_sample]{sits_sample()}},
\code{\link[=sits_sampling_design]{sits_sampling_design()}},
\code{\link[=sits_stratified_sampling]{sits_stratified_sampling()}},
\code{\link[=sits_uncertainty_sampling]{sits_uncertainty_sampling()}}
}
\author{
Victor Maus, \email{vwmaus1@gmail.com}

Gilberto Camara, \email{gilberto.camara@inpe.br}

Rolf Simoes, \email{rolfsimoes@gmail.com}
}
\concept{training data}
