% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_TWDTW.R
\name{sits_patterns_fromTWDTW}
\alias{sits_patterns_fromTWDTW}
\title{Transform patterns from .RData to JSON}
\usage{
sits_patterns_fromTWDTW(patterns, coverage = cov_name.global)
}
\description{
Transform patterns from .RData to JSON
}
