% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_patterns.R
\name{sits_patterns_gam}
\alias{sits_patterns_gam}
\title{Create temporal patterns using a generalised additive model (gam)}
\usage{
sits_patterns_gam(samples.tb, bands, freq = 8, from = NULL, to = NULL,
  formula = y ~ s(x))
}
\arguments{
\item{samples.tb}{a table in SITS format with time series to be classified using TWTDW}

\item{bands}{the bands used to obtain the pattern}

\item{freq}{int - the interval in days for the estimates to be generated}

\item{from}{starting date of the estimate (month-day)}

\item{to}{end data of the estimated (month-day)}

\item{formula}{the formula to be applied in the estimate}
}
\value{
patterns.tb   a SITS table with the patterns
}
\description{
This function takes a set of time series samples as input
estimates a set of patterns. The patterns are calculated based in a GAM model.
The idea is to use a formula of type y ~ s(x), where x is a temporal
reference and y if the value of the signal. For each time, there will be as many predictions
as there are sample values. The GAM model predicts a suitable
approximation that fits the assumptions of the statistical model.
By default, the gam methods  produces an approximation based on a smooth function.

This method is based on the "createPatterns" method of the dtwSat package, which is also
described in the reference paper.
}
\references{
Maus V, Camara G, Cartaxo R, Sanchez A, Ramos FM, de Queiroz GR (2016).
A Time-Weighted Dynamic Time Warping Method for Land-Use and Land-Cover Mapping.
IEEE Journal of Selected Topics in Applied Earth Observations and Remote Sensing, 9(8):3729-3739,
August 2016. ISSN 1939-1404. doi:10.1109/JSTARS.2016.2517118.
}
\author{
Victor Maus, \email{vwmaus1@gmail.com}

Gilberto Camara, \email{gilberto.camara@inpe.br}
}
