% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_plot.R
\name{sits_plot}
\alias{sits_plot}
\title{Plot a set of time series for the same spatio-temporal reference}
\usage{
sits_plot(data.tb, type = "joined", colours = "Dark2")
}
\arguments{
\item{colours}{the color pallete to be used (default is "Set1")}

\item{ts}{a tibble the list of time series to be plotted}
}
\value{
none
}
\description{
\code{sits_plot_table} plot one time series for a single place and interval
}
\details{
Plot one SITS time series for a single place and interval
}
\examples{
sits_plot  (ts, "Set1")


}
\seealso{
Other STIS auxiliary functions: \code{\link{sits_cluster_hierarchical}},
  \code{\link{sits_cluster_partitional}}
}
\keyword{STIS}
