% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_plot.R
\name{sits_plot_all_samples}
\alias{sits_plot_all_samples}
\title{\code{sits_plot_all_samples} plots all samples for a given table and band}
\usage{
sits_plot_all_samples(ts, bands, label, number)
}
\arguments{
\item{ts}{the list of time series to be plotted (a list of tibbles)}

\item{bands}{thes bands to be plotted}
}
\value{
none
}
\description{
\code{sits_plot_all_samples} plots all samples for a given table and band
}
