% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_plot.R
\name{sits_plot_dendrogram}
\alias{sits_plot_dendrogram}
\title{Plot a dendrogram}
\usage{
sits_plot_dendrogram(data.tb, cluster_obj, cutree_height = NULL,
  colors = "RdYlGn")
}
\arguments{
\item{data.tb}{SITS tibble to extract labels}

\item{cluster_obj}{cluster object. Usually stored by `sits_cluster` function in `.sits_last_object`}

\item{cutree_height}{A dashed horizontal line to be drawed indicating the height of dendrogram cutting.}

\item{colors}{a color scheme as showed in `sits_color_name` function}
}
\description{
Plot an enhanced dendrogram based on a cluster object usually found in `.sits_last_cluster`
}
