% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_plot_distances.R
\name{sits_plot_distances}
\alias{sits_plot_distances}
\title{Make a kernel density plot of samples distances.}
\usage{
sits_plot_distances(samples_tb, roi_sf, n)
}
\arguments{
\item{samples_tb}{A `sits` tibble.}

\item{roi_sf}{A `sf` object (polygon).}
}
\value{
A `ggplot2` object.
}
\description{
Produce a figure showing the kernel density estimates of the distances
among the training samples and between the samples and random points in the
given region of interest.
}
\author{
Alber Sanchez, \email{alber.sanchez@inpe.br}

Rolf Simoes, \email{rolf.simoes@inpe.br}

Felipe Carvalho, \email{felipe.carvalho@inpe.br}

Gilberto Camara, \email{gilberto.camara@inpe.br}
}
