% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_plot_cluster.R
\name{sits_plot_kohonen}
\alias{sits_plot_kohonen}
\title{Plot the SOM grid with neurons labeled}
\usage{
sits_plot_kohonen(koh, type = "codes")
}
\arguments{
\item{koh}{Data to be plotted (must be a kohonen object).}

\item{type}{Type of plot. "codes" is the weight of neuron (time series) and "mapping" is the number of samples allocated in a neuron.}
}
\description{
Given a kohonen object with a set of time neurons, plot them.

The plot function produces different plots based on the input data:
\itemize{
 \item{"codes" - }{Plot in each neuron the vector weight which corresponds to it }
 \item{"mapping" - }{shows where samples are mapped.}

}
The sits_plot_kohonen function plots a classified kohonen map.A set of neuron
with same category corresponds to a same cluster.
}
\author{
Lorena Santos \email{lorena.santos@inpe.br}
}
