% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_plot.R
\name{sits_plot_one_by_one}
\alias{sits_plot_one_by_one}
\title{Plot  time intervals of one time series separately (one-by-one)}
\usage{
sits_plot_one_by_one(data.tb, colors)
}
\arguments{
\item{data.tb}{one or more time series (stored in a SITS tibble)}

\item{colors}{the color pallete to be used (default is "Dark2")}
}
\description{
plots each row of a data set separately
}
