% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_plot_cluster.R
\name{sits_plot_som}
\alias{sits_plot_som}
\title{Plot the SOM grid with neurons labeled}
\usage{
sits_plot_som(koh, type = "codes", whatmap = 1)
}
\arguments{
\item{koh}{Kohonen map produced by "sits_som_map" function}

\item{type}{Type of plot. "codes" is the weight of neuron (time series) and "mapping" is the number of samples allocated in a neuron.}

\item{whatmap}{What data layer will be plotted.}
}
\description{
Given a kohonen object with a set of time neurons, plot them.

The plot function produces different plots based on the input data:
\itemize{
 \item{"codes": }{Plot in each neuron the vector weight which corresponds to it.}
 \item{"mapping": }{Shows where samples are mapped.}
}
Function sits_plot_som plots a classified kohonen map. A set of neurons
with same category corresponds to a same cluster.
}
\examples{
\donttest{
# Produce a cluster map
som_cluster <- sits_som_map(prodes_226_064)
# Plot kohonen map with vector of weight
sits_plot_som(som_cluster, type = "codes")
# Plot kohonen map showing where the samples were allocated
sits_plot_som(som_cluster, type = "mapping")
}

}
\author{
Lorena Santos \email{lorena.santos@inpe.br}
}
