% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_plot_cluster.R
\name{sits_plot_som_clusters}
\alias{sits_plot_som_clusters}
\title{Plot information about clusters}
\usage{
sits_plot_som_clusters(data, text_title = " Cluster ")
}
\arguments{
\item{data}{Table containing the percentage of mixture between the clusters
(produced by \code{\link[sits]{sits_evaluate_cluster}})}

\item{text_title}{Title of plot. Default is "Cluster".}
}
\description{
Plot a bar graph with informations about each cluster.
The percentage of mixture between the clusters.
}
\examples{
\donttest{
# Produce a cluster map
som_cluster <- sits_som_map(prodes_226_064)
# Evaluate the clusters
cluster_overall <- sits_evaluate_cluster(som_cluster)
# Plot confusion between the clusters
sits_plot_som_clusters(cluster_overall, "Confusion by cluster")
}
}
\author{
Lorena Santos \email{lorena.santos@inpe.br}
}
