% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_plot_cluster.R
\name{sits_plot_subgroups}
\alias{sits_plot_subgroups}
\title{Plot the patterns of subgroups}
\usage{
sits_plot_subgroups(neurons_subgroup)
}
\arguments{
\item{neurons_subgroup}{The list contain the EVI and NDVI time series (weight of each neuron) by class.}
}
\description{
Plot the average pattern of subgroups from  weights of neurons labelled.
Each neuron has a weight which can be represent a set of time series samples. Neurons of same
category or label form a cluster, but the neurons within a cluster can have different patterns.
The plots are also saved into files named "class_neurons<class_neuron>_plot_EVI<subgroup>.png".
}
\author{
Lorena Santos \email{lorena.santos@inpe.br}
}
