% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_colors.R, R/sits_plot.R
\name{sits_plot_title}
\alias{sits_plot_title}
\alias{sits_color_name}
\alias{sits_plot_title}
\alias{.sits_plot_title}
\title{Brewer color schemes}
\usage{
sits_color_name(name = NULL)

.sits_plot_title(row)
}
\arguments{
\item{name}{name of the brewer color set}

\item{row}{data row with <longitude, latitude, label> information}
}
\value{
name   string - a common color set name

title   string - the title to be used in the plot
}
\description{
brewer color schemes names to be used with `sits_plot_dendrogram`

creates a plot title from row information
}
