% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_plot.R
\name{sits_plot_together}
\alias{sits_plot_together}
\title{Plot a set of time series for the same spatio-temporal reference}
\usage{
sits_plot_together(data.tb, colors = "Dark2")
}
\arguments{
\item{data.tb}{tibble - a SITS table with the list of time series to be plotted}

\item{colours}{the color pallete to be used (default is "Set1")}
}
\value{
data.tb    tibble - the input SITS table (useful for chaining functions)
}
\description{
\code{sits_plot} plot one time series for a single place and interval
}
\details{
Plot one SITS time series for a single place and interval
}
\examples{
sits_plot  ts <- (ts, type = "grouped", colors = "Set1")
}
\seealso{
Other STIS main functions: \code{\link{sits_centroids}},
  \code{\link{sits_dendogram}},
  \code{\link{sits_fromJSON}},
  \code{\link{sits_personal_clust}},
  \code{\link{sits_plot}}, \code{\link{sits_toJSON}}
}
\keyword{STIS}

