% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_TWDTW_plot.R
\name{sits_plot_TWDTW_matches}
\alias{sits_plot_TWDTW_matches}
\title{Plot matches between a label pattern and a time series using the dtwSat package}
\usage{
sits_plot_TWDTW_matches(matches, patterns.tb, n_matches = 4)
}
\arguments{
\item{matches}{a list of dtwSat S4 matches objects produced by sits_TWDTW_matches}

\item{patterns.tb}{a set of known temporal signatures for the chosen classes}

\item{n_matches}{number of matches of a given label to be displayed}
}
\description{
Plots the matches from TWDTW classification for one label
}
\author{
Victor Maus, \email{vwmaus1@gmail.com}

Gilberto Camara, \email{gilberto.camara@inpe.br}
}
