% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_machine_learning.R
\name{sits_predict_stack}
\alias{sits_predict_stack}
\title{Predict class based on the trained models}
\usage{
sits_predict_stack(x, model.fit, labels, filename = "", progress = "text",
  parallel = TRUE, ...)
}
\arguments{
\item{x}{a raster \code{\link[raster]{stack}} object with TWDTW distances for 
each class in the layers}

\item{model.fit}{a model trained by \code{\link[sits]{sits_train}}
\code{\link[sits]{sits_TWDTW_matches}}}

\item{labels}{class labels in the order of the raster stack layers}

\item{filename}{a character with the file name. Optional}

\item{progress}{a character. See \code{\link[raster]{pbCreate}}. Default is \code{'text'}}

\item{parallel}{perform parallel processing. Default is TRUE}

\item{...}{other arguments to pass to  \code{\link[raster]{beginCluster}} and 
\code{\link[raster]{writeStart}}}
}
\description{
Given a raster \code{\link[raster]{stack}} object and a model trained 
by \code{\link[sits]{sits_train}}, returns a raster stack object with the classification.
}
\author{
Alexandre Xavier Ywata de Carvalho, \email{alexandre.ywata@ipea.gov.br}

Victor Maus, \email{vwmaus1@gmail.com}

Gilberto Camara, \email{gilberto.camara@inpe.br}
}
