% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_predictors.R
\name{sits_predictors}
\alias{sits_predictors}
\title{Obtain predictors for time series samples}
\usage{
sits_predictors(samples)
}
\arguments{
\item{samples}{Time series in sits format (tibble of class "sits")}
}
\value{
The predictors for the sample: a data.frame with one row per sample.
}
\description{
Predictors are X-Y values required for machine learning
algorithms, organized as a data table where each row corresponds
to a training sample. The first two columns of the predictors table
are categorical (\code{label_id} and \code{label}). The other columns are
the values of each band and time, organized first by band and then by time.
}
\examples{
if (sits_run_examples()) {
    # Include a new machine learning function (naive bayes)

    sits_naive_bayes <- function(samples = NULL){
        train_fun <- function(samples) {
            # Data normalization
            ml_stats <- sits_stats(samples)
            train_samples <- sits_predictors(samples)
            train_samples <- sits_pred_normalize(
                pred = train_samples,
                stats = ml_stats
            )
            # call method and return the trained model
            nb_model <- e1071::naiveBayes(
                x = sits_pred_features(train_samples),
                y = as.factor(sits_pred_references(train_samples))
            )

            # construct model predict closure function and returns
            predict_fun <- function(values) {
                # retrieve the prediction (values and probs)
                prediction <- tibble::as_tibble(
                    stats::predict(nb_model,
                        newdata = values,
                        type = "raw"
                    )
                )
                return(prediction)
            }
            class(predict_fun) <- c("sits_model", "naiveBayes", class(predict_fun))
            return(predict_fun)
        }
        result <- sits_factory_function(samples, train_fun)
        return(result)
    }
    # create an naiveBayes model using a set of samples
    nb_model <- sits_train(samples_modis_ndvi, sits_naive_bayes)
    # classify a point
    point_ndvi <- sits_select(point_mt_6bands, bands = "NDVI")
    point_class <- sits_classify(point_ndvi, nb_model, multicores = 1)
    plot(point_class)
}

}
\seealso{
Other predictor helpers:
\code{\link[=sits_pred_features]{sits_pred_features()}},
\code{\link[=sits_pred_normalize]{sits_pred_normalize()}},
\code{\link[=sits_pred_references]{sits_pred_references()}},
\code{\link[=sits_pred_sample]{sits_pred_sample()}},
\code{\link[=sits_stats]{sits_stats()}}
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
\concept{predictor helpers}
