% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_tibble_functions.R
\name{sits_prune}
\alias{sits_prune}
\title{Prunes dates of time series to fit a timeline}
\usage{
sits_prune(data.tb, timeline)
}
\arguments{
\item{data.tb}{tibble - input SITS table}

\item{timeline}{a timeline of dates for which data is available}
}
\value{
pruned.tb    tibble - the converted SITS table
}
\description{
prunes the times series contained a set of sits tables
to a timeline. This function is useful to constrain the different samples
of land cover to a timeline of available input
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}

Rolf Simoes, \email{rolf.simoes@inpe.br}
}
