% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_machine_learning.R
\name{sits_qda}
\alias{sits_qda}
\title{Train SITS classifiction models with Quadratic Discriminant Analysis}
\usage{
sits_qda(distances.tb = NULL, formula = sits_formula_logref(), ...)
}
\arguments{
\item{distances.tb}{a time series with a set of distance measures for each training sample}

\item{formula}{a symbolic description of the model to be fit. SITS offers a set of such formulas (default: sits_formula_logref)}

\item{...}{other parameters to be passed to MASS::lda function}
}
\value{
result          a model function to be passed in sits_predict
}
\description{
This function receives a tibble with a set of attributes X for each observation Y
These attributes are distance metrics between patterns and observations, obtained by a distance
function in SITS (either sits_distances or sits_TWTDTW_distances).
The function performs a quadratic discriminant analysis (qda) to obtain a predictive model.
This function is a front-end to the "qda" method in the "MASS" package.
Please refer to the documentation in that package for more details.
}
\author{
Alexandre Xavier Ywata de Carvalho, \email{alexandre.ywata@ipea.gov.br}

Rolf Simoes, \email{rolf.simoes@inpe.br}

Gilberto Camara, \email{gilberto.camara@inpe.br}
}
