% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_sample_functions.R
\name{sits_random_sampling}
\alias{sits_random_sampling}
\title{Sampling random points in a data cube}
\usage{
sits_random_sampling(
  cube,
  n_samples = 10000,
  multicores = 2L,
  memsize = 2L,
  progress = TRUE
)
}
\arguments{
\item{cube}{Data cube}

\item{n_samples}{Number of points to be sampled}

\item{multicores}{Number of cores that will be used to
sample the images in parallel.}

\item{memsize}{Memory available for sampling.}

\item{progress}{Show progress bar? Default is TRUE.}
}
\value{
samples              sf objects with sample locations
}
\description{
Takes a random sample of locations in a data cube
}
\examples{
if (sits_run_examples()) {
    # create a data cube from local files
    data_dir <- system.file("extdata/raster/mod13q1", package = "sits")
    cube <- sits_cube(
        source = "BDC",
        collection = "MOD13Q1-6.1",
        data_dir = data_dir
    )
    # sample for data cube
    samples <- sits_random_sampling(
        cube = cube,
        n_samples = 100
    )
}
}
\seealso{
Other training data:
\code{\link[=sits_confidence_sampling]{sits_confidence_sampling()}},
\code{\link[=sits_geo_dist]{sits_geo_dist()}},
\code{\link[=sits_get_data]{sits_get_data()}},
\code{\link[=sits_patterns]{sits_patterns()}},
\code{\link[=sits_reduce_imbalance]{sits_reduce_imbalance()}},
\code{\link[=sits_sample]{sits_sample()}},
\code{\link[=sits_sampling_design]{sits_sampling_design()}},
\code{\link[=sits_stratified_sampling]{sits_stratified_sampling()}},
\code{\link[=sits_uncertainty_sampling]{sits_uncertainty_sampling()}}
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}

Rolf Simoes, \email{rolfsimoes@gmail.com}
}
\concept{training data}
