% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_tibble_functions.R
\name{sits_random_tibble}
\alias{sits_random_tibble}
\title{Simulate a sits_tibble made of random observations}
\usage{
sits_random_tibble(n_samples, label = "label_A", lon_mean = -65,
  lon_sd = 1, lat_mean = -5, lat_sd = 1, date_start = "2000/01/01",
  date_end = "2000/12/31", obs_freq = 23)
}
\arguments{
\item{label}{A character. The labels for returned tibble. The default is "label_A".}

\item{lon_mean}{A length-one numeric. The central longitude of the random samples. The default is 65 degrees est.}

\item{lon_sd}{A length-one numeric. The standard deviation of the longitude of the random samples. The default is 1.}

\item{lat_mean}{A length-one numeric. The central latitude of the random samples. The default is 5 degrees south.}

\item{date_start}{A length-one character.The start date. The default is "2000/01/01".}

\item{date_end}{A length-one character.The start date. The default is "2000/12/31".}

\item{obs_freq}{A length-one numeric. The number of observations in the period. The default are 23.}

\item{samples}{A length-one numeric. The number of samples per label.}

\item{lat_mean}{A length-one numeric. The standard deviation of the latitude of the random samples. The default is 1.}
}
\value{
random_st    A sits_tibble
}
\description{
this function normalizes the time series using the mean and
standard deviation of all the time series.
}
\author{
Alber Sanchez, \email{alber.ipia@inpe.br}
}
