% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_machine_learning.R
\name{sits_ranger}
\alias{sits_ranger}
\title{Train a sits classifiction model using fast random forest algorithm}
\usage{
sits_ranger(data = NULL, num_trees = 2000, importance = "impurity", ...)
}
\arguments{
\item{data}{Time series with the training samples.}

\item{num_trees}{Number of trees to grow. This should not be set
to too small a number,
to ensure that every input row gets predicted
at least a few times. (default: 2000).}

\item{importance}{Variable importance mode, one of 'none',
 'impurity', 'impurity_corrected', 'permutation'.
The 'impurity' measure is the Gini index.}

\item{...}{Other \code{\link[ranger]{ranger}}  parameters}
}
\value{
Model fitted to input data
                    (to be passed to \code{\link[sits]{sits_classify}})
}
\description{
Use Fast Random Forest algorithm to classify data.
This function is a front-end to the "ranger" method in the "ranger" package.
Please refer to the documentation in that package for more details.
}
\examples{
# Retrieve the set of samples for Mato Grosso  (provided by EMBRAPA)
samples_ndvi <- sits_select(samples_mt_6bands, bands = c("NDVI"))

# Build a machine learning model based on deep learning
ml_model <- sits_train(samples_ndvi, sits_ranger(num_trees = 100))

# get a point and classify the point with the ml_model
class <- sits_classify(point_ndvi, ml_model)
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}

Alexandre Ywata de Carvalho, \email{alexandre.ywata@ipea.gov.br}

Rolf Simoes, \email{rolf.simoes@inpe.br}
}
