% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_raster.R
\name{sits_raster}
\alias{sits_raster}
\title{Relabel raster stack object}
\usage{
sits_raster(file = NULL, timeline = NULL, bands = NULL)
}
\arguments{
\item{file}{file name of a raster file}

\item{timeline}{a vector of dates with the timeline of the bands}

\item{bands}{the bands contained in the file}
}
\description{
This function creates a tibble to store the information
about a raster time series
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
