% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_raster.R
\name{sits_raster_tibble}
\alias{sits_raster_tibble}
\title{Relabel raster stack object}
\usage{
sits_raster_tibble()
}
\description{
This function creates a raster tibble
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
