% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_table.R
\name{sits_relabel_conv}
\alias{sits_relabel_conv}
\title{creates a conversion identity list for sits_relabel}
\usage{
sits_relabel_conv(file = NULL)
}
\arguments{
\item{file}{a JSON file contaning the result of a validation procedure}
}
\value{
conv.lst       a conversion list
}
\description{
Given a confusion matrix obtained in the validation
procedure, returns a identity conversion list of labels.
}
\author{
Rolf Simoes, \email{rolf.simoes@inpe.br}
}
