% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_table.R
\name{sits_rename}
\alias{sits_rename}
\title{Finds the names of the bands of a time series}
\usage{
sits_rename(in.tb, bands_new)
}
\arguments{
\item{in.tb}{a SITS table with a list of SITS time series}

\item{bands}{a list of new band names}
}
\value{
out.tb  a SITS table with a list of renamed bands for the time series
}
\description{
\code{sits_renames} renames the bands of a sits table
}
