% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_utils.R
\name{sits_run_examples}
\alias{sits_run_examples}
\title{Informs if sits examples should run}
\usage{
sits_run_examples()
}
\value{
A logical value
}
\description{
This function informs if sits examples should run.
To run the examples, set "SITS_RUN_EXAMPLES" to TRUE using
Sys.setenv("SITS_RUN_EXAMPLES" = TRUE)
To come back to the default behaviour, please set
Sys.setenv("SITS_RUN_EXAMPLES" = FALSE)
}
\seealso{
Other utilities:
\code{\link[=sits_config]{sits_config()}},
\code{\link[=sits_config_show]{sits_config_show()}},
\code{\link[=sits_config_user_file]{sits_config_user_file()}},
\code{\link[=sits_factory_function]{sits_factory_function()}},
\code{\link[=sits_list_collections]{sits_list_collections()}},
\code{\link[=sits_parallel]{sits_parallel()}},
\code{\link[=sits_run_tests]{sits_run_tests()}}
}
\concept{utilities}
