% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_WTSPS.R
\name{sits_runProcess_WTSPS}
\alias{sits_runProcess_WTSPS}
\title{Runs a process in the WTSPS service}
\usage{
sits_runProcess_WTSPS(URL = "http://www.dpi.inpe.br/tws/wtsps", ...)
}
\arguments{
\item{URL}{the URL for the WTSPS service}

\item{...}{algorithm name and its parameters}
}
\description{
uses the WTSPS services to run a process in the server
}
