% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_sample_functions.R
\name{sits_sample.cube}
\alias{sits_sample.cube}
\alias{sits_sample.eo_cube}
\title{Sampling points in a data cube}
\usage{
\method{sits_sample}{eo_cube}(
  data,
  ...,
  npoints = 10000,
  start_date = NULL,
  end_date = NULL,
  bands = NULL,
  impute_fn = impute_linear(),
  multicores = 2L,
  memsize = 2L,
  progress = TRUE
)
}
\arguments{
\item{data}{Data cube}

\item{...}{Specific parameters for method}

\item{npoints}{Number of points to be sampled}

\item{start_date}{Initial date of time series}

\item{end_date}{End date of time series}

\item{bands}{Bands to be included in data cube}

\item{impute_fn}{Imputation function to remove NA.}

\item{multicores}{Number of cores that will be used to
sample the images in parallel.}

\item{memsize}{Memory available for sampling.}

\item{progress}{Show progress bar? Default is TRUE.}
}
\value{
samples              SITS tibble with time series
}
\description{
Takes a class cube with different labels and a sampling
design with a number of samples per class and allocates a set of
locations for each class
}
\examples{
if (sits_run_examples()) {
    # create a data cube from local files
    data_dir <- system.file("extdata/raster/mod13q1", package = "sits")
    cube <- sits_cube(
        source = "BDC",
        collection = "MOD13Q1-6.1",
        data_dir = data_dir
    )
    # sample for data cube
    ts_samples <- sits_sample(
        data = cube,
        npoints = 100
    )
}
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}

Rolf Simoes, \email{rolfsimoes@gmail.com}
}
