% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_sample_functions.R
\name{sits_sample}
\alias{sits_sample}
\title{Sample a time series}
\usage{
sits_sample(data, frac = 0.2, oversample = TRUE)
}
\arguments{
\item{data}{Sits time series tibble}

\item{frac}{Percentage of samples to extract
(range: 0.0 to 2.0, default = 0.2)}

\item{oversample}{Logical: oversample classes with small number of samples?
(TRUE/FALSE)}
}
\value{
A sits tibble with a fixed quantity of samples.
}
\description{
Takes samples from a sits tibble
             returns a new tibble. For a given field as a group criterion,
             this new tibble contains a percentage
             of the total number of samples per group.
             If frac > 1 , all sampling will be done with replacement.
}
\examples{
# Retrieve a set of time series with 2 classes
data(cerrado_2classes)
# Print the labels of the resulting tibble
summary(cerrado_2classes)
# Sample by fraction
data_02 <- sits_sample(cerrado_2classes, frac = 0.2)
# Print the labels
summary(data_02)
}
\seealso{
Other training data:
\code{\link[=sits_confidence_sampling]{sits_confidence_sampling()}},
\code{\link[=sits_geo_dist]{sits_geo_dist()}},
\code{\link[=sits_get_data]{sits_get_data()}},
\code{\link[=sits_patterns]{sits_patterns()}},
\code{\link[=sits_random_sampling]{sits_random_sampling()}},
\code{\link[=sits_reduce_imbalance]{sits_reduce_imbalance()}},
\code{\link[=sits_sampling_design]{sits_sampling_design()}},
\code{\link[=sits_stratified_sampling]{sits_stratified_sampling()}},
\code{\link[=sits_uncertainty_sampling]{sits_uncertainty_sampling()}}
}
\author{
Rolf Simoes, \email{rolfsimoes@gmail.com}
}
\concept{training data}
