% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_tibble_functions.R
\name{sits_sample_shp}
\alias{sits_sample_shp}
\title{Get random samples from a polygon sf object}
\usage{
sits_sample_shp(shp.sf, label_field, nsamples = 500, border_offset = 50,
  proj_crs = 29101, min_area = 50000)
}
\arguments{
\item{shp.sf}{a polygon sf object.}

\item{label_field}{a length-one character. The name of the label field in shp.sf}

\item{nsamples}{a length-one numeric. The expected number of samples per class.}

\item{border_offset}{a length-one numeric. An offset distance (in meters) to avoid sapling near the borders}

\item{proj_crs}{a length-one character. The name of an intermediary CRS used when the shp.sf is given in longitude and latitude.}

\item{min_area}{a length-one numeric. The minimum area (in square meters) of a polygon included during the sampling.}
}
\value{
shp_samples  a sits_tibble
}
\description{
This function takes raonm samples of the polygons given as a
sf object.
}
\examples{
nc <- sf::st_read(system.file("shape/nc.shp", package="sf"))
pnt_samples <- sits_sample_shp(shp.sf = nc, label_field = "NAME",
                               nsamples = 500, border_offset = 0.01)

}
\author{
Alber Sanchez, \email{alber.ipia@inpe.br}
}
