% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_save.R
\name{sits_save}
\alias{sits_save}
\title{Save the contents of a SITS table in a JSON file}
\usage{
sits_save(source, file_name = NULL)
}
\arguments{
\item{source}{tibble - an existing SITS table}

\item{file_name}{string - file name where compressed data will be saved}
}
\value{
table       the same table (for chaining functions)
}
\description{
stores a SITS table in a compressed JSON file
The set of time series from a SITS table can be saved in a compressed JSON format
and later retrieved for further use
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
