% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_indexes.R
\name{sits_savi}
\alias{sits_savi}
\title{Builds soil-adjusted vegetation index}
\usage{
sits_savi(data.tb)
}
\arguments{
\item{data.tb}{Valid sits tibble.}
}
\value{
A sits tibble with the SAVI index.
}
\description{
Adds new tasseled cap bands.
}
\examples{
\donttest{
# Retrieve data for time series with label samples in Mato Grosso in Brazil
data (samples_MT_9classes)
# Generate a new image with the tasseled cap
savi.tb <- sits_savi(samples_MT_9classes)
}
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
