% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_segmentation.R
\name{sits_segmentation}
\alias{sits_segmentation}
\alias{sits_segment}
\title{Segment an image}
\usage{
sits_segment(cube, tiles, bands, date, seg_fn, ...)
}
\arguments{
\item{cube}{Regular data cube}

\item{tiles}{Tiles to be segmented}

\item{bands}{Bands to include in the segmentation}

\item{date}{Date to select the image to be segmented}

\item{seg_fn}{Function to apply the segmentation}

\item{...}{Other params to be passed to segmentation function}
}
\description{
Apply a segmentation on a data cube based on the "supercells" package.
This is an adaptation and extension to remote sensing data of the
SLIC superpixels algorithm proposed by Achanta et al. (2012).
See references for more details.
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}

Rolf Simoes, \email{rolf.simoes@inpe.br}

Felipe Carvalho, \email{felipe.carvalho@inpe.br}
}
