% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_data_generics.R
\name{sits_select.cube}
\alias{sits_select.cube}
\title{Filter bands on a data cube}
\usage{
\method{sits_select}{cube}(data, bands)
}
\arguments{
\item{data}{data cube}

\item{bands}{vector with the names of the bands}
}
\description{
Returns a data cube with the selected bands.
}
\author{
Rolf Simoes, \email{rolf.simoes@inpe.br}
}
