% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_data_generics.R
\name{sits_select.patterns}
\alias{sits_select.patterns}
\title{Filter bands on a data set (tibble or cube)}
\usage{
\method{sits_select}{patterns}(data, bands)
}
\arguments{
\item{data}{A sits tibble metadata and data on time series.}

\item{bands}{Character vector with the names of the bands}
}
\value{
A tibble in sits format with the selected bands.
}
\description{
Returns a sits tibble with the selected bands.
}
\author{
Rolf Simoes, \email{rolf.simoes@inpe.br}
}
