% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_tibble_functions.R
\name{sits_select_bands_}
\alias{sits_select_bands_}
\title{Filter bands on a sits tibble}
\usage{
sits_select_bands_(data.tb, bands)
}
\arguments{
\item{data.tb}{A sits tibble metadata and data on time series.}

\item{bands}{The selcted bands.}
}
\value{
A tibble in sits format with the selected bands.
}
\description{
Returns a sits tibble with the selected bands.
}
\examples{
# Retrieve a set of time series with 2 classes
data(cerrado_2classes)
# Print the original bands
sits_bands(cerrado_2classes)
# Select only the "ndvi" band
band_ndvi <- "ndvi"
data.tb <- sits_select_bands_(cerrado_2classes, bands = band_ndvi)
# Print the labels of the resulting table
sits_bands(data.tb)
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
