% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_separability.R
\name{sits_separability_experiments}
\alias{sits_separability_experiments}
\title{Do a set of separability experiments based on a cross combination of all elements of its input parameters.}
\usage{
sits_separability_experiments(data.tb, method = "kohonen-dendogram",
  times = 1, bands = list("evi", "ndvi", c("evi", "ndvi")),
  n_clusters = 5, grouping_method = c("average", "ward.D2", "complete",
  "single"), koh_grid_length = 7, koh_rlen = 1200, koh_alpha_from = 0.2,
  koh_alpha_to = 0.04, .multiproc = 1)
}
\arguments{
\item{data.tb}{a SITS table with the samples to be used in experiments.}

\item{method}{string - either 'dendogram', 'centroids', 'kohonen', or 'kohonen-dendogram'.}

\item{times}{how many times to run an parameterized experiment.}

\item{bands}{the bands to be clusterized.}

\item{n_clusters}{the number of clusters to be croped from hierarchical clustering (ignored in `kohonen` method). Default is 2.}

\item{grouping_method}{the agglomeration method to be used. Any `hclust` method (see `hclust`) (ignored in `kohonen` method). Default is 'ward.D2'.}

\item{koh_grid_length}{grid side dimension of the SOM grid (used only in `kohonen` or `kohonen-dendogram` methods). Defaul is 5.}

\item{koh_rlen}{the number of times the complete data set will be presented to the SOM grid
(used only in `kohonen` or `kohonen-denddogram` methods). Default is 100.}

\item{koh_alpha_from}{starting learning rate. Default is to decline linearly from 0.05 to `koh_alpha_to` over rlen updates.}

\item{koh_alpha_to}{ending learning rate. Default is to decline linearly from `koh_alpha_from` to 1.0 over rlen updates.}

\item{.multiproc}{(Linux only) numbers of cores to be used in multiprocessing.}
}
\value{
result.tb      a tibble with all cross combinations of parameters' values.
}
\description{
Runs a set of separability experiments according to informed parameters and returns a tibble with experiment's result.
}
\author{
Rolf Simoes, \email{rolf.simoes@inpe.br}
}
