% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_separability.R
\name{sits_separability_measure}
\alias{sits_separability_measure}
\title{computes a separability measure from a clusterized SITS table data.}
\usage{
sits_separability_measure(data.tb, measure = "entropy", per_cluster = FALSE)
}
\arguments{
\item{data.tb}{a SITS table with the samples to be validated.}

\item{measure}{an separability measure. For now, only "entropy" is implemented.}

\item{per_cluster}{(boolean) should return a total average separability measure, or a per cluster measure?}
}
\value{
result         a separability measure.
}
\description{
Computes an measure of seggregation from SITS table based on mean relative Shannon Entropy,
where maximum entropy is 1.0 and minimum entropy is 0.0, from each cluster label against
original_label occorrences in it. To obtain this measure, first we compute Shannon
information entropy from counts of each cluster label. To get the relative entropy, we
divide this value by the maximum entropy. After that, we compute the weighted mean relative
to the frequency of each cluster label. The final values represents an average of
maximum entropy fraction. This is useful to assess the separability of samples for a given clustering algorithm.
}
\author{
Rolf Simoes, \email{rolf.simoes@inpe.br}
}
