% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_filters.R
\name{sits_sgolay}
\alias{sits_sgolay}
\title{Smooth the time series using Savitsky-Golay filter (based on signal package)}
\usage{
sits_sgolay(data.tb, order = 3, scale = 1, bands_suffix = "sg")
}
\arguments{
\item{data.tb}{The SITS tibble containing the original time series}

\item{order}{filter order}

\item{scale}{time scaling}

\item{bands_suffix}{the suffix to be appended to the smoothed filters}
}
\value{
output.tb a tibble with smoothed sits time series
}
\description{
The algorithm searches for an optimal polynomial describing the warping.
The degree of smoothing depends on smoothing factor lambda (usually from 0.5 to 10.0)
Use lambda = 0.5 for very slight smoothing and lambda = 5.0 for strong smoothing
}
