% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_config.R
\name{sits_show_config}
\alias{sits_show_config}
\title{Shows the contents of the SITS configuration file}
\usage{
sits_show_config()
}
\value{
config_sits  list with the configuration parameters used by SITS
}
\description{
Displays the contents of SITS configuration file. For more details
on how to set the configuration file, please use \code{\link[sits]{sits_config}}.
}
\examples{
config_sits <- sits_config()
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
