% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_utils.R
\name{sits_show_prediction}
\alias{sits_show_prediction}
\title{Shows the predicted labels for a classified tibble}
\usage{
sits_show_prediction(class)
}
\arguments{
\item{class}{A SITS tibble that has been classified.}
}
\value{
Tibble with the columns "from", "to", "class"
}
\description{
This function takes a tibble with a classified time series
by a machine learning method and displays the result.
}
\examples{
if (sits_run_examples()) {
    # Retrieve the samples for Mato Grosso
    # train an SVM model
    ml_model <- sits_train(samples_modis_ndvi, ml_method = sits_svm)
    # classify the point
    point_ndvi <- sits_select(point_mt_6bands, bands = "NDVI")
    point_class <- sits_classify(
        data = point_ndvi, ml_model = ml_model
    )
    sits_show_prediction(point_class)
}

}
\seealso{
Other classification:
\code{\link[=sits_classify]{sits_classify()}},
\code{\link[=sits_clean]{sits_clean()}},
\code{\link[=sits_combine_predictions]{sits_combine_predictions()}},
\code{\link[=sits_get_class]{sits_get_class()}},
\code{\link[=sits_get_probs]{sits_get_probs()}},
\code{\link[=sits_label_classification]{sits_label_classification()}},
\code{\link[=sits_reclassify]{sits_reclassify()}},
\code{\link[=sits_smooth]{sits_smooth()}},
\code{\link[=sits_uncertainty]{sits_uncertainty()}},
\code{\link[=sits_variance]{sits_variance()}}
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
\concept{classification}
\keyword{internal}
