% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_csv.R
\name{sits_shp_to_csv}
\alias{sits_shp_to_csv}
\title{Export a shapefile with points to a CSV file for later processing}
\usage{
sits_shp_to_csv(shpfile, csvfile, label, timeline, start_date, end_date,
  interval = "12 month")
}
\arguments{
\item{shpfile}{A point shapefile.}

\item{csvfile}{The name of the exported CSV file.}

\item{label}{Label associated to the samples.}

\item{timeline}{The timeline of the data set.}

\item{start_date}{Starting date for which the samples are valid.}

\item{end_date}{End date for which the samples are valid.}

\item{interval}{Interval between two samples of the same place.}
}
\value{
Status of the operation.
}
\description{
Converts points from a shapefile to a CSV file. The CSV file will not contain the actual time
             series. Its columns will be the same as those of a CSV file used to retrieve data from
             ground information ("latitude", "longitude", "start_date", "end_date", "label").
}
\examples{
\donttest{
# select the cube
wtss_cube <- sits_cube(service = "WTSS", name    = "MOD13Q1")
#  get the timeline from the cube
cube_timeline <- sits_timeline(wtss_cube)
# define the input shapefile (consisting of POINTS)
shpfile <- system.file("extdata/shapefiles/cerrado_forested.shp", package = "sits")
# set the start and end dates for the validity of the labels of the points in the shapefile
start_date <- lubridate::ymd("2002-08-29")
end_date   <- lubridate::ymd("2013-08-13")
# define the output csv file
csvfile <- paste0("cerrado_forested.csv")
#' # define the label
label <- "Cerrado_Forested"
# read the points in the shapefile and produce a CSV file
sits_shp_to_csv(shpfile, csvfile, label, cube_timeline, start_date, end_date, interval = "12 month")
# read the first three samples from the CSV file
csv_data <- sits_get_data(cube = wtss_cube, file = csvfile, .n_max = 3)
csv_data
}
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
