% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_csv.R
\name{sits_shp_toCSV}
\alias{sits_shp_toCSV}
\title{Export a shapefile with points to a CSV file for later processing}
\usage{
sits_shp_toCSV(shpfile, csvfile, label, timeline, start_date, end_date,
  interval = "12 month", n_samples = 500, min_area = 100, min_dist = 50,
  border_offset = 50, is_density = FALSE)
}
\arguments{
\item{shpfile}{a POINT or POLYGON shapefile}

\item{csvfile}{the name of the exported CSV file}

\item{label}{it can be either the label associated to the samples (POINT) or the name of a field in the shapefile (POLYGON)}

\item{timeline}{the timeline of the data set}

\item{start_date}{starting date for which the samples are valid}

\item{end_date}{end date for which the samples are valid}

\item{interval}{interval between two samples of the same place}

\item{n_samples}{A length-one numeric. The number of samples requested. If is_density = TRUE, nsamples is the number of samples per unit of area. The default is 500.}

\item{min_area}{A length-one numeric. The minimum area of a sampled polygon. The default is 100.}

\item{min_dist}{A length-one numeric. The minimum disatnces between samples. The default is 50.}

\item{border_offset}{A length-one numeric. The minimum distance from the samples to the polygon's borders. The default is 50.}

\item{is_density}{A length-one logical. Is this a density sampling? The dafault is FALSE}
}
\value{
status     the status of the operation
}
\description{
Converts points from a shapefile to a CSV file. The CSV file will not contain the actual time
             series. Its columns will be the same as those of a CSV file used to retrieve data from
             ground information ("latitude", "longitude", "start_date", "end_date", "coverage", "label").
}
\examples{
\donttest{
# set the timeline
data("timeline_2000_2017")
# set the start and end dates
start_date <- lubridate::ymd("2002-08-29")
end_date   <- lubridate::ymd("2013-08-13")
# define the input shapefile
shpfile <- system.file ("extdata/shapefiles/cerrado_forested.shp", package = "sits")
# define the output csv file
csvfile <- paste0("./cerrado_forested.csv")
# define the label
label <- "Cerrado_Forested"
# read the points in the shapefile and produce a CSV file
sits_shp_toCSV(shpfile, csvfile, label, timeline_2000_2017, start_date, end_date)
}
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
