% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_table.R
\name{sits_significant_labels}
\alias{sits_significant_labels}
\title{Get the significant labeled time series among all others labels.}
\usage{
sits_significant_labels(data.tb, min_label_frac = 0.05)
}
\arguments{
\item{data.tb}{a SITS table with the data to be extracted}

\item{min_label_frac}{a decimal between 0 and 1. The minimum percentagem of valid cluster members,
with reference to the total number of samples.}
}
\value{
result.tb      a SITS table with the filtered data.
}
\description{
Given a SITS table, select only those whose time series
exceed a minimum percentage
}
\author{
Rolf Simoes, \email{rolf.simoes@inpe.br}
}
