% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_smooth.R
\name{sits_smooth.bayes}
\alias{sits_smooth.bayes}
\title{Post-process a classified data raster probs using Bayesian smoothing}
\usage{
\method{sits_smooth}{bayes}(
  cube,
  type = "bayes",
  ...,
  window_size = 5,
  smoothness = 20,
  covar = FALSE,
  multicores = 1,
  memsize = 1,
  output_dir = getwd(),
  version = "v1"
)
}
\arguments{
\item{cube}{Probability data cube}

\item{type}{Type of smoothing}

\item{...}{Parameters for specific functions}

\item{window_size}{Size of the neighbourhood.}

\item{smoothness}{Estimated variance of logit of class probabilities
(Bayesian smoothing parameter). It can be either
a matrix or a scalar.}

\item{covar}{a logical argument indicating if a covariance
matrix must be computed as the prior covariance.}

\item{multicores}{Number of process to run the Bayesian smoothing in
snow subprocess.}

\item{memsize}{Maximum overall memory (in GB) to run the Bayesian
smoothing.}

\item{output_dir}{Output directory where to out the file}

\item{version}{Version of resulting image
(in the case of multiple tests)}
}
\value{
A tibble with metadata about the output raster objects.
}
\description{
Takes a set of raster bricks with probabilities,
             whose metadata is created by \code{\link[sits]{sits_cube}},
             and apply a bayesian smoothing process.
             If \code{covar} is \code{FALSE} only main diagonal of
             covariance prior matrix will be computed.
}
\examples{
\dontrun{
# Retrieve the samples for Mato Grosso
# select band "ndvi"

samples_ndvi <- sits_select(samples_mt_4bands, bands = "NDVI")

# select a random forest model
rfor_model <- sits_train(samples_ndvi, sits_rfor(num_trees = 500))

# create a data cube based on files
data_dir <- system.file("extdata/raster/mod13q1", package = "sits")
cube <- sits_cube(
    source = "LOCAL",
    name = "sinop-2014",
    satellite = "TERRA",
    sensor = "MODIS",
    data_dir = data_dir,
    delim = "_",
    parse_info = c("X1", "X2", "band", "date")
)

# classify the raster image
probs_cube <- sits_classify(cube,
    ml_model = rfor_model,
    output_dir = tempdir(),
    memsize = 4, multicores = 2
)

# label the classification and smooth the result with a Bayesian filter
bayes_cube <- sits_smooth(probs_cube, output_dir = tempdir())
}

}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
