% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_smooth.R
\name{sits_smooth.bilinear}
\alias{sits_smooth.bilinear}
\title{Post-process a classified data raster probs using bilinear smoothing}
\usage{
\method{sits_smooth}{bilinear}(
  cube,
  type = "bilinear",
  ...,
  window_size = 5,
  sigma = 1,
  tau = 0.25,
  multicores = 1,
  memsize = 1,
  output_dir = getwd(),
  version = "v1"
)
}
\arguments{
\item{cube}{Probability data cube}

\item{type}{Type of smoothing}

\item{...}{Parameters for specific functions}

\item{window_size}{Size of the neighbourhood.}

\item{sigma}{Standard deviation of the spatial gaussian kernel}

\item{tau}{Standard deviation of the class probs value}

\item{multicores}{Number of process to run the Bayesian smoothing in
snow subprocess.}

\item{memsize}{Maximum overall memory (in GB) to run the Bayesian
smoothing.}

\item{output_dir}{Output directory}

\item{version}{Version of resulting image
(in the case of multiple tests)}
}
\value{
A tibble with metadata about the output raster objects.
}
\description{
Takes a set of classified raster layers with probabilities,
             whose metadata is]created by \code{\link[sits]{sits_cube}},
             and apply bilinear smoothing process.
}
\examples{
\dontrun{
# Retrieve the samples for Mato Grosso
# select band "ndvi"

samples_ndvi <- sits_select(samples_mt_4bands, bands = "NDVI")

# select a random forest model
rfor_model <- sits_train(samples_ndvi, sits_rfor(num_trees = 500))

# create a data cube based on files
data_dir <- system.file("extdata/raster/mod13q1", package = "sits")
cube <- sits_cube(
    type = "STACK",
    name = "sinop-2014",
    satellite = "TERRA",
    sensor = "MODIS",
    data_dir = data_dir,
    delim = "_",
    parse_info = c("X1", "X2", "band", "date")
)

# classify the raster image
probs_cube <- sits_classify(cube,
    ml_model = rfor_model,
    output_dir = tempdir(),
    memsize = 4, multicores = 2
)

# smooth the result with a bilinear filter
bil_cube <- sits_smooth(probs_cube,
    type = "bilinear",
    output_dir = tempdir()
    )
}

}
\references{
K. Schindler, "An Overview and Comparison of Smooth Labeling
            Methods for Land-Cover Classification",
            IEEE Transactions on Geoscience and Remote Sensing,
            50 (11), 4534-4545, 2012.
}
\author{
@author Gilberto Camara, \email{gilberto.camara@inpe.br}
}
