% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_som.R
\name{sits_som_cluster}
\alias{sits_som_cluster}
\title{Clustering a set of satellite image time series using SOM}
\usage{
sits_som_cluster(
  data,
  grid_xdim = 10,
  grid_ydim = 10,
  alpha = 1,
  distance = "euclidean",
  rlen = 100,
  prior_threshold = 0.6,
  posterior_threshold = 0.6,
  som_radius = 2
)
}
\arguments{
\item{data}{A tibble with samples to be clustered.}

\item{grid_xdim}{X dimension of the SOM grid (default = 25).}

\item{grid_ydim}{Y dimension of the SOM grid.}

\item{alpha}{Starting learning rate,
which decreases according to number of iterations.}

\item{distance}{The similarity measure (distance).}

\item{rlen}{How many times dataset will be presented to the SOM.}

\item{prior_threshold}{Threshold of priot probability
(frequency of samples assigned to a same SOM neuron)}

\item{posterior_threshold}{Threshold of posterior probability
(influenced by the SOM neighborhood)}

\item{som_radius}{Radius of neighborhood on the SOM map
(controls the size of the neighbourhood)}
}
\value{
A sits tibble with an evaluation column indicating if
                     each samples is clean, should be analyzed or
                     should be removed, and with a new column indicating
                     the posterior probability of the sample
}
\description{
This function uses self-organized maps to find clusters in
satellite image time series for quality control of the  samples.
Calls \code{\link[sits]{sits_som_map}} to generate the som map and
\code{\link[sits]{sits_som_clean_samples}} to produce a clean set of samples.
The parameters "grid_xdim", "grid_ydim", "rlen", "distance", "alpha", and
"iterations" are used by \code{\link[sits]{sits_som_map}} to control
how the Kohonen map is generated.
The parameters "prior_threshold" and "posterior_threshold" control
how the good quality samples are selected, based on the Kohonen map.
}
\examples{
\dontrun{
# Evaluate the quality of the samples using SOM clustering
new_samples <- sits_som_cluster(samples_mt_4bands)
}
}
\references{
`kohonen` package (https://CRAN.R-project.org/package=kohonen)
}
\author{
Lorena Alves, \email{lorena.santos@inpe.br}

Karine Ferreira. \email{karine.ferreira@inpe.br}
}
