% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_kohonen.R
\name{sits_som_cluster_neighbourhood}
\alias{sits_som_cluster_neighbourhood}
\alias{.sits_som_cluster_neighbourhood}
\title{SOM neighbourhood}
\usage{
.sits_som_cluster_neighbourhood(som_cluster)
}
\arguments{
\item{som_cluster}{Kohonen map and associated information created by
\code{\link[sits]{sits_som_map}}}
}
\value{
Returns a sits tibble with a overall percentage
         of the majority neighbourhood of each sample of all iterations.
}
\description{
This fuction presents the overall statistics of neigbourhood for each sample
}
\author{
Lorena Santos, \email{lorena.santos@inpe.br}
}
