% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_som.R
\name{sits_som_evaluate_cluster}
\alias{sits_som_evaluate_cluster}
\title{Evaluate cluster}
\usage{
sits_som_evaluate_cluster(som_map)
}
\arguments{
\item{som_map}{A SOM map produced by the som_map() function}
}
\value{
A tibble stating the purity for each cluster
}
\description{
\code{sits_som_evaluate_cluster()} produces a tibble with the clusters
found by the SOM map. For each cluster, it provides the percentage
of classes inside it.
}
\examples{
if (sits_run_examples()) {
    # create a som map
    som_map <- sits_som_map(samples_modis_ndvi)
    # plot the som map
    plot(som_map)
    # evaluate the som map and create clusters
    clusters_som <- sits_som_evaluate_cluster(som_map)
    # plot the cluster evaluation
    plot(clusters_som)
    # clean the samples
    new_samples <- sits_som_clean_samples(som_map)
}
}
\seealso{
Other clustering and SOM:
\code{\link[=sits_cluster_clean]{sits_cluster_clean()}},
\code{\link[=sits_cluster_dendro]{sits_cluster_dendro()}},
\code{\link[=sits_cluster_frequency]{sits_cluster_frequency()}},
\code{\link[=sits_som_clean_samples]{sits_som_clean_samples()}},
\code{\link[=sits_som_map]{sits_som_map()}},
\code{\link[=sits_som_remove_samples]{sits_som_remove_samples()}}
}
\author{
Lorena Alves, \email{lorena.santos@inpe.br}

Karine Ferreira. \email{karine.ferreira@inpe.br}

Gilberto Camara, \email{gilberto.camara@inpe.br}
}
\concept{clustering and SOM}
