% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_kohonen.R
\name{sits_som_evaluate_subgroups}
\alias{sits_som_evaluate_subgroups}
\title{Create new groups from kohonen maps}
\usage{
sits_som_evaluate_subgroups(som_cluster)
}
\arguments{
\item{som_cluster}{Kohonen map and associated information created by \code{\link[sits]{sits_som_map}}}
}
\value{
Returns a sits tibble with a new column presenting the subgroups that each sample belongs.
}
\description{
Create subgroups from a cluster created by SOM.
}
\examples{
\donttest{
# Create a Kohonen map
som_cluster <- sits_som_map(prodes_226_064)
# Create subgroups
subgroups.tb <- sits_som_evaluate_subgroups(som_cluster)
}
}
\author{
Lorena Santos, \email{lorena.santos@inpe.br}
}
