% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_kohonen.R
\name{sits_som_plot_subgroups}
\alias{sits_som_plot_subgroups}
\title{Plot the patterns of SOM subgroups (only NDVI and EVI time series)}
\usage{
sits_som_plot_subgroups(subgroups, label, band = "ndvi")
}
\arguments{
\item{subgroups}{The list contain the EVI and NDVI time series (weight of each neuron) by class.}

\item{label}{Label to be plotted}

\item{band}{Band to be plotted}
}
\description{
Plot the average pattern of subgroups from  weights of neurons labelled.
Each neuron has a weight which can be represent a set of time series samples. Neurons of same
category or label form a cluster, but the neurons within a cluster can have different patterns.
}
\examples{
\donttest{
# Produce a cluster map
som_cluster <- sits_som_map(prodes_226_064)
# Evaluate and produce information on the subgroups
subgroups <- sits_som_evaluate_subgroups(som_cluster)
# Plot the resulting subgroups
sits_som_plot_subgroups(subgroups, cluster = "Forest", band = "ndvi")
}

}
\author{
Lorena Santos \email{lorena.santos@inpe.br}
}
