% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_table.R
\name{sits_spread_matches}
\alias{sits_spread_matches}
\title{Spread matches from a sits matches tibble}
\usage{
sits_spread_matches(data.tb)
}
\arguments{
\item{data.tb}{a SITS matches tibble}
}
\value{
a tibble where whose columns have the reference label and the TWDTW distances for each temporal pattern
}
\description{
Given a SITS tibble with a matches, returns a tibble whose columns have
the reference label and the TWDTW distances for each temporal pattern.
}
\author{
Victor Maus, \email{vwmaus1@gmail.com}

Gilberto Camara, \email{gilberto.camara@inpe.br}

Alexandre Xavier Ywata de Carvalho, \email{alexandre.ywata@ipea.gov.br}

Rolf Simoes, \email{rolf.simoes@inpe.br}
}
