% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_distances.R
\name{sits_spread_time_series}
\alias{sits_spread_time_series}
\title{Use time series values from a sits tibble as distances for training patterns}
\usage{
sits_spread_time_series(data.tb = NULL)
}
\arguments{
\item{data.tb}{a SITS tibble with original data}
}
\value{
distances.tb  a tibble where columns have the reference label and the time series values as distances
}
\description{
This function allows using a set of labelled time series as
input to the machine learning models. Instead of first estimating a set
of idealised patterns and then computing distances from these patterns,
the attributes used to train the model are the series themselves.
This function then extracts the time series from a SITS tibble and
"spreads" them in time to produce a tibble with distances.
}
\author{
Rolf Simoes, \email{rolf.simoes@inpe.br}
}
