% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_distances.R
\name{sits_spread_ts_as_distance}
\alias{sits_spread_ts_as_distance}
\title{Spread time series values from a sits tibble as distances in a sits distance tibble}
\usage{
sits_spread_ts_as_distance(data.tb = NULL)
}
\arguments{
\item{data.tb}{a SITS tibble}
}
\value{
result.tb  a tibble where columns have the reference label and the time series bands as distances
}
\description{
Given a SITS tibble with a set of time series values, returns a tibble whose columns have
the reference label and the time series bands as distances to be used as input in Machine Learning functions.
}
\author{
Rolf Simoes, \email{rolf.simoes@inpe.br}
}
