% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_stack_proc.R
\name{sits_stack_relabel}
\alias{sits_stack_relabel}
\title{Relabel raster stack object}
\usage{
sits_stack_relabel(x, old_values, new_values, filename = "",
  progress = "text", parallel = TRUE, ...)
}
\arguments{
\item{x}{raster stack .... tbl raster or wms???}

\item{old_values}{an integer vector with raster values (label) to be updated}

\item{new_values}{an integer vector with new values (label). It must have the same length as \code{old_label}}

\item{filename}{a character with the file name. Optional}

\item{progress}{a character. See \code{\link[raster]{pbCreate}}. Default is \code{'text'}}

\item{parallel}{perform parallel processing. Default is TRUE}

\item{...}{other arguments to pass to  \code{\link[raster]{beginCluster}} and \code{\link[raster]{writeStart}}}
}
\description{
This function updates categorical raster values (labels).
}
\author{
Victor Maus, \email{vwmaus1@gmail.com}
}
