% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_stack_proc.R
\name{sits_stack_transition_relabel}
\alias{sits_stack_transition_relabel}
\title{Relabel raster stack based on temporal transitions}
\usage{
sits_stack_transition_relabel(x, A, filename = "", progress = "text",
  parallel = TRUE, ...)
}
\arguments{
\item{x}{raster stack object}

\item{A}{transition matrix. This gives the possible transitions among classes.
The first line and the first column of the table has the raster class followed
by the class label, such that <class_number>.<class_label>. The rows are the
Class(t) and columns Class(t-1), the cobination between a row (Class(t)) and a
column (Class(t-1)) gives the updated class number.}

\item{filename}{a character with the file name. Optional}

\item{progress}{a character. See \code{\link[raster]{pbCreate}}. Default is \code{'text'}}

\item{parallel}{perform parallel processing. Default is TRUE}

\item{...}{other arguments to pass to \code{\link[raster]{beginCluster}} and \code{\link[raster]{writeStart}}}
}
\description{
This function updates categorical raster values (labels) based on the temporal transitions,
assuming Class(t) = f(Class(t-1), Class(t)) subject to the initial condition Class(t=1) = Class(t=1)
}
\author{
Victor Maus, \email{vwmaus1@gmail.com}
}
