% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_WTSPS.R
\name{sits_statusProcess_WTSPS}
\alias{sits_statusProcess_WTSPS}
\title{Provides information about a process running in the WTSPS service}
\usage{
sits_statusProcess_WTSPS(URL = "http://www.dpi.inpe.br/tws/wtsps",
  uuid = NULL)
}
\arguments{
\item{URL}{the URL for the WTSPS service}

\item{uuid}{the id of the process}
}
\description{
uses the WTSPS services to get the status of a running process in the server
}
