% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_raster.R
\name{sits_STRaster}
\alias{sits_STRaster}
\title{Create a metadata tibble to store the description of a spatio-temporal raster dataset}
\usage{
sits_STRaster(files, timeline, bands, scale_factors)
}
\arguments{
\item{files}{Vector with the file paths of the raster files}

\item{timeline}{Vector of dates with the timeline of the bands}

\item{bands}{The bands contained in the Raster Brick set (in the same order as the files)}

\item{scale_factors}{Scale factors to convert each band to [0..1] range (in the same order as the files)}
}
\value{
raster.tb   A tibble with metadata information about a raster data set
}
\description{
This function creates a tibble containing the metadata for
              a set of spatio-temporal raster files, organized as a set of "Raster Bricks".
              These files should be of the same size and
              projection. Each raster brick file should contain one band
              per time step. Different bands are archived in different raster files.
}
\examples{
# read a raster file and put it into a vector
files  <- c(system.file ("extdata/raster/mod13q1/sinop-crop-ndvi.tif", package = "sits"))

# define the timeline
data(timeline_mod13q1)
timeline <- lubridate::as_date (timeline_mod13q1$V1)
# define the bands
bands <- c("ndvi")
# define the scale factor
scale_factors <-  c(0.0001)

# create a raster metadata file based on the information about the files
raster.tb <- sits_STRaster (files, timeline, bands, scale_factors)

}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
