% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_kohonen.R
\name{sits_subgroup}
\alias{sits_subgroup}
\title{Create new groups from kohonen maps}
\usage{
sits_subgroup(koh)
}
\arguments{
\item{koh}{An object with informations about kohonen clustering and samples.}
}
\value{
Returns a sits tibble with subgroups generated by hierarchical clustering.
}
\description{
Create new groups to identify variations in a same group.
}
\examples{
\donttest{
data(samples_mt_6bands)
# generate a SOM map to find clusters of samples
koh <- sits_kohonen(samples_mt_6bands, bands = NULL, grid_xdim = 15, grid_ydim = 15)
# Divide groups according to variations
subgroups <- sits_subgroup(koh)
samples_subgroup <- subgroups$samples_subgroup.tb
neurons_subgroup <- subgroups$neurons_subgroup.lst
number_of_subgroup <- lengths(neurons_subgroup)
}
}
\author{
Lorena Santos, \email{lorena.santos@inpe.br}
}
