% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_table.R
\name{sits_summary}
\alias{sits_summary}
\title{returns the labels' count of a sits table}
\usage{
sits_summary(data.tb)
}
\arguments{
\item{data.tb}{a valid sits table}
}
\value{
result.tb  a tibble with the names of the labels and its absolute and relative frequency
}
\description{
returns the labels and its respective counting and frequency.
}
\author{
Rolf Simoes, \email{rolf.simoes@inpe.br}
}
