% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_segmentation.R
\name{sits_supercells}
\alias{sits_supercells}
\title{Segment an image using supercells}
\usage{
sits_supercells(
  cube,
  tiles = cube[1, "tile"],
  bands,
  date,
  step = 50,
  compactness = 1,
  iter = 10,
  minarea = 30,
  chunks = TRUE,
  future = FALSE,
  multicores = 4
)
}
\arguments{
\item{cube}{Regular data cube}

\item{tiles}{Tiles to be segmented}

\item{bands}{Bands to include in the segmentation}

\item{date}{Date to select the image to be segmented}

\item{step}{Distance (in number of cells) between initial supercells' centers.}

\item{compactness}{A compactness value. Larger values cause clusters to
be more compact/even (square).}

\item{iter}{Number of iterations to create the output.}

\item{minarea}{Specifies the minimal size of a supercell (in cells).}

\item{chunks}{Should the input (x) be split into chunks before deriving
supercells? Either FALSE,
TRUE (default - only large input objects are split),
or a numeric value (representing the side length of the chunk
in the number of cells).}

\item{future}{Should the future package be used for parallelization
of the calculations?}

\item{multicores}{Number of cores for parallel processing}
}
\description{
Apply a segmentation on a data cube based on the "supercells" package. This is an
adaptation and extention to remote sensing data of the SLIC superpixels
algorithm proposed by Achanta et al. (2012). See references for more details.
}
\examples{
# example code
if (sits_run_examples()) {
    # Example of classification of a data cube
    # create a data cube from local files
    data_dir <- system.file("extdata/raster/mod13q1", package = "sits")
    cube <- sits_cube(
        source = "BDC",
        collection = "MOD13Q1-6",
        data_dir = data_dir
    )
    date <- sits_timeline(cube)[1]
    segments <- sits_segment(
        cube = cube,
        bands = "NDVI",
        date = sits_timeline(cube)[1],
        step = 20
    )
}
}
\references{
Achanta, Radhakrishna, Appu Shaji, Kevin Smith, Aurelien Lucchi,
                     Pascal Fua, and Sabine Süsstrunk. 2012. “SLIC Superpixels Compared
                     to State-of-the-Art Superpixel Methods.” IEEE Transactions on
                     Pattern Analysis and Machine Intelligence 34 (11): 2274–82.

                     Nowosad, Jakub, and Tomasz F. Stepinski. 2022. “Extended SLIC
                     Superpixels Algorithm for Applications to Non-Imagery Geospatial
                     Rasters.” International Journal of Applied Earth Observation
                     and Geoinformation 112 (August): 102935.
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}

Rolf Simoes, \email{rolf.simoes@inpe.br}

Felipe Carvalho, \email{felipe.carvalho@inpe.br}
}
