% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_table.R
\name{sits_table_result}
\alias{sits_table_result}
\title{Create a sits table to store the result of TWDTW classification}
\usage{
sits_table_result()
}
\value{
table  a tibble in SITS format
}
\description{
\code{sits_table_result} returns an empty sits table
A sits table is a tibble with pre-defined columns that
has the metadata and data for each time series. The columns are
<longitude, latitude, start_date, end_date, label, coverage, time_series, distances, matches>
}
\details{
Most functions on the sits package use a sits table as input (with additional parameters)
and a sits table as output. This allows for chaining of operation on time series
}
\seealso{
Other STIS table functions: \code{\link{sits_align}},
  \code{\link{sits_bands}}, \code{\link{sits_dates}},
  \code{\link{sits_group_bylatlong}},
  \code{\link{sits_rename}}, \code{\link{sits_select}},
  \code{\link{sits_table}}, \code{\link{sits_values_cols}},
  \code{\link{sits_values_rows}}
}
