% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_assessment.R
\name{sits_test_patterns}
\alias{sits_test_patterns}
\title{Evaluates the accuracy of a set of patterns}
\usage{
sits_test_patterns(data.tb, patterns.tb, bands, alpha = -0.1, beta = 100,
  theta = 0.5, span = 0, start_date = NULL, end_date = NULL,
  interval = "12 month", overlap = 0.5)
}
\arguments{
\item{data.tb}{A sits tibble containing a set of samples with known and trusted labels}

\item{patterns.tb}{A sits tibble containing a set of patterns}

\item{bands}{The bands used for classification}

\item{alpha}{(double)  - the steepness of the logistic function used for temporal weighting}

\item{beta}{(integer) - the midpoint (in days) of the logistic function}

\item{theta}{(double)  - the relative weight of the time distance compared to the dtw distance}

\item{span}{(integer) - minimum number of days between two matches of the same pattern in the time series (approximate)}

\item{start_date}{date - the start of the classification period}

\item{end_date}{date - the end of the classification period}

\item{interval}{date - the period between two classifications}

\item{overlap}{(double) minimum overlapping between one match and the interval of classification}
}
\value{
assess         an assessment of validation
}
\description{
Tests the accuracy of TWDTW classification
of set of labelled samples using a set of patterns.
This function should be used when the patterns are not directly derived from the samples.
It provides an initial assessment of the validity of using this set of pattern
to classify an area whose samples are given.
This function returns the Overall Accuracy, User's Accuracy,
Producer's Accuracy, error matrix (confusion matrix), and Kappa values.
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
