% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_testWTSS.R
\name{sits_testWTSS}
\alias{sits_testWTSS}
\title{Verify if the information concerning WTSS service has been configured}
\usage{
sits_testWTSS()
}
\description{
\code{sits_testWTSS} tests if the WTSS service has been correctly configured
 The function verifies that the WTSS service is running.
}
\details{
The Web Time Seris Service (WTSS) is a lightweight web service the allow remote access to satellite
 image time series and provides three operations:

1. list coverages: this operation allows clients to retrieve the capabilities provided
    by any server that implements WTSS. It returns a list of coverage
    names available in a server instance.

2. describe coverage: this operation returns the metadata for a given coverage
    identified by its name. It includes its range in the spatial and temporal dimensions.

3. time series: this operation requests the time series of values of a coverage attribute
   at a given location.
}
