% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_tibble.R
\name{sits_tibble}
\alias{sits_tibble}
\title{Create a sits table to store the time series information}
\usage{
sits_tibble()
}
\value{
result.tb  a tibble in SITS format
}
\description{
this function returns an empty sits table.
SITS tibbles are the main structures of the "sits" package.
They contain both the satellite image time series and its metadata.
A sits tibble is a tibble with pre-defined columns that
has the metadata and data for each time series. The columns are
<longitude, latitude, start_date, end_date, label, coverage, time_series>
Most functions on the sits package use a sits tibble as input (with additional parameters)
and a sits tibble as output. This allows for chaining of operation on time series.
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
