% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_tibble.R
\name{sits_tibble_coverage}
\alias{sits_tibble_coverage}
\title{Create an empty tibble to store the metadata of a coverage}
\usage{
sits_tibble_coverage()
}
\value{
coverage.tb   a tibble to store the metadata
}
\description{
Create an empty tibble to store the metadata about a coverage
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}

Rolf Simoes, \email{rolf.simoes@inpe.br}
}
