% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_tibble.R
\name{sits_tibble_distance}
\alias{sits_tibble_distance}
\title{Create an empty distance tibble to store the results of distance metrics}
\usage{
sits_tibble_distance(patterns.tb)
}
\arguments{
\item{patterns.tb}{a SITS tibble with a set of patterns}
}
\value{
distances.tb   a tibble to store the distances between a time series and a set of patterns
}
\description{
Create an empty distance tibble to store the results of distance metrics
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}

Rolf Simoes, \email{rolf.simoes@inpe.br}
}
