% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_tibble.R
\name{sits_tibble_patterns}
\alias{sits_tibble_patterns}
\title{Create an empty tibble to store the patterns used for classification}
\usage{
sits_tibble_patterns()
}
\value{
result.tb   a tibble to store the result of classifications
}
\description{
Create an empty tibble to store a set of patterns
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}

Rolf Simoes, \email{rolf.simoes@inpe.br}
}
