% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_tibble.R
\name{sits_tibble_prediction}
\alias{sits_tibble_prediction}
\title{Create an empty tibble to store the results of predictions}
\usage{
sits_tibble_prediction()
}
\value{
distances.tb   a tibble to store the predictions
}
\description{
Create an empty tibble to store the results of predictions
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}

Rolf Simoes, \email{rolf.simoes@inpe.br}
}
