% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_tibble.R
\name{sits_tibble_raster}
\alias{sits_tibble_raster}
\title{Create one line of metadata tibble to store the description of a spatio-temporal raster}
\usage{
sits_tibble_raster(raster.obj, band, timeline, scale_factor)
}
\arguments{
\item{raster.obj}{Valid Raster object (associated to filename)}

\item{band}{Name of band (either raw or classified)}

\item{timeline}{Timeline of data collection}

\item{scale_factor}{Scale factor to correct data}
}
\value{
raster.tb     A tibble for storing metadata about a spatio-temporal raster
}
\description{
This function creates one line of tibble containing the metadata for
              a set of spatio-temporal raster files.
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
