% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_grid_systems.R
\name{sits_tiles_to_roi}
\alias{sits_tiles_to_roi}
\title{Convert tile information to ROI in WGS84}
\usage{
sits_tiles_to_roi(tiles, grid_system = "MGRS")
}
\arguments{
\item{tiles}{Character vector with names of tiles from the
selected \code{grid_system}.}

\item{grid_system}{Grid system that the \code{tiles} belong to.
Currently supported grid systems are the
MGRS grid (\code{"MGRS"}, default) and those
used by the Brazil Data Cube
(\code{"BDC_LG_V2"}, \code{"BDC_MD_V2"} and
\code{"BDC_SM_V2"}).}
}
\value{
roi         Valid ROI to use in other SITS functions
}
\description{
Takes a list of tiles from a given grid system and produces a ROI
(region of interest) in WGS84 covering them.
}
\examples{
if (sits_run_examples()) {
# Convert MGRS tiles to a ROI
roi <- sits_tiles_to_roi(c("22KGA", "22KGV"))

# Convert Brazil Data Cube (large grid) tiles to a ROI
roi <- sits_tiles_to_roi(
  c("003004", "003005"), grid_system = "BDC_LG_V2"
)
}
}
\seealso{
Other data cube management:
\code{\link[=sits_add_base_cube]{sits_add_base_cube()}},
\code{\link[=sits_cube]{sits_cube()}},
\code{\link[=sits_cube_copy]{sits_cube_copy()}},
\code{\link[=sits_cube_replace_path]{sits_cube_replace_path()}},
\code{\link[=sits_merge]{sits_merge()}},
\code{\link[=sits_mosaic]{sits_mosaic()}},
\code{\link[=sits_regularize]{sits_regularize()}},
\code{\link[=sits_roi_to_tiles]{sits_roi_to_tiles()}}
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}

Rolf Simoes, \email{rolf.simoes@gmail.com}
}
\concept{data cube management}
