% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_table.R
\name{sits_time_interval}
\alias{sits_time_interval}
\title{Sample a percentage of a time series}
\usage{
sits_time_interval(data.tb)
}
\arguments{
\item{data.tb}{a sits tibble}
}
\value{
ndays   number of days covered by the time series
}
\description{
finds out the temporal interval of the time series of the data
}
\author{
Rolf Simoes, \email{rolf.simoes@inpe.br}
}
