% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_tibble_functions.R
\name{sits_time_series_dates}
\alias{sits_time_series_dates}
\title{Retrieve the dates of time series for a row of a sits tibble}
\usage{
sits_time_series_dates(data)
}
\arguments{
\item{data}{A sits tibble with one or more time series.}
}
\value{
A vector of dates
}
\description{
Returns the dates of the time series associated to a sits tibble
}
\examples{
# Retrieve a set of time series with 2 classes
data(cerrado_2classes)
# Retrieve the dates of the first time series
sits_time_series_dates(cerrado_2classes)
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
