% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_timeline.R
\name{sits_timeline.classified_image}
\alias{sits_timeline.classified_image}
\title{Obtains the timeline for a classified data cube}
\usage{
\method{sits_timeline}{classified_image}(data)
}
\description{
Obtains the timeline for a classified data cube
}
