% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_timeline.R
\name{sits_timeline.cube}
\alias{sits_timeline.cube}
\title{Obtains the timeline for a data cube}
\usage{
\method{sits_timeline}{cube}(data, index = 1)
}
\arguments{
\item{data}{A sits tibble (either a sits tibble or a raster metadata).}

\item{index}{The index to obtain the timeline}
}
\description{
Obtains the timeline for a data cube
}
