% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_timeline.R
\name{sits_timeline.probs_cube}
\alias{sits_timeline.probs_cube}
\title{Obtains the timeline for a probs data cube}
\usage{
\method{sits_timeline}{probs_cube}(data)
}
\description{
Obtains the timeline for a probs data cube
}
