% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_timeline.R
\name{sits_timeline.satveg_cube}
\alias{sits_timeline.satveg_cube}
\title{Obtains the timeline for a raster data cube}
\usage{
\method{sits_timeline}{satveg_cube}(data)
}
\description{
Obtains the timeline for a raster data cube
}
