% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_timeline.R
\name{sits_timeline.sits}
\alias{sits_timeline.sits}
\title{Obtains the timeline for a set of time series}
\usage{
\method{sits_timeline}{sits}(data)
}
\arguments{
\item{data}{A sits tibble}
}
\description{
Obtains the timeline for a set of time series
}
