% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_timeline.R
\name{sits_timeline_match}
\alias{sits_timeline_match}
\title{Find dates in the input data cube that match those of the patterns}
\usage{
sits_timeline_match(timeline, ref_start_date, ref_end_date,
  interval = "12 month")
}
\arguments{
\item{timeline}{Timeline of input observations (vector).}

\item{ref_start_date}{The day of the year to be taken as reference for starting the classification.}

\item{ref_end_date}{The day of the year to be taken as reference for end the classification.}

\item{interval}{Period between two classification.}
}
\value{
A list of breaks that will be applied to the input data set.
}
\description{
For correct classification, the time series of the input data set
             should be aligned to that of the reference data set (usually a set of patterns).
             This function aligns these data sets so that shape matching works correctly
}
\examples{
# get a timeline for MODIS data
data("timeline_2000_2017")
# get a set of subsets for a period of 10 years
ref_start_date <- lubridate::ymd("2000-08-28")
ref_end_date <- lubridate::ymd("2000-08-13")
subset_dates.lst <- sits_timeline_match(timeline_2000_2017, ref_start_date, ref_end_date)
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
