% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_csv.R
\name{sits_to_csv}
\alias{sits_to_csv}
\alias{sits_to_csv.sits}
\alias{sits_to_csv.tbl_df}
\alias{sits_to_csv.default}
\title{Export a sits tibble metadata to the CSV format}
\usage{
sits_to_csv(data, file = NULL)

\method{sits_to_csv}{sits}(data, file = NULL)

\method{sits_to_csv}{tbl_df}(data, file)

\method{sits_to_csv}{default}(data, file)
}
\arguments{
\item{data}{Time series (tibble of class "sits").}

\item{file}{Full path of the exported CSV file
(valid file name with extension ".csv").}
}
\value{
Return data.frame with CSV columns (optional)
}
\description{
Converts metadata from a sits tibble to a CSV file.
             The CSV file will not contain the actual time
             series. Its columns will be the same as those of a
             CSV file used to retrieve data from
             ground information ("latitude", "longitude", "start_date",
             "end_date", "cube", "label").
             If the file is NULL, returns a data.frame as an object
}
\examples{
csv_file <- paste0(tempdir(), "/cerrado_2classes.csv")
sits_to_csv(cerrado_2classes, file = csv_file)
}
\seealso{
Other data conversion:
\code{\link[=sits_as_sf]{sits_as_sf()}},
\code{\link[=sits_as_stars]{sits_as_stars()}},
\code{\link[=sits_as_terra]{sits_as_terra()}},
\code{\link[=sits_config_to_hf]{sits_config_to_hf()}},
\code{\link[=sits_sf_to_tibble]{sits_sf_to_tibble()}},
\code{\link[=sits_timeseries_to_csv]{sits_timeseries_to_csv()}},
\code{\link[=sits_to_hf]{sits_to_hf()}},
\code{\link[=sits_to_xlsx]{sits_to_xlsx()}}
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
\concept{data conversion}
