% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_tibble_functions.R
\name{sits_to_shp}
\alias{sits_to_shp}
\title{Export from sits_tibble to sf object.}
\usage{
sits_to_shp(data.tb, crs = 4326)
}
\arguments{
\item{data.tb}{A sits table.}

\item{crs}{A length-one numeric. The Coordinate Reference System identificator. The default is WGS84.}
}
\value{
sf_obj       A sf object.
}
\description{
Export the given sits_tibble to a sf object of points.
}
\examples{
# get a random sits_tibble
data.tb <- sits_random_tibble(n_samples = 10, n_vi = 3)
# export to sf object
sf_obj <- sits_to_shp(data.tb)
# save as shapefile
sf::st_write(sf_obj, dsn = file.path(tempdir(), "sf_obj.shp"), driver = "ESRI Shapefile")

}
\author{
Alber Sanchez, \email{alber.ipia@inpe.br}
}
