% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_csv.R
\name{sits_toCSV}
\alias{sits_toCSV}
\title{Export a tibble data to the CSV format}
\usage{
sits_toCSV(data.tb, file)
}
\arguments{
\item{data.tb}{a SITS time series}

\item{file}{the name of the CSV file to be exported}
}
\value{
status     the status of the operation
}
\description{
Converts data from a SITS tibble to a CSV file
}
\examples{
\donttest{
# read a tibble with 400 samples of Cerrado and 346 samples of Pasture
data(cerrado_2classes)
# export a time series to zoo
sits_toCSV (cerrado_2classes, file = "./cerrado_2classes.csv")
}
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
