% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_JSON.R
\name{sits_toJSON}
\alias{sits_toJSON}
\title{Store the contents of an in-memory database in a JSON file}
\usage{
sits_toJSON(json_file, table)
}
\arguments{
\item{json_file}{name of the JSON file}

\item{table}{a string with the name of a global database (previously created)}
}
\value{
an updated in-memory database
}
\description{
\code{sits_toJSON} gets a SITS table from a JSON file
}
\details{
The set of time series from a SITS table can be saved in JSON format
and later retrieved for further use
}
\examples{
sits_toJSON ("myjson_file.json", database = "my.db")
}
\seealso{
Other STIS: \code{\link{sits_fromJSON}}
}
\keyword{STIS}
