% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_tuning.R
\name{sits_torch_optim_tuning}
\alias{sits_torch_optim_tuning}
\title{Tuning torch optmizer's hyper-parameters}
\usage{
sits_torch_optim_tuning(
  samples,
  ...,
  samples_validation = NULL,
  validation_split = 0.2,
  ml_fns = list(sits_tempcnn, sits_lighttae),
  opt_fns = list(torch::optim_adam, optim_adamw),
  lr = c(0.01, 0.005, 0.001),
  eps = c(1e-06, 1e-07, 1e-08),
  weight_decay = c(0, 1e-05, 1e-06),
  multicores = 2
)
}
\arguments{
\item{...}{}
}
\value{
A list containing the best torch optimizer and its parameters.
}
\description{
This function combines all parameters and computes torch models to
parameter combination, do a validation using validation samples or
splitting samples using validation_split. The function returns the
best hyper-parameters in a list.
}
\examples{
\donttest{
# read a set of samples
data(cerrado_2classes)
# two fold validation with random forest
hparams <- sits_torch_optim_tuning(cerrado_2classes)
}
}
