% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_toTWDTW.R
\name{sits_toTWDTW}
\alias{sits_toTWDTW}
\title{Convert a SITS table to TWDTW format}
\usage{
sits_toTWDTW(data.tb)
}
\arguments{
\item{data.tb}{a tibble in SITS format with time series for different bands}
}
\value{
twdtw.ts   a list of time series in TWDTW format
}
\description{
\code{sits_toTWDTW} returns a TWDTW time series for classification
}
\details{
A sits table has the metadata and data for each time series
<longitude, latitude, from, to, label, coverage, time_series>
}

