% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_export.R
\name{sits_toTWDTW_matches}
\alias{sits_toTWDTW_matches}
\title{Export data to be used by the dtwSat package}
\usage{
sits_toTWDTW_matches(data.tb, patterns.tb)
}
\arguments{
\item{data.tb}{a table in SITS format with time series to be converted to TWTDW time series}

\item{patterns.tb}{patterns SITS tibble used to matching}
}
\value{
ts.twdtw      a time series in TWDTW format (an object of the twdtwTimeSeries class)
}
\description{
Converts data from a SITS table to an instance of a TWDTW matches class,
Returns a dtwSat::twdtwMatches object (S4)
}
\author{
Rolf Simoes, \email{rolf.simoes@inpe.br}
}
