% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_getdata.R
\name{sits_tozoo}
\alias{sits_tozoo}
\title{Export data to be used to the zoo format}
\usage{
sits_tozoo(ts.tb, band)
}
\arguments{
\item{ts.tb}{a tibble with a the indexes and values of a SITS time series}

\item{band}{the name of the band to}
}
\value{
ts.zoo        a time series in TWDTW format (an object of the twdtwTimeSeries class)
}
\description{
Converts data from a SITS table to an instance of a zoo series,
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
