% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_transitions.R
\name{sits_transitions}
\alias{sits_transitions}
\alias{sits_transitions.class_cube}
\alias{sits_transitions.default}
\title{Reclassify a classified cube}
\usage{
sits_transitions(cube, ...)

\method{sits_transitions}{class_cube}(
  cube,
  ...,
  rules = NULL,
  memsize = 4L,
  multicores = 2L,
  output_dir,
  version = "v1",
  progress = TRUE
)

\method{sits_transitions}{default}(cube, ...)
}
\arguments{
\item{cube}{Image cube to be reclassified (class = "class_cube")}

\item{...}{Other parameters for specific functions.}

\item{rules}{Expressions to be evaluated (named list).}

\item{memsize}{Memory available for classification in GB
(integer, min = 1, max = 16384).}

\item{multicores}{Number of cores to be used for classification
(integer, min = 1, max = 2048).}

\item{output_dir}{Directory where files will be saved
(character vector of length 1 with valid location).}

\item{version}{Version of resulting image (character).}

\item{progress}{Set progress bar??}
}
\value{
An object of class "class_cube" (reclassified cube).
}
\description{
Apply a set of named expressions to reclassify a classified image.
The expressions should use character values to refer to labels in
logical expressions.
}
\note{
Reclassification of a remote sensing map refers
to changing the classes assigned to different pixels in the image.
Reclassification involves assigning new classes to pixels based
on additional information from a reference map.
Users define rules according to the desired outcome.
These rules are then applied to the classified map to produce
a new map with updated classes.

\code{sits_reclassify()} allow any valid R expression to compute
reclassification. User should refer to \code{cube} and \code{mask}
to construct logical expressions.
Users can use can use any R expression that evaluates to logical.
\code{TRUE} values will be relabeled to expression name.
Updates are done in asynchronous manner, that is, all expressions
are evaluated using original classified values. Expressions are
evaluated sequentially and resulting values are assigned to
output cube. Last expressions has precedence over first ones.
}
\author{
Felipe Carlos, \email{efelipecarlos@gmail.com}

Felipe Carvalho, \email{felipe.carvalho@inpe.br}

Gilberto Camara, \email{gilberto.camara@inpe.br}
}
