% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_tibble_functions.R
\name{sits_transmute_bands}
\alias{sits_transmute_bands}
\title{Add new sits bands and drops existing.}
\usage{
sits_transmute_bands(data.tb, ...)
}
\arguments{
\item{data.tb}{A sits tibble.}

\item{...}{Pair expressions in the format `name = value`. See \code{\link[dplyr]{mutate}} help for more details.}
}
\value{
A sits tibble with same samples and the new bands.
}
\description{
Adds new bands and drops existing in the time series of a sits tibble using dplyr::transmute function.
}
\examples{
\donttest{
# Retrieve data for time series with label samples in Mato Grosso in Brazil
data(samples_mt_9classes)
# Generate a new image with the SAVI (Soil-adjusted vegetation index)
savi.tb <- sits_transmute_bands(samples_mt_9classes, savi = (1.5*(nir - red)/(nir + red + 0.5)))
}
}
\author{
Rolf Simoes, \email{rolf.simoes@inpe.br}
}
