% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_tsne.R
\name{sits_tsne.torch_model_mae_encoder}
\alias{sits_tsne.torch_model_mae_encoder}
\title{Run t-SNE on sits torch models}
\usage{
\method{sits_tsne}{torch_model_mae_encoder}(
  model,
  samples,
  remove_duplicates = TRUE,
  perplexity = 30,
  rounds = 1000,
  ...
)
}
\arguments{
\item{model}{A sits torch model (output of sits_train()).}

\item{samples}{A sits tibble with time series samples.}

\item{remove_duplicates}{Logical. Should duplicated embeddings be removed before running t-SNE? Default is TRUE.}

\item{perplexity}{Numeric; Rtsne perplexity (default \code{30}). Will be clamped to
\code{floor((N - 1) / 3)} based on the number of (possibly deduped) points.}

\item{rounds}{Integer; number of optimization iterations in t-SNE
(maps to \code{Rtsne::Rtsne(max_iter=)}; default \code{1000}).}

\item{...}{Other parameters (passed to methods).}
}
\description{
Applies t-SNE for masked autoencoder models using the MAE's encoder outputs as embeddings.
}
