% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_tsne.R
\name{sits_tsne.torch_model_mlp}
\alias{sits_tsne.torch_model_mlp}
\title{Run t-SNE on sits torch models}
\usage{
\method{sits_tsne}{torch_model_mlp}(model, samples, remove_duplicates = TRUE, ...)
}
\arguments{
\item{model}{A sits torch model (output of sits_train()).}

\item{samples}{A sits tibble with time series samples.}

\item{remove_duplicates}{Logical. Should duplicate embeddings be removed before running t-SNE? Default is TRUE.}

\item{...}{Other parameters (passed to methods).}
}
\description{
Applies t-SNE for Multilayer Perceptron (MLP) models using temporal encoder outputs as embeddings.
}
