% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_twdtw_classify.R
\name{sits_twdtw_classify}
\alias{sits_twdtw_classify}
\title{Classify a sits tibble using TWDTW}
\usage{
sits_twdtw_classify(samples.tb, patterns.tb, bands, alpha = -0.1,
  beta = 100)
}
\arguments{
\item{bands}{string - the bands to be used for classification}

\item{data.tb}{a tibble in SITS format with time series for different bands}
}
\value{
table   a tibble in SITS format with values
}
\description{
\code{sits_twdtw_classify} returns a sits table with values only
}
\details{
A sits table has the metadata and data for each time series
<longitude, latitude, from, to, label, coverage, time_series>
}

