% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_TWDTW.R
\name{sits_TWDTW_classify_tibble}
\alias{sits_TWDTW_classify_tibble}
\title{Classify a sits tibble using the matches found by the TWDTW methods}
\usage{
sits_TWDTW_classify_tibble(data.tb, patterns.tb, start_date = NULL,
  end_date = NULL, interval = "12 month", overlap = 0.5)
}
\arguments{
\item{data.tb}{a table in SITS format with the matches that have been produced by TWTDW}

\item{patterns.tb}{patterns SITS tibble used to matching}

\item{start_date}{date - the start of the classification period}

\item{end_date}{date - the end of the classification period}

\item{interval}{the period between two classifications}

\item{overlap}{minimum overlapping between one match and the interval of classification}
}
\value{
matches       a SITS table with the information on matches for the data
}
\description{
Returns a sits table with the results of the TWDTW classifier.
The TWDTW classifier uses the result of the sits_TWDTW_matches function to
find the best alignments of these matches in intervals chosen by the user
}
\references{
Maus V, Camara G, Cartaxo R, Sanchez A, Ramos FM, de Queiroz GR (2016).
A Time-Weighted Dynamic Time Warping Method for Land-Use and Land-Cover Mapping. IEEE
 Journal of Selected Topics in Applied Earth Observations and Remote Sensing, 9(8):3729-3739,
 August 2016. ISSN 1939-1404. doi:10.1109/JSTARS.2016.2517118.
}
\author{
Victor Maus, \email{vwmaus1@gmail.com}

Gilberto Camara, \email{gilberto.camara@inpe.br}
}
