% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_spectral_unmixing.R
\name{sits_unmixing}
\alias{sits_unmixing}
\title{Create a multiple endmember spectral mixture analyses fractions images}
\usage{
sits_unmixing(
  cube,
  endmembers_spectra,
  memsize = 1,
  multicores = 2,
  output_dir = getwd(),
  progress = TRUE
)
}
\arguments{
\item{cube}{...}

\item{memsize}{...}

\item{multicores}{...}

\item{output_dir}{...}

\item{progress}{...}

\item{endmembers}{...}
}
\value{
a cube ...
}
\description{
A multiple endmember spectral mixture analyses...
}
