% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_validate.R
\name{sits_validadeCluster}
\alias{sits_validadeCluster}
\title{validades clusters against original labels}
\usage{
sits_validadeCluster(data.tb)
}
\arguments{
\item{data.tb}{a SITS table with the samples to be validated}
}
\description{
Given a sits table with `original_label` column, computes a confusion matrix
between the original labels (`original_label` column) and new labels
}
\author{
Rolf Simoes, \email{rolf.simoes@inpe.br}
}
