% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_table.R
\name{sits_value_cols}
\alias{sits_value_cols}
\alias{sits_values_cols}
\title{Return the values of one band of a SITS table colwise organised}
\usage{
sits_values_cols(data.tb, band)
}
\arguments{
\item{data.tb}{a tibble in SITS format with time series for different bands}

\item{band}{string - a band whose values are to be extracted}
}
\value{
table   a tibble  with values
}
\description{
returns a sits table with values only (colwise organised)
}
\seealso{
Other STIS table functions: \code{\link{sits_table_result}}
}
\author{
Rolf Simoes, \email{rolf.simoes@inpe.br}
}
