% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_table.R
\name{sits_value_rows}
\alias{sits_value_rows}
\alias{sits_values_rows}
\title{Return the values of one band of a SITS table}
\usage{
sits_values_rows(data.tb, band)
}
\arguments{
\item{data.tb}{a tibble in SITS format with time series for different bands}

\item{band}{string - a band whose values are to be extracted}
}
\value{
table      a tibble in SITS format with values
}
\description{
this function returns only the values of a sits table (rowwise organized).
This function is useful to use packages such as ggplot and dtwclust that
require values that are rowwise organised
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
