% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_tibble.R
\name{sits_values.bands_cases_dates}
\alias{sits_values.bands_cases_dates}
\title{Return the values of a given sits tibble using "bands_cases_dates"}
\usage{
\method{sits_values}{bands_cases_dates}(data, bands = NULL, format)
}
\arguments{
\item{data}{A sits tibble with time series for different bands.}

\item{bands}{A string with a group of bands whose
values are to be extracted. If no bands are informed
extract ALL bands}

\item{format}{A string with either "cases_dates_bands"
or "bands_cases_dates" or "bands_dates_cases"}
}
\value{
A sits tibble with values.
}
\description{
This function returns only the values of a sits tibble
(according a specified format).
This function is useful to use packages such as ggplot2, dtwclust, or kohonen
that require values that are rowwise or colwise organised.
}
\examples{
# Retrieve a set of time series with 2 classes
data(cerrado_2classes)
# retrieve the values split by bands
sits_values(cerrado_2classes[1:2, ], format = "bands_cases_dates")
}
\author{
Rolf Simoes, \email{rolf.simoes@inpe.br}
}
