% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_values.R
\name{sits_values_cols}
\alias{sits_values_cols}
\title{Return the values of one band of a SITS table (colwise organised)}
\usage{
sits_values_cols(data, band)
}
\arguments{
\item{band}{string - a band whose values are to be extracted}

\item{data.tb}{a tibble in SITS format with time series for different bands}
}
\value{
table   a tibble in SITS format with values
}
\description{
\code{sits_value_cols} returns a sits table with values only (colwise organised)
}
\details{
A sits table has the metadata and data for each time series
<longitude, latitude, start_date, end_date, label, coverage, time_series>
}
