% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_source_api.R
\name{source_bands}
\alias{source_bands}
\title{Source bands functions}
\arguments{
\item{source}{Valid data source.}

\item{collection}{Image collection.
source.}

\item{fn_filter}{A \code{function} that will be applied in each band
to filter selection. The provided function must have an input parameter to
receive band object and return a \code{logical} value.}

\item{add_cloud}{Should the cloud band be returned?}

\item{key}{Key of a band object.}

\item{bands}{Bands to be retrieved}

\item{default}{Value to be returned if an attribute or key is not
found.}
}
\value{
The values returned by each function are described as follows.
}
\description{
These functions provide an API to handle/retrieve data from bands.
}
\keyword{internal}
