% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_source_api.R
\name{source_collection}
\alias{source_collection}
\title{Source collection functions}
\arguments{
\item{source}{Data source.}

\item{collection}{Image collection.}

\item{tiles}{Tile names}

\item{start_date}{Start date.}

\item{end_date}{End date.}
}
\value{
The values returned by each function are described as follows.
}
\description{
These functions provide an API to handle/retrieve data from source's
collections.
}
\keyword{internal}
