% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_source_api.R
\name{source_item_get_functions}
\alias{source_item_get_functions}
\alias{.source_item_get_date}
\alias{.source_item_get_hrefs}
\alias{.source_item_get_bands}
\alias{.source_item_get_resolutions}
\title{Function for retrieving information from a single item}
\usage{
.source_item_get_date(source, item, ..., collection = NULL)

.source_item_get_hrefs(source, item, ..., collection = NULL)

.source_item_get_bands(source, item, ..., collection = NULL)

.source_item_get_resolutions(source, item, ..., collection = NULL)
}
\arguments{
\item{source}{Data source (one of "SATVEG", "LOCAL", "BDC", "AWS", "USGS"
"DEAFRICA", "PROBS").}

\item{...}{Additional parameters.}

\item{collection}{Collection to be searched in the data source.}

\item{items}{a \code{STACItemCollection} object returned by rstac.}
}
\value{
an atomic \code{vector} with the required information.
}
\description{
Generic function for retrieving information in item returned by
 APIs. The concept of items is used in STAC APIs, where each item corresponds
 to a satellite image (scene), with a single time and multiple bands.
}
\keyword{internal}
