% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_source_api.R
\name{source_items_get_functions}
\alias{source_items_get_functions}
\alias{.source_items_tiles_group}
\alias{.source_items_get_sensor}
\alias{.source_items_get_satellite}
\title{Function for retrieving information from multiples items}
\usage{
.source_items_tiles_group(source, items, ..., collection = NULL)

.source_items_get_sensor(source, items, ..., collection = NULL)

.source_items_get_satellite(source, items, ..., collection = NULL)
}
\arguments{
\item{source}{Data source (one of "SATVEG", "LOCAL", "BDC", "AWS", "USGS"
"DEAFRICA", "PROBS").}

\item{items}{A \code{STACItemCollection} object returned by rstac or
\code{character} vector with directories to be search.}

\item{...}{Additional parameters.}

\item{collection}{Collection to be searched in the data source.}
}
\value{
an \code{atomic} vector with informations retrived from items.
}
\description{
Retrieves information from the STACItemCollection object in the
rstac package. Generic function created to handle different STAC providers.
}
\keyword{internal}
