% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_source_api.R
\name{source_tile_items_get_functions}
\alias{source_tile_items_get_functions}
\alias{.source_items_tile_get_name}
\alias{.source_items_tile_get_bbox}
\alias{.source_items_tile_get_size}
\title{...}
\usage{
.source_items_tile_get_name(source, tile_items, ..., collection = NULL)

.source_items_tile_get_bbox(source, tile_items, ..., collection = NULL)

.source_items_tile_get_size(source, tile_items, ..., collection = NULL)
}
\arguments{
\item{source}{Name of the STAC provider}

\item{...}{Aditional parameters}

\item{collection}{...}

\item{items}{...}
}
\value{
...
}
\description{
TODO: add
}
\keyword{internal}
