% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_stars.R
\name{st_as_stars.raster_cube}
\alias{st_as_stars.raster_cube}
\title{Extension to stars for exporting sits cubes as stars objects}
\usage{
st_as_stars.raster_cube(.x, ...)
}
\arguments{
\item{.x}{A sits cube.}

\item{...}{Other parameters for st_as_stars}
}
\value{
A space-time stars object.
}
\description{
Uses the information about files, bands and dates
in a data cube to produce an object of class \code{stars}.
User has to select a tile from the data cube. By default,
all bands and dates are included in the \code{stars} object.
Users can select bands and dates.
}
\note{
By default, the \code{stars} object will be loaded in memory. This
can result in heavy memory usage. To produce a \code{stars.proxy} object,
uses have to select a single date, since \code{stars} does not allow
proxy objects to be created with two dimensions.
}
\examples{
if (sits_run_examples()) {
    library(stars)
    # convert sits cube to an sf object (polygon)
    data_dir <- system.file("extdata/raster/mod13q1", package = "sits")
    cube <- sits_cube(
        source = "BDC",
        collection = "MOD13Q1-6.1",
        data_dir = data_dir
    )
    stars_object <- st_as_stars(cube)
}
}
\author{
Gilberto Camara, \email{gilberto.camara.inpe@gmail.com}
}
