% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_transitions.R
\name{Starts}
\alias{Starts}
\title{Starts transition function}
\usage{
Starts(target_class)
}
\arguments{
\item{target_class}{Target class name for the transition}
}
\value{
expression that creates a STARTS transition call
}
\description{
Starts transition function
}
\author{
Felipe Carlos, \email{efelipecarlos@gmail.com}

Felipe Carvalho, \email{felipe.carvalho@inpe.br}

Gilberto Camara, \email{gilberto.camara@inpe.br}
}
\keyword{internal}
