% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_WTSPS.R
\docType{methods}
\name{statusProcessing}
\alias{statusProcessing}
\alias{statusProcessing,WTSPS-method}
\title{Returns the status of a processing algorithm}
\usage{
statusProcessing(object, id)

\S4method{statusProcessing}{WTSPS}(object, id)
}
\arguments{
\item{object}{A WTSPS object}

\item{id}{a process identifier}
}
\description{
Returns the status of a processing algorithm
}
