% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_data_generics.R
\name{`sits_bands<-.cube`}
\alias{`sits_bands<-.cube`}
\alias{sits_bands<-.cube}
\title{Replaces the names of the bands of a cube}
\usage{
\method{sits_bands}{cube}(x) <- value
}
\arguments{
\item{x}{Valid sits tibble (time series or a cube)}

\item{value}{Vector of bands}
}
\value{
An updated data cube with the new bands.
}
\description{
Finds the names of the bands of a cube
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}

Rolf Simoes, \email{rolf.simoes@inpe.br}
}
