% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_data_generics.R
\name{`sits_bands<-.predicted`}
\alias{`sits_bands<-.predicted`}
\alias{sits_bands<-.predicted}
\title{Replaces the names of the bands of a set of predictions}
\usage{
\method{sits_bands}{predicted}(x) <- value
}
\arguments{
\item{x}{Valid sits tibble (time series or a cube)}

\item{value}{Vector of bands}
}
\value{
An updated data set with the new bands.
}
\description{
Replaces the names of the bands of time series in a sits tibble
}
\examples{
# Replace the name of bands for the samples for Mato Grosso
sits_bands(samples_mt_4bands) <- c("ndvi", "evi", "nir", "mir")
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}

Rolf Simoes, \email{rolf.simoes@inpe.br}
}
