% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_table.R
\name{`sits_bands<-`}
\alias{`sits_bands<-`}
\alias{sits_bands<-}
\title{names of the bands of a time series}
\usage{
sits_bands(data.tb) <- value
}
\arguments{
\item{data.tb}{a valid sits table}

\item{value}{string vector with the new bands' names}
}
\value{
invisible(data.tb) the input data invisible
}
\description{
set the names of the bands of time series in a sits table
}
\author{
Rolf Simoes, \email{rolf.simoes@inpe.br}
}
