% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_table.R
\name{`sits_labels<-`}
\alias{`sits_labels<-`}
\alias{sits_labels<-}
\title{names of the labels of sits table}
\usage{
sits_labels(data.tb) <- value
}
\arguments{
\item{data.tb}{a valid sits table}

\item{value}{string vector with the new label names}
}
\value{
invisible(data.tb) the input data invisible
}
\description{
set the names of the labels of a given sits table
}
\author{
Rolf Simoes, \email{rolf.simoes@inpe.br}
}
