% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_tile_api.R
\name{tile_accessors}
\alias{tile_accessors}
\alias{.tile_source}
\alias{.tile_collection}
\alias{.tile_name}
\alias{.tile_name<-}
\alias{.tile_ncols}
\alias{.tile_nrows}
\alias{.tile_size}
\alias{.tile_labels}
\alias{.tile_labels<-}
\alias{.tile_file_info}
\alias{.tile_file_info<-}
\title{Tile field accessors}
\usage{
.tile_source(tile)

.tile_collection(tile)

.tile_name(tile)

.tile_name(tile, name) <- value

.tile_ncols(tile)

.tile_nrows(tile)

.tile_size(tile)

.tile_labels(tile)

.tile_labels(tile) <- value

.tile_file_info(tile)

.tile_file_info(tile) <- value
}
\arguments{
\item{tile}{A \code{tile} or \code{cube} (only first row will be considered).}

\item{value}{A value to set in a \code{tile} field.}
}
\value{
Respective \code{tile} field value.
}
\description{
These functions are accessors of \code{tile} objects. A
\code{tile} is a only-one-row \code{tibble} \code{cube} that stores
metadata of a cube.
}
\section{Functions}{
\itemize{
\item \code{.tile_source()}: Get \code{'source'} field of a \code{tile}.

\item \code{.tile_collection()}: Get \code{'collection'} field of a \code{tile}.

\item \code{.tile_name()}: Get \code{'tile'} field of a \code{tile}.

\item \code{.tile_name(tile, name) <- value}: Set \code{'labels'} field of a \code{tile}.

\item \code{.tile_ncols()}: Get number of image columns from \code{ncols}
field (if it exists) or from first \code{ncols} in file_info.

\item \code{.tile_nrows()}: Get number of image columns from \code{nrows}
field (if it exists) or from first \code{nrows} in file_info.

\item \code{.tile_size()}: Get the size of tile from \code{ncols} and
\code{nrows} fields (if they exist) or from first entry in file_info.

\item \code{.tile_labels()}: Get \code{'labels'} field of a \code{tile}.

\item \code{.tile_labels(tile) <- value}: Set \code{'labels'} field of a \code{tile}.

\item \code{.tile_file_info()}: Get \code{'file_info'} field of a \code{tile}.

\item \code{.tile_file_info(tile) <- value}: Set \code{'file_info'} field of a \code{tile}.

}}
\seealso{
Other accessors: 
\code{\link{band_accessors}},
\code{\link{bbox_accessors}},
\code{\link{block_accessors}},
\code{\link{chunk_accessors}}
}
\concept{accessors}
\keyword{internal}
