% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{timeline_2000_2017}
\alias{timeline_2000_2017}
\title{The timeline for the sequence of images for MOD13Q1 collection 6}
\format{A vector with 407 dates in YYYY-MM-DD format.}
\usage{
data(timeline_2000_2017)
}
\description{
The timeline for the time series used in the examples of the SITS package
   that use data from MODIS collection 6.
   There are 407 instances from 2000-02-18 until 2017-10-16.
}
\keyword{datasets}
