% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{timeline_2013_2014}
\alias{timeline_2013_2014}
\title{The timeline for the sequence of images one year (2013 to 2014)}
\format{A vector with 23 dates in YYYY-MM-DD format.}
\usage{
data(timeline_2013_2014)
}
\description{
The timeline for the time series used in the examples of the sits package
   that use data for one year classification.
   There are 23 instances from 2013-09-14 until 2014-08-29.
}
\keyword{datasets}
