% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{timeline_mod13q1}
\alias{timeline_mod13q1}
\title{The timeline for the sequence of images for MOD13Q1 collection 5}
\format{A data frame with 392 lines and 1 column}
\usage{
data(timeline_mod13q1)
}
\description{
The timeline for the time series used in the examples of the SITS package.
   There are 392 instances from 2000-02-18 until 2017-02-18.
}
\keyword{datasets}
