% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{timeline_modis_392}
\alias{timeline_modis_392}
\title{The timeline for the sequence of images for MOD13Q1 collection 5}
\format{
A vector with 392 dates in YYYY-MM-DD format.
}
\usage{
data(timeline_modis_392)
}
\description{
The timeline for the time series used in the examples of the sits package
   that use data from MODIS collection 5.
   There are 392 instances from 2000-02-18 until 2017-02-18.
}
\keyword{datasets}
