% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wtss_cube}
\alias{wtss_cube}
\title{Description of a WTSS cube}
\format{
A tibble with information about a WTSS cube
}
\usage{
data(wtss_cube)
}
\description{
The metadata for a cube derived from Web Time Series Services.
}
\keyword{datasets}
