% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_tile_api.R
\name{bbox_accessors}
\alias{bbox_accessors}
\alias{.xmin}
\alias{.xmin<-}
\alias{.xmax}
\alias{.xmax<-}
\alias{.ymin}
\alias{.ymin<-}
\alias{.ymax}
\alias{.ymax<-}
\alias{.as_crs}
\alias{.crs}
\alias{.crs<-}
\title{Bbox accessors}
\usage{
.xmin(x)

.xmin(x) <- value

.xmax(x)

.xmax(x) <- value

.ymin(x)

.ymin(x) <- value

.ymax(x)

.ymax(x) <- value

.as_crs(x)

.crs(x, default_crs = NULL)

.crs(x) <- value
}
\arguments{
\item{x}{Object to get/set field value.}

\item{value}{Value to set on object field.}

\item{default_crs}{Default crs value if object doesn't have one.}
}
\value{
Getters return respective field value or \code{NULL}, if it doesn't
  exist. Setters return the updated \code{x} object.
}
\description{
These functions are accessors of \code{bbox} fields of a \code{vector}.
Getters functions returns the respective field values with the expected
data type. Setters functions convert value to expected data type and
store it in respective fields on a given object. If value has no length
and the \code{vector} is not \code{atomic}, it is removed from the object.
}
\details{
\code{.xmin()}, \code{.xmax()}, \code{.ymin()}, \code{.ymax()},
and \code{.crs()} get/set, respectively, \code{"xmin"}, \code{"xmax"},
\code{"ymin"}, \code{"ymax"}, and \code{"crs"} fields.
}
\section{Functions}{
\itemize{
\item \code{.xmin()}: Get \code{'xmin'} field.

\item \code{.xmin(x) <- value}: Set \code{'xmin'} field as numeric.

\item \code{.xmax()}: Get \code{'xmax'} field.

\item \code{.xmax(x) <- value}: Set \code{'xmax'} field as numeric.

\item \code{.ymin()}: Get \code{'ymin'} field.

\item \code{.ymin(x) <- value}: Set \code{'ymin'} field as numeric.

\item \code{.ymax()}: Get \code{'ymax'} field.

\item \code{.ymax(x) <- value}: Set \code{'ymax'} field as numeric.

\item \code{.as_crs()}: Convert a CRS numeric value to \code{character},
appending it after \code{'EPSG:'}.

\item \code{.crs()}: Get \code{'crs'} field.

\item \code{.crs(x) <- value}: Set \code{'crs'} field as \code{character} string.

}}
\examples{
if (sits_run_examples()) {
x <- c(xmax = "123")
.xmax(x) # 123 as number
x <- list(xmin = 1, xmax = 2, ymin = 3, ymax = 4)
.crs(x) <- 4326
x # with 'crs' field
.as_crs(3857) # EPSG:3857
}

}
\seealso{
Other accessors: 
\code{\link{band_accessors}},
\code{\link{block_accessors}},
\code{\link{chunk_accessors}},
\code{\link{point_accessors}},
\code{\link{tile_accessors}}
}
\concept{accessors}
\keyword{internal}
