% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_check.R
\name{check_functions}
\alias{check_functions}
\alias{.check_num}
\alias{.check_chr}
\alias{.check_lst}
\title{Auxiliary check functions}
\usage{
.check_num(
  x,
  allow_na = FALSE,
  min = -Inf,
  max = Inf,
  min_len = 0,
  max_len = 2^31,
  allow_null = FALSE,
  msg = NULL
)

.check_chr(
  x,
  allow_na = FALSE,
  allow_empty = TRUE,
  choices = NULL,
  min_len = 0,
  max_len = 2^31,
  allow_null = FALSE,
  msg = NULL
)

.check_lst(
  x,
  allow_unnamed = FALSE,
  min_len = 0,
  max_len = 2^31,
  allow_null = FALSE,
  msg = NULL
)
}
\arguments{
\item{x}{A \code{object} that will be check. That can be a
\code{numeric} or \code{character} vectors or a \code{list}.}

\item{allow_na}{A \code{logical} indicating if the check permits empty
NA values in check. Default is FALSE.}

\item{min}{A atomic \code{vector} of numeric indicating the
minimum value that the user can provide in function parameter. Only works for
numeric check. By default is \code{-Inf}.}

\item{max}{A atomic \code{vector} of numeric indicating the
maximum value that the user can provide in function parameter. Only works for
numeric check. By default is \code{Inf}.}

\item{min_len}{A \code{numeric} indicating the minimum length of vector
or list users provides for functions. Default is \code{0}.}

\item{max_len}{A \code{numeric} indicating the maximum length of vector
or list users provides for functions. Default is \code{2^31}.}

\item{allow_null}{A \code{logical} indicating if the check permits empty
NULL values in check. Default is FALSE.}

\item{msg}{A \code{character} with the error message that will show
to the user.}

\item{allow_empty}{A \code{logical} indicating if the check permits empty
list in check. Default is FALSE.}

\item{choices}{A atomic \code{vector} of characters indicating the
choices of user can provide in function parameter. Only works for character
check.}

\item{allow_unnamed}{A \code{logical} indicating if the check permits empty
unnamed list in check. Default is FALSE.}
}
\description{
Functions to check parameters from a specific type, for example,
for \code{.check_chr} functions the check is exclusive for character
type.
}
\keyword{internal}
