% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_check.R
\name{check_functions}
\alias{check_functions}
\alias{.check_that}
\alias{.check_null}
\alias{.check_na}
\alias{.check_names}
\alias{.check_length}
\alias{.check_apply}
\alias{.check_lgl_type}
\alias{.check_num_type}
\alias{.check_int_type}
\alias{.check_num_range}
\alias{.check_chr_type}
\alias{.check_chr_empty}
\alias{.check_chr_choices}
\alias{.check_lst_type}
\alias{.check_lgl}
\alias{.check_num}
\alias{.check_chr}
\alias{.check_lst}
\alias{.check_file}
\alias{.check_warn}
\title{Auxiliary check functions}
\usage{
.check_that(x, ..., local_msg = NULL, msg = NULL)

.check_null(x, ..., allow_null = FALSE, msg = NULL)

.check_na(x, ..., allow_na = FALSE, msg = NULL)

.check_names(x, ..., is_named = TRUE, msg = NULL)

.check_length(x, ..., len_min = 0, len_max = 2^31, msg = NULL)

.check_apply(x, fn_check, ..., msg = NULL)

.check_lgl_type(x, ..., msg = NULL)

.check_num_type(x, ..., msg = NULL)

.check_int_type(x, ..., msg = NULL)

.check_num_range(x, ..., min = -Inf, max = Inf, allow_zero = TRUE, msg = NULL)

.check_chr_type(x, ..., msg = NULL)

.check_chr_empty(x, ..., msg = NULL)

.check_chr_choices(x, choices, ..., case_sensitive = TRUE, msg = NULL)

.check_lst_type(x, ..., msg = NULL)

.check_lgl(
  x,
  ...,
  allow_na = FALSE,
  len_min = 0,
  len_max = 2^31,
  allow_null = FALSE,
  is_named = FALSE,
  msg = NULL
)

.check_num(
  x,
  ...,
  allow_na = FALSE,
  min = -Inf,
  max = Inf,
  allow_zero = TRUE,
  len_min = 0,
  len_max = 2^31,
  allow_null = FALSE,
  is_integer = FALSE,
  is_named = FALSE,
  msg = NULL
)

.check_chr(
  x,
  ...,
  allow_na = FALSE,
  allow_empty = TRUE,
  choices = NULL,
  len_min = 0,
  len_max = 2^31,
  allow_null = FALSE,
  is_named = FALSE,
  msg = NULL
)

.check_lst(
  x,
  ...,
  min_len = 0,
  max_len = 2^31,
  allow_null = FALSE,
  is_named = TRUE,
  fn_check = NULL,
  msg = NULL
)

.check_file(x, ..., msg = NULL)

.check_warn(expr)
}
\arguments{
\item{x}{A \code{object} that will be check. That can be a
\code{numeric} or \code{character} vectors or a \code{list}.}

\item{...}{Additional parameters for \code{fn_check} function.}

\item{local_msg}{A \code{character} with the generic error message that
will show to the user.}

\item{msg}{A \code{character} with the error message that will show
to the user.}

\item{allow_null}{A \code{logical} indicating if the check permits empty
NULL values. Default is FALSE.}

\item{allow_na}{A \code{logical} indicating if the check permits empty
NA values. Default is FALSE.}

\item{is_named}{A \code{logical} indicating if the check permits unnamed
list.}

\item{len_min}{A \code{numeric} indicating the minimum length of vector
or list users provides for functions. Default is \code{0}.}

\item{len_max}{A \code{numeric} indicating the maximum length of vector
or list users provides for functions. Default is \code{2^31}.}

\item{fn_check}{A \code{function} used to test the object elements.}

\item{min}{A atomic \code{vector} of numeric indicating the
minimum value that the user can provide in function parameter. Only works for
numeric check. By default is \code{-Inf}.}

\item{max}{A atomic \code{vector} of numeric indicating the
maximum value that the user can provide in function parameter. Only works for
numeric check. By default is \code{Inf}.}

\item{allow_zero}{A \code{logical} indicating if the check permits zero
values. Default is TRUE.}

\item{choices}{A atomic \code{vector} of characters indicating the
choices of user can provide in function parameter. Only works for character
check.}

\item{case_sensitive}{A \code{logical} indicating if check will be case
sensitive. Default is TRUE.}

\item{is_integer}{A \code{logical} indicating if the value must be
integer.}

\item{allow_empty}{A \code{logical} indicating if the check permits empty
list. Default is TRUE.}

\item{min_len}{A \code{numeric} indicating the minimum length of vector
or list users provides for functions. Default is \code{0}.}

\item{max_len}{A \code{numeric} indicating the maximum length of vector
or list users provides for functions. Default is \code{2^31}.}

\item{expr}{A \code{expression} to be evaluated.}
}
\description{
Functions to check parameters from a specific type, for example,
for \code{.check_chr} functions the check is exclusive for character
type.
}
