% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_tile_api.R
\name{chunk_accessors}
\alias{chunk_accessors}
\alias{.xres}
\alias{.yres}
\title{Chunk accessors}
\usage{
.xres(x)

.yres(x)
}
\arguments{
\item{x}{Object to get field value.}

\item{value}{Value to set on object field.}
}
\value{
Spatial resolution.
}
\description{
These functions are read-only accessors of \code{chunk} fields in a
\code{vector}.
}
\details{
\code{.xres()} and \code{.yres()} computes, respectively, \code{"xres"} and
\code{"yres"} values from chunk fields. The values are computed as
\itemize{
\item \eqn{xres=(xmax - xmin)/ncols}
\item \eqn{yres=(ymax - ymin)/nrows}
}
}
\section{Functions}{
\itemize{
\item \code{.xres()}: Computes \code{x} resolution of a \code{chunk}.

\item \code{.yres()}: Computes \code{y} resolution of a \code{chunk}.

}}
\examples{
if (sits_run_examples()) {
x <- c(nrows = 100, ymin = 1, ymax = 10)
.yres(x) # 0.09
}

}
\seealso{
Other accessors: 
\code{\link{band_accessors}},
\code{\link{bbox_accessors}},
\code{\link{block_accessors}},
\code{\link{point_accessors}},
\code{\link{tile_accessors}}
}
\concept{accessors}
\keyword{internal}
