% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_cube.R
\name{cube_functions}
\alias{cube_functions}
\alias{.cube_satellite}
\alias{.cube_sensor}
\alias{.cube_collection}
\alias{.cube_name}
\alias{.cube_labels}
\alias{.cube_bands}
\alias{.cube_source}
\alias{.cube_timeline}
\alias{.cube_tiles}
\alias{.cube_tile_check}
\alias{.cube_tile_get_fields}
\alias{.cube_tile_crs}
\alias{.cube_tile_bbox}
\alias{.cube_tile_resolution}
\alias{.cube_tile_size}
\alias{.cube_bands_missing_value}
\alias{.cube_bands_minimum_value}
\alias{.cube_bands_maximum_value}
\alias{.cube_bands_scale_factor}
\alias{.cube_bands_offset_value}
\alias{.cube_bands_resampling}
\alias{.cube_bands_resolutions}
\alias{.cube_bands_band_name}
\alias{.cube_has_cloud}
\title{Auxiliary functions for sits cubes}
\usage{
.cube_satellite(cube)

.cube_sensor(cube)

.cube_collection(cube)

.cube_name(cube)

.cube_labels(cube)

.cube_bands(cube)

.cube_source(cube)

.cube_timeline(cube)

.cube_tiles(cube)

.cube_tile_check(cube, tile)

.cube_tile_get_fields(cube, tile, fields)

.cube_tile_crs(cube, ..., tile = 1)

.cube_tile_bbox(cube, ..., tile = 1)

.cube_tile_resolution(cube, ..., tile = 1)

.cube_tile_size(cube, ..., tile = 1)

.cube_bands_missing_value(cube, ..., bands = NULL)

.cube_bands_missing_value(cube, ..., bands = NULL)

.cube_bands_minimum_value(cube, ..., bands = NULL)

.cube_bands_maximum_value(cube, ..., bands = NULL)

.cube_bands_scale_factor(cube, ..., bands = NULL)

.cube_bands_offset_value(cube, ..., bands = NULL)

.cube_bands_resampling(cube, ..., bands = NULL)

.cube_bands_resolutions(cube, ..., bands = NULL)

.cube_bands_band_name(cube, ..., bands = NULL)

.cube_has_cloud(cube)
}
\arguments{
\item{cube}{Data cube from where data is to be retrieved.}

\item{tile}{Tiles from the collection to be included in the data cube}

\item{bands}{Bands to be included.}

\item{....}{Additional parameters to be included.}
}
\value{
a \code{vector} for get attributes functions and NULL or error for
check parameters functions.
}
\description{
Auxiliary functions for extracting information from sits cubes.
}
\keyword{internal}
