% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_cube_aux_functions.R
\name{.cube_bands_check}
\alias{.cube_bands_check}
\title{Check if bands are part of a data cube}
\usage{
.cube_bands_check(cube, bands, add_cloud = TRUE)
}
\arguments{
\item{cube}{Data cube}

\item{bands}{Bands to be check}

\item{add_cloud}{Include the cloud band?}
}
\value{
No return value, called for side effects.
}
\description{
Check if bands are part of a data cube
}
\keyword{internal}
