% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_cube.R
\name{cube_functions}
\alias{cube_functions}
\alias{.cube_check}
\alias{.cube_satellite}
\alias{.cube_sensor}
\alias{.cube_collection}
\alias{.cube_name}
\alias{.cube_labels}
\alias{.cube_bands}
\alias{.cube_bands_check}
\alias{.cube_source}
\alias{.cube_timeline}
\alias{.cube_tiles}
\alias{.cube_tile_check}
\alias{.cube_tile_get_fields}
\alias{.cube_tile_crs}
\alias{.cube_tile_bbox}
\alias{.cube_tile_resolution}
\alias{.cube_tile_size}
\alias{.cube_band_missing_value}
\alias{.cube_band_minimum_value}
\alias{.cube_band_maximum_value}
\alias{.cube_band_scale_factor}
\alias{.cube_band_offset_value}
\alias{.cube_band_resolutions}
\alias{.cube_has_cloud}
\alias{.cube_s3class}
\title{Auxiliary functions for sits cubes}
\usage{
.cube_check(cube)

.cube_satellite(cube)

.cube_sensor(cube)

.cube_collection(cube)

.cube_name(cube)

.cube_labels(cube)

.cube_bands(cube, ..., add_cloud = TRUE)

.cube_bands_check(cube, bands, ..., add_cloud = TRUE)

.cube_source(cube)

.cube_timeline(cube)

.cube_tiles(cube)

.cube_tile_check(cube, tile)

.cube_tile_get_fields(cube, tile, fields = NULL)

.cube_tile_crs(cube, ..., tile = 1)

.cube_tile_bbox(cube, ..., tile = 1)

.cube_tile_resolution(cube, ..., tile = 1)

.cube_tile_size(cube, ..., tile = 1)

.cube_band_missing_value(cube, band)

.cube_band_minimum_value(cube, band)

.cube_band_maximum_value(cube, band)

.cube_band_scale_factor(cube, band)

.cube_band_offset_value(cube, band)

.cube_band_resolutions(cube, band)

.cube_has_cloud(cube)

.cube_s3class(cube)
}
\arguments{
\item{cube}{Data cube from where data is to be retrieved.}

\item{bands}{Bands to be included.}

\item{tile}{Tiles from the collection to be included in the data cube}

\item{....}{Additional parameters to be included.}
}
\value{
a \code{vector} for get attributes functions and NULL or error for
check parameters functions.
}
\description{
Auxiliary functions for extracting information from sits cubes.
}
\keyword{internal}
