% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_apply.R
\name{.apply_across}
\alias{.apply_across}
\title{Apply an expression across all bands}
\usage{
.apply_across(data, fn, ...)
}
\arguments{
\item{data}{Tile name.}
}
\value{
A sits tibble with all processed bands.
}
\description{
Apply an expression across all bands
}
\keyword{internal}
