% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_apply.R
\name{.apply_missing_band}
\alias{.apply_missing_band}
\title{Finds the missing bands in a cube}
\usage{
.apply_missing_band(cube, band)
}
\arguments{
\item{cube}{Data cube.}

\item{band}{Band name.}
}
\value{
List of combination among tiles, bands, and dates
                  that are missing from the cube.
}
\description{
Finds the missing bands in a cube
}
\keyword{internal}
