% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_apply.R, R/sits_spectral_unmixing.R
\name{.apply_missing_band}
\alias{.apply_missing_band}
\alias{.mesma_get_jobs_lst}
\title{Finds the missing bands in a cube}
\usage{
.apply_missing_band(cube, band)

.mesma_get_jobs_lst(cube, fractions)
}
\arguments{
\item{cube}{Data cube.}

\item{band}{Band name.}

\item{fractions}{Name of output fractions.}
}
\value{
List of combination among tiles, bands, and dates
                  that are missing from the cube.

List of combination among tiles, endmembers, and dates
                  that are missing from the cube.
}
\description{
Finds the missing bands in a cube

Create a list of jobs
}
\keyword{internal}
