% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_source_api_aws.R
\name{.aws_bands}
\alias{.aws_bands}
\title{...}
\usage{
.aws_bands(source, collection, s2_resolution)
}
\arguments{
\item{source}{Name of the STAC provider}

\item{collection}{...}

\item{s2_resolution}{...}
}
\value{
...
}
\description{
TODO: document
}
\keyword{internal}
