% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_check.R
\name{.check_cubes_match}
\alias{.check_cubes_match}
\title{Check if uncert cubes have the same organization}
\usage{
.check_cubes_match(cube1, cube2)
}
\arguments{
\item{cubes}{list of input data cubes}
}
\value{
TRUE/FALSE
}
\description{
Check if uncert cubes have the same organization
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
\keyword{internal}
