% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_check.R
\name{.check_cubes_same_bbox}
\alias{.check_cubes_same_bbox}
\title{Check if  cubes have the same bbox}
\usage{
.check_cubes_same_bbox(cube1, cube2)
}
\arguments{
\item{cubes}{list of input data cubes}
}
\value{
TRUE/FALSE
}
\description{
Check if  cubes have the same bbox
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
\keyword{internal}
