% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_check.R
\name{.check_cubes_same_timeline}
\alias{.check_cubes_same_timeline}
\title{Check if cubes have the same timeline}
\usage{
.check_cubes_same_timeline(cube1, cube2)
}
\arguments{
\item{cube1}{input data cube}

\item{cube2}{input data cube}
}
\value{
No return value, called for side effects.
}
\description{
Check if cubes have the same timeline
}
\keyword{internal}
