% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_check.R
\name{.check_error_matrix_area}
\alias{.check_error_matrix_area}
\title{Check if errox matrix and area are correct}
\usage{
.check_error_matrix_area(error_matrix, area)
}
\description{
Check if errox matrix and area are correct
}
\keyword{internal}
