% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_check.R
\name{.check_predicted}
\alias{.check_predicted}
\title{Does the input data contain a set of predicted values?}
\usage{
.check_predicted(data)
}
\arguments{
\item{data}{a sits tibble}
}
\value{
No return value, called for side effects.
}
\description{
Does the input data contain a set of predicted values?
}
\keyword{internal}
