% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_config.R
\name{.config_data_meta_type}
\alias{.config_data_meta_type}
\title{Check metatype associated to the data}
\usage{
.config_data_meta_type(data)
}
\arguments{
\item{data}{Time series or data cube.}
}
\value{
The meta data type associated to a sits object.
}
\description{
associates a valid SITS class to the data
}
\keyword{internal}
