% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_config.R
\name{.config_palette_colors}
\alias{.config_palette_colors}
\title{Get colors associated to the labels}
\usage{
.config_palette_colors(
  labels = labels,
  ...,
  palette = palette,
  brewer_palette = brewer_palette,
  brewer_order = brewer_order
)
}
\arguments{
\item{labels}{labels associated to the training classes}

\item{palette}{name of palette available in the config file}

\item{brewer_palette}{palette from RColorBrewer that replaces default palette
when labels are not included in the config palette}

\item{brewer_order}{invert the order of brewer colors?}
}
\value{
colors required to display the labels
}
\description{
Get colors associated to the labels
}
\keyword{internal}
