% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_cube_aux_functions.R
\name{.cube_create}
\alias{.cube_create}
\title{Creates the description of a data cube}
\usage{
.cube_create(
  source,
  collection = NA_character_,
  satellite,
  sensor,
  tile = NA_character_,
  bands,
  labels = NA_character_,
  xmin,
  xmax,
  ymin,
  ymax,
  crs,
  file_info = NULL
)
}
\arguments{
\item{source}{Source of data}

\item{collection}{Image collection}

\item{satellite}{Name of satellite}

\item{sensor}{Name of sensor}

\item{tile}{Tile of the image collection}

\item{bands}{Vector with the names of the bands.}

\item{labels}{Vector with labels (only for classified data).}

\item{xmin}{Spatial extent (xmin).}

\item{xmax}{Spatial extent (xmax).}

\item{ymin}{Spatial extent (ymin).}

\item{ymax}{Spatial extent (ymin).}

\item{crs}{CRS for cube (EPSG code or PROJ4 string).}

\item{file_info}{Tibble with information about stacks (for stacks)}
}
\value{
A tibble containing a data cube
}
\description{
Print information and save metadata about a data cube.
}
\keyword{internal}
