% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_tile_api.R
\name{.cube_foreach_tile}
\alias{.cube_foreach_tile}
\title{Cube API}
\usage{
.cube_foreach_tile(cube, fn, ...)
}
\arguments{
\item{cube}{A cube.}

\item{fn}{A function.}

\item{...}{Additional arguments to be passed to \code{fn}.}
}
\value{
cube
}
\description{
Iterates over each cube tile, passing tile to function's
first argument.
}
