% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_cube_aux_functions.R
\name{.cube_token_generator}
\alias{.cube_token_generator}
\title{...}
\usage{
.cube_token_generator(cube)
}
\arguments{
\item{cube}{input data cube}
}
\value{
A cube ...
}
\description{
...
}
\keyword{internal}
