% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_gdalcubes.R
\name{.gc_arrange_images}
\alias{.gc_arrange_images}
\title{Images arrangement in sits cube}
\usage{
.gc_arrange_images(cube, agg_method, duration, ...)
}
\arguments{
\item{cube}{A sits data cube}

\item{agg_method}{A \code{character} with method that will be applied by
\code{gdalcubes} for aggregation. Options: \code{min}, \code{max},
\code{mean}, \code{median} and \code{first}. Default is \code{median}.}

\item{duration}{A \code{Duration} object from lubridate package.}

\item{...}{Additional parameters.}
}
\value{
A sits cube with the images arranged according to some criteria.
}
\description{
Images arrangement in sits cube
}
\keyword{internal}
