% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_gdalcubes.R
\name{.gc_create_cube_view}
\alias{.gc_create_cube_view}
\title{Create a cube_view object}
\usage{
.gc_create_cube_view(tile, period, res, roi, date, agg_method, resampling)
}
\arguments{
\item{tile}{Data cube tile}

\item{period}{Period of time in which it is desired to apply in the cube,
must be provided based on ISO8601, where 1 number and a
unit are provided, for example "P16D".}

\item{res}{Spatial resolution of the image that
will be aggregated.}

\item{roi}{Region of interest.}

\item{agg_method}{Aggregation method.}

\item{resampling}{Resampling method.
Options: \code{near}, \code{bilinear}, \code{bicubic} or
others supported by gdalwarp
(see https://gdal.org/programs/gdalwarp.html).
Default is "bilinear".}

\item{toi}{Timeline of intersection.}
}
\value{
\code{Cube_view} object from gdalcubes.
}
\description{
Create a cube_view object
}
\keyword{internal}
