% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_gdalcubes.R
\name{.gc_missing_tiles}
\alias{.gc_missing_tiles}
\title{Finds the missing tiles in a regularized cube}
\usage{
.gc_missing_tiles(cube, local_cube, timeline)
}
\arguments{
\item{cube}{original cube to be regularized}

\item{timeline}{timeline used by gdalcube for regularized cube}

\item{gc_cube}{regularized cube (may be missing tile)}

\item{period}{period of timeline regularization}
}
\value{
tiles that are missing from the regularized cube
}
\description{
Finds the missing tiles in a regularized cube
}
\keyword{internal}
