% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_gdalcubes.R
\name{.gc_new_cube}
\alias{.gc_new_cube}
\title{Save the images based on an aggregation method.}
\usage{
.gc_new_cube(
  tile,
  cv,
  img_col,
  path_db,
  output_dir,
  cloud_mask,
  ...,
  version = "v1"
)
}
\arguments{
\item{tile}{A data cube tile}

\item{cv}{A \code{list} 'cube_view' with values from cube.}

\item{img_col}{A \code{object} 'image_collection' containing information
about the images metadata.}

\item{path_db}{Database to be created by gdalcubes}

\item{output_dir}{Directory where the aggregated images will be written.}

\item{cloud_mask}{A \code{logical} corresponds to the use of the cloud band
for aggregation.}

\item{...}{Additional parameters that can be included. See
'?gdalcubes::write_tif'.}

\item{version}{A \code{character} with version of the output files.}
}
\value{
A data cube tile with information used in its creation.
}
\description{
Save the images based on an aggregation method.
}
\keyword{internal}
