% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_gdalcubes.R
\name{.gc_new_cube}
\alias{.gc_new_cube}
\title{Save the images based on an aggregation method.}
\usage{
.gc_new_cube(
  tile,
  cv,
  roi,
  fill_method,
  img_col,
  path_db,
  output_dir,
  cloud_mask,
  multicores,
  ...
)
}
\arguments{
\item{tile}{A data cube tile}

\item{cv}{A \code{list} 'cube_view' with values from cube.}

\item{roi}{A region of interest.}

\item{fill_method}{A \code{character} indicating which interpolation method
will be applied. Options: \code{near} for nearest neighbor; \code{linear}
for linear interpolation; \code{locf} for ast observation carried forward,
or \code{nocb} for next observation carried backward.
Default is \code{near}..}

\item{img_col}{A \code{object} 'image_collection' containing information
about the images metadata.}

\item{path_db}{Database to be created by gdalcubes}

\item{output_dir}{Directory where the aggregated images will be written.}

\item{cloud_mask}{A \code{logical} corresponds to the use of the cloud band
for aggregation.}

\item{multicores}{A \code{numeric} with the number of cores will be used in
the regularize. By default is used 1 core.}

\item{...}{Additional parameters that can be included. See
'?gdalcubes::write_tif'.}
}
\value{
A data cube tile with information used in its creation.
}
\description{
Save the images based on an aggregation method.
}
\keyword{internal}
