% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_mixture_model.R
\name{.mesma_scale_endmembers}
\alias{.mesma_scale_endmembers}
\title{Scale the endmembers spectra}
\usage{
.mesma_scale_endmembers(cube, endmembers_spectra)
}
\arguments{
\item{cube}{Data cube.}

\item{endmembers_spectra}{Tibble with endmembers spectra values.}
}
\value{
a matrix with spectral values scaled.
}
\description{
Scale the endmembers spectra
}
\keyword{internal}
