% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_metrics_generate.R
\name{.metrics_blocks_estimate}
\alias{.metrics_blocks_estimate}
\title{Estimate the size of blocks}
\usage{
.metrics_blocks_estimate(cube, nbands, memsize)
}
\arguments{
\item{cube}{A data cube}

\item{nbands}{Number of bands}

\item{memsize}{Memory size}
}
\value{
a list with information about the beginning and end of each block
}
\description{
This function estimates the number of blocks using the number of
bands and the memory size
}
