% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_mvp.R
\name{.mvp_best_split_drop}
\alias{.mvp_best_split_drop}
\title{Get best split drop for MVP}
\usage{
.mvp_best_split_drop(X, dir = NULL, min_seg = 2L)
}
\arguments{
\item{X}{A sits time series tibble}

\item{dir}{The "direction" of the decrease in the time series value. With NDVI, the ts should be split when
the sharpest decrease occurs. With SWIR, the opposite holds true. Default: NULL; will split on sharpest decrease.}

\item{min_seg}{The smaller size eithr ts segment can have after the split Default 2L.}
}
\value{
A list Containing the max split score and the time step in which it occurred.
}
\description{
Scores all time series' time steps to determine the most
             likely time for an event like deforestation.
}
\author{
Alexandre Assuncao \email{alexcarssuncao@gmail.com}
}
