% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_raster_api.R
\name{.raster_crop}
\alias{.raster_crop}
\title{Raster package internal crop raster function}
\usage{
.raster_crop(r_obj, ..., block = NULL, bbox = NULL)
}
\arguments{
\item{r_obj}{raster package object to be written}

\item{...}{additional parameters to be passed to raster package}

\item{block}{numeric vector with names "first_col", "ncols", "first_row",
"nrows".}

\item{bbox}{numeric vector with names "xmin", "xmax", "ymin", "ymax".}
}
\value{
Subset of a raster object as defined by either block
               or bbox parameters
}
\description{
Raster package internal crop raster function
}
\note{
block starts at (1, 1)
}
\author{
Rolf Simoes, \email{rolf.simoes@inpe.br}
}
\keyword{internal}
