% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_raster_api.R
\name{.raster_merge}
\alias{.raster_merge}
\title{Merge all input files into one raster file}
\usage{
.raster_merge(
  in_files,
  out_file,
  format,
  gdal_datatype,
  multicores = 2,
  overwrite = TRUE
)
}
\arguments{
\item{in_files}{Input file paths}

\item{out_file}{Output raster file path}

\item{format}{Format to write the file}

\item{gdal_datatype}{Data type in gdal format}

\item{multicores}{Number of cores to process merging}

\item{overwrite}{Should the output file be overwritten?}
}
\value{
No return value, called for side effects.
}
\description{
Merge all input files into one raster file
}
\author{
Rolf Simoes, \email{rolf.simoes@inpe.br}
}
\keyword{internal}
