% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_raster_api.R
\name{.raster_merge}
\alias{.raster_merge}
\title{Merge all input files into one raster file}
\usage{
.raster_merge(
  in_files,
  out_file,
  format,
  gdal_datatype,
  gdal_options,
  overwrite
)
}
\arguments{
\item{in_files}{Input file paths}

\item{out_file}{Output raster file path}

\item{format}{Format to write the file}

\item{gdal_datatype}{Data type in gdal format}

\item{gdal_options}{Compression method to be used}

\item{overwrite}{Overwrite the file?}
}
\description{
Merge all input files into one raster file
}
\author{
Rolf Simoes, \email{rolf.simoes@inpe.br}
}
\keyword{internal}
