% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_raster_api.R
\name{.raster_read_rast}
\alias{.raster_read_rast}
\title{Raster package internal read raster file function}
\usage{
.raster_read_rast(files, ..., block = NULL, method = "bilinear")
}
\arguments{
\item{...}{additional parameters to be passed to raster package}

\item{block}{a valid block with (\code{col}, \code{row},
\code{ncols}, \code{nrows}).}

\item{method}{method used to resample pixels (used only in case of
out_size parameter is informed)}

\item{file}{path to raster file(s) to be read}
}
\value{
Numeric matrix read from file based on parameter block
}
\description{
Raster package internal read raster file function
}
\author{
Rolf Simoes, \email{rolf.simoes@inpe.br}
}
\keyword{internal}
