% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_raster_api.R
\name{.raster_read_rast}
\alias{.raster_read_rast}
\title{Raster package internal read raster file function}
\usage{
.raster_read_rast(file, ..., block = NULL)
}
\arguments{
\item{file}{raster file to be read}

\item{...}{additional parameters to be passed to raster package}

\item{block}{numeric vector with names "first_col", "ncols", "first_row", "nrows".}
}
\value{
numeric matrix
}
\description{
Raster package internal read raster file function
}
\author{
Rolf Simoes, \email{rolf.simoes@inpe.br}
}
\keyword{internal}
