% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_raster_api.R
\name{.raster_read_stack}
\alias{.raster_read_stack}
\title{Raster package internal read raster stack file function}
\usage{
.raster_read_stack(files, block = NULL, ...)
}
\arguments{
\item{files}{raster files to be read}

\item{block}{numeric vector with names "col", "ncols", "row", "nrows".}

\item{...}{additional parameters to be passed to raster package}
}
\value{
numeric matrix
}
\description{
Raster package internal read raster stack file function
}
\author{
Rolf Simoes, \email{rolf.simoes@inpe.br}
}
\keyword{internal}
