% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_raster_api.R
\name{.raster_read_stack}
\alias{.raster_read_stack}
\title{Raster package internal read raster stack file function}
\usage{
.raster_read_stack(
  files,
  ...,
  block = NULL,
  out_size = NULL,
  method = "bilinear"
)
}
\arguments{
\item{files}{raster files to be read}

\item{...}{additional parameters to be passed to raster package}

\item{block}{numeric vector with names "first_col", "ncols",
"first_row", "nrows".}

\item{out_size}{a numeric vector with names "nrows" and "ncols"}

\item{method}{method used to resample pixels (used only in case of
out_size parameter is informed)}
}
\value{
Numeric matrix read from file based on parameter block
}
\description{
Raster package internal read raster stack file function
}
\author{
Rolf Simoes, \email{rolf.simoes@inpe.br}
}
\keyword{internal}
