% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_raster_api.R
\name{.raster_resample_method}
\alias{.raster_resample_method}
\alias{.raster_resample_methods}
\title{Convert sits internal resample methods}
\usage{
.raster_resample_method(method)

.raster_resample_methods()
}
\value{
character string
}
\description{
Convert sits internal resample methods
}
\author{
Rolf Simoes, \email{rolf.simoes@inpe.br}
}
\keyword{internal}
