% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_regularize.R
\name{.reg_composite_image}
\alias{.reg_composite_image}
\title{Generate a composite image based on tile interval.}
\usage{
.reg_composite_image(
  tile_band_period,
  res,
  blocks,
  ratio_band_out,
  ratio_cloud_out,
  date,
  output_dir,
  multithreads
)
}
\arguments{
\item{res}{A \code{numeric} with spatial resolution of the image
that will be aggregated.}

\item{blocks}{A \code{list} with: blocks of output band,
blocks of input band, and blocks of cloud band}

\item{date}{A \code{Date} of reference to generate the image
composite}

\item{output_dir}{A \code{character} with a valid directory where the
regularized images will be written.}

\item{multithreads}{A \code{integer} with number of cores to process each
image in the interval}

\item{tile_period_band}{A unique tile from \code{sits_cube} object}
}
\value{
A data cube tile with information used in its creation.
}
\description{
Generate a composite image based on tile interval.
}
\keyword{internal}
