% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_regularize_aux_functions.R
\name{.reg_filename}
\alias{.reg_filename}
\title{Create the merge image filename}
\usage{
.reg_filename(tile, band, date, output_dir, ..., block = NULL)
}
\arguments{
\item{tile}{A unique tile from \code{sits_cube} object.}

\item{band}{Band to be written.}

\item{date}{Vector with two position, first one is
the start date and second one is the end date.}

\item{output_dir}{Directory where the regularized images will be written.}

\item{block}{Vector with information about a block.}
}
\value{
File name of resampled image.
}
\description{
Create the merge image filename
}
\keyword{internal}
