% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_gdalcubes.R
\name{.reg_new_cube}
\alias{.reg_new_cube}
\title{Save the images based on an aggregation method.}
\usage{
.reg_new_cube(
  tile_period_band,
  res,
  blocks,
  date_period,
  resampling,
  roi,
  output_dir,
  multithreads
)
}
\arguments{
\item{tile_period_band}{A unique tile from \code{sits_cube} object}

\item{res}{A \code{numeric} with spatial resolution of the image
that will be aggregated.}

\item{blocks}{...}

\item{date_period}{...}

\item{resampling}{A \code{character} with method to be used  for resampling in mosaic operation.
Options: \code{near}, \code{bilinear}, \code{bicubic}, \code{cubicspline},
and \code{lanczos}.
Default is bilinear.}

\item{roi}{A named \code{numeric} vector with a region of interest.}

\item{output_dir}{A \code{character} with a valid directory where the
regularized images will be written.}

\item{multithreads}{...}

\item{period}{A \code{character} vector with two position, first one is
the start date and second one is the end date.}

\item{agg_method}{A \code{character} with method that will be applied by
\code{gdalcubes} for aggregation. Options: \code{median} and
\code{least_cc_first}. The default aggregation method is
\code{least_cc_first}. See more above.}
}
\value{
A data cube tile with information used in its creation.
}
\description{
Save the images based on an aggregation method.
}
\keyword{internal}
