% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_regularize.R
\name{.reg_preprocess_block}
\alias{.reg_preprocess_block}
\title{Preprocessing steps of sits regularize}
\usage{
.reg_preprocess_block(
  tile_fid,
  band,
  out_size,
  band_block,
  cloud_block,
  output_block,
  ratio_band_out,
  ratio_cloud_out,
  data_type,
  output_file
)
}
\arguments{
\item{tile}{A unique tile from \code{sits_cube} object}

\item{band_paths}{Paths for one band}

\item{resolution}{Spatial resolution of the image
that will be aggregated.}

\item{resampling}{Resampling method.
\code{cubicspline}, and \code{lanczos}.}

\item{block}{Information about a block.}
}
\value{
A \code{SpatRast} object resampled.
}
\description{
Preprocessing steps of sits regularize
}
\keyword{internal}
