% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_gdalcubes.R
\name{.reg_preprocess_rast}
\alias{.reg_preprocess_rast}
\alias{.reg_preprocess}
\title{Preprocessing steps of sits regularize}
\usage{
.reg_preprocess(
  tile,
  band,
  band_path,
  cloud_path,
  cloud_block,
  band_block,
  resolution,
  resampling,
  datatype,
  filename,
  cloud_interp,
  missing_value
)
}
\arguments{
\item{tile}{A unique tile from \code{sits_cube} object}

\item{resolution}{A \code{numeric} with spatial resolution of the image
that will be aggregated.}

\item{resampling}{A \code{character} with method to be used  for resampling
in mosaic operation. Options: \code{near}, \code{bilinear}, \code{bicubic},
\code{cubicspline}, and \code{lanczos}. Default is bilinear.}

\item{band_paths}{A \code{character} with paths for a unique band}

\item{block}{A \code{numeric} vector with information about a block}
}
\value{
A \code{SpatRast} object resampled
}
\description{
Preprocessing steps of sits regularize
}
\keyword{internal}
