% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_regularize_aux_functions.R
\name{.reg_timeline}
\alias{.reg_timeline}
\title{Compute the timeline of the regularized cube}
\usage{
.reg_timeline(cube, period)
}
\arguments{
\item{cube}{Data cube from where data is to be retrieved.}

\item{period}{A \code{character} with ISO8601 time period for regular
data cubes produced by \code{gdalcubes}, with number and unit, e.g., "P16D"
for 16 days. Use "D", "M" and "Y" for days, month and year.}
}
\value{
a \code{vector} with all timeline values.
}
\description{
Compute the timeline of the regularized cube
}
\keyword{internal}
