% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_aws.R
\name{.sits_aws_check_access}
\alias{.sits_aws_check_access}
\title{Check access rights on AWS}
\usage{
.sits_aws_check_access(
  type,
  access_key = NULL,
  secret_key = NULL,
  region = NULL
)
}
\arguments{
\item{type}{cube type}

\item{access_key}{AWS access key}

\item{secret_key}{AWS secret key}

\item{region}{AWS region}
}
\value{
TRUE if access is granted
}
\description{
Check access rights on AWS
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
\keyword{internal}
