% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_bbox.R
\name{.sits_bbox_to_sf}
\alias{.sits_bbox_to_sf}
\title{Convert a bounding box to a sf object (polygon)}
\usage{
.sits_bbox_to_sf(xmin, xmax, ymin, ymax, crs)
}
\arguments{
\item{xmin, xmax, ymin, ymax}{a numeric value with bound box values}

\item{crs}{a valid crs value}
}
\value{
a sf object
}
\description{
Convert a bounding box to a sf object (polygon)
}
\keyword{internal}
