% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_space_time_operations.R
\name{.sits_binary_search}
\alias{.sits_binary_search}
\title{Implement a binary search to find the nearest date}
\usage{
.sits_binary_search(values, val)
}
\arguments{
\item{values}{A vector of ordered integers.}

\item{val}{A reference value.}
}
\value{
The input value closest to the reference one.
}
\description{
Performs a binary search in a ordered set of integer values and
             returns the values closest to a reference.
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
