% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_raster_utils.R
\name{.sits_check_classify_params}
\alias{.sits_check_classify_params}
\title{Check clasification parameters}
\usage{
.sits_check_classify_params(file, cube, ml_model)
}
\arguments{
\item{file}{Vector of file names to store the output (one file per classified year).}

\item{cube}{Tibble with information about a data cube.}

\item{ml_model}{An R model trained by \code{\link[sits]{sits_train}}.}
}
\value{
Tests succeeded?
}
\description{
Verify that required parameters are correct.
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
