% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_raster_utils.R
\name{.sits_check_results}
\alias{.sits_check_results}
\title{Check the results of the classification with the input data}
\usage{
.sits_check_results(prediction, nrows_DT)
}
\arguments{
\item{prediction}{Prototype with predictions.}

\item{nrows_DT}{Number of rows of distance matrix.}
}
\value{
TRUE if check is OK.
}
\description{
Verify that classification results have the right size.
}
\author{
Rolf Simoes, \email{rolf.simoes@inpe.br}
}
