% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_classification.R
\name{.sits_classify_interval}
\alias{.sits_classify_interval}
\title{Classify one interval of data}
\usage{
.sits_classify_interval(DT, time, ml_model, first_row, multicores)
}
\arguments{
\item{DT}{A data.table with distance values.}

\item{time}{Time interval to be processed.}

\item{ml_model}{Machine learning model to be applied.}

\item{first_row}{Initial row of the output layer to write block.}

\item{multicores}{Number of cores to process the time series.}
}
\value{
A data table with predicted values of probs
}
\description{
Classify one interval of data
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
