% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_classify_cube.R
\name{.sits_classify_multicores}
\alias{.sits_classify_multicores}
\title{Classify a chunk of raster data  using multicores}
\usage{
.sits_classify_multicores(
  cube,
  samples,
  ml_model,
  roi,
  filter,
  memsize,
  multicores,
  output_dir,
  version
)
}
\arguments{
\item{cube}{data cube}

\item{samples}{samples used for training the classification model.}

\item{ml_model}{model trained by \code{\link[sits]{sits_train}}.}

\item{roi}{region of interest}

\item{filter}{smoothing filter to be applied to the data.}

\item{memsize}{memory available for classification (in GB).}

\item{multicores}{number of cores.}

\item{output_dir}{output directory}

\item{version}{version of result}
}
\value{
List of the classified raster layers.
}
\description{
Classifies a block of data using multicores. It breaks
the data into horizontal blocks and divides them between the available cores.

Reads data using terra, cleans the data for NAs and missing values.
The clean data is stored in a data table with the time instances
for all pixels of the block. The algorithm then classifies data on
an year by year basis. For each year, extracts the sub-blocks for each band.

After all cores process their blocks, it joins the result and then writes it
in the classified images for each corresponding year.
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
\keyword{internal}
