% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_cube_classification.R
\name{.sits_classify_multicores}
\alias{.sits_classify_multicores}
\title{Classify a raster chunk using multicores}
\usage{
.sits_classify_multicores(cube, cube_class, samples, ml_model, interval,
  filter, memsize, multicores)
}
\arguments{
\item{cube}{Tibble with metadata for a data cube derived from a raster brick.}

\item{cube_class}{Raster layer objects to be written.}

\item{samples}{Tibble with samples used for training the classification model.}

\item{ml_model}{A model trained by \code{\link[sits]{sits_train}}.}

\item{interval}{Classification interval.}

\item{filter}{Smoothing filter to be applied to the data.}

\item{memsize}{Memory available for classification (in GB).}

\item{multicores}{Number of cores.}
}
\value{
List of the classified raster layers.
}
\description{
Classifies a block of data using multicores. It breaks
the data into horizontal blocks and divides them between the available cores.

Reads data from a file using Rgdal, then cleans the data for NAs and missing values. The clean
data is stored in a data table that has all the time instances for all pixels of
the block. The algorithm then classifies data on an year by year basis.
For each year, it extracts the sub-blocks for each band.

After all cores process their blocks, it joins the result and then writes it
in the classified images for each corresponding year.
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
