% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_classify_ts.R
\name{.sits_classify_ts}
\alias{.sits_classify_ts}
\title{Classify a distances tibble using machine learning models}
\usage{
.sits_classify_ts(samples, ml_model, filter_fn, multicores, progress)
}
\arguments{
\item{samples}{a tibble with sits samples}

\item{ml_model}{model trained by \code{\link[sits]{sits_train}}.}

\item{filter_fn}{Smoothing filter to be applied (if desired).}

\item{multicores}{number of threads to process the time series.}

\item{progress}{Show progress bar?}
}
\value{
A data.table with the predicted labels.
}
\description{
Returns a sits tibble with the results of the ML classifier.
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
\keyword{internal}
