% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_clouds.R
\name{.sits_clouds_interpolate}
\alias{.sits_clouds_interpolate}
\title{Create the output of the cloud estimation procedure}
\usage{
.sits_clouds_interpolate(cube, data_dir, blocks, impute_fn, multicores)
}
\arguments{
\item{cube}{input data cube}

\item{data_dir}{directory where data is to be stored}

\item{blocks}{block information}

\item{impute_fn}{imputation function to remove NA}

\item{multicores}{number of cores to use}
}
\value{
a tibble with date, band and path information.
}
\description{
Create the output of the cloud estimation procedure
}
\keyword{internal}
