% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_config.R
\name{.sits_config_default_sensor}
\alias{.sits_config_default_sensor}
\title{Retrieve the default sensor for the satellite}
\usage{
.sits_config_default_sensor(satellite)
}
\arguments{
\item{satellite}{Name of the satellite}
}
\value{
A best guess for the sensor
}
\description{
Based on the satellite, find the default sensor
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
