% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_config.R
\name{.sits_config_maximum_values}
\alias{.sits_config_maximum_values}
\title{Retrieve the maximum values for a given band}
\usage{
.sits_config_maximum_values(sensor, bands)
}
\arguments{
\item{sensor}{Name of the sensor}

\item{bands}{Vector of bands.}
}
\value{
The maximum values.
}
\description{
Retrieve the maximum values for a given band
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
