% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_config.R
\name{.sits_config_sensors}
\alias{.sits_config_sensors}
\title{List the sensors supported per satellite}
\usage{
.sits_config_sensors(satellite)
}
\arguments{
\item{satellite}{Name of the satellite}
}
\value{
List of sensors associated to a satellite that are supported by SITS
}
\description{
List the sensors supported per satellite
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
