% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_config.R
\name{.sits_config_sensors}
\alias{.sits_config_sensors}
\title{Retrieve the default sensor for the satellite}
\usage{
.sits_config_sensors(satellite)

.sits_config_sensors(satellite)
}
\arguments{
\item{satellite}{Name of the satellite}
}
\value{
Sensors associated to the satellite

List of sensors associated to a satellite that are supported by SITS
}
\description{
Based on the satellite, find the default sensor
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}

Gilberto Camara, \email{gilberto.camara@inpe.br}
}
