% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_stac.R
\name{.sits_config_stac_values}
\alias{.sits_config_stac_values}
\title{Get the metadata values from STAC.}
\usage{
.sits_config_stac_values(collection_info, bands)
}
\arguments{
\item{collection_info}{a \code{STACCollection} object returned by rstac.
package.}

\item{bands}{a \code{character} with the bands names to be
filtered.}
}
\value{
a \code{list} with the information of scale factors,
 missing, minimum, and maximum values.
}
\description{
Get the metadata values from STAC.
}
\keyword{internal}
