% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_space_time_operations.R
\name{.sits_convert_resolution}
\alias{.sits_convert_resolution}
\title{Convert resolution from projection values to lat/long}
\usage{
.sits_convert_resolution(cube)
}
\arguments{
\item{cube}{Metadata about a data cube.}
}
\value{
A matrix with resolution in WGS84 coordinates.
}
\description{
Transform a latitude and longitude coordinate to a XY projection coordinate
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
