% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_coverage.R
\name{.sits_coverage_SATVEG}
\alias{.sits_coverage_SATVEG}
\title{Provides information about one coverage of the SATVEG time series service}
\usage{
.sits_coverage_SATVEG(name, timeline)
}
\arguments{
\item{name}{name of the coverage}

\item{timeline}{timeline of the coverage}
}
\description{
creates a tibble with metadata about a given coverage
}
