% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_coverage.R
\name{.sits_coverageWTSS}
\alias{.sits_coverageWTSS}
\title{Provides information about one coverage of the WTSS service}
\usage{
.sits_coverageWTSS(wtss.obj, service, name)
}
\arguments{
\item{wtss.obj}{R WTSS object associated to the service}

\item{service}{name of the service}

\item{name}{name of the coverage}
}
\description{
uses the WTSS services to print information and save metadata about a
chosen coverage
}
