% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_coverage.R
\name{.sits_create_coverage}
\alias{.sits_create_coverage}
\title{Creates a coverage metadata}
\usage{
.sits_create_coverage(r_objs.lst, name, service, bands.vec, labels.vec,
  scale_factors.vec, missing_values.vec, minimum_values.vec, timeline.lst,
  nrows, ncols, xmin, xmax, ymin, ymax, xres, yres, crs, files.vec)
}
\arguments{
\item{r_objs.lst}{List of raster objects contained in the coverage.}

\item{name}{Name of the coverage.}

\item{service}{Name of the time series service.}

\item{bands.vec}{Vector with the names of the bands.}

\item{labels.vec}{Vector with labels (only valid for classified data).}

\item{scale_factors.vec}{Vector with scale factor for each band.}

\item{missing_values.vec}{Vector with missing values for each band.}

\item{minimum_values.vec}{Vector with minimum values for each band.}

\item{timeline.lst}{List with vectors of valid timelines for each band.}

\item{nrows}{Number of rows in the coverage.}

\item{ncols}{Number of columns in the coverage.}

\item{xmin}{Spatial extent (xmin).}

\item{xmax}{Spatial extent (xmax).}

\item{ymin}{Spatial extent (ymin).}

\item{ymax}{Spatial extent (ymin).}

\item{xres}{Spatial resolution (x dimension).}

\item{yres}{Spatial resolution (y dimension).}

\item{crs}{CRS for coverage.}

\item{files.vec}{Vector with associated files.}
}
\description{
Uses the configuration file to print information and save metadata about a
chosen coverage.
}
