% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_metrics_generate.R
\name{.sits_create_metrics_bands}
\alias{.sits_create_metrics_bands}
\title{Create the images metrics}
\usage{
.sits_create_metrics_bands(
  cube,
  output_dir,
  blocks,
  list_metrics,
  impute_fn,
  multicores
)
}
\arguments{
\item{cube}{input data cube}

\item{output_dir}{directory where data is to be stored}

\item{blocks}{block information}

\item{list_metrics}{a named \code{list} with bands name and metrics to be
generated}

\item{impute_fn}{imputation function to remove NA}

\item{multicores}{number of cores to use}
}
\value{
a tibble with date, band and path information.
}
\description{
Create the images metrics
}
\keyword{internal}
