% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_coverage.R
\name{.sits_create_raster_coverage}
\alias{.sits_create_raster_coverage}
\title{Creates a tibble with information about a set of raster bricks}
\usage{
.sits_create_raster_coverage(raster.lst, service, name, timeline.lst,
  bands.vec, labels.vec, scale_factors.vec, missing_values.vec,
  minimum_values.vec, files.vec)
}
\arguments{
\item{raster.lst}{List of Raster objects associated with the raster coverages.}

\item{service}{Time series service.}

\item{name}{Name of the coverage.}

\item{timeline.lst}{List of coverage timelines.}

\item{bands.vec}{Vector with names of bands.}

\item{labels.vec}{Vector of labels for classified image.}

\item{scale_factors.vec}{Vector of scale factors.}

\item{missing_values.vec}{Vector of missing values.}

\item{minimum_values.vec}{Vector of minimum values.}

\item{files.vec}{Vector of names of raster files where the data is stored.}
}
\description{
Creates a tibble with metadata about a given coverage.
}
