% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_csv.R
\name{.sits_csv_check}
\alias{.sits_csv_check}
\title{Check if a CSV tibble is valid}
\usage{
.sits_csv_check(csv)
}
\arguments{
\item{csv}{Tibble read from a CSV file}
}
\value{
A logical value
}
\description{
Check if a CSV tibble is valid
}
\keyword{internal}
