% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_cube_aux_functions.R
\name{.sits_cube_classified}
\alias{.sits_cube_classified}
\title{Create a set of RasterLayer objects to store
data cube classification results (only the probs)}
\usage{
.sits_cube_classified(cube, samples, sub_image, output_dir, version)
}
\arguments{
\item{cube}{input data cube.}

\item{samples}{samples used for training the classification model.}

\item{sub_image}{bounding box of the ROI}

\item{output_dir}{prefix of the output files.}

\item{version}{version of the output files}
}
\value{
output data cube
}
\description{
Take a tibble containing metadata about a data cube
containing time series (each Brick has information for one band) and create a
set of RasterLayers to store the classification result.
Each RasterLayer corresponds to one time step.
The time steps are specified in a list of dates.
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
\keyword{internal}
