% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_cube_aux_functions.R
\name{.sits_cube_create}
\alias{.sits_cube_create}
\title{Creates the description of a data cube}
\usage{
.sits_cube_create(
  name,
  source,
  collection = NA_character_,
  satellite,
  sensor,
  tile = NA_character_,
  bands,
  labels = NA_character_,
  nrows,
  ncols,
  xmin,
  xmax,
  ymin,
  ymax,
  xres,
  yres,
  crs,
  file_info = NULL
)
}
\arguments{
\item{name}{Name of the data cube (mandatory)}

\item{source}{Source of data}

\item{collection}{Image collection}

\item{satellite}{Name of satellite}

\item{sensor}{Name of sensor}

\item{tile}{Tile of the image collection}

\item{bands}{Vector with the names of the bands.}

\item{labels}{Vector with labels (only for classified data).}

\item{nrows}{Number of rows in the cube.}

\item{ncols}{Number of columns in the cube.}

\item{xmin}{Spatial extent (xmin).}

\item{xmax}{Spatial extent (xmax).}

\item{ymin}{Spatial extent (ymin).}

\item{ymax}{Spatial extent (ymin).}

\item{xres}{Spatial resolution (x dimension).}

\item{yres}{Spatial resolution (y dimension).}

\item{crs}{CRS for cube (EPSG code or PROJ4 string).}

\item{file_info}{Tibble with information about stacks (for stacks)}
}
\value{
A tibble containing a data cube
}
\description{
Print information and save metadata about a data cube.
}
\keyword{internal}
