% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_cube.R
\name{.sits_cube_create}
\alias{.sits_cube_create}
\title{Creates the description of a data cube}
\usage{
.sits_cube_create(
  service,
  URL,
  satellite,
  sensor,
  name,
  bands,
  labels,
  scale_factors,
  missing_values,
  minimum_values,
  maximum_values,
  timelines,
  nrows,
  ncols,
  xmin,
  xmax,
  ymin,
  ymax,
  xres,
  yres,
  crs,
  files = NULL
)
}
\arguments{
\item{service}{Web service with metadata about the cube.}

\item{URL}{URL of the provider}

\item{satellite}{Name of satellite}

\item{sensor}{Name of sensor}

\item{name}{Name of the data cube.}

\item{bands}{Vector with the names of the bands.}

\item{labels}{Vector with labels (only for classified data).}

\item{scale_factors}{Vector with scale factor for each band.}

\item{missing_values}{Vector with missing values for each band.}

\item{minimum_values}{Vector with minimum values for each band.}

\item{maximum_values}{Vector with maximum values for each band.}

\item{timelines}{List with vectors of valid timelines for each band.}

\item{nrows}{Number of rows in the cube.}

\item{ncols}{Number of columns in the cube.}

\item{xmin}{Spatial extent (xmin).}

\item{xmax}{Spatial extent (xmax).}

\item{ymin}{Spatial extent (ymin).}

\item{ymax}{Spatial extent (ymin).}

\item{xres}{Spatial resolution (x dimension).}

\item{yres}{Spatial resolution (y dimension).}

\item{crs}{CRS for cube.}

\item{files}{Vector with associated files.}
}
\description{
Print information and save metadata about a data cube.
}
