% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_cube.R
\name{.sits_cube_eocubes}
\alias{.sits_cube_eocubes}
\title{Uses the EOCUBES service to provide information about a data cube}
\usage{
.sits_cube_eocubes(remote.obj, service, URL, name, bands, tiles, geom,
  from, to)
}
\arguments{
\item{remote.obj}{Remote object.}

\item{service}{Name of the data service.}

\item{URL}{URL of the service provider.}

\item{name}{Name of the cube}

\item{bands}{Bands of the cube.}

\item{tiles}{Filter tiles by prefix name.}

\item{geom}{Geometry to filter tiles.}

\item{from}{Start date to be filtered.}

\item{to}{End date to be filtered.}
}
\description{
Creates a tibble with metadata about a data cube.
}
