% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_tibble_aux_functions.R
\name{.sits_extract}
\alias{.sits_extract}
\title{Extract a subset of the data based on dates}
\usage{
.sits_extract(data, start_date, end_date)
}
\arguments{
\item{data}{A sits tibble.}

\item{start_date}{Starting date of the time series segment.}

\item{end_date}{End date of the time series segment.}
}
\value{
A tibble in sits format with the chosen subset.
}
\description{
Returns a vector containing the dates of a sits tibble.
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
\keyword{internal}
