% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_get_data.R
\name{.sits_from_csv}
\alias{.sits_from_csv}
\title{Obtain timeSeries from a data cube, based on a CSV file.}
\usage{
.sits_from_csv(csv_file, cube, bands, prefilter, .n_start, .n_max, .n_save)
}
\arguments{
\item{csv_file}{Name of a CSV file with information <id, latitude, longitude, from, end, label>.}

\item{cube}{A tibble with metadata about the data cube which contains data to be retrieved.}

\item{bands}{A string vector with the names of the bands to be retrieved.}

\item{prefilter}{String ("0" - none, "1" - no data correction, "2" - cloud correction, "3" - no data and cloud correction).}

\item{.n_start}{Row on the CSV file to start reading (optional).}

\item{.n_max}{Maximum number of samples to be read.}

\item{.n_save}{Number of samples to save as intermediate files (used for long reads).}
}
\value{
A sits tibble.
}
\description{
reads descriptive information about a set of
spatio-temporal locations from a CSV file. Then, it retrieve the time series from a data cube,
and stores the time series on a sits tibble for later use.
The CSV file should have the following column names:
"longitude", "latitude", "start_date", "end_date", "label"
}
