% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_getdata.R
\name{.sits_from_raster}
\alias{.sits_from_raster}
\title{Extract a time series from a ST raster data set}
\usage{
.sits_from_raster(coverage, longitude, latitude, start_date = NULL,
  end_date = NULL, label = "NoClass")
}
\arguments{
\item{coverage}{A tibble with metadata describing a raster coverage.}

\item{longitude}{A double with the longitude of the chosen location.}

\item{latitude}{A double with the latitude of the chosen location.}

\item{start_date}{A date with the start of the period.}

\item{end_date}{A date with the end of the period.}

\item{label}{A string with the label to attach to the time series.}
}
\value{
A sits tibble with the time series.
}
\description{
Reads metadata about a raster data set to retrieve a set of
time series.
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
