% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_get_data.R
\name{.sits_from_raster}
\alias{.sits_from_raster}
\title{Extract a time series from a ST raster data set}
\usage{
.sits_from_raster(
  cube,
  longitude,
  latitude,
  start_date,
  end_date,
  bands,
  label = "NoClass"
)
}
\arguments{
\item{cube}{Metadata describing a raster data cube.}

\item{longitude}{Longitude of the chosen location.}

\item{latitude}{Latitude of the chosen location.}

\item{start_date}{Start of the period.}

\item{end_date}{End of the period.}

\item{bands}{Bands to be retrieved.}

\item{label}{Label to attach to the time series.}
}
\value{
A sits tibble with the time series.
}
\description{
Retrieve a set of time series for a raster data cube.
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
