% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_get_data.R
\name{.sits_from_service}
\alias{.sits_from_service}
\title{Obtain timeSeries from a web service associated to data cubes}
\usage{
.sits_from_service(cube, longitude, latitude, start_date, end_date, bands,
  label = "NoClass", .prefilter = "1")
}
\arguments{
\item{cube}{Data cube metadata.}

\item{longitude}{Longitude of the chosen location.}

\item{latitude}{Latitude of the chosen location).}

\item{start_date}{Optional start date of the period.}

\item{end_date}{Optional end date of the period.}

\item{bands}{Optional string vector with the names of the bands to be retrieved.}

\item{label}{String with the label to attach to the time series.}

\item{.prefilter}{String (only for SATVEG) ("0" - none, "1" - no data correction, "2" - cloud correction, "3" - no data and cloud correction).}
}
\value{
A sits tibble.
}
\description{
Obtains a time series from a time series service.
}
