% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_shp.R
\name{.sits_from_shp}
\alias{.sits_from_shp}
\title{Obtain timeSeries from WTSS server, based on a SHP file.}
\usage{
.sits_from_shp(
  shp_file,
  cube,
  start_date,
  end_date,
  bands,
  label,
  shp_attr,
  .n_shp_pol,
  .n_shp_pts,
  .prefilter
)
}
\arguments{
\item{shp_file}{SHP file which provides the boundaries of a region.}

\item{cube}{Data cube metadata.}

\item{start_date}{Start date of the period.}

\item{end_date}{End date of the period.}

\item{bands}{Names of the bands to be retrieved.}

\item{label}{Label to attach to the time series.}

\item{shp_attr}{Attribute in the shapefile used as a polygon label}

\item{.n_shp_pol}{Number of samples per polygon to be read
(for POLYGON or MULTIPOLYGON shapes)}

\item{.n_shp_pts}{Number of points to be read (for POINT shapes)}

\item{.prefilter}{Filtering for SATVEG cube
("0" - none, "1" - no data correction,
"2" - cloud correction,
"3" - no data and cloud correction).}
}
\value{
A sits tibble.
}
\description{
reads a shapefile and retrieves a sits tibble
containing time series from a data cube that are inside the SHP file.
The script uses the WTSS service, taking information about spatial and
temporal resolution from the WTSS configuration.
}
