% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_get_data.R
\name{.sits_from_shp}
\alias{.sits_from_shp}
\title{Obtain timeSeries from WTSS server, based on a SHP file.}
\usage{
.sits_from_shp(shp_file, coverage, start_date = NULL, end_date = NULL,
  bands = NULL, prefilter = "1", label = "NoClass")
}
\arguments{
\item{shp_file}{Name of a SHP file which provides the boundaries of a region of interest.}

\item{coverage}{A tibble with metadata about the coverage.}

\item{start_date}{The start date of the period.}

\item{end_date}{The end date of the period.}

\item{bands}{A string vector with the names of the bands to be retrieved.}

\item{prefilter}{A string related to data correction ("0" - none, "1" - no data correction, "2" - cloud correction, "3" - no data and cloud correction).}

\item{label}{A string with the label to attach to the time series.}
}
\value{
A sits tibble.
}
\description{
reads a shapefile and retrieves a sits tibble
containing time series from a coverage that are inside the SHP file.
The script uses the WTSS service, taking information about coverage, spatial and
temporal resolution from the WTSS configuration.
}
