% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_getdata.R
\name{.sits_fromCSV}
\alias{.sits_fromCSV}
\title{Obtain timeSeries from time series server, based on a CSV file.}
\usage{
.sits_fromCSV(csv_file, coverage, bands, prefilter, .n_start, .n_max,
  .n_save)
}
\arguments{
\item{csv_file}{Name of a CSV file with information <id, latitude, longitude, from, end, label>.}

\item{coverage}{A tibble with metadata about coverage which contains data to be retrieved.}

\item{bands}{A string vector with the names of the bands to be retrieved.}

\item{prefilter}{String ("0" - none, "1" - no data correction, "2" - cloud correction, "3" - no data and cloud correction).}

\item{.n_start}{Row on the CSV file to start reading (optional).}

\item{.n_max}{Maximum number of samples to be read.}

\item{.n_save}{Number of samples to save as intermediate files (used for long reads).}
}
\value{
A sits tibble.
}
\description{
reads descriptive information about a set of
spatio-temporal locations from a CSV file. Then, it uses the WTSS time series service
to retrieve the time series, and stores the time series on a sits tibble for later use.
The CSV file should have the following column names:
"longitude", "latitude", "start_date", "end_date", "label"
}
