% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_gdalcubes.R
\name{.sits_gc_cloud_mask}
\alias{.sits_gc_cloud_mask}
\title{Create an object image_mask with information about mask band}
\usage{
.sits_gc_cloud_mask(cube)
}
\arguments{
\item{cube}{Data cube from where data is to be retrieved.}
}
\value{
A \code{object} 'image_mask' from gdalcubes containing information
 about the mask band.
}
\description{
Create an object image_mask with information about mask band
}
\keyword{internal}
