% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_gdalcubes.R
\name{.sits_gc_compose}
\alias{.sits_gc_compose}
\title{Save the images based on an aggregation method.}
\usage{
.sits_gc_compose(
  cube,
  name,
  cv_list,
  img_col,
  path_db,
  path_images,
  cloud_mask,
  ...,
  version = "v1"
)
}
\arguments{
\item{cube}{Data cube from where data is to be retrieved.}

\item{name}{Name of the new data cube}

\item{cv_list}{A \code{list} 'cube_view' with values from cube.}

\item{img_col}{A \code{object} 'image_collection' containing information
about the images metadata.}

\item{path_db}{A \code{character} with the path and name where the
database will be create. E.g. "my/path/gdalcubes.db"}

\item{path_images}{A \code{character} with the path where the aggregated
images will be writed.}

\item{cloud_mask}{A \code{logical} corresponds to the use of the cloud band
for aggregation.}

\item{...}{Additional parameters that can be included. See
'?gdalcubes::write_tif'.}

\item{version}{A \code{character} with version of the output files.}
}
\value{
A data cube with information used in its creation.
}
\description{
Save the images based on an aggregation method.
}
\keyword{internal}
