% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_config.R
\name{.sits_get_projection}
\alias{.sits_get_projection}
\title{Retrieve the projection for the product available at service}
\usage{
.sits_get_projection(service, name)
}
\arguments{
\item{service}{Name of the time series service.}

\item{name}{Name of the cube.}
}
\value{
CRS PROJ4 infomation.
}
\description{
Retrieve the projection for the product available at service
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
