% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_config.R
\name{.sits_get_scale_factors}
\alias{.sits_get_scale_factors}
\title{Retrieve the scale factor for a given band for an image product}
\usage{
.sits_get_scale_factors(service, name, bands)
}
\arguments{
\item{service}{name of the service}

\item{name}{name of the coverage}

\item{bands}{vector of bands}
}
\value{
sf            vector of scale factors
}
\description{
Retrieve the scale factor for a given band for an image product
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
