% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_config.R
\name{.sits_get_size}
\alias{.sits_get_size}
\title{Retrieve the size of the product for a given data cube}
\usage{
.sits_get_size(service, name, r_obj = NA)
}
\arguments{
\item{service}{Name of the time series service.}

\item{name}{Name of the cube.}

\item{r_obj}{R object associated with the cube.}
}
\value{
Vector of (nrows, ncols).
}
\description{
Retrieve the size of the product for a given data cube
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
