% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_get_data.R
\name{.sits_get_ts}
\alias{.sits_get_ts}
\title{Dispatch function to get time series from data cubes and cloud services}
\usage{
.sits_get_ts(
  cube,
  samples,
  ...,
  bands = NULL,
  impute_fn,
  multicores,
  output_dir,
  progress
)
}
\arguments{
\item{cube}{Data cube from where data is to be retrieved.}

\item{samples}{Samples to be retrieved.}

\item{bands}{Bands to be retrieved (optional).}

\item{impute_fn}{Imputation function for NA values.}

\item{multicores}{Number of threads to process the time series.}

\item{progress}{A logical value indicating if a progress bar
should be shown. Default is \code{FALSE}.}
}
\description{
Dispatch function to get time series from data cubes and cloud services
}
\author{
Gilberto Camara
}
\keyword{internal}
