% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_plot.R
\name{.sits_ggplot_together}
\alias{.sits_ggplot_together}
\title{Plot many timeSeries together using ggplot}
\usage{
.sits_ggplot_together(melted.tb, means.tb, plot_title)
}
\arguments{
\item{melted.tb}{A tibble with the time series (already melted).}

\item{means.tb}{Means and std deviations of the time series.}

\item{plot_title}{The title for the plot.}
}
\description{
Plots a set of  time series together.
}
