% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_label_classification.R
\name{.sits_label_cube}
\alias{.sits_label_cube}
\title{Create a set of RasterLayer objects
       to store data cube classification results (labelled classes)}
\usage{
.sits_label_cube(cube_probs, output_dir, version)
}
\arguments{
\item{cube_probs}{Metadata about the input data cube (probability).}

\item{output_dir}{Output directory where to put the files}

\item{version}{Name of the version of the result}
}
\value{
Metadata about the output RasterLayer objects.
}
\description{
Takes a tibble containing metadata about a data cube wuth
classification probabilites and and creates a
set of RasterLayers to store the classification result. Each RasterLayer is
to one time step. The time steps are specified in a list of dates.
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
\keyword{internal}
