% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_signal.R
\name{.sits_MASS_ginv}
\alias{.sits_MASS_ginv}
\title{Generalized Inverse of a Matrix}
\usage{
.sits_MASS_ginv(X, tol = sqrt(.Machine$double.eps))
}
\arguments{
\item{X}{Matrix for which the Moore-Penrose inverse is required.}

\item{tol}{A relative tolerance to detect zero singular values.}
}
\value{
A MP generalized inverse matrix for X.
}
\description{
Calculates the Moore-Penrose generalized inverse of a matrix X.
}
\references{
Venables, W. N. and Ripley, B. D. (1999)
Modern Applied Statistics with S-PLUS.
}
\keyword{internal}
