% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_machine_learning.R
\name{.sits_normalize_matrix}
\alias{.sits_normalize_matrix}
\title{Normalize the time series values in the case of a matrix}
\usage{
.sits_normalize_matrix(data.mx, stats, band, multicores, progress = TRUE)
}
\arguments{
\item{data.mx}{Matrix of values.}

\item{stats}{Statistics for normalization.}

\item{band}{Band to be normalized.}

\item{multicores}{Number of cores.}

\item{progress}{Show progress bar? Default is TRUE.}
}
\value{
A normalized matrix.
}
\description{
this function normalizes one band of the values read from a raster brick
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
