% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_parallel.R
\name{.sits_parallel_map}
\alias{.sits_parallel_map}
\title{Maps a function to a given list in parallel or sequential processing}
\usage{
.sits_parallel_map(x, fn, ..., progress = FALSE, n_retries = 3, sleep = 0)
}
\arguments{
\item{x}{List to be passed to a function.}

\item{fn}{Function to be applied to each list element.}

\item{progress}{Show progress bar?}

\item{n_retries}{Number of retries before fail.}

\item{sleep}{Number in seconds to wait before trying again.}
}
\value{
List with the function results in the same order
                      as the input list
}
\description{
Maps a function to a given list in parallel or sequential processing
}
\author{
Rolf Simoes, \email{rolf.simoes@inpe.br}
}
\keyword{internal}
