% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_plot.R
\name{.sits_plot_classified_image}
\alias{.sits_plot_classified_image}
\title{Plot a raster classified image}
\usage{
.sits_plot_classified_image(cube, time, title, legend, palette)
}
\arguments{
\item{cube}{A tibble with the metadata for a labelled data cube.}

\item{time}{Temporal reference for plot.}

\item{title}{Title of the plot}

\item{legend}{named vector that associates labels to colors.}

\item{palette}{palette provided in the configuration file}
}
\description{
plots a raster using ggplot. This function is used
for showing the same lat/long location in a series of time steps.
}
\keyword{internal}
