% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_plot.R
\name{.sits_plot_classified_image}
\alias{.sits_plot_classified_image}
\title{Plot a raster classified image}
\usage{
.sits_plot_classified_image(cube, time, title, legend, palette, rev)
}
\arguments{
\item{cube}{metadata for a labelled data cube.}

\item{time}{temporal reference for plot.}

\item{title}{title of the plot}

\item{legend}{named vector that associates labels to colors.}

\item{palette}{palette (one of grDevices::hcl.pals())}

\item{rev}{revert the order of hcl palette (TRUE/FALSE)}
}
\description{
plots a raster using ggplot. This function is used
for showing the same lat/long location in a series of time steps.
}
\keyword{internal}
