% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_plot.R
\name{.sits_plot_dendrogram}
\alias{.sits_plot_dendrogram}
\title{Plot a dendrogram}
\usage{
.sits_plot_dendrogram(
  data,
  cluster.obj,
  cutree_height = NULL,
  colors = "RdYlGn"
)
}
\arguments{
\item{data}{A sits tibble with data used to extract the dendrogram.}

\item{cluster.obj}{Cluster object produced by `sits_cluster` function.}

\item{cutree_height}{A dashed horizontal line to be drawn
indicating the height of dendrogram cutting.}

\item{colors}{A color scheme as showed in `sits_color_name` function.}
}
\description{
Plot a dendrogram
}
\author{
Rolf Simoes, \email{rolf.simoes@inpe.br}
}
