% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_plot.R
\name{.sits_plot_som_confusion}
\alias{.sits_plot_som_confusion}
\title{Plot information about confusion between clusters}
\usage{
.sits_plot_som_confusion(data, title)
}
\arguments{
\item{data}{Table containing the percentage of mixture between the clusters
(produced by \code{\link[sits]{sits_som_evaluate_cluster}})}

\item{title}{Title of plot.}
}
\value{
ggplot2 object
}
\description{
Plot a bar graph with informations about each cluster.
The percentage of mixture between the clusters.
}
\author{
Lorena Santos \email{lorena.santos@inpe.br}
}
