% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_plot.R
\name{.sits_plot_TWDTW_alignments}
\alias{.sits_plot_TWDTW_alignments}
\title{Plot classification alignments using the dtwSat package}
\usage{
.sits_plot_TWDTW_alignments(matches)
}
\arguments{
\item{matches}{a list of dtwSat S4 matches objects produced by sits_TWDTW_matches}
}
\description{
plots the alignments from TWDTW classification (uses dtwSat)
}
\author{
Victor Maus, \email{vwmaus1@gmail.com}

Gilberto Camara, \email{gilberto.camara@inpe.br}
}
