% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_accuracy.R
\name{.sits_pred_ref}
\alias{.sits_pred_ref}
\title{Obtains the predicted value of a reference set}
\usage{
.sits_pred_ref(class.tb)
}
\arguments{
\item{class.tb}{A sits tibble containing a set of classified samples whose labels are known}
}
\value{
pred_ref.tb    A tibble with predicted and reference values
}
\description{
Obtains a tibble of predicted and reference values from a classified data set
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
