% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_machine_learning.R
\name{.sits_predict}
\alias{.sits_predict}
\title{Predict class based on the trained models}
\usage{
.sits_predict(distances_DT = NULL, ml_model, ...)
}
\arguments{
\item{distances_DT}{Set of distance metrics to each of the classes.}

\item{ml_model}{Model trained by \code{\link[sits]{sits_train}}.}

\item{...}{Other parameters to be passed to the model function.}
}
\value{
A vector of predicted labels.
}
\description{
Given a sits tibble time series and a model trained by \code{\link[sits]{sits_train}},
  returns a predicted label. Note that this function is
  called inside \code{\link[sits]{sits_classify}}, so the user does not need
  to explicitly use it. Please see the above-mentioned classification functions.
}
\author{
Alexandre Xavier Ywata de Carvalho, \email{alexandre.ywata@ipea.gov.br}

Rolf Simoes, \email{rolf.simoes@inpe.br}
}
