% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_space_time_operations.R
\name{.sits_proj_to_latlong}
\alias{.sits_proj_to_latlong}
\title{Coordinate transformation (X/Y to lat/long)}
\usage{
.sits_proj_to_latlong(x, y, crs)
}
\arguments{
\item{x}{X coordinate of the chosen location.}

\item{y}{Y coordinate of the chosen location.}

\item{crs}{Projection definition to be converted from.}
}
\value{
Matrix with latlong coordinates.
}
\description{
Transform a XY coordinate to a latitude and longitude
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
\keyword{internal}
