% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_raster_api.R
\name{.sits_raster_api_extract}
\alias{.sits_raster_api_extract}
\title{Raster package internal extract values function}
\usage{
.sits_raster_api_extract(r_obj, xy, ...)
}
\arguments{
\item{r_obj}{raster package object}

\item{xy}{numeric matrix with coordinates}

\item{...}{additional parameters to be passed to raster package}
}
\value{
numeric matrix with raster values for each coordinate
}
\description{
Raster package internal extract values function
}
\author{
Rolf Simoes, \email{rolf.simoes@inpe.br}
}
\keyword{internal}
