% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_raster_api.R
\name{.sits_raster_api_merge}
\alias{.sits_raster_api_merge}
\title{Merge all input files into one raster file}
\usage{
.sits_raster_api_merge(
  in_files,
  out_file,
  datatype,
  format = "GTiff",
  compress = "LZW",
  overwrite = TRUE
)
}
\arguments{
\item{in_files}{Input file paths}

\item{out_file}{Output raster file path}

\item{datatype}{Data type}

\item{format}{Format to write the file}

\item{compress}{Compression method to be used}

\item{overwrite}{Overwrite the file?}

\item{filename}{File name of the raster image file.}
}
\value{
Output file path
}
\description{
Merge all input files into one raster file
}
\author{
Rolf Simoes, \email{rolf.simoes@inpe.br}
}
\keyword{internal}
