% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_raster_api.R
\name{.sits_raster_api_params_file}
\alias{.sits_raster_api_params_file}
\title{Determine the file params to write in the metadata}
\usage{
.sits_raster_api_params_file(file)
}
\arguments{
\item{file}{A valid raster image}
}
\value{
A tibble with the cube params
}
\description{
Based on the R object associated to a raster object,
                determine its params
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
\keyword{internal}
