% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_raster_blocks.R
\name{.sits_raster_block_list}
\alias{.sits_raster_block_list}
\title{Calculate a list of blocks to be read from disk to memory}
\usage{
.sits_raster_block_list(nblocks, sub_image)
}
\arguments{
\item{nblocks}{Number of blocks to read from each image.}

\item{sub_image}{Area of interest in the image.}
}
\value{
List with named vectors with (\code{col},
                        \code{row}, \code{ncols}, \code{nrows}).
}
\description{
Calculate a list of blocks to be read from disk to memory
}
\keyword{internal}
