% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_raster_blocks.R
\name{.sits_raster_blocks}
\alias{.sits_raster_blocks}
\title{Define a reasonable block size to process an image subset}
\usage{
.sits_raster_blocks(cube, ml_model, sub_image, memsize, multicores)
}
\arguments{
\item{cube}{input data cube.}

\item{ml_model}{machine learning model.}

\item{sub_image}{bounding box of the ROI}

\item{memsize}{memory available for classification (in GB).}

\item{multicores}{number of threads to process the time series.}
}
\value{
list with three attributes: n (number of blocks),
                        rows (list of rows to begin),
                        nrows (number of rows to read at each iteration).
}
\description{
Defines the size of the block of an image to be read.
For example, a Raster Brick
with 500 rows and 500 columns and 400 time instances will have a total pixel size
of 800 Mb if pixels are 64-bit.
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
\keyword{internal}
