% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_raster.R
\name{.sits_raster_blocks_estimate}
\alias{.sits_raster_blocks_estimate}
\title{Estimate the number of blocks}
\usage{
.sits_raster_blocks_estimate(
  ml_model,
  nbands,
  nrows,
  ncols,
  timeline,
  interval,
  memsize,
  multicores
)
}
\arguments{
\item{ml_model}{Machine learning model.}

\item{nbands}{Number of bands.}

\item{nrows}{Number of rows per brick.}

\item{ncols}{Number of cols per brick.}

\item{timeline}{Timeline of the brick.}

\item{interval}{Classification interval.}

\item{memsize}{Memory available for classification (in GB).}

\item{multicores}{Number of threads to process the time series.}
}
\value{
Number of blocks to be read.
}
\description{
Defines the number of blocks of a Raster Brick
             to be read into memory.
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
\keyword{internal}
