% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_raster_blocks.R
\name{.sits_raster_blocks_estimate}
\alias{.sits_raster_blocks_estimate}
\title{Estimate the number of blocks}
\usage{
.sits_raster_blocks_estimate(cube, ml_model, sub_image, memsize, multicores)
}
\arguments{
\item{cube}{input data cube}

\item{ml_model}{machine learning model}

\item{sub_image}{area of interest in the image}

\item{memsize}{Memory available for classification (in GB).}

\item{multicores}{Number of threads to process the time series.}
}
\value{
Number of blocks to be read.
}
\description{
Defines the number of blocks of a set of images
             to be read into memory.
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
\keyword{internal}
