% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_apply.R, R/sits_raster_blocks.R
\name{.sits_raster_blocks_estimate_apply}
\alias{.sits_raster_blocks_estimate_apply}
\alias{.apply_raster_blocks_estimate}
\title{Estimate the number of blocks}
\usage{
.apply_raster_blocks_estimate(nbands, sub_image, memsize, multicores)

.sits_raster_blocks_estimate_apply(tile, sub_image, memsize, multicores)
}
\arguments{
\item{nbands}{Number of bands to open.}

\item{sub_image}{Area of interest in the image.}

\item{memsize}{Memory available for classification (in GB).}

\item{multicores}{Number of threads to process the time series.}

\item{tile}{Tile of data cube.}
}
\value{
Number of blocks to be read.

Number of blocks to be read.
}
\description{
Defines the number of blocks of a Raster Brick
             to be read into memory.

Defines the number of blocks of a Raster Brick
             to be read into memory.
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
\keyword{internal}
