% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_raster_data.R
\name{.sits_raster_brick_cube}
\alias{.sits_raster_brick_cube}
\title{Create a raster brick data cube}
\usage{
.sits_raster_brick_cube(satellite, sensor, name, timeline, bands, files)
}
\arguments{
\item{satellite}{Name of satellite}

\item{sensor}{Name of sensor}

\item{name}{Name of the data cube.}

\item{timeline}{Vector of dates with the timeline of the bands.}

\item{bands}{Vector of bands contained in the Raster Brick
set (in the same order as the files).}

\item{files}{Vector with the file paths of the raster files.}
}
\value{
A tibble with metadata information about a raster data set.
}
\description{
Builds a BRICK cube
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
