% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_raster_data.R
\name{.sits_raster_check_bricks}
\alias{.sits_raster_check_bricks}
\title{Check if the raster files are bricks}
\usage{
.sits_raster_check_bricks(files)
}
\arguments{
\item{files}{files associated to the raster data}
}
\value{
TRUE         true if filles are acessible
}
\description{
Check if the raster files are bricks
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
