% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_raster_data.R
\name{.sits_raster_check_webfiles}
\alias{.sits_raster_check_webfiles}
\title{Check if the raster files are on the web}
\usage{
.sits_raster_check_webfiles(files)
}
\arguments{
\item{files}{files associated to the raster data}
}
\value{
files        Updated files with  information for GDAL access
}
\description{
Check if the raster files are on the web
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
\keyword{internal}
