% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_raster_data.R
\name{.sits_raster_data_get_ts}
\alias{.sits_raster_data_get_ts}
\title{Extract a time series from raster}
\usage{
.sits_raster_data_get_ts(
  cube,
  points,
  bands,
  cld_band = NULL,
  impute_fn = sits_impute_linear()
)
}
\arguments{
\item{cube}{Metadata describing a raster data cube.}

\item{points}{tibble with points}

\item{bands}{Bands to be retrieved.}

\item{cld_band}{Cloud band (if available)}

\item{impute_fn}{Imputation function for NA values}
}
\value{
A sits tibble with the time series.
}
\description{
Retrieve a set of time series for a raster data cube.
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
\keyword{internal}
