% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_raster_data.R
\name{.sits_raster_data_preprocess}
\alias{.sits_raster_data_preprocess}
\title{Preprocess a set of values retrieved from a raster}
\usage{
.sits_raster_data_preprocess(
  cube,
  band_cube,
  extent,
  filter_fn = NULL,
  stats = NULL,
  impute_fn,
  multicores,
  .verbose = FALSE
)
}
\arguments{
\item{cube}{data cube being processed}

\item{band_cube}{band to be processed}

\item{extent}{extent to be read}

\item{filter_fn}{smoothing filter to be applied.}

\item{stats}{normalization parameters.}

\item{impute_fn}{imputing function to be applied to replace NA}

\item{multicores}{number of cores to process the time series.}

\item{.verbose}{prints information about processing times}
}
\value{
Matrix with pre-processed values.
}
\description{
Preprocess a set of values retrieved from a raster
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
\keyword{internal}
