% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_raster_data.R
\name{.sits_raster_data_read}
\alias{.sits_raster_data_read}
\title{Read a block of values retrieved from a set of raster images}
\usage{
.sits_raster_data_read(
  cube,
  samples,
  extent,
  stats,
  filter_fn,
  impute_fn,
  interp_fn,
  compose_fn
)
}
\arguments{
\item{cube}{input data cube.}

\item{samples}{tibble with samples.}

\item{extent}{bounding box in (i,j) coordinates}

\item{stats}{normalization parameters.}

\item{filter_fn}{smoothing filter to be applied.}

\item{impute_fn}{impute function to replace NA}

\item{interp_fn}{function to interpolate points from cube to match samples}

\item{compose_fn}{function to compose points from cube to match samples}
}
\value{
A data.table with values for classification.
}
\description{
Read a block of values retrieved from a set of raster images
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
\keyword{internal}
