% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_metrics_processing.R
\name{.sits_raster_metrics}
\alias{.sits_raster_metrics}
\title{Filter the time series values in the case of a matrix}
\usage{
.sits_raster_metrics(values, multicores)
}
\arguments{
\item{values}{matrix of values.}

\item{multicores}{Number of cores.}
}
\value{
Filtered matrix.
}
\description{
This function filters a matrix.
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
\keyword{internal}
