% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_raster.R
\name{.sits_raster_params}
\alias{.sits_raster_params}
\title{Determine the cube params to write in the metadata}
\usage{
.sits_raster_params(r_obj)
}
\arguments{
\item{r_obj}{An R object associated to a Raster (Layer, Brick or Stack)}
}
\value{
A tibble with the cube params
}
\description{
Based on the R object associated to a raster object,
                determine its params
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
