% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_raster.R
\name{.sits_raster_preprocess_data}
\alias{.sits_raster_preprocess_data}
\title{Preprocess a set of values retrived from a raster brick}
\usage{
.sits_raster_preprocess_data(
  values.mx,
  band,
  clouds.mx,
  cld_index,
  missing_value,
  minimum_value,
  maximum_value,
  scale_factor,
  stats,
  filter,
  impute_fn,
  multicores
)
}
\arguments{
\item{values.mx}{matrix of values retrieved from a raster object}

\item{band}{band to be processed}

\item{clouds.mx}{matrix of cloud values (optional)}

\item{cld_index}{index of values indicating clouds in the cloud band}

\item{missing_value}{missing value for the band.}

\item{minimum_value}{minimum values for the band.}

\item{scale_factor}{scale factor for each band (only for raster data).}

\item{stats}{normalization parameters.}

\item{filter}{smoothing filter to be applied.}

\item{impute_fn}{imputing function to be applied to replace NA}

\item{multicores}{number of cores to process the time series.}

\item{maximum_values}{maximum values for the band}
}
\value{
Matrix with pre-processed values.
}
\description{
Preprocess a set of values retrived from a raster brick
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
\keyword{internal}
