% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_raster.R
\name{.sits_raster_preprocess_data}
\alias{.sits_raster_preprocess_data}
\title{Preprocess a set of values retrived from a raster brick}
\usage{
.sits_raster_preprocess_data(
  cube,
  values.mx,
  band,
  clouds.mx,
  cld_index,
  impute_fn,
  multicores,
  .verbose
)
}
\arguments{
\item{cube}{data cube being processed}

\item{values.mx}{matrix of values retrieved from a raster object}

\item{band}{band to be processed}

\item{clouds.mx}{matrix of cloud values (optional)}

\item{cld_index}{index of values indicating clouds in the cloud band}

\item{impute_fn}{imputing function to be applied to replace NA}

\item{multicores}{number of cores to process the time series.}

\item{.verbose}{prints information about processing times}
}
\value{
Matrix with pre-processed values.
}
\description{
Preprocess a set of values retrived from a raster brick
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
\keyword{internal}
