% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_raster.R
\name{.sits_raster_read_data}
\alias{.sits_raster_read_data}
\title{Read a block of values retrived from a set of raster bricks}
\usage{
.sits_raster_read_data(
  cube,
  samples,
  ml_model,
  first_row,
  nrows_block,
  first_col,
  ncols_block,
  stats,
  filter,
  multicores
)
}
\arguments{
\item{cube}{input data cube.}

\item{samples}{tibble with samples.}

\item{ml_model}{machine learning model.}

\item{first_row}{first row to start reading.}

\item{nrows_block}{number of rows in the block.}

\item{first_col}{first column to start reading}

\item{ncols_block}{number of columns in the block}

\item{stats}{normalization parameters.}

\item{filter}{smoothing filter to be applied.}

\item{multicores}{number of cores to process the time series.}
}
\value{
A data.table with values for classification.
}
\description{
Read a block of values retrived from a set of raster bricks
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
\keyword{internal}
