% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_raster.R
\name{.sits_raster_read_data}
\alias{.sits_raster_read_data}
\title{Read a block of values retrived from a set of raster bricks}
\usage{
.sits_raster_read_data(
  cube,
  samples,
  ml_model,
  first_row,
  n_rows_block,
  stats,
  filter,
  multicores
)
}
\arguments{
\item{cube}{Input data cube.}

\item{samples}{Tibble with samples.}

\item{ml_model}{Machine learning model.}

\item{first_row}{First row to start reading.}

\item{n_rows_block}{Number of rows in the block.}

\item{stats}{Normalization parameters.}

\item{filter}{Smoothing filter to be applied.}

\item{multicores}{Number of cores to process the time series.}
}
\value{
A data.table with values for classification.
}
\description{
Read a block of values retrived from a set of raster bricks
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
