% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_raster.R
\name{.sits_raster_read_data}
\alias{.sits_raster_read_data}
\title{Read a block of values retrived from a set of raster bricks}
\usage{
.sits_raster_read_data(
  cube,
  samples,
  t_obj.lst,
  extent,
  stats,
  filter,
  multicores
)
}
\arguments{
\item{cube}{input data cube.}

\item{samples}{tibble with samples.}

\item{t_obj.lst}{list of terra objects to be read}

\item{extent}{bounding box in (i,j) coordinates}

\item{stats}{normalization parameters.}

\item{filter}{smoothing filter to be applied.}

\item{multicores}{number of cores to process the time series.}
}
\value{
A data.table with values for classification.
}
\description{
Read a block of values retrived from a set of raster bricks
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
\keyword{internal}
