% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_raster.R
\name{.sits_raster_scale_data}
\alias{.sits_raster_scale_data}
\title{Scale the time series values in the case of a matrix}
\usage{
.sits_raster_scale_data(values.mx, scale_factor, multicores,
  progress = TRUE)
}
\arguments{
\item{values.mx}{Matrix of values.}

\item{scale_factor}{Scaling factor.}

\item{multicores}{Number of cores.}

\item{progress}{Show progress bar? Default is TRUE.}
}
\value{
A scaled matrix.
}
\description{
This function normalizes one band of the values read from a raster brick.
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
