% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_raster_sub_image.R
\name{.sits_raster_sub_image_from_block}
\alias{.sits_raster_sub_image_from_block}
\title{Extract a sub_image from a valid block}
\usage{
.sits_raster_sub_image_from_block(
  block,
  source,
  collection,
  tile = NULL,
  ...,
  xmin = NULL,
  xmax = NULL,
  ymin = NULL,
  ymax = NULL,
  nrows = NULL,
  ncols = NULL,
  crs = NULL
)
}
\arguments{
\item{block}{a valid block with \code{first_col}, \code{first_row},
\code{nrows}, and \code{ncols}.}

\item{source}{a \code{character} with the cube source.}

\item{collection}{a \code{character} with the cube collection.}

\item{tile}{a \code{tibble} with a sits cube object.}

\item{xmin}{a \code{numeric} with the xmin value.}

\item{xmax}{a \code{numeric} with the xmax value.}

\item{ymin}{a \code{numeric} with the ymin value.}

\item{ymax}{a \code{numeric} with the ymax value.}

\item{nrows}{a \code{numeric} with the nrows value.}

\item{ncols}{a \code{numeric} with the nclols value.}

\item{crs}{a \code{character} with the crs value.}
}
\value{
sub_image with additional info on first row, first col, nrows,
ncols, and crs.
}
\description{
Extract a sub_image from a valid block
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
\keyword{internal}
