% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_roi.R
\name{.sits_roi_bbox}
\alias{.sits_roi_bbox}
\title{Find the bounding box for a spatial ROI in a data cube}
\usage{
.sits_roi_bbox(roi, cube)
}
\arguments{
\item{roi}{spatial region of interest}

\item{cube}{input data cube.}
}
\value{

}
\description{
Find the bounding box for a spatial ROI in a data cube
}
\keyword{internal}
