% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_filters.R
\name{.sits_signal_sgolayfilt}
\alias{.sits_signal_sgolayfilt}
\title{Savitzky-Golay smoothing filter}
\usage{
.sits_signal_sgolayfilt(x, p = 3, n = p + 3 - p\%\%2, m = 0, ts = 1)
}
\arguments{
\item{x}{Time series vector.}

\item{p}{Filter order (integer).}

\item{n}{Filter length (must be odd)}

\item{m}{Derivative to calculate (default = 0)}

\item{ts}{Time scaling (integer).}
}
\value{
A time series with filtered values.
}
\description{
Smooth the data in x with a Savitzky-Golay smoothing filter of
  polynomial order p and length n, n odd, n > p.  By default, p=3
   and n=p+2 or n=p+3 if p is even. This filters is particularly good
   at preserving lineshape while
   removing high frequency squiggles
}
\keyword{internal}
