% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_kohonen.R
\name{.sits_som_neighbor_neurons}
\alias{.sits_som_neighbor_neurons}
\title{Get the neighbor of neurons}
\usage{
.sits_som_neighbor_neurons(
  class_vector,
  kohonen_obj,
  duplicated_id_neuron,
  radius = 1
)
}
\arguments{
\item{class_vector}{A vector contained the labels of each neuron.}

\item{kohonen_obj}{An object returned by kohonen package.}

\item{duplicated_id_neuron}{A vector containing the ids of each neuron
that has two majority labels.}

\item{radius}{radius of neigbourhood of each neuron.}
}
\value{
A sits tibble with information about
                             the vinicity of each neuron.
}
\description{
This function creates a table contained the information about all
neighbor of one neuron of radius 1.
}
\author{
Lorena Santos, \email{lorena.santos@inpe.br}
}
\keyword{internal}
