% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_kohonen.R
\name{.sits_som_paint_neurons}
\alias{.sits_som_paint_neurons}
\title{Paint neurons}
\usage{
.sits_som_paint_neurons(kohonen_obj)
}
\arguments{
\item{kohonen_obj}{Object Kohonen, this object contains all parameters of SOM provided by package Kohonen}
}
\value{
Returns kohonen_obj with a new parameter with the colour of the neuron.
}
\description{
This function paint all neurons of the last iteration of SOM in function sits_cluster_som
}
\author{
Lorena Santos, \email{lorena.santos@inpe.br}
}
