% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_smooth_aux_functions.R
\name{.sits_split_cluster}
\alias{.sits_split_cluster}
\title{parallel raster brick}
\usage{
.sits_split_cluster(file, n_tiles, pad_rows, fun, args = NULL, cl = NULL, ...)
}
\arguments{
\item{file}{a path to raster file.}

\item{n_tiles}{number of horizontal chunks to break the raster.}

\item{pad_rows}{number of overlaping rows of each chunk.}

\item{fun}{a function that receives RasterBrick and
returns any Raster*.}

\item{args}{additional arguments to pass to \code{fun} function.}

\item{cl}{snow cluster object}

\item{...}{optional arguments to merge final raster
(see \link[raster]{writeRaster} function)}
}
\value{
RasterBrick object
}
\description{
Process chunks of raster brick in parallel.
}
\keyword{internal}
