% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_stac.R
\name{.sits_stac_items}
\alias{.sits_stac_items}
\title{Get information from item}
\usage{
.sits_stac_items(
  url = NULL,
  collection = NULL,
  tiles = NULL,
  bbox = NULL,
  start_date = NULL,
  end_date = NULL,
  ...
)
}
\arguments{
\item{url}{a \code{character} representing a URL for the BDC catalog.}

\item{collection}{a \code{character} with the collection to be searched.}

\item{tiles}{a \code{character} with the names of the tiles.}

\item{bbox}{a \code{numeric} vector with only features that have a
geometry that intersects the bounding box are selected. The bounding box is
provided as four or six numbers, depending on whether the coordinate
reference system includes a vertical axis (elevation or depth):
\itemize{ \item Lower left corner, coordinate axis 1
          \item Lower left corner, coordinate axis 2
          \item Lower left corner, coordinate axis 3 (optional)
          \item Upper right corner, coordinate axis 1
          \item Upper right corner, coordinate axis 2
          \item Upper right corner, coordinate axis 3 (optional) }

The coordinate reference system of the values is WGS84 longitude/latitude
(\url{http://www.opengis.net/def/crs/OGC/1.3/CRS84}). The values are in
most cases the sequence of minimum longitude, minimum latitude, maximum
longitude and maximum latitude. However, in cases where the box spans the
antimeridian the first value (west-most box edge) is larger than the third
value (east-most box edge).}

\item{ids}{a \code{character} vector with the items features ids.}

\item{datetime}{a \code{character} with a date-time or an interval. Date
and time strings needs to conform RFC 3339. Intervals are expressed by
separating two date-time strings by \code{'/'} character. Open intervals are
expressed by using \code{'..'} in place of date-time.}

\item{intersects}{a \code{character} value expressing GeoJSON geometries
objects as specified in RFC 7946. Only returns items that intersect with
the provided polygon.}

\item{limit}{an \code{integer} defining the maximum number of results
to return. If not informed it defaults to the service implementation.}
}
\value{
a \code{STACItemCollection} object representing the search
 by rstac.
}
\description{
Get information from item
}
\keyword{internal}
