% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_stac.R
\name{.sits_stac_tile_cube}
\alias{.sits_stac_tile_cube}
\title{Get the STAC information corresponding to a tile.}
\usage{
.sits_stac_tile_cube(url, name, collection, items, cube, file_info)
}
\arguments{
\item{url}{a \code{character} representing URL for the BDC STAC.}

\item{name}{a \code{character} representing the output data cube.}

\item{collection}{a \code{STACCollection} object returned by rstac.}

\item{items}{a \code{STACItemCollection} object returned by rstac.}

\item{cube}{a \code{character} with name input data cube in BDC.}

\item{file_info}{a \code{tbl_df} with the information from STAC.}
}
\value{
a \code{tibble} with metadata information about a
 raster data set.
}
\description{
Get the STAC information corresponding to a tile.
}
\keyword{internal}
