% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_timeline.R
\name{.sits_timeline_match_indexes}
\alias{.sits_timeline_match_indexes}
\title{Find indexes in a timeline that match the reference dates}
\usage{
.sits_timeline_match_indexes(timeline, ref_dates)
}
\arguments{
\item{timeline}{Timeline of input observations (vector).}

\item{ref_dates}{List of breaks to be applied to the input data set.}
}
\value{
A list of indexes that match the reference dates
                     to the timelines.
}
\description{
For correct classification, the time series of the input data
             should be aligned to that of the reference data set
             (usually a set of patterns).
             This function aligns these data sets so that shape
             matching works correctly
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
\keyword{internal}
