% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_tibble_aux_functions.R
\name{.sits_transmute_bands}
\alias{.sits_transmute_bands}
\title{Add new sits bands and drops existing.}
\usage{
.sits_transmute_bands(data, ...)
}
\arguments{
\item{data}{A sits tibble.}

\item{...}{Pair expressions in the format `name = value`.
See \code{\link[dplyr]{mutate}} help for more details.}
}
\value{
A sits tibble with same samples and the new bands.
}
\description{
Adds new bands and drops existing in the time series
              of a sits tibble using dplyr::transmute function.
}
\examples{
\donttest{
# Retrieve data for time series with label samples in Mato Grosso
data(samples_mt_6bands)
# Generate a new image with the SAVI (Soil-adjusted vegetation index)
savi.tb <- sits:::.sits_transmute_bands(samples_mt_6bands,
                                SAVI = (1.5*(NIR - RED)/(NIR + RED + 0.5)))
}

}
\author{
Rolf Simoes, \email{rolf.simoes@inpe.br}
}
\keyword{internal}
