% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_satveg.R
\name{.sits_ts_from_satveg}
\alias{.sits_ts_from_satveg}
\title{Retrieve a time series from the SATVEG service}
\usage{
.sits_ts_from_satveg(longitude, latitude, name)
}
\arguments{
\item{longitude}{The longitude of the chosen location.}

\item{latitude}{The latitude of the chosen location.}

\item{name}{Name of the desired data cube in SATVEG}
}
\value{
A tibble containing a time series
}
\description{
Retrieves a time series from the SATVEG service.
}
\author{
Julio Esquerdo, \email{julio.esquerdo@embrapa.br}

Gilberto Camara, \email{gilberto.camara@inpe.br}
}
\keyword{internal}
