% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_raster_read_write.R
\name{.sits_write_raster_values}
\alias{.sits_write_raster_values}
\title{Write the values and probs into raster files}
\usage{
.sits_write_raster_values(output.lst, prediction_DT, labels, int_labels,
  time, first_row, multicores)
}
\arguments{
\item{output.lst}{list with value layers and probability bricks}

\item{prediction_DT}{data.table with predicted probabilities for each class}

\item{labels}{class labels}

\item{int_labels}{integer values corresponding to labels}

\item{time}{interval to be written to file}

\item{first_row}{initial row of the output layer to write block}

\item{multicores}{number of cores to process the time series}
}
\value{
output.lst       updated list with value layers and probability bricks
}
\description{
write the raster values to the outout files
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
