% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_wtss.R
\name{.sits_wtss_check}
\alias{.sits_wtss_check}
\title{Check that the URL of WTSS service is working}
\usage{
.sits_wtss_check(URL, name)
}
\arguments{
\item{URL}{URL of the WTSS service.}

\item{name}{name of the converage}
}
\value{
check     TRUE or FALSE
}
\description{
Check that the URL of WTSS service is working
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
\keyword{internal}
