% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_wtss.R
\name{.sits_wtss_cube}
\alias{.sits_wtss_cube}
\title{Provides information about one cube of the WTSS service}
\usage{
.sits_wtss_cube(URL, name)
}
\arguments{
\item{URL}{URL of the service provider.}

\item{name}{Name of the cube.}
}
\description{
Uses the WTSS services to print information and save metadata
about a chosen cube.
}
\keyword{internal}
