% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_smooth_aux_functions.R
\name{.smth_block_filename}
\alias{.smth_block_filename}
\alias{.smth_filename}
\title{Create a filename to store an output block from smoothing operation}
\usage{
.smth_filename(tile, output_dir, block, hash = NULL, create_dir = FALSE)
}
\arguments{
\item{tile}{Output tile to be written.}

\item{output_dir}{Directory where block will be written.}

\item{block}{Block designation.}

\item{hash}{Additional hash to be add in file name.}

\item{create_dir}{A boolean indicating if the output dir should be created.}
}
\value{
returns a filename
}
\description{
Create a filename to store an output block from smoothing operation
}
\keyword{internal}
