% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_source_api.R
\name{.source_access_test}
\alias{.source_access_test}
\title{...}
\usage{
.source_access_test(source, collection, ...)
}
\arguments{
\item{source}{Name of the provider}
}
\value{
...
}
\description{
...
}
\keyword{internal}
