% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_source_api.R
\name{.source_bands_to_sits}
\alias{.source_bands_to_sits}
\title{Convert bands names as sits bands}
\usage{
.source_bands_to_sits(source, collection, bands)
}
\arguments{
\item{source}{Data source (one of "SATVEG", "LOCAL", "BDC", "AWS",
"USGS", "DEAFRICA", "PROBS").}

\item{collection}{Collection to be searched in the data source}

\item{bands}{Bands requested to be read}
}
\value{
an \code{character} with converted bands.
}
\description{
Convert bands used by the origin data cube to the name used by
 SITS
}
\keyword{internal}
