% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_source_api.R
\name{.source_cube}
\alias{.source_cube}
\title{Create a data cube}
\usage{
.source_cube(source, ...)
}
\arguments{
\item{source}{Data source (one of "SATVEG", "LOCAL", "BDC", "AWS",
"USGS", "DEAFRICA", "PROBS").}

\item{...}{Additional parameters.}
}
\value{
a sits cube.
}
\description{
Generic function responsible for creating data cubes in sits.
}
\keyword{internal}
