% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_source_api.R
\name{.source_items_bands_select}
\alias{.source_items_bands_select}
\title{Item selection from Bands}
\usage{
.source_items_bands_select(source, collection, items, bands, ...)
}
\arguments{
\item{source}{Data source (one of "SATVEG", "LOCAL", "BDC", "AWS", "USGS"
"DEAFRICA", "PROBS").}

\item{collection}{Collection to be searched in the data source.}

\item{items}{A \code{STACItemCollection} object returned by rstac.}

\item{bands}{A \code{character} with bands to be select in items object.}

\item{...}{Additional parameters.}
}
\value{
A \code{STACItemCollection} object returned by rstac with items
 selected.
}
\description{
Selection of items from specific bands by the user.
}
\keyword{internal}
