% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_source_api.R
\name{.source_items_new}
\alias{.source_items_new}
\title{...}
\usage{
.source_items_new(
  source,
  collection,
  name,
  bands,
  tiles,
  bbox,
  start_date,
  end_date,
  ...
)
}
\arguments{
\item{source}{Name of the STAC provider}

\item{collection}{...}

\item{name}{...}

\item{bands}{...}

\item{tiles}{...}

\item{bbox}{...}

\item{start_date}{...}

\item{end_date}{...}

\item{...}{...}
}
\value{
...
}
\description{
TODO: document
}
\keyword{internal}
