% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_stac.R
\name{.stac_items_query}
\alias{.stac_items_query}
\alias{.stac_create_items_query}
\title{Creates a query to send to STAC api}
\usage{
.stac_create_items_query(
  source,
  collection,
  ...,
  roi = NULL,
  start_date = NULL,
  end_date = NULL,
  limit = NULL
)
}
\arguments{
\item{source}{Name of the STAC provider}

\item{collection}{Collection to be searched in the data source}

\item{...}{Other parameters to be passed for specific types.}

\item{roi}{Region of interest.}

\item{start_date}{Initial date for the cube (optional).}

\item{end_date}{Final date for the cube  (optional).}

\item{limit}{limit items to be returned in requisition.}
}
\value{
an \code{RSTACQuery} object.
}
\description{
Creates a query using rstac package to send to STAC API.
}
\keyword{internal}
