% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_tile_api.R
\name{.tile_ncols}
\alias{.tile_ncols}
\title{Tile API}
\usage{
.tile_ncols(tile)
}
\arguments{
\item{tile}{A tile.}
}
\value{
integer
}
\description{
Get number of image columns from \code{ncols} field (if it exists)
or from unique \code{ncols} in file_info (which can be more than one if tile
images has multiples \code{ncols}).
}
