% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_tile_api.R
\name{.tile_nrows}
\alias{.tile_nrows}
\title{Tile API}
\usage{
.tile_nrows(tile)
}
\value{
integer
}
\description{
Get number of image columns from 'nrows' field (if it exists)
or from unique 'nrows' in file_info (which can be more than one if tile
images has multiples 'nrows').
}
