% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_tile_api.R
\name{.tile_read_block}
\alias{.tile_read_block}
\title{Tile API}
\usage{
.tile_read_block(tile, band, block)
}
\value{
numeric
}
\description{
Read and preprocess a 'block' of 'band' values from
file_info rasters.
}
