% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_tile_api.R
\name{.tile_read_block}
\alias{.tile_read_block}
\title{Tile API}
\usage{
.tile_read_block(tile, band, block)
}
\arguments{
\item{tile}{A tile.}

\item{band}{Band character vector.}

\item{block}{A block list with (col, row, ncols, nrows).}
}
\value{
numeric
}
\description{
Read and preprocess a \code{block} of \code{band} values from
file_info rasters.
}
\details{
\code{eo_cube} preprocess is slightly different from
\code{derived_cube}: values outside the range of minimum and maximum for
a band are replaced by \code{NA} in \code{eo_cube}. In \code{derived_cube},
values outside allowed range are clamped and replaced by minimum or maximum
values.
}
