% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_metrics_generate.R
\name{.transform_bands_to_metrics}
\alias{.transform_bands_to_metrics}
\title{Create the output of the cloud estimation procedure}
\usage{
.transform_bands_to_metrics(bands, list_metrics)
}
\arguments{
\item{bands}{...}

\item{list_metrics}{...}
}
\value{
....
}
\description{
Create the output of the cloud estimation procedure
}
\keyword{internal}
