% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_source_api_usgs.R
\name{.usgs_format_tiles}
\alias{.usgs_format_tiles}
\title{Format tile parameter provided by users}
\usage{
.usgs_format_tiles(tiles)
}
\arguments{
\item{tiles}{a \code{character} vector with the tiles provided by users.}
}
\value{
a \code{tibble} with attributes of wrs path and row.
}
\description{
Format tile parameter provided by users
}
\keyword{internal}
