% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_view.R
\name{.view_get_colors}
\alias{.view_get_colors}
\title{Return the colors associated to the classified image}
\usage{
.view_get_colors(labels, legend, palette)
}
\arguments{
\item{labels}{Labels of the classified cube.}

\item{legend}{Named vector that associates labels to colors.}

\item{palette}{Palette provided in the configuration file.}
}
\value{
Colors for legend of classified image.
}
\description{
Return the colors associated to the classified image
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
\keyword{internal}
