% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_view.R
\name{.view_resample_size}
\alias{.view_resample_size}
\title{Return the cell size for the image to be resamples}
\usage{
.view_resample_size(nrows, ncols, ntiles)
}
\arguments{
\item{nrows}{number of rows in the input image}

\item{ncols}{number of cols in the input image}

\item{ntiles}{number of tiles in the input image}
}
\value{
vector with cell size for x and y coordinates
}
\description{
Return the cell size for the image to be resamples
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
\keyword{internal}
