% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_file_info.R
\name{file_info_functions}
\alias{file_info_functions}
\alias{.file_info}
\alias{.file_info_nrows}
\alias{.file_info_ncols}
\alias{.file_info_path}
\alias{.file_info_paths}
\alias{.file_info_xres}
\alias{.file_info_yres}
\alias{.file_info_fids}
\alias{.file_info_timeline}
\alias{.file_info_timeline_wtss}
\alias{.file_info_start_date}
\alias{.file_info_end_date}
\alias{.file_info_bands}
\title{Functions to work with file info tibble}
\usage{
.file_info(cube, bands = NULL, fid = NULL, start_date = NULL, end_date = NULL)

.file_info_nrows(cube, bands = NULL)

.file_info_ncols(cube, bands = NULL)

.file_info_path(cube)

.file_info_paths(cube, bands = NULL)

.file_info_xres(cube)

.file_info_yres(cube)

.file_info_fids(cube)

.file_info_timeline(cube)

.file_info_timeline_wtss(cube)

.file_info_start_date(cube)

.file_info_end_date(cube)

.file_info_bands(cube)
}
\arguments{
\item{cube}{input data cube}

\item{bands}{bands to be filtered}

\item{fid}{feature id (fid) to be filtered}

\item{dates}{dates to be filtered}
}
\value{
vector with requested information
}
\description{
Functions to work with file info tibble
}
\details{
Return the file info for a cube with a single tile
Filter by bands if required

Return number of rows for a given tile
Throws an error if rows are not equal

Returns number of cols for a given tile
Throws an error if cols are not equal

Returns a single path to a file
Throws an error if there is more than one path

Returns a single path to a file
Throws an error if there is more than one path

Returns the X resolution for a single tiled cube
Throws an error if resolution is not unique

Returns the Y resolution for a single tiled cube
Throws an error if resolution is not unique

Returns the file ids for a single tiled cube

Returns timeline  for a single tiled cube

Returns timeline  for a single tiled cube

Returns start date  for a single tiled cube
Throws an error if cube is not a processed one

Returns start date  for a single tiled cube
Throws an error if cube is not a processed one

Returns the bands for the cube
Throws an error bands are NULL
}
\author{
Rolf Simoes, \email{rolf.simoes@inpe.br}
}
\keyword{internal}
