% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_histogram.R
\name{hist.sits}
\alias{hist.sits}
\title{Histogram}
\usage{
\method{hist}{sits}(x, ...)
}
\arguments{
\item{x}{Object of classes "sits".}

\item{...}{Further specifications for \link{hist}.}
}
\value{
A summary of the sits tibble.
}
\description{
This is a generic function. Parameters depend on the specific
type of input.
}
\examples{
if (sits_run_examples()) {
    hist(samples_modis_ndvi)
}

}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
