% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_imputation.R
\name{impute_mean}
\alias{impute_mean}
\title{Replace NA values by Mean Value}
\usage{
impute_mean(data = NULL)
}
\arguments{
\item{data}{A time series vector or matrix}
}
\value{
A set of filtered time series using
                      the imputation function.
}
\description{
Remove NA by Mean Value
}
