% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_imputation.R
\name{impute_median}
\alias{impute_median}
\title{Remove NA using median}
\usage{
impute_median(data = NULL)
}
\arguments{
\item{data}{A time series vector or matrix}
}
\value{
A set of filtered time series using
                      the imputation function.
}
\description{
Remove NA using median
}
\author{
Felipe Carlos, \email{efelipecarlos@gmail.com}
}
