% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_plot.R
\name{plot.classified_image}
\alias{plot.classified_image}
\title{Plot classified images}
\usage{
\method{plot}{classified_image}(
  x,
  y,
  ...,
  tiles = NULL,
  title = "Classified Image",
  legend = NULL,
  palette = "Spectral",
  rev = TRUE
)
}
\arguments{
\item{x}{Object of class "classified_image".}

\item{y}{Ignored.}

\item{...}{Further specifications for \link{plot}.}

\item{tiles}{Tiles to be plotted.}

\item{title}{Title of the plot.}

\item{legend}{Named vector that associates labels to colors.}

\item{palette}{Alternative palette that uses grDevices::hcl.pals().}

\item{rev}{Invert the order of hcl palette?}
}
\description{
plots a classified raster using ggplot.
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
