% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_plot.R
\name{plot.classified_image}
\alias{plot.classified_image}
\title{Plot classified images}
\usage{
\method{plot}{classified_image}(
  x,
  y,
  ...,
  tiles = NULL,
  title = "Classified Image",
  legend = NULL,
  palette = "Spectral",
  rev = TRUE
)
}
\arguments{
\item{x}{object of class "classified_image"}

\item{y}{ignored}

\item{...}{further specifications for \link{plot}.}

\item{tiles}{tiles to be plotted}

\item{title}{title of the plot}

\item{legend}{named vector that associates labels to colors}

\item{palette}{alternative palette that uses grDevices::hcl.pals()}

\item{rev}{invert the order of hcl palette (TRUE/FALSE)}
}
\description{
plots a classified raster using ggplot.
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
