% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_plot.R
\name{plot.dem_cube}
\alias{plot.dem_cube}
\title{Plot DEM cubes}
\usage{
\method{plot}{dem_cube}(
  x,
  ...,
  band = "ELEVATION",
  tile = x[["tile"]][[1]],
  palette = "RdYlGn",
  rev = TRUE,
  scale = 0.75,
  style = "log10",
  max_cog_size = 1024
)
}
\arguments{
\item{x}{Object of class "dem_cube".}

\item{...}{Further specifications for \link{plot}.}

\item{band}{Band for plotting grey images.}

\item{tile}{Tile to be plotted.}

\item{palette}{An RColorBrewer palette}

\item{rev}{Reverse the color order in the palette?}

\item{scale}{Scale to plot map (0.4 to 1.0)}

\item{style}{Style for plotting continuous objects}

\item{max_cog_size}{Maximum size of COG overviews (lines or columns)}
}
\value{
A plot object with a DEM cube
                      or a B/W image on a color scale
}
\description{
Plot RGB raster cube
}
\note{
Use \code{scale} parameter for general output control.

The following optional parameters are available to allow for detailed
      control over the plot output:
\itemize{
\item \code{graticules_labels_size}: size of coordinates labels (default = 0.8)
\item \code{legend_title_size}: relative size of legend title (default = 1.0)
\item \code{legend_text_size}: relative size of legend text (default = 1.0)
\item \code{legend_bg_color}: color of legend background (default = "white")
\item \code{legend_bg_alpha}: legend opacity (default = 0.5)
\item \code{legend_width}: relative width of legend (default = 1.0)
\item \code{legend_position}: 2D position of legend (default = c("left", "bottom"))
\item \code{legend_height}: relative height of legend (default = 1.0)
}
}
\examples{
if (sits_run_examples()) {
# obtain the DEM cube
    dem_cube_19HBA <- sits_cube(
        source = "MPC",
        collection = "COP-DEM-GLO-30",
        bands = "ELEVATION",
        tiles = "19HBA"
        )
# plot the DEM reversing the palette
    plot(dem_cube_19HBA, band = "ELEVATION")
}
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
