% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_plot.R
\name{plot.keras_model}
\alias{plot.keras_model}
\title{Generic interface for plotting a Keras model}
\usage{
\method{plot}{keras_model}(x, y, ...)
}
\arguments{
\item{x}{Object of class "keras_model"}

\item{y}{ignored}

\item{...}{further specifications for \link{plot}.}
}
\value{
The plot itself.
}
\description{
plots a deep learning model
}
\examples{
\donttest{
# Get a set of samples
samples_ndvi_evi <- sits_select(samples_mt_4bands, bands = c("NDVI", "EVI"))

# train a deep learning model
dl_model <- sits_train(samples_ndvi_evi, ml_method = sits_deeplearning(
    layers = c(512, 512, 512),
    activation = "relu",
    dropout_rates = c(0.50, 0.40, 0.35),
    epochs = 100,
    batch_size = 128,
    validation_split = 0.2
))
plot(dl_model)
}
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
