% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_plot.R
\name{plot.patterns}
\alias{plot.patterns}
\title{Plot patterns that describe classes}
\usage{
\method{plot}{patterns}(x, y, ..., bands = NULL, year_grid = FALSE)
}
\arguments{
\item{x}{Object of class "patterns".}

\item{y}{Ignored.}

\item{...}{Further specifications for \link{plot}.}

\item{bands}{Bands to be viewed (optional).}

\item{year_grid}{Plot a grid of panels using labels as columns and
years as rows. Default is FALSE.}
}
\value{
A plot object produced by ggplot2
                      with one average pattern per label.
}
\description{
Plots the patterns to be used for classification

Given a sits tibble with a set of patterns, plot them.
}
\note{
Please refer to the sits documentation available in
<https://e-sensing.github.io/sitsbook/> for detailed examples.
This code is reused from the dtwSat package by Victor Maus.
}
\examples{
if (sits_run_examples()) {
    # plot patterns
    plot(sits_patterns(cerrado_2classes))
}
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}

Victor Maus, \email{vwmaus1@gmail.com}
}
