% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_plot.R
\name{plot.predicted}
\alias{plot.predicted}
\title{Plot time series predictions}
\usage{
\method{plot}{predicted}(x, y, ..., bands = "NDVI", palette = "Harmonic")
}
\arguments{
\item{x}{Object of class "predicted".}

\item{y}{Ignored.}

\item{...}{Further specifications for \link{plot}.}

\item{bands}{Bands for visualization.}

\item{palette}{HCL palette used for visualization
in case classes are not in the default sits palette.}
}
\value{
The plot itself.
}
\description{
Given a sits tibble with a set of predictions, plot them
}
\note{
Please refer to the sits documentation available in
<https://e-sensing.github.io/sitsbook/> for detailed examples.
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
