% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_plot.R
\name{plot.probs_cube}
\alias{plot.probs_cube}
\title{Generic interface for plotting probability cubes}
\usage{
\method{plot}{probs_cube}(x, y, ..., time = 1, title = "Probabilities for Classes", colors = "YlGn")
}
\arguments{
\item{x}{Object of class "classified_image"}

\item{y}{Ignored}

\item{...}{Further specifications for \link{plot}.}

\item{time}{Temporal reference for plot.}

\item{title}{A string.}

\item{colors}{Color pallete.}
}
\description{
plots a probability cube using stars
}
\examples{
\donttest{
# Retrieve the samples for Mato Grosso

# select the bands "ndvi", "evi"
samples_ndvi <- sits_select_bands(samples_mt_4bands, ndvi)

#select a random forest model

rfor_model <- sits_train(samples_ndvi, ml_method = sits_rfor())
# Classify a raster file with 23 instances for one year
files <- c(system.file("extdata/raster/mod13q1/sinop-crop-ndvi.tif",
           package = "sits"))
# create a data cube based on the information about the files
sinop <- sits_cube(type = "BRICK", satellite = "TERRA", sensor = "MODIS",
         name = "Sinop-crop", timeline = timeline_modis_392,
         bands = "ndvi", files = files)

# classify the raster image
sinop_probs <- sits_classify(sinop, ml_model = rfor_model,
                             memsize = 2, multicores = 1)

# plot the raster image
plot(sinop_probs)

# remove the files (cleanup)
file.remove(unlist(sinop_probs$files))
}
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
