% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_plot.R
\name{plot.probs_cube}
\alias{plot.probs_cube}
\title{Generic interface for plotting probability cubes}
\usage{
\method{plot}{probs_cube}(
  x,
  y,
  ...,
  time = 1,
  title = "Probabilities for Classes",
  breaks = "kmeans",
  colors = "YlGnBu",
  n_colors = 10
)
}
\arguments{
\item{x}{object of class "probs_image"}

\item{y}{ignored}

\item{...}{further specifications for \link{plot}.}

\item{time}{temporal reference for plot.}

\item{title}{string.}

\item{breaks}{type of breaks}

\item{colors}{color palette.}

\item{n_colors}{number of colors.}
}
\value{
plot
}
\description{
plots a probability cube using stars
}
\examples{
\donttest{
# Retrieve the samples for Mato Grosso
# select the bands for classification
samples_ndvi_evi <- sits_select(samples_mt_4bands, bands = c("EVI", "NDVI"))
# build the classification model
rfor_model <- sits_train(samples_ndvi_evi, ml_method = sits_rfor(num_trees = 2000))

# select the bands "ndvi", "evi" from the "inSitu" package
evi_file <- system.file("extdata/Sinop", "Sinop_evi_2014.tif", package = "inSitu")
ndvi_file <- system.file("extdata/Sinop", "Sinop_ndvi_2014.tif", package = "inSitu")

files <- c(ndvi_file, evi_file)
# define the timeline
time_file <- system.file("extdata/Sinop", "timeline_2014.txt", package = "inSitu")
timeline_2013_2014 <- scan(time_file, character())
# create a raster metadata file based on the information about the files
sinop <- sits_cube(type = "RASTER",
                   satellite = "TERRA",
                   sensor  = "MODIS",
                   name = "Sinop",
                   timeline = timeline_2013_2014,
                   bands = c("NDVI", "EVI"),
                   files = files)

# classify the raster image
sinop_probs <- sits_classify(sinop, ml_model = rfor_model, memsize = 8, multicores = 2)
plot(sinop_probs)
# smooth the result with a bayesian filter
sinop_bayes <- sits_label_classification(sinop_probs, smoothing = "bayesian")

map_1 <- plot(sinop, red = "evi", green = "ndvi", blue = "evi", time = 23)
# plot the smoothened image
plot(sinop_bayes, map = map_1, time = 1, title = "Sinop-Bayes")

# remove the files (cleanup)
file.remove(unlist(sinop_probs$file_info[[1]]$path))
file.remove(unlist(sinop_bayes$file_info[[1]]$path))
}
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
