% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_plot.R
\name{plot.probs_cube}
\alias{plot.probs_cube}
\title{Plot probability cubes}
\usage{
\method{plot}{probs_cube}(
  x,
  ...,
  tiles = NULL,
  labels = NULL,
  breaks = "pretty",
  n_colors = 20,
  palette = "Terrain"
)
}
\arguments{
\item{x}{Object of class "probs_image".}

\item{...}{Further specifications for \link{plot}.}

\item{tiles}{Tiles to be plotted.}

\item{labels}{Labels to plot (optional).}

\item{breaks}{Type of class intervals.}

\item{n_colors}{Number of colors to plot.}

\item{palette}{HCL palette used for visualization.}
}
\value{
The plot itself.
}
\description{
plots a probability cube using stars
}
\note{
\itemize{Possible class intervals
 \item{"sd":} {intervals based on the average and standard deviation.}
 \item{"equal": } {divides the range of the variable into n parts.}
 \item{"pretty": } {number of breaks likely to be legible.}
 \item{"quantile": } {quantile breaks}
 \item{"log": }{logarithm plot}
 }

The function accepts color palettes are defined in grDevices::hcl.pals()
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
