% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_plot.R
\name{plot.probs_cube}
\alias{plot.probs_cube}
\title{Generic interface for plotting probability cubes}
\usage{
\method{plot}{probs_cube}(
  x,
  y,
  ...,
  time = 1,
  title = "Probabilities for Classes",
  breaks = "kmeans",
  colors = "YlGnBu",
  n_colors = 10,
  labels = NULL
)
}
\arguments{
\item{x}{object of class "probs_image"}

\item{y}{ignored}

\item{...}{further specifications for \link{plot}.}

\item{time}{temporal reference for plot.}

\item{title}{string.}

\item{breaks}{type of breaks}

\item{colors}{color palette.}

\item{n_colors}{number of colors.}

\item{labels}{labels to plot (optional)}
}
\value{
The plot itself.
}
\description{
plots a probability cube using stars
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
