% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_plot.R
\name{plot.probs_vector_cube}
\alias{plot.probs_vector_cube}
\title{Plot probability vector cubes}
\usage{
\method{plot}{probs_vector_cube}(
  x,
  ...,
  tile = x[["tile"]][[1L]],
  roi = NULL,
  labels = NULL,
  palette = "YlGn",
  rev = FALSE,
  quantile = NULL,
  scale = 1,
  max_cog_size = 512L,
  seg_color = "black",
  line_width = 0.5,
  legend_position = "outside",
  legend_title = "probs"
)
}
\arguments{
\item{x}{Object of class "probs_vector_cube".}

\item{...}{Further specifications for \link{plot}.}

\item{tile}{Tile to be plotted.}

\item{roi}{Region of interest (see notes below).}

\item{labels}{Labels to plot}

\item{palette}{RColorBrewer or "cols4all" palette}

\item{rev}{Reverse order of colors in palette?}

\item{quantile}{Minimum quantile to plot}

\item{scale}{Scale to plot map (0.4 to 1.0)}

\item{max_cog_size}{Maximum size of COG overviews (lines or columns)}

\item{seg_color}{Color for segment borders (default = "black")}

\item{line_width}{Line width for segment borders (default = 0.5)}

\item{legend_position}{Where to place the legend (default = "outside")}

\item{legend_title}{Title of legend (default = "probs")}
}
\value{
A plot containing probabilities associated
                      to each class with segment overlay.
}
\description{
Plots a probability vector cube, which result from
first running a segmentation \code{\link{sits_segment}} and then
running a machine learning classification model. The result is
a probability raster overlaid with segment vector boundaries.
}
\note{
To see which color palettes are supported, please run cols4all::c4a_gui().
To define a \code{roi} use one of:
     \itemize{
       \item{A path to a shapefile with polygons;}
       \item{A \code{sfc} or \code{sf} object from \code{sf} package;}
       \item{A \code{SpatExtent} object from \code{terra} package;}
       \item{A named \code{vector} (\code{"lon_min"},
            \code{"lat_min"}, \code{"lon_max"}, \code{"lat_max"}) in WGS84;}
       \item{A named \code{vector} (\code{"xmin"}, \code{"xmax"},
             \code{"ymin"}, \code{"ymax"}) with XY coordinates.}
}
}
\examples{
if (sits_run_examples()) {
    # create a random forest model
    rfor_model <- sits_train(samples_modis_ndvi, sits_rfor())
    # create a data cube from local files
    data_dir <- system.file("extdata/raster/mod13q1", package = "sits")
    cube <- sits_cube(
        source = "BDC",
        collection = "MOD13Q1-6.1",
        data_dir = data_dir
    )
    # segment the image
    segments <- sits_segment(
        cube = cube,
        seg_fn = sits_snic(
            grid_seeding = "diamond",
            spacing = 7,
            compactness = 0.5,
            padding = 0
        ),
        output_dir = tempdir()
    )
    # classify a data cube
    probs_vector_cube <- sits_classify(
        data = segments,
        ml_model = rfor_model,
        output_dir = tempdir()
    )
    # plot the resulting probability cube with segment overlay
    plot(probs_vector_cube)
}

}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
