% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_plot.R
\name{plot.probs_vector_cube}
\alias{plot.probs_vector_cube}
\title{Plot probability vector cubes}
\usage{
\method{plot}{probs_vector_cube}(
  x,
  ...,
  tile = x$tile[[1]],
  labels = NULL,
  palette = "YlGnBu",
  rev = FALSE,
  tmap_options = NULL
)
}
\arguments{
\item{x}{Object of class "probs_vector_cube".}

\item{...}{Further specifications for \link{plot}.}

\item{tile}{Tile to be plotted.}

\item{labels}{Labels to plot (optional).}

\item{palette}{RColorBrewer palette}

\item{rev}{Reverse order of colors in palette?}

\item{tmap_options}{Named vector with optional tmap parameters
max_cells (default: 1e+06)
scale (default: 0.5)
graticules_labels_size (default: 0.7)
legend_title_size (default: 1.0)
legend_text_size (default: 1.0)
legend_bg_color (default: "white")
legend_bg_alpha (default: 0.5)}
}
\value{
A plot containing probabilities associated
                      to each class for each pixel.
}
\description{
plots a probability cube using stars
}
\examples{
if (sits_run_examples()) {
    # create a random forest model
    rfor_model <- sits_train(samples_modis_ndvi, sits_rfor())
    # create a data cube from local files
    data_dir <- system.file("extdata/raster/mod13q1", package = "sits")
    cube <- sits_cube(
        source = "BDC",
        collection = "MOD13Q1-6",
        data_dir = data_dir
    )
    # segment the image
    segments <- sits_segment(
        cube = cube,
        seg_fn = sits_slic(step = 5,
                           compactness = 1,
                           dist_fun = "euclidean",
                           avg_fun = "median",
                           iter = 20,
                           minarea = 10,
                           verbose = FALSE),
        output_dir = tempdir()
    )
    # classify a data cube
    probs_vector_cube <- sits_classify(
        data = segments,
        ml_model = rfor_model,
        output_dir = tempdir()
    )
    # plot the resulting probability cube
    plot(probs_vector_cube)
}

}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
