% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_plot.R
\name{plot.raster_cube}
\alias{plot.raster_cube}
\title{Plot RGB data cubes}
\usage{
\method{plot}{raster_cube}(
  x,
  ...,
  band = NULL,
  red = NULL,
  green = NULL,
  blue = NULL,
  tiles = NULL,
  date = NULL,
  roi = NULL
)
}
\arguments{
\item{x}{Object of class "raster_cube".}

\item{...}{Further specifications for \link{plot}.}

\item{band}{Band for plotting grey images.}

\item{red}{Band for red color.}

\item{green}{Band for green color.}

\item{blue}{Band for blue color.}

\item{tiles}{Tiles to be plotted.}

\item{date}{Date to be plotted.}

\item{roi}{sf object giving a region of interest.}
}
\value{
Plot object.
}
\description{
Uses mapview to visualize raster cube and classified images
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
