% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_plot.R
\name{plot.raster_cube}
\alias{plot.raster_cube}
\title{Generic interface for plotting stack cubes}
\usage{
\method{plot}{raster_cube}(x, y, ...)
}
\arguments{
\item{x}{object of class "raster_cube"}

\item{y}{ignored}

\item{...}{further specifications for \link{plot}.}
}
\description{
This function is deprecated and replaced by sits_view
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
