% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_plot.R
\name{plot.raster_cube}
\alias{plot.raster_cube}
\title{Plot RGB data cubes}
\usage{
\method{plot}{raster_cube}(
  x,
  ...,
  band = NULL,
  red = NULL,
  green = NULL,
  blue = NULL,
  tiles = x$tile[[1]],
  dates = NULL,
  colors = rev(grDevices::terrain.colors(30, alpha = NULL))
)
}
\arguments{
\item{x}{Object of class "raster_cube".}

\item{...}{Further specifications for \link{plot}.}

\item{band}{Band for plotting grey images.}

\item{red}{Band for red color.}

\item{green}{Band for green color.}

\item{blue}{Band for blue color.}

\item{tiles}{Tiles to be plotted.}

\item{dates}{Dates to be plotted.}

\item{colors}{An R color palette}

\item{samples}{samples to be plotted together with image}

\item{sample_color}{RColorBrewer name to display the samples}
}
\value{
A plot object produced by the terra package
                      with an RGB image or a B/W image on a color
                      scale chosen from the RColorBrewer palette.
}
\description{
Plot RGB raster cube
}
\examples{
if (sits_run_examples()) {
    # create a data cube from local files
    data_dir <- system.file("extdata/raster/mod13q1", package = "sits")
    cube <- sits_cube(
        source = "BDC",
        collection = "MOD13Q1-6",
        data_dir = data_dir,
        delim = "_",
        parse_info = c("X1", "tile", "band", "date")
    )
    # plot NDVI band of the second date date of the data cube
    plot(cube, band = "NDVI", dates = sits_timeline(cube)[c(1:2)])
}
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
