% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_plot.R
\name{plot.raster_cube}
\alias{plot.raster_cube}
\title{Generic interface for plotting stack cubes}
\usage{
\method{plot}{raster_cube}(x, y, ..., band, time = 1)
}
\arguments{
\item{x}{object of class "raster_cube"}

\item{y}{ignored}

\item{...}{further specifications for \link{plot}.}

\item{band}{band to be plotted}

\item{time}{time instance}
}
\description{
Generic interface for plotting stack cubes
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
