% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_plot.R
\name{plot.raster_cube}
\alias{plot.raster_cube}
\title{Plot RGB data cubes}
\usage{
\method{plot}{raster_cube}(x, ..., band = NULL, red, green, blue, tile = 1, time = 1, roi = NULL)
}
\arguments{
\item{x}{object of class "raster_cube" or "classified image"}

\item{...}{further specifications for \link{sits_view}.}

\item{band}{for plotting grey images}

\item{red}{band for red color.}

\item{green}{band for green color.}

\item{blue}{band for blue color.}

\item{tile}{tile number to be plotted}

\item{time}{temporal instances to be plotted.}

\item{roi}{sf object giving a region of interest.}
}
\value{
plot object
}
\description{
Uses mapview to visualize raster cube and classified images
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
