% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_plot.R
\name{plot.raster_cube}
\alias{plot.raster_cube}
\title{Plot RGB data cubes}
\usage{
\method{plot}{raster_cube}(
  x,
  ...,
  band = NULL,
  red = NULL,
  green = NULL,
  blue = NULL,
  tiles = NULL,
  date = NULL,
  roi = NULL
)
}
\arguments{
\item{x}{object of class "raster_cube" or "classified image"}

\item{...}{further specifications for \link{sits_view}.}

\item{band}{for plotting grey images}

\item{red}{band for red color.}

\item{green}{band for green color.}

\item{blue}{band for blue color.}

\item{tiles}{tiles to be plotted}

\item{date}{date to be plotted.}

\item{roi}{sf object giving a region of interest.}
}
\value{
plot object
}
\description{
Uses mapview to visualize raster cube and classified images
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
