% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_plot.R
\name{plot.raster_cube}
\alias{plot.raster_cube}
\title{Plot RGB data cubes}
\usage{
\method{plot}{raster_cube}(
  x,
  ...,
  band = NULL,
  red = NULL,
  green = NULL,
  blue = NULL,
  tile = x[["tile"]][[1]],
  dates = NULL,
  palette = "RdYlGn",
  rev = FALSE,
  scale = 1.2
)
}
\arguments{
\item{x}{Object of class "raster_cube".}

\item{...}{Further specifications for \link{plot}.}

\item{band}{Band for plotting grey images.}

\item{red}{Band for red color.}

\item{green}{Band for green color.}

\item{blue}{Band for blue color.}

\item{tile}{Tile to be plotted.}

\item{dates}{Dates to be plotted.}

\item{palette}{An RColorBrewer palette}

\item{rev}{Reverse the color order in the palette?}

\item{scale}{Scale to plot map (0.4 to 1.0)}
}
\value{
A plot object with an RGB image
                      or a B/W image on a color scale
}
\description{
Plot RGB raster cube
}
\note{
Use \code{scale} parameter for general output control.
      The \code{dates} parameter indicates the date allows plotting of different dates when
      a single band and three dates are provided, `sits` will plot a
      multi-temporal RGB image for a single band (useful in the case of
      SAR data). For RGB bands with multi-dates, multiple plots will be
      produced.
}
\examples{
if (sits_run_examples()) {
    # create a data cube from local files
    data_dir <- system.file("extdata/raster/mod13q1", package = "sits")
    cube <- sits_cube(
        source = "BDC",
        collection = "MOD13Q1-6",
        data_dir = data_dir
    )
    # plot NDVI band of the second date date of the data cube
    plot(cube, band = "NDVI", dates = sits_timeline(cube)[1])
    # plot NDVI band as an RGB composite for the three bands
}
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
