% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_plot.R
\name{plot.sits_cluster}
\alias{plot.sits_cluster}
\title{Plot a dendrogram cluster}
\usage{
\method{plot}{sits_cluster}(x, ..., cluster, cutree_height, palette)
}
\arguments{
\item{x}{sits tibble with cluster indexes.}

\item{cluster}{cluster object produced by `sits_cluster` function.}

\item{cutree_height}{dashed horizontal line to be drawn
indicating the height of dendrogram cutting.}

\item{palette}{hcl color palette.}
}
\value{
The dendrogram object.
}
\description{
Plot a dendrogram
}
\author{
Rolf Simoes, \email{rolf.simoes@inpe.br}
}
