% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_plot.R
\name{plot.som_evaluate_cluster}
\alias{plot.som_evaluate_cluster}
\title{Plot information about confunsion between clusters}
\usage{
\method{plot}{som_evaluate_cluster}(x, y, ..., name_cluster = NULL, title = "Confusion by cluster")
}
\arguments{
\item{x}{object of class "plot.som_evaluate_cluster"}

\item{y}{ignored}

\item{...}{further specifications for \link{plot}.}

\item{name_cluster}{Choose the cluster to plot}

\item{title}{title of plot. default is ""Confusion by cluster"".}
}
\value{
The plot itself.
}
\description{
Plot a bar graph with informations about each cluster.
The percentage of mixture between the clusters.
}
\examples{
\dontrun{
# Produce a cluster map

samples_mt_2bands <- sits_select(samples_mt_6bands, bands = c("NDVI", "EVI"))
som_map <- sits_som_map(samples_mt_2bands)
# Evaluate the clusters
cluster_overall <- sits_som_evaluate_cluster(som_map)
# Plot confusion between the clusters
plot(cluster_overall)
}
}
\author{
Lorena Santos \email{lorena.santos@inpe.br}
}
