% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_plot.R
\name{plot.stack_cube}
\alias{plot.stack_cube}
\title{Generic interface for plotting probability cubes}
\usage{
\method{plot}{stack_cube}(x, y, ..., red = 1, green = 1, blue = 1)
}
\arguments{
\item{x}{Object of class "classified_image"}

\item{y}{Ignored}

\item{...}{Further specifications for \link{plot}.}

\item{red}{Band for red color.}

\item{green}{Band for green color.}

\item{blue}{Band for blue color.}
}
\description{
plots a stack cube using terra
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
