% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_plot.R
\name{plot.stack_cube}
\alias{plot.stack_cube}
\title{Generic interface for plotting stack cubes}
\usage{
\method{plot}{stack_cube}(x, y, ..., red, green, blue, time = 1, roi = NULL)
}
\arguments{
\item{x}{object of class "stack_cube"}

\item{y}{ignored}

\item{...}{further specifications for \link{plot}.}

\item{red}{band for red color.}

\item{green}{band for green color.}

\item{blue}{band for blue color.}

\item{time}{temporal instances to be plotted.}

\item{roi}{sf object giving a region of interest.}
}
\value{
mapview object
}
\description{
plots a stack cube using terra
}
\examples{
\dontrun{
data_dir <- system.file("extdata/raster/cbers", package = "sits")

cbers_022024 <- sits_cube(
    type = "STACK",
    name = "cbers_022024",
    satellite = "CBERS-4",
    sensor = "AWFI",
    resolution = "64m",
    data_dir = data_dir,
    parse_info = c("X1", "X2", "band", "date")
)
# plot the data cube
plot(cbers_022024, red = "B15", green = "B16", blue = "B13", time = 1)
}

}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
