% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_plot.R
\name{plot.uncertainty_cube}
\alias{plot.uncertainty_cube}
\title{Plot uncertainty cubes}
\usage{
\method{plot}{uncertainty_cube}(x, y, ..., tile = 1, n_colors = 20, palette = "Blues")
}
\arguments{
\item{x}{object of class "probs_image"}

\item{y}{ignored}

\item{...}{further specifications for \link{plot}.}

\item{tile}{tile number to be plotted}

\item{n_colors}{number of colors to plot}

\item{palette}{hcl palette used for visualisation}
}
\value{
The plot itself.
}
\description{
plots a probability cube using stars
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
