% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_plot.R
\name{plot.uncertainty_cube}
\alias{plot.uncertainty_cube}
\title{Plot uncertainty cubes}
\usage{
\method{plot}{uncertainty_cube}(
  x,
  ...,
  tiles = NULL,
  n_colors = 14,
  intervals = "log",
  palette = "YlOrRd"
)
}
\arguments{
\item{x}{Object of class "probs_image".}

\item{...}{Further specifications for \link{plot}.}

\item{tiles}{Tiles to be plotted.}

\item{n_colors}{Number of colors to plot.}

\item{intervals}{Type of class intervals.}

\item{palette}{HCL palette used for visualization.}
}
\value{
The plot itself.
}
\description{
plots a probability cube using stars
}
\note{
\itemize{Possible class intervals
 \item{"sd":} {intervals based on the average and standard deviation.}
 \item{"equal": } {divides the range of the variable into n parts.}
 \item{"quantile": } {quantile breaks}
 \item{"pretty": } {number of breaks likely to be legible.}
 \item{"log" :} {logarithm plot.}
 }
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
