% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_plot.R
\name{plot.uncertainty_cube}
\alias{plot.uncertainty_cube}
\title{Plot uncertainty cubes}
\usage{
\method{plot}{uncertainty_cube}(
  x,
  y,
  ...,
  tiles = NULL,
  n_colors = 10,
  breaks = "pretty",
  palette = "Blues"
)
}
\arguments{
\item{x}{Object of class "probs_image".}

\item{y}{Ignored.}

\item{...}{Further specifications for \link{plot}.}

\item{tiles}{Tiles to be plotted.}

\item{n_colors}{Number of colors to plot.}

\item{breaks}{Yype of class intervals.}

\item{palette}{HCL palette used for visualisation.}
}
\value{
The plot itself.
}
\description{
plots a probability cube using stars
}
\note{
\itemize{Possible class intervals
 \item{"sd":} {intervals based on the average and standard deviation.}
 \item{"equal": } {divides the range of the variable into n parts.}
 \item{"pretty": } {number of breaks likely to be legible.}
 \item{"quantile": } {quantile breaks}
 \item{"kmeans" :} {uses kmeans to generate the breaks.}
 \item{"hclust" :} {breaks defined by hierarchical clustering.}
 \item{"bclust" :} {breaks defined by bagged clustering.}
 \item{"fisher" :} {method proposed by Fischer (1958).}
 \item{"jenks" :} {method proposed by Jenks.}
 \item{"dpih" :} {based on the bin width of a histogram.}
 \item{"headtails" :} {algorithm proposed by Bin Jiang (2013)}
 }
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
