% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_plot.R
\name{plot.vector_cube}
\alias{plot.vector_cube}
\title{Plot RGB vector data cubes}
\usage{
\method{plot}{vector_cube}(
  x,
  ...,
  band = sits_bands(x)[1],
  red = NULL,
  green = NULL,
  blue = NULL,
  tile = x$tile[[1]],
  date = NULL,
  seg_color = "black",
  line_width = 1,
  palette = "RdYlGn",
  style = "cont",
  n_colors = 10,
  rev = FALSE,
  scale = 0.8
)
}
\arguments{
\item{x}{Object of class "raster_cube".}

\item{...}{Further specifications for \link{plot}.}

\item{band}{Band for plotting grey images.}

\item{red}{Band for red color.}

\item{green}{Band for green color.}

\item{blue}{Band for blue color.}

\item{tile}{Tile to be plotted.}

\item{date}{Date to be plotted.}

\item{seg_color}{Color to show the segment boundaries}

\item{line_width}{Line width to plot the segments boundary (in pixels)}

\item{palette}{An RColorBrewer palette}

\item{n_colors}{Number of colors to be shown}

\item{rev}{Reverse the color order in the palette?}

\item{scale}{Scale to plot map (0.4 to 1.0)}
}
\value{
A plot object with an RGB image
                      or a B/W image on a color
                      scale using the pallete
}
\description{
Plot RGB raster cube
}
\note{
To see which color palettes are supported, please run
}
\examples{
if (sits_run_examples()) {
    # create a data cube from local files
    data_dir <- system.file("extdata/raster/mod13q1", package = "sits")
    cube <- sits_cube(
        source = "BDC",
        collection = "MOD13Q1-6",
        data_dir = data_dir
    )
    # Segment the cube
    segments <- sits_segment(
        cube = cube,
        output_dir = tempdir(),
        multicores = 2,
        memsize = 4
    )
    # plot NDVI band of the second date date of the data cube
    plot(segments, band = "NDVI", date = sits_timeline(cube)[1])
}
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
