% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits-package.R
\docType{package}
\name{sits-package}
\alias{sits-package}
\alias{_PACKAGE}
\alias{sits}
\title{sits}
\description{
Satellite Image Time Series Analysis
             for Earth Observation Data Cubes
}
\section{Purpose}{


The SITS package provides a set of tools for analysis,
visualization and classification of satellite image time series.
It includes methods for filtering, clustering, classification,
and post-processing.
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/e-sensing/sits/}
  \item \url{https://e-sensing.github.io/sitsbook/}
  \item Report bugs at \url{https://github.com/e-sensing/sits/issues}
}

}
\author{
\strong{Maintainer}: Gilberto Camara \email{gilberto.camara.inpe@gmail.com} [thesis advisor]

Authors:
\itemize{
  \item Rolf Simoes \email{rolf.simoes@inpe.br}
  \item Felipe Souza \email{felipe.carvalho@inpe.br}
  \item Felipe Carlos \email{efelipecarlos@gmail.com}
}

Other contributors:
\itemize{
  \item Lorena Santos \email{lorena.santos@inpe.br} [contributor]
  \item Karine Ferreira \email{karine.ferreira@inpe.br} [contributor, thesis advisor]
  \item Charlotte Pelletier \email{charlotte.pelletier@univ-ubs.fr} [contributor]
  \item Pedro Andrade \email{pedro.andrade@inpe.br} [contributor]
  \item Alber Sanchez \email{alber.ipia@inpe.br} [contributor]
  \item Gilberto Queiroz \email{gilberto.queiroz@inpe.br} [contributor]
}

}
