% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_accuracy.R
\name{sits_accuracy}
\alias{sits_accuracy}
\title{Evaluates the accuracy of classification}
\usage{
sits_accuracy(conf.tb, conv.lst = NULL, pred_sans_ext = FALSE)
}
\arguments{
\item{conf.tb}{A tibble containing pairs of reference and predicted values}

\item{conv.lst}{A list conversion list of labels. If NULL no conversion is done.}

\item{pred_sans_ext}{(Boolean) remove all label extension (i.e. every string after last '.' character) from predictors before compute assesment.}
}
\value{
caret_assess  a confusion matrix assessment produced by the caret package
}
\description{
Evaluates the accuracy of classification stored in two vectors.
Returns a confusion matrix used by the "caret" package
}
\author{
Rolf Simoes, \email{rolf.simoes@inpe.br}

Gilberto Camara, \email{gilberto.camara@inpe.br}
}
