% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_assessment.R
\name{sits_accuracy}
\alias{sits_accuracy}
\title{Evaluates the accuracy of classification}
\usage{
sits_accuracy(pred.vec, ref.vec, pred_sans_ext = FALSE, conv.lst = NULL)
}
\arguments{
\item{pred.vec}{A vector of all predicted labels.}

\item{ref.vec}{A vector of all reference labels.}

\item{pred_sans_ext}{(Boolean) remove all label extension (i.e. every string after last '.' character) from predictors before compute assesment.}

\item{conv.lst}{A list conversion list of labels. If NULL no conversion is done.}
}
\value{
result.lst     a list with accuracy measures and confusion matrix
}
\description{
Evaluates the accuracy of classification stored in two vectors.
Returns the overall accuracy, producers and users accuracy, and confusion matrix.
This algorith was inspired by `rfUtilities::accuracy` function, and fix the user and
producer accuracy computation inconsistency.
}
\author{
Rolf Simoes, \email{rolf.simoes@inpe.br}
}
