% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_assessment.R
\name{sits_accuracy}
\alias{sits_accuracy}
\title{Post-classification accuracy assessment of classified maps (non-weigthed)}
\usage{
sits_accuracy(results.tb)
}
\arguments{
\item{results.tb}{a sits table with a set of lat/long/time locations with known and trusted labels and
with the result of a classification method}
}
\value{
assessment  a list containing overall accuracy, producers and users accuracy, and confusion matrix.
}
\description{
To use this function, the input table should be a set of results containing
both the label assigned by the user and the classification result.
Accuracy assessment set us a confusion matrix to determine the accuracy of your classified result.
This function does not use an area-weigthed technique and should be used only as a
first check of accuracy. When the area of each class for the region of interest is available,
please use sits_accuracy_area instead
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
