% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_accuracy.R
\name{sits_accuracy}
\alias{sits_accuracy}
\title{Area-weighted classification accuracy assessment}
\usage{
sits_accuracy(label_cube, validation_csv)
}
\arguments{
\item{label_cube}{A data cube with classified images.}

\item{validation_csv}{A CSV file path with validation data}
}
\value{
A list of lists: The error_matrix, the class_areas, the unbiased
estimated areas, the standard error areas, confidence interval 95% areas,
and the accuracy (user, producer, and overall), or NULL if the data is empty.
}
\description{
To use this function the input table should be
a set of results containing
both the label assigned by the user and the classification result.
Accuracy assessment set us a confusion matrix to determine the accuracy
of your classified result.
This function uses an area-weighted technique proposed by Olofsson et al. to
produce more reliable accuracy estimates at 95% confidence level.

This function performs an accuracy assessment of the classified, including
Overall Accuracy, User's Accuracy, Producer's Accuracy
and error matrix (confusion matrix) according to [1-3].
}
\examples{
\dontrun{
# get the samples for Mato Grosso for bands NDVI and EVI
samples_mt_ndvi <- sits_select(samples_mt_4bands, bands = c("NDVI"))

# filter the samples for three classes (to simplify the example)
samples_mt_ndvi <- dplyr::filter(samples_mt_ndvi, label \%in\%
  c("Forest", "Pasture", "Soy_Corn"))

# build an extreme gradient boosting model
xgb_model <- sits_train(
  samples_mt_ndvi,
  sits_xgboost(nrounds = 10, verbose = FALSE)
)

# create a data cube based on files
data_dir <- system.file("extdata/raster/mod13q1", package = "sits")
cube <- sits_cube(
    source = "LOCAL",
    name = "sinop-2014",
    satellite = "TERRA",
    sensor = "MODIS",
    data_dir = data_dir,
    delim = "_",
    parse_info = c("X1", "X2", "band", "date")
)


# classify the data cube with xgb model
probs_cube <- sits_classify(cube,
  xgb_model,
  output_dir = tempdir(),
  memsize = 4,
  multicores = 1
)
# label the classification
label_cube <- sits_label_classification(probs_cube,
  output_dir = tempdir()
)
# get ground truth points
ground_truth <- system.file("extdata/samples/samples_sinop_crop.csv",
  package = "sits"
)
# calculate accuracy according to Olofsson's method
accuracy <- suppressWarnings(sits_accuracy(label_cube, ground_truth))
}
}
\references{
[1] Olofsson, P., Foody, G.M., Stehman, S.V., Woodcock, C.E. (2013).
Making better use of accuracy data in land change studies: Estimating
accuracy and area and quantifying uncertainty using stratified estimation.
Remote Sensing of Environment, 129, pp.122-131.

[2] Olofsson, P., Foody G.M., Herold M., Stehman, S.V.,
Woodcock, C.E., Wulder, M.A. (2014)
Good practices for estimating area and assessing accuracy of land change.
Remote Sensing of Environment, 148, pp. 42-57.

[3] FAO, Map Accuracy Assessment and Area Estimation: A Practical Guide.
National forest monitoring assessment working paper No.46/E, 2016.
}
\author{
Rolf Simoes, \email{rolf.simoes@inpe.br}

Alber Sanchez, \email{alber.ipia@inpe.br}
}
