% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_accuracy.R
\name{sits_accuracy_area}
\alias{sits_accuracy_area}
\title{Area-weighted classification accuracy assessment}
\usage{
sits_accuracy_area(class.tb, area = NULL)
}
\arguments{
\item{class.tb}{A sits tibble with a set of lat/long/time locations
with known and trusted labels and
with the result of classification method.}

\item{area}{A list with the area of each label.}
}
\description{
To use this function, the input table should be
a set of results containing
both the label assigned by the user and the classification result.
Accuracy assessment set us a confusion matrix to determine the accuracy
of your classified result.
This function uses an area-weighted technique proposed by Olofsson et al. to
produce more reliable accuracy estimates at 95% confidence level.

This function performs an accuracy assessment of the classified, including
Overall Accuracy, User's Accuracy, Producer's Accuracy
and error matrix (confusion matrix) according to [1-2].
}
\examples{
\donttest{
# Retrieve the set of samples for the Mato Grosso (provided by EMBRAPA)
samples_2bands <- sits_select_bands(samples_mt_6bands, ndvi, evi)
set.seed(42)

# Fit a randon forest model to the samples.
rfor_model <- sits_train(samples_2bands, ml_method = sits_rfor())

# Classify samples.
class.tb <- sits_classify(samples_2bands, rfor_model)


# Simulate the area of each label in a reference map.
n_labels <- nrow(sits_labels(class.tb))
area <- sample(1:100 * 10^3, size = n_labels)
names(area) <- unique(class.tb$label)

# Compute the accuracy.
sits_accuracy_area(class.tb, area)
}
}
\references{
[1] Olofsson, P., Foody, G.M., Stehman, S.V., Woodcock, C.E. (2013).
Making better use of accuracy data in land change studies: Estimating
accuracy and area and quantifying uncertainty using stratified estimation.
Remote Sensing of Environment, 129, pp.122-131.

[2] Olofsson, P., Foody G.M., Herold M., Stehman, S.V.,
Woodcock, C.E., Wulder, M.A. (2014)
Good practices for estimating area and assessing accuracy of land change.
Remote Sensing of
Environment, 148, pp. 42-57.
}
\author{
Alber Sanchez, \email{alber.ipia@inpe.br}
}
