% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_accuracy.R
\name{sits_accuracy_summary}
\alias{sits_accuracy_summary}
\title{Print the ssumary of the accuracy}
\usage{
sits_accuracy_summary(
  x,
  mode = "sens_spec",
  digits = max(3, getOption("digits") - 3)
)
}
\arguments{
\item{x}{An object of class \code{sits_assessment}.}

\item{mode}{A single character string either "sens_spec",
"prec_recall", or "everything".}

\item{digits}{Number of significant digits when printed.}
}
\value{
\code{x}   is invisibly returned.
}
\description{
Adaptation of the caret::print.confusionMatrix method
             for the more common usage in Earth Observation.
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
\keyword{internal}
