% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_utils.R
\name{sits_active_tests}
\alias{sits_active_tests}
\title{Informs if sits tests should run}
\usage{
sits_active_tests()
}
\value{
TRUE/FALSE
}
\description{
This function informs if sits test should run.
Useful to avoid running slow tests in CRAN environment.
Behaviour controlled by environmental variable R_CONFIG_ACTIVE_TESTS
}
\examples{
# unset active tests
Sys.setenv("R_CONFIG_ACTIVE_TESTS"="NO")
# result should be false
isFALSE(sits_active_tests())
# set active tests
Sys.setenv("R_CONFIG_ACTIVE_TESTS"="YES")
# result should be true
isTRUE(sits_active_tests())

}
