% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_tibble_functions.R
\name{sits_adjust}
\alias{sits_adjust}
\title{Adjust a SITS tibble based on information about ref dates}
\usage{
sits_adjust(data.tb, class_info.tb)
}
\arguments{
\item{data.tb}{a tibble in SITS format with time series}

\item{class_info.tb}{a tibble with information on timeline, dates and interval for classification}
}
\value{
subset.tb      a tibble in SITS format with the subsets of the input data that match the classification requirements
}
\description{
The reference dates for the classification are an important information,
since they will constrain the data used for training the machine learning methods.
It is important that the all data used for ML have the same amount of samples. For
this reason, it is a good practice to adjust the training samples with the
expected dates for classification (defined by the function \code{\link[sits]{sits_class_info}} )
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
