% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_tibble_functions.R
\name{sits_align}
\alias{sits_align}
\title{Align a data set to the dates of a set of samples, given an interval}
\usage{
sits_align(data.tb, samples.tb, interval = "12 month")
}
\arguments{
\item{data.tb}{a valid sits tibble}

\item{samples.tb}{the samples used in the classification}

\item{interval}{the interval between the classificationa}
}
\description{
This function aligns an input data set to the dates used by a set of samples
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
