% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_table.R
\name{sits_align}
\alias{sits_align}
\title{Aligns dates of time series to a reference date}
\usage{
sits_align(data.tb, from = NULL, to = NULL, sample_freq = 16)
}
\arguments{
\item{data.tb}{tibble - input SITS table (useful for chaining functions)}

\item{from}{starting date of the estimate in month-day}

\item{to}{end data of the estimated in month-day}

\item{sample_freq}{int - the interval in days for the samples}
}
\value{
data1.tb      tibble - the converted SITS table (useful for chaining functions)
}
\description{
converts the time indexes of a set of sits tables to a single reference year.
This function is useful to join many time series from different years to a single year,
which is required by methods that combine many time series, such as clustering methods.
The reference year is taken from the date of the start of the time series
available in the coverage.
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
