% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_table.R
\name{sits_align}
\alias{sits_align}
\title{Aligns dates of time series to a reference date}
\usage{
sits_align(data.tb, ref_date)
}
\arguments{
\item{data.tb}{tibble - input SITS table (useful for chaining functions)}

\item{ref_date}{date   - a reference date where all series will start}
}
\value{
data1.tb   tibble - the converted SITS table (useful for chaining functions)
}
\description{
\code{sits_align} convert the index of a sits time to a reference year
The reference year is taken from the date of the start of the time series
available in the coverage.
}
\seealso{
Other STIS table functions: \code{\link{sits_bands}},
  \code{\link{sits_dates}},
  \code{\link{sits_group_bylatlong}},
  \code{\link{sits_rename}}, \code{\link{sits_select}},
  \code{\link{sits_table}}, \code{\link{sits_values_cols}},
  \code{\link{sits_values_rows}}
}
