% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_table.R
\name{sits_apply}
\alias{sits_apply}
\title{Apply a function over SITS bands.}
\usage{
sits_apply(data.tb, fun, fun_index = NULL, bands_suffix = "")
}
\arguments{
\item{data.tb}{a valid sits table}

\item{fun}{a function with one parameter as input and a vector or list of vectors as output.}

\item{fun_index}{a function with one parameter as input and a Date vector as output.}

\item{bands_suffix}{a string informing the resulting bands name's suffix.}
}
\value{
result.tb    a sits_table with same samples and the new bands
}
\description{
`sits_apply` returns a sits_table with the same samples points and new bands computed by `fun`,
`fun_index` functions. These functions must be defined inline and are called by `sits_ts_apply` for each band,
whose vector values is passed as the function argument.

`fun` function may either return a vector or a list of vectors. In the first case, the vector will be the new values
of the corresponding band. In the second case, the returned list must have names, and each element vector will
generate a new band which name composed by concatenating original band name and the corresponding list element name.

If a suffix is provided in `bands_suffix`, all resulting bands names will end with provided suffix separated by a ".".

The boolean parameter `original` controls whether the original bands should be also part of
the out SITS table.
}
\author{
Rolf Simoes, \email{rolf.simoes@inpe.br}
}
