% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_sf.R
\name{sits_as_sf}
\alias{sits_as_sf}
\title{Return a sits_tibble as a point sf object.}
\usage{
sits_as_sf(samples)
}
\arguments{
\item{samples}{A sits tibble with one or more time series.}
}
\value{
Return a sits_tibble as a sf object of point geometry.
}
\description{
Return a sits_tibble as a sf object of point geometry.
}
\author{
Felipe Carvalho, \email{felipe.carvalho@inpe.br}

Alber Sanchez, \email{alber.ipia@inpe.br}
}
