% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_assert.R
\name{sits_assert}
\alias{sits_assert}
\title{Verify if the sits table has been created}
\usage{
sits_assert(table_obj)
}
\arguments{
\item{table_name}{a string - name of table (must exist)}
}
\description{
\code{sits_assert} tests if the sits table exists
This is an utility function used to test if a table exists
}

