% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_assess.R
\name{sits_assess}
\alias{sits_assess}
\title{Assess time series classification}
\usage{
sits_assess(results.tb, area, conf.int = 0.95, rm.nosample = FALSE)
}
\arguments{
\item{results.tb}{an object of "sits_table" with the results of a time series classification}

\item{area}{a numeric vector with the area for each class.}

\item{conf.int}{specifies the confidence level (0-1).}

\item{rm.nosample}{if sum of columns and sum of rows of the error matrix are zero
then remove class. Default is TRUE.}
}
\description{
This function calls \code{\link[dtwSat]{twdtwAssess}} from \pkg{dtwSat}.
\code{\link[dtwSat]{twdtwAssess}} performs an accuracy assessment of the classified, including
Overall Accuracy, User's Accuracy, Produce's Accuracy, error matrix (confusion matrix),
and estimated area according to [1-2].
}
\references{
[1] Olofsson, P., Foody, G.M., Stehman, S.V., Woodcock, C.E. (2013).
Making better use of accuracy data in land change studies: Estimating
accuracy and area and quantifying uncertainty using stratified estimation.
Remote Sensing of Environment, 129, pp.122-131.

[2] Olofsson, P., Foody G.M., Herold M., Stehman, S.V., Woodcock, C.E., Wulder, M.A. (2014)
Good practices for estimating area and assessing accuracy of land change. Remote Sensing of
Environment, 148, pp. 42-57.
}
\author{
Victor Maus, \email{vwmaus1@gmail.com}

Gilberto Camara, \email{gilberto.camara@inpe.br}
}
