% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_bands.R
\name{sits_bands}
\alias{sits_bands}
\title{Informs the names of the bands}
\usage{
sits_bands(x)
}
\arguments{
\item{x}{Valid sits tibble (time series or a cube)}
}
\value{
A string vector with the names of the bands.
}
\description{
Finds/replaces the names of the bands of
              a set of time series or of a data cube
}
\examples{
{
# Retrieve the set of samples for Mato Grosso (provided by EMBRAPA)
# show the bands
sits_bands(samples_mt_6bands)
}

}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}

Rolf Simoes, \email{rolf.simoes@inpe.br}
}
