% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_tibble_functions.R
\name{sits_bands}
\alias{sits_bands}
\title{Informs the names of the bands of a time series}
\usage{
sits_bands(data)
}
\arguments{
\item{data}{Valid sits tibble (time series or a cube)}
}
\value{
A string vector with the names of the bands.
}
\description{
Finds the names of the bands of time series in a sits tibble
              or in a metadata cube
}
\examples{
# Retrieve the set of samples for the Mato Grosso region (provided by EMBRAPA)
data(samples_mt_6bands)
# print the bands
sits_bands(samples_mt_6bands)
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}

Rolf Simoes, \email{rolf.simoes@inpe.br}
}
