% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_data_generics.R
\name{sits_bands}
\alias{sits_bands}
\title{Informs the names of the bands}
\usage{
sits_bands(data)
}
\arguments{
\item{data}{Valid sits tibble (time series or a cube)}
}
\value{
A string vector with the names of the bands.
}
\description{
Finds the names of the bands of time series in a sits tibble
              or in a metadata cube
              For details see:
\itemize{
 \item{"time series": }{see \code{\link{sits_bands.sits}}}
 \item{"data cube": }{see \code{\link{sits_bands.cube}}}
}
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}

Rolf Simoes, \email{rolf.simoes@inpe.br}
}
