% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_bands.R
\name{sits_bands}
\alias{sits_bands}
\title{Finds the names of the bands of a time series}
\usage{
sits_bands(sits.tb)
}
\arguments{
\item{sits.tb}{a valid sits table}
}
\value{
names      a string vector with the names of the bands
}
\description{
\code{sits_bands} finds the names of the bands of time series in a sits table
}
