% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_table.R
\name{sits_bands}
\alias{sits_bands}
\title{returns the names of the bands of a time series}
\usage{
sits_bands(data.tb)
}
\arguments{
\item{data.tb}{a valid sits table}
}
\value{
names      a string vector with the names of the bands
}
\description{
finds the names of the bands of time series in a sits table
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
