% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_bands.R
\name{sits_bands}
\alias{sits_bands}
\alias{sits_bands.sits}
\alias{sits_bands.sits_cube}
\alias{sits_bands.patterns}
\alias{sits_bands.sits_model}
\title{Get the names of the bands}
\usage{
sits_bands(x)

\method{sits_bands}{sits}(x)

\method{sits_bands}{sits_cube}(x)

\method{sits_bands}{patterns}(x)

\method{sits_bands}{sits_model}(x)
}
\arguments{
\item{x}{Valid sits tibble (time series or a cube)}
}
\value{
A vector with the names of the bands.
}
\description{
Finds the names of the bands of
              a set of time series or of a data cube
}
\examples{
bands <- sits_bands(samples_modis_4bands)
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}

Rolf Simoes, \email{rolf.simoes@inpe.br}
}
