% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_tibble_functions.R
\name{sits_bands}
\alias{sits_bands}
\title{names of the bands of a time series}
\usage{
sits_bands(data.tb)
}
\arguments{
\item{data.tb}{a valid sits table}
}
\value{
result.vec  a string vector with the names of the bands
}
\description{
finds the names of the bands of time series in a sits table
              or sets the names of the bands if a set of values is given
}
\examples{
# Retrieve the set of samples for the Mato Grosso region (provided by EMBRAPA)
data(samples_MT_9classes)
# print the bands
sits_bands(samples_MT_9classes)
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}

Rolf Simoes, \email{rolf.simoes@inpe.br}
}
