% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_bands.R
\name{sits_bands}
\alias{sits_bands}
\alias{sits_bands.sits}
\alias{sits_bands.raster_cube}
\alias{sits_bands.patterns}
\alias{sits_bands.sits_model}
\alias{sits_bands<-}
\alias{sits_bands<-.sits}
\alias{sits_bands<-.raster_cube}
\title{Get the names of the bands}
\usage{
sits_bands(x)

\method{sits_bands}{sits}(x)

\method{sits_bands}{raster_cube}(x)

\method{sits_bands}{patterns}(x)

\method{sits_bands}{sits_model}(x)

sits_bands(x) <- value

\method{sits_bands}{sits}(x) <- value

\method{sits_bands}{raster_cube}(x) <- value
}
\arguments{
\item{x}{Valid sits tibble (time series or a cube)}
}
\value{
A vector with the names of the bands.
}
\description{
Finds the names of the bands of a set of time series or of a data cube
}
\examples{
bands <- sits_bands(point_mt_6bands)

}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}

Rolf Simoes, \email{rolf.simoes@inpe.br}
}
