% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_bayes_postprocess.R
\name{sits_bayes_postprocess}
\alias{sits_bayes_postprocess}
\title{Post-process a classified data raster with bayesian filter}
\usage{
sits_bayes_postprocess(raster_class, window = matrix(1, nrow = 3, ncol =
  3, byrow = TRUE), noise = 100, file)
}
\arguments{
\item{raster_class}{output raster coverage}

\item{window}{(matrix) neighborhood window to compute bayesian smooth.
The central element index (i, j) is given by
i = floor(nrows(window)/2)+1 and j = floor(ncols(window)/2)+1.
Elements '0' are excluded from window.}

\item{noise}{bayesian smoothing parameter}

\item{file}{file to save the post processed raster}
}
\value{
raster_layers.tb  tibble with metadata about the output RasterLayer objects
}
\description{
Takes a set of classified raster layers, whose metadata is
             described by tibble (created by \code{\link[sits]{sits_coverage}}),
             and a noise value, to do a bayesian smoothing process.
}
\author{
Rolf Simoes, \email{rolf.simoes@inpe.br}
}
