% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_bbox.R
\name{sits_bbox}
\alias{sits_bbox}
\title{Get the bounding box of the data}
\usage{
sits_bbox(data, ...)
}
\arguments{
\item{data}{Valid sits tibble (time series or a cube).}

\item{...}{Additional parameters.}

\item{wgs84}{Show the bbox of a data cube in WGS84 (EPSG:4326)
projection.}
}
\value{
named vector with bounding box in WGS84 for time series and
        on the cube projection for a data cube unless wgs84 parameter
        is TRUE.
}
\description{
Obtain a vector of limits (either on lat/long for time series
              or in projection coordinates in the case of cubes)
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}

Rolf Simoes, \email{rolf.simoes@inpe.br}
}
