% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_bbox.R
\name{sits_bbox}
\alias{sits_bbox}
\title{Get the bounding box of the data}
\usage{
sits_bbox(data)
}
\arguments{
\item{data}{Valid sits tibble (time series or a cube).}
}
\value{
named vector with bounding box in WGS 84 for time series and
        on the cube projection for a data cube.
}
\description{
Obtain a vector of limits (either on lat/long for time series
              or in projection coordinates in the case of cubes)
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}

Rolf Simoes, \email{rolf.simoes@inpe.br}
}
