% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_data_generics.R
\name{sits_bbox}
\alias{sits_bbox}
\title{Get the bounding box of the data}
\usage{
sits_bbox(data)
}
\arguments{
\item{data}{Valid sits tibble (time series or a cube).}
}
\value{
named vector with bounding box ("lon_min", "lon_max",
     "lat_min", "lat_max").
}
\description{
Obtain a vector of limits (either on lat/long for time series
              or in projection coordinates in the case of cubes)
\itemize{
 \item{"time series": }{see \code{\link{sits_bbox.sits}}}
 \item{"data cube": }{see \code{\link{sits_bbox.cube}}}
}
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}

Rolf Simoes, \email{rolf.simoes@inpe.br}
}
