% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_bbox.R
\name{sits_bbox}
\alias{sits_bbox}
\alias{sits_bbox.sits}
\alias{sits_bbox.raster_cube}
\alias{sits_bbox.mpc_cube_sentinel-1-grd}
\alias{sits_bbox.tbl_df}
\alias{sits_bbox.default}
\title{Get the bounding box of the data}
\usage{
sits_bbox(data, crs = "EPSG:4326", as_crs = NULL)

\method{sits_bbox}{sits}(data, crs = "EPSG:4326", as_crs = NULL)

\method{sits_bbox}{raster_cube}(data, crs = "EPSG:4326", as_crs = NULL)

\method{sits_bbox}{`mpc_cube_sentinel-1-grd`}(data, crs = "EPSG:4326", as_crs = NULL)

\method{sits_bbox}{tbl_df}(data, crs = "EPSG:4326", as_crs = NULL)

\method{sits_bbox}{default}(data, crs = "EPSG:4326", as_crs = NULL)
}
\arguments{
\item{data}{samples (class "sits") or \code{cube}.}

\item{crs}{CRS of the samples points (single char)}

\item{as_crs}{CRS to project the resulting \code{bbox}.}
}
\value{
A \code{bbox}.
}
\description{
Obtain a vector of limits (either on lat/long for time series
              or in projection coordinates in the case of cubes)
}
\examples{
if (sits_run_examples()) {
    # get the bbox of a set of samples
    sits_bbox(samples_modis_ndvi)
    # get the bbox of a cube in WGS84
    data_dir <- system.file("extdata/raster/mod13q1", package = "sits")
    cube <- sits_cube(
        source = "BDC",
        collection = "MOD13Q1-6",
        data_dir = data_dir
    )
    sits_bbox(cube, as_crs = "EPSG:4326")
}
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}

Rolf Simoes, \email{rolf.simoes@inpe.br}
}
