% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_bbox.R
\name{sits_bbox}
\alias{sits_bbox}
\alias{sits_bbox.sits}
\alias{sits_bbox.sits_cube}
\title{Bounding box (bbox)}
\usage{
sits_bbox(data, wgs84 = FALSE, ...)

\method{sits_bbox}{sits}(data, ...)

\method{sits_bbox}{sits_cube}(data, wgs84 = FALSE, ...)
}
\arguments{
\item{data}{valid sits tibble or sits cube objects (inherited from `tbl`).}

\item{wgs84}{reproject resulting bbox to EPSG:4326
(take effect only for sits data cubes).}

\item{...}{Additional parameters (not implemented).}
}
\value{
return a `bbox` object (inherited from `tbl`) containing
`("xmin", "xmax", "ymin", "ymax", "crs")` variables
}
\description{
Get the bounding box of a given data.

`sits_bbox.sits()`: obtain a bbox of time series limits (in EPSG:4326).

`sits_bbox.sits_cube()`: obtain a bbox of a data cube. If data cube have
more than one crs the result will be converted to EPSG:4326
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}

Rolf Simoes, \email{rolf.simoes@inpe.br}
}
