% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_tibble_functions.R
\name{sits_break}
\alias{sits_break}
\title{Break a data set in segments to match the dates of a set of samples, given an interval}
\usage{
sits_break(data.tb, samples.tb, interval = "12 month")
}
\arguments{
\item{data.tb}{a valid sits tibble}

\item{samples.tb}{the samples used in the classification}

\item{interval}{the interval between the classificationa}
}
\description{
This function aligns an input data set to the dates used by a set of samples
}
\examples{
# Retrieve the set of samples for the Mato Grosso region (provided by EMBRAPA)
samples.tb <- readRDS(system.file("extdata/time_series/embrapa_mt.rds", package = "sits"))
# get a point
point.tb <- readRDS(system.file("extdata/time_series/point.rds", package = "sits"))
# break the point to match the samples (breaks a long time series into intervals)
point2.tb <- sits_break(point.tb, samples.tb)
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
