% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_cluster.R
\name{sits_centroids}
\alias{sits_centroids}
\title{Cluster a set of time series using partitional clustering}
\usage{
sits_centroids(data.tb, n_clusters = 4, perc = 0.1, show = TRUE)
}
\arguments{
\item{data.tb}{a tibble the list of time series to be clustered}

\item{n_clusters}{the number of clusters to be identified}

\item{perc}{the minimum percentage for a cluster to be valid}

\item{show}{(boolean) should the results be shown?}
}
\value{
clusters.tb a SITS tibble with the clusters
}
\description{
Partitional clustering assigns the data to one and only
one cluster out of k total clusters.
First, k centroids are randomly initialized, usually by choosing k objects
from the dataset at random. these are assigned to individual clusters.
The distance between all objects in the data and all centroids
is calculated, and each object is assigned to the cluster of its closest centroid.
A prototyping function is applied to each cluster to update the corresponding centroid.
Then, distances and centroids are updated iteratively until a certain number of iterations have elapsed,
or no object changes clusters anymore (description extracted from dtwclust package)
}
\references{
"dtwclust" package (https://CRAN.R-project.org/package=dtwclust)
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}

Rolf Simoes, \email{rolf.simoes@inpe.br}
}
