% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_centroids.R
\name{sits_centroids}
\alias{sits_centroids}
\title{Cluster a set of time series using partitional clustering (uses dtwclust package)}
\usage{
sits_centroids(data.tb, n_clusters = 6)
}
\arguments{
\item{data.tb}{a tibble the list of time series to be clustered}

\item{n_clusters}{the number of clusters to be identified}
}
\value{
none
}
\description{
\code{sits_centroids} find the k centroids of a set of time series clusters
}
\details{
Partitional clustering assigns the data
to one and only one cluster out of k total clusters. The total number of
desired clusters must be specified beforehand.
Partitional clustering algorithms commonly work in the following way.
First, k centroids are randomly initialized, usually by choosing k objects
from the dataset at random. these are assigned to individual clusters.
The distance between all objects in the data and all centroids
is calculated, and each object is assigned to the cluster of its closest centroid.
A prototyping function is applied to each cluster to update the corresponding centroid.
Then, distances and centroids are updated iteratively until a certain number of iterations have elapsed,
or no object changes clusters anymore (extracted from dtwclust package)
}
\examples{
sits_plot  (data.tb, n_clusters = 6)
}
\seealso{
Other STIS main functions: \code{\link{sits_dendogram}},
  \code{\link{sits_fromJSON}},
  \code{\link{sits_plot_together}},
  \code{\link{sits_plot}}, \code{\link{sits_toJSON}}
}
\keyword{STIS}

