% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_classification.R
\name{sits_classify.raster_cube}
\alias{sits_classify.raster_cube}
\title{Classify a data cube using multicore machines}
\usage{
\method{sits_classify}{raster_cube}(
  data,
  ml_model,
  ...,
  roi = NULL,
  filter = NULL,
  impute_fn = sits_impute_linear(),
  interp_fn = NULL,
  compose_fn = NULL,
  start_date = NULL,
  end_date = NULL,
  memsize = 8,
  multicores = 2,
  output_dir = "./",
  version = "v1"
)
}
\arguments{
\item{data}{data cube}

\item{ml_model}{R model trained by \code{\link[sits]{sits_train}}.}

\item{...}{other parameters to be passed to specific functions}

\item{roi}{a region of interest (see above)}

\item{filter}{smoothing filter to be applied (if desired).}

\item{impute_fn}{impute function to replace NA}

\item{interp_fn}{function to interpolate points from cube to match samples}

\item{compose_fn}{function to compose points from cube to match samples}

\item{start_date}{starting date for the classification}

\item{end_date}{end date for the classification}

\item{memsize}{memory available for classification (in GB).}

\item{multicores}{number of cores to be used for classification.}

\item{output_dir}{directory for output file}

\item{version}{version of the output (for multiple classifications)}
}
\value{
cube with the metadata of a brick of probabilities.
}
\description{
Takes a data cube, whose metadata is
   described by tibble (created by \code{\link[sits]{sits_cube}}),
   a set of samples used for training a classification model,
   a prediction model (created by \code{\link[sits]{sits_train}}),
   and produces a classified set of RasterLayers.

   The "roi" parameter defines a region of interest. It can be
   an sf_object, a shapefile, or a bounding box vector with
   named XY values ("xmin", "xmax", "ymin", "ymax") or
   named lat/long values ("lat_min", "lat_max", "long_min", "long_max")

   The "memsize" and "multicores" parameters define the
   The "multicores" parameter defines the
   number of cores used for processing. The "memsize" parameter  controls
   the amount of memory available for classification.
}
\examples{
\dontrun{

# create a data cube based on files
data_dir <- system.file("extdata/raster/mod13q1", package = "sits")
cube <- sits_cube(
    type = "STACK",
    name = "sinop-2014",
    satellite = "TERRA",
    sensor = "MODIS",
    data_dir = data_dir,
    delim = "_",
    parse_info = c("X1", "X2", "band", "date")
)

# select band "NDVI"
samples <- sits_select(samples_mt_4bands, bands = "NDVI")

# select a random forest model
rfor_model <- sits_train(samples, ml_method = sits_rfor(num_trees = 300))

# classify the raster image
probs_cube <- sits_classify(cube,
    ml_model = rfor_model,
    output_dir = tempdir(),
    memsize = 4, multicores = 2
)

# label the classified image
label_cube <- sits_label_classification(probs_cube, output_dir = tempdir())
}

}
\author{
Rolf Simoes, \email{rolf.simoes@inpe.br}

Gilberto Camara, \email{gilberto.camara@inpe.br}
}
