% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_classification.R
\name{sits_classify}
\alias{sits_classify}
\title{Classify a sits tibble using machine learning models}
\usage{
sits_classify(data.tb = NULL, samples.tb = NULL, ml_model = NULL,
  interval = "12 month", multicores = 1)
}
\arguments{
\item{data.tb}{SITS tibble time series (cleaned)}

\item{samples.tb}{The samples used for training the classification model}

\item{ml_model}{a model trained by \code{\link[sits]{sits_train}}}

\item{interval}{The interval used for classification}

\item{multicores}{Number of threads to process the time series.}
}
\value{
data.tb         SITS tibble with the predicted labels for each input segment
}
\description{
This is the main classification function for time series data
organized as a SITS tibble (one row per time series). It considers that
the machine learning model used to classify the data set has been trained
with samples which match the full dimension of the time series. This means that,
if the sample has four bands with 20 time instances,  the
machine learning model is 80-dimension. Each time instance is considered as
a dimension.

In terms of the SITS package, the sits_classify_raster function assumes that, when
tranining the model, the user has called the \code{\link[sits]{sits_distances_from_data}} function.
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
