% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_TWDTW.R
\name{sits_classify}
\alias{sits_classify}
\title{Classify a sits tibble using TWDTW (using the dtwSat package)}
\usage{
sits_classify(point.tb, patterns.tb, bands, alpha = -0.1, beta = 100,
  theta = 0.5, start_date = as.Date("2000-09-01"),
  end_date = as.Date("2016-08-31"), interval = "12 month")
}
\arguments{
\item{patterns.tb}{a set of known temporal signatures for the chosen classes}

\item{bands}{string - the bands to be used for classification}

\item{alpha}{(double) - the steepness of the logistic function used for temporal weighting}

\item{beta}{(integer) - the midpoint (in days) of the logistic function}

\item{theta}{(double)  - the relative weight of the time distance compared to the dtw distance}

\item{start_date}{date - the starting date of the classification}

\item{end_date}{date - the end date of the classification}

\item{interval}{the period between two classifications}

\item{samples.tb}{a table in SITS format with time series to be classified using TWTDW}
}
\value{
matches       a SITS table with the information on matches for the data
}
\description{
\code{sits_classify} returns a sits table with values only
}
\details{
A sits table has the metadata and data for each time series
<longitude, latitude, start_date, end_date, label, coverage, time_series>

The TWDTW (time-weighted dynamical time warping) is a version of the
Dynamic Time Warping method for land use and land cover mapping using a sequence
of multi-band satellite images. Methods based on dynamic time warping are flexible to
handle irregular sampling and out-of-phase time series, and they have achieved significant
results in time series analysis.
In contrast to standard DTW, the TWDTW method is sensitive to seasonal
changes of natural and cultivated vegetation types. It also considers inter-annual climatic and
seasonal variability.

Reference: Maus V, Camara G, Cartaxo R, Sanchez A, Ramos FM, de Queiroz GR (2016).
A Time-Weighted Dynamic Time Warping Method for Land-Use and Land-Cover Mapping. IEEE
 Journal of Selected Topics in Applied Earth Observations and Remote Sensing, 9(8):3729-3739,
 August 2016. ISSN 1939-1404. doi:10.1109/JSTARS.2016.2517118.
}
