% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_classification.R
\name{sits_classify}
\alias{sits_classify}
\title{Classify a sits tibble using machine learning models}
\usage{
sits_classify(data.tb = NULL, train_samples.tb = NULL,
  ml_method = sits_svm(kernel = "radial", cost = 10, coef0 = 0, tolerance =
  0.001, epsilon = 0.1, cross = 4), dist_method = sits_distances_from_data(),
  interval = "12 month", multicores = 1)
}
\arguments{
\item{data.tb}{SITS tibble time series (cleaned)}

\item{train_samples.tb}{The samples used for training the classification model}

\item{ml_method}{A machine learning method (see \code{\link[sits]{sits_train}})}

\item{dist_method}{The method to obtain the values to be used from training and classification}

\item{interval}{The interval used for classification}

\item{multicores}{Number of threads to process the time series.}
}
\value{
data.tb           SITS tibble with the predicted labels for each input segment
}
\description{
This function classifies a set of time series, given
a set of training samples, an inference model, and an interval.
To perform the classification, users should provide a set of
labelled samples. Each samples should be associated to one spatial location
(latitude/longitude), one time interval and a label.

After defining the training samples, the users need to provide a machine learning model.
Currenly, sits supports the following models:
'svm' (see \code{\link[sits]{sits_svm}}), 'random forest' (see \code{\link[sits]{sits_rfor}}),
'boosting' (see \code{\link[sits]{sits_gbm}}), 'lda' (see \code{\link[sits]{sits_lda}}),
'qda' (see \code{\link[sits]{sits_qda}}), multinomial logit' (see \code{\link[sits]{sits_mlr}}),
'lasso' (see \code{\link[sits]{sits_mlr}}), and 'ridge' (see \code{\link[sits]{sits_mlr}}).

The default is to use an SVM with a radial kernel, but users are encouraged to test
alternatives.
}
\examples{
\donttest{
# read a training data set
# Retrieve the set of samples for the Mato Grosso region (provided by EMBRAPA)
data(samples_MT_ndvi)
# Retrieve a point
data(point_ndvi)
# classify the point
class_ndvi.tb <-  sits_classify (point_ndvi, samples_MT_ndvi)
# plot the classification
sits_plot (class_ndvi.tb)

# Read a point from the WTSS server
point2.tb <- sits_getdata (longitude = -46.4070, latitude = -10.8630)
# select the ndvi
point2.tb <- sits_select (point2.tb, bands = c("ndvi"))
# classify the point
class2.tb <-  sits_classify (point2.tb, samples_MT_ndvi)
# plot the classification
sits_plot (class2.tb)
}

}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
