% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_classification.R
\name{sits_classify}
\alias{sits_classify}
\title{Classify time series or data cube using machine learning models}
\usage{
sits_classify(data, ml_model, ...)
}
\arguments{
\item{data}{Tibble with time series metadata and data.}

\item{ml_model}{Pre-built machine learning model
(see \code{\link[sits]{sits_train}}).}

\item{...}{Other parameters to be passed to specific functions}
}
\value{
Predicted data
}
\description{
This function classifies a set of time series or data cube given
a set of training samples, an inference model, and an interval.
To perform the classification, users should provide a set of
labelled samples. Each samples should be associated to one spatial location
(latitude/longitude), one time interval and a label.
This is a generic function. The following specific functions are available:
\itemize{
 \item{"sits tibble": }{see \code{\link{sits_classify.sits}}}
 \item{"cube": }{see \code{\link{sits_classify.raster_cube}}}
}
SITS supports the following models:
\itemize{
 \item{support vector machines: }         {see \code{\link[sits]{sits_svm}}}
 \item{random forests: }                  {see \code{\link[sits]{sits_rfor}}}
 \item{linear discriminant analysis: }    {see \code{\link[sits]{sits_lda}}}
 \item{quadratic discriminant analysis: } {see \code{\link[sits]{sits_qda}}}
 \item{multinomial logit: }               {see \code{\link[sits]{sits_mlr}}}
 \item{extreme gradient boosting: }       {see \code{\link[sits]{sits_xgboost}}}
 \item{multi-layer perceptrons: }         {see \code{\link[sits]{sits_deeplearning}}}
 \item{1D convolutional neural networks: }{see \code{\link[sits]{sits_FCN}}}
 \item{mixed 1D and MLP networks: }       {see \code{\link[sits]{sits_TempCNN}}}
 \item{1D version of ResNet: }            {see \code{\link[sits]{sits_ResNet}}}
 \item{combined LSTM-FCN model: }         {see \code{\link[sits]{sits_LSTM_FCN}}}
 }
The model should be precomputed using \code{\link[sits]{sits_train}}
and then passed to the "sits_classify" function using parameter "ml_model".
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
