% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_classification.R
\name{sits_classify.sits}
\alias{sits_classify.sits}
\title{Classify a set of time series using machine learning models}
\usage{
\method{sits_classify}{sits}(data, ml_model, ..., filter_fn = NULL, multicores = 2)
}
\arguments{
\item{data}{Tibble with time series metadata and data.}

\item{ml_model}{Pre-built machine learning model
(see \code{\link[sits]{sits_train}}).}

\item{...}{Other parameters to be passed to specific functions}

\item{filter_fn}{Smoothing filter to be applied (if desired).}

\item{multicores}{Number of cores to be used for classification.}
}
\value{
A tibble with the predicted labels for each input segment.
}
\description{
This function classifies a set of time series, given
a set of training samples, an inference model, and an interval.
To perform the classification, users should provide a set of
labelled samples. Each samples should be associated to one spatial location
(latitude/longitude), one time interval and a label.

The model should be precomputed by the user. This model should be
passed to the function using the parameter "ml_model".
}
\examples{

# Retrieve the samples for Mato Grosso
# select band "NDVI"
samples_ndvi <- sits_select(samples_mt_4bands, bands = "NDVI")

# select a random forest model
rfor_model <- sits_train(samples_ndvi,
    ml_method = sits_rfor(num_trees = 300)
)

# classify the point
point_class <- sits_classify(point_ndvi, rfor_model)

}
\author{
Rolf Simoes, \email{rolf.simoes@inpe.br}

Gilberto Camara, \email{gilberto.camara@inpe.br}
}
