% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_classification.R
\name{sits_classify_distances}
\alias{sits_classify_distances}
\title{Classify a sits distances tibble using machine learning models}
\usage{
sits_classify_distances(distances.tb, timeline, dates_index.lst,
  ml_model = NULL, interval = "12 month", multicores = 2)
}
\arguments{
\item{distances.tb}{a tibble with distances}

\item{timeline}{the timeline of the data set}

\item{dates_index.lst}{List with the indexes of the timeline to break the time series for classification}

\item{ml_model}{a model trained by \code{\link[sits]{sits_train}}}

\item{interval}{the period between two classifications}

\item{multicores}{Number of threads to process the time series.}
}
\value{
class.tb        a SITS tibble with the predicted labels for each input segment
}
\description{
Returns a sits table with the results of the ML classifier.
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
