% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_classification.R
\name{sits_classify_raster}
\alias{sits_classify_raster}
\title{Classify a set of spatio-temporal raster bricks using machine learning models}
\usage{
sits_classify_raster(file = NULL, raster.tb, samples.tb, ml_model = NULL,
  interval = "12 month", blocksize = 250000, multicores = 2)
}
\arguments{
\item{file}{a set of file names to store the output (one file per classified year)}

\item{raster.tb}{a tibble with information about a set of space-time raster bricks}

\item{samples.tb}{The samples used for training the classification model}

\item{ml_model}{a model trained by \code{\link[sits]{sits_train}}}

\item{interval}{The interval between two sucessive classification}

\item{blocksize}{Default size of the block (rows * cols) (see function .sits_raster_block_size)}

\item{multicores}{Number of threads to process the time series.}

\item{...}{other parameters to be passed to the distance function}
}
\value{
raster_class.tb a SITS tibble with the metadata for the set of RasterLayers
}
\description{
Takes a set of spatio-temporal raster bricks, whose metadata is
             described by tibble (created by \code{\link[sits]{sits_fromRaster}}),
             the information on classification intervals (created by \code{\link[sits]{sits_class_info}}),
             a prediction model (created by \code{\link[sits]{sits_train}}), and
             a method to extract shape attributes from time_series (used by  \code{\link[sits]{sits_distances_from_data}} ),
             and produces a classified set of RasterLayers. This function is similar to
              \code{\link[sits]{sits_classify}} which is applied to time series stored in a SITS tibble.

The sits_classify_raster function assumes that, when
tranining the model, the user has called the \code{\link[sits]{sits_distances_from_data}} function.
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
