% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_classification.R
\name{sits_classify_raster}
\alias{sits_classify_raster}
\title{Classify a set of spatio-temporal raster bricks using machine learning models}
\usage{
sits_classify_raster(file = NULL, raster.tb, samples.tb,
  ml_method = sits_svm(), dist_method = sits_distances_from_data(),
  adj_fun = function(x) {     x + 3 }, interval = "12 month",
  blocksize = 250000, multicores = 2)
}
\arguments{
\item{file}{a set of file names to store the output (one file per classified year)}

\item{raster.tb}{a tibble with information about a set of space-time raster bricks}

\item{samples.tb}{The samples used for training the classification model}

\item{ml_method}{a model trained by \code{\link[sits]{sits_train}}}

\item{dist_method}{The method to obtain the values to be used from training and classification}

\item{adj_fun}{Function to adjust the values (required for ML classification)}

\item{interval}{The interval between two sucessive classification}

\item{blocksize}{Default size of the block (rows * cols) (see function .sits_raster_block_size)}

\item{multicores}{Number of threads to process the time series.}

\item{...}{other parameters to be passed to the distance function}
}
\value{
raster_class.tb a SITS tibble with the metadata for the set of RasterLayers
}
\description{
Takes a set of spatio-temporal raster bricks, whose metadata is
             described by tibble (created by \code{\link[sits]{sits_fromRaster}}),
             a set of samples used for training a classification model,
             a prediction model (created by \code{\link[sits]{sits_train}}),
             and produces a classified set of RasterLayers. This function is similar to
              \code{\link[sits]{sits_classify}} which is applied to time series stored in a SITS tibble.
}
\examples{
\donttest{
# Retrieve the set of samples for the Mato Grosso region (provided by EMBRAPA)
data(samples_MT_ndvi)

# read a raster file and put it into a vector
files  <- c(system.file ("extdata/raster/mod13q1/sinop-crop-ndvi.tif", package = "sits"))

# define the timeline
data(timeline_mod13q1)
timeline <- lubridate::as_date (timeline_mod13q1$V1)

# create a raster metadata file based on the information about the files
raster.tb <- sits_coverageRaster(product = "MOD13Q1", coverage = "Sinop-Crop",
             timeline = timeline, bands = c("ndvi"), files = files)

# classify the raster file
raster_class.tb <- sits_classify_raster (file = "./raster-class", raster.tb, samples_MT_ndvi,
   ml_method = sits_svm(), blocksize = 300000, multicores = 1)
}

}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
