% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_raster_classification.R
\name{sits_classify_raster}
\alias{sits_classify_raster}
\title{Classify a set of spatio-temporal raster bricks using machine learning models}
\usage{
sits_classify_raster(file = NULL, raster.tb, samples.tb, ml_model = NULL,
  ml_method = sits_svm(), adj_fun = sits_adjust(), interval = "12 month",
  blocksize = 2000, multicores = 2, verbose = FALSE)
}
\arguments{
\item{file}{vector of file names to store the output (one file per classified year)}

\item{raster.tb}{tibble with information about a set of space-time raster bricks}

\item{samples.tb}{tibble with samples used for training the classification model}

\item{ml_model}{an R model trained by \code{\link[sits]{sits_train}}}

\item{ml_method}{an R machine learning method such as SVM, Random Forest or Deep Learning}

\item{adj_fun}{adjustment function to be applied to the data}

\item{interval}{interval between two sucessive classifications, expressed in months}

\item{blocksize}{size of the block to be read to build a block for classification}

\item{multicores}{number of threads to process the time series.}

\item{verbose}{logical: run function in verbose mode? (useful for working with big data sets)}
}
\value{
raster_class.tb tibble with the metadata for the vector of classified RasterLayers
}
\description{
Takes a set of spatio-temporal raster bricks, whose metadata is
             described by tibble (created by \code{\link[sits]{sits_coverage}}),
             a set of samples used for training a classification model,
             a prediction model (created by \code{\link[sits]{sits_train}}),
             and produces a classified set of RasterLayers. This function is similar to
              \code{\link[sits]{sits_classify}} which is applied to time series stored in a SITS tibble.
              There are two parameters for optimizing processing of large data sets. These
              parameters are "blocksize" and "multicores". The "multicores" parameter defines the
              number of cores used for processing. The "blocksize" parameter  controls
              the size of each image block to be read into memory.
              The thumb rule to determine the blockisze is:

              blocksize = (available memory for use) / (20 * bands * times).

              For example, for a server with 60 GB available, and an image with
              2 bands and 500 instances of time, the blocksize can be up to 3000000 (3e+06).
              The factor of 20 is an estimate of how many bytes are used for each pixel,
              since R stores everything in double precision and makes copies of data.
}
\examples{
\donttest{
# Retrieve the set of samples for the Mato Grosso region (provided by EMBRAPA)
data(samples_MT_ndvi)

# read a raster file and put it into a vector
files  <- c(system.file ("extdata/raster/mod13q1/sinop-crop-ndvi.tif", package = "sits"))

# define the timeline
data(timeline_modis_392)

# create a raster metadata file based on the information about the files
#' # create a raster coverage file based on the information about the files
raster.tb <- sits_coverage(service = "RASTER", name  = "Sinop-crop",
             timeline = timeline_modis_392, bands = c("ndvi"), files = files)

# classify the raster file
raster_class.tb <- sits_classify_raster (file = "./raster-class", raster.tb, samples_MT_ndvi,
   ml_method = sits_svm(), blocksize = 250, multicores = 1)
}

}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
