% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_clean.R
\name{sits_clean}
\alias{sits_clean}
\title{Cleans a classified map using a windowed mode}
\usage{
sits_clean(
  cube,
  window_size,
  memsize = 4,
  multicores = 2,
  output_dir,
  version = "v1",
  progress = TRUE
)
}
\arguments{
\item{cube}{A classified data cube}

\item{window_size}{An odd number representing the size of the
sliding window of the modal function.}

\item{memsize}{Memory available for classification (in GB).}

\item{multicores}{Number of cores to be used for classification.}

\item{output_dir}{Directory where files will be saved.}

\item{version}{Version of the output file.}

\item{progress}{Show progress bar?}
}
\value{
A data cube with an classified map cleaned.
}
\description{
Applies a windowed mode function to clean up possible noisy pixels keeping
the most frequently values within the neighborhood.
In a tie, the first value of the vector is considered.
}
\examples{
if (sits_run_examples()) {
rf_model <- sits_train(samples_modis_ndvi, ml_method = sits_rfor)
# create a data cube from local files
data_dir <- system.file("extdata/raster/mod13q1", package = "sits")
cube <- sits_cube(
    source = "BDC",
    collection = "MOD13Q1-6",
    data_dir = data_dir
)
# classify a data cube
probs_cube <- sits_classify(
    data = cube, ml_model = rf_model, output_dir = tempdir()
)
# label the probability cube
label_cube <- sits_label_classification(
    probs_cube, output_dir = tempdir()
)
# apply a mode function in the labelled cube
clean_cube <- sits_clean(
    label_cube, window_size = 5,  output_dir = tempdir()
)
}

}
\author{
Felipe Carvalho, \email{felipe.carvalho@inpe.br}

Gilberto Camara, \email{gilberto.camara@inpe.br}

Rolf Simoes, \email{rolf.simoes@inpe.br}
}
