% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_filters.R
\name{sits_cloud_filter}
\alias{sits_cloud_filter}
\title{Cloud filter}
\usage{
sits_cloud_filter(data.tb, cutoff = -0.25, p = 0, d = 0, q = 3,
  bands_suffix = ".cf", apply_whit = TRUE, lambda_whit = 1)
}
\arguments{
\item{data.tb}{a valid sits tibble containing the "ndvi" band}

\item{cutoff}{a numeric value for the maximum acceptable value of a NDVI difference}

\item{p}{the order (number of time lags) of the autoregressive model,}

\item{d}{the degree of differencing (the number of times the data have had past values subtracted),}

\item{q}{the order of the moving-average model.}

\item{bands_suffix}{the suffix to rename the filtered bands}

\item{apply_whit}{apply the whittaker smoother after filtering}

\item{lambda_whit}{lambda parameter of the whittaker smoother}
}
\value{
result.tb    a sits tibble with same samples and the new bands
}
\description{
This function tries to remove clouds in the satellite image time series
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
