% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_clouds.R
\name{sits_cloud_remove}
\alias{sits_cloud_remove}
\title{Clean data cube to improve quality}
\usage{
sits_cloud_remove(
  cube,
  data_dir,
  name,
  impute_fn = sits_impute_linear(),
  memsize = 8,
  multicores = 2
)
}
\arguments{
\item{cube}{input data cube}

\item{data_dir}{data directory where output data is written}

\item{name}{name of the output data cube}

\item{impute_fn}{imputing function to be applied to replace NA}

\item{memsize}{size of memory}

\item{multicores}{number of cores}
}
\value{
new data cube with interpolated cloud data
}
\description{
Interpolate data over time to fill cloud pixels.
}
\examples{
# define a data cube of CBERS-4 AWFI data
data_dir <- system.file("extdata/raster/cbers", package = "sits")

cbers_022024 <- sits_cube(
    type = "STACK",
    name = "cbers_022024",
    satellite = "CBERS-4",
    sensor = "AWFI",
    resolution = "64m",
    data_dir = data_dir,
    parse_info = c("X1", "X2", "band", "date")
)

cbers_022024_no_clds <- sits_cloud_remove(
    cube = cbers_022024,
    data_dir = tempdir(),
    name = "cbers_022024_no_cld"
)
}
