% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_cluster.R
\name{sits_cluster}
\alias{sits_cluster}
\title{Cluster a set of time series}
\usage{
sits_cluster(data.tb, type = "dendogram", n_clusters = 6)
}
\arguments{
\item{data.tb}{a tibble the list of time series to be clustered}

\item{n_clusters}{the number of clusters to be identified}
}
\value{
none
}
\description{
uses the dtwclust package to cluster time series. s
See "dtwclust" documentation for more details
}
