% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_cluster.R
\name{sits_cluster}
\alias{sits_cluster}
\title{Cluster a set of time series}
\usage{
sits_cluster(data.tb, type = "dendogram", n_clusters = 4)
}
\arguments{
\item{data.tb}{a SITS tibble the list of time series to be clustered}

\item{type}{string - either "dendogram" or "centroids"}

\item{n_clusters}{the number of clusters to be identified}
}
\value{
clusters.tb a SITS tibble with the clusters
}
\description{
\code{sits_cluster} cluster time series with dtw distance
}
\details{
This function uses package "dtwclust" to do time series clustering.
See "dtwclust" documentation for more details
}
\seealso{
Other STIS cluster functions: \code{\link{sits_centroids}},
  \code{\link{sits_dendogram}}
}
