% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_cluster.R
\name{sits_cluster}
\alias{sits_cluster}
\title{Cuts a cluster tree produced by sits_dendrogram}
\usage{
sits_cluster(data.tb, clusters, k = NULL, height = NULL)
}
\arguments{
\item{data.tb}{a tibble with input data of dtwclust.}

\item{clusters}{a cluster structure returned from dtwclust.}

\item{k}{the desired number of clusters}

\item{height}{the height of dendrogram cutting}
}
\value{
result.tb       a SITS table with the clusters or clusters' members
}
\description{
reads a list of clusters provided by the dtwclust
package and produces a sits table.
}
\references{
"dtwclust" package (https://CRAN.R-project.org/package=dtwclust)
}
\author{
Rolf Simoes, \email{rolf.simoes@inpe.br}
}
