% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_cluster.R
\name{sits_cluster}
\alias{sits_cluster}
\title{Cluster a set of satellite image time series}
\usage{
sits_cluster(data.tb, label_groups = NULL, band_groups = NULL,
  method = "dendogram", n_clusters = 4, perc = 0.1, show = TRUE)
}
\arguments{
\item{data.tb}{a SITS tibble the list of time series to be clustered}

\item{label_groups}{a list containing groups of labels to be clusterized. A group of labels may be a label alone (str) or a vector of labels.}

\item{band_groups}{a list containing groups of bands to be clusterized. A group of bands may be a band alone (str) or a vector of bands.}

\item{method}{string - either "dendogram" or "centroids"}

\item{n_clusters}{the number of clusters to be identified}

\item{perc}{the minimum percentage for a cluster to be valid}

\item{show}{(boolean) should the results be shown?}
}
\value{
clusters.tb a SITS tibble with the clusters
}
\description{
This function uses package "dtwclust" to do time series clustering.
There are two options: "dendogram" (hierarchical clustering) and "controids" (positional
clustering)
}
\references{
"dtwclust" package (https://CRAN.R-project.org/package=dtwclust)
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}

Rolf Simoes, \email{rolf.simoes@inpe.br}
}
