% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_cluster.R
\name{sits_cluster_clean}
\alias{sits_cluster_clean}
\title{Cluster cleaner}
\usage{
sits_cluster_clean(data.tb)
}
\arguments{
\item{data.tb}{Tibble with `cluster` column.}
}
\value{
A tibble with all selected samples.
}
\description{
Removes sits tibble samples of labels that are minority in each cluster.
}
\examples{
\donttest{
# Load the "dtwclust" package
# library(dtwclust)
# load a simple data set with two classes
data(cerrado_2classes)
# calculate the dendrogram and the best clusters
cluster.tb <- sits_cluster (cerrado_2classes, bands = c("ndvi", "evi"))
# show clusters samples frequency
sits_cluster_frequency(cluster.tb)
# remove cluster 3 from the samples
clusters_new.tb <- dplyr::filter(cluster.tb, cluster != 3)
# show clusters samples frequency
sits_cluster_frequency(clusters_new.tb)
# clean all remaining clusters
cleaned.tb <- sits_cluster_clean(clusters_new.tb)
# show clusters samples frequency
sits_cluster_frequency(cleaned.tb)
}
}
\author{
Rolf Simoes, \email{rolf.simoes@inpe.br}
}
