% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_cluster.R
\name{sits_cluster_clean}
\alias{sits_cluster_clean}
\title{Cluster cleaner}
\usage{
sits_cluster_clean(samples)
}
\arguments{
\item{samples}{Tibble with `cluster` column.}
}
\value{
A tibble with all selected samples.
}
\description{
Removes labels that are minority in each cluster.
}
\examples{
\dontrun{
# Load the "dtwclust" package
# library(dtwclust)
# calculate the dendrogram and the best clusters
clusters <- sits_cluster_dendro(cerrado_2classes, bands = c("NDVI", "EVI"))
# show clusters samples frequency
sits_cluster_frequency(clusters)
# remove cluster 3 from the samples
clusters_new <- dplyr::filter(clusters, cluster != 3)
# show clusters samples frequency
sits_cluster_frequency(clusters_new)
# clean all remaining clusters
cleaned <- sits_cluster_clean(clusters_new)
# show clusters samples frequency
sits_cluster_frequency(cleaned)
}
}
\author{
Rolf Simoes, \email{rolf.simoes@inpe.br}
}
