% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_cluster.R
\name{sits_cluster_clean}
\alias{sits_cluster_clean}
\title{Cluster cleaner}
\usage{
sits_cluster_clean(data.tb, min_perc = 0.05, method = "intracluster")
}
\arguments{
\item{data.tb}{tibble with `cluster` column.}

\item{min_perc}{minimum percentage of label inside a cluster for the label to remain in cluster.}

\item{method}{string with "intracluster" or "intercluster" or both}
}
\value{
result.tb        tibble with all selected samples
}
\description{
Removes SITS tibble samples of labels that are minority in each cluster.
The function removes samples according to a percentage threshold "min_perc".
If the method "intracluster" is chosen, the "min_perc" parameter
controls the relative percentage of labels inside each cluster. If the number of samples
of a given label inside a cluster are below this limit, then all those label samples are
removed from that cluster. If the method "intercluster" is set, the "min_perc"
parameter is a threshold that controls the minimum
percentage of each label in all clusters. If the percentage of samples of a label in a cluster
is less than this "min_perc", all thise label samples are removed from that cluster.
}
\author{
Rolf Simoes, \email{rolf.simoes@inpe.br}
}
