% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_cluster.R
\name{sits_cluster_cleaner}
\alias{sits_cluster_cleaner}
\title{Cluster cleaner}
\usage{
sits_cluster_cleaner(data.tb, min_clu_perc)
}
\arguments{
\item{data.tb}{a SITS tibble with `cluster` column.}

\item{min_clu_perc}{minimum percentage of representativeness inside a cluster to remain in cluster.}
}
\value{
result.tb       a SITS tibble with all selected samples
}
\description{
Removes SITS tibble samples of labels that are minority in each cluster.
This function needs as input a SITS tibble with `cluster` column.
}
\author{
Rolf Simoes, \email{rolf.simoes@inpe.br}
}
