% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_cluster.R
\name{sits_cluster_dendro}
\alias{sits_cluster_dendro}
\title{Clusters a set of time series
       using aglomerative hierarchical clustering}
\usage{
sits_cluster_dendro(
  samples = NULL,
  bands = NULL,
  dist_method = "dtw_basic",
  linkage = "ward.D2",
  k = NULL,
  colors = "RdYlGn",
  .plot = TRUE,
  ...
)
}
\arguments{
\item{samples}{A tibble with input set of time series}

\item{bands}{Bands to be used in the clustering}

\item{dist_method}{String with one of the supported distances.}

\item{linkage}{String with agglomeration method to be used.
Can be any `hclust` method (see `hclust`).
Default is 'ward.D2'.}

\item{k}{Desired number of clusters (overrides default value)}

\item{colors}{Color scheme as per `sits_color_name` function.}

\item{.plot}{Plot the dendrogram?}

\item{...}{Additional parameters to be passed
to dtwclust::tsclust() function.}
}
\value{
A tibble with the clusters or clusters' members.
}
\description{
Takes a SITS tibble and produces a sits tibble
with an added "cluster" column.
This is done in several steps:
1. Calculation of the dendrogram;
2. Get validity index for best cluster using the adjusted Rand Index;
3. Cut the dendrogram using the chosen validity index.
}
\examples{
\dontrun{
# Load the "dtwclust" package
# library(dtwclust)
# load a simple data set with two classes
data(cerrado_2classes)
# calculate the dendrogram and the best clusters
clusters <- sits_cluster_dendro(cerrado_2classes, bands = c("NDVI", "EVI"))
}
}
\references{
"dtwclust" package (https://CRAN.R-project.org/package=dtwclust)
}
\author{
Rolf Simoes, \email{rolf.simoes@inpe.br}
}
