% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_cluster.R
\name{sits_cluster_frequency}
\alias{sits_cluster_frequency}
\title{Cluster contigency table}
\usage{
sits_cluster_frequency(data.tb)
}
\arguments{
\item{data.tb}{A tibble with `cluster` column.}
}
\value{
A matrix containing all frequencies of labels in clusters.
}
\description{
Computes the contingency table between labels and clusters.
This function needs as input a sits tibble with `cluster` column.
}
\examples{
\donttest{
# Load the "dtwclust" package
library(dtwclust)
# load a simple data set with two classes
data(cerrado_2classes)
# create clusters by cutting a dendrogram
clusters.tb <- sits_cluster(cerrado_2classes, bands = c("ndvi", "evi"))
# show clusters samples frequency
sits_cluster_frequency(clusters.tb)
}
}
\author{
Rolf Simoes, \email{rolf.simoes@inpe.br}
}
