% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_cluster.R
\name{sits_cluster_frequency}
\alias{sits_cluster_frequency}
\title{Cluster frequency}
\usage{
sits_cluster_frequency(data.tb)
}
\arguments{
\item{data.tb}{a SITS tibble with `cluster` column.}
}
\value{
result.mtx      matrix containing all frequencies of labels in clusters
}
\description{
Computes the frequency of labels in each cluster.
This function needs as input a SITS tibble with `cluster` column.
}
\author{
Rolf Simoes, \email{rolf.simoes@inpe.br}
}
