% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_cluster.R
\name{sits_cluster_frequency}
\alias{sits_cluster_frequency}
\title{Show label frequency in each cluster produced by dendrogram analysis}
\usage{
sits_cluster_frequency(samples)
}
\arguments{
\item{samples}{Tibble with input set of time series with additional
cluster information produced
by \code{link[sits]{sits_cluster_dendro}}.}
}
\value{
A matrix containing frequencies
                       of labels in clusters.
}
\description{
Show label frequency in each cluster produced by dendrogram analysis
}
\examples{
if (sits_run_examples()) {
    clusters <- sits_cluster_dendro(cerrado_2classes)
    freq <- sits_cluster_frequency(clusters)
    freq
}
}
\author{
Rolf Simoes, \email{rolf.simoes@inpe.br}
}
