% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_kohonen.R
\name{sits_cluster_som}
\alias{sits_cluster_som}
\title{Clustering a set of satellite image time series using SOM}
\usage{
sits_cluster_som(data, grid_xdim = 10, grid_ydim = 10, alpha = 1,
  distance = "euclidean", iterations = 50, prob_label_change = 0.8,
  min_cluster_prob = 0.6)
}
\arguments{
\item{data}{A tibble with samples to be clustered.}

\item{grid_xdim}{X dimension of the SOM grid (default = 25).}

\item{grid_ydim}{Y dimension of the SOM grid.}

\item{alpha}{Starting learning rate, which decreases according to number of iterations.}

\item{distance}{The similarity measure (distance).}

\item{iterations}{The number of time to run the SOM cluster.}

\item{prob_label_change}{Threshold of probability to change the label}

\item{min_cluster_prob}{Minimum probability for a good sample to belong to a cluster}
}
\value{
Returns a sits tibble with a new subset of samples and a new column
presenting the probability of each sample belongs to a class described in column label.
}
\description{
This function uses self-organized maps to find clusters in
satellite image time series for quality control of the  samples.
It calls the function \code{\link[sits]{sits_som_map}} to generate the som map and
\code{\link[sits]{sits_som_clean_samples}} to produce a cleaned set of samples.
The parameters "grid_xdim", "grid_ydim", "rlen", "distance", "alpha", and
"iterations are used by \code{\link[sits]{sits_som_map}} to control how the Kohonen map
is generated. The parameters "prob_label_change" and "min_cluster_prob" control how the
good quality samples are selected, based on the Kohonen map.
}
\examples{
\donttest{
# Get a new subset of samples evaluated by SOM clustering methods
new_samples.tb <- sits_cluster_som(prodes_226_064)
}
}
\references{
`kohonen` package (https://CRAN.R-project.org/package=kohonen)
}
\author{
Lorena Alves, \email{lorena.santos@inpe.br}

Karine Ferreira. \email{karine.ferreira@inpe.br}
}
