% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_cluster.R
\name{sits_cluster_validity}
\alias{sits_cluster_validity}
\title{Cluster validity indices}
\usage{
sits_cluster_validity(data.tb, type = "valid")
}
\arguments{
\item{data.tb}{a SITS tibble with `cluster` column.}

\item{type}{character vector indicating which indices are to be computed. (Default "valid")}
}
\value{
result.vec      vectors with chosen CVIs
}
\description{
Compute different cluster validity indices. This function needs
as input a SITS tibble with `cluster` column. It is a front-end to
`dtwclust::cvi` function. Please refer to the documentation in that package for more details.
}
\references{
"dtwclust" package (https://CRAN.R-project.org/package=dtwclust)
}
\author{
Rolf Simoes, \email{rolf.simoes@inpe.br}
}
