% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_colors.R
\name{sits_color_names}
\alias{sits_color_names}
\title{Function to show color names in SITS}
\usage{
sits_color_names()
}
\value{
A data.table with color names and HEX values
}
\description{
Returns a color palette for plotting

Shows the default SITS colors
}
\examples{
if (sits_run_examples()) {
    # show the names of the colors supported by SITS
    sits_color_names()
}
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
