% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_colors.R
\name{sits_colors}
\alias{sits_colors}
\title{Brewer color schemes}
\usage{
sits_colors(brewer, n = 3)
}
\arguments{
\item{brewer}{name of the brewer color set}

\item{n}{number of desirable colors (default 3)}
}
\value{
list     list of rgb colors
}
\description{
Number of colors available in brewer color schemes
}
