% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_colors.R
\name{sits_colors}
\alias{sits_colors}
\title{Function to show colors in SITS}
\usage{
sits_colors(name)
}
\value{
No return, called for side effects
}
\description{
Returns a color palette for plotting

Shows the default SITS colors
}
\examples{
if (sits_run_examples()) {
    # show the names of the colors supported by SITS
    sits_colors()
}
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
