% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_colors.R
\name{sits_colors}
\alias{sits_colors}
\title{Function to retrieve sits color table}
\usage{
sits_colors()
}
\value{
A data.table with color names and HEX values
}
\description{
Returns a color table
}
\examples{
if (sits_run_examples()) {
    # show the names of the colors supported by SITS
    sits_colors()
}
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
