% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_colors.R
\name{sits_colors}
\alias{sits_colors}
\title{Function to retrieve sits color table}
\usage{
sits_colors(legend = NULL)
}
\arguments{
\item{legend}{One of the accepted legends in sits}
}
\value{
A tibble with color names and values
}
\description{
Returns a color table
}
\examples{
if (sits_run_examples()) {
    # return the names of all colors supported by SITS
    sits_colors()
}
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
