% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_colors.R
\name{sits_colors_reset}
\alias{sits_colors_reset}
\title{Function to reset sits color table}
\usage{
sits_colors_reset()
}
\value{
No return, called for side effects
}
\description{
Resets the color table
}
\examples{
if (sits_run_examples()) {
    # reset the default colors supported by SITS
    sits_colors_reset()
}
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
