% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_colors.R
\name{sits_colors_reset}
\alias{sits_colors_reset}
\title{Function to reset sits color table}
\usage{
sits_colors_reset()
}
\value{
The original sits color table
}
\description{
Resets the color table
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
