% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_colors.R
\name{sits_colors_set}
\alias{sits_colors_set}
\title{Function to set sits color table}
\usage{
sits_colors_set(color_tb)
}
\arguments{
\item{color_tb}{New color table}
}
\value{
A modified sits color table
}
\description{
Sets a color table
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
