% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_colors.R
\name{sits_colors_show}
\alias{sits_colors_show}
\title{Show one SITS color palette}
\usage{
sits_colors_show(name, n_colors = 32)
}
\arguments{
\item{name}{Name of sits color palette}

\item{n_colors}{Number of colors to be displayed}
}
\value{
Called for side effects
}
\description{
Shows a barplot with the requested color palette
}
\examples{
if (sits_run_examples()) {
    # show the names of the colors supported by SITS
    sits_color_names()
    # show a color palette
    sits_colors_show("Blues", n_colors = 16)
}
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
