% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_colors.R
\name{sits_colors_show}
\alias{sits_colors_show}
\title{Function to show colors in SITS}
\usage{
sits_colors_show(legend = NULL, font_family = "roboto")
}
\arguments{
\item{legend}{One of the accepted legends in sits}

\item{font_family}{A font family loaded in SITS}
}
\value{
no return, called for side effects
}
\description{
Shows the default SITS colors
}
\examples{
if (sits_run_examples()) {
    # show the colors supported by SITS
    sits_colors_show()
}
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
