% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_accuracy.R
\name{sits_conf_matrix}
\alias{sits_conf_matrix}
\title{Assessment of the accuracy of classification based on a confusion matrix}
\usage{
sits_conf_matrix(class.tb, conv.lst = NULL, pred_sans_ext = FALSE)
}
\arguments{
\item{class.tb}{A sits tibble containing a set of classified samples whose labels are known}

\item{conv.lst}{A list conversion list of labels. If NULL no conversion is done.}

\item{pred_sans_ext}{(boolean) remove all label extension (i.e. every string after last '.' character) from predictors before compute assesment.}
}
\value{
caret_assess   a confusion matrix assessment produced by the caret package
}
\description{
Evaluates the confusion matrix based on "reference" and "predicted" values
provided in a SITS tibble that has been classified. This function takes two kinds of input:
(a) The output of the \code{\link[sits]{sits_classify}} function (a tibble with a list of predicted values)
(b) The output of the \code{\link[sits]{sits_kfold_validate}} function (a tibble with two columns - predicted and reference)
This function returns the Overall Accuracy, User's Accuracy,
Producer's Accuracy, error matrix (confusion matrix), and Kappa value.
}
\examples{
# read a tibble with 400 samples of Cerrado and 346 samples of Pasture
cerrado2.tb <- readRDS(system.file("extdata/time_series/cerrado_2classes.rds", package = "sits"))
# perform a 2 fold validation of this sample file
pred_ref.tb <-  sits_kfold_validate(cerrado2.tb, folds = 2)
# calculate and print the confusion matrix
conf.mx <- sits_conf_matrix(pred_ref.tb)


}
\author{
Rolf Simoes, \email{rolf.simoes@inpe.br}

Gilberto Camara, \email{gilberto.camara@inpe.br}
}
