% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_accuracy.R
\name{sits_conf_matrix}
\alias{sits_conf_matrix}
\title{Assessment of the accuracy of classification based on a confusion matrix}
\usage{
sits_conf_matrix(class.tb, conv.lst = NULL, pred_sans_ext = FALSE)
}
\arguments{
\item{class.tb}{A sits tibble containing a set of classified samples whose labels are known}

\item{conv.lst}{A list conversion list of labels. If NULL no conversion is done.}

\item{pred_sans_ext}{(Boolean) remove all label extension (i.e. every string after last '.' character) from predictors before compute assesment.}
}
\value{
caret_assess  a confusion matrix assessment produced by the caret package
}
\description{
Evaluates the confusion matrix based on "reference" and "predicted" values
This function returns the Overall Accuracy, User's Accuracy,
Producer's Accuracy, error matrix (confusion matrix), and Kappa value.
}
\author{
Rolf Simoes, \email{rolf.simoes@inpe.br}

Gilberto Camara, \email{gilberto.camara@inpe.br}
}
