% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_config.R
\name{sits_config}
\alias{sits_config}
\title{Reads a configuration file and loads it in the main environment}
\usage{
sits_config()
}
\value{
config_sits  A list with the configuration parameters used by SITS
}
\description{
Reads a user-specified configuration file, located in a "config.yml" file
in the working directory. If this file is not found, reads a default package configuration file.
By default, the SITS configuration file "config.yml" is located at the directory "extdata" of the
package. The configuration file is an YAML file that should provide at least the following parameters:

default:
   ts_servers     :
       - "WTSS"
       - "SATVEG"
   WTSS_server    : "http://www.dpi.inpe.br/tws/wtss"
   SATVEG_server  : "https://www.satveg.cnptia.embrapa.br"
   SATVEG_account : "/satvegws/ws/perfil/ZW46IXzr4pRzJlX/"

To see the contents of the configuration file, please use \code{\link[sits]{sits_show_config}}.
}
\examples{
config_sits <- sits_config()
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
