% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_config.R
\name{sits_config}
\alias{sits_config}
\title{Reads a configuration file and loads it in the main environment}
\usage{
sits_config()
}
\value{
A list with the configuration parameters used by sits.
}
\description{
Reads a user-specified configuration file,
located in a "config.yml" file in the working directory.
If this file is not found, reads a default package configuration file.
By default, the sits configuration file "config.yml" is located at
the directory "extdata" of the package.

Users can provide additional configuration files, by specifying the
location of their file in the environmental variable
SITS_USER_CONFIG_FILE

To see the contents of the configuration file,
please use \code{\link[sits]{sits_config_show}}.
}
\examples{
# create configuration file
config_sits <- sits_config()
# show configuration file
sits_config_show()
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
