% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_config.R
\name{sits_config_satveg_bands}
\alias{sits_config_satveg_bands}
\alias{.sits_config_satveg_bands}
\title{Retrieve the bands associated to SATVEG}
\usage{
.sits_config_satveg_bands(name)
}
\arguments{
\item{name}{Name of SATVEG cube}
}
\description{
Retrieve the cubes associated a service.
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
