% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_config.R
\name{sits_config_show}
\alias{sits_config_show}
\title{Shows the contents of the sits configuration file}
\usage{
sits_config_show()
}
\value{
List with the configuration parameters used by sits.
}
\description{
Displays the contents of sits configuration file. For more details
on how to set the configuration file, please use \code{\link[sits]{sits_config}}.
}
\examples{
sits_config_show()
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
