% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_config.R
\name{sits_config_show}
\alias{sits_config_show}
\title{Show current sits configuration}
\usage{
sits_config_show(source = NULL, collection = NULL, colors = FALSE)
}
\arguments{
\item{source}{Data source to be shown in detail.}

\item{collection}{Collection key entry to be shown in detail.}

\item{colors}{Show colors?}
}
\value{
A \code{list} containing the respective
configuration printed in the console.
}
\description{
Prints the current sits configuration options.
To show specific configuration options for
a source, a collection, or a palette, users can inform the corresponding
keys to \code{source}, \code{collection}, and \code{colors} parameters.
}
\examples{
sits_config_show()
}
