% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_config.R
\name{sits_config_show}
\alias{sits_config_show}
\title{Show current sits configuration}
\usage{
sits_config_show(source = NULL, collection = NULL)
}
\arguments{
\item{source}{Data source (character vector).}

\item{collection}{Collection (character vector).}
}
\value{
No return value, called for side effects.
}
\description{
Prints the current sits configuration options.
To show specific configuration options for
a source, a collection, or a palette, users can inform the corresponding
keys to \code{source} and \code{collection}.
}
\examples{
sits_config_show(source = "BDC")
sits_config_show(source = "BDC", collection = "CBERS-WFI-16D")
}
