% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_config.R
\name{sits_config_user_file}
\alias{sits_config_user_file}
\title{List the cloud collections supported by sits}
\usage{
sits_config_user_file(file_path, overwrite = FALSE)
}
\arguments{
\item{file_path}{file to store the user configuration file}

\item{overwrite}{replace current configurarion file?}
}
\value{
Called for side effects
}
\description{
Creates a user configuration file.
}
\examples{
user_file <- paste0(tempdir(), "/my_config_file.yml")
sits_config_user_file(user_file)
}
