% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_contrastive_training.R
\name{sits_contrastive_network}
\alias{sits_contrastive_network}
\title{Contrastive neural net pre-training for sits}
\usage{
sits_contrastive_network(
  samples = NULL,
  margin = 0.001,
  triplet_smp_method = "random",
  num_triplets = NULL,
  encoder_model = "tempcnn",
  epochs = 150L,
  batch_size = 128L,
  validation_split = 0.2,
  optimizer = torch::optim_adamw,
  opt_hparams = list(lr = 5e-04, eps = 1e-08, weight_decay = 1e-06),
  lr_decay_epochs = 1,
  lr_decay_rate = 0.95,
  patience = 20,
  min_delta = 0.01,
  bands_prefix = "EMB",
  verbose = FALSE,
  seed = 10L
)
}
\arguments{
\item{margin}{Margin for triplet loss.}

\item{epochs}{Number of training epochs.}

\item{batch_size}{Batch size for training.}

\item{verbose}{Whether to print training progress.}

\item{lr}{Learning rate.}

\item{log_every}{Period of epochs in which the model logs partial results}
}
\value{
A function for \code{\link[sits]{sits_pre_train}} (to be used as the \code{deep_method} parameter).
}
\description{
Self-supervised pre-training using contrastive triplet loss and torch encoder.
}
\examples{
if (sits_run_examples()) {
    model <- sits_pre_train(samples_modis_ndvi, sits_contrastive_network())
}

}
