% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_coverage.R
\name{sits_coverage}
\alias{sits_coverage}
\title{Provides information about one coverage used to retrieve data}
\usage{
sits_coverage(service = "WTSS", product = "MOD13Q1",
  coverage = "mod13q1_512", timeline = NULL, bands = NULL,
  file_names = NA)
}
\arguments{
\item{service}{the name of the time series service}

\item{product}{the name of the product}

\item{coverage}{the name of the coverage}

\item{timeline}{timeline of the coverage}

\item{bands}{list of bands (for raster data)}

\item{file_names}{file names (for raster data)}
}
\description{
uses the configuration file to print information and save metadata about a
chosen coverage:
 bands          - the information about the bands of the data to be retrieved from the WTSS
 start_date     - the start date for the time series data in the coverage
 end_date       - the end date for the time series data in the coverage
 xres           - spatial resolution (x dimension)
 yres           - spatial resolution (y dimension)
 start_date     - initial date of the coverage time series
 end_date       - final date of the coverage time series
 xmin           - spatial extent (xmin)
 ymin           - spatial extent (ymin)
 xmax           - spatial extent (xmax)
 ymax           - spatial extent (ymin)
 scale_factor   - scale factor to covert bands to a [0..1] scale
 crs            - Projection crs
}
\examples{
# Retrieve information about a WTSS coverage
coverage.tb <- sits_coverage(service = "WTSS", product = "MOD13Q1", coverage = "mod13q1_512")
}
