% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_coverage.R
\name{sits_coverage}
\alias{sits_coverage}
\title{Provides information about one coverage used to retrieve data}
\usage{
sits_coverage(service = "RASTER", name, timeline = NULL,
  bands = NULL, missing_values = NULL, scale_factors = NULL,
  minimum_values = NULL, maximum_values = NULL, files = NA,
  tiles = NULL, from = NULL, to = NULL)
}
\arguments{
\item{service}{Name of the time series service.}

\item{name}{Name of the coverage.}

\item{timeline}{Vector with the timeline of the coverage.}

\item{bands}{Vector of bands.}

\item{missing_values}{Vector of missing values for each band.}

\item{scale_factors}{Vector with the scale factor for each band.}

\item{minimum_values}{Vector of minimum values for each band.}

\item{maximum_values}{Vector of maximum values for each band.}

\item{files}{Vector of file names for each band (only for raster data).}

\item{tiles}{A string with tile name or \code{sfc} object to filter tiles from a cube.}

\item{from}{A date value to filter cube's layers by date.}

\item{to}{A date value to filter cube's layers by date.}
}
\description{
Defines a coverage to retrieve data. As default, it uses the metadata about a
chosen coverage stored in an yml configuration file.
\itemize{
\item{service: }{Name of time series service that provides the coverage (e.g., "WTSS", "SATVEG", "RASTER").}
\item{name: }{Name of the coverage (must be unique).}
\item{bands: }{Vector of bands.}
\item{scale_factor: }{Vector of scale factors.}
\item{missing_values: }{Vector of missing values.}
\item{minimum_values: }{Vector of minimum values.}
\item{maximum_values: }{Vector of maximum values.}
\item{timeline: }{The timelines of the coverage (more than one if data has been classified).}
\item{xmin: }{Spatial extent (xmin).}
\item{ymin: }{Spatial extent (ymin).}
\item{xmax: }{Spatial extent (xmax).}
\item{ymax: }{Spatial extent (ymin).}
\item{xres: }{Spatial resolution (x dimension).}
\item{yres: }{Spatial resolution (y dimension).}
\item{crs: }{Projection CRS.}
\item{files: }{Files associated with the coverage (in case of raster data).}
}
}
\examples{
\donttest{
# Example 1. Retrieve information about a WTSS coverage
coverage.tb <- sits_coverage(service = "WTSS-INPE", name = "MOD13Q1")

# Example 2. Create a raster coverage with metadata
# read a raster file and put it into a vector
files <- c(system.file("extdata/raster/mod13q1/sinop-crop-ndvi.tif", package = "sits"))

# create a raster coverage file based on the information about the files
raster.tb <- sits_coverage(service = "RASTER", name  = "Sinop-crop",
             timeline = timeline_modis_392, bands = "ndvi", files = files)
}
}
\seealso{
To see the available values for the parameters above use \code{\link{sits_services}}, \code{\link{sits_config}} or \code{\link{sits_show_config}}.
}
