% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_coverage.R
\name{sits_coverage}
\alias{sits_coverage}
\title{Provides information about one coverage used to retrieve data}
\usage{
sits_coverage(service = "WTSS", product = "MOD13Q1",
  coverage = "mod13q1_512", timeline = NULL, bands = NULL, files = NA)
}
\arguments{
\item{service}{the name of the time series service}

\item{product}{the name of the product}

\item{coverage}{the name of the coverage}

\item{timeline}{timeline of the coverage}

\item{bands}{list of bands (for raster data)}

\item{files}{file names (for raster data)}
}
\description{
uses the configuration file to print information and save metadata about a
chosen coverage:
 r_objs.lst     - list of R objects associated with the coverage
 coverage       = name of the coverage (must be unique)
 service        - name of time series service that provides the coverage (e.g., "WTSS", "SATVEG", "RASTER")
 product        - name of the product associated with coverage (e.g., "MOD13Q1")
 band_info      - tibble with information about the bands (name, scale_factor and missing_value)
 start_date     - the start date for the time series data in the coverage
 end_date       - the end date for the time series data in the coverage
 timeline       - the timeline of the coverage
 xmin           - spatial extent (xmin)
 ymin           - spatial extent (ymin)
 xmax           - spatial extent (xmax)
 ymax           - spatial extent (ymin)
 xres           - spatial resolution (x dimension)
 yres           - spatial resolution (y dimension)
 crs            - Projection crs
 files          - Files associated with the coverage (in the case of service == "RASTER")
}
\examples{
# Retrieve information about a WTSS coverage
coverage.tb <- sits_coverage(service = "WTSS", product = "MOD13Q1", coverage = "mod13q1_512")
}
