% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_coverage.R
\name{sits_coverageSATVEG}
\alias{sits_coverageSATVEG}
\title{Retrieve a coverage name from the SATVEG service}
\usage{
sits_coverageSATVEG(product = "MOD13Q1", coverage = "terra",
  timeline = NULL)
}
\arguments{
\item{product}{the SATVEG product}

\item{coverage}{the name of the coverage ("terra", "aqua", "comb")}

\item{timeline}{the timeline for the SATVEG product (optional)}
}
\value{
coverage.tb       metadata about the coverage
}
\description{
Retrieves a coverage name based on the capabilities of the SATVEG service
}
\examples{
coverage.tb <- sits_coverageSATVEG(product = "MOD13Q1", coverage = "terra")
}
\author{
Julio Esquerdo, \email{julio.esquerdo@embrapa.br}

Gilberto Camara, \email{gilberto.camara@inpe.br}
}
