% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_WTSS.R
\name{sits_coverageWTSS}
\alias{sits_coverageWTSS}
\title{Provides information about one coverage of the WTSS service}
\usage{
sits_coverageWTSS(coverage = NULL, .show = TRUE)
}
\arguments{
\item{coverage}{the name of the coverage}

\item{.show}{show information about the coverage (Default: TRUE)}
}
\description{
uses the WTSS services to print information and save metadata about a
chosen coverage:
 bands          - the information about the bands of the data to be retrieved from the WTSS
 start_date     - the start date for the time series data in the coverage
 end_date       - the end date for the time series data in the coverage
 xres           - spatial resolution (x dimension)
 yres           - spatial resolution (y dimension)
 start_date     - initial date of the coverage time series
 end_date       - final date of the coverage time series
 xmin           - spatial extent (xmin)
 ymin           - spatial extent (ymin)
 xmax           - spatial extent (xmax)
 ymax           - spatial extent (ymin)
 scale_factor   - scale factor to covert bands to a [0..1] scale
 crs            - Projection crs
}
\examples{
# Retrieve information about a WTSS coverage
coverage.tb <- sits_coverageWTSS("mod13q1_512")
}
