% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_WTSS.R
\name{sits_coverageWTSS}
\alias{sits_coverageWTSS}
\title{Provides information about one coverage of the WTSS service}
\usage{
sits_coverageWTSS(URL = "http://www.dpi.inpe.br/tws/wtss", coverage = NULL)
}
\arguments{
\item{URL}{the URL for the WTSS time series service}

\item{coverage}{the name of the coverage}
}
\description{
\code{sits_coverageWTSS} uses the WTSS services to print information about a
chosen coverage:
}
\details{
start_date     - the start date for the time series data in the coverage
 end_date       - the end date for the time series data in the coverage
 res_x          - spatial resolution (x dimension)
 res_y          - spatial resolution (y dimension)
 start_date     - initial date of the coverage time series
 end_date       - final date of the coverage time series
 x_min          - spatial extent (xmin)
 y_min          - spatial extent (ymin)
 x_max          - spatial extent (xmax)
 y_max          - spatial extent (ymin)
 bands          - the bands of the data to be retrieved from the WTSS
}
