% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_assessment.R
\name{sits_create_partitions}
\alias{sits_create_partitions}
\alias{.sits_create_partitions}
\title{Create partitions of a data set}
\usage{
.sits_create_partitions(data.tb, times, frac)
}
\arguments{
\item{data.tb}{a SITS table to be partitioned}

\item{times}{number of iterations}

\item{frac}{fraction of original data to be extracted. Value must be between 0 and 1.}
}
\description{
Create a list of partitions of a SITS table, based on a percentage and
a number of iterations
}
\author{
Rolf Simoes, \email{rolf.simoes@inpe.br}
}
