% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_table.R
\name{sits_cross}
\alias{sits_cross}
\title{Cross join bands of two satellite image time series}
\usage{
sits_cross(sits1.tb, sits2.tb)
}
\arguments{
\item{sits1.tb}{the first SITS table in wich entries will be crossed with ts2 entries}

\item{sits2.tb}{the second SITS table entries}
}
\value{
new.tb    a cross-joined SITS tibble with a nested set of time series
}
\description{
\code{sits_cross} Cross-join two SITS tables with the same spatio-temporal references
}
\details{
#' To cross two series' bands, we consider that they contain DIFFERENT
attributes (bands) but refer to the same coverage. We DON'T make any assumptions about their spatio-temporal location.
So be careful with this function as it will cross join every time series' bands between two SITS tables entries.

This function is useful to create different bands of clusters centroids time series.
For example, one may want to put evi and ndvi bands centroids together in an cross joined fashion in order to
generate patterns combinations.0
}
\seealso{
Other SITS table functions: \code{\link{sits_merge}}
}
\keyword{SITS}
