% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_cube.R
\name{sits_cube.bdc_cube}
\alias{sits_cube.bdc_cube}
\title{Defines a data cube for a BDC TILE}
\usage{
\method{sits_cube}{bdc_cube}(
  type = "BDC_TILE",
  ...,
  name = NULL,
  satellite = NULL,
  sensor = NULL,
  bands = NULL,
  cube = NULL,
  tile = NULL,
  version = "v001",
  data_access = "web",
  start_date = NULL,
  end_date = NULL,
  .local = NULL,
  .web = NULL,
  .cloud_band = FALSE
)
}
\arguments{
\item{type}{Type of cube}

\item{...}{Other parameters to be passed for specific types}

\item{name}{Name of the output data cube.}

\item{satellite}{Name of satellite}

\item{sensor}{Name of sensor}

\item{bands}{Vector of bands.}

\item{cube}{Name of the input data cube (or image collection)}

\item{tile}{Name of the tile}

\item{version}{Version of the cube}

\item{data_access}{Type of access (local or web)}

\item{start_date}{Starting date of the cube}

\item{end_date}{Ending date of the cube}

\item{.local}{Directory for local access to the input cube (optional)}

\item{.web}{Directory for web access to the input cube (optional)}

\item{.cloud_band}{Include cloud band? (TRUE/FALSE)}
}
\value{
A brick data cube
}
\description{
Defines a cube to retrieve data from the Brazil Data Cube (BDC). The retrieval
             is based on tiles of a given cube. Allows local or web access. For
             local access, the user should be logged in the BDC.
             For more on BDC, please see http://brazildatacube.dpi.inpe.br/
}
\examples{
\donttest{

# create a raster cube file based on the information about the files
cbers_bdc_tile <- sits_cube(type       = "BDC_TILE",
                            name       = "022024",
                            satellite  = "CBERS-4",
                            sensor     = "AWFI",
                            cube       = "CB4_64_16D_STK",
                            tile       = "022024",
                            version    = "v001",
                            data_access = "web",
                            bands       = c("NDVI", "EVI"),
                            start_date  = as.Date("2018-08-29"),
                            end_date    = as.Date("2019-08-13"))
}

}
