% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_cube.R
\name{sits_cube.bdc_cube}
\alias{sits_cube.bdc_cube}
\title{Defines a data cube for a BDC STAC}
\usage{
\method{sits_cube}{bdc_cube}(
  type = "BDC",
  name = "bdc_cube",
  ...,
  url = NULL,
  collection = NULL,
  tiles = NULL,
  bands = NULL,
  roi = NULL,
  start_date = NULL,
  end_date = NULL
)
}
\arguments{
\item{type}{Type of cube.}

\item{name}{Name of the output data cube.}

\item{...}{Other parameters to be passed for specific types.}

\item{url}{URL for the BDC catalog (mandatory).}

\item{collection}{BDC collection to be searched (mandatory).}

\item{tiles}{Tile names to be searched (optional).}

\item{bands}{Bands names to be filtered (optional).}

\item{roi}{Region of interest (optional), expressed either as
an \code{sfc} or \code{sf} object from sf package, a
a GeoJSON following the rules from RFC 7946, or a
bounding box with named XY values ("xmin", "xmax", "ymin", "ymax").}

\item{start_date}{Initial date for the cube files (optional).}

\item{end_date}{Final date for the cube files (optional).}
}
\value{
A data cube.
}
\description{
Defines a cube to retrieve data from the Brazil Data Cube (BDC)
             STAC. The retrieval is based on tiles of a given cube.
             For more on BDC, please see http://brazildatacube.dpi.inpe.br/
}
\examples{
\dontrun{
# this example requires access to an external service, so should not be run
# by CRAN

# Provide your BDC credentials as enviroment variables
# Sys.setenv(
# "BDC_ACCESS_KEY" = <your_bdc_access_key>
# )

# create a raster cube file based on the information about the files
cbers_tile <- sits_cube(
    type = "BDC",
    name = "cbers_022024",
    bands = c("NDVI", "EVI"),
    tiles = "022024",
    collection = "CB4_64_16D_STK-1",
    start_date = "2018-09-01",
    end_date = "2019-08-28"
)
}
}
\references{
`rstac` package (https://github.com/brazil-data-cube/rstac)
}
