% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_cube.R
\name{sits_cube.bdc_stac}
\alias{sits_cube.bdc_stac}
\title{Defines a data cube for a BDC STAC}
\usage{
\method{sits_cube}{bdc_stac}(
  type = "BDC_STAC",
  ...,
  name = NULL,
  tiles = NULL,
  bands = NULL,
  url = "http://brazildatacube.dpi.inpe.br/stac/",
  collection = NULL,
  roi = NULL,
  start_date = NULL,
  end_date = NULL
)
}
\arguments{
\item{type}{a \code{character} with the type of cube.}

\item{...}{other parameters to be passed for specific types.}

\item{name}{a \code{character} representing the output data cube.}

\item{tiles}{a \code{character} representing the names of the tiles.}

\item{bands}{a \code{character} with the bands names to be filtered.}

\item{url}{a \code{character} representing a URL for the BDC catalog.}

\item{collection}{a \code{character} with the collection to be searched.}

\item{roi}{the "roi" parameter defines a region of interest. It can be
an \code{sfc} or \code{sf} object from sf package, a \code{character} with
a GeoJSON following the rules from RFC 7946, or a \code{vector}
bounding box \code{vector} with named XY values
("xmin", "xmax", "ymin", "ymax").}

\item{start_date}{a \code{character} corresponds to the initial date when the
cube will be created.}

\item{end_date}{a \code{character} corresponds to the final date when the
cube will be created.}
}
\value{
A \code{raster_cube} object with the information of the
 created cube.
}
\description{
Defines a cube to retrieve data from the Brazil Data Cube (BDC)
             STAC. The retrieval is based on tiles of a given cube.
             For more on BDC, please see http://brazildatacube.dpi.inpe.br/
}
\examples{
\dontrun{
# this example requires access to an external service, so should not be run
# by CRAN

# create a raster cube file based on the information about the files
cbers_stac_tile <- sits_cube(type        = "BDC_STAC",
                             name        = "cbers_stac",
                             bands       = c("NDVI", "EVI"),
                             url         = "http://brazildatacube.dpi.inpe.br/stac/",
                             collection  = "CB4_64_16D_STK-1",
                             start_date  = "2018-09-01",
                             end_date    = "2019-08-28")
}
}
\references{
`rstac` package (https://github.com/brazil-data-cube/rstac)
}
