% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_cube.R
\name{sits_cube.bdc_stac}
\alias{sits_cube.bdc_stac}
\title{Defines a data cube for a BDC STAC}
\usage{
\method{sits_cube}{bdc_stac}(
  type = "BDC_STAC",
  name = NULL,
  tiles = NULL,
  bands = NULL,
  url = NULL,
  collection = NULL,
  ids = NULL,
  bbox = NULL,
  datetime = NULL,
  intersects = NULL,
  limit = NULL
)
}
\arguments{
\item{type}{a \code{character} with the type of cube.}

\item{name}{a \code{character} representing the output data cube.}

\item{tiles}{a \code{character} representing the names of the tiles.}

\item{bands}{a \code{character} with the bands names to be filtered.}

\item{url}{a \code{character} representing a URL for the BDC catalog.}

\item{collection}{a \code{character} with the collection to be searched.}

\item{ids}{a \code{character} vector with the items features ids.}

\item{bbox}{a \code{numeric} vector with only features that have a
geometry that intersects the bounding box are selected. The bounding box is
provided as four or six numbers, depending on whether the coordinate
reference system includes a vertical axis (elevation or depth):
\itemize{ \item Lower left corner, coordinate axis 1
          \item Lower left corner, coordinate axis 2
          \item Lower left corner, coordinate axis 3 (optional)
          \item Upper right corner, coordinate axis 1
          \item Upper right corner, coordinate axis 2
          \item Upper right corner, coordinate axis 3 (optional) }

The coordinate reference system of the values is WGS84 longitude/latitude
(\url{http://www.opengis.net/def/crs/OGC/1.3/CRS84}). The values are in
most cases the sequence of minimum longitude, minimum latitude, maximum
longitude and maximum latitude. However, in cases where the box spans the
antimeridian the first value (west-most box edge) is larger than the third
value (east-most box edge).}

\item{datetime}{a \code{character} with a date-time or an interval. Date
and time strings needs to conform RFC 3339. Intervals are expressed by
separating two date-time strings by \code{'/'} character. Open intervals are
expressed by using \code{'..'} in place of date-time.}

\item{intersects}{a \code{character} value expressing GeoJSON geometries
objects as specified in RFC 7946. Only returns items that intersect with
the provided polygon.}

\item{limit}{an \code{integer} defining the maximum number of results
to return. If not informed it defaults to the service implementation.}
}
\value{
A \code{raster_cube} object with the information of the
 created cube.
}
\description{
Defines a cube to retrieve data from the Brazil Data Cube (BDC)
             STAC. The retrieval is based on tiles of a given cube.
             For more on BDC, please see http://brazildatacube.dpi.inpe.br/
}
\examples{
\donttest{

# create a raster cube file based on the information about the files
}
}
\references{
`rstac` package (https://github.com/brazil-data-cube/rstac)
}
