% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_cube.R
\name{sits_cube}
\alias{sits_cube}
\title{Defines a data cube}
\usage{
sits_cube(service = "RASTER", provider = NULL, name = NULL,
  tiles_names = NULL, geom = NULL, from = NULL, to = NULL,
  timeline = NULL, bands = NULL, missing_values = NULL,
  scale_factors = NULL, minimum_values = NULL, maximum_values = NULL,
  files = NA)
}
\arguments{
\item{service}{Name of the data service.}

\item{provider}{Name of the service provider.}

\item{name}{Name of the image collection.}

\item{tiles_names}{A string with tile names to be filtered.}

\item{geom}{A \code{sfc} object to filter tiles that intersects the given geometry.}

\item{from}{A date value to filter cube's layers by date.}

\item{to}{A date value to filter cube's layers by date.}

\item{timeline}{Vector with the timeline of the collection.}

\item{bands}{Vector of bands.}

\item{missing_values}{Vector of missing values for each band.}

\item{scale_factors}{Vector with the scale factor for each band.}

\item{minimum_values}{Vector of minimum values for each band.}

\item{maximum_values}{Vector of maximum values for each band.}

\item{files}{Vector of file names for each band (only for raster data).}
}
\description{
Defines a cube to retrieve data. Cubes are associated to
data services. The following services are available:
\itemize{
 \item{"WTSS": }{Web Time Series Service - used to get time series}
 \item{"EOCUBES": }{EOCUBES service - used for cloud processing of data cubes}
 \item{"RASTER": }{Raster Brick files, local or remote}
 \item{"STACK": }{Raster Stack files, local or remote}
}
}
\examples{
\donttest{
# Example 1. Retrieve information about a WTSS collection
cube.tb <- sits_cube(service = "WTSS", name = "MOD13Q1")

# Example 2. Create a raster cube with metadata
# read a raster file and put it into a vector
files <- c(system.file("extdata/raster/mod13q1/sinop-crop-ndvi.tif", package = "sits"))

# create a raster cube file based on the information about the files
raster.tb <- sits_cube(service = "RASTER", name  = "Sinop-crop",
             timeline = timeline_modis_392, bands = "ndvi", files = files)
}
}
\seealso{
To see the available values for the parameters above use \code{\link{sits_services}}, \code{\link{sits_config}} or \code{\link{sits_show_config}}.
}
