% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_cube.R
\name{sits_cube}
\alias{sits_cube}
\title{Defines a data cube}
\usage{
sits_cube(type = "RASTER", ...)
}
\arguments{
\item{type}{Type of cube (one of "WTSS", "RASTER", "BDC_TILE",
"S2_L2A_AWS", "PROBS", "CLASSIFIED")}

\item{...}{Other parameters to be passed for specific types}
}
\description{
Defines a cube to retrieve data. This is a generic function.
Cubes can be of the following types. See each function description for the
required parameters:
\itemize{
 \item{"WTSS": }{Web Time Series Service - see \code{\link{sits_cube.wtss_cube}}}
 \item{"SATVEG": }{ SATVEG Time Series Service - see \code{\link{sits_cube.satveg_cube}}}
 \item{"RASTER": }{Raster files - see \code{\link{sits_cube.raster_cube}}}
 \item{"BDC_TILE"}{A tile from the Brazil Data Cube - see \code{\link{sits_cube.bdc_cube}}}
 \item{"S2_L2A_AWS"}{A tile of Sentinel-2 data in AWS - see \code{\link{sits_cube.s2_l2a_aws_cube}}}
}
}
\examples{
\donttest{
# Example 1. Create a data cube based on a WTSS service
cube_wtss <- sits_cube(type = "WTSS",
          name = "MOD13Q1",
          URL = "http://www.esensing.dpi.inpe.br/wtss/")

# Example 2. Create a data cube based on the SATVEG service
cube_satveg <- sits_cube(type = "SATVEG",
                         name = "terra")

# Example 3. Create a raster cube
# inform the files that make up a raster brick with 392 time instances
files <- c(system.file("extdata/raster/mod13q1/sinop-crop-ndvi.tif",
           package = "sits"))

# create a raster cube file based on the information about the files
raster.tb <- sits_cube(type = "RASTER",
                       name      = "Sinop-crop",
                       satellite = "TERRA",
                       sensor    = "MODIS",
                       timeline  = timeline_modis_392,
                       bands     = "NDVI",
                       files     = files)

}
}
