% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_cube.R
\name{sits_cube}
\alias{sits_cube}
\title{Defines a data cube}
\usage{
sits_cube(service = "LOCALHOST", URL = NULL, name = NULL,
  satellite = NULL, sensor = NULL, tiles_names = NULL, geom = NULL,
  from = NULL, to = NULL, timeline = NULL, bands = NULL,
  files = NA)
}
\arguments{
\item{service}{Name of the data service.}

\item{URL}{URL of the service provider.}

\item{name}{Name of the data cube in the remote service.}

\item{satellite}{Name of satellite}

\item{sensor}{Name of sensor}

\item{tiles_names}{A string with tile names to be filtered (for EOCUBES service)}

\item{geom}{An \code{sf} object to filter tiles (for EOCUBES service)}

\item{from}{Starting date for the cube to be extracted}

\item{to}{End date for the cube to be extracted}

\item{timeline}{Vector with the timeline of the collection (only for local files)}

\item{bands}{Vector of bands.}

\item{files}{Vector of file names for each band (only for raster data).}
}
\description{
Defines a cube to retrieve data. Cubes are associated to
data services. The following services are available:
\itemize{
 \item{"WTSS": }{Web Time Series Service - used to get time series}
 \item{"EOCUBES": }{EOCUBES service - used for cloud processing of data cubes}
 \item{"AWS": }{AWS service - used for cloud processing of data cubes}
 \item{"LOCALHOST": }{Raster Brick files, local}
 \item{"STACK": }{Raster Stack files, local}
}
}
\examples{
\donttest{
# Example 1. Retrieve information about a WTSS collection
cube.tb <- sits_cube(service = "WTSS", name = "MOD13Q1")

# Example 2. Create a raster cube with metadata
# read a raster file and put it into a vector
files <- c(system.file("extdata/raster/mod13q1/sinop-crop-ndvi.tif", package = "sits"))

# create a raster cube file based on the information about the files
raster.tb <- sits_cube(name  = "Sinop-crop",
             timeline = timeline_modis_392, bands = "ndvi", files = files)
}
}
\seealso{
To see the available values for the parameters above use \code{\link{sits_services}}, \code{\link{sits_config}} or \code{\link{sits_show_config}}.
}
