% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_cube.R
\name{sits_cube}
\alias{sits_cube}
\title{Defines a data cube}
\usage{
sits_cube(type = "RASTER", ...)
}
\arguments{
\item{type}{Type of cube (one of "WTSS", "RASTER", "BDC_TILE",
"S2_L2A_AWS", "PROBS", "CLASSIFIED")}

\item{...}{Other parameters to be passed for specific types}
}
\description{
Defines a cube to retrieve data. This is a generic function.
Cubes can be of the following types. See each function description for the
required parameters:
\itemize{
 \item{"WTSS": }{see \code{\link{sits_cube.wtss_cube}}}
 \item{"SATVEG": }{ see \code{\link{sits_cube.satveg_cube}}}
 \item{"RASTER": }{see \code{\link{sits_cube.raster_cube}}}
 \item{"BDC_TILE"}{Brazil Data Cube - see \code{\link{sits_cube.bdc_cube}}}
 \item{"BDC_STAC"}{STAC Brazil Data Cube - see \code{\link{sits_cube.bdc_stac}}}
 \item{"S2_L2A_AWS"}{Sentinel-2 data in AWS - see \code{\link{sits_cube.s2_l2a_aws_cube}}}
}
}
