% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_cube.R
\name{sits_cube}
\alias{sits_cube}
\title{Defines a data cube}
\usage{
sits_cube(
  type = NULL,
  name,
  URL = NULL,
  satellite = NULL,
  sensor = NULL,
  timeline = NULL,
  bands = NULL,
  files = NULL,
  cube = NULL,
  tile = NULL,
  data_access = "local",
  start_date = NULL,
  end_date = NULL,
  service = NULL,
  .local = NULL,
  .web = NULL
)
}
\arguments{
\item{type}{Type of cube (one of "WTSS", "SATVEG", "BRICK", "BDC-TILE")}

\item{name}{Name of the output data cube.}

\item{URL}{URL of the service provider (for WTSS).}

\item{satellite}{Name of satellite}

\item{sensor}{Name of sensor}

\item{timeline}{Vector with the timeline of the collection
(only for local files)}

\item{bands}{Vector of bands.}

\item{files}{Vector of file names for each band
(only for BRICK data cube).}

\item{cube}{Name of the input data cube (or image collection)}

\item{tile}{Name of the tile}

\item{data_access}{Type of access (local or web)}

\item{start_date}{Starting date of the cube}

\item{end_date}{Ending date of the cube}

\item{service}{Name of the data service (deprecated)}

\item{.local}{Directory for local access to the input cube (optional)}

\item{.web}{Directory for web access to the input cube (optional)}
}
\description{
Defines a cube to retrieve data. '
Cubes can be of the following types:
\itemize{
 \item{"WTSS": }{Web Time Series Service - used to get time series}
 \item{"SATVEG": }{ SATVEG Time Series Service - used to get time series}
 \item{"BRICK": }{Raster Brick files}
 \item{"BDC_TILE"}{A tile from the Brazil Data Cube}
}
}
\examples{
\donttest{
# Example 1. Create a data cube based on a WTSS service
cube_wtss <- sits_cube(type = "WTSS",
          name = "MOD13Q1",
          URL = "http://www.esensing.dpi.inpe.br/wtss/")

# Example 2. Create a data cube based on the SATVEG service
cube_satveg <- sits_cube(type = "SATVEG",
                         name = "terra")

# Example 3. Create a raster cube based on bricks
# inform the files that make up a raster brick with 392 time instances
files <- c(system.file("extdata/raster/mod13q1/sinop-crop-ndvi.tif",
           package = "sits"))

# create a raster cube file based on the information about the files
raster.tb <- sits_cube(type = "BRICK",
                       name      = "Sinop-crop",
                       satellite = "TERRA",
                       sensor    = "MODIS",
                       timeline  = timeline_modis_392,
                       bands     = "ndvi",
                       files     = files)

}
}
\seealso{
To see the available values for the parameters above
use \code{\link{sits_config}} or \code{\link{sits_config_show}}.
}
