% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_cube_local.R
\name{sits_cube.results_cube}
\alias{sits_cube.results_cube}
\title{Create a results cube from local files}
\usage{
\method{sits_cube}{results_cube}(
  source,
  collection,
  ...,
  data_dir,
  bands = NULL,
  labels,
  parse_info = c("X1", "X2", "tile", "start_date", "end_date", "band", "version"),
  version = "v1",
  delim = "_",
  multicores = 2L,
  progress = TRUE
)
}
\arguments{
\item{source}{Data source: one of \code{"AWS"}, \code{"BDC"},
\code{"CDSE"}, \code{"DEAFRICA"}, \code{"DEAUSTRALIA"},
\code{"HLS"}, \code{"PLANETSCOPE"}, \code{"MPC"},
\code{"SDC"} or \code{"USGS"}. This is the source
from which the data has been downloaded.}

\item{collection}{Image collection in data source.
To find out the supported collections,
use \code{\link{sits_list_collections}()}).}

\item{...}{Other parameters to be passed for specific types.}

\item{data_dir}{Local directory where images are stored}

\item{bands}{Results bands to be retrieved
("probs", "bayes", "variance", "class", "uncertainty")}

\item{labels}{Labels associated to the classes
(Named character vector for cubes of
classes "probs_cube" or "class_cube")}

\item{parse_info}{Parsing information for local files
(see notes below).}

\item{version}{Version of the classified and/or labelled files.}

\item{delim}{Delimiter for parsing local files
(default = "_")}

\item{multicores}{Number of workers for parallel processing
(integer, min = 1, max = 2048).}

\item{progress}{Logical: show a progress bar?}
}
\value{
A \code{tibble} describing the contents of a data cube.
}
\description{
Creates a data cube from local files produced by sits operations
that produces results (such as probs_cubs and class_cubes)
}
\note{
This function creates result cubes from local files produced by
classification or post-classification algorithms. In this case, the
\code{parse_info} is specified differently, and additional parameters
are required.
The parameter \code{bands} should be a single character vector with
the name associated to the type of result:
\itemize{
\item{\code{"probs"}, for probability cubes produced by
  \code{\link[sits]{sits_classify}}.}
\item{\code{"bayes"}, for smoothed cubes produced by
   \code{\link[sits]{sits_smooth}}.}
\item{\code{"entropy"} when using \code{\link[sits]{sits_uncertainty}} to measure
  entropy in pixel classification.}
\item{\code{"margin"} when using \code{\link[sits]{sits_uncertainty}} to measure
  probability margin in pixel classification.}
\item{\code{"least"} when using \code{\link[sits]{sits_uncertainty}} to measure
  difference between 100\% and most probable class in pixel classification.}
\item{\code{"class"} for cubes produced by
         \code{\link[sits]{sits_label_classification}}.}
}
For cubes of type \code{"probs"}, \code{"bayes"}, \code{"class"}, the
    \code{labels} parameter should be named vector associated to the
    classification results. For \code{"class"} cubes, its names should be
    integers associated to the values of the raster files that represent
    the classified cube.

Parameter \code{parse_info} should contain parsing information
  to deduce the values of \code{tile}, \code{start_date},
  \code{end_date} and \code{band} from the file name.
  Default is c("X1", "X2", "tile", "start_date", "end_date", "band").
}
