% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_cube.R
\name{sits_cube.s2_l2a_aws_cube}
\alias{sits_cube.s2_l2a_aws_cube}
\title{Defines a data cube for a Sentinel-2 L2A AWS cube}
\usage{
\method{sits_cube}{s2_l2a_aws_cube}(
  type = "S2_L2A_AWS",
  ...,
  name = NULL,
  url = NULL,
  collection = NULL,
  tiles = NULL,
  bands = NULL,
  s2_resolution = NULL,
  roi = NULL,
  start_date = NULL,
  end_date = NULL
)
}
\arguments{
\item{type}{type of cube}

\item{...}{other parameters to be passed for specific types}

\item{name}{output data cube.}

\item{url}{URL for the STAC AWS catalog (optional). By default,
'https://earth-search.aws.element84.com/v0/' will be used.}

\item{collection}{AWS collection to be searched (mandatory).}

\item{tiles}{Tile names to be searched (optional).}

\item{bands}{Bands names to be filtered (optional).}

\item{s2_resolution}{resolution of S2 images ("10m", "20m" or "60m");}

\item{roi}{selects images (tiles) that intersect according to the
region of interest provided. Expressed either as an \code{sfc} or \code{sf}
object from sf package, or a \code{vector} with bounding box named XY values
in WGS 84 ("xmin", "xmax", "ymin", "ymax").}

\item{start_date}{Initial date for the cube files (optional).}

\item{end_date}{Final date for the cube files (optional).}
}
\value{
data cube
}
\description{
Defines a cube to retrieve data from the Sentinel-2 L2A data
             available in AWS. Users need to be an AWS
             user and provide her access key and secret key. These keys may
             be passed as environment variables. The AWS bands in
             10m resolution are "B02", "B03", "B04", and "B08".
             The  20m bands are "B02", "B03", "B04", "B05", "B06", "BO7",
             "B08", "B8A", "B11", and "B12".
             All 12 bands are available at 60m resolution.

   Users need to provide AWS credentials in their environmental variables\cr
Sys.setenv(\cr
     "AWS_ACCESS_KEY_ID"     = <your_access_key>, \cr
     "AWS_SECRET_ACCESS_KEY" = <your_secret_access_key>,\cr
     "AWS_DEFAULT_REGION"    = <your AWS region>, \cr
     "AWS_ENDPOINT" = "sentinel-s2-l2a.s3.amazonaws.com", \cr
     "AWS_REQUEST_PAYER"     = "requester" \cr
)
}
\examples{
\dontrun{
# this example requires access to an external service, so should not be run
# by CRAN

# Provide your AWS credentials here
# Sys.setenv(
# "AWS_ACCESS_KEY_ID"     = <your_access_key>,
# "AWS_SECRET_ACCESS_KEY" = <your_secret_access_key>,
# "AWS_DEFAULT_REGION"    = <your AWS region>,
# "AWS_ENDPOINT" = "sentinel-s2-l2a.s3.amazonaws.com",
# "AWS_REQUEST_PAYER"     = "requester"
# )

cube_s2 <- sits::sits_cube(type = "S2_L2A_AWS",
                           name = "s2_cube",
                           collection = "sentinel-s2-l2a",
                           bands = c("B04", "B08"),
                           s2_resolution = "20m",
                           roi = c("xmin" = 17.379,
                                   "ymin" = 1.1573,
                                   "xmax" = 17.410,
                                   "ymax" = 1.1910),
                           start_date = "2019-01-01",
                           end_date = "2019-10-28")

}

}
