% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_cube_local.R
\name{sits_cube.vector_cube}
\alias{sits_cube.vector_cube}
\title{Create a vector cube from local files}
\usage{
\method{sits_cube}{vector_cube}(
  source,
  collection,
  ...,
  raster_cube,
  vector_dir,
  vector_band = NULL,
  parse_info = c("X1", "X2", "tile", "start_date", "end_date", "band", "version"),
  version = "v1",
  delim = "_",
  multicores = 2L,
  progress = TRUE
)
}
\arguments{
\item{source}{Data source: one of \code{"AWS"}, \code{"BDC"},
\code{"CDSE"}, \code{"DEAFRICA"}, \code{"DEAUSTRALIA"},
\code{"HLS"}, \code{"PLANETSCOPE"}, \code{"MPC"},
\code{"SDC"} or \code{"USGS"}. This is the source
from which the data has been downloaded.}

\item{collection}{Image collection in data source.
To find out the supported collections,
use \code{\link{sits_list_collections}()}).}

\item{...}{Other parameters to be passed for specific types.}

\item{raster_cube}{Raster cube to be merged with vector data.}

\item{vector_dir}{Local directory where vector files are stored.}

\item{vector_band}{Band for vector cube ("segments", "probs", "class").
This parameter is deprecated and will be removed in the
next versions. Now, the type vector data cube loaded
is defined based on raster_cube object.}

\item{parse_info}{Parsing information for local vector files.}

\item{version}{Version of the classified and/or labelled files.}

\item{delim}{Delimiter for parsing local vector files.
(default = "_")}

\item{multicores}{Number of workers for parallel processing
(integer, min = 1, max = 2048).}

\item{progress}{Logical: show a progress bar?}
}
\value{
A \code{tibble} describing the contents of a vector data cube.
}
\description{
Creates a data cube from local files which include a vector file
produced by a segmentation algorithm.
}
\note{
This function creates vector cubes from local files produced by
\code{\link[sits]{sits_segment}} and by the operations applied to
segmented cubes. It can also be applied to merge external
segment (polygon) files with raster data cubes.
In this case,
\code{parse_info} is differently as \code{c("X1", "X2", "tile",
"start_date", "end_date", "band", "version")}, where band is one of
"segments", "probs", or "class".
The parameter \code{vector_dir} is the directory where the segments
produced by \code{\link[sits]{sits_segment}} are stored.

The type of vector cube which is loaded is defined by \code{raster_cube}:
\itemize{
\item{a raster cube produces a \code{segs_cube}.}
\item{a probability cube produced by
   \code{\link{sits_classify}} results in a
   \code{probs_vector_cube}.}
\item{a classified cube produced by
   \code{\link{sits_label_classification}} results in a
   \code{class_vector_cube}.}
\item{an uncertainty cube produced by
   \code{\link{sits_uncertainty}} results in an
   \code{uncertainty_vector_cube}.}
\item{a variance cube produced by
   \code{\link{sits_variance}} results in a \code{variance_vector_cube}.}
}
}
\examples{
if (sits_run_examples()) {
    # load local cube
    data_dir <- system.file("extdata/raster/mod13q1", package = "sits")
    modis_cube <- sits_cube(
        source = "BDC",
        collection = "MOD13Q1-6.1",
        data_dir = data_dir,
        parse_info = c("satellite", "sensor", "tile", "band", "date")
    )
    # segment the raster cube
    segs_cube <- sits_segment(
        cube = modis_cube,
        seg_fn = sits_snic(
            grid_seeding = "rectangular",
            spacing = 15,
            compactness = 0.4,
            padding = 2
        ),
        output_dir = tempdir()
    )
    plot(segs_cube)

    # recover the local segmented cube
    local_segs_cube <- sits_cube(
        source = "BDC",
        collection = "MOD13Q1-6.1",
        raster_cube = modis_cube,
        vector_dir = tempdir()
    )
    # plot the recovered cube and compare
    plot(local_segs_cube)

    # classify the segments
    # create a random forest model
    rfor_model <- sits_train(samples_modis_ndvi, sits_rfor())
    probs_vector_cube <- sits_classify(
        data = segs_cube,
        ml_model = rfor_model,
        output_dir = tempdir()
    )
    # name the labels of the local results cubes
    labels <- sits_labels(rfor_model)
    names(labels) <- seq_along(labels)

    # recover the probability cube and attach the segments to it
    local_probs_cube <- sits_cube(
        source = "BDC",
        collection = "MOD13Q1-6.1",
        data_dir = tempdir(),
        bands = "probs",
        labels = labels
    )
    local_probs_vector_cube <- sits_cube(
        source = "BDC",
        collection = "MOD13Q1-6.1",
        raster_cube = local_probs_cube,
        vector_dir = tempdir()
    )
    plot(local_probs_vector_cube)

    # label the segments
    class_vector_cube <- sits_label_classification(
        cube = probs_vector_cube,
        label_method = "mean",
        output_dir = tempdir()
    )

    # recover the classified cube and attach the segments to it
    local_class_cube <- sits_cube(
        source = "BDC",
        collection = "MOD13Q1-6.1",
        data_dir = tempdir(),
        bands = "class",
        labels = labels
    )
    local_class_vector_cube <- sits_cube(
        source = "BDC",
        collection = "MOD13Q1-6.1",
        raster_cube = local_class_cube,
        vector_dir = tempdir()
    )
    plot(local_class_vector_cube)

    # measure the uncertainty of the segments
    uncert_vector_cube <- sits_uncertainty(
        cube = probs_vector_cube,
        type = "entropy",
        output_dir = tempdir()
    )

    # recover the uncertainty cube and attach the segments to it
    local_uncert_cube <- sits_cube(
        source = "BDC",
        collection = "MOD13Q1-6.1",
        data_dir = tempdir(),
        bands = "entropy"
    )
    local_uncert_vector_cube <- sits_cube(
        source = "BDC",
        collection = "MOD13Q1-6.1",
        raster_cube = local_uncert_cube,
        vector_dir = tempdir()
    )

    # measure the variance of the segments
    var_vector_cube <- sits_variance(
        cube = probs_vector_cube,
        output_dir = tempdir()
    )

    # recover the variance cube and attach the segments to it
    local_var_cube <- sits_cube(
        source = "BDC",
        collection = "MOD13Q1-6.1",
        data_dir = tempdir(),
        bands = "variance",
        labels = labels
    )
    local_var_vector_cube <- sits_cube(
        source = "BDC",
        collection = "MOD13Q1-6.1",
        raster_cube = local_var_cube,
        vector_dir = tempdir()
    )
}

}
