% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_table.R
\name{sits_dates}
\alias{sits_dates}
\title{Return the dates of a sits table}
\usage{
sits_dates(data.tb)
}
\arguments{
\item{data.tb}{a tibble in SITS format with time series for different bands}
}
\value{
table   a tibble in SITS format with values of time indexes
}
\description{
\code{sits_dates} returns a sits table with dates only (colwise organised)
}
\details{
A sits table has the metadata and data for each time series
<longitude, latitude, start_date, end_date, label, coverage, time_series>
}
\seealso{
Other STIS table functions: \code{\link{sits_align}},
  \code{\link{sits_bands}},
  \code{\link{sits_group_bylatlong}},
  \code{\link{sits_rename}}, \code{\link{sits_select}},
  \code{\link{sits_table_result}},
  \code{\link{sits_table}}, \code{\link{sits_values_cols}},
  \code{\link{sits_values_rows}}
}
