% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_tibble_functions.R
\name{sits_dates}
\alias{sits_dates}
\title{Return the dates of a sits table}
\usage{
sits_dates(data.tb)
}
\arguments{
\item{data.tb}{tibble in SITS format with time series for different bands}
}
\value{
table    tibble  with values of time indexes
}
\description{
returns a vector containing the dates of a sits table
}
\examples{
# get a point
data (ts_2000_2016)
# return a vector of values
sits_dates (ts_2000_2016)
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
