% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_table.R
\name{sits_dates}
\alias{sits_dates}
\title{Return the dates of a sits table}
\usage{
sits_dates(data.tb)
}
\arguments{
\item{data.tb}{a tibble in SITS format with time series for different bands}
}
\value{
table   a tibble in SITS format with values of time indexes
}
\description{
returns a vector containing the dates of a sits table
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
