% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_db.R
\name{sits_db_info}
\alias{sits_db_info}
\title{List the time series and data cubes stored in an SQLite database}
\usage{
sits_db_info(conn)
}
\arguments{
\item{conn}{SQLite connection}
}
\value{
Description
}
\description{
This functions describes the time series and data cubes
stored RSQLite database.
}
\examples{
\donttest{
# create a data base
conn <- sits_db_create("sits.sql")
# write a set of time series
conn <- sits_db_write(conn, "cerrado_2classes", cerrado_2classes)
# describe the data available in the database
desc <-  sits_db_info(conn)
}
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
