% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_deafrica.R
\name{sits_deafrica_bands}
\alias{sits_deafrica_bands}
\alias{.sits_deafrica_bands}
\title{Get bands names from items}
\usage{
.sits_deafrica_bands(items, bands)
}
\arguments{
\item{items}{a \code{STACItemCollection} object returned by rstac package.
grouped.}

\item{bands}{a \code{character} vector with the bands name.}
}
\value{
a \code{STACItemCollection} object representing the search
             by rstac.
}
\description{
Get bands names from items
}
\keyword{internal}
