% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_deep_learning.R
\name{sits_deeplearning}
\alias{sits_deeplearning}
\title{Train a classification model using a multi-layer perceptron.}
\usage{
sits_deeplearning(data = NULL, layers = c(512, 512, 512, 512, 512),
  activation = "elu", dropout_rates = c(0.5, 0.4, 0.35, 0.3, 0.2),
  optimizer = keras::optimizer_adam(lr = 0.001), epochs = 500,
  batch_size = 128, validation_split = 0.2, verbose = 1,
  binary_classification = FALSE)
}
\arguments{
\item{data}{Time series with the training samples.}

\item{layers}{Vector with the number of hidden nodes in each hidden layer.}

\item{activation}{Vector with the names of activation functions. Valid values are {'relu', 'elu', 'selu', 'sigmoid'}.}

\item{dropout_rates}{Vector with the dropout rates (0,1) for each layer to the next layer.}

\item{optimizer}{Function with a pointer to the optimizer function (default is optimization_adam()).
Options are optimizer_adadelta(), optimizer_adagrad(), optimizer_adam(),
optimizer_adamax(), optimizer_nadam(), optimizer_rmsprop(), optimizer_sgd()}

\item{epochs}{Number of iterations to train the model.}

\item{batch_size}{Number of samples per gradient update.}

\item{validation_split}{Number between 0 and 1. Fraction of the training data to be used as validation data.
The model will set apart this fraction of the training data, will not train on it,
and will evaluate the loss and any model metrics on this data at the end of each epoch.
The validation data is selected from the last samples in the x and y data provided,
before shuffling.}

\item{verbose}{Verbosity mode (0 = silent, 1 = progress bar, 2 = one line per epoch).}

\item{binary_classification}{A lenght-one logical indicating if this is a binary classification. If it is so,
the number of unique labels in the training data must be two as well.}
}
\value{
A model fitted to input data to be passed to \code{\link[sits]{sits_classify}}
}
\description{
Use a multi-layer perceptron to classify data. USers can define
the number and size of the hidden layers, and the dropout rates and activation
functions for each layer.
This function is a front-end to the "keras" method R package.
Please refer to the documentation in that package for more details.
}
\examples{
\donttest{
# Retrieve the set of samples for the Mato Grosso region (provided by EMBRAPA)
samples_2bands <- sits_select_bands(samples_mt_6bands, ndvi, evi)

# Build a machine learning model based on deep learning
dl_model <- sits_train (samples_2bands,
                        sits_deeplearning(layers = c(512, 512, 512),
                                          dropout_rates = c(0.50, 0.40, 0.35),
                                          epochs = 100))

# get a point and classify the point with the ml_model
point.tb <- sits_select_bands(point_mt_6bands, ndvi, evi)
class.tb <- sits_classify(point.tb, dl_model)
sits_plot(class.tb)
}
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}

Alexandre Xavier Ywata de Carvalho, \email{alexandre.ywata@ipea.gov.br}

Rolf Simoes, \email{rolf.simoes@inpe.br}
}
