% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_detect_change.R
\name{sits_detect_change}
\alias{sits_detect_change}
\alias{sits_detect_change.sits}
\alias{sits_detect_change.default}
\title{Detect changes in time series}
\usage{
sits_detect_change(
  data,
  cd_method,
  ...,
  filter_fn = NULL,
  multicores = 2L,
  progress = TRUE
)

\method{sits_detect_change}{sits}(
  data,
  cd_method,
  ...,
  filter_fn = NULL,
  multicores = 2L,
  progress = TRUE
)

\method{sits_detect_change}{default}(data, cd_method, ...)
}
\arguments{
\item{data}{Set of time series.}

\item{cd_method}{Change detection method (with parameters).}

\item{...}{Other relevant parameters.}

\item{filter_fn}{Smoothing filter function to be applied to the data.}

\item{multicores}{Number of threads to process the time series.}

\item{progress}{Show progress bar?}
}
\value{
Set of time series where significant change has been
                    detected.
}
\description{
Given a set of time series or an image, this function compares
each time series with a set of change/no-change patterns, and indicates
places and dates where change has been detected.
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}

Felipe Carlos, \email{efelipecarlos@gmail.com}

Felipe Carvalho, \email{felipe.carvalho@inpe.br}
}
