% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_detect_change.R
\name{sits_detect_change}
\alias{sits_detect_change}
\alias{sits_detect_change.sits}
\alias{sits_detect_change.raster_cube}
\alias{sits_detect_change.default}
\title{Detect changes in time series}
\usage{
sits_detect_change(
  data,
  dc_method,
  ...,
  filter_fn = NULL,
  multicores = 2L,
  progress = TRUE
)

\method{sits_detect_change}{sits}(
  data,
  dc_method,
  ...,
  filter_fn = NULL,
  multicores = 2L,
  progress = TRUE
)

\method{sits_detect_change}{raster_cube}(
  data,
  dc_method,
  ...,
  roi = NULL,
  filter_fn = NULL,
  impute_fn = identity,
  start_date = NULL,
  end_date = NULL,
  memsize = 8L,
  multicores = 2L,
  output_dir,
  version = "v1",
  verbose = FALSE,
  progress = TRUE
)

\method{sits_detect_change}{default}(data, dc_method, ...)
}
\arguments{
\item{data}{Set of time series.}

\item{dc_method}{Change detection method (with parameters).}

\item{...}{Other parameters for specific functions.}

\item{filter_fn}{Smoothing filter to be applied - optional
(closure containing object of class "function").}

\item{multicores}{Number of cores to be used for classification
(integer, min = 1, max = 2048).}

\item{progress}{Logical: Show progress bar?}

\item{roi}{Region of interest (either an sf object, shapefile,
or a numeric vector with named XY values
("xmin", "xmax", "ymin", "ymax") or
named lat/long values
("lon_min", "lat_min", "lon_max", "lat_max").}

\item{impute_fn}{Imputation function to remove NA.}

\item{start_date}{Start date for the classification
(Date in YYYY-MM-DD format).}

\item{end_date}{End date for the classification
(Date in YYYY-MM-DD format).}

\item{memsize}{Memory available for classification in GB
(integer, min = 1, max = 16384).}

\item{output_dir}{Valid directory for output file.
(character vector of length 1).}

\item{version}{Version of the output
(character vector of length 1).}

\item{verbose}{Logical: print information about processing time?}
}
\value{
Time series with detection labels for
                          each point (tibble of class "sits")
                          or a data cube indicating detections in each pixel
                          (tibble of class "detections_cube").
}
\description{
Given a set of time series or an image, this function compares
each time series with a set of change/no-change patterns, and indicates
places and dates where change has been detected.
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}

Felipe Carlos, \email{efelipecarlos@gmail.com}

Felipe Carvalho, \email{felipe.carvalho@inpe.br}
}
