% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_distances.R
\name{sits_distances}
\alias{sits_distances}
\title{Calculate a set of distance measures for satellite image time series}
\usage{
sits_distances(data.tb, patterns.tb,
  dist_method = sits_TWDTW_distances(data.tb = NULL, patterns.tb = NULL,
  by_bands = TRUE, alpha = -0.1, beta = 100, theta = 0.5, span = 0))
}
\arguments{
\item{data.tb}{a SITS tibble time series}

\item{patterns.tb}{a set of patterns obtained from training samples}

\item{dist_method}{a method for calculating distances between time series}
}
\value{
result          a set of distance metrics
}
\description{
This function allows the user to select different alternatives to define a set of
distances between a set of satellite image time series and a set of patterns.
 The alternatives are:
"TWTDTW"    - uses the TWDTW (time-weighted dynamic time warping)
other methods as used by the TSdist package
}
\author{
Rolf Simoes, \email{rolf.simoes@inpe.br}

Gilberto Camara, \email{gilberto.camara@inpe.br}
}
