% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_distances.R
\name{sits_distances}
\alias{sits_distances}
\title{Use time series values from a sits tibble as distances for training patterns}
\usage{
sits_distances(data.tb, dist_method = sits_distances_from_data())
}
\arguments{
\item{data.tb}{a SITS tibble with original data}

\item{dist_method}{method used to compute the distances}
}
\value{
distances.tb  a tibble where columns have the reference label and the time series values as distances
}
\description{
This function allows using a set of labelled time series as
input to the machine learning models. The values used as input for the training methods
are computed by the function defined by the parameter dist_method
}
\examples{
# Retrieve the set of samples for the Cerrado and Pasture classes
data(cerrado_2classes)
# estimate distances from the data
distances.tb <- sits_distances(cerrado_2classes)
}
\author{
Rolf Simoes, \email{rolf.simoes@inpe.br}

Gilberto Camara, \email{gilberto.camara@inpe.br}
}
