% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_dtw.R
\name{sits_dtw}
\alias{sits_dtw}
\title{Dynamic Time Warping for Detect changes.}
\usage{
sits_dtw(samples = NULL, ..., threshold = NULL, window = NULL)
}
\arguments{
\item{samples}{Time series with the training samples.}

\item{...}{Other relevant parameters.}

\item{threshold}{Threshold used to define if an event was detected.}

\item{window}{ISO8601-compliant time period used to define the
DTW moving window, with number and unit,
where "D", "M" and "Y" stands for days, month and
year; e.g., "P16D" for 16 days. This parameter is not
used in operations with data cubes.}
}
\value{
Change detection method prepared to be passed to
                      \code{\link[sits]{sits_detect_change_method}}
}
\description{
Create a Dynamic Time Warping (DTW) method for the
\code{\link[sits]{sits_detect_change_method}}.
}
\author{
Felipe Carlos, \email{efelipecarlos@gmail.com}

Felipe Carvalho, \email{felipe.carvalho@inpe.br}

Gilberto Camara, \email{gilberto.camara@inpe.br}

Rolf Simoes, \email{rolf.simoes@inpe.br}
}
