% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_detect_change_method.R
\name{sits_dtw}
\alias{sits_dtw}
\title{Dynamic Time Warping for Detect changes.}
\usage{
sits_dtw(
  samples = NULL,
  distance_threshold,
  window_size_left = 2,
  window_size_right = 2,
  window_step = 4,
  window_complete = T
)
}
\arguments{
\item{samples}{Time series with the training samples.}

\item{distance_threshold}{Threshold used to define if an event was detected.}

\item{window_size_left}{Size of window in the left side of the current
position.}

\item{window_size_right}{Size of window in the right side of the current
position.}

\item{window_step}{Window step size.}

\item{window_complete}{Flag indicating if only complete window, including
all values defined in `window_size_left` and
`window_size_right`, should be used.}
}
\value{
Change detection method prepared
                         to be passed to
                         \code{\link[sits]{sits_detect_change_method}}
}
\description{
Create a Dynamic Time Warping (DTW) method for the
\code{\link[sits]{sits_detect_change_method}}.
}
\author{
Felipe Carlos, \email{efelipecarlos@gmail.com}

Gilberto Camara, \email{gilberto.camara@inpe.br}

Felipe Carvalho, \email{felipe.carvalho@inpe.br}
}
