% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_encode.R
\name{sits_encode}
\alias{sits_encode}
\alias{sits_encode.tbl_df}
\alias{sits_encode.derived_cube}
\alias{sits_encode.default}
\title{encode time series or data cubes}
\usage{
sits_encode(data, dl_model, ...)

\method{sits_encode}{tbl_df}(data, dl_model, ...)

\method{sits_encode}{derived_cube}(data, dl_model, ...)

\method{sits_encode}{default}(data, dl_model, ...)
}
\arguments{
\item{data}{Data cube (tibble of class "raster_cube")}

\item{dl_model}{torch model trained by \code{\link[sits]{sits_pre_train}}}

\item{...}{Other parameters for specific functions.}
}
\value{
Time series with encoded labels for
                          each point (tibble of class "sits")
                          or a data cube with embeddings
                          (tibble of class "embeddings_cube").
}
\description{
This function encodes a set of time series or data cube using
a trained model prediction model created by \code{\link[sits]{sits_pre_train}}.

The \code{sits_encode} function takes two types of data as input
   and produce two types of output. Users should call
   \code{\link[sits]{sits_encode}} but be aware that the parameters
   are different for each type of input.
\itemize{
   \item{\code{\link[sits]{sits_encode.sits}} is called when the input is
   a set of time series. The output is the same set
   with the additional column \code{predicted}.}
   \item{\code{\link[sits]{sits_encode.raster_cube}} is called when the
   input is a regular raster data cube. The output is an embeddings cube,
   which has the same tiles as the raster cube. Each tile contains
   a multiband image; each band contains a single embedding dimension
   Embeddings cubes are objects of class "embeddings_cube".}
   }
}
\note{
The main \code{sits} encoding workflow has the following steps:
\enumerate{
     \item{\code{\link[sits]{sits_cube}}: selects a ARD image collection from
         a cloud provider.}
     \item{\code{\link[sits]{sits_cube_copy}}: copies an ARD image collection
         from a cloud provider to a local directory for faster processing.}
     \item{\code{\link[sits]{sits_regularize}}: create a regular data cube
         from an ARD image collection.}
     \item{\code{\link[sits]{sits_apply}}: create new indices by combining
         bands of a  regular data cube (optional).}
     \item{\code{\link[sits]{sits_get_data}}: extract time series
         from a regular data cube based on user-provided labelled samples.}
     \item{\code{\link[sits]{sits_pre_train}}: pre train a deep learning encoder
         model based on image time series.}
     \item{\code{\link[sits]{sits_encode}}: encode a data cube
         using a machine learning model and obtain an embeddings cube.}
}

SITS supports the following pre-trained encoders:
\itemize{
     \item{masked autoencoder: \code{\link[sits]{sits_mae}};}
}

   Please refer to the sits documentation available in
   \url{https://e-sensing.github.io/sitsbook/} for detailed examples.
}
\author{
Rolf Simoes, \email{rolfsimoes@gmail.com}

Alexandre Assuncao, \email{alexcarssuncao@gmail.com}
}
