% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_encode.R
\name{sits_encode}
\alias{sits_encode}
\alias{sits_encode.tbl_df}
\alias{sits_encode.derived_cube}
\alias{sits_encode.default}
\title{Encode time series or data cubes}
\usage{
sits_encode(data, encoder, ...)

\method{sits_encode}{tbl_df}(data, encoder, ...)

\method{sits_encode}{derived_cube}(data, encoder, ...)

\method{sits_encode}{default}(data, encoder, ...)
}
\arguments{
\item{data}{Input data to be encoded. Either a set of time series
(tibble of class \code{"sits"}) or a regular raster data cube
(tibble of class \code{"raster_cube"}).}

\item{encoder}{Torch model trained by
\code{\link[sits]{sits_pre_train}}.}

\item{...}{Additional arguments passed to the corresponding
\code{sits_encode} method, depending on the class of \code{data}.}
}
\value{
A tibble of encoded time series (class \code{"sits"}) or an embeddings
data cube (class \code{"embeddings_cube"}), depending on the input
type.
}
\description{
Encodes time series or raster data cubes using a pre-trained deep
learning encoder created by
\code{\link[sits]{sits_pre_train}}.

The function dispatches on the class of \code{data} and supports two
main workflows:
\itemize{
  \item For time series input (objects of class \code{"sits"}), the
  method \code{\link[sits]{sits_encode.sits}} is used. The output is a
  tibble with the original samples augmented by their encoded
  representations.
  \item For raster data cubes (objects of class
  \code{"raster_cube"}), the method
  \code{\link[sits]{sits_encode.raster_cube}} is used. The output is an
  embeddings cube, with the same tiling as the input cube and one band
  per embedding dimension.
}
}
\note{
A typical \pkg{sits} encoding workflow includes:
\enumerate{
  \item Selecting an ARD image collection with
    \code{\link[sits]{sits_cube}}.
  \item Optionally copying data locally using
    \code{\link[sits]{sits_cube_copy}}.
  \item Creating a regular data cube with
    \code{\link[sits]{sits_regularize}}.
  \item Optionally deriving additional indices with
    \code{\link[sits]{sits_apply}}.
  \item Extracting time series with
    \code{\link[sits]{sits_get_data}}.
  \item Training an encoder using
    \code{\link[sits]{sits_pre_train}}.
  \item Encoding data with \code{sits_encode}.
}

Currently supported pre-trained encoders include:
\itemize{
  \item Masked autoencoder:
    \code{\link[sits]{sits_mae}}.
  \item Barlow Twins self-supervised encoder:
    \code{\link[sits]{sits_barlow_twins}}.
  \item Contrastive learning (triplet-based) encoder:
    \code{\link[sits]{sits_contrastive_learning}}.
}

For detailed examples and usage, see the \pkg{sits} documentation at
\url{https://e-sensing.github.io/sitsbook/}.
}
\author{
Alexandre Assuncao, \email{alexcarssuncao@gmail.com}

Rolf Simoes, \email{rolfsimoes@gmail.com}
}
