% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_filters.R
\name{sits_envelope}
\alias{sits_envelope}
\title{Envelope filter}
\usage{
sits_envelope(data = NULL, operations = "UULL", bands_suffix = "env")
}
\arguments{
\item{data}{A tibble with time series data and metadata.}

\item{operations}{A character sequence for the sequence operations.
("U" for upper filter, "L" for lower filter).}

\item{bands_suffix}{Suffix of the resulting data.}
}
\value{
A tibble with filtered time series values.
}
\description{
This function computes the envelope of a time series using the
streaming algorithm proposed by Lemire (2009).
This functions calls `dtwclust::compute_envelope` function.
}
\examples{
\donttest{
# Read a set of samples of forest/non-forest in Amazonia
# This is an area full of clouds
data(prodes_226_064)
# Select the NDVI band of the first point
point_ndvi.tb <- sits_select_bands(prodes_226_064[1,], ndvi)
# Apply the envelope filter
point_env.tb <- sits_envelope(point_ndvi.tb)
# Merge the filtered with the raw data
point2.tb <- sits_merge (point_ndvi.tb, point_env.tb)
# Plot the result
plot(point2.tb)
}
}
\author{
Rolf Simoes, \email{rolf.simoes@inpe.br}
}
