% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_kohonen.R
\name{sits_evaluate_cluster}
\alias{sits_evaluate_cluster}
\title{Cluster evaluation}
\usage{
sits_evaluate_cluster(som_cluster)
}
\arguments{
\item{som_cluster}{Kohonen map and associated information created by \code{\link[sits]{sits_som_map}}}
}
\value{
Returns the confusion matrix and a table with percentage of mixture between the clusters.
}
\description{
This function extracts metrics about the clusters calculating
the percentage of mixture between a cluster and others.
}
\examples{
\donttest{
# Produce a Kohonen map for the time series samples
som_cluster <- sits_som_map(prodes_226_064)
# Extract metrics about the clusters
confusion_by_cluster <- sits_evaluate_cluster(som_cluster)
# Show confusion matrix
confusion_matrix <- confusion_by_cluster$confusion_matrix
}
}
\author{
Lorena Santos, \email{lorena.santos@inpe.br}
}
