% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_kohonen.R
\name{sits_evaluate_samples}
\alias{sits_evaluate_samples}
\title{Evaluate samples}
\usage{
sits_evaluate_samples(data.tb, time_series, grid_xdim = 5, grid_ydim = 5,
  rlen = 100, alpha = c(1), radius = 6, distance = "euclidean",
  iterations = 1, mode = "online")
}
\arguments{
\item{data.tb}{a SITS tibble with info of samples}

\item{time_series}{the time series extracted from SITS tibble}

\item{grid_xdim}{x dimension of the SOM grid (default = 5)}

\item{grid_ydim}{y dimension of the SOM grid}

\item{rlen}{number of times the complete data set will be presented to the SOM grid}

\item{alpha}{learning rate, a vector of two numbers indicating the amount of change.}

\item{radius}{the radius of kohonen map.}

\item{distance}{the similiraty measure (Euclidean or DTW)}

\item{iterations}{the number of time to run the SOM cluster}
}
\value{
returns a SITS tibble with a new column of label and a table with information about the sammples
}
\description{
This function evaluate the samples and extract metrics from SOM cluster.
The user can choose the amount of time that the kohonen process will be run.
This function allows check if the sample reliable or not
}
\author{
Lorena Santos, \email{lorena.santos@inpe.br}
}
