% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_table.R
\name{sits_filter}
\alias{sits_filter}
\title{Filter bands on a SITS table}
\usage{
sits_filter(data.tb, ...)
}
\arguments{
\item{data.tb}{a sits table with the time series of the selected bands}

\item{...}{`name=value` pairs expressions, where `name` is any SITS column name.
See `dplyr::filter` help for more details.}
}
\value{
result.tb   a tibble in SITS format with the selected bands
}
\description{
returns a sits table with the filtered bands
}
\author{
Rolf Simoes, \email{rolf.simoes@inpe.br}
}
