% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_filters.R
\name{sits_filter}
\alias{sits_filter}
\title{General function for filtering}
\usage{
sits_filter(data.tb, filter = sits_whittaker())
}
\arguments{
\item{data.tb}{Set of time series}

\item{filter}{Filter to be applied to the data.}
}
\value{
A set of filtered time series
}
\description{
Given a tibble with a set of time series, filter the tibble with one of
the available filtering algorithms:
}
\examples{
\donttest{
# Test the diffeent filters
sits_plot(point_ndvi)

# Apply the cloud filter
# Merge the filtered with the raw data
# Plot the result
point_ndvi \%>\%
      sits_filter(filter = sits_cloud_filter()) \%>\%
      sits_merge (point_ndvi, .) \%>\%
      sits_plot ()


# Apply the envelope filter
# Merge the filtered with the raw data
# Plot the result
point_ndvi \%>\%
      sits_filter(filter = sits_envelope()) \%>\%
      sits_merge (point_ndvi, .) \%>\%
      sits_plot ()

# Apply the ARIMA filter
# Merge the filtered with the raw data
# Plot the result
point_ndvi \%>\%
      sits_filter(filter = sits_ndvi_arima_filter()) \%>\%
      sits_merge (point_ndvi, .) \%>\%
      sits_plot ()

# Apply the Whittaker smoother
# Merge the filtered with the raw data
# Plot the result
point_ndvi \%>\%
      sits_filter(filter = sits_whittaker (lambda = 3.0)) \%>\%
      sits_merge (point_ndvi, .) \%>\%
      sits_plot ()

# Apply the Savitsky-Golay smoother
# Merge the filtered with the raw data
# Plot the result
point_ndvi \%>\%
      sits_filter(filter = sits_sgolay()) \%>\%
      sits_merge (point_ndvi, .) \%>\%
      sits_plot ()

# interpolate three times more points
# find out how many time instances are there in the time series
n_times <- NROW(sits_time_series(point_ndvi))
point_int <- sits_filter(point_ndvi, sits_linear_interp(n = 3*n_times))
# plot the result
sits_plot (point_int)

}
}
\author{
Rolf Simoes, \email{rolf.simoes@inpe.br}

Gilberto Camara, \email{gilberto.camara@inpe.br}
}
