% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_filter.R
\name{sits_filter}
\alias{sits_filter}
\title{Filter a set of satellite image time series}
\usage{
sits_filter(data.tb = NULL, type = "whittaker", bands = NULL,
  lambda = 0.5, window_size = 1)
}
\arguments{
\item{data.tb}{tibble - a SITS table with the list of time series to be plotted}

\item{type}{string - the type of plot to be generated}

\item{bands}{a vector of strings with band's names.}

\item{lambda}{double   - the smoothing factor to be applied}

\item{window_size}{an integer informing the window size for envelop calculation. See compute_envelop details.}
}
\value{
data.tb    tibble - the input SITS table (useful for chaining functions)
}
\description{
\code{sits_filter} filters one or more time series from a SITS table
}
\details{
This function provides access to different types of filters
}
