% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_filters.R
\name{sits_filter}
\alias{sits_filter}
\title{General function for filtering}
\usage{
sits_filter(data, filter = sits_whittaker())
}
\arguments{
\item{data}{Set of time series}

\item{filter}{Filter to be applied to the data.}
}
\value{
A set of filtered time series
}
\description{
Given a set of time series, filter them with one of
the available filtering algorithms:
}
\examples{
\donttest{
# Test the diffeent filters
plot(point_ndvi)

# Apply the cloud filter
# Merge the filtered with the raw data
# Plot the result
point_ndvi \%>\%
      sits_filter(filter = sits_cloud_removal(band_suffix = "cf")) \%>\%
      sits_merge (point_ndvi, .) \%>\%
      plot()


# Apply the envelope filter
# Merge the filtered with the raw data
# Plot the result
point_ndvi \%>\%
      sits_filter(filter = sits_envelope(band.suffix = "env")) \%>\%
      sits_merge (point_ndvi, .) \%>\%
      plot()

# Apply the ARIMA filter
# Merge the filtered with the raw data
# Plot the result
point_ndvi \%>\%
      sits_filter(filter = sits_ndvi_arima(band_suffix = "arima")) \%>\%
      sits_merge (point_ndvi, .) \%>\%
      plot()

# Apply the Whittaker smoother
# Merge the filtered with the raw data
# Plot the result
point_ndvi \%>\%
      sits_filter(filter =
                  sits_whittaker (lambda = 3.0, band_suffix = "whit")) \%>\%
      sits_merge (point_ndvi, .) \%>\%
      plot()

# Apply the Savitsky-Golay smoother
# Merge the filtered with the raw data
# Plot the result
point_ndvi \%>\%
      sits_filter(filter = sits_sgolay(band_suffix = "sg")) \%>\%
      sits_merge (point_ndvi, .) \%>\%
      plot()

# interpolate three times more points
# find out how many time instances are there in the time series
n_times <- NROW(sits_time_series(point_ndvi))
point_int <- sits_filter(point_ndvi, sits_linear_interp(n = 3*n_times))
# plot the result
plot(point_int)

}
}
\author{
Rolf Simoes, \email{rolf.simoes@inpe.br}

Gilberto Camara, \email{gilberto.camara@inpe.br}
}
