% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_filter.R
\name{sits_filter}
\alias{sits_filter}
\title{Filter a set of satellite image time series}
\usage{
sits_filter(data.tb = NULL, type = "whittaker", lambda = 0.5,
  window_size = 1)
}
\arguments{
\item{data.tb}{tibble - a SITS table with the list of time series to be plotted}

\item{type}{string - the type of filter to be used}

\item{lambda}{double   - the smoothing factor to be applied for Whittaker smoothing}

\item{window_size}{an integer informing the window size for envelop calculation. See compute_envelop details.}
}
\value{
output.tb     a SITS table with the filtered time series
}
\description{
This function provides access to different types of filters
}
\author{
Rolf Simoes, \email{rolf.simoes@inpe.br}

Gilberto Camara, \email{gilberto.camara@inpe.br}
}
