% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_filters.R
\name{sits_filter}
\alias{sits_filter}
\title{General function for filtering}
\usage{
sits_filter(data, filter = sits_whittaker())
}
\arguments{
\item{data}{Set of time series}

\item{filter}{Filter to be applied to the data.}
}
\value{
A set of filtered time series
}
\description{
Given a set of time series, filter them with one of
the available filtering algorithms:

\itemize{
 \item{Whittaker smoother - see \code{\link{sits_whittaker}}}
 \item{Savitsky-Golay filter - see \code{\link{sits_sgolay}}}
 \item{Envelope filter - see \code{\link{sits_envelope}}}
 \item{Interpolation filter - see \code{\link{sits_interp}}}
 \item{Kalman filter - see \code{\link{sits_kalman}}}
}
}
\author{
Rolf Simoes, \email{rolf.simoes@inpe.br}

Gilberto Camara, \email{gilberto.camara@inpe.br}
}
