% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_filters.R
\name{sits_filter}
\alias{sits_filter}
\title{Filter time series with smoothing filter}
\usage{
sits_filter(data, filter = sits_whittaker())
}
\arguments{
\item{data}{Time series or matrix.}

\item{filter}{Filter function to be applied.}
}
\description{
Applies a filter to all bands, using a filter function
 such as `sits_whittaker()` or `sits_sgolay()`.
}
