% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_machine_learning.R
\name{sits_formula_linear}
\alias{sits_formula_linear}
\title{Define a linear formula for classification models}
\usage{
sits_formula_linear(predictors_index = -2:0)
}
\arguments{
\item{predictors_index}{Index of the valid columns
whose names are used to compose formula (default: -2:0).}
}
\value{
A function that computes a valid formula.
}
\description{
Provides a symbolic description of a fitting model.
Tells the model to do a linear transformation of the input values.
The `predictors_index` parameter informs the positions of fields
corresponding to formula independent variables.
If no value is given,  that all fields will be used as predictors.
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}

Alexandre Ywata de Carvalho, \email{alexandre.ywata@ipea.gov.br}

Rolf Simoes, \email{rolf.simoes@inpe.br}
}
