% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_machine_learning.R
\name{sits_formula_logref}
\alias{sits_formula_logref}
\title{Train SITS classifiction models}
\usage{
sits_formula_logref(predictors_index = -4:0)
}
\arguments{
\item{predictors_index}{the index of the valid columns whose names are used to compose formula (default: NULL)}
}
\value{
result_fun       a function that computes a valid formula
}
\description{
A function to be used as a symbolic description of some fitting models such as svm.
`predictors_index` parameter informs the positions of `tb` fields corresponding to formula independent variables.
If no value is given, the default is NULL, a value indicating that all fields will be used as predictors.
}
\author{
Alexandre Xavier Ywata de Carvalho, \email{alexandre.ywata@ipea.gov.br}

Rolf Simoes, \email{rolf.simoes@inpe.br}
}
