% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_machine_learning.R
\name{sits_formula_smooth}
\alias{sits_formula_smooth}
\title{Define a smoothing formula for classification models}
\usage{
sits_formula_smooth(predictors_index = -2:0)
}
\arguments{
\item{predictors_index}{Index of the valid columns whose names are used to compose formula (default: NULL).}
}
\value{
A function that computes a valid formula.
}
\description{
A function to be used as a symbolic description of some fitting models such as gam.
This function instructs the model to do a smoothing transformation (via splines) of the input values.
`predictors_index` parameter informs the positions of `tb` fields corresponding to formula independent variables.
If no value is given, the default is NULL, a value indicating that all fields will be used as predictors.
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
