% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_import.R
\name{sits_fromCSV}
\alias{sits_fromCSV}
\title{Obtain timeSeries from WTSS server, based on a CSV file.}
\usage{
sits_fromCSV(csv_file, URL, coverage, bands, n_max = Inf,
  ignore_dates = FALSE)
}
\arguments{
\item{csv_file}{string  - name of a CSV file with information <id, latitude, longitude, from, end, label>}

\item{URL}{string - the URL of WTSS (Web Time Series Service)}

\item{coverage}{string - the name of the coverage to be retrieved}

\item{bands}{string vector - the names of the bands to be retrieved}

\item{n_max}{integer - the maximum number of samples to be read}

\item{ignore_dates}{whether to use the dates of the coverage and not those specified in the file}
}
\value{
data.tb        a SITS tibble
}
\description{
reads descriptive information about a set of
spatio-temporal locations from a CSV file. Then, it uses the WTSS time series service
to retrieve the time series, and stores the time series on a SITS tibble for later use.
The CSV file should have the following column names:
"longitude", "latitude", "start_date", "end_date", "label"
}
\examples{
\donttest{
#' # Read a set of points defined in a CSV file from a WTSS server
csv_file <- system.file ("extdata/samples/samples_import.csv", package = "sits")
points.tb <- sits_getdata (file = csv_file, URL = "http://www.dpi.inpe.br/tws/wtss",
             coverage = "mod13q1_512")
}
}
