% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_csv.R
\name{sits_fromCSV}
\alias{sits_fromCSV}
\title{Obtain timeSeries from time series server, based on a CSV file.}
\usage{
sits_fromCSV(csv_file, service = "WTSS", coverage = "mod13q1_512",
  bands = NULL, prefilter = "1", n_save = 0, n_max = Inf)
}
\arguments{
\item{csv_file}{string  - name of a CSV file with information <id, latitude, longitude, from, end, label>}

\item{service}{string - name of the time series service (options are "WTSS" or "SATVEG")}

\item{coverage}{string - the name of the coverage to be retrieved}

\item{bands}{string vector - the names of the bands to be retrieved}

\item{prefilter}{string ("0" - none, "1" - no data correction, "2" - cloud correction, "3" - no data and cloud correction)}

\item{n_save}{number of samples to save as intermediate files (used for long reads)}

\item{n_max}{the maximum number of samples to be read}
}
\value{
data.tb        a SITS tibble
}
\description{
reads descriptive information about a set of
spatio-temporal locations from a CSV file. Then, it uses the WTSS time series service
to retrieve the time series, and stores the time series on a SITS tibble for later use.
The CSV file should have the following column names:
"longitude", "latitude", "start_date", "end_date", "label"
}
\examples{
\donttest{
#' # Read a set of points defined in a CSV file from a WTSS server
csv_file <- system.file ("extdata/samples/samples_import.csv", package = "sits")
points.tb <- sits_fromCSV (file = csv_file)
}
}
