% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_fromCSV.R
\name{sits_fromCSV}
\alias{sits_fromCSV}
\title{Obtain timeSeries from WTSS server, based on a CSV file.}
\usage{
sits_fromCSV(csv_file, n_max = Inf)
}
\arguments{
\item{csv_file}{string  - name of a CSV file with information <id, latitude, longitude, from, end, label>}

\item{n_max}{integer - the maximum number of samples to be read}
}
\value{
table     tibble  - a SITS table
}
\description{
\code{sits_fromCSV} reads descriptive information about a set of
spatio-temporal locations from a CSV file. Then, it uses the WTSS time series service
to retrieve the time series, and stores the time series on a SITS table for later use.
}
\details{
#' The CSV file should have the following column names:
"longitude", "latitude", "start_date", "end_date", "label"

A sits table has the metadata and data for each time series
<longitude, latitude, start_date, end_date, label, coverage, time_series>
}
\examples{
sits_fromCSV ("mysamples.csv", n_max = 5)
}
\seealso{
Other SITS main functions: \code{\link{sits_fromClusters}},
  \code{\link{sits_fromWTSS}}, \code{\link{sits_merge}}
}
\keyword{SITS}

