% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_getdata.R
\name{sits_fromJSON}
\alias{sits_fromJSON}
\title{Obtain timeSeries from a JSON file.}
\usage{
sits_fromJSON(file)
}
\arguments{
\item{file}{string  - name of a JSON file with sits data and metadata}
}
\value{
data.tb   a SITS tibble
}
\description{
reads a set of data and metadata for satellite image time series from a JSON file
}
