% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_JSON.R
\name{sits_fromJSON}
\alias{sits_fromJSON}
\title{Retrieve a set of time series from a JSON file}
\usage{
sits_fromJSON(json_file)
}
\arguments{
\item{json_file}{string - name of the JSON file}
}
\value{
table     tibble - a new SITS table
}
\description{
\code{sits_from_JSON} gets a SITS database from a JSON file
}
\details{
The set of time series from the SITS database can be saved in JSON format
and later retrieved for further use
}
\examples{
table <- sits_fromJSON ("myjson_file.json")
}
\seealso{
Other STIS main functions: \code{\link{sits_align}},
  \code{\link{sits_centroids}},
  \code{\link{sits_dendogram}},
  \code{\link{sits_plot_together}},
  \code{\link{sits_plot}}, \code{\link{sits_toJSON}}
}
\keyword{STIS}
