% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_fromSHP.R
\name{sits_fromSHP}
\alias{sits_fromSHP}
\title{Obtain timeSeries from WTSS server, based on a SHP file.}
\usage{
sits_fromSHP(shp_file,
  crs = "+proj=longlat +ellps=WGS84 +datum=WGS84 +no_defs")
}
\arguments{
\item{shp_file}{string  - name of a SHP file which provides the boundaries of a region of interest}

\item{crs}{string  - the coordinate reference system used by the shapefile}
}
\value{
table     tibble  - a SITS table
}
\description{
\code{sits_fromSHP} reads a shapefile and retrieves a SITS table
containing time series from a coverage that are inside the SHP file.
The script uses the WTSS service, taking information about coverage, spatial and
temporal resolution from the WTSS configuration.
}
\examples{
sits_fromSHP ("municipality.shp")
}
\seealso{
Other SITS main functions: \code{\link{sits_fromCSV}},
  \code{\link{sits_fromClusters}},
  \code{\link{sits_fromWTSS}}, \code{\link{sits_merge}}
}
\keyword{SITS}
