% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_fromTWDTW.R
\name{sits_fromTWDTW}
\alias{sits_fromTWDTW}
\title{Transform patterns from TWDTW format
\code{sits_fromTWDTW} reads a set of TWDTW patterns
transforms them into a SITS table}
\usage{
sits_fromTWDTW(patterns)
}
\arguments{
\item{patterns}{- a TWDTW object containing a set of patterns to be used for classification}
}
\value{
sits.tb  - a SITS table containing the patterns
}
\description{
Transform patterns from TWDTW format
\code{sits_fromTWDTW} reads a set of TWDTW patterns
transforms them into a SITS table
}

