% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_getdata.R
\name{sits_fromWTSS}
\alias{sits_fromWTSS}
\title{Obtain one timeSeries from WTSS server and load it on a sits table}
\usage{
sits_fromWTSS(longitude, latitude, start_date, end_date, label, wtss.obj, cov,
  bands, ignore_dates = FALSE)
}
\arguments{
\item{longitude}{double - the longitude of the chosen location}

\item{latitude}{double - the latitude of the chosen location}

\item{start_date}{date - the start of the period}

\item{end_date}{date - the end of the period}

\item{label}{string - the label to attach to the time series (optional)}

\item{wtss.obj}{an R object that represents the WTSS server}

\item{cov}{a list containing information about the coverage from which data is to be recovered}

\item{bands}{list of string - a list of the names of the bands of the coverage}

\item{ignore_dates}{whether to use the dates of the coverage and not those specified in the file}
}
\value{
data.tb        tibble  - a SITS table
}
\description{
Returns one set of time series provided by a WTSS server
Given a location (lat/long), and start/end period, and the WTSS server information
retrieve a time series and include it on a stis table.
A Web Time Series Service (WTSS) is a light-weight service that
retrieves one or more time series in JSON format from a data base.
}
\references{
Lubia Vinhas, Gilberto Queiroz, Karine Ferreira, Gilberto Camara,
Web Services for Big Earth Observation Data.
In: XVII Brazilian Symposium on Geoinformatics, 2016, Campos do Jordao.
Proceedings of GeoInfo 2016. Sao Jose dos Campos: INPE/SBC, 2016. v.1. p.166-177
}
