% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_getdata.R
\name{sits_fromZOO}
\alias{sits_fromZOO}
\title{Export data to be used to the zoo format}
\usage{
sits_fromZOO(ts.tb, band)
}
\arguments{
\item{ts}{a zoo time series}
}
\value{
ts       a time series in SITS format
}
\description{
Converts data from a SITS table to an instance of a zoo series,
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
