% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_machine_learning.R
\name{sits_gbm}
\alias{sits_gbm}
\title{Train a SITS classification model with a gradient boosting machine}
\usage{
sits_gbm(distances.tb = NULL, formula = sits_formula_logref(),
  distribution = "multinomial", n.trees = 500, interaction.depth = 2,
  shrinkage = 0.001, cv.folds = 5, n.cores = 1, ...)
}
\arguments{
\item{distances.tb}{a time series with a set of distance measures for each training sample}

\item{formula}{a symbolic description of the model to be fit. SITS offers a set of such formulas (default: sits_formula_logref)}

\item{distribution}{The name of the distribution - use "multinomial" for classification}

\item{n.trees}{Number of trees to fit. This should not be set to too small a number,
to ensure that every input row gets predicted at least a few times. (default: 5000)}

\item{interaction.depth}{The maximum depth of variable interactions. 1 implies an additive model, 2 implies a model with up to 2-way interactions.}

\item{shrinkage}{a shrinkage parameter applied to each tree in the expansion.
Also known as the learning rate or step-size reduction.}

\item{cv.folds}{number of cross-validations to run}

\item{n.cores}{number of cores to run}

\item{...}{other parameters to be passed to `gbm::gbm` function}
}
\value{
result          a model function to be passed in sits_predict or an function prepared that can be called further to compute multinom training model
}
\description{
This function implements the generalized boosted modeling framework.
Boosting is the process of iteratively adding basis functions in a greedy fashion
so that each additional basis function further reduces the selected loss function.
This function is a front-end to the "gbm" method in the "gbm" package.
Please refer to the documentation in that package for more details.
}
\examples{
\donttest{
# Retrieve the set of samples for the Mato Grosso region (provided by EMBRAPA)
samples.tb <- readRDS(system.file("extdata/time_series/embrapa_mt.rds", package = "sits"))
samples.tb <- sits_select (samples.tb, bands = c("ndvi", "evi", "nir"))
# get a point with a 16 year time series
point.tb <- readRDS(system.file("extdata/time_series/point.rds", package = "sits"))
point.tb <- sits_select (point.tb, bands = c("ndvi", "evi", "nir"))
# classify the point
class.tb <- sits_classify (point.tb, samples.tb,
       ml_method = sits_gbm())
}
}
\author{
Alexandre Xavier Ywata de Carvalho, \email{alexandre.ywata@ipea.gov.br}

Rolf Simoes, \email{rolf.simoes@inpe.br}

Gilberto Camara, \email{gilberto.camara@inpe.br}
}
