% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_plot_distances.R
\name{sits_geo_dist}
\alias{sits_geo_dist}
\title{Compute the minimum distances among samples and prediction points.}
\usage{
sits_geo_dist(samples, roi = NULL, n = 1000)
}
\arguments{
\item{roi}{A `sf` object (polygon).}

\item{n}{Maximum number of samples to consider.}

\item{samples_tb}{A `sits` tibble.}
}
\value{
A tibble.
}
\description{
Compute the minimum distances among samples and samples to prediction
points.
}
\author{
Alber Sanchez, \email{alber.sanchez@inpe.br}

Rolf Simoes, \email{rolf.simoes@inpe.br}

Felipe Carvalho, \email{felipe.carvalho@inpe.br}

Gilberto Camara, \email{gilberto.camara@inpe.br}
}
