% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_config.R
\name{sits_get_color}
\alias{sits_get_color}
\alias{.sits_get_color}
\title{Retrieve the color associated to a class}
\usage{
.sits_get_color(label)
}
\arguments{
\item{label}{a class label}
}
\description{
retrieve the color associated a class label
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
