% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_get_data.R
\name{sits_get_data.csv_raster_cube}
\alias{sits_get_data.csv_raster_cube}
\title{Obtain time series from brick based on CSV file}
\usage{
\method{sits_get_data}{csv_raster_cube}(cube, file, ..., bands = NULL)
}
\arguments{
\item{cube}{Data cube from where data is to be retrived.}

\item{file}{File with information on the data to be retrieved}

\item{...}{Other parameters to be passed for specific types}

\item{bands}{Bands to be retrieved (optional)}
}
\value{
A tibble with time series data and metadata.
}
\description{
Obtain time series from brick based on CSV file
}
