% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_get_data.R
\name{sits_get_data.csv_raster_cube}
\alias{sits_get_data.csv_raster_cube}
\title{Obtain time series from brick based on CSV file}
\usage{
\method{sits_get_data}{csv_raster_cube}(
  cube,
  file,
  ...,
  bands = NULL,
  impute_fn = sits_impute_linear(),
  .n_pts_csv = NULL
)
}
\arguments{
\item{cube}{Data cube from where data is to be retrieved.}

\item{file}{File with information on the data to be retrieved.}

\item{...}{Other parameters to be passed for specific types.}

\item{bands}{Bands to be retrieved (optional).}

\item{impute_fn}{Imputation function for NA values.}

\item{.n_pts_csv}{Number of points from CSV file to be retrieved.}
}
\value{
A tibble with time series data and metadata.
}
\description{
Obtain time series from brick based on CSV file
}
\examples{
#' Read a CSV in a Raster Brick
# define the file that has the raster brick
ndvi_file <- c(system.file("extdata/raster/mod13q1/sinop-ndvi-2014.tif",
               package = "sits"))

evi_file <- c(system.file("extdata/raster/mod13q1/sinop-evi-2014.tif",
              package = "sits"))

# define the data cube based on files and known timeline
sinop_2014 <- sits_cube(
    type = "BRICK",
    name = "sinop-2014",
    timeline = timeline_2013_2014,
    satellite = "TERRA",
    sensor = "MODIS",
    bands = c("ndvi", "evi"),
    files = c(ndvi_file, evi_file)
)
# read data from a CSV file
csv_file <- system.file("extdata/samples/samples_sinop_crop.csv",
    package = "sits"
)
points <- sits_get_data(sinop_2014, file = csv_file)

# show the points

plot(points)
}
