% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_get_data.R
\name{sits_get_data.raster_cube}
\alias{sits_get_data.raster_cube}
\title{Obtain time series from raster cube}
\usage{
\method{sits_get_data}{raster_cube}(
  cube,
  file = NULL,
  ...,
  longitude = NULL,
  latitude = NULL,
  start_date = NULL,
  end_date = NULL,
  bands = NULL,
  label = "NoClass",
  impute_fn = sits_impute_linear()
)
}
\arguments{
\item{cube}{Data cube from where data is to be retrieved.}

\item{file}{File with information on the data to be retrieved.}

\item{...}{Other parameters to be passed for specific types}

\item{longitude}{Longitude of the chosen location.}

\item{latitude}{Latitude of the chosen location.}

\item{start_date}{Start of the interval for the time series
in "YYYY-MM-DD" format (optional).}

\item{end_date}{End of the interval for the time series in
"YYYY-MM-DD" format (optional).}

\item{bands}{Bands to be retrieved (optional).}

\item{label}{Label to be assigned to the time series (optional).}

\item{impute_fn}{Imputation function for NA values.}
}
\value{
A tibble with time series data and metadata.
}
\description{
Obtain time series from raster cube
}
\examples{
# Read a point in a Raster Brick
# define the file that has the raster brick
files <- c(system.file("extdata/raster/mod13q1/sinop-crop-ndvi.tif",
    package = "sits"
))
# define the timeline
data(timeline_modis_392)
# create a data cube based on the information about the files
raster_cube <- sits_cube(
    type = "BRICK", satellite = "TERRA",
    sensor = "MODIS", name = "Sinop-crop",
    timeline = timeline_modis_392,
    bands = c("NDVI"), files = files
)

# read the time series of the point from the raster
point_ts <- sits_get_data(raster_cube,
    longitude = -55.554,
    latitude = -11.525
)
plot(point_ts)
}
