% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_get_data.R
\name{sits_get_data}
\alias{sits_get_data}
\title{Get time series from data cubes and cloud services}
\usage{
sits_get_data(
  cube,
  file = NULL,
  samples = NULL,
  longitude = NULL,
  latitude = NULL,
  start_date = NULL,
  end_date = NULL,
  label = "NoClass",
  bands = NULL,
  impute_fn = sits_impute_linear(),
  shp_attr = NULL,
  .n_pts_csv = NULL,
  .n_shp_pol = 30,
  multicores = 1
)
}
\arguments{
\item{cube}{Data cube from where data is to be retrieved.}

\item{file}{File with information on the data to be retrieved.}

\item{samples}{Data.frame with samples location in spacetime.}

\item{longitude}{Longitude of the chosen location.}

\item{latitude}{Latitude of the chosen location.}

\item{start_date}{Start of the interval for the time series
in "YYYY-MM-DD" format (optional).}

\item{end_date}{End of the interval for the time series in
"YYYY-MM-DD" format (optional).}

\item{label}{Label to be assigned to the time series (optional).}

\item{bands}{Bands to be retrieved (optional).}

\item{impute_fn}{Imputation function for NA values.}

\item{shp_attr}{Attribute in the shapefile to be used
as a polygon label.}

\item{.n_pts_csv}{Number of points from CSV file to be retrieved.}

\item{.n_shp_pol}{Number of samples per polygon to be read
(for POLYGON or MULTIPOLYGON shapefile).}

\item{multicores}{Number of threads to process the time series.}
}
\value{
A tibble with the metadata and data for each time series
<longitude, latitude, start_date, end_date, label, cube, time_series>.
}
\description{

}
\examples{
\donttest{
# -- Read a point in a raster data cube

# Create a data cube based on files
data_dir <- system.file("extdata/raster/mod13q1", package = "sits")
raster_cube <- sits_cube(
    source = "BDC",
    collection = "MOD13Q1-6",
    data_dir = data_dir,
    delim = "_",
    parse_info = c("X1", "X2", "tile", "band", "date")
)

# read the time series of the point from the raster
point_ts <- sits_get_data(raster_cube,
    longitude = -55.554,
    latitude = -11.525
)

# --- Read a set of points described by a CSV file

# read data from a CSV file
csv_file <- system.file("extdata/samples/samples_sinop_crop.csv",
    package = "sits"
)
points_csv <- sits_get_data(raster_cube, file = csv_file)
}
}
\author{
Gilberto Camara
}
