% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_get_data.R
\name{sits_get_data}
\alias{sits_get_data}
\title{Get time series from data cubes and cloud services}
\usage{
sits_get_data(
  cube,
  file = NULL,
  samples = NULL,
  longitude = NULL,
  latitude = NULL,
  start_date = NULL,
  end_date = NULL,
  label = "NoClass",
  bands = NULL,
  impute_fn = sits_impute_linear(),
  shp_attr = NULL,
  .n_shp_pol = 30,
  .shp_avg = FALSE,
  .shp_id = NULL,
  multicores = 1,
  output_dir = ".",
  progress = FALSE
)
}
\arguments{
\item{cube}{Data cube from where data is to be retrieved.}

\item{file}{File with information on the data to be retrieved.}

\item{samples}{Data.frame with samples location in spacetime.}

\item{longitude}{Longitude of the chosen location.}

\item{latitude}{Latitude of the chosen location.}

\item{start_date}{Start of the interval for the time series
in "YYYY-MM-DD" format (optional).}

\item{end_date}{End of the interval for the time series in
"YYYY-MM-DD" format (optional).}

\item{label}{Label to be assigned to the time series (optional).}

\item{bands}{Bands to be retrieved (optional).}

\item{impute_fn}{Imputation function for NA values.}

\item{shp_attr}{Attribute in the shapefile to be used
as a polygon label.}

\item{.n_shp_pol}{Number of samples per polygon to be read
(for POLYGON or MULTIPOLYGON shapefile).}

\item{.shp_avg}{Logical value to summarize samples for a same polygon.}

\item{.shp_id}{ID attribute for polygons shapefile.}

\item{multicores}{Number of threads to process the time series.}

\item{output_dir}{Directory where the time series will be saved as rds.
Default is the current path.}

\item{progress}{A logical value indicating if a progress bar
should be shown. Default is \code{FALSE}.}
}
\value{
A tibble with the metadata and data for each time series
<longitude, latitude, start_date, end_date, label, cube, time_series>.
}
\description{
Retrieve a set of time series from a data cube or from
a time series service. Data cubes and puts it in a "sits tibble".
Sits tibbles are the main structures of sits package.
They contain both the satellite image time series and their metadata.
There are four ways of specifying data to be retrieved:
\itemize{
\item{CSV file:}{Provide a CSV file with columns
"longitude", "latitude", "start_date", "end_date" and "label" for
each sample}
\item{SHP file:}{Provide a shapefile in POINT or POLYGON geometry
containing the location of the samples and an attribute to be
used as label. Also, provide start and end date for the time series.}
\item{samples:}{A data.frame with with columns
"longitude", "latitude", "start_date", "end_date" and "label" for
each sample}
\item{single point:}{Provide the values
"longitude", "latitude", "start_date", "end_date" and "label" to
obtain a time series for a spacetime location}
}
}
\note{
Please refer to the sits documentation available in
<https://e-sensing.github.io/sitsbook/> for detailed examples.
}
\examples{
if (sits_run_examples()) {
   # reading a lat/long from a local cube
   # create a cube from local files
   data_dir <- system.file("extdata/raster/mod13q1", package = "sits")
   raster_cube <- sits_cube(
     source  = "BDC",
     collection = "MOD13Q1-6",
     data_dir = data_dir,
     delim = "_",
     parse_info = c("X1", "X2", "tile", "band", "date")
  )
  point_ndvi <- sits_get_data(raster_cube,
       longitude = -55.66738,
       latitude = -11.76990
  )
  # reading samples from a cube based on a  CSV file
  csv_file <- system.file("extdata/samples/samples_sinop_crop.csv",
                            package = "sits"
  )
  points <- sits_get_data(cube = raster_cube, file = csv_file)

  # reading a shapefile from BDC (Brazil Data Cube)
  # needs a BDC access key that can be obtained
  # for free by registering in the BDC website
  if (nchar(Sys.getenv("BDC_ACCESS_KEY")) > 0 ){
      # create a data cube from the BDC
      bdc_cube <- sits_cube(
                   source = "BDC",
                   collection = "CB4_64_16D_STK-1",
                   bands = c("NDVI", "EVI"),
                   tiles = c("022024", "022025"),
                   start_date = "2018-09-01",
                   end_date = "2018-10-28"
     )
     # define a shapefile to be read from the cube
     shp_path <- system.file("extdata/shapefiles/bdc-test/samples.shp",
                             package = "sits"
                             )
     # get samples from the BDC based on the shapefile
     time_series_bdc <- sits::sits_get_data(
                  cube = bdc_cube
                  file = shp_path
     )
  }
}

}
\author{
Gilberto Camara
}
