% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_get_data.R
\name{sits_get_data}
\alias{sits_get_data}
\title{Obtain time series from different sources}
\usage{
sits_get_data(cube, file = NULL, ...)
}
\arguments{
\item{cube}{Data cube from where data is to be retrieved.}

\item{file}{File with information on the data to be retrieved.}

\item{...}{Other parameters to be passed for specific types.}
}
\value{
A tibble with the metadata and data for each time series
<longitude, latitude, start_date, end_date, label, cube, time_series>.
}
\description{
Retrieve a set of time series and puts it in a "sits tibble".
Sits tibbles are the main structures of sits package.
They contain both the satellite image time series and their metadata.
A sits tibble is a tibble with pre-defined columns that
has the metadata and data for each time series. The columns are
<longitude, latitude, start_date, end_date, label, cube, time_series>.
There are many ways of retrieving time series:
\itemize{
\item{SATVEG: }{Retrieve data from SATVEG service using a lat/long point
 (\code{\link[sits]{sits_get_data.satveg_cube}}),
 a CSV file (\code{\link[sits]{sits_get_data.csv_satveg_cube}})
 or a SHP file (\code{\link[sits]{sits_get_data.shp_satveg_cube}})}

\item{RASTER: }{Retrieve data from a RASTER cube using a lat/long point
 (\code{\link[sits]{sits_get_data.raster_cube}}),
 a CSV file (\code{\link[sits]{sits_get_data.csv_raster_cube}})
 or a SHP file (\code{\link[sits]{sits_get_data.shp_raster_cube}})}

}

The URL and other parameters for accessing the time series services
are defined in the package configuration file. This file is "config.yml".
Please see the \code{\link[sits]{sits_config}} for more information.

Before using this service, the user should create a valid description
of a data cube using the \code{\link[sits]{sits_cube}} function.
}
\references{
Lubia Vinhas, Gilberto Queiroz, Karine Ferreira, Gilberto Camara,
Web Services for Big Earth Observation Data.
In: XVII Brazilian Symposium on Geoinformatics, 2016, Campos do Jordao.
Proceedings of GeoInfo 2016. Sao Jose dos Campos: INPE/SBC, 2016. p.166-177.
}
\author{
Gilberto Camara
}
