% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_WTSS.R
\name{sits_getcovWTSS}
\alias{sits_getcovWTSS}
\title{Obtain information about one coverage of the WTSS service}
\usage{
sits_getcovWTSS(URL = "http://www.dpi.inpe.br/tws/wtss", coverage = NULL)
}
\arguments{
\item{URL}{the URL for the WTSS time series service}

\item{coverage}{the name of the coverage}
}
\value{
cov       a list with descriptive information about the coverage
}
\description{
uses the WTSS services to retrieve information about a given coverage:
 name            - coverage name
 description     - description
 detail          - more information (source)
 attributes      - spectral bands (name, description, datatype, valid_range.min, valid_range.max, scale_factor, missing_value)
 spatial_extent  - bounding box in space (xmin, ymin, xmax, ymax)
 crs             - Projection CRS
 timeline        - dates of images contained in the coverage
}
