% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_getdata.R
\name{sits_getdata}
\alias{sits_getdata}
\title{Obtain timeSeries}
\usage{
sits_getdata(file = NULL, table = NULL, longitude = NULL,
  latitude = NULL, URL = "http://www.dpi.inpe.br/tws/wtss",
  coverage = NULL, bands = NULL, n_max = Inf)
}
\arguments{
\item{file}{the name of a file with information on the data to be retrieved (options - CSV, JSON, SHP)}

\item{table}{an R object ("sits_table")}

\item{longitude}{double - the longitude of the chosen location}

\item{latitude}{double - the latitude of the chosen location)}

\item{URL}{string - the URL of WTSS (Web Time Series Service)}

\item{coverage}{string - the name of the coverage to be retrieved}

\item{bands}{string vector - the names of the bands to be retrieved}

\item{n_max}{integer - the maximum number of samples to be read (optional)}
}
\value{
data.tb        tibble  - a SITS table
}
\description{
\code{sits_getdata} reads descriptive information about a data source to retrive a set of
time series. The following options are available:
}
\details{
(a) The source is a SITS table - retrieves the metadata from the sits table and the data
from the WTSS service
(b) The source is a CSV file - retrieves the metadata from the CSV file and the data
from the WTSS service
(c) The source is a JSON file - retrieves the metadata and data from the JSON file.
(d) No source is given - it retrives the data based on <long, lat, wtss>
A sits table has the metadata and data for each time series
<longitude, latitude, start_date, end_date, label, coverage, time_series>
}
\keyword{SITS}
