% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_import.R
\name{sits_getdata}
\alias{sits_getdata}
\title{Obtain time series from different sources}
\usage{
sits_getdata(raster.tb = NULL, file = NULL, longitude = NULL,
  latitude = NULL, start_date = NULL, end_date = NULL,
  URL = "http://www.dpi.inpe.br/tws/wtss", coverage = "mod13q1_512",
  bands = NULL, label = "NoClass", n_max = Inf, ignore_dates = FALSE)
}
\arguments{
\item{raster.tb}{(optional) an STRaster  object (tibble with raster information)}

\item{file}{(optional) the name of a file with information on the data to be retrieved (options - CSV, JSON, SHP)}

\item{longitude}{double - the longitude of the chosen location}

\item{latitude}{double - the latitude of the chosen location)}

\item{start_date}{(optional) date - the start of the period}

\item{end_date}{(optional) date - the end of the period}

\item{URL}{string - the URL of WTSS (Web Time Series Service)}

\item{coverage}{string - the name of the coverage to be retrieved}

\item{bands}{(optional) vector - the names of the bands to be retrieved}

\item{label}{(optional) string - the label to be assigned to the time series}

\item{n_max}{(optional) integer - the maximum number of samples to be read}

\item{ignore_dates}{(optional) use the start and end dates from the coverage instead of the time series}
}
\value{
data.tb        a SITS tibble
}
\description{
Retrieve a set of time series. There are two main ways of retrieving time series:
using the WTSS service (Web Time Series Service) and from a Raster Brick. The Web Time Series Service (WTSS)
is a light-weight service that provides access to time series.
Please see \code{link[sits]{sits_infoWTSS}} for more information on thw WTSS service.
The following options are available:
\enumerate{
\item No input file is given - it retrieves the data and metadata based on the latitude/longitude location
and on the information provided by the WTSS server.
\item The source is a CSV file - retrieves the metadata from the CSV file and the time series
from the WTSS service.
\item The source is a SHP file - retrives all points inside the shapefile from the WTSS service.
\item The source is a RasterBrick - retrieves the point based on lat/long from the RasterBrick.
\item The source is a ZOO time series - retrieves the time series from the ZOO file, and
the user should supply the metadata.
}
<longitude, latitude, start_date, end_date, label, coverage, time_series>
}
\examples{
# Read a single lat long point from a WTSS server
point.tb <- sits_getdata (longitude = -55.50563, latitude = -11.71557,
         URL = "http://www.dpi.inpe.br/tws/wtss", coverage    = "mod13q1_512")

# Read a set of points defined in a CSV file from a WTSS server
csv_file <- system.file ("extdata/samples/samples_import.csv", package = "sits")
points.tb <- sits_getdata (file = csv_file, URL = "http://www.dpi.inpe.br/tws/wtss",
             coverage = "mod13q1_512")

# Read a point in a Raster Brick
# define the file that has the raster brick
files  <- c(system.file ("extdata/raster/mod13q1/sinop_ndvi_sample.tif", package = "sits"))
# define the timeline
timeline <- read.csv(system.file("extdata/raster/mod13q1/mod13Q1-timeline-2000-2017.csv", package = "sits"), header = FALSE)
timeline <- lubridate::as_date (timeline$V1)
# create a raster metadata file based on the information about the files
raster.tb <- sits_STRaster (files, timeline, bands = c("ndvi"), scale_factors = c(0.0001))
# read the point from the raster
point.tb <- sits_getdata(raster.tb, longitude = -55.50563, latitude = -11.71557)

}
\references{
Lubia Vinhas, Gilberto Queiroz, Karine Ferreira, Gilberto Camara,
Web Services for Big Earth Observation Data.
In: XVII Brazilian Symposium on Geoinformatics, 2016, Campos do Jordao.
Proceedings of GeoInfo 2016. Sao Jose dos Campos: INPE/SBC, 2016. v.1. p.166-177.
}
\author{
Gilberto Camara
}
