% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_ggplot.R
\name{sits_ggplot_series}
\alias{sits_ggplot_series}
\title{Plot one or more timeSeries using ggplot}
\usage{
sits_ggplot_series(tb, colors = "Dark2")
}
\arguments{
\item{tb}{a SITS table with the time series to be plotted}

\item{colors}{string - the set of Brewer colors to be used for plotting}
}
\description{
\code{sits_ggplot_series} plots a set of time series using ggplot
}
\details{
This function is used for showing the same lat/long location in
a series of time steps.
}
\keyword{SITS}

