% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_plot.R
\name{sits_ggplot_series}
\alias{sits_ggplot_series}
\title{Plot one or more timeSeries using ggplot}
\usage{
sits_ggplot_series(data.ts, plot_title, colors = "Dark2")
}
\arguments{
\item{data.ts}{the time series to be plotted (tibble in SITS format)}

\item{plot_title}{string - the plot title}

\item{colors}{string - the set of Brewer colors to be used for plotting}
}
\description{
\code{sits_ggplot_series} plots a set of  time series
}
\seealso{
Other SITS auxiliary functions: \code{\link{sits_fromCSV}},
  \code{\link{sits_fromWTSS}},
  \code{\link{sits_ggplot_together}},
  \code{\link{sits_make_plot_title}},
  \code{\link{sits_merge}}
}
\keyword{SITS}
