% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_machine_learning.R
\name{sits_glm}
\alias{sits_glm}
\title{Train a SITS classification models using generalised linear models (lasso and ridge)}
\usage{
sits_glm(distances.tb = NULL, family = "multinomial", alpha = 1,
  lambda_kfolds = 10, ...)
}
\arguments{
\item{distances.tb}{a time series with a set of distance measures for each training sample}

\item{family}{Response type. Can be either "gaussian", "binomial", "poisson", "multinomial", "cox", or "mgaussian". (default: "multinomial")}

\item{alpha}{the elasticnet mixing parameter, with 0<=alpha<=1.
Set alpha = 0.0 to obtain ridge model, and alpha = 1.0 to obtain lasso model.
(refer to `glmnet::cv.glmnet` function for more details)}

\item{lambda_kfolds}{number of folds to find best lambda parameter (default: 10)}

\item{...}{other parameters to be passed to `glmnet::cv.glmnet` function}
}
\value{
result          a model function to be passed in sits_predict
}
\description{
Use generalized liner model (glm) via penalized maximim likelihood to classify data.
These attributes are distance metrics between patterns and observations, obtained by a distance
function in SITS (either sits_distances or sits_TWTDTW_distances).
This function is a front-end to the "cv.glmnet" method in the "glmnet" package.
Please refer to the documentation in that package for more details.
}
\examples{
\donttest{
# Retrieve the set of samples for the Mato Grosso region (provided by EMBRAPA)
samples.tb <- readRDS(system.file("extdata/time_series/embrapa_mt.rds", package = "sits"))
samples.tb <- sits_select (samples.tb, bands = c("ndvi", "evi", "nir"))
# get a point with a 16 year time series
point.tb <- readRDS(system.file("extdata/time_series/point.rds", package = "sits"))
point.tb <- sits_select (point.tb, bands = c("ndvi", "evi", "nir"))
# classify the point
class.tb <- sits_classify (point.tb, samples.tb,
       ml_method = sits_glm(alpha = 1.0))
}
}
\author{
Alexandre Xavier Ywata de Carvalho, \email{alexandre.ywata@ipea.gov.br}

Rolf Simoes, \email{rolf.simoes@inpe.br}

Gilberto Camara, \email{gilberto.camara@inpe.br}
}
