% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_machine_learning.R
\name{sits_glm}
\alias{sits_glm}
\title{Train SITS classifiction models with Generalised Linear Models}
\usage{
sits_glm(distances.tb = NULL, family = "multinomial", alpha = 1,
  lambda_kfolds = 10, ...)
}
\arguments{
\item{distances.tb}{a time series with a set of distance measures for each training sample}

\item{family}{Response type. Can be either "gaussian", "binomial", "poisson", "multinomial", "cox", or "mgaussian". (default: "multinomial")}

\item{alpha}{the elasticnet mixing parameter, with 0<=alpha<=1. Set alpha = 0.0 to obtain ridge model, and alpha = 1.0 to obtain lasso model).
(refer to `glmnet::cv.glmnet` function for more details)}

\item{lambda_kfolds}{number of folds to find best lambda parameter (default: 10)}

\item{...}{other parameters to be passed to `glmnet::cv.glmnet` function}
}
\value{
result          either an model function to be passed in sits_predict or an function prepared that can be called further to compute multinom training model
}
\description{
Use generalized liner model (glm) via penalized maximim likelihood to classify data.
This function is a front-end to the "cv.glmnet" method in the "glmnet" package.
Please refer to the documentation in that package for more details.
}
\author{
Alexandre Xavier Ywata de Carvalho, \email{alexandre.ywata@ipea.gov.br}

Rolf Simoes, \email{rolf.simoes@inpe.br}
}
