% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_active_learning.R
\name{sits_increase_samples}
\alias{sits_increase_samples}
\title{Suggest high confidence samples to increase the training set.}
\usage{
sits_increase_samples(
  probs_cube,
  n = 20,
  min_margin = 0.9,
  min_dist_pixels = 10
)
}
\arguments{
\item{probs_cube}{A `sits` probability cube. See `sits_classify`.}

\item{n}{Number of suggested points per class.}

\item{min_margin}{Minimum margin of confidence to select a sample}

\item{min_dist_pixels}{Minimum distance among suggested points (in pixels).}
}
\value{
A data.frame with longitude & latitude in WGS84.
}
\description{
Suggest points for increasing the training set. These points are labelled
with high confidence so they can be added to the training set.
They need to have a satisfactory margin of confidence to be selected.
}
\author{
Alber Sanchez, \email{alber.sanchez@inpe.br}

Rolf Simoes, \email{rolf.simoes@inpe.br}

Felipe Carvalho, \email{felipe.carvalho@inpe.br}

Gilberto Camara, \email{gilberto.camara@inpe.br}
}
