% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_WTSS.R
\name{sits_infoWTSS}
\alias{sits_infoWTSS}
\title{Provides information about WTSS service}
\usage{
sits_infoWTSS(URL = "http://www.dpi.inpe.br/tws/wtss")
}
\arguments{
\item{URL}{the URL for the WTSS time series service}
}
\value{
wtss.obj       an R object containing the information about the WTSS server
}
\description{
obtains information about the WTSS server
and about the coverages.

The Web Time Series Service is a lightweight web service the allow remote access to satellite
 image time series and provides three operations:

 1. list coverages: this operation allows clients to retrieve the capabilities provided
    by any server that implements WTSS. It returns a list of coverage
    names available in a server instance.

 2. describe coverage: this operation returns the metadata for a given coverage
    identified by its name. It includes its range in the spatial and temporal dimensions.

 3. time series: this operation requests the time series of values of a coverage attribute
   at a given location.
}
\examples{
\donttest{
# Set the URL of the WTSS service
URL <- "http://www.dpi.inpe.br/tws/wtss"
# Obtain information about the coverages available
wtss_inpe <- sits_infoWTSS(URL)
}
}
\author{
Gilberto Camara
}
