% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_init.R
\name{sits_init}
\alias{sits_init}
\alias{.onAttach}
\title{Initialization commands for SITS package}
\usage{
.onAttach(lib, pkg)
}
\description{
provides the information for the dependencies of the sits package
}
