% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_filters.R
\name{sits_interp}
\alias{sits_interp}
\title{Interpolate values in time series}
\usage{
sits_interp(
  data,
  fun = stats::approx,
  n = 2 * length(sits_timeline(data)),
  ...
)
}
\arguments{
\item{data}{A tibble with time series data and metadata.}

\item{fun}{Interpolation function (by default, stats::approx())}

\item{n}{Number of time series elements to be created
between start date and end date.}

\item{...}{Additional parameters to be used by the fun function.}
}
\value{
A tibble with interpolated samples.
}
\description{
Computes the interpolated bands using a user-defined
function (by default the R base function approx)
}
\examples{
# Retrieve a time series with values of NDVI
point_ndvi <- sits_select(point_mt_6bands, bands = "NDVI")
# find out how many time instances are there
n_times <- length(sits_timeline(point_ndvi))
# interpolate three times more points
point_int.tb <- sits_interp(point_ndvi, fun = stats::spline, n = 3 * n_times)
# plot the result
plot(point_int.tb)
}
\author{
Rolf Simoes, \email{rolf.simoes@inpe.br}
}
