% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_filters.R
\name{sits_interp}
\alias{sits_interp}
\title{Interpolation function of the time series of a sits_tibble}
\usage{
sits_interp(data = NULL, fun = stats::approx, n = base::length, ...)
}
\arguments{
\item{data}{A tibble with time series data and metadata.}

\item{fun}{Interpolation function.}

\item{n}{Number of time series elements to be created
between start date and end date.
When a class function is passed to `n`,
it is evaluated with each band time series as
an argument, e.g. n(band) (default: `length` function).}

\item{...}{Additional parameters to be used by the fun function.}
}
\value{
A tibble with same samples and the new bands.
}
\description{
Computes the linearly interpolated bands
              using the R base function approx.
}
\examples{
# Retrieve a time series with values of NDVI
data(point_ndvi)
# find out how many time instances are there
n_times <- NROW(sits_time_series(point_ndvi))
# interpolate three times more points
point_int.tb <- sits_interp(point_ndvi, fun = stats::spline, n = 3 * n_times)
# plot the result
plot(point_int.tb)
}
\author{
Rolf Simoes, \email{rolf.simoes@inpe.br}
}
