% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_segmentation.R
\name{sits_join_segments}
\alias{sits_join_segments}
\title{Return segments from a classified set of time series}
\usage{
sits_join_segments(data, segments)
}
\arguments{
\item{data}{A sits tibble with predicted values}

\item{segments}{Output coordinate reference system.}
}
\value{
An sf object of polygon geometry with an additional class attribute
}
\description{
Return a sits_tibble or raster_cube as an sf object.
}
\author{
Felipe Carvalho, \email{felipe.carvalho@inpe.br}
}
