% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_deep_learning.R
\name{sits_keras_diagnostics}
\alias{sits_keras_diagnostics}
\title{Provides access to diagnostic information about a Keras deep learning model}
\usage{
sits_keras_diagnostics(dl_model)
}
\arguments{
\item{dl_model}{A valid keras model.}
}
\value{
This function returns NULL. It only prints the model diagnostics.
}
\description{
After the Keras deeplearning model is compiled and fit, this
             function provides access to the history plot and the evaluation results.
}
\examples{
\donttest{
# Retrieve the set of samples for the Mato Grosso region (provided by EMBRAPA)
data(cerrado_2classes)
 # obtain a DL model
dl_model <- sits_train(cerrado_2classes,
     sits_deeplearning(units = c(512, 512), dropout_rates = c(0.45, 0.25), epochs = 100))
# run the keras diagnostics
sits_keras_diagnostics(dl_model)
}
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
