% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_validate.R
\name{sits_kfold_validate}
\alias{sits_kfold_validate}
\title{Cross-validate temporal patterns}
\usage{
sits_kfold_validate(data.tb, folds = 5, ml_method = sits_svm())
}
\arguments{
\item{data.tb}{sits tibble}

\item{folds}{number of partitions to create.}

\item{ml_method}{machine learning method}
}
\value{
pred_ref.tb    tibble containing pairs of reference and predicted values
}
\description{
Splits the set of time series into training and validation and
perform k-fold cross-validation.
Cross-validation is a model validation technique for assessing how the results
of a statistical analysis will generalize to an independent data set.
It is mainly used in settings where the goal is prediction,
and one wants to estimate how accurately a predictive model will perform in practice.
One round of cross-validation involves partitioning a sample of data
into complementary subsets, performing the analysis on one subset
(called the training set), and validating the analysis on the other subset
(called the validation set or testing set).

The k-fold cross validation method involves splitting the dataset
into k-subsets. For each subset is held out while the model is trained
on all other subsets. This process is completed until accuracy
is determine for each instance in the dataset, and an overall
accuracy estimate is provided.

This function returns the confusion matrix, and Kappa values.
}
\examples{
\donttest{
# read a set of samples
data (cerrado_2classes)

# perform a five fold validation with the SVM machine learning method
conf_matrix1.mx <- sits_kfold_validate (cerrado_2classes, ml_method = sits_svm())
}
}
\author{
Rolf Simoes, \email{rolf.simoes@inpe.br}

Gilberto Camara, \email{gilberto.camara@inpe.br}
}
