% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_validate.R
\name{sits_kfold_validate}
\alias{sits_kfold_validate}
\alias{sits_validate}
\title{Cross-validate time series samples}
\usage{
sits_kfold_validate(data, folds = 5, ml_method = sits_rfor(), multicores = 2)

sits_validate(
  data,
  ...,
  samples_validation = NULL,
  validation_split = 0.2,
  ml_method = sits_rfor()
)
}
\arguments{
\item{data}{Time series.}

\item{folds}{Number of partitions to create.}

\item{ml_method}{Machine learning method.}

\item{multicores}{Number of cores to process in parallel.}

\item{samples_validation}{Validation time series.}
}
\value{
A tibble containing pairs of reference and predicted values.
}
\description{
Splits the set of time series into training and validation and
perform k-fold cross-validation.
Cross-validation is a technique for assessing how the results
of a statistical analysis will generalize to an independent data set.
It is mainly used in settings where the goal is prediction,
and one wants to estimate how accurately a predictive model will perform.
One round of cross-validation involves partitioning a sample of data
into complementary subsets, performing the analysis on one subset
(called the training set), and validating the analysis on the other subset
(called the validation set or testing set).

The k-fold cross validation method involves splitting the dataset
into k-subsets. For each subset is held out while the model is trained
on all other subsets. This process is completed until accuracy
is determine for each instance in the dataset, and an overall
accuracy estimate is provided.

This function returns the confusion matrix, and Kappa values.
}
\examples{
\donttest{
# read a set of samples
data(cerrado_2classes)
# two fold validation with random forest
acc <- sits_kfold_validate(cerrado_2classes,
  folds = 2,
  ml_method = sits_rfor(num_trees = 100)
)
}
}
\author{
Rolf Simoes, \email{rolf.simoes@inpe.br}

Gilberto Camara, \email{gilberto.camara@inpe.br}
}
