% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_kohonen.R
\name{sits_kohonen}
\alias{sits_kohonen}
\title{Clustering a set of satellite image time series using SOM}
\usage{
sits_kohonen(data.tb, time_series, bands = NULL, grid_xdim = 25,
  grid_ydim = 25, rlen = 100, dist.fcts = "euclidean", alpha = 1,
  neighbourhood.fct = "bubble", ...)
}
\arguments{
\item{data.tb}{A tibble with samples to be clustered.}

\item{time_series}{Time series extracted from tibble data.}

\item{bands}{Bands to be clustered.}

\item{grid_xdim}{X dimension of the SOM grid (default = 5).}

\item{grid_ydim}{Y dimension of the SOM grid.}

\item{rlen}{Number of times the complete data set will be presented to the SOM grid}

\item{dist.fcts}{The similiraty measure (distance).}

\item{alpha}{Learning rate, a vector of two numbers indicating the amount of change.}

\item{neighbourhood.fct}{Type of neighbourhood function (bubble or gaussian).}

\item{...}{Additional parameters to be passed to kohonen::supersom function.}
}
\value{
A tibble with the clusters time series or cluster' members time series according to return_member parameter.
If return_members are FALSE, the returning tibble will contain a new collumn called `n_members` informing how many members has each cluster.
}
\description{
This function uses package self-organized maps to find clusters in
satellite image time series to cluster samples.
}
\references{
`kohonen` package (https://CRAN.R-project.org/package=kohonen)
}
\author{
Lorena Alves, \email{lorena.santos@inpe.br}
}
