% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_kohonen.R
\name{sits_kohonen}
\alias{sits_kohonen}
\title{Cluster a set of satellite image time series using SOM}
\usage{
sits_kohonen(data.tb, bands = NULL, grid_xdim = 5, grid_ydim = 5,
  rlen = 100, alpha = c(0.05, 0.01), ...)
}
\arguments{
\item{data.tb}{a tibble with time series to be clustered}

\item{bands}{bands to be clustered.}

\item{grid_xdim}{x dimension of the SOM grid (default = 5)}

\item{grid_ydim}{y dimension of the SOM grid}

\item{rlen}{number of times the complete data set will be presented to the SOM grid}

\item{alpha}{learning rate, a vector of two numbers indicating the amount of change.
Default is to decline linearly from 0.05 to 0.01 over rlen updates.}

\item{...}{additional parameters to be passed to kohonen::supersom function}
}
\value{
kohonen.obj   a full kohonen object for data analysis
}
\description{
This function uses package "kohonen" (self-organized maps) to find clusters in
satellite image time series.
}
\examples{
\donttest{

# Retrieve a time series to be clusterized
data("samples_MT_9classes")
# compute a kohonen object
koh_obj <- sits_kohonen(samples_MT_9classes)
# generate a dataset
cluster.tb <- sits_kohonen_cluster(samples_MT_9classes, koh_obj)
# plot the result
sits_cluster_frequency(cluster.tb)
}
}
\references{
`kohonen` package (https://CRAN.R-project.org/package=kohonen)
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}

Rolf Simoes, \email{rolf.simoes@inpe.br}

Lorena Alves, \email{lorena.santos@inpe.br}
}
