% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_kohonen.R
\name{sits_kohonen_cluster}
\alias{sits_kohonen_cluster}
\title{Cluster a set of satellite image time series using SOM}
\usage{
sits_kohonen_cluster(data.tb, kohonen.obj)
}
\arguments{
\item{data.tb}{a tibble with time series to be clustered}

\item{kohonen.obj}{a full kohonen object for data analysis}
}
\value{
data.tb       tibble with the clusters or clusters' members
}
\description{
This function uses package "kohonen" (self-organized maps) to find clusters in
satellite image time series.
}
\examples{
\donttest{

# Retrieve a time series to be clusterized
data("samples_MT_9classes")
# compute a kohonen object
koh_obj <- sits_kohonen(samples_MT_9classes)
# generate a dataset
cluster.tb <- sits_kohonen_cluster(samples_MT_9classes, koh_obj)
# plot the result
sits_cluster_frequency(cluster.tb)
}
}
\references{
`kohonen` package (https://CRAN.R-project.org/package=kohonen)
}
\author{
Rolf Simoes, \email{rolf.simoes@inpe.br}

Lorena Alves, \email{lorena.santos@inpe.br}
}
