% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_label_classification.R
\name{sits_label_classification}
\alias{sits_label_classification}
\title{Build a labelled image from a probability cube}
\usage{
sits_label_classification(
  cube,
  multicores = 2,
  memsize = 4,
  output_dir = ".",
  version = "v1"
)
}
\arguments{
\item{cube}{Classified image data cube.}

\item{multicores}{Number of workers to label the classification in
parallel.}

\item{memsize}{maximum overall memory (in GB) to label the
classification.}

\item{output_dir}{Output directory for classified files.}

\item{version}{Version of resulting image
(in the case of multiple runs).}
}
\value{
A data cube
}
\description{
Takes a set of classified raster layers with probabilities,
             and label them based on the maximum probability for each pixel.
}
\note{
Please refer to the sits documentation available in
<https://e-sensing.github.io/sitsbook/> for detailed examples.
}
\author{
Rolf Simoes, \email{rolf.simoes@inpe.br}
}
