% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_labels.R
\name{sits_labels<-.sits}
\alias{sits_labels<-.sits}
\title{Change the labels of a set of time series}
\usage{
\method{sits_labels}{sits}(data) <- value
}
\description{
Change the labels of a set of time series
}
