% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_table.R
\name{sits_labels}
\alias{sits_labels}
\title{Finds the labels of all occurences in a sits table}
\usage{
sits_labels(data.tb)
}
\arguments{
\item{data.tb}{a valid sits table}
}
\value{
names      a string vector with the labels
}
\description{
\code{sits_labels} finds and returns the labels tagged in a sits table
}
\seealso{
Other SITS table functions: \code{\link{sits_bands}},
  \code{\link{sits_cross}}, \code{\link{sits_merge}}
}
