% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_labels.R
\name{sits_labels}
\alias{sits_labels}
\title{Returns the information about labels of a data set (tibble or cube)}
\usage{
sits_labels(data)
}
\arguments{
\item{data}{Valid sits tibble (time series or a cube)}
}
\value{
A string vector with the labels.
}
\description{
Finds labels in a sits tibble or data cube
              For details see:
\itemize{
 \item{"time series": }{see \code{\link{sits_labels.sits}}}
 \item{"data cube": }{see \code{\link{sits_labels.cube}}}
}
}
\author{
Rolf Simoes, \email{rolf.simoes@inpe.br}
}
