% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_table.R
\name{sits_labels}
\alias{sits_labels}
\title{returns the labels' count of a sits table}
\usage{
sits_labels(data.tb, from_members = FALSE)
}
\arguments{
\item{data.tb}{a valid sits table}

\item{from_members}{(Boolean) if n_members collumn is present, compute label's statistics from it.}
}
\value{
result.tb  a tibble with the names of the labels and its absolute and relative frequency
}
\description{
returns the labels and its respective counting and frequency.
}
\author{
Rolf Simoes, \email{rolf.simoes@inpe.br}
}
