% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_table.R
\name{sits_labels}
\alias{sits_labels}
\title{returns the labels' count of a sits table}
\usage{
sits_labels(data.tb)
}
\arguments{
\item{data.tb}{a valid sits table}
}
\value{
labels.tb  a tibble with the names of the labels and its counting
}
\description{
returns the labels and its counting
}
\author{
Rolf Simoes, \email{rolf.simoes@inpe.br}
}
