% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_labels.R
\name{sits_labels}
\alias{sits_labels}
\title{Returns the information about labels  of a sits tibble}
\usage{
sits_labels(data)
}
\arguments{
\item{data}{A tibble with time series data and metadata.}
}
\value{
A tibble with the names of the labels and their absolute and relative frequency.
}
\description{
Returns the labels and its respective counting and frequency.
}
\examples{
# read a tibble with 400 samples of Cerrado and 346 samples of Pasture
data(cerrado_2classes)
# print the labels
sits_labels(cerrado_2classes)
}
\author{
Rolf Simoes, \email{rolf.simoes@inpe.br}
}
