% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_labels.R
\name{sits_labels}
\alias{sits_labels}
\title{Get labels associated to a data set}
\usage{
sits_labels(data)
}
\arguments{
\item{data}{Time series or a cube.}
}
\value{
Labels.
}
\description{
Finds labels in a sits tibble or data cube
}
\examples{
# read a tibble with 400 samples of Cerrado and 346 samples of Pasture
data(cerrado_2classes)
# print the labels
sits_labels(cerrado_2classes)
}
\author{
Rolf Simoes, \email{rolf.simoes@inpe.br}
}
