% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_labels.R
\name{sits_labels}
\alias{sits_labels}
\title{returns the labels' count of a sits table}
\usage{
sits_labels(data.tb)
}
\arguments{
\item{data.tb}{a valid sits table}
}
\value{
result.tb  a tibble with the names of the labels and its absolute and relative frequency
}
\description{
returns the labels and its respective counting and frequency.
}
\examples{
# read a tibble with 400 samples of Cerrado and 346 samples of Pasture
cerrado2.tb <- readRDS(system.file("extdata/time_series/cerrado_2classes.rds", package = "sits"))
# print the labels
sits_labels (cerrado2.tb)

}
\author{
Rolf Simoes, \email{rolf.simoes@inpe.br}
}
