% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_labels.R
\name{sits_labels}
\alias{sits_labels}
\alias{sits_labels.sits}
\alias{sits_labels.raster_cube}
\alias{sits_labels.patterns}
\alias{sits_labels.sits_model}
\alias{sits_labels.tbl_df}
\alias{sits_labels.default}
\alias{sits_labels<-.tbl_df}
\title{Get labels associated to a data set}
\usage{
sits_labels(data)

\method{sits_labels}{sits}(data)

\method{sits_labels}{raster_cube}(data)

\method{sits_labels}{patterns}(data)

\method{sits_labels}{sits_model}(data)

\method{sits_labels}{tbl_df}(data)

\method{sits_labels}{default}(data)

\method{sits_labels}{tbl_df}(data) <- value
}
\arguments{
\item{data}{Data.frame containing time series or a data cube.}

\item{value}{A character vector used to convert labels. Labels will
be renamed to the respective value positioned at the
labels order returned by \code{\link{sits_labels}}.}
}
\value{
The labels associated to a set of time series or to
                 a data cube (character vector).
}
\description{
Finds labels in a sits tibble or data cube
}
\examples{
# read a tibble with 400 samples of Cerrado and 346 samples of Pasture
data(cerrado_2classes)
# print the labels
sits_labels(cerrado_2classes)
}
\author{
Rolf Simoes, \email{rolf.simoes@inpe.br}
}
