% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_machine_learning.R
\name{sits_lda}
\alias{sits_lda}
\title{Train a SITS classification model using linear discriminant analysis}
\usage{
sits_lda(distances.tb = NULL, formula = sits_formula_logref(), ...)
}
\arguments{
\item{distances.tb}{a time series with a set of distance measures for each training sample}

\item{formula}{a symbolic description of the model to be fit. SITS offers a set of such formulas (default: sits_formula_logref)}

\item{...}{other parameters to be passed to MASS::lda function}
}
\value{
result          a model function to be passed in sits_predict
}
\description{
This function receives a tibble with a set of attributes X for each observation Y
These attributes are distance metrics between patterns and observations, obtained by a distance
function in SITS (see \code{link[sits]{sits_distances}}).
The method performs a linear discriminant analysis (lda) to obtain a predictive model.
This function is a front-end to the "lda" method in the "MASS" package.
Please refer to the documentation in that package for more details.
}
\examples{
\donttest{
# Retrieve the set of samples for the Mato Grosso region (provided by EMBRAPA)
data(samples_MT_9classes)
# Select 3 bands for classification
samples.tb <- sits_select (samples_MT_9classes, bands = c("ndvi", "evi", "nir"))
# get a point with a 16 year time series
data(ts_2000_2016)
# Select 3 bands for classification
point.tb <- sits_select (ts_2000_2016, bands = c("ndvi", "evi", "nir"))
# classify the point
class.tb <- sits_classify (point.tb, samples.tb, ml_method = sits_lda())
}
}
\author{
Alexandre Xavier Ywata de Carvalho, \email{alexandre.ywata@ipea.gov.br}

Rolf Simoes, \email{rolf.simoes@inpe.br}

Gilberto Camara, \email{gilberto.camara@inpe.br}
}
