% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_filters.R
\name{sits_linear_interp}
\alias{sits_linear_interp}
\title{Interpolation function of sits_table's time series}
\usage{
sits_linear_interp(data.tb, n = 23)
}
\arguments{
\item{data.tb}{a valid sits table}

\item{n}{the number of time series elements to be created between start date and end date}
}
\value{
result.tb    a sits_table with same samples and the new bands
}
\description{
Computes the linearly interpolated bands for a given resolution
              using the R base function approx
}
\author{
Rolf Simoes, \email{rolf.simoes@inpe.br}
}
