% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_config.R
\name{sits_list_collections}
\alias{sits_list_collections}
\title{List the cloud collections supported by sits}
\usage{
sits_list_collections(source = NULL)
}
\arguments{
\item{source}{Data source to be shown in detail.}
}
\value{
Prints collections available in
        each cloud service supported by sits.
}
\description{
Prints the collections available
in each cloud service supported by sits.
Users can select to get information
only for a single service by using the \code{source} parameter.
}
