% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_time_handling.R
\name{sits_match_dates}
\alias{sits_match_dates}
\title{Adjust the dates of an input data set with a reference data set}
\usage{
sits_match_dates(input_start_date, input_end_date, ref_start_date, ref_end_date,
  interval = "12 month")
}
\arguments{
\item{input_start_date}{the starting date of the input data set}

\item{input_end_date}{the end date of the input data set}

\item{ref_start_date}{the starting date of the reference data set}

\item{ref_end_date}{the end date of the reference data set}

\item{interval}{the interval that shape matching will be performed}
}
\value{
breaks            the breaks that will be applied to the input data set
}
\description{
For correct classification, the time series of the input data set
             should be aligned to that of the reference data set (usually a set of patterns).
             This function aligns these data sets so that shape matching works correctly
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
