% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_tibble_utils.R
\name{sits_match_dates}
\alias{sits_match_dates}
\title{Adjust the dates of an input data set with a reference data set}
\usage{
sits_match_dates(data.tb, patterns.tb)
}
\arguments{
\item{data.tb}{the input data set}

\item{patterns.tb}{the reference data set}
}
\value{
breaks            the breaks that will be applied to the input data set
}
\description{
For correct classification, the time series of the input data set
             should be aligned to that of the reference data set (usually a set of patterns).
             This function aligns these data sets so that shape matching works correctly
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
