% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_tibble_functions.R
\name{sits_match_dates}
\alias{sits_match_dates}
\title{Adjust the dates of an input data set with a reference data set}
\usage{
sits_match_dates(timeline, ref_start_date, ref_end_date,
  interval = "12 month")
}
\arguments{
\item{timeline}{Timeline of input observations (vector)}

\item{ref_start_date}{Reference starting date}

\item{ref_end_date}{Reference end date}

\item{interval}{Period to match the data to the patterns}
}
\value{
subset_dates.lst     A list of breaks that will be applied to the input data set
}
\description{
For correct classification, the time series of the input data set
             should be aligned to that of the reference data set (usually a set of patterns).
             This function aligns these data sets so that shape matching works correctly
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
