% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_tibble_utils.R
\name{sits_match_dates}
\alias{sits_match_dates}
\title{Adjust the dates of an input data set with a reference data set}
\usage{
sits_match_dates(data.tb, patterns.tb, start_date = NULL, end_date = NULL,
  interval = "12 month")
}
\arguments{
\item{data.tb}{Input data set}

\item{patterns.tb}{Reference data set}

\item{start_date}{Starting date of the distance matching between time series and patterns}

\item{end_date}{End date of the distance matching between time series and patternss}

\item{interval}{Period to match the data to the patterns}
}
\value{
subset_dates.lst            the breaks that will be applied to the input data set
}
\description{
For correct classification, the time series of the input data set
             should be aligned to that of the reference data set (usually a set of patterns).
             This function aligns these data sets so that shape matching works correctly
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
