% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_tibble_functions.R
\name{sits_merge}
\alias{sits_merge}
\title{Merge two satellite image time series}
\usage{
sits_merge(data1.tb, data2.tb)
}
\arguments{
\item{data1.tb}{the first SITS table to be merged}

\item{data2.tb}{the second SITS table to be merged}
}
\value{
result.tb    a merged SITS tibble with a nested set of time series
}
\description{
This function merges the time series of two STIS tables.
To merge two series, we consider that they contain different
attributes but refer to the same coverage, and spatio-temporal location.
This function is useful to merge different bands of the same spatio-temporal locations.
For example, one may want to put the raw and smoothed bands for the same set of locations
in the same table.
}
\examples{
# Read a point
point.tb <- readRDS(system.file("extdata/time_series/point.rds", package = "sits"))
# Select the NDVI band
point.tb <- sits_select (point.tb, bands = c("ndvi"))
# Filter the point using the whittaker smoother
point_ws.tb <- sits_whittaker (point.tb, lambda = 3.0)
# Merge the two tibbles
point_merge.tb <- sits_merge(point.tb, point_ws.tb)
# Plot the merged time series to see the smoothing effect
sits_plot(point_merge.tb)
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
