% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_table.R
\name{sits_merge}
\alias{sits_merge}
\title{Merge two satellite image time series}
\usage{
sits_merge(sits1.tb, sits2.tb)
}
\arguments{
\item{sits1.tb}{the first SITS table to be merged}

\item{sits2.tb}{the second SITS table to be merged}
}
\value{
new.tb    a merged SITS tibble with a nested set of time series
}
\description{
This function merges the time series of two STIS tables.
To merge two series, we consider that they contain different
attributes but refer to the same coverage, and spatio-temporal location.
This function is useful to merge different bands of the same spatio-temporal locations.
For example, one may want to put the raw and smoothed bands for the same set of locations
in the same table.
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
