% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_merge.R
\name{sits_merge}
\alias{sits_merge}
\title{Merge two data sets (time series or cubes)}
\usage{
sits_merge(data1, data2, ..., suffix = c(".1", ".2"))
}
\arguments{
\item{data1}{Time series or cube to be merged.}

\item{data2}{Time series or cube to be merged.}

\item{...}{Additional parameters}

\item{suffix}{If there are duplicate bands in data1 and data2
these suffixes will be added.}
}
\value{
merged data sets
}
\description{
To merge two series, we consider that they contain different
attributes but refer to the same data cube, and spatiotemporal location.
This function is useful to merge different bands of the same locations.
For example, one may want to put the raw and smoothed bands
for the same set of locations in the same tibble.

To merge data cubes, they should share the same sensor, resolution,
bounding box, timeline, and have different bands.
}
\examples{
if (sits_active_tests()) {
# Retrieve a time series with values of NDVI
point_ndvi <- sits_select(point_mt_6bands, bands = "NDVI")
# Apply Savitsky-Golay filter on NDVI
point_ndvi_sg <- point_ndvi \%>\%
  sits_filter(sits_sgolay())

# Merge time series back
point <- sits_merge(point_ndvi, point_ndvi_sg, suffix = c("", ".SG"))
}

}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
