% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_table.R
\name{sits_merge}
\alias{sits_merge}
\title{Merge two satellite image time series}
\usage{
sits_merge(sits1.tb, sits2.tb)
}
\arguments{
\item{ts1}{the first SITS table to be merged}

\item{ts2}{the second SITS table to be merged}
}
\value{
new.tb    a merged SITS tibble with a nested set of time series
}
\description{
\code{sits_merge} merges two STIS tables with the same spatio-temporal references
}
\details{
#' To merge two series, we consider that they contain different
attributes but refer to the same coverage, and spatio-temporal location.

This function is useful to merge different bands of the same spatio-temporal locations.
For example, one may want to put the raw and smoothed bands for the same set of locations
in the same table.
}
\examples{
merged.tb <- merge_WTSS (sits1.tb, sits2.tb)

}
\seealso{
Other SITS main functions: \code{\link{sits_fromClusters}},
  \code{\link{sits_fromSHP}}, \code{\link{sits_getdata}}
}
\keyword{SITS}
