% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_merge.R
\name{sits_merge}
\alias{sits_merge}
\title{Merge two data sets (time series or cubes)}
\usage{
sits_merge(data1, data2)
}
\arguments{
\item{data1}{sits tibble or cube to be merged.}

\item{data2}{sits tibble or cube to be merged.}
}
\value{
merged data sets
}
\description{
To merge two series, we consider that they contain different
attributes but refer to the same data cube, and spatio-temporal location.
This function is useful to merge different bands of the same locations.
For example, one may want to put the raw and smoothed bands
for the same set of locations in the same tibble.

To merge data cubes, they should share the same sensor, resolution,
bounding box, timeline, and have different bands.
}
\examples{
{
# Retrieve a time series with values of NDVI
point_ndvi <- sits_select(point_mt_6bands, bands = "NDVI")
# Filter the point using the whittaker smoother
point_ws.tb <- sits_whittaker(point_ndvi, lambda = 3.0)
# Plot the two points to see the smoothing effect
plot(sits_merge(point_ndvi, point_ws.tb))
}

}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
