% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_data_generics.R
\name{sits_merge}
\alias{sits_merge}
\title{Merge two data sets (time series or cubes)}
\usage{
sits_merge(data1, data2)
}
\arguments{
\item{data1}{sits tibble or cube to be merged.}

\item{data2}{sits tibble or cube to be merged.}
}
\value{
merged data sets
}
\description{
For details see:
\itemize{
 \item{"time series": }{see \code{\link{sits_merge.sits}}}
 \item{"data cube": }{see \code{\link{sits_merge.cube}}}
}
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
