% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_csv.R
\name{sits_metadata_to_csv}
\alias{sits_metadata_to_csv}
\title{Export a sits tibble metadata to the CSV format}
\usage{
sits_metadata_to_csv(data, file)
}
\arguments{
\item{data}{Time series.}

\item{file}{Name of the exported CSV file.}
}
\value{
The status of the operation.
}
\description{
Converts metadata from a sits tibble to a CSV file.
             The CSV file will not contain the actual time
             series. Its columns will be the same as those of a
             CSV file used to retrieve data from
             ground information ("latitude", "longitude", "start_date",
             "end_date", "cube", "label").
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
