% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_kohonen.R
\name{sits_metrics_by_cluster}
\alias{sits_metrics_by_cluster}
\title{Metrics by cluster}
\usage{
sits_metrics_by_cluster(info_sample_cluster.tb)
}
\arguments{
\item{data.tb}{obj contained the clusterized samples}

\item{size_grid}{the size of kohonen map}
}
\value{
returns TRUE if data.tb has data.
}
\description{
Create subgroups to identify variations in a same group
}
\author{
Lorena Santos, \email{lorena.santos@inpe.br}
}
