% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_filters.R
\name{sits_missing_values}
\alias{sits_missing_values}
\title{Remove missing values}
\usage{
sits_missing_values(data.tb, miss_value)
}
\arguments{
\item{data.tb}{a valid sits table}

\item{miss_value}{a number indicating missing values in a time series.}
}
\value{
result.tb  a sits_table with same samples and the new bands
}
\description{
This function removes the missing values from an image time series by substituting them by NA
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
