% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_spectral_unmixing.R
\name{sits_mixture_model}
\alias{sits_mixture_model}
\title{Create a multiple endmember spectral mixture analyses fractions images}
\usage{
sits_mixture_model(
  cube,
  endmembers_spectra,
  memsize = 1,
  multicores = 2,
  output_dir = getwd(),
  progress = TRUE
)
}
\arguments{
\item{cube}{A sits data cube.}

\item{memsize}{Memory available for mixture model (in GB).}

\item{multicores}{Number of cores to be used for generate the mixture model.}

\item{output_dir}{Directory for output file.}

\item{progress}{Show progress bar?}

\item{endmembers}{Reference endmembers spectra}
}
\value{
a cube with the generated endmembers.
}
\description{
A multiple endmember spectral mixture analyses...
}
