% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_machine_learning.R
\name{sits_mlp}
\alias{sits_mlp}
\title{Train a SITS classifiction model using multi-layer perceptron (neural networks and deep learning)}
\usage{
sits_mlp(distances.tb = NULL, hidden_node = c(400, 200, 100),
  learning.rate = 0.001, dropout = NULL, activation = "sigmoid",
  out_activation = "softmax", optimizer = "adam", num.round = 5000,
  batch.size = 32, stop.metric = 0.98, device = mxnet::mx.cpu(), ...)
}
\arguments{
\item{distances.tb}{a time series with a set of distance measures for each training sample}

\item{hidden_node}{a vector containing the number of hidden nodes in each hidden layer}

\item{learning.rate}{the learning rate of the MLP (default = 0.001)}

\item{dropout}{a number in (0,1) containing the dropout ratio from the last hidden layer to the output layer}

\item{activation}{either a single string or a vector containing the names of activation functions. Valid values are {'relu', 'sigmoid', 'softrelu', 'tanh'}}

\item{out_activation}{a single string containing the name of the output activation function. Valid values are {'rmse', 'sofrmax', 'logistic'}}

\item{optimizer}{string - default is 'adm', and valid values are 'sgd', 'adam', 'rmsprop', 'adagrad'}

\item{num.round}{number of iterations to train the model}

\item{batch.size}{batch size used for array training}

\item{stop.metric}{precision where iteration stops}

\item{device}{whether to train on mx.cpu (default) or mx.gpu}

\item{...}{other parameters to be passed to `mx.model.FeedForward.create` function}
}
\value{
result          either an model function to be passed in sits_predict or an function prepared that can be called further to compute multinom training model
}
\description{
Use a multi-layer perceptron algorithm to classify data.
This function is a front-end to the "mx.mlp" method in the "mxnet" R package.
Please refer to the documentation in that package for more details.
}
\examples{
\donttest{
# Retrieve the set of samples for the Mato Grosso region (provided by EMBRAPA)
data(samples_MT_9classes)
# Select 3 bands for classification
samples.tb <- sits_select (samples_MT_9classes, bands = c("ndvi", "evi", "nir"))
# get a point with a 16 year time series
data(ts_2000_2016)
# Select 3 bands for classification
point.tb <- sits_select (ts_2000_2016, bands = c("ndvi", "evi", "nir"))
# Classify the point
class.tb <- sits_classify (point.tb, samples.tb,
       ml_method = sits_mlp())
}
}
\author{
Alexandre Xavier Ywata de Carvalho, \email{alexandre.ywata@ipea.gov.br}

Rolf Simoes, \email{rolf.simoes@inpe.br}

Gilberto Camara, \email{gilberto.camara@inpe.br}
}
