% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_mlp.R
\name{sits_mlp}
\alias{sits_mlp}
\title{Train multi-layer perceptron models using torch}
\usage{
sits_mlp(
  samples = NULL,
  samples_validation = NULL,
  layers = c(512, 512, 512),
  dropout_rates = c(0.2, 0.3, 0.4),
  learning_rate = 0.001,
  epochs = 100,
  batch_size = 64,
  validation_split = 0.2,
  patience = 20,
  min_delta = 0.01,
  verbose = FALSE
)
}
\arguments{
\item{samples}{Time series with the training samples.}

\item{samples_validation}{Time series with the validation samples. if the
\code{samples_validation} parameter is provided,
the \code{validation_split} parameter is ignored.}

\item{layers}{Vector with number of hidden nodes in each layer.}

\item{dropout_rates}{Vector with the dropout rates (0,1)
for each layer.}

\item{learning_rate}{Learning rate of the optimizer}

\item{epochs}{Number of iterations to train the model.}

\item{batch_size}{Number of samples per gradient update.}

\item{validation_split}{Number between 0 and 1.
Fraction of the training data for validation.
The model will set apart this fraction
and will evaluate the loss and any model metrics
on this data at the end of each epoch.}

\item{patience}{Number of epochs without improvements until
training stops.}

\item{min_delta}{Minimum improvement to reset the patience counter.}

\item{verbose}{Verbosity mode (TRUE/FALSE). Default is FALSE.}
}
\value{
Either a model to be passed in sits_predict
                          or a function prepared to be called further.
}
\description{
Use a multi-layer perceptron algorithm to classify data.
This function uses the R "torch" and "luz" packages.
Please refer to the documentation of those package for more details.
}
\note{
The parameters for the MLP have been chosen based on the work by Wang et al. 2017
that takes multilayer perceptrons as the baseline for time series classifications:
(a) Three layers with 512 neurons each, specified by the parameter `layers`;
(b) dropout rates of 10%, 20%, and 30% for the layers;
(c) the "optimizer_adam" as optimizer (default value);
(d) a number of training steps (`epochs`) of 100;
(e) a `batch_size` of 64, which indicates how many time series
are used for input at a given steps;
(f) a validation percentage of 20%, which means 20% of the samples
will be randomly set side for validation.
(g) The "relu" activation function.

#' @references

Zhiguang Wang, Weizhong Yan, and Tim Oates,
"Time series classification from scratch with deep neural networks:
 A strong baseline",
 2017 international joint conference on neural networks (IJCNN).
}
\examples{
\dontrun{
# Retrieve the set of samples for the Mato Grosso region
# Build a machine learning model based on deep learning
dl_model <- sits_train(samples_modis_4bands, sits_mlp())
# get a point with a 16 year time series
point_4classes <- sits_select(point_mt_6bands,
                              bands = c("NDVI", "EVI", "NIR", "MIR"))
# classify the point
point_class <- sits_classify(point_4classes, dl_model)
# plot the classified point
plot(point_class)
}
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}

Rolf Simoes, \email{rolf.simoes@inpe.br}

Felipe Souza, \email{lipecaso@gmail.com}

Alber Sanchez, \email{alber.ipia@inpe.br}
}
