% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_machine_learning.R
\name{sits_mlr}
\alias{sits_mlr}
\title{Train a SITS classifiaction model using multinomial log-linear fitting}
\usage{
sits_mlr(distances.tb = NULL, formula = sits_formula_logref(), ...)
}
\arguments{
\item{distances.tb}{a time series with a set of distance measures for each training sample}

\item{formula}{a symbolic description of the model to be fit. SITS offers a set of such formulas (default: sits_formula_logref)}

\item{...}{other parameters to be passed to nnet::multinom function}
}
\value{
result          a model function to be passed in sits_predict
}
\description{
Use multinomial log-linear (mlr) fitting model via neural networks to classify data.
These attributes are distance metrics between patterns and observations, obtained by a distance
function in SITS (see \code{link[sits]{sits_distances}}).
This function is a front-end to the "multinom" method in the "nnet" package.
Please refer to the documentation in that package for more details.
}
\examples{
\donttest{
# Retrieve the set of samples for the Mato Grosso region (provided by EMBRAPA)
samples.tb <- readRDS(system.file("extdata/time_series/embrapa_mt.rds", package = "sits"))
samples.tb <- sits_select (samples.tb, bands = c("ndvi", "evi", "nir"))
# get a point with a 16 year time series
point.tb <- readRDS(system.file("extdata/time_series/point.rds", package = "sits"))
point.tb <- sits_select (point.tb, bands = c("ndvi", "evi", "nir"))
# classify the point
class.tb <- sits_classify (point.tb, samples.tb,
       ml_method = sits_mlr())
}
}
\author{
Alexandre Xavier Ywata de Carvalho, \email{alexandre.ywata@ipea.gov.br}

Rolf Simoes, \email{rolf.simoes@inpe.br}

Gilberto Camara, \email{gilberto.camara@inpe.br}
}
