% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_tibble_utils.R
\name{sits_mutate}
\alias{sits_mutate}
\title{Add new SITS bands.}
\usage{
sits_mutate(data.tb, ...)
}
\arguments{
\item{data.tb}{a valid sits tibble}

\item{...}{`name=value` pairs expressions. See `dplyr::mutate` help for more details.}
}
\value{
data.tb      a sits tibble with same samples and the new bands
}
\description{
Adds new bands and preserves existing in the time series of a sits tibble using dplyr::mutate function
}
\author{
Rolf Simoes, \email{rolf.simoes@inpe.br}
}
