% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_tibble.R
\name{sits_mutate_bands}
\alias{sits_mutate_bands}
\title{Add new sits bands.}
\usage{
sits_mutate_bands(data, ...)
}
\arguments{
\item{data}{Valid sits tibble.}

\item{...}{Expressions written as `name = value`.
See \code{dplyr::mutate()} help for more details.}
}
\value{
A sits tibble with same samples and the selected bands.
}
\description{
Adds new bands and preserves existing in the time series
             of a sits tibble using \code{dplyr::mutate} function.
}
\examples{
\donttest{
# Retrieve data for time series with label samples in Mato Grosso in Brazil
data(samples_mt_6bands)
# Generate a new image with the SAVI (Soil-adjusted vegetation index)
savi.tb <- sits_mutate_bands(samples_mt_6bands,
           SAVI = (1.5 * (NIR - RED) / (NIR + RED + 0.5)))
}
}
\author{
Rolf Simoes, \email{rolf.simoes@inpe.br}
}
\keyword{internal}
