% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_tibble_functions.R
\name{sits_normalize_ts}
\alias{sits_normalize_ts}
\title{Normalize the time series in the given sits_tibble}
\usage{
sits_normalize_ts(data.tb, use_IQR = TRUE, multicores = 1)
}
\arguments{
\item{data.tb}{a tibble in SITS format}

\item{use_IQR}{(logical): should we use inter-quartile ranges?}

\item{multicores}{number of threads to process the time series.}
}
\value{
data.tb    a normalized sits tibble
}
\description{
this function normalizes the time series using the mean and
standard deviation of all the time series.
}
\examples{
\donttest{
#' sits_tibble_normalized <- sits_normalize_ts(samples_MT_9classes)
}

}
\author{
Alber Sanchez, \email{alber.ipia@inpe.br}
}
