% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_parallel.R
\name{sits_parallel_fault_tolerant}
\alias{sits_parallel_fault_tolerant}
\alias{.sits_parallel_recv_one_data}
\alias{.sits_parallel_recv_one_result}
\alias{.sits_parallel_cluster_apply}
\title{Fault tolerant version of some parallel functions}
\usage{
.sits_parallel_recv_one_data()

.sits_parallel_recv_one_result()

.sits_parallel_cluster_apply(x, fn, ..., pb = NULL)
}
\arguments{
\item{x}{a given list to be passed to a function}

\item{fn}{a function to be applied to each list element}

\item{...}{additional arguments to fn function}
}
\value{
List with IDs of workers and their values

List with values and nodes

No value, called for side effect.
}
\description{
These internal functions are a reimplementation of a fault tolerant
version of snow package functions \code{recv_one_data()},
\code{recv_one_result()}, and \code{cluster_apply()}
from Luke Tierney, A. J. Rossini, Na Li, H. Sevcikova.
snow package is licensed as GPL-2 | GPL-3.
This re-implementation allows `sits` cope with massive volume of data
processing over networks without compromise overall results.
}
\author{
Rolf Simoes, \email{rolf.simoes@inpe.br}
}
\keyword{internal}
