% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_patterns.R
\name{sits_patterns}
\alias{sits_patterns}
\title{Find time series patterns for classification
\code{sits_patterns} returns a sits table with a list of patterns}
\usage{
sits_patterns(samples.tb, method = "gam", ...)
}
\arguments{
\item{samples.tb}{a table in SITS format with a set of labelled time series}

\item{method}{the method to be used for classification}
}
\value{
patterns.tb   a SITS table with the patterns
}
\description{
A sits table has the metadata and data for each time series
<longitude, latitude, start_date, end_date, label, coverage, time_series>
}
