% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_patterns.R
\name{sits_patterns}
\alias{sits_patterns}
\title{Create time series patterns for classification}
\usage{
sits_patterns(samples.tb, method = "gam", ...)
}
\arguments{
\item{samples.tb}{a table in SITS format with a set of labelled time series}

\item{method}{the method to be used for classification}
}
\value{
patterns.tb   a SITS table with the patterns
}
\description{
This function allows the user to select different alternatives to define a set of
patterns, given his samples. The alternatives are:
"gam" - uses a generalised additive model to approximate a smooth spline for each pattern
"dendogram" - uses a herarchical clustering method to group the patterns
"centroids" - uses a positional clustering method to group the patterns
}
\author{
Victor Maus, \email{vwmaus1@gmail.com}

Gilberto Camara, \email{gilberto.camara@inpe.br}

Rolf Simoes, \email{rolf.simoes@inpe.br}
}
