% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_patterns.R
\name{sits_patterns}
\alias{sits_patterns}
\title{Find time series patterns to use using TWDTW (using the dtwSat package)}
\usage{
sits_patterns(samples.tb, bands, method = "gam")
}
\arguments{
\item{samples.tb}{a table in SITS format with time series to be classified using TWTDW}

\item{bands}{string - the bands to be used for classification}

\item{method}{the method to be used for classification}
}
\value{
patterns.tb   a SITS table with the patterns
}
\description{
\code{sits_patterns} returns a sits table with a list of patterns based on samples
}
\details{
A sits table has the metadata and data for each time series
<longitude, latitude, start_date, end_date, label, coverage, time_series>

The patterns are calculated based on a statistical model that tries to find a suitable
for a set of samples

This method is based on the "createPatterns" method of the dtwSat package

Reference: Maus V, Camara G, Cartaxo R, Sanchez A, Ramos FM, de Queiroz GR (2016).
“A Time-Weighted Dynamic Time Warping Method for Land-Use and Land-Cover Mapping.” IEEE
 Journal of Selected Topics in Applied Earth Observations and Remote Sensing, 9(8):3729-3739,
 August 2016. ISSN 1939-1404. doi:10.1109/JSTARS.2016.2517118.
}
