% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_patterns.R
\name{sits_patterns}
\alias{sits_patterns}
\title{Create temporal patterns using a generalised additive model (gam)}
\usage{
sits_patterns(data.tb = NULL, timeline = NULL, start_date = NULL,
  end_date = NULL, freq = 8, formula = y ~ s(x), ...)
}
\arguments{
\item{data.tb}{A tibble in sits format with time series.}

\item{timeline}{A timeline vector with the all dates for the cube.}

\item{start_date}{Starting date (month-day).}

\item{end_date}{End date (month-day).}

\item{freq}{Interval in days for the estimates to be generated.}

\item{formula}{Formula to be applied in the estimate.}

\item{...}{Any additional parameters.}
}
\value{
A sits tibble with the patterns.
}
\description{
This function takes a set of time series samples as input
estimates a set of patterns. The patterns are calculated based in a GAM model.
The idea is to use a formula of type y ~ s(x), where x is a temporal
reference and y if the value of the signal. For each time, there will be as many predictions
as there are sample values. The GAM model predicts a suitable
approximation that fits the assumptions of the statistical model.
By default, the GAM method produces an approximation based on a smooth function.

This method is based on the "createPatterns" method of the dtwSat package, which is also
described in the reference paper.
}
\examples{
\donttest{
# Read a set of samples for two classes
data(cerrado_2classes)
# Estimate a set of patterns (one for each label)
patterns.tb <- sits_patterns(cerrado_2classes)
# Show the patterns
sits_plot(patterns.tb)

# Read a set of samples for the state of Mato Grosso, Brazil, provided by EMBRAPA
data(samples_mt_6bands)
# Estimate a set of patterns (one for each label)
patterns.tb <- sits_patterns(samples_mt_6bands)
# Show the patterns
sits_plot(patterns.tb)
}
}
\references{
Maus V, Camara G, Cartaxo R, Sanchez A, Ramos FM, de Queiroz GR (2016).
A Time-Weighted Dynamic Time Warping Method for Land-Use and Land-Cover Mapping.
IEEE Journal of Selected Topics in Applied Earth Observations and Remote Sensing, 9(8):3729-3739,
August 2016. ISSN 1939-1404. doi:10.1109/JSTARS.2016.2517118.
}
\author{
Victor Maus, \email{vwmaus1@gmail.com}

Gilberto Camara, \email{gilberto.camara@inpe.br}

Rolf Simoes, \email{rolf.simoes@inpe.br}
}
