% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_patterns.R
\name{sits_patterns}
\alias{sits_patterns}
\title{Create time series patterns for classification}
\usage{
sits_patterns(samples.tb, method = "gam", freq = 8, from = NULL,
  to = NULL, formula = y ~ s(x))
}
\arguments{
\item{samples.tb}{a table in SITS format with time series to be classified using TWTDW}

\item{method}{the method to be used for classification}

\item{freq}{int - the interval in days for the estimates to be generated}

\item{from}{starting date of the estimate (month-day)}

\item{to}{end data of the estimated (month-day)}

\item{formula}{the formula to be applied in the estimate}
}
\value{
patterns.tb   a SITS table with the patterns
}
\description{
This function allows the user to select different alternatives to define a set of
patterns, given his samples. The alternatives are:
"gam" - uses a generalised additive model to approximate a smooth spline for each pattern
"dendogram" - uses a herarchical clustering method to group the patterns
"centroids" - uses a positional clustering method to group the patterns
}
\author{
Victor Maus, \email{vwmaus1@gmail.com}

Gilberto Camara, \email{gilberto.camara@inpe.br}

Rolf Simoes, \email{rolf.simoes@inpe.br}
}
