% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_patterns.R
\name{sits_patterns}
\alias{sits_patterns}
\title{Create time series patterns for classification}
\usage{
sits_patterns(data.tb, timeline, pt_method = sits_gam(data.tb = NULL, timeline
  = NULL, from = NULL, to = NULL, freq = 8, formula = y ~ s(x), interval =
  "12 month"))
}
\arguments{
\item{data.tb}{a SITS tibble time series with an alignment column}

\item{timeline}{the valid timeline for the data}

\item{pt_method}{a pattern fitting method}

\item{timeline}{The timeline for the coverage (all acquisition dates)}
}
\value{
result          a model fitted into input data given by train_method parameter
}
\description{
This function allows the user to select different alternatives to define a set of
patterns, given his samples. The alternatives are:
"gam" - uses a generalised additive model to approximate a smooth spline for each pattern
"dendogram" - uses a herarchical clustering method to group the patterns
"centroids" - uses a positional clustering method to group the patterns
}
\author{
Rolf Simoes, \email{rolf.simoes@inpe.br}

Gilberto Camara, \email{gilberto.camara@inpe.br}
}
