% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_plot.R
\name{sits_plot}
\alias{sits_plot}
\title{Plot a set of satellite image time series}
\usage{
sits_plot(data.tb, clusters = NULL, cutree_height = NULL, type = NULL,
  band = NULL, colors = "Set1")
}
\arguments{
\item{data.tb}{A SITS tibble with the list of time series to be plotted}

\item{clusters}{The output of \code{\link[sits]{sits_dendrogram}}. Required for dendrogram plotting.}

\item{cutree_height}{A dashed horizontal line to be drawed indicating the height of dendrogram cutt}

\item{type}{Type of plot to be generated}

\item{band}{The band used for visualisation (optional for sits_plot_classification)}

\item{colors}{Color pallete to be used (based on Color Brewer - default is "Dark2")}
}
\value{
data.tb      Input SITS table (useful for chaining functions)
}
\description{
Given a SITS tibble with a set of time series, plot them.

The following types of plot are supported (parameter "type):
\itemize{
 \item{"all years"}{Plot all samples from the same location together}
 \item{"one by one"}{Plot each time series separately}
 \item{"patterns"}{Plot the patterns for a given set of classes}
 \item{"classification"}{Plot the results of a classification}
 \item{"dendogram"}{Plot a cluster object}
 \item{"together"}{Plot all samples of the same band and label together}
}
If the user does not specify the type of plot (parameter "type"),
the sits_plot function makes an educated guess of what plot is required,
based on the input data. If the input data has less than 30 samples, it
will default to "all years". If there is only one sample per class, it will
default to "patterns". If there are more than 30 samples, it will default to
"together". If there is a cluster object as a parameter, it will plot the dendrogram.
If the input data has predicted values resulting from a classification, it will
plot the classification.
}
\examples{
# Read a set of samples with 2 classes ("Cerrado" and "Pasture")
samples.tb <- readRDS (system.file("extdata/time_series/cerrado_2classes.rds", package = "sits"))
# Plot all the samples together
sits_plot (samples.tb)
# Plot the first 20 samples (defaults to "allyears")
sits_plot (samples.tb[1:20,])
# Plot the patterns
sits_plot (sits_patterns(samples.tb))
\donttest{
# Generate and plot a dendrogram
library (dtwclust)
clusters <- sits_dendrogram (samples.tb, bands = c("ndvi"))
sits_plot (samples.tb, clusters)
}
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
