% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_plot.R
\name{sits_plot}
\alias{sits_plot}
\title{Plot a set of satellite image time series}
\usage{
sits_plot(data.tb, patterns.tb = NULL, band = NULL, colors = "Dark2",
  n_max = 30)
}
\arguments{
\item{data.tb}{A SITS tibble with the list of time series to be plotted}

\item{patterns.tb}{Patterns used for classification (required for sits_plot_predicted)}

\item{band}{Band for displaying the results (useful for sits_plot_predicted)}

\item{colors}{Color pallete to be used (default is "Set1")}

\item{n_max}{Maximum number to visualise individual series
For inputs with more than this number of rows,
input data is plotted together}
}
\value{
data.tb      Input SITS table (useful for chaining functions)
}
\description{
Given a SITS tibble with a set of time series, plot them
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
