% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_plot.R
\name{sits_plot}
\alias{sits_plot}
\title{Plot a set of time series}
\usage{
sits_plot(data.tb, type = "allyears", colors = "Dark2", patterns = NULL)
}
\arguments{
\item{data.tb}{tibble - a SITS table with the list of time series to be plotted}

\item{type}{string - the type of plot to be generated}

\item{colors}{the color pallete to be used (default is "Set1")}
}
\value{
data.tb    tibble - the input SITS table (useful for chaining functions)
}
\description{
\code{sits_plot} plot one or more time series
}
\details{
Plot one SITS time series for a single place and interval
}
\keyword{STIS}

