% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_plot.R
\name{sits_plot}
\alias{sits_plot}
\title{Plot a set of satellite image time series}
\usage{
sits_plot(data.tb, patterns.tb = NULL, distances.tb = NULL, bands = NULL,
  type = "allyears", colors = "RdYlGreen", cluster_obj = NULL,
  cutree_height = NULL)
}
\arguments{
\item{data.tb}{A SITS tibble with the list of time series to be plotted}

\item{patterns.tb}{Patterns used for classification (required for sits_plot_classification)}

\item{distances.tb}{Distances from each series to each attribute (generated by sits_distances)}

\item{bands}{Bands for displaying the results (useful for sits_plot_classification)}

\item{type}{Type of plot to be generated}

\item{colors}{Color pallete to be used (default is "RdYlGreen")}

\item{cluster_obj}{Useful for plotting a dendrogram.}

\item{cutree_height}{A dashed horizontal line to be drawed indicating the height of dendrogram cutting.}
}
\value{
data.tb      Input SITS table (useful for chaining functions)
}
\description{
Given a SITS tibble with a set of time series, plot them
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
