% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_plot.R
\name{sits_plot}
\alias{sits_plot}
\title{Plot a set of satellite image time series}
\usage{
sits_plot(data, band = "ndvi", colors = "Dark2")
}
\arguments{
\item{data}{The data to be plotted (can be a SITS tibble, clusters, or TWDTW matches)}

\item{band}{(optional) The band used for visualisation (optional for sits_plot_classification)}

\item{colors}{(optional) Color pallete to be used (based on Color Brewer - default is "Dark2")}
}
\value{
data          Input SITS table (useful for chaining functions)
}
\description{
Given a SITS tibble with a set of time series, plot them.

The following types of plot are supported:
\itemize{
 \item{"all years" - }{Plot all samples from the same location together (default)}
 \item{"patterns" - }{Plot the patterns for a given set of classes}
 \item{"classification" - }{Plot the results of a classification}
 \item{"dendrogram" - }{Plot a cluster object}
 \item{"together" - }{Plot all samples of the same band and label together}
 \item{"alignments" - }{Plot the alignments (used by TWDTW method)}
 \item{"matches" = }{Plot the matches produced by TWTDW matches}
}
If the user does not specify the type of plot (parameter "type"),
the sits_plot function makes an educated guess of what plot is required,
based on the input data. If the input data has less than 30 samples, it
will default to "all years". If there is only one sample per class, it will
default to "patterns". If there are more than 30 samples, it will default to
"together". If there is a cluster object as a parameter, it will plot the dendrogram.
If the input data has predicted values resulting from a classification, it will
plot the classification.
}
\examples{
\donttest{
# Read a set of samples with 2 classes ("Cerrado" and "Pasture")
data ("cerrado_2classes")
# Plot all the samples together
sits_plot (cerrado_2classes)
# Plot the first 20 samples (defaults to "allyears")
sits_plot (cerrado_2classes[1:20,])
# Plot the patterns
sits_plot (sits_patterns(cerrado_2classes))
}
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
