% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_plot_cluster.R
\name{sits_plot_cluster_info}
\alias{sits_plot_cluster_info}
\title{Plot information about clusters}
\usage{
sits_plot_cluster_info(data, text_title = " Cluster ")
}
\arguments{
\item{data}{Table containing the percentage of mixture between the clusters.}
}
\description{
Plot a bar graph with informations about each cluster.
The percentage of mixture between the clusters.
}
\author{
Lorena Santos \email{lorena.santos@inpe.br}
}
