% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_plot.R
\name{sits_plot_dendrogram}
\alias{sits_plot_dendrogram}
\title{Plot a dendrogram}
\usage{
sits_plot_dendrogram(data.tb, dendro.obj, cutree_height = NULL,
  colors = "RdYlGn")
}
\arguments{
\item{data.tb}{SITS tibble with data used to extract the dendrogram}

\item{dendro.obj}{cluster object. Usually stored by `sits_cluster` function in `.sits_last_object`}

\item{cutree_height}{A dashed horizontal line to be drawed indicating the height of dendrogram cutting.}

\item{colors}{a color scheme as showed in `sits_color_name` function}
}
\description{
Plot an enhanced dendrogram based on the result of \code{\link[sits]{sits_dendrogram}}
}
\examples{
\donttest{
# Read a set of samples with 2 classes ("Cerrado" and "Pasture")
data ("cerrado_2classes")
# Generate and plot a dendrogram
dendro.obj <- sits_dendrogram (cerrado_2classes, bands = c("ndvi"))
sits_plot_dendrogram (cerrado_2classes, dendro.obj)
}
}
\author{
Rolf Simoes, \email{rolf.simoes@inpe.br}
}
