% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_plot.R
\name{sits_plot_raster}
\alias{sits_plot_raster}
\title{Plot a raster classified images}
\usage{
sits_plot_raster(raster_class.tb, title, colors = NULL)
}
\arguments{
\item{raster_class.tb}{a tibble with the metadata for a classified raster object}

\item{title}{string}

\item{colors}{color pallete}
}
\description{
plots a raster using ggplot. This function is used
for showing the same lat/long location in a series of time steps.
}
