% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_ggplot.R
\name{sits_plot_title}
\alias{sits_plot_title}
\title{Create a plot title to use with ggplot}
\usage{
sits_plot_title(row)
}
\arguments{
\item{row}{data row with <longitude, latitude, label> information}
}
\value{
title   string - the title to be used in the plot
}
\description{
\code{sits_plot_title} creates a plot title from row information
}
\seealso{
Other SITS plotting functions: \code{\link{sits_ggplot_together}}
}
\keyword{SITS}

