% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_plot_together.R
\name{sits_plot_together}
\alias{sits_plot_together}
\title{\code{sits_plot_together} plots all samples of a SITS table having the same label}
\usage{
sits_plot_together(data.tb, colors = "Dark2")
}
\arguments{
\item{data.tb}{tibble - a table with a set of time series}
}
\value{
none
}
\description{
\code{sits_plot_together} plots all samples of a SITS table having the same label
}
