% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_pre_train.R
\name{sits_pre_train}
\alias{sits_pre_train}
\title{Pre-train deep learning models for sits}
\usage{
sits_pre_train(samples, deep_method)
}
\arguments{
\item{samples}{Time series samples (as a sits tibble). Labels are optional and
may or may not be used depending on the pre-training method.}

\item{deep_method}{A function defining the deep learning pre-training method.}
}
\value{
A \code{sits_model} object containing the pre-trained deep model.
}
\description{
Self-supervised pre-training for Earth observation time series.
Allows users to pre-train deep learning models using unlabeled or weakly labeled data.
}
\examples{
if (sits_run_examples()) {
    # Example of MAE pre-training
    mae_model <- sits_pre_train(samples_modis_ndvi, sits_mae_pretrain(mask_ratio = 0.5))
}

}
\author{
Alexandre Assuncao \email{alexcarssuncao@gmail.com}
}
