% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_machine_learning.R
\name{sits_predict}
\alias{sits_predict}
\title{Predict class based on the trained models}
\usage{
sits_predict(data.tb, model)
}
\arguments{
\item{data.tb}{a SITS tibble time series}

\item{model}{a model trained by \code{\link[sits]{sits_train_svm}}}

\item{...}{other parameters to pass to \code{\link[sits]{sits_patterns}} and
\code{\link[sits]{sits_TWDTW_matches}}}
}
\description{
Given a SITS tibble time series and a model trained by \code{\link[sits]{sits_train_svm}},
returns a SITS tibble with the classification.
}
\author{
Alexandre Xavier Ywata de Carvalho, \email{alexandre.ywata@ipea.gov.br}

Victor Maus, \email{vwmaus1@gmail.com}

Gilberto Camara, \email{gilberto.camara@inpe.br}

Rolf Simoes, \email{rolf.simoes@inpe.br}
}
