% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_machine_learning.R
\name{sits_predict}
\alias{sits_predict}
\title{Predict class based on the trained models}
\usage{
sits_predict(data.tb = NULL, distances.tb = NULL, ml_model, ...)
}
\arguments{
\item{data.tb}{a SITS tibble time series}

\item{distances.tb}{a tibble with a set of distance metrics to each of the classes}

\item{ml_model}{a model trained by \code{\link[sits]{sits_train}}}

\item{...}{other parameters to be passed to the model function}
}
\value{
data.tb      a SITS tibble with the predicted label
}
\description{
Given a SITS tibble time series and a model trained by \code{\link[sits]{sits_train}},
returns a SITS tibble with the classification.
}
\author{
Alexandre Xavier Ywata de Carvalho, \email{alexandre.ywata@ipea.gov.br}

Rolf Simoes, \email{rolf.simoes@inpe.br}
}
