% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_table.R
\name{sits_prune}
\alias{sits_prune}
\title{Prunes dates of time series to fit an interval}
\usage{
sits_prune(data.tb, interval = "365 days")
}
\arguments{
\item{data.tb}{tibble - input SITS table}

\item{period}{a string describing the interval (in days)}
}
\value{
data1.tb   tibble - the converted SITS table
}
\description{
prunes the times series contained a set of sits tables
to an interval. This function is useful to constrain the different samples
of land cover to the same interval (usually, one year)
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
