% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_machine_learning.R
\name{sits_qda}
\alias{sits_qda}
\title{Train a sits classification model using quadratic discriminant analysis}
\usage{
sits_qda(data.tb = NULL, formula = sits_formula_logref(), ...)
}
\arguments{
\item{data.tb}{Time series with the training samples.}

\item{formula}{Symbolic description of the model to be fit. Package sits offers a set of such formulas (default: sits_formula_logref).}

\item{...}{Other parameters to be passed to MASS::qda function.}
}
\value{
A model function to be passed in sits_predict.
}
\description{
This function receives a tibble with a set of attributes X for each observation Y
These attributes are distance metrics between patterns and observations, obtained by a distance
function in sits (see \code{\link[sits]{.sits_distances}}).
The function performs a quadratic discriminant analysis (qda) to obtain a predictive model.
This function is a front-end to the "qda" method in the "MASS" package.
Please refer to the documentation in that package for more details.
}
\examples{
\donttest{
# Retrieve the set of samples for the Mato Grosso region (provided by EMBRAPA)
# Install the inSitu library
devtools::install_github("e-sensing/inSitu")
library(inSitu)

# select the bands for classification
samples <- inSitu::br_mt_1_8K_9classes_6bands

# find a training model based on the distances and default values (SVM model)
samples_4bands <- sits_select_bands(samples, ndvi, evi, nir, mir)

# Build a machine learning model based on deep learning
ml_model <- sits_train (samples_4bands, sits_qda())

# get a point and classify the point with the ml_model
point.tb <- sits_select_bands(point_mt_6bands, ndvi, evi, nir, mir)
class.tb <- sits_classify(point.tb, ml_model)
sits_plot(class.tb)
}
}
\author{
Alexandre Xavier Ywata de Carvalho, \email{alexandre.ywata@ipea.gov.br}

Rolf Simoes, \email{rolf.simoes@inpe.br}

Gilberto Camara, \email{gilberto.camara@inpe.br}
}
