% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_machine_learning.R
\name{sits_qda}
\alias{sits_qda}
\title{Train a classification model using quadratic discriminant analysis}
\usage{
sits_qda(data = NULL, formula = sits_formula_logref(), ...)
}
\arguments{
\item{data}{Time series with the training samples.}

\item{formula}{Symbolic description of the model to be fit.
(default: sits_formula_logref).}

\item{...}{Other parameters to be passed to MASS::qda function.}
}
\value{
Model fitted to input data
                 (to be passed to \code{\link[sits]{sits_classify}})
}
\description{
This function receives a tibble with a set of attributes X
for each observation Y. These attributes are the values of the time series
for each band. The function performs a quadratic discriminant analysis (qda)
to obtain a predictive model.
This function is a front-end to the "qda" method in the "MASS" package.
Please refer to the documentation in that package for more details.
}
\examples{
# Retrieve the set of samples for Mato Grosso region (provided by EMBRAPA)
# Select the NDVI band
samples_mt_ndvi <- sits_select(samples_mt_4bands, bands = "NDVI")
# Train a QDA model
qda_model <- sits_train(samples_mt_ndvi, sits_qda())
# Classify a point
class <- sits_classify(point_ndvi, qda_model)
}
\author{
Alexandre Ywata de Carvalho, \email{alexandre.ywata@ipea.gov.br}

Rolf Simoes, \email{rolf.simoes@inpe.br}

Gilberto Camara, \email{gilberto.camara@inpe.br}
}
