% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_tibble.R
\name{sits_reduce_imbalance}
\alias{sits_reduce_imbalance}
\title{Reduce imbalance in a set of samples}
\usage{
sits_reduce_imbalance(
  samples,
  oversample = scutr::oversample_smote,
  undersample = scutr::undersample_kmeans,
  n_samples_over = 100,
  n_samples_under = 400
)
}
\arguments{
\item{samples}{Sample set to rebalance}

\item{oversample}{Method used to oversample classes with
limited number of samples}

\item{undersample}{Method used to undersample classes with
large number of samples}

\item{n_samples_over}{Number of samples to oversample
for classes with samples less than this number}

\item{n_samples_under}{Number of samples to undersample
for classes with samples more than this number}
}
\value{
A sits tibble with a fixed quantity of samples.
}
\description{
Takes a sits tibble with different labels and
             returns a new tibble. Deals with class imbalance
             using a synthetic minority oversampling technique (SMOTE).
             By default, does undersampling and oversampling on all classes
             using the techniques avaliable in the SCUTR package.

             By default, all classes will have m/n samples, where m
             is the total number of samples in the original dataset,
             and n is the number of classes.

             Please see the function scutr::SCUT() for more details.
}
\examples{
# Retrieve a set of time series with 2 classes
data(samples_modis_4bands)
# Print the labels of the resulting tibble
sits_labels_summary(samples_modis_4bands)
# Samples the data set
new_data <- sits_reduce_imbalance(samples_modis_4bands)
# Print the labels of the resulting tibble
sits_labels_summary(new_data)
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
