% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_regularize.R
\name{sits_regularize}
\alias{sits_regularize}
\title{Creates a regularized data cube from an irregular one}
\usage{
sits_regularize(
  cube,
  name,
  path_images,
  path_db = NULL,
  period = NULL,
  agg_method = NULL,
  resampling = "bilinear",
  cloud_mask = TRUE
)
}
\arguments{
\item{cube}{A cube whose spacing of observation times is not constant
and will be regularized by the "gdalcubes" packges}

\item{name}{Name of the output data cube}

\item{path_images}{Directory where the regularized images will be
written by \code{gdalcubes}.}

\item{path_db}{Path and name where the \code{gdalcubes}
database will be create. E.g. "my/path/gdalcubes.db"}

\item{period}{ISO8601 time period for regular data cubes
produced by \code{gdalcubes},
with number and unit, e.g., "P16D" for 16 days.
Use "D", "M" and "Y" for days, month and year..}

\item{agg_method}{Method that will be applied by \code{gdalcubes}
for aggregation. Options: "min", "max", "mean",
"median" and "first".}

\item{resampling}{Method to be used by \code{gdalcubes}
for resampling in mosaic operation.
Options: "near", "bilinear", "bicubic"
or others supported by gdalwarp
(see https://gdal.org/programs/gdalwarp.html).}

\item{cloud_mask}{Use cloud band for aggregation by \code{gdalcubes}? (TRUE/FALSE)}
}
\description{
Creates cubes with regular time intervals
 using the gdalcubes package. Cubes are composed using "min", "max", "mean",
"median" or "first" functions. Users need to provide an
 time interval which is used by the composition function.
}
\references{
APPEL, Marius; PEBESMA, Edzer. On-demand processing of data cubes
 from satellite image collections with the gdalcubes library. Data, v. 4,
 n. 3, p. 92, 2019. DOI: 10.3390/data4030092.
}
