% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_regularize.R
\name{sits_regularize}
\alias{sits_regularize}
\title{Build a regular data cube from an irregular one}
\usage{
sits_regularize(
  cube,
  period,
  res,
  output_dir,
  multicores = 1,
  memsize = 4,
  progress = TRUE,
  use_gdalcubes = FALSE,
  ...
)
}
\arguments{
\item{cube}{\code{sits_cube} object whose observation
period and/or spatial resolution is not constant.}

\item{period}{ISO8601-compliant time period for regular
data cubes, with number and unit, where
"D", "M" and "Y" stand for days, month and year;
 e.g., "P16D" for 16 days.}

\item{res}{spatial resolution of regularized images (in meters).}

\item{output_dir}{valid directory for storing regularized images.}

\item{multicores}{number of cores used for regularization;
used for parallel processing of input.}

\item{memsize}{memory available for regularization (in GB).}

\item{progress}{show progress bar?}

\item{use_gdalcubes}{use gdalcubes package? (see details).}

\item{...}{deprecated parameters are controlled with ellipses.}
}
\value{
A \code{sits_cube} object with aggregated images.
}
\description{
Produces regular data cubes for analysis-ready data (ARD)
image collections. Analysis-ready data (ARD) collections available in
AWS, MSPC, USGS and DEAfrica are not regular in space and time.
Bands may have different resolutions,
images may not cover the entire time, and time intervals are not regular.
For this reason, subsets of these collection need to be converted to
regular data cubes before further processing and data analysis.

This function requires users to include the cloud band in their ARD-based
data cubes.
}
\note{
As of sits version 0.16.3 and \code{gdalcubes} version 0.5.1,
      \code{gdalcubes} is having some inconsistent behavior when
      dealing with large data sets. For this reason, \code{sits}
      provides an alternative algorithm for regularization, which is slower
      but is more reliable for large data sets than \code{gdalcubes}.

The aggregation method used in \code{sits_regularize}
      sorts the images based on cloudcover, where images with the fewest
      clouds at the top of the stack. Once
      the stack of images is sorted, the method uses the first valid value to
      create the temporal aggregation.

The input (non-regular) ARD cube needs to include the cloud band for
      the regularization to work.
}
\examples{
{
\dontrun{

# --- Access to the AWS STAC

# define an AWS data cube
  s2_cube <- sits_cube(
      source = "AWS",
      collection = "sentinel-s2-l2a-cogs",
      bands = c("B08", "SCL"),
      tiles = c("20LKP"),
      start_date = "2018-07-18",
      end_date = "2018-08-18"
  )

# create a directory to store the resulting images
dir.create(paste0(tempdir(),"/images/"))

# Build a data cube of equal intervals using the "gdalcubes" package
gc_cube <- sits_regularize(
    cube       = s2_cube,
    output_dir = paste0(tempdir(),"/images/"),
    period     = "P1M",
    res        = 320)
}
}
}
\references{
APPEL, Marius; PEBESMA, Edzer. On-demand processing of data cubes
 from satellite image collections with the gdalcubes library. Data, v. 4,
 n. 3, p. 92, 2019. DOI: 10.3390/data4030092.
}
