% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_labels.R
\name{sits_relabel}
\alias{sits_relabel}
\title{Relabels a sits tibble}
\usage{
sits_relabel(data, conv.lst = list())
}
\arguments{
\item{data}{A sits tibble.}

\item{conv.lst}{A named list used to convert labels to a new value.
Actual labels must be the names of the list elements.
An empty list produces no difference.}
}
\value{
A new sits tibble with modified labels.
}
\description{
Given a sits tibble with a set of labels, and a conversion list
between the original labels and new labels,
returns a new sits tibble whose labels are changed.
}
\examples{
# Read a set of time series with information on deforestation
data(samples_mt_4bands)
# Print the labels
sits_labels(samples_mt_4bands)
# Create a conversion list.
# Three classes will be converted to "Cropland".
conv.lst = list(Soy_Corn = "Cropland",
                Soy_Cotton  = "Cropland",
                Soy_Fallow  = "Cropland",
                Soy_Millet  = "Cropland",
                Soy_Sunflower  = "Cropland",
                Fallow_Cotton  = "Cropland")
# relabel the data
new_data  <- sits_relabel(samples_mt_4bands, conv.lst)
# show the new labels
sits_labels(new_data)
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}

Rolf Simoes, \email{rolf.simoes@inpe.br}
}
