% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_table.R
\name{sits_relabel}
\alias{sits_relabel}
\title{relabels a sits tibble}
\usage{
sits_relabel(data.tb = NULL, conv = NULL)
}
\arguments{
\item{data.tb}{a SITS tibble}

\item{conv}{a conversion of label names for the classes (optional))}
}
\value{
assess         an assessment of validation
}
\description{
Given a SITS tibble with a set of labels, and a conversion list
between the original labels and
new labels, returns a new SITS tibble whose labels are changed..
}
\author{
Victor Maus, \email{vwmaus1@gmail.com}

Gilberto Camara, \email{gilberto.camara@inpe.br}

Rolf Simoes, \email{rolf.simoes@inpe.br}
}
