% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_validate.R
\name{sits_relabel}
\alias{sits_relabel}
\title{re-label classification results}
\usage{
sits_relabel(file, conv)
}
\arguments{
\item{file}{a JSON file contaning the result of a validation procedure}

\item{conv}{a conversion of label names for the classes (optional))}
}
\value{
assess         an assessment of validation
}
\description{
Given a confusion matrix obtained in the validation
procedure, and a conversion list between the original labels and
new labels, returns a new confusion matrix
where classes have been merged.
}
\author{
Victor Maus, \email{vwmaus1@gmail.com}

Gilberto Camara, \email{gilberto.camara@inpe.br}

Rolf Simoes, \email{rolf.simoes@inpe.br}
}
