% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_labels.R
\name{sits_relabel}
\alias{sits_relabel}
\title{relabels a sits tibble}
\usage{
sits_relabel(data.tb, conv.lst = list())
}
\arguments{
\item{data.tb}{a SITS tibble}

\item{conv.lst}{a named list used to convert labels to a new value. Actual labels must be the names of the conv.lst elements.
An empty list produces no difference.}
}
\value{
result.tb      an assessment of validation
}
\description{
Given a SITS tibble with a set of labels, and a conversion list
between the original labels and new labels, returns a new SITS tibble whose labels are changed.
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}

Rolf Simoes, \email{rolf.simoes@inpe.br}
}
