% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_labels.R
\name{sits_relabel}
\alias{sits_relabel}
\title{Relabels a sits tibble}
\usage{
sits_relabel(data, conv.lst = list())
}
\arguments{
\item{data}{A sits tibble.}

\item{conv.lst}{A named list used to convert labels to a new value. Actual labels must be the names of the conv.lst elements.
an empty list produces no difference.}
}
\value{
A new sits tibble with modified labels.
}
\description{
Given a sits tibble with a set of labels, and a conversion list
between the original labels and new labels, returns a new sits tibble whose labels are changed.
}
\examples{
\donttest{
# Read a set of time series with information on deforestation
data(prodes_226_064)
# Print the labels
sits_labels(prodes_226_064)
# Create a conversion list.
# Three classes will be converted to "NonForest".
conv.lst = list(Deforestation_2014 = "NonForest",
                Deforestation_2015 = "NonForest",
                Pasture = "NonForest")
# relabel the data
new_data  <- sits_relabel(prodes_226_064, conv.lst)
# show the new labels
sits_labels(new_data)
}
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}

Rolf Simoes, \email{rolf.simoes@inpe.br}
}
