% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_tibble.R
\name{sits_rename}
\alias{sits_rename}
\title{Names of the bands of a time series}
\usage{
sits_rename(data, names)
}
\arguments{
\item{data}{Valid sits tibble.}

\item{names}{String vector with the names of the new bands.}
}
\value{
A sits tibble with the new names for the bands.
}
\description{
Set the names of the bands of time series in a sits tibble.
}
\examples{
# Retrieve a time series with one band
data(point_ndvi)
# Rename the band
ndvi1.tb <- sits_rename(point_ndvi, names = c("VEG_INDEX"))
# print the names of the new band
sits_bands(ndvi1.tb)
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}

Rolf Simoes, \email{rolf.simoes@inpe.br}
}
