% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_rename.R
\name{sits_rename}
\alias{sits_rename}
\title{Finds the names of the bands of a time series}
\usage{
sits_rename(in.ts, bands_new)
}
\arguments{
\item{in.ts}{a list of SITS time series}

\item{bands}{a list of new band names}
}
\value{
out.ts  a list of renamed SITS time series
}
\description{
\code{sits_bands} finds the names of the bands of time series in a sits table
}

