% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_table.R
\name{sits_rename}
\alias{sits_rename}
\title{Rename bands of a sits table}
\usage{
sits_rename(data.tb, bands_new)
}
\arguments{
\item{data.tb}{a SITS table with a list of SITS time series}

\item{bands_new}{a list of new band names}
}
\value{
out.tb      a SITS table with a list of renamed bands for the time series
}
\description{
replaces the names of the bands of a satellite image time series
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
