% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_tibble_utils.R
\name{sits_rename}
\alias{sits_rename}
\title{names of the bands of a time series}
\usage{
sits_rename(data.tb, names)
}
\arguments{
\item{data.tb}{a valid sits tibble}

\item{names}{string vector with the new bands' names}
}
\value{
data.tb     the sits tibble with the new names for the bands
}
\description{
set the names of the bands of time series in a sits table
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}

Rolf Simoes, \email{rolf.simoes@inpe.br}
}
