% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_machine_learning.R
\name{sits_rfor}
\alias{sits_rfor}
\title{Train a SITS classifiction model using random forest algorithm}
\usage{
sits_rfor(distances.tb = NULL, ntree = 500, ...)
}
\arguments{
\item{distances.tb}{a time series with a set of distance measures for each training sample}

\item{ntree}{Number of trees to grow. This should not be set to too small a number,
to ensure that every input row gets predicted at least a few times. (default: 500)}

\item{...}{other parameters to be passed to `randomForest::randomForest` function}
}
\value{
result          either an model function to be passed in sits_predict or an function prepared that can be called further to compute multinom training model
}
\description{
Use Random Forest algorithm to classify data.
This function is a front-end to the "randomForest" method in the "randomForest" package.
Please refer to the documentation in that package for more details.
}
\examples{
\donttest{
# Retrieve the set of samples for the Mato Grosso region (provided by EMBRAPA)
samples.tb <- readRDS(system.file("extdata/time_series/embrapa_mt.rds", package = "sits"))
samples.tb <- sits_select (samples.tb, bands = c("ndvi", "evi", "nir"))
# get a point with a 16 year time series
point.tb <- readRDS(system.file("extdata/time_series/point.rds", package = "sits"))
point.tb <- sits_select (point.tb, bands = c("ndvi", "evi", "nir"))
# classify the point
class.tb <- sits_classify (point.tb, samples.tb,
       ml_method = sits_rfor())
}
}
\author{
Alexandre Xavier Ywata de Carvalho, \email{alexandre.ywata@ipea.gov.br}

Rolf Simoes, \email{rolf.simoes@inpe.br}

Gilberto Camara, \email{gilberto.camara@inpe.br}
}
