% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_machine_learning.R
\name{sits_rfor}
\alias{sits_rfor}
\title{Train a sits classifiction model using fast random forest algorithm}
\usage{
sits_rfor(data = NULL, num_trees = 2000, importance = "impurity",
  ...)
}
\arguments{
\item{data}{Time series with the training samples.}

\item{num_trees}{Number of trees to grow. This should not be set to too small a number,
to ensure that every input row gets predicted at least a few times. (default: 2000).}

\item{importance}{Variable importance mode, one of 'none', 'impurity', 'impurity_corrected', 'permutation'.
The 'impurity' measure is the Gini index for classification.}

\item{...}{Other parameters to be passed to \code{\link[ranger]{ranger}} function.}
}
\value{
A model fitted to input data to be passed to \code{\link[sits]{sits_classify}}
}
\description{
Use Fast Random Forest algorithm to classify data.
This function is a front-end to the "ranger" method in the "ranger" package.
Please refer to the documentation in that package for more details.
}
\examples{
\donttest{
# Retrieve the set of samples for the Mato Grosso region (provided by EMBRAPA)
samples_2bands <- sits_select_bands(samples_mt_6bands, ndvi, evi)

# Build a machine learning model based on deep learning
ml_model <- sits_train (samples_2bands, sits_rfor())

# get a point and classify the point with the ml_model
point.tb <- sits_select_bands(point_mt_6bands, ndvi, evi)
class.tb <- sits_classify(point.tb, ml_model)
plot(class.tb, bands = c("ndvi", "evi"))
}
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}

Alexandre Xavier Ywata de Carvalho, \email{alexandre.ywata@ipea.gov.br}

Rolf Simoes, \email{rolf.simoes@inpe.br}
}
