% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_machine_learning.R
\name{sits_rfor}
\alias{sits_rfor}
\title{Train SITS classifiction models}
\usage{
sits_rfor(distances.tb = NULL, ntree = 500, ...)
}
\arguments{
\item{distances.tb}{a time series with a set of distance measures for each training sample}

\item{ntree}{Number of trees to grow. This should not be set to too small a number,
to ensure that every input row gets predicted at least a few times. (default: 500)}

\item{...}{other parameters to be passed to `randomForest::randomForest` function}
}
\value{
result          either an model function to be passed in sits_predict or an function prepared that can be called further to compute multinom training model
}
\description{
Use Random Forest algorithm to classify data.
This function is a front-end to the "randomForest" method in the "randomForest" package.
Please refer to the documentation in that package for more details.
}
\author{
Alexandre Xavier Ywata de Carvalho, \email{alexandre.ywata@ipea.gov.br}

Rolf Simoes, \email{rolf.simoes@inpe.br}

Gilberto Camara, \email{gilberto.camara@inpe.br}
}
