% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_machine_learning.R
\name{sits_rfor}
\alias{sits_rfor}
\title{Train a sits classifiction model using fast random forest algorithm}
\usage{
sits_rfor(data.tb = NULL, num.trees = 2000, ...)
}
\arguments{
\item{data.tb}{Time series with the training samples.}

\item{num.trees}{Number of trees to grow. This should not be set to too small a number,
to ensure that every input row gets predicted at least a few times. (default: 2000).}

\item{...}{Other parameters to be passed to \code{\link[ranger]{ranger}} function.}
}
\value{
Either a model function to be passed in sits_predict or an function prepared that can be called further to compute multinom training model.
}
\description{
Use Fast Random Forest algorithm to classify data.
This function is a front-end to the "ranger" method in the "ranger" package.
Please refer to the documentation in that package for more details.
}
\examples{
\donttest{
# Retrieve the set of samples for the Mato Grosso region (provided by EMBRAPA)
# Install the inSitu library
devtools::install_github("e-sensing/inSitu")
library(inSitu)

# select the bands for classification
samples <- inSitu::br_mt_1_8K_9classes_6bands

# find a training model based on the distances and default values (SVM model)
samples_4bands <- sits_select_bands(samples, ndvi, evi, nir, mir)

# Build a machine learning model based on deep learning
ml_model <- sits_train (samples_4bands, sits_rfor())

# get a point and classify the point with the ml_model
point.tb <- sits_select_bands(point_mt_6bands, ndvi, evi, nir, mir)
class.tb <- sits_classify(point.tb, ml_model)
sits_plot(class.tb)
}
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}

Alexandre Xavier Ywata de Carvalho, \email{alexandre.ywata@ipea.gov.br}

Rolf Simoes, \email{rolf.simoes@inpe.br}
}
