% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_machine_learning.R
\name{sits_rfor}
\alias{sits_rfor}
\title{Train random forest models}
\usage{
sits_rfor(samples = NULL, num_trees = 200, nodesize = 1, ...)
}
\arguments{
\item{samples}{Time series with the training samples.}

\item{num_trees}{Number of trees to grow.
This should not be set to too small a number,
to ensure that every input row gets predicted
at least a few times (default: 200).}

\item{nodesize}{Minimum size of terminal nodes
(default 1 for classification).}

\item{...}{Other parameters to be passed
to `randomForest::randomForest` function.}
}
\value{
Model fitted to input data
                        (to be passed to \code{\link[sits]{sits_classify}}).
}
\description{
Use Random Forest algorithm to classify samples.
This function is a front-end to the "randomForest" package.
Please refer to the documentation in that package for more details.
}
\examples{
# Retrieve the set of samples for the Mato Grosso region
samples_MT_ndvi <- sits_select(samples_modis_4bands, bands = "NDVI")
# Build a random forest model
rfor_model <- sits_train(samples_MT_ndvi, sits_rfor(num_trees = 200))
# get a point with a 16 year time series
point_ndvi <- sits_select(point_mt_6bands, bands = "NDVI")
# classify the point
class.tb <- sits_classify(point_ndvi, rfor_model)
}
\author{
Alexandre Ywata de Carvalho, \email{alexandre.ywata@ipea.gov.br}

Rolf Simoes, \email{rolf.simoes@inpe.br}

Gilberto Camara, \email{gilberto.camara@inpe.br}
}
