% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_machine_learning.R
\name{sits_rfor}
\alias{sits_rfor}
\title{Train a sits classifiction model using fast random forest algorithm}
\usage{
sits_rfor(data.tb = NULL, num.trees = 2000, ...)
}
\arguments{
\item{data.tb}{Time series with the training samples.}

\item{num.trees}{Number of trees to grow. This should not be set to too small a number,
to ensure that every input row gets predicted at least a few times. (default: 2000).}

\item{...}{Other parameters to be passed to \code{\link[ranger]{ranger}} function.}
}
\value{
Either a model function to be passed in sits_predict or an function prepared that can be called further to compute multinom training model.
}
\description{
Use Fast Random Forest algorithm to classify data.
This function is a front-end to the "ranger" method in the "ranger" package.
Please refer to the documentation in that package for more details.
}
\examples{
\donttest{
# Retrieve the set of samples for the Mato Grosso region (provided by EMBRAPA)
data(samples_mt_ndvi)
# Build a random forest model
rfor_model <- sits_train(samples_mt_ndvi, sits_rfor())
# get a point with a 16 year time series
data(point_ndvi)
# classify the point
class.tb <- sits_classify (point_ndvi, rfor_model)
# plot the classification
sits_plot(class.tb)
}
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}

Alexandre Xavier Ywata de Carvalho, \email{alexandre.ywata@ipea.gov.br}

Rolf Simoes, \email{rolf.simoes@inpe.br}
}
