% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_machine_learning.R
\name{sits_rfor}
\alias{sits_rfor}
\title{Train a SITS classifiction model using random forest algorithm}
\usage{
sits_rfor(data.tb = NULL, normalize = FALSE, ntree = 2000, nodesize = 1,
  multicores = 2, ...)
}
\arguments{
\item{data.tb}{time series with the training samples}

\item{normalize}{should the data be normalized?}

\item{ntree}{number of trees to grow. This should not be set to too small a number,
to ensure that every input row gets predicted at least a few times. (default: 2000)}

\item{nodesize}{minimum size of terminal nodes (default 1 for classification)}

\item{multicores}{number of cores (for parallel processing)}

\item{...}{other parameters to be passed to `randomForest::randomForest` function}
}
\value{
result          either an model function to be passed in sits_predict or an function prepared that can be called further to compute multinom training model
}
\description{
Use Random Forest algorithm to classify data.
This function is a front-end to the "randomForest" method in the "randomForest" package.
Please refer to the documentation in that package for more details.
}
\examples{
\donttest{
# Retrieve the set of samples for the Mato Grosso region (provided by EMBRAPA)
data(samples_MT_ndvi)
# Build a random forest model
rfor_model <- sits_train(samples_MT_ndvi, sits_rfor())
# get a point with a 16 year time series
data(point_ndvi)
# classify the point
class.tb <- sits_classify (point_ndvi, rfor_model)
}
}
\author{
Alexandre Xavier Ywata de Carvalho, \email{alexandre.ywata@ipea.gov.br}

Rolf Simoes, \email{rolf.simoes@inpe.br}

Gilberto Camara, \email{gilberto.camara@inpe.br}
}
