% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_utils.R
\name{sits_run_examples}
\alias{sits_run_examples}
\title{Informs if sits examples should run}
\usage{
sits_run_examples()
}
\value{
A logical value
}
\description{
This function informs if sits examples should run.
This is useful to avoid running slow examples in CRAN environment.
}
\examples{
# unset active tests
Sys.setenv("R_CONFIG_RUN_EXAMPLES"="NO")
# result should be false
isFALSE(sits_run_examples())
# set active tests
Sys.setenv("R_CONFIG_RUN_EXAMPLES"="YES")
# result should be true
isTRUE(sits_run_examples())
}
