% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_utils.R
\name{sits_run_examples}
\alias{sits_run_examples}
\title{Informs if sits examples should run}
\usage{
sits_run_examples()
}
\value{
A logical value
}
\description{
This function informs if sits examples should run.
To run the examples, set "SITS_RUN_EXAMPLES" to "YES" using
Sys.setenv("SITS_RUN_EXAMPLES" = "YES")
To come back to the default behaviour, please set
Sys.setenv("SITS_RUN_EXAMPLES" = "NO")
}
\examples{
if (sits_run_examples()) {
    message("use devtools::run_examples() to run the examples")
} else {
    message(paste0(
        "Set the envvar SITS_RUN_EXAMPLES to YES",
        "to run the examples"
    ))
}
}
