% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_utils.R
\name{sits_run_tests}
\alias{sits_run_tests}
\title{Informs if sits tests should run}
\usage{
sits_run_tests()
}
\value{
TRUE/FALSE
}
\description{
To run the tests, set "SITS_RUN_TESTS" environment to TRUE using
Sys.setenv("SITS_RUN_TESTS" = TRUE)
To come back to the default behaviour, please set
Sys.setenv("SITS_RUN_TESTS" = FALSE)
}
