% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_utils.R
\name{sits_run_tests}
\alias{sits_run_tests}
\title{Informs if sits tests should run}
\usage{
sits_run_tests()
}
\value{
TRUE/FALSE
}
\description{
This function informs if sits test should run.
Useful to avoid running slow tests in CRAN environment.
Behaviour controlled by environmental variable R_CONFIG_ACTIVE_TESTS
}
\examples{
if (sits_run_examples()) {
# recover config state
config_tests <- sits_run_tests()
# set active tests to FALSE
sits_config(run_tests = FALSE)
isFALSE(sits_run_tests())
# recover config state
# set active tests
sits_config(run_tests = TRUE)
# result should be true
isTRUE(sits_run_tests())
# restore previous state
sits_config(run_tests = config_tests)
}

}
