% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_utils.R
\name{sits_run_tests}
\alias{sits_run_tests}
\title{Informs if sits tests should run}
\usage{
sits_run_tests()
}
\value{
TRUE/FALSE
}
\description{
To run the tests, set "SITS_RUN_TESTS" environment to "YES" using
Sys.setenv("SITS_RUN_TESTS" = "YES")
To come back to the default behaviour, please set
Sys.setenv("SITS_RUN_TESTS" = "NO")
}
\examples{
if (sits_run_tests()) {
    message("Tests will be run")
} else {
    message(paste(
        "Set the environmental variable SITS_RUN_TESTS to YES",
        "to run the tests or use devtools::run_tests()"
    ))
}

}
