% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_utils.R
\name{sits_run_tests}
\alias{sits_run_tests}
\title{Informs if sits tests should run}
\usage{
sits_run_tests()
}
\value{
TRUE/FALSE
}
\description{
This function informs if sits examples should run.
To run the examples, set "SITS_RUN_TESTS" environment
variable to "YES" using
Sys.setenv("SITS_RUN_TESTS" = "YES")
To come back to the default behaviour, please unset
the enviroment variable
Sys.unsetenv("SITS_RUN_TESTS")
}
\examples{
#' if (sits_run_tests()){
    message("use devtools::run_tests()")
}
else{
    message("Set the environmental variable "SITS_RUN_TESTS" to "YES"
    to run the tests or use devtools::run_tests()")
}

}
