% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_sample_functions.R
\name{sits_sample}
\alias{sits_sample}
\alias{sits_sample.default}
\title{Sample a time series or a data cube}
\usage{
sits_sample(data, ...)

\method{sits_sample}{default}(data, ...)
}
\arguments{
\item{data}{Sits time series tibble or data cube}

\item{...}{Specific parameters for method}
}
\value{
A sits tibble.
}
\description{
Takes samples from
             returns a new tibble. For a given field as a group criterion,
             this new tibble contains a percentage
             of the total number of samples per group.
             If frac > 1 , all sampling will be done with replacement.
}
\author{
Rolf Simoes, \email{rolfsimoes@gmail.com}
}
