% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_tibble_functions.R
\name{sits_sample}
\alias{sits_sample}
\title{Sample a percentage of a time series}
\usage{
sits_sample(data.tb, n = NULL, frac = NULL)
}
\arguments{
\item{data.tb}{input SITS table}

\item{n}{the quantity of samples to pick from a given group of data.}

\item{frac}{the percentage of samples to pick from a given group of data.}
}
\value{
result.tb  the new SITS table with a fixed quantity of samples of informed labels and all other
}
\description{
takes a sits table with different labels and
             returns a new table. For a given field as a group criterion, this new table contains a given number or percentage
             of the total number of samples per group. Parameter n indicantes the number of random samples with reposition.
             Parameter frac indicates a fraction of random samples without reposition. If frac > 1, the sampling is taken with reposition.
}
\author{
Rolf Simoes, \email{rolf.simoes@inpe.br}
}
