% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_save.R
\name{sits_save}
\alias{sits_save}
\title{Save the contents of an in-memory database in a JSON file}
\usage{
sits_save(source, json_file = NULL)
}
\arguments{
\item{source}{tibble - an existing SITS table}

\item{json_file}{string - name of the JSON file to be written}
}
\value{
table       the same table (for chaining functions)
}
\description{
\code{sits_toJSON} stores a SITS table in a JSON file
}
\details{
The set of time series from a SITS table can be saved in JSON format
and later retrieved for further use
}
\keyword{STIS}
