% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_save.R
\name{sits_save}
\alias{sits_save}
\title{Save the contents of an in-memory database in a JSON file}
\usage{
sits_save(source, json_file = NULL)
}
\arguments{
\item{json_file}{string - name of the JSON file to be written}

\item{table}{tibble - an existing SITS table}
}
\value{
table      the same table (for chaining functions)
}
\description{
\code{sits_toJSON} stores a SITS table in a JSON file
}
\details{
The set of time series from a SITS table can be saved in JSON format
and later retrieved for further use
}
\examples{
table <- sits_toJSON (table = "my.db", json_file = "myjson_file.json")
}
\seealso{
Other STIS main functions: \code{\link{sits_align}},
  \code{\link{sits_centroids}},
  \code{\link{sits_dendogram}},
  \code{\link{sits_plot_together}}, \code{\link{sits_plot}}
}
\keyword{STIS}
