% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_indexes.R
\name{sits_savi}
\alias{sits_savi}
\title{Builds soil-adjusted vegetation index}
\usage{
sits_savi(data)
}
\arguments{
\item{data}{Valid sits tibble.}
}
\value{
A sits tibble with the SAVI index.
}
\description{
Adds new tasseled cap bands.
}
\examples{
# Retrieve data for time series with label samples in Mato Grosso in Brazil
data (samples_mt_6bands)
# Generate a new image with the tasseled cap
savi.tb <- sits_savi(samples_mt_6bands)

}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
