% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_time_series.R
\name{sits_select}
\alias{sits_select}
\title{Filter bands on a SITS table}
\usage{
sits_select(data.tb, ...)
}
\arguments{
\item{data.tb}{a sits table with the time series of the selected bands}

\item{...}{one or more band names separated by commas. See `dplyr::select` help for more details.}
}
\value{
data.tb     a tibble in SITS format with the selected bands
}
\description{
returns a sits table with the selected bands
}
\author{
Rolf Simoes, \email{rolf.simoes@inpe.br}

Gilberto Camara, \email{gilberto.camara@inpe.br}
}
