% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_tibble_functions.R
\name{sits_select}
\alias{sits_select}
\title{General selection criteria for subsetting a SITS table}
\usage{
sits_select(data.tb, ...)
}
\arguments{
\item{data.tb}{a sits tibble with the time series}

\item{...}{`name == value` logical expressions, where `name` is any SITS column name
or 'bands = c(band names)' for selection of a subset of bands
See `dplyr::filter` help for more details..}
}
\value{
data.tb     a tibble in SITS format with the selected bands
}
\description{
This is a general selection function for extracting any subset of a sits tibble
             It supports spatial selection by lat/long bounding boxes, names of labels,
             start and end dates, and subset of bands
}
\author{
Rolf Simoes, \email{rolf.simoes@inpe.br}

Gilberto Camara, \email{gilberto.camara@inpe.br}
}
