% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_table.R
\name{sits_select}
\alias{sits_select}
\title{Filter bands on a SITS table}
\usage{
sits_select(data.tb, bands)
}
\arguments{
\item{data.tb}{a sits table with the time series of the selected bands}

\item{bands}{a vector of bands}
}
\value{
table  a tibble in SITS format with the selected bands
}
\description{
returns a sits table with the selected bands
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
