% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_table.R
\name{sits_select}
\alias{sits_select}
\title{Filter bands on a SITS table}
\usage{
sits_select(data.tb, bands)
}
\arguments{
\item{data.tb}{a sits table with the time series of the selected bands}

\item{bands}{a vector of bands}
}
\value{
table  a tibble in SITS format with the selected bands
}
\description{
\code{sits_select} returns a sits table with the selected bands
}
\seealso{
Other STIS table functions: \code{\link{sits_align}},
  \code{\link{sits_bands}}, \code{\link{sits_dates}},
  \code{\link{sits_group_bylatlong}},
  \code{\link{sits_rename}}, \code{\link{sits_table}},
  \code{\link{sits_values_cols}},
  \code{\link{sits_values_rows}}
}

