% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_tibble.R
\name{sits_select}
\alias{sits_select}
\title{Filter bands on a sits tibble}
\usage{
sits_select(data, bands)
}
\arguments{
\item{data}{A sits tibble metadata and data on time series.}

\item{bands}{Character vector with the names of the bands}
}
\value{
A tibble in sits format with the selected bands.
}
\description{
Returns a sits tibble with the selected bands.
}
\examples{
# Retrieve a set of time series with 2 classes
data(cerrado_2classes)
# Print the original bands
sits_bands(cerrado_2classes)
# Select only the NDVI band
data <- sits_select (cerrado_2classes, bands = c("NDVI"))
# Print the labels of the resulting tibble
sits_bands(data)
}
\author{
Rolf Simoes, \email{rolf.simoes@inpe.br}
}
