% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_data_generics.R
\name{sits_select}
\alias{sits_select}
\title{Filter bands on a data set (tibble or cube)}
\usage{
sits_select(data, bands)
}
\arguments{
\item{data}{A sits tibble or data cube}

\item{bands}{Character vector with the names of the bands}
}
\value{
data sets with only the bands selected
}
\description{
For details see:
\itemize{
 \item{"time series": }{see \code{\link{sits_select.sits}}}
 \item{"data cube": }{see \code{\link{sits_select.cube}}}
}
}
\author{
Rolf Simoes, \email{rolf.simoes@inpe.br}
}
