% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_tibble.R
\name{sits_select_bands}
\alias{sits_select_bands}
\title{Filter bands on a sits tibble}
\usage{
sits_select_bands(data, ...)
}
\arguments{
\item{data}{A sits tibble metadata and data on time series.}

\item{...}{Names of the selected bands.}
}
\value{
A tibble in sits format with the selected bands.
}
\description{
Returns a sits tibble with the selected bands.
}
\examples{
# Retrieve a set of time series with 2 classes
data(cerrado_2classes)
# Print the original bands
sits_bands(cerrado_2classes)
# Select only the NDVI band
data <- sits_select_bands(cerrado_2classes, NDVI)
# Print the labels of the resulting tibble
sits_bands(data)
}
\author{
Rolf Simoes, \email{rolf.simoes@inpe.br}
}
