% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_tibble_functions.R
\name{sits_select_bands}
\alias{sits_select_bands}
\title{Filter bands on a SITS table}
\usage{
sits_select_bands(data.tb, bands)
}
\arguments{
\item{data.tb}{a sits table with the time series of the selected bands}

\item{bands}{a vector of bands}
}
\value{
result.tb   a tibble in SITS format with the selected bands
}
\description{
returns a sits table with the selected bands
}
\examples{
# Retrieve a set of time series with 2 classes
data (cerrado_2classes)
# Print the original bands
sits_bands (cerrado_2classes)
# Select only the "ndvi" band
data.tb <- sits_select_bands (cerrado_2classes, bands = c("ndvi"))
# Print the labels of the resulting table
sits_bands (data.tb)
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
