% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_services.R
\name{sits_services}
\alias{sits_services}
\title{Provides information about the time series services available}
\usage{
sits_services()
}
\description{
Uses the configuration file to print information about the services, products and coverages.
}
