% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_filters.R
\name{sits_sgolay}
\alias{sits_sgolay}
\title{Smooth the time series using Savitsky-Golay filter}
\usage{
sits_sgolay(data.tb = NULL, order = 3, length = 5, scaling = 1,
  bands_suffix = "sg")
}
\arguments{
\item{data.tb}{A tibble with time series data and metadata.}

\item{order}{Filter order (integer).}

\item{length}{Filter length (must be odd)}

\item{scaling}{Time scaling (integer).}

\item{bands_suffix}{Suffix to be appended to the smoothed filters.}
}
\value{
A tibble with smoothed sits time series.
}
\description{
The algorithm searches for an optimal polynomial describing the warping.
The degree of smoothing depends on the filter order (usually 3.0).
The user can set the order of the polynomial using the parameter `order` (default = 3), +
the size of the temporal window with the parameter `length` (default = 5),
and the temporal expansion with the parameter `scaling`.
}
\examples{
\donttest{
#' # Retrieve a time series with values of NDVI
data(point_ndvi)
# Filter the point using the Savitsky Golay smoother
point_sg.tb <- sits_sgolay (point_ndvi, order = 3, length  = 5)
# Plot the two points to see the smoothing effect
sits_plot(sits_merge(point_ndvi, point_sg.tb))
}
}
