% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_filters.R
\name{sits_sgolay}
\alias{sits_sgolay}
\title{Smooth the time series using Savitsky-Golay filter}
\usage{
sits_sgolay(data.tb, order = 3, scale = 1, bands_suffix = "sg")
}
\arguments{
\item{data.tb}{The SITS tibble containing the original time series}

\item{order}{filter order}

\item{scale}{time scaling}

\item{bands_suffix}{the suffix to be appended to the smoothed filters}
}
\value{
output.tb a tibble with smoothed sits time series
}
\description{
The algorithm searches for an optimal polynomial describing the warping.
The degree of smoothing depends on the filter order (usually 3.0)
Use lambda = 0.5 for very slight smoothing and lambda = 5.0 for strong smoothing
}
\examples{
# Retrieve a time series with values of NDVI
data(point_ndvi)
# Filter the point using the Savitsky Golay smoother
point_sg.tb <- sits_sgolay (point_ndvi, order = 3, scale = 2)
# Plot the two points to see the smoothing effect
sits_plot(sits_merge(point_ndvi, point_sg.tb))

}
