% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_filters.R
\name{sits_sgolay}
\alias{sits_sgolay}
\title{Smooth the time series using Savitsky-Golay filter}
\usage{
sits_sgolay(data = NULL, order = 3, length = 5, scaling = 1, bands_suffix = "")
}
\arguments{
\item{data}{A tibble with time series data and metadata.}

\item{order}{Filter order (integer).}

\item{length}{Filter length (must be odd)}

\item{scaling}{Time scaling (integer).}

\item{bands_suffix}{Suffix to be appended to the smoothed filters.}
}
\value{
A tibble with smoothed sits time series.
}
\description{
An optimal polynomial for warping a time series.
The degree of smoothing depends on the filter order (usually 3.0).
The order of the polynomial uses the parameter `order` (default = 3),
the size of the temporal window uses the parameter `length` (default = 5),
and the temporal expansion uses the parameter `scaling`.
}
\examples{
\donttest{
#' # Retrieve a time series with values of NDVI
data(point_ndvi)
# Filter the point using the Savitsky Golay smoother
point_sg <- sits_filter(point_ndvi, sits_sgolay (order = 3,
                                    length  = 5, bands_suffix = "sg"))
# Plot the two points to see the smoothing effect
plot(sits_merge(point_ndvi, point_sg))
}
}
\references{
A. Savitzky, M. Golay, "Smoothing and Differentiation of Data by
Simplified Least Squares Procedures".
Analytical Chemistry, 36 (8): 1627–39, 1964.
}
