% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_classify_ts.R
\name{sits_show_prediction}
\alias{sits_show_prediction}
\title{Shows the predicted labels for a classified tibble}
\usage{
sits_show_prediction(class)
}
\arguments{
\item{class}{A SITS tibble that has been classified}
}
\value{
returns a tibble with the columns "from", "to", "class"
}
\description{
This function takes a tibble with a classified time series
by a machine learning method and displays the result.
}
\examples{
# Retrieve the samples for Mato Grosso
# select band "NDVI"
samples_ndvi <- sits_select(samples_modis_4bands, bands = "NDVI")
# select a random forest model
rfor_model <- sits_train(samples_ndvi, sits_rfor(num_trees = 200))
# classify the point
point_ndvi <- sits_select(point_mt_6bands, bands = "NDVI")
point_class <- sits_classify(point_ndvi, rfor_model)
# show the prediction
sits_show_prediction(point_class)
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
\keyword{internal}
