% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_classification.R
\name{sits_show_prediction}
\alias{sits_show_prediction}
\title{Shows the predicted labels for a classified tibble}
\usage{
sits_show_prediction(class.tb)
}
\arguments{
\item{class.tb}{A SITS tibble that has been classified}
}
\value{
returns a tibble with the columns "from", "to", "class"
}
\description{
This function takes a tibble with a classified time series
by a machine learning method and displays the result.
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
