% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_smooth.R
\name{sits_smooth}
\alias{sits_smooth}
\title{Smooth the time series using Whittaker smoother (based on PTW package)}
\usage{
sits_smooth(table_in, lambda = 0.5)
}
\arguments{
\item{table_in}{a string - the name of the database with original time series  (must exist)}

\item{lambda}{double   - the smoothing factor to be applied}
}
\value{
a tibble with smoothed sits time series
}
\description{
\code{sits_smooth} returns a database with raw smoothed sits time series
}
\details{
The input and the output are two databases. The input database must exist.

The algorithm searches for an optimal polynomial describing the warping.
It is possible to align one sample to a reference,
several samples to the same reference, or several samples to several references.

The degree of smoothing depends on smoothing factor lambda (usually from 0.5 to 10.0)
Use lambda = 0.5 for very slight smoothing and lambda = 5.0 for strong smoothing
}
\keyword{sits}
