% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_smooth.R
\name{sits_smooth}
\alias{sits_smooth}
\alias{sits_smooth.bayes}
\alias{sits_smooth.gaussian}
\alias{sits_smooth.bilateral}
\title{Smooth probability cubes with spatial predictors}
\usage{
sits_smooth(cube, type = "bayes", ...)

\method{sits_smooth}{bayes}(
  cube,
  type = "bayes",
  ...,
  window_size = 5,
  smoothness = 20,
  covar = FALSE,
  multicores = 2,
  memsize = 4,
  output_dir = ".",
  version = "v1"
)

\method{sits_smooth}{gaussian}(
  cube,
  type = "gaussian",
  ...,
  window_size = 5,
  sigma = 1,
  multicores = 2,
  memsize = 4,
  output_dir = ".",
  version = "v1"
)

\method{sits_smooth}{bilateral}(
  cube,
  type = "bilateral",
  ...,
  window_size = 5,
  sigma = 8,
  tau = 0.1,
  multicores = 2,
  memsize = 4,
  output_dir = ".",
  version = "v1"
)
}
\arguments{
\item{cube}{Probability data cube}

\item{type}{Type of smoothing}

\item{...}{Parameters for specific functions}

\item{window_size}{Size of the neighbourhood.}

\item{smoothness}{Estimated variance of logit of class probabilities
(Bayesian smoothing parameter). It can be either
a matrix or a scalar.}

\item{covar}{a logical argument indicating if a covariance
matrix must be computed as the prior covariance
for bayesian smoothing.}

\item{multicores}{Number of cores to run the smoothing function}

\item{memsize}{Maximum overall memory (in GB) to run the
smoothing.}

\item{output_dir}{Output directory for image files}

\item{version}{Version of resulting image
(in the case of multiple tests)}

\item{sigma}{Standard deviation of the spatial Gaussian kernel
(for gaussian and bilateral smoothing)}

\item{tau}{Standard deviation of the class probs value
(for bilateral smoothing)}
}
\value{
A tibble with metadata about the output raster objects.
}
\description{
Takes a set of classified raster layers with probabilities,
             whose metadata is]created by \code{\link[sits]{sits_cube}},
             and applies a smoothing function. There are three options,
             defined by the "type" parameter:
\itemize{
   \item{"bayes": }{Use a bayesian smoother}
   \item{"gaussian": }{Use a gaussian smoother}
   \item{"bilateral: }{Use a bilateral smoother}

}
}
\examples{
\dontrun{
# Retrieve the samples for Mato Grosso
# select band "ndvi"

samples_ndvi <- sits_select(samples_modis_4bands, bands = "NDVI")

# select a random forest model
rfor_model <- sits_train(samples_ndvi, sits_rfor(num_trees = 500))

# create a data cube based on the information about the files
data_dir <- system.file("extdata/raster/mod13q1", package = "sits")
cube <- sits_cube(
  source = "BDC",
  collection = "MOD13Q1-6",
  data_dir = data_dir,
  delim = "_",
  parse_info = c("X1", "X2", "tile", "band", "date")
)

# classify the raster image
probs_cube <- sits_classify(cube,
  ml_model = rfor_model,
  output_dir = tempdir(),
  memsize = 4, multicores = 2
)

# smooth the result with a bayesian filter
bayes_cube <- sits_smooth(probs_cube,
  type = "bayes", output_dir = tempdir()
)

# smooth the result with a gaussian filter
gauss_cube <- sits_smooth(probs_cube,
  type = "gaussian", output_dir = tempdir()
)

# smooth the result with a bilateral filter
bil_cube <- sits_smooth(probs_cube,
  type = "bilateral", output_dir = tempdir()
)
}

}
\references{
K. Schindler, "An Overview and Comparison of Smooth Labeling
            Methods for Land-Cover Classification",
            IEEE Transactions on Geoscience and Remote Sensing,
            50 (11), 4534-4545, 2012 (for gaussian and bilateral smoothing)
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}

Rolf Simoes, \email{rolf.simoes@inpe.br}
}
