% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_kohonen.R
\name{sits_som_clean_samples}
\alias{sits_som_clean_samples}
\title{Clean samples}
\usage{
sits_som_clean_samples(
  som_cluster,
  prob_label_change = 0.8,
  min_cluster_prob = 0.6
)
}
\arguments{
\item{som_cluster}{Result of \code{\link[sits]{sits_som_map}}}

\item{prob_label_change}{Threshold of probability to change the label}

\item{min_cluster_prob}{Minimum probability for a good sample
to belong to a cluster}
}
\value{
Returns a sits tibble with a
                              new subset of samples and a new column
                              with the probability that each sample belongs
                              to the class described in label.
}
\description{
This function remove sample that do not have good quality based
on the statistics of the result of the clustering using SOM.
}
\examples{
\donttest{
# Read a set of samples
# Get a new subset of samples evaluated by clustering methods
som_cluster <- sits_som_map(prodes_226_064, grid_xdim = 10, grid_ydim = 10,
       distance = "euclidean", iterations = 50)
new_samples <- sits_som_clean_samples(som_cluster)
}
}
\author{
Lorena Santos, \email{lorena.santos@inpe.br}

Karine Ferreira. \email{karine.ferreira@inpe.br}
}
