% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_kohonen.R
\name{sits_som_clean_samples}
\alias{sits_som_clean_samples}
\title{Clean samples}
\usage{
sits_som_clean_samples(
  som_map,
  conditonal_threshold = 0.6,
  posterior_threshold = 0.6
)
}
\arguments{
\item{som_map}{An object returned
by \code{\link[sits]{sits_som_map}}}

\item{conditonal_threshold}{Threshold of conditional probability
(frequency of samples assigned to the same SOM neuron)}

\item{posterior_threshold}{Threshold of posterior probability
(influenced by the SOM neighborhood)}
}
\value{
A list with two sits tibbles
                                The first tibble has clean samples
                                The second has samples that need to be analysed
}
\description{
This function remove sample that do not have good quality based
on the statistics of the result of the clustering using SOM. It also provides
samples that should be analysed.
}
\examples{
\dontrun{
# Read a set of samples
# Get a new subset of samples evaluated by clustering methods
samples_mt <- sits_select(samples_mt_4bands, bands = c("NDVI"))
som_map <- sits_som_map(samples_mt, grid_xdim = 5, grid_ydim = 5,
       distance = "euclidean", iterations = 10)
new_samples <- sits_som_clean_samples(som_map)
}
}
\author{
Lorena Santos, \email{lorena.santos@inpe.br}

Karine Ferreira. \email{karine.ferreira@inpe.br}
}
