% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_som.R
\name{sits_som_clean_samples}
\alias{sits_som_clean_samples}
\title{Clean samples}
\usage{
sits_som_clean_samples(
  som_map,
  prior_threshold = 0.6,
  posterior_threshold = 0.6,
  keep = c("clean", "analyze")
)
}
\arguments{
\item{som_map}{An object returned
by \code{\link[sits]{sits_som_map}}}

\item{prior_threshold}{Threshold of conditional probability
(frequency of samples assigned to the same SOM neuron)}

\item{posterior_threshold}{Threshold of posterior probability
(influenced by the SOM neighborhood)}

\item{keep}{Samples evaluation to be maintained in the data}
}
\value{
List with two sits tibbles.
        The first tibble has clean samples
        The second has samples that need to be analysed.
}
\description{
This function evaluate the quality of the samples
based on the results of the SOM map. It produces
a sits tibble with an evaluation column indicating if
each sample is clean, should be analyzed or
should be removed, and with a new column indicating
the posterior probability of the sample
}
\note{
The algorithm identifies noisy samples, using `prior_threshold` for
    the prior probability and `posterior_threshold` for the posterior probability.
    Each sample receives an evaluation tag, according to the following rule:
    (a) If the prior probability is $ < `prior_threshold`$, the sample is tagged as "remove";
    (b) If the prior probability is $ \geq `prior_threshold`$ and the posterior probability
    is $ \geq `posterior_threshold`$, the sample is tagged as "clean";
    (c) If the prior probability is $\geq `posterior_threshold`$ and
    the posterior probability is $< `posterior_threshold`$, the sample is tagged as "analyze"
    for further inspection.

    The user can define which tagged samples will be returned using the "keep"
    parameter, with the following options: "clean", "analyze", "remove", "all"
}
\examples{
\dontrun{
# Read a set of samples
# Get a new subset of samples evaluated by clustering methods
som_map <- sits_som_map(samples_modis_4bands,
    grid_xdim = 10, grid_ydim = 10,
    distance = "euclidean"
)
new_samples <- sits_som_clean_samples(som_map)
}
}
\author{
Lorena Santos, \email{lorena.santos@inpe.br}

Karine Ferreira. \email{karine.ferreira@inpe.br}
}
