% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_som.R
\name{sits_som_clean_samples}
\alias{sits_som_clean_samples}
\title{Clean samples}
\usage{
sits_som_clean_samples(
  som_map,
  prior_threshold = 0.6,
  posterior_threshold = 0.6
)
}
\arguments{
\item{som_map}{An object returned
by \code{\link[sits]{sits_som_map}}}

\item{prior_threshold}{Threshold of conditional probability
(frequency of samples assigned to the same SOM neuron)}

\item{posterior_threshold}{Threshold of posterior probability
(influenced by the SOM neighborhood)}
}
\value{
List with two sits tibbles.
        The first tibble has clean samples
        The second has samples that need to be analysed.
}
\description{
This function evaluate the quality of the samples
based on the results of the SOM map. It produces
a sits tibble with an evaluation column indicating if
each sample is clean, should be analyzed or
should be removed, and with a new column indicating
the posterior probability of the sample
}
\examples{
\dontrun{
# Read a set of samples
# Get a new subset of samples evaluated by clustering methods
som_map <- sits_som_map(samples_mt_4bands,
    grid_xdim = 10, grid_ydim = 10,
    distance = "euclidean"
)
new_samples <- sits_som_clean_samples(som_map)
}
}
\author{
Lorena Santos, \email{lorena.santos@inpe.br}

Karine Ferreira. \email{karine.ferreira@inpe.br}
}
