% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_som.R
\name{sits_som_evaluate_cluster}
\alias{sits_som_evaluate_cluster}
\title{Evaluate cluster}
\usage{
sits_som_evaluate_cluster(som_map)
}
\arguments{
\item{som_map}{A SOM map produced by the som_map() function}
}
\value{
A tibble stating the purity for each cluster
}
\description{
\code{sits_som_evaluate_cluster()} produces a tibble with the clusters
found by the SOM map. For each cluster, it provides the percentage
of classes inside it.
}
