% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_kohonen.R
\name{sits_som_evaluate_cluster}
\alias{sits_som_evaluate_cluster}
\title{Cluster evaluation}
\usage{
sits_som_evaluate_cluster(som_map)
}
\arguments{
\item{som_map}{SOM map and associated information
created by \code{\link[sits]{sits_som_map}}}
}
\value{
The confusion matrix and
                    a table with percentage of mixture between the clusters.
}
\description{
This function extracts metrics about the clusters calculating
the percentage of mixture between a cluster and others.
}
\examples{
\donttest{
# Produce a Kohonen map for the time series samples
samples_mt <- sits_select(samples_mt_4bands, bands = c("NDVI", "EVI"))
som_map <- sits_som_map(samples_mt)
# Extract metrics about the clusters
confusion_by_cluster <- sits_som_evaluate_cluster(som_map)
# Show confusion matrix
confusion_matrix <- confusion_by_cluster$confusion_matrix
# Plot cluster confusion
plot(confusion_by_cluster)
}
}
\author{
Lorena Santos, \email{lorena.santos@inpe.br}
}
