% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_som.R
\name{sits_som_evaluate_cluster}
\alias{sits_som_evaluate_cluster}
\title{Evaluate cluster}
\usage{
sits_som_evaluate_cluster(som_map)
}
\arguments{
\item{som_map}{An object returned
by \code{\link[sits]{sits_som_map}}}
}
\value{
A tibble with the cluster and the percentage of classes mixtured in each cluster.
}
\description{
This function evaluate the clusters created by
SOM. Each cluster is a neuron or a set of neuron categorized with same label.
It produces a sits tibble indicating the percentage of mixture
of classes in each cluster.
}
\examples{
\dontrun{
# Read a set of samples
# Get a new subset of samples evaluated by clustering methods
som_map <- sits_som_map(samples_mt_4bands,
    grid_xdim = 10, grid_ydim = 10,
    distance = "euclidean"
)
cluster_purity <- sits_som_evaluate_cluster(som_map)
}
}
\author{
Lorena Santos, \email{lorena.santos@inpe.br}

Karine Ferreira. \email{karine.ferreira@inpe.br}
}
