% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_kohonen.R
\name{sits_som_map}
\alias{sits_som_map}
\title{Generate a Kohonen map for quality control of the samples and calculate parameters}
\usage{
sits_som_map(
  data,
  grid_xdim = 10,
  grid_ydim = 10,
  alpha = 1,
  rlen = 100,
  distance = "euclidean",
  iterations = 1
)
}
\arguments{
\item{data}{A tibble with samples to be clustered.}

\item{grid_xdim}{X dimension of the SOM grid (default = 25).}

\item{grid_ydim}{Y dimension of the SOM grid.}

\item{alpha}{Starting learning rate, which decreases according to number of iterations.}

\item{rlen}{Number of iterations to produce the SOM.}

\item{distance}{The type of similarity measure (distance).}

\item{iterations}{The number of time to run the SOM cluster.}
}
\value{
list of sits tibbles containing statistics
                      about the samples and the neuron in each iteration.
}
\description{
This function uses package self-organized maps to find clusters in
satellite image time series to cluster the  samples. Besides that, this function
evaluates the quality of each sample through SOM properties, such as evaluate each
sample allocated in aneuron-cluster and its neighbourhood.

The output day contains additional columns:
(a) the probability of a sample belongs to a cluster based on a frequency
that a sample was allocated in a neuron;
(b) the total probability of a sample belongs
to a cluster using data of neighbourhood.
}
\examples{
\donttest{
# Produce a cluster map
som_cluster <- sits_som_map(prodes_226_064)
# Clean the samples to get better quality ones
clean_samples <- sits_som_clean_samples(som_cluster)
}
}
\references{
`kohonen` package (https://CRAN.R-project.org/package=kohonen)
}
\author{
Lorena Alves, \email{lorena.santos@inpe.br}

Karine Ferreira. \email{karine.ferreira@inpe.br}
}
