% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_raster.R
\name{sits_STRaster}
\alias{sits_STRaster}
\title{Create a metadata tibble to store the description of a spatio-temporal raster dataset}
\usage{
sits_STRaster(files, timeline, bands, scale_factors)
}
\arguments{
\item{files}{Vector with the file paths of the raster files}

\item{timeline}{Vector of dates with the timeline of the bands}

\item{bands}{The bands contained in the Raster Brick set (in the same order as the files)}

\item{scale_factors}{Scale factors to convert each band to [0..1] range (in the same order as the files)}
}
\value{
raster.tb   A tibble with metadata information about a raster data set
}
\description{
This function creates a tibble containing the metadata for
              a set of spatio-temporal raster files, defined as a set of "Raster Bricks".
              These files should be of the same size and
              projection. Each raster file should contain a data set, one layer
              per time step. Different bands are archived in different raster files.

This function creates a tibble to store the information
about a raster time series
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
