% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_sample_functions.R
\name{sits_stratified_sampling}
\alias{sits_stratified_sampling}
\title{Allocation of sample size to strata}
\usage{
sits_stratified_sampling(cube, sampling_design, alloc, shp_name = NULL)
}
\arguments{
\item{cube}{Classified cube}

\item{sampling_design}{Result of sits_sampling_design}

\item{alloc}{Allocation method chosen}

\item{shp_name}{Name of shapefile to be saved (optional)}
}
\value{
samples              Point sf object with required samples
}
\description{
Takes a class cube with different labels and a sampling
design with a number of samples per class and allocates a set of
locations for each class
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
