% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_active_learning.R
\name{sits_suggest_samples}
\alias{sits_suggest_samples}
\title{Suggest samples for enhancing classification accuracy}
\usage{
sits_suggest_samples(cube, n = 100, min_dist_pixels = 10)
}
\arguments{
\item{cube}{A `sits` uncertainty cube. See `sits_uncertainty`.}

\item{n}{Number of suggested points.}

\item{min_dist_pixels}{Minimum distance among suggested points (in pixels).}
}
\value{
A data.frame with longitude & latitude in WGS84.
}
\description{
Suggest points with higher chances of increasing classification accuracy.
These points don't have labels and are meant to be labelled by experts
outside R and then used to increase the classification's training set.
}
\author{
Alber Sanchez, \email{alber.sanchez@inpe.br}

Rolf Simoes, \email{rolf.simoes@inpe.br}

Felipe Carvalho, \email{felipe.carvalho@inpe.br}

Gilberto Camara, \email{gilberto.camara@inpe.br}
}
