% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_summary.R
\name{sits_summary}
\alias{sits_summary}
\alias{sits_summary.sits}
\title{Returns the information about labels of a tibble data set}
\usage{
sits_summary(data)

\method{sits_summary}{sits}(data)
}
\arguments{
\item{data}{Valid sits tibble}
}
\value{
A tibble with labels frequency.
}
\description{
Finds labels in a sits tibble
}
\examples{
# read a tibble with 400 samples of Cerrado and 346 samples of Pasture
data(cerrado_2classes)
# print the labels
sits_summary(cerrado_2classes)

}
\author{
Rolf Simoes, \email{rolf.simoes@inpe.br}
}
