% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_table.R
\name{sits_table}
\alias{sits_table}
\title{Create a sits table to store the time series information}
\usage{
sits_table()
}
\value{
table  a tibble in SITS format
}
\description{
\code{sits_table} returns an empty sits table
A sits table has the metadata and data for each time series
<longitude, latitude, start_date, end_date, label, coverage, time_series>
}
