% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_table.R
\name{sits_table_result}
\alias{sits_table_result}
\title{Create a sits table to store the result of TWDTW classification}
\usage{
sits_table_result()
}
\value{
table  a tibble in SITS format
}
\description{
A sits table is a tibble with pre-defined columns that
has the metadata and data for each time series.
To include the results of the classification, this basic structure is extended with two new
columns: distances (a list of distances to each class for each interval) and matches
(the output of the TWDTW classifier). The resulting columns will be
<longitude, latitude, start_date, end_date, label, coverage, time_series, distances, matches>.
}
\references{
Please see the documentation of the dtwSat package
}
\seealso{
Other STIS table functions: \code{\link{sits_value_cols}},
  \code{\link{sits_value_rows}}
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}

Victor Maus, \email{vwmaus1@gmail.com}
}
