% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_indexes.R
\name{sits_tasseled_cap}
\alias{sits_tasseled_cap}
\title{Builds tasseled cap bands}
\usage{
sits_tasseled_cap(data.tb, satellite = "MODIS")
}
\arguments{
\item{data.tb}{valid sits tibble}

\item{satellite}{satellite name}
}
\value{
data.tb      sits tibble with same samples and the new bands
}
\description{
Adds new tasseled cap bands
}
\examples{
\donttest{
# Retrieve data for time series with label samples in Mato Grosso in Brazil
data (samples_MT_9classes)
# Generate a new image with the tasseled cap
tc.tb <- sits_tasseled_cap(samples_MT_9classes, satellite = "MODIS")
}
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
