% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_deep_learning.R
\name{sits_tempCNN}
\alias{sits_tempCNN}
\title{Train a model using the Temporal Convolutional Neural Network}
\usage{
sits_tempCNN(data = NULL, conv_units = c(64, 64, 64),
  conv_kernels = c(7, 7, 7), conv_activation = "relu",
  conv_L2_rate = 1e-06, conv_dropout_rates = c(0.5, 0.5, 0.5),
  node_units = c(512, 512, 512), node_activation = "elu",
  node_dropout_rates = c(0.5, 0.45, 0.4),
  optimizer = keras::optimizer_adam(lr = 0.001), epochs = 150,
  batch_size = 128, validation_split = 0.2, verbose = 1,
  binary_classification = FALSE)
}
\arguments{
\item{data}{Time series with the training samples.}

\item{conv_units}{Vector with the number of 1D convolutional filters.}

\item{conv_kernels}{Vector with the size of the 1D convolutional kernels.}

\item{conv_activation}{Activation function for 1D convolution.
Valid values are {'relu', 'elu', 'selu', 'sigmoid'}.}

\item{conv_L2_rate}{Regularization rate for 1D convolution.
Valid values are {'relu', 'elu', 'selu', 'sigmoid'}.}

\item{conv_dropout_rates}{Vector with dropout rates for 1D convolutional filters.}

\item{node_units}{Vector with the number of hidden nodes in each 2D layer.}

\item{node_activation}{Names of 2D activation functions. Valid values are {'relu', 'elu', 'selu', 'sigmoid'}.}

\item{node_dropout_rates}{Vector with the dropout rates (0,1) for each layer to the next layer.}

\item{optimizer}{Function with a pointer to the optimizer function (default is optimization_adam()).
Options are optimizer_adadelta(), optimizer_adagrad(), optimizer_adam(),
optimizer_adamax(), optimizer_nadam(), optimizer_rmsprop(), optimizer_sgd()}

\item{epochs}{Number of iterations to train the model.}

\item{batch_size}{Number of samples per gradient update.}

\item{validation_split}{Number between 0 and 1. Fraction of the training data to be used as validation data.
The model will set apart this fraction of the training data, will not train on it,
and will evaluate the loss and any model metrics on this data at the end of each epoch.
The validation data is selected from the last samples in the x and y data provided,
before shuffling.}

\item{verbose}{Verbosity mode (0 = silent, 1 = progress bar, 2 = one line per epoch).}

\item{binary_classification}{A lenght-one logical indicating if this is a binary classification. If it is so,}
}
\value{
A model fitted to input data to be passed to \code{\link[sits]{sits_classify}}
}
\description{
Use a tempCNN algorithm to classify data. The tempCNN algorithm has
two stages: a 1D CNN is applied to the input time series data is combined with a
multi-layer perceptron. Users can define the depth of the 1D network, as well as
the number of perceptron layers.

This function is based on the paper by Charlotte Pelletier referenced below
and the code made available on github (https://github.com/charlotte-pel/temporalCNN)
If you use this method, please cite the original tempCNN paper.
}
\examples{
\donttest{
# Retrieve the set of samples for the Mato Grosso region (provided by EMBRAPA)
# Install the inSitu library
# devtools::install_github("e-sensing/inSitu")
# library(inSitu)

# select the bands for classification
samples <- inSitu::br_mt_1_8K_9classes_6bands

# find a training model based on the distances and default values (SVM model)
samples_4bands <- sits_select_bands(samples, ndvi, evi, nir, mir)

# Build a machine learning model based on deep learning
dl_model <- sits_train (samples_4bands, sits_tempCNN())

# get a point and classify the point with the ml_model
point.tb <- sits_select_bands(point_mt_6bands, ndvi, evi, nir, mir)
class.tb <- sits_classify(point.tb, dl_model)
sits_plot(class.tb)
}
}
\references{
Charlotte Pelletier, Geoffrey Webb and François Petitjean,
"Temporal Convolutional Neural Network for the Classification of Satellite Image Time Series",
Remote Sensing, 11,523, 2019. DOI: 10.3390/rs11050523.
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}

Alexandre Xavier Ywata de Carvalho, \email{alexandre.ywata@ipea.gov.br}

Rolf Simoes, \email{rolf.simoes@inpe.br}
}
