% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_tibble.R
\name{sits_tibble}
\alias{sits_tibble}
\title{Create a sits tibble to store the time series information}
\usage{
sits_tibble()
}
\value{
A sits tibble.
}
\description{
This function returns an empty sits tibble.
Sits tibbles are the main structures of sits package.
They contain both the satellite image time series and its metadata.
A sits tibble is a tibble with pre-defined columns that
has the metadata and data for each time series. The columns are
<longitude, latitude, start_date, end_date, label, coverage, time_series>.
Most functions of sits package get a sits tibble as input (with additional parameters)
and return another sits tibble as output. This allows chaining functions over sits tibbles.
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
