% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_tibble.R
\name{sits_tibble_prediction}
\alias{sits_tibble_prediction}
\title{Create an empty tibble to store the results of predictions}
\usage{
sits_tibble_prediction(distances.tb, timeline, time_index.lst)
}
\arguments{
\item{distances.tb}{a tibble with distances}

\item{timeline}{the timeline of the data set}

\item{time_index.lst}{The subsets of the timeline}
}
\value{
distances.tb   a tibble to store the predictions
}
\description{
Create atibble to store the results of predictions
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}

Rolf Simoes, \email{rolf.simoes@inpe.br}
}
