% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_tibble.R
\name{sits_time_series}
\alias{sits_time_series}
\title{Get the time series for a row of a sits tibble}
\usage{
sits_time_series(data)
}
\arguments{
\item{data}{A sits tibble with one or more time series.}
}
\value{
A tibble in sits format with the time series.
}
\description{
Returns the time series associated to a row of the a sits tibble
}
\examples{
if (sits_active_tests()) {
# Retrieve a set of time series with 2 classes
data(cerrado_2classes)
# Retrieve the first time series
sits_time_series(cerrado_2classes)
}

}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
