% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_timeline.R
\name{sits_timeline}
\alias{sits_timeline}
\title{Obtains the timeline}
\usage{
sits_timeline(data, index = 1)
}
\arguments{
\item{data}{A sits tibble (either a sits tibble or a raster metadata).}

\item{index}{The index to obtain the timeline (in case of raster bricks)}
}
\description{
This function returns the timeline for a given cube.
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
