% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_timeline.R
\name{sits_timeline}
\alias{sits_timeline}
\title{Obtains the timeline}
\usage{
sits_timeline(data)
}
\arguments{
\item{data}{either a sits tibble, a data cube, or a trained model.}
}
\description{
This function returns the timeline for a given data set, either
             a set of time series, a data cube, or a trained model.
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
