% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_timeline.R
\name{sits_timeline}
\alias{sits_timeline}
\title{Obtains the timeline}
\usage{
sits_timeline(data)
}
\arguments{
\item{data}{either a sits tibble or data cube}
}
\description{
This function returns the timeline for a given data set
             For details see:
\itemize{
 \item{"time series": }{see \code{\link{sits_timeline.sits}}}
 \item{"data cube": }{see \code{\link{sits_timeline.cube}}}
}
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
