% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_timeline.R
\name{sits_timeline}
\alias{sits_timeline}
\title{Get timeline of a cube or a set of time series}
\usage{
sits_timeline(data)
}
\arguments{
\item{data}{either a sits tibble, a data cube, or a trained model.}
}
\value{
Timeline of sample set or of data cube.
}
\description{
This function returns the timeline for a given data set, either
             a set of time series, a data cube, or a trained model.
}
\examples{
sits_timeline(samples_modis_ndvi)
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
