% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_csv.R
\name{sits_to_csv}
\alias{sits_to_csv}
\title{Export a sits tibble metadata to the CSV format}
\usage{
sits_to_csv(data, file)
}
\arguments{
\item{data}{Sits time series.}

\item{file}{Full path of the exported CSV file (character).}
}
\value{
Called for side effects.
}
\description{
Converts metadata from a sits tibble to a CSV file.
             The CSV file will not contain the actual time
             series. Its columns will be the same as those of a
             CSV file used to retrieve data from
             ground information ("latitude", "longitude", "start_date",
             "end_date", "cube", "label").
}
\examples{
csv_file <- paste0(tempdir(), "/cerrado_2classes.csv")
sits_to_csv(cerrado_2classes, file = csv_file)
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
