% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_xlsx.R
\name{sits_to_xlsx}
\alias{sits_to_xlsx}
\title{Save accuracy assessments as Excel files}
\usage{
sits_to_xlsx(acc_lst, file, data = NULL)
}
\arguments{
\item{acc_lst}{A list of accuracy statistics}

\item{file}{The file where the XLSX data is to be saved.}

\item{data}{(optional) Print information about the samples}
}
\description{
Saves confusion matrices as Excel spreadsheets. This function
takes the a list of accuracy assessments generated
by \code{\link[sits]{sits_accuracy}}
and saves them in an Excel spreadsheet.
}
\note{
Please refer to the sits documentation available in
<https://e-sensing.github.io/sitsbook/> for detailed examples.
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
