% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_export_data.R
\name{sits_toGZ}
\alias{sits_toGZ}
\title{Export the contents of a SITS tibble in a compressed JSON file}
\usage{
sits_toGZ(data.tb, file = NULL)
}
\arguments{
\item{data.tb}{a SITS tibble}

\item{file}{string - file name where compressed data will be saved}
}
\value{
data.tb      the input tibble (for chaining functions)
}
\description{
stores a SITS tibble in a compressed JSON file
The set of time series from a SITS tibble can be saved in a compressed JSON format
and later retrieved for further use
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
