% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_export.R
\name{sits_toJSON}
\alias{sits_toJSON}
\title{Save data in a compressed JSON file}
\usage{
sits_toJSON(data.tb, file = NULL)
}
\arguments{
\item{data.tb}{a SITS tibble}

\item{file}{string - name of the JSON file to be written}
}
\description{
stores data in a JSON file
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
