% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_save.R
\name{sits_toJSON}
\alias{sits_toJSON}
\title{Save data in a JSON file}
\usage{
sits_toJSON(source, file = NULL)
}
\arguments{
\item{source}{an R object}

\item{json_file}{string - name of the JSON file to be written}
}
\description{
stores data in a JSON file
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
