% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_TWDTW.R
\name{sits_toTWDTW}
\alias{sits_toTWDTW}
\title{Export data to be used by the dtwSat package}
\usage{
sits_toTWDTW(data.tb)
}
\arguments{
\item{data.tb}{a table in SITS format with time series to be converted to TWTDW time series}
}
\value{
ts.tw         a time series in TWDTW format (an object of the twdtwTimeSeries class)
}
\description{
\code{sits_toTWDTW} returns a twdtwTimeSeries object (S4)
}
\details{
Converts data from a SITS table to an instance of a TWDTW time series class

Reference: Maus V, Camara G, Cartaxo R, Sanchez A, Ramos FM, de Queiroz GR (2016).
A Time-Weighted Dynamic Time Warping Method for Land-Use and Land-Cover Mapping. IEEE
 Journal of Selected Topics in Applied Earth Observations and Remote Sensing, 9(8):3729-3739,
 August 2016. ISSN 1939-1404. doi:10.1109/JSTARS.2016.2517118.
}
