% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_export_data.R
\name{sits_toZOO}
\alias{sits_toZOO}
\title{Export data to be used to the zoo format}
\usage{
sits_toZOO(ts.tb, band = NULL)
}
\arguments{
\item{ts.tb}{a tibble with the indexes and values of a SITS time series
the tibble should have only the time series and should not be a full tibble}

\item{band}{the name of the band to be exported (if NULL all bands are exported)}
}
\value{
ts.zoo   a time series in zoo format
}
\description{
Converts data from a SITS tibble to an instance of a zoo series,
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
