% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_export.R
\name{sits_toZOO}
\alias{sits_toZOO}
\title{Export data to be used to the zoo format}
\usage{
sits_toZOO(data.tb, band = NULL)
}
\arguments{
\item{data.tb}{a SITS time series}

\item{band}{the name of the band to be exported (if NULL all bands are exported)}
}
\value{
zoo.lst    a list of time series in zoo format
}
\description{
Converts data from a SITS tibble to a list of a zoo series.
}
\examples{

# read a tibble with 400 samples of Cerrado and 346 samples of Pasture
data(cerrado_2classes)
# export a time series to zoo
zoo.lst <- sits_toZOO (cerrado_2classes[1:5,])

}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
