% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_machine_learning.R
\name{sits_train}
\alias{sits_train}
\title{Train classifiction models using the TWDTW distances for each class}
\usage{
sits_train(data.tb, model_class, ...)
}
\arguments{
\item{data.tb}{a SITS tibble time series}

\item{...}{other parameters to pass to \code{\link[sits]{sits_patterns}} and 
\code{\link[sits]{sits_TWDTW_matches}}}

\item{model}{a vector of character with the models to be tested. The options are: XXXXXXXX}
}
\description{
Given a SITS tibble time series, returns trained models using Machine Learning.
}
\author{
Alexandre Xavier Ywata de Carvalho, \email{alexandre.ywata@ipea.gov.br}

Victor Maus, \email{vwmaus1@gmail.com}

Gilberto Camara, \email{gilberto.camara@inpe.br}
}
