% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_machine_learning.R
\name{sits_train}
\alias{sits_train}
\title{Train SITS classification models}
\usage{
sits_train(data.tb, ml_method = sits_svm(),
  dist_method = sits_distances_from_data())
}
\arguments{
\item{data.tb}{a time series with the training samples}

\item{ml_method}{the machine learning method that returns a model for prediction}

\item{dist_method}{the method used to compute the distances used to build the model}
}
\value{
result          a model fitted into input data given by train_method parameter
}
\description{
Given a tibble with a set of distance measures,
returns trained models using support vector machines.
#' After defining the training samples, the users need to provide a machine learning model.
Currenly, sits supports the following models:
'svm' (see \code{\link[sits]{sits_svm}}), 'random forest' (see \code{\link[sits]{sits_rfor}}),
'boosting' (see \code{\link[sits]{sits_gbm}}), 'lda' (see \code{\link[sits]{sits_lda}}),
'qda' (see \code{\link[sits]{sits_qda}}), multinomial logit' (see \code{\link[sits]{sits_mlr}}),
'lasso' (see \code{\link[sits]{sits_mlr}}), 'ridge' (see \code{\link[sits]{sits_mlr}}),
"multi-layer perceptrons" also known as "deep learning" (see \code{\link[sits]{sits_mlp}})

The sits_train function is called inside \code{\link[sits]{sits_classify}}
and \code{\link[sits]{sits_classify_raster}}, so the user does not need
to explicitly use this function. Please see the above-mention classification functions.
}
\examples{

\donttest{
# Retrieve the set of samples for the Mato Grosso region (provided by EMBRAPA)
data(samples_MT_ndvi)
# find a training model based on the distances and default values (SVM model)
ml_model <- sits_train (samples_MT_ndvi)
# get a point
data(ts_2000_2016)
point.tb <- sits_select (ts_2000_2016, bands = c("ndvi"))
# classify the point with the ml_model
class.tb <- sits_classify_model (point.tb, samples_MT_ndvi, ml_model)
}
}
\author{
Rolf Simoes, \email{rolf.simoes@inpe.br}

Alexandre Xavier Ywata de Carvalho, \email{alexandre.ywata@ipea.gov.br}

Gilberto Camara, \email{gilberto.camara@inpe.br}
}
