% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_machine_learning.R
\name{sits_train}
\alias{sits_train}
\title{Train SITS classification models}
\usage{
sits_train(data.tb, ml_method = sits_svm())
}
\arguments{
\item{data.tb}{time series with the training samples}

\item{ml_method}{machine learning method that returns a model for prediction}
}
\value{
result           model fitted to input data given by train_method parameter
}
\description{
Given a tibble with a set of distance measures,
   returns trained models. Currenly, sits supports the following models:
'svm' (see \code{\link[sits]{sits_svm}}), 'random forest' (see \code{\link[sits]{sits_rfor}}),
'boosting' (see \code{\link[sits]{sits_gbm}}), 'lda' (see \code{\link[sits]{sits_lda}}),
'qda' (see \code{\link[sits]{sits_qda}}), multinomial logit' (see \code{\link[sits]{sits_mlr}}),
'lasso' (see \code{\link[sits]{sits_mlr}}), 'ridge' (see \code{\link[sits]{sits_mlr}}),
and "deep learning" (see \code{\link[sits]{sits_deeplearning}})
The sits_train function is called inside \code{\link[sits]{sits_classify}}
and \code{\link[sits]{sits_classify_raster}}, so the user does not need
to explicitly use this function. Please see the above-mention classification functions.
}
\examples{

\donttest{
# Retrieve the set of samples for the Mato Grosso region (provided by EMBRAPA)
# find a training model based on the distances and default values (SVM model)
samples.tb <- sits_select(samples_MT_9classes, bands = c("ndvi", "evi", "nir", "mir"))
ml_model <- sits_train (samples.tb)
# get a point and classify the point with the ml_model
point.tb <- sits_select(point_MT_6bands, bands = c("ndvi", "evi", "nir", "mir"))
class.tb <- sits_classify(point_MT_6bands, samples_MT_9classes, ml_model)
sits_plot(class.tb)
}
}
\author{
Rolf Simoes, \email{rolf.simoes@inpe.br}

Gilberto Camara, \email{gilberto.camara@inpe.br}

Alexandre Xavier Ywata de Carvalho, \email{alexandre.ywata@ipea.gov.br}
}
