% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_machine_learning.R
\name{sits_train_svm}
\alias{sits_train_svm}
\title{Train SITS classifiction models using support vector machines}
\usage{
sits_train_svm(data.tb, type = "C-classification", kernel = "linear",
  degree = 3, coef0 = 0, cost = 100, tolerance = 0.001, epsilon = 0.1)
}
\arguments{
\item{data.tb}{a SITS tibble time series with an alignment column}

\item{type}{whether svm is used for classification or for regression}

\item{kernel}{the kernel used in training and predicting (options = linear, polynomial, radial basis, sigmoid)}

\item{degree}{exponential of polynomial type kernel}

\item{coef0}{parameter needed for kernels of type polynomial and sigmoid (default: 0)}

\item{cost}{cost of constraints violation (default: 1): it is the Cconstant of the regularization term in the Lagrange formulation.}

\item{tolerance}{tolerance of termination criterion (default: 0.001)}

\item{epsilon}{epsilon in the insensitive-loss function (default: 0.1)}
}
\value{
result.svm an svm model fit for the input data
}
\description{
Given a SITS tibble time series with DTW matches produced by TWDTW,
             returns trained models using support vector machines. This function will
             use the TWDTW alignment information for all classes as the attributes
             of the SVM. Please use this function in the following way:
             1. call sits_patterns to produce a set a labelled patterns from a reference data set
             2. call sits_TWDTW_matches to produce a set of alignements
             3. use the alignment response as an input to the training function

             This function is a front-end to the "svm" method in the "e1071" package.
             In the context of time series classification, the only relevant types are "C-classification" and "nu-classification".
             Please refer to the documentation in that package for more details.
}
\details{
Functions for machine learning associated to the SITS package
The attributes for the training functions are the DTW distances
computed by the TWTDW function (see documentation on sits_TWDTW_matches)

models supported: 'svm', 'random forests', 'boosting', 'lda',
                  'multinomial logit', 'lasso', 'ridge', 'elnet', 'best model'
}
\author{
Alexandre Xavier Ywata de Carvalho, \email{alexandre.ywata@ipea.gov.br}

Victor Maus, \email{vwmaus1@gmail.com}

Gilberto Camara, \email{gilberto.camara@inpe.br}

Rolf Simoes, \email{rolf.simoes@inpe.br}
}
