% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_time_series.R
\name{sits_transmute}
\alias{sits_transmute}
\title{Add new SITS bands and drops existing.}
\usage{
sits_transmute(data.tb, ...)
}
\arguments{
\item{data.tb}{a valid sits table}

\item{...}{`name=value` pairs expressions. See `dplyr::transmute` help for more details.}
}
\value{
data.tb      a sits_table with same samples and the new bands
}
\description{
Adds new bands and drops existing in a sits_table's time series,
              using dplyr::transmute function
}
\author{
Rolf Simoes, \email{rolf.simoes@inpe.br}
}
