% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_tsne.R
\name{sits_tsne}
\alias{sits_tsne}
\title{Run t-SNE on sits torch models}
\usage{
sits_tsne(
  model,
  samples,
  remove_duplicates = TRUE,
  perplexity = 30,
  rounds = 1000,
  ...
)
}
\arguments{
\item{model}{A sits torch model (output of sits_train()).}

\item{samples}{A sits tibble with time series samples.}

\item{...}{Other parameters (passed to methods).}
}
\value{
A list with:
\itemize{
  \item \code{tsne}: The t-SNE result (object of class \code{"Rtsne"}).
  \item \code{labels}: A character vector of labels, aligned with rows in \code{tsne$Y}.
}
}
\description{
Generic function that Applies t-SNE dimensionality reduction to the feature embeddings produced
by a sits torch model (e.g., LTAE, TempCNN) for a given set of samples.
Automatically handles duplicate embeddings by removing them before running t-SNE.
}
\examples{
if (sits_run_examples()) {
    # load samples
    samples <- samples_modis_ndvi
    # sits TAE
    model <- sits_train(
    samples   = samples,
    ml_method = sits_tae(epochs = 2)
    )
    # run tsne on a deep learning model
    tsne <- sits_tsne(model, samples, perplexity = 30, rounds = 100)
    plot(tsne)
}

}
