% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_TWDTW.R
\name{sits_TWDTW}
\alias{sits_TWDTW}
\title{Classify a sits tibble using TWDTW (using the dtwSat package)}
\usage{
sits_TWDTW(series.tb, patterns.tb, bands, alpha = -0.1, beta = 100,
  theta = 0.5, start_date = as.Date("2000-09-01"),
  end_date = as.Date("2016-08-31"), interval = "12 month", span = 250)
}
\arguments{
\item{series.tb}{a table in SITS format with a time series to be classified using TWTDW}

\item{patterns.tb}{a set of known temporal signatures for the chosen classes}

\item{bands}{string - the bands to be used for classification}

\item{alpha}{(double) - the steepness of the logistic function used for temporal weighting}

\item{beta}{(integer) - the midpoint (in days) of the logistic function}

\item{theta}{(double)  - the relative weight of the time distance compared to the dtw distance}

\item{start_date}{date - the starting date of the classification}

\item{end_date}{date - the end date of the classification}

\item{interval}{the period between two classifications}

\item{span}{the minimum period for a match between a pattern and a signal)}
}
\value{
matches       a SITS table with the information on matches for the data
}
\description{
\code{sits_TWDTW} returns a sits table with the results of the TWDTW classifier.
}
\details{
The TWDTW classifier compares the values of a satellite image time series with
the values of known patters and tries to match each pattern to a part of the time series

The TWDTW (time-weighted dynamical time warping) is a version of the
Dynamic Time Warping method for land use and land cover mapping using a sequence
of multi-band satellite images. Methods based on dynamic time warping are flexible to
handle irregular sampling and out-of-phase time series, and they have achieved significant
results in time series analysis.
In contrast to standard DTW, the TWDTW method is sensitive to seasonal
changes of natural and cultivated vegetation types. It also considers inter-annual climatic and
seasonal variability.

Reference: Maus V, Camara G, Cartaxo R, Sanchez A, Ramos FM, de Queiroz GR (2016).
A Time-Weighted Dynamic Time Warping Method for Land-Use and Land-Cover Mapping. IEEE
 Journal of Selected Topics in Applied Earth Observations and Remote Sensing, 9(8):3729-3739,
 August 2016. ISSN 1939-1404. doi:10.1109/JSTARS.2016.2517118.
}
