% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_TWDTW.R
\name{sits_TWDTW_distances}
\alias{sits_TWDTW_distances}
\title{Find distance between a set of SITS patterns and segments of sits tibble using TWDTW}
\usage{
sits_TWDTW_distances(data.tb = NULL, patterns.tb = NULL, bands = NULL,
  dist.method = "euclidean", alpha = -0.1, beta = 100, theta = 0.5,
  span = 250, keep = FALSE)
}
\arguments{
\item{data.tb}{a table in SITS format with a time series to be classified using TWTDW}

\item{patterns.tb}{a set of known temporal signatures for the chosen classes}

\item{bands}{string - the bands to be used for classification}

\item{dist.method}{A character. Method to derive the local cost matrix.}

\item{alpha}{(double) - the steepness of the logistic function used for temporal weighting}

\item{beta}{(integer) - the midpoint (in days) of the logistic function}

\item{theta}{(double)  - the relative weight of the time distance compared to the dtw distance}

\item{span}{minimum number of days between two matches of the same pattern in the time series (approximate)}

\item{keep}{keep internal values for plotting matches}
}
\value{
matches       a SITS table with the information on matches for the data
}
\description{
Returns a SITS table with distances to be used for training in ML methods
This is a front-end to the sits_TWDTW_matches whose outout is trimmed down to contain just distances
}
\author{
Rolf Simoes, \email{rolf.simores@inpe.br}

Gilberto Camara, \email{gilberto.camara@inpe.br}
}
