% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_distances.R
\name{sits_TWDTW_distances}
\alias{sits_TWDTW_distances}
\title{Find distances between a set of SITS patterns and segments of sits tibble using TWDTW}
\usage{
sits_TWDTW_distances(data.tb = NULL, patterns.tb = NULL,
  interval = "12 month", dist.method = "euclidean", alpha = -0.1,
  beta = 100, theta = 0.5, span = 250, keep = FALSE, multicores = 1)
}
\arguments{
\item{data.tb}{Table in SITS format with a time series to be classified using TWTDW}

\item{patterns.tb}{Set of known temporal signatures for the chosen classes}

\item{interval}{Period to match the data to the patterns}

\item{dist.method}{Method to derive the local cost matrix.}

\item{alpha}{Steepness of the logistic function used for temporal weighting}

\item{beta}{Midpoint (in days) of the logistic function}

\item{theta}{Relative weight of the time distance compared to the dtw distance}

\item{span}{Minimum number of days between two matches of the same pattern in the time series (approximate)}

\item{keep}{Keep internal values for plotting matches}

\item{multicores}{Number of threads to process the validation (Linux only). Each process will run a
whole partition validation.}
}
\value{
distances.tb       a SITS table with the information on distances for the data
}
\description{
Returns a SITS table with distances to be used for training in ML methods
This is a front-end to the sits_TWDTW_matches whose outout is trimmed down to contain just distances
}
\author{
Rolf Simoes, \email{rolf.simores@inpe.br}

Gilberto Camara, \email{gilberto.camara@inpe.br}
}
