% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_uncertainty.R
\name{sits_uncertainty}
\alias{sits_uncertainty}
\alias{sits_uncertainty.entropy}
\alias{sits_uncertainty.least}
\alias{sits_uncertainty.margin}
\title{Estimate classification uncertainty based on probs cube}
\usage{
sits_uncertainty(
  cube,
  type = "least",
  ...,
  multicores = 2,
  memsize = 8,
  output_dir = ".",
  version = "v1"
)

\method{sits_uncertainty}{entropy}(
  cube,
  type = "entropy",
  ...,
  window_size = 5,
  window_fn = "median",
  multicores = 2,
  memsize = 4,
  output_dir = ".",
  version = "v1"
)

\method{sits_uncertainty}{least}(
  cube,
  type = "least",
  ...,
  multicores = 2,
  memsize = 8,
  output_dir = ".",
  version = "v1"
)

\method{sits_uncertainty}{margin}(
  cube,
  type = "margin",
  ...,
  multicores = 2,
  memsize = 8,
  output_dir = ".",
  version = "v1"
)
}
\arguments{
\item{cube}{Probability data cube.}

\item{type}{Method to measure uncertainty. See details.}

\item{...}{Parameters for specific functions.}

\item{multicores}{Number of cores to run the function.}

\item{memsize}{Maximum overall memory (in GB) to run the
function.}

\item{output_dir}{Output directory for image files.}

\item{version}{Version of resulting image.
(in the case of multiple tests)}
}
\value{
An uncertainty data cube
}
\description{
Calculate the uncertainty cube based on the probabilities
produced by the classifier. Takes a probability cube as input.
The uncertainty measure is relevant in the context of active leaning,
and helps the increase the quantity and quality of training samples by
providing information about the confidence of the model.

The supported types of uncertainty n_are \code{entropy}, \code{least}, and
\code{margin}:
 \itemize{
 \item{\code{entropy}: }{difference between all predictions
   expressed as entropy.}
 \item{\code{least}: }{difference between 100% and most
   confident prediction.}
 \item{\code{margin}: }{difference between the two most
   confident predictions.}
 }
}
\examples{
if (sits_active_tests()) {
data_dir <- system.file("extdata/raster/mod13q1", package = "sits")
out_dir <- tempdir()
cube <- sits_cube(
    source = "BDC",
    collection = "MOD13Q1-6",
    data_dir = data_dir,
    delim = "_",
    parse_info = c("X1", "X2", "tile", "band", "date")
)
samples_2bands <- sits_select(
    sits::samples_modis_4bands,
    bands = c("NDVI")
)
rfor_model <- sits_train(samples_2bands,
                        ml_method = sits_rfor(verbose = FALSE)
)
probs_cube <- sits_classify(
    cube,
    ml_model = rfor_model,
    output_dir = tempdir(),
    memsize = 4, multicores = 1
)
cube <- sits_uncertainty(probs_cube,
                         type = "entropy",
                         output_dir = out_dir)

}

}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}

Rolf Simoes, \email{rolf.simoes@inpe.br}
}
