% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_uncertainty.R
\name{sits_uncertainty}
\alias{sits_uncertainty}
\alias{sits_uncertainty.entropy}
\title{Estimate classification uncertainty based on probs cube}
\usage{
sits_uncertainty(
  cube,
  type = "entropy",
  ...,
  multicores = 2,
  memsize = 8,
  output_dir = ".",
  version = "v1"
)

\method{sits_uncertainty}{entropy}(
  cube,
  type = "entropy",
  ...,
  multicores = 2,
  memsize = 8,
  output_dir = ".",
  version = "v1"
)
}
\arguments{
\item{cube}{Probability data cube.}

\item{type}{Method to measure uncertainty.}

\item{...}{Parameters for specific functions.}

\item{multicores}{Number of cores to run the function.}

\item{memsize}{Maximum overall memory (in GB) to run the
function.}

\item{output_dir}{Output directory for image files.}

\item{version}{Version of resulting image.
(in the case of multiple tests)}
}
\value{
An uncertainty data cube
}
\description{
Estimate classification uncertainty based on probs cube
}
\note{
Please refer to the sits documentation available in
<https://e-sensing.github.io/sitsbook/> for detailed examples.
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}

Rolf Simoes, \email{rolf.simoes@inpe.br}
}
