% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_uncertainty.R
\name{sits_uncertainty}
\alias{sits_uncertainty}
\alias{sits_uncertainty.entropy}
\alias{sits_uncertainty.least}
\alias{sits_uncertainty.margin}
\title{Estimate classification uncertainty based on probs cube}
\usage{
sits_uncertainty(
  cube,
  type = "least",
  ...,
  multicores = 2,
  memsize = 8,
  output_dir = ".",
  version = "v1"
)

\method{sits_uncertainty}{entropy}(
  cube,
  type = "entropy",
  ...,
  window_size = 5,
  window_fn = "median",
  multicores = 2,
  memsize = 4,
  output_dir = ".",
  version = "v1"
)

\method{sits_uncertainty}{least}(
  cube,
  type = "least",
  ...,
  multicores = 2,
  memsize = 8,
  output_dir = ".",
  version = "v1"
)

\method{sits_uncertainty}{margin}(
  cube,
  type = "margin",
  ...,
  multicores = 2,
  memsize = 8,
  output_dir = ".",
  version = "v1"
)
}
\arguments{
\item{cube}{Probability data cube.}

\item{type}{Method to measure uncertainty. See details.}

\item{...}{Parameters for specific functions.}

\item{multicores}{Number of cores to run the function.}

\item{memsize}{Maximum overall memory (in GB) to run the
function.}

\item{output_dir}{Output directory for image files.}

\item{version}{Version of resulting image.
(in the case of multiple tests)}
}
\value{
An uncertainty data cube
}
\description{
Calculate the uncertainty cube based on the probabilities
produced by the classifier. Takes a probability cube as input.
The uncertainity measure is relevant in the context of active leaning,
and helps the increase the quantity and quality of training samples by
providing information about the confidence of the model.
The supported types of uncertainty are 'entropy', 'least', and 'margin'.
"entropy" is the difference between all predictions expressed as entropy,
"least" is the difference between 100% and most confident prediction, and
"margin" is the difference between the two most confident predictions.
}
\note{
Please refer to the sits documentation available in
<https://e-sensing.github.io/sitsbook/> for detailed examples.
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}

Rolf Simoes, \email{rolf.simoes@inpe.br}
}
