% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_uncertainty.R
\name{sits_uncertainty}
\alias{sits_uncertainty}
\alias{sits_uncertainty.entropy}
\alias{sits_uncertainty.least}
\alias{sits_uncertainty.margin}
\title{Estimate classification uncertainty based on probs cube}
\usage{
sits_uncertainty(
  cube,
  type = "entropy",
  ...,
  multicores = 2,
  memsize = 8,
  output_dir = ".",
  version = "v1"
)

\method{sits_uncertainty}{entropy}(
  cube,
  type = "entropy",
  ...,
  window_size = 5,
  window_fn = "median",
  multicores = 2,
  memsize = 4,
  output_dir = ".",
  version = "v1"
)

\method{sits_uncertainty}{least}(
  cube,
  type = "least",
  ...,
  multicores = 2,
  memsize = 8,
  output_dir = ".",
  version = "v1"
)

\method{sits_uncertainty}{margin}(
  cube,
  type = "margin",
  ...,
  multicores = 2,
  memsize = 8,
  output_dir = ".",
  version = "v1"
)
}
\arguments{
\item{cube}{Probability data cube.}

\item{type}{Method to measure uncertainty. See details.}

\item{...}{Parameters for specific functions.}

\item{multicores}{Number of cores to run the function.}

\item{memsize}{Maximum overall memory (in GB) to run the
function.}

\item{output_dir}{Output directory for image files.}

\item{version}{Version of resulting image.
(in the case of multiple tests)}
}
\value{
An uncertainty data cube
}
\description{
Estimate classification uncertainty based on probs cube
}
\details{
The supported types are 'entropy', 'least', and 'margin'. Entropy
is the information entropy computed (in nats) for each pixel in a
probability cube. Least confidence is the amount missing from total
confidence to the label with the largest probabilities. Margin of confidence
is the amount missing from total confidence to the difference between the
probabilities of the top 2 labels. The larger the metric (either entropy,
least confidence, or margin of confidence) the larger the uncertainty
regarding the label of the pixels; that is, the probabilities of each label
are similar.
}
\note{
Please refer to the sits documentation available in
<https://e-sensing.github.io/sitsbook/> for detailed examples.
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}

Rolf Simoes, \email{rolf.simoes@inpe.br}
}
