% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_filters.R
\name{sits_upper_bound}
\alias{sits_upper_bound}
\title{Upper bound filter}
\usage{
sits_upper_bound(data.tb, window_size = 1)
}
\arguments{
\item{data.tb}{a valid sits table}

\item{window_size}{an integer informing the window size for envelop calculation. See compute_envelop details.}
}
\value{
result.tb    a sits_table with same samples and the new bands
}
\description{
This function computes the envelope of a time series using the
streaming algorithm proposed by Lemire (2009). This functions calls `dtwclust::compute_envelop` function.
}
\author{
Rolf Simoes, \email{rolf.simoes@inpe.br}
}
