% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_validate.R
\name{sits_validate}
\alias{sits_validate}
\title{validate temporal patterns}
\usage{
sits_validate(data.tb, times = 10, p = 0.1)
}
\arguments{
\item{data.tb}{a SITS tibble}

\item{times}{Number of partitions to create.}

\item{p}{the percentage of data that goes to training.}
}
\description{
Splits the set of time series into training and validation and
compute accuracy metrics. The function uses stratified sampling and a simple
random sampling for each stratum. For each data partition this function
performs a TWDTW analysis and returns the Overall Accuracy, User's Accuracy,
Produce's Accuracy, error matrix (confusion matrix), and a \code{\link[base]{data.frame}}
with the classification (Predicted), the reference classes (Reference),
and the results of the TWDTW analysis.
}
\author{
Victor Maus, \email{vwmaus1@gmail.com}
}
