% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_validate.R
\name{sits_validate}
\alias{sits_validate}
\title{validate temporal patterns}
\usage{
sits_validate(data.tb, bands = NULL, method = "gam", times = 100,
  p = 0.1, conversion.lst = NULL, ...)
}
\arguments{
\item{data.tb}{a SITS tibble}

\item{bands}{the bands used for classification}

\item{method}{method to create patterns ("gam", "dendogram" or "centroids")}

\item{times}{Number of partitions to create.}

\item{p}{the percentage of data that goes to training.}

\item{reference.lst}{a conversion of label names for the reference classes (optional)}

\item{predicted.lst}{a conversion of label names for the predicted classes (optional)}
}
\description{
Splits the set of time series into training and validation and
compute accuracy metrics. The function uses stratified sampling and a simple
random sampling for each stratum. For each data partition this function
performs a TWDTW analysis and returns the Overall Accuracy, User's Accuracy,
Produce's Accuracy, error matrix (confusion matrix), and a \code{\link[base]{data.frame}}
with the classification (Predicted), the reference classes (Reference),
and the results of the TWDTW analysis.
}
\author{
Victor Maus, \email{vwmaus1@gmail.com}

Gilberto Camara, \email{gilberto.camara@inpe.br}

Rolf Simoes, \email{rolf.simoes@inpe.br}
}
