% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_tibble_functions.R
\name{sits_values}
\alias{sits_values}
\title{Return the values of a given SITS table as a list of matrices according to a specified format.}
\usage{
sits_values(data.tb, bands = NULL, format = "cases_dates_bands")
}
\arguments{
\item{data.tb}{a tibble in SITS format with time series for different bands}

\item{bands}{string - a group of bands whose values are to be extracted. If no bands is informed extract ALL bands.}

\item{format}{string - either "cases_dates_bands" or "bands_cases_dates" or "bands_dates_cases"}
}
\value{
table   a tibble in SITS format with values
}
\description{
this function returns only the values of a sits table (according a specified format).
This function is useful to use packages such as ggplot, dtwclust, or kohonen that
require values that are rowwise or colwise organised.
}
\examples{
# Retrieve a set of time series with 2 classes
data (cerrado_2classes)
# retrieve the values split by bands
sits_values (cerrado_2classes[1:2,], format = "bands_dates_cases")

}
\author{
Rolf Simoes, \email{rolf.simoes@inpe.br}

Gilberto Camara, \email{gilberto.camara@inpe.br}
}
