% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_values.R
\name{sits_values}
\alias{sits_values}
\title{Return the values of one band of a SITS table}
\usage{
sits_values(data.tb, band)
}
\arguments{
\item{data.tb}{a tibble in SITS format with time series for different bands}

\item{band}{string - a band whose values are to be extracted}
}
\value{
table   a tibble in SITS format with values
}
\description{
\code{sits_value} returns a sits table with values only
}
\details{
A sits table has the metadata and data for each time series
<longitude, latitude, start_date, end_date, label, coverage, time_series>
}
