% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_table.R
\name{sits_values_cols}
\alias{sits_values_cols}
\title{Return the values of one band of a SITS table (colwise organised)}
\usage{
sits_values_cols(data.tb, band)
}
\arguments{
\item{data.tb}{a tibble in SITS format with time series for different bands}

\item{band}{string - a band whose values are to be extracted}
}
\value{
table   a tibble  with values
}
\description{
\code{sits_value_cols} returns a sits table with values only (colwise organised)
}
\details{
A sits table has the metadata and data for each time series
<longitude, latitude, start_date, end_date, label, coverage, time_series>
}
\seealso{
Other STIS table functions: \code{\link{sits_align}},
  \code{\link{sits_bands}}, \code{\link{sits_dates}},
  \code{\link{sits_group_bylatlong}},
  \code{\link{sits_rename}}, \code{\link{sits_select}},
  \code{\link{sits_table}}, \code{\link{sits_values_rows}}
}

