% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_view.R
\name{sits_view}
\alias{sits_view}
\alias{sits_view.sits}
\alias{sits_view.raster_cube}
\alias{sits_view.classified_image}
\title{View data cubes and samples in leaflet}
\usage{
sits_view(x, ...)

\method{sits_view}{sits}(x, ..., legend = NULL, palette = "Harmonic")

\method{sits_view}{raster_cube}(
  x,
  ...,
  band = NULL,
  red = NULL,
  green = NULL,
  blue = NULL,
  tiles = NULL,
  times = c(1),
  class_cube = NULL,
  legend = NULL,
  palette = "default"
)

\method{sits_view}{classified_image}(x, ..., tile, legend = NULL, palette = "default")
}
\arguments{
\item{x}{object of class "sits", "raster_cube" or "classified image"}

\item{...}{further specifications for \link{sits_view}.}

\item{legend}{named vector that associates labels to colors}

\item{palette}{palette provided in the configuration file}

\item{band}{for plotting grey images}

\item{red}{band for red color.}

\item{green}{band for green color.}

\item{blue}{band for blue color.}

\item{tiles}{tiles to be plotted in raster cubes
(in case of a multi-tile cube).}

\item{times}{temporal instances to be plotted.}

\item{class_cube}{classified cube to be overlayed on top on image}

\item{tile}{tile to be plotted in classified cubes.}
}
\value{
leaflet object
}
\description{
Uses leaflet to visualize time series, raster cube and classified images
}
\examples{
\donttest{
# view a collection of time series
sits_view(samples_modis_4bands)

# view a temporal instance of a cube

data_dir <- system.file("extdata/raster/mod13q1", package = "sits")

modis_cube <- sits_cube(
    source = "BDC",
    collection = "MOD13Q1-6",
    bands = c("NDVI", "EVI"),
    data_dir = data_dir,
    parse_info = c("X1", "X2", "tile", "band", "date")
)

# plot the data cube
sits_view(modis_cube, red = "EVI", green = "NDVI", blue = "EVI", times = 1)
}

}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
