% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_view.R
\name{sits_view}
\alias{sits_view}
\alias{sits_view.raster_cube}
\alias{sits_view.classified_image}
\title{Generic interface for visualization of data cube}
\usage{
sits_view(x, ...)

\method{sits_view}{raster_cube}(x, ..., red, green, blue, time = 1, roi = NULL)

\method{sits_view}{classified_image}(x, ..., map = NULL, time = 1, legend = NULL)
}
\arguments{
\item{x}{object of class "raster_cube" or "classified image"}

\item{...}{further specifications for \link{sits_view}.}

\item{red}{band for red color.}

\item{green}{band for green color.}

\item{blue}{band for blue color.}

\item{time}{temporal instances to be plotted.}

\item{roi}{sf object giving a region of interest.}

\item{map}{map to overlay (mapview object)}

\item{legend}{named vector that associates labels to colors}
}
\value{
mapview object
}
\description{
Uses mapview to visualize raster cube and classified images
}
\examples{
\donttest{
data_dir <- system.file("extdata/raster/cbers", package = "sits")

cbers_022024 <- sits_cube(
    source = "LOCAL",
    name = "cbers_022024",
    satellite = "CBERS-4",
    sensor = "AWFI",
    resolution = 64,
    data_dir = data_dir,
    parse_info = c("X1", "X2", "band", "date")
)
# plot the data cube
sits_view(cbers_022024, red = "B15", green = "B16", blue = "B13", time = 1)
}

}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
