% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_view.R
\name{sits_view}
\alias{sits_view}
\alias{sits_view.sits}
\alias{sits_view.som_map}
\alias{sits_view.raster_cube}
\alias{sits_view.class_cube}
\alias{sits_view.probs_cube}
\alias{sits_view.default}
\title{View data cubes and samples in leaflet}
\usage{
sits_view(x, ...)

\method{sits_view}{sits}(x, ..., legend = NULL, palette = "Harmonic")

\method{sits_view}{som_map}(
  x,
  ...,
  label,
  prob_max = 1,
  prob_min = 0.7,
  legend = NULL,
  palette = "Harmonic"
)

\method{sits_view}{raster_cube}(
  x,
  ...,
  band = NULL,
  red = NULL,
  green = NULL,
  blue = NULL,
  tiles = x$tile,
  dates = sits_timeline(x[1, ])[1],
  class_cube = NULL,
  legend = NULL,
  palette = "default"
)

\method{sits_view}{class_cube}(x, ..., tiles = NULL, legend = NULL, palette = "default")

\method{sits_view}{probs_cube}(x, ...)

\method{sits_view}{default}(x, ...)
}
\arguments{
\item{x}{Object of class "sits",
"raster_cube" or "classified image".}

\item{...}{Further specifications for \link{sits_view}.}

\item{legend}{Named vector that associates labels to colors.}

\item{palette}{Palette provided in the configuration file.}

\item{label}{Label from the SOM map to be shown.}

\item{prob_max}{Maximum a posteriori probability for SOM neuron
samples to be shown}

\item{prob_min}{Minimum a posteriori probability for SOM neuron
samples to be shown}

\item{band}{For plotting grey images.}

\item{red}{Band for red color.}

\item{green}{Band for green color.}

\item{blue}{Band for blue color.}

\item{tiles}{Tiles to be plotted (in case of a multi-tile cube).}

\item{dates}{Dates to be plotted.}

\item{class_cube}{Classified cube to be overlayed on top on image.}
}
\value{
A leaflet object containing either samples or
                      data cubes embedded in a global map that can
                      be visualized directly in an RStudio viewer.
}
\description{
Uses leaflet to visualize time series, raster cube and
classified images
}
\note{
Please refer to the sits documentation available in
<https://e-sensing.github.io/sitsbook/> for detailed examples.
}
\examples{
if (sits_run_examples()) {
    sits_view(cerrado_2classes)

    data_dir <- system.file("extdata/raster/mod13q1", package = "sits")

    modis_cube <- sits_cube(
        source = "BDC",
        collection = "MOD13Q1-6",
        data_dir = data_dir,
        parse_info = c("X1", "tile", "band", "date")
    )
    # view the data cube
    sits_view(modis_cube,
        band = "NDVI"
    )
    # train a model
    rf_model <- sits_train(samples_modis_ndvi, sits_rfor())

    modis_probs <- sits_classify(
        data = modis_cube,
        ml_model = rf_model,
        output_dir = tempdir(),
    )
    modis_label <- sits_label_classification(
        modis_probs, output_dir = tempdir()
    )

    sits_view(modis_label)

    sits_view(modis_cube,
        band = "NDVI",
        class_cube = modis_label,
        dates = sits_timeline(modis_cube)[[1]]
    )
}
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
