% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_filters.R
\name{sits_whittaker}
\alias{sits_whittaker}
\title{Smooth the time series using Whittaker smoother (based on PTW package)}
\usage{
sits_whittaker(data.tb, lambda = 1, differences = 3,
  bands_suffix = "whit")
}
\arguments{
\item{data.tb}{The SITS tibble containing the original time series}

\item{lambda}{double   - the smoothing factor to be applied (default 1.0)}

\item{differences}{an integer indicating the order of differences of contiguous elements (default 3)}

\item{bands_suffix}{the suffix to be appended to the smoothed filters (default "whit")}
}
\value{
output.tb a tibble with smoothed sits time series
}
\description{
The algorithm searches for an optimal polynomial describing the warping.
The degree of smoothing depends on smoothing factor lambda (usually from 0.5 to 10.0)
Use lambda = 0.5 for very slight smoothing and lambda = 5.0 for strong smoothing
}
