% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_source_api.R
\name{source_cube}
\alias{source_cube}
\title{Functions to instantiate a new cube from a source}
\arguments{
\item{source}{Data source.}

\item{...}{Additional parameters.}

\item{items}{Images that compose a cube.}

\item{asset}{A \code{raster} object to retrieve information.}

\item{collection}{Image collection.}

\item{data_dir}{Directory where local files are stored}

\item{file_info}{A \code{tibble} that organizes the metadata about each
file in the tile: date, band, resolution, and path (or URL).}

\item{bands}{Bands to be selected in the collection.}

\item{progress}{Show a progress bar?}
}
\value{
The values returned by each function are described as follows.
}
\description{
These functions provide an API to instantiate a new cube object and
access/retrieve information from services or local files to fill
cube attributes.

A cube is formed by images (items) organized in tiles. To create a sits
cube object (a \code{tibble}), a set of functions are called in order
to retrieve metadata.
}
\keyword{internal}
