% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_source_api.R
\name{source_item_get_functions}
\alias{source_item_get_functions}
\alias{.source_item_get_date}
\alias{.source_item_get_hrefs}
\alias{.source_item_get_bands}
\alias{.source_item_get_resolutions}
\title{...}
\usage{
.source_item_get_date(source, item, ..., collection = NULL)

.source_item_get_hrefs(source, item, ..., collection = NULL)

.source_item_get_bands(source, item, ..., collection = NULL)

.source_item_get_resolutions(source, item, ..., collection = NULL)
}
\arguments{
\item{source}{Name of the STAC provider}

\item{item}{...}

\item{...}{Aditional parameters}

\item{collection}{...}
}
\value{
...
}
\description{
TODO: add
}
\keyword{internal}
