% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_source_api.R
\name{source_items_get_functions}
\alias{source_items_get_functions}
\alias{.source_items_tiles_group}
\alias{.source_items_get_sensor}
\alias{.source_items_get_satellite}
\alias{.source_items_tile_get_crs}
\title{...}
\usage{
.source_items_tiles_group(source, items, ..., collection = NULL)

.source_items_get_sensor(source, items, ..., collection = NULL)

.source_items_get_satellite(source, items, ..., collection = NULL)

.source_items_tile_get_crs(source, tile_items, ..., collection = NULL)
}
\arguments{
\item{source}{Name of the STAC provider}

\item{items}{...}

\item{...}{Aditional parameters}

\item{collection}{...}
}
\value{
...
}
\description{
TODO: add
}
\keyword{internal}
