% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_source_api.R
\name{source_tile_items_get_functions}
\alias{source_tile_items_get_functions}
\alias{.source_items_tile_get_crs}
\alias{.source_items_tile_get_name}
\alias{.source_items_tile_get_bbox}
\alias{.source_items_tile_get_size}
\title{#' @title Function for retrieving information from multiples items
related to a single tile.}
\usage{
.source_items_tile_get_crs(source, tile_items, ..., collection = NULL)

.source_items_tile_get_name(source, tile_items, ..., collection = NULL)

.source_items_tile_get_bbox(source, tile_items, ..., collection = NULL)

.source_items_tile_get_size(source, tile_items, ..., collection = NULL)
}
\arguments{
\item{source}{Data source (one of "SATVEG", "LOCAL", "BDC", "AWS", "USGS"
"DEAFRICA", "PROBS").}

\item{...}{Additional parameters.}

\item{collection}{Collection to be searched in the data source.}

\item{items}{A \code{STACItemCollection} object returned by rstac or
\code{character} vector with directories to be search.}
}
\value{
an \code{atomic} vector with informations retrived from items.
}
\description{
Function to retrieves informations from the STACItemCollection
object in the rstac package, but related from a single tile.
}
\keyword{internal}
