% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_bands_set.R
\name{`sits_bands<-`}
\alias{`sits_bands<-`}
\alias{sits_bands<-}
\title{Replaces the names of the bands}
\usage{
sits_bands(x) <- value
}
\arguments{
\item{x}{Valid sits tibble (time series or a cube)}

\item{value}{Vector of bands}
}
\description{
Replaces the names of the bands
}
\examples{
# Replace the name of bands for the samples for Mato Grosso
sits_bands(samples_modis_4bands) <- c("ndvi", "evi", "nir", "mir")
}
