% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_data_generics.R
\name{`sits_bands<-`}
\alias{`sits_bands<-`}
\alias{sits_bands<-}
\title{Replaces the names of the bands}
\usage{
sits_bands(x) <- value
}
\arguments{
\item{x}{Valid sits tibble (time series or a cube)}

\item{value}{Vector of bands}
}
\value{
An updated data set with the new bands.
}
\description{
Replace the names of the bands of time series in a sits tibble
              or in a metadata cube
              For details see:
\itemize{
 \item{"time series": }{see \code{\link{sits_bands<-.sits}}}
 \item{"data cube": }{see \code{\link{sits_bands<-.cube}}}
}
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}

Rolf Simoes, \email{rolf.simoes@inpe.br}
}
