% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_labels.R
\name{`sits_labels<-`}
\alias{`sits_labels<-`}
\alias{sits_labels<-}
\title{Change the labels of a set of time series}
\usage{
sits_labels(data) <- value
}
\arguments{
\item{data}{Data cube or time series.}

\item{value}{A character vector used to convert labels. Labels will
be renamed to the respective value positioned at the
labels order returned by \code{\link{sits_labels}}.}
}
\value{
A sits tibble with modified labels.
}
\description{
Given a sits tibble with a set of labels, renames the labels
to the specified in value.
}
\author{
Rolf Simoes, \email{rolf.simoes@inpe.br}
}
