% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_labels_set.R
\name{`sits_labels<-`}
\alias{`sits_labels<-`}
\alias{sits_labels<-}
\title{Change labels of a sits tibble}
\usage{
sits_labels(data) <- value
}
\arguments{
\item{data}{A sits tibble.}

\item{value}{A character vector used to convert labels. Labels will
be renamed to the respective value positioned at the
labels order returned by \code{\link{sits_labels}}.}
}
\value{
A sits tibble with modified labels.
}
\description{
Given a sits tibble with a set of labels, renames the labels
to the specified in value.
}
\examples{
# Read a set of time series with information on deforestation
data("samples_modis_4bands")
# Print the labels
sits_labels(samples_modis_4bands)
# Create a conversion list.
# relabel the data
sits_labels(samples_modis_4bands) <- c("Natural", "Natural",
                                       "Anthropic", "Anthropic")
# show the new labels
sits_labels(samples_modis_4bands)

}
\author{
Rolf Simoes, \email{rolf.simoes@inpe.br}
}
