% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_labels.R
\name{`sits_labels<-`}
\alias{`sits_labels<-`}
\alias{sits_labels<-}
\alias{sits_labels<-.sits}
\alias{sits_labels<-.probs_cube}
\alias{sits_labels<-.class_cube}
\alias{sits_labels<-.default}
\title{Change the labels of a set of time series}
\usage{
sits_labels(data) <- value

\method{sits_labels}{sits}(data) <- value

\method{sits_labels}{probs_cube}(data) <- value

\method{sits_labels}{class_cube}(data) <- value

\method{sits_labels}{default}(data) <- value
}
\arguments{
\item{data}{Data cube or time series.}

\item{value}{A character vector used to convert labels. Labels will
be renamed to the respective value positioned at the
labels order returned by \code{\link{sits_labels}}.}
}
\value{
A sits tibble or data cube with modified labels.

A probs or class_cube cube with modified labels.
}
\description{
Given a sits tibble with a set of labels, renames the labels
to the specified in value.
}
\examples{
# show original samples ("Cerrado" and "Pasture")
sits_labels(cerrado_2classes)
# rename label samples to "Savanna" and "Grasslands"
sits_labels(cerrado_2classes) <- c("Savanna", "Grasslands")
# see the change
sits_labels(cerrado_2classes)
}
\author{
Rolf Simoes, \email{rolf.simoes@inpe.br}
}
